Function style::animation::start_transitions_if_applicable

pub fn start_transitions_if_applicable<Impl: SelectorImplExt>(new_animations_sender: &Mutex<Sender<Animation<Impl>>>, node: OpaqueNode, old_style: &Impl::ComputedValues, new_style: &mut Arc<Impl::ComputedValues>) -> bool

Inserts transitions into the queue of running animations as applicable for the given style difference. This is called from the layout worker threads. Returns true if any animations were kicked off and false otherwise.