Function style::animation::update_style_for_animation

pub fn update_style_for_animation<Damage, Impl>(context: &SharedStyleContext<Impl>, animation: &Animation<Impl>, style: &mut Arc<Damage::ConcreteComputedValues>, damage: Option<&mut Damage>) where Impl: SelectorImplExt, Damage: TRestyleDamage<ConcreteComputedValues=Impl::ComputedValues>

Updates a single animation and associated style based on the current time. If damage is provided, inserts the appropriate restyle damage.