Struct style::animation::PropertyAnimation

pub struct PropertyAnimation {
    property: AnimatedProperty,
    timing_function: TransitionTimingFunction,
    duration: Time,
}

Fields
property: AnimatedProperty timing_function: TransitionTimingFunction duration: Time
Methods

impl PropertyAnimation

fn from_transition(transition_index: usize, old_style: &ServoComputedValues, new_style: &mut ServoComputedValues) -> Vec<PropertyAnimation>

Creates a new property animation for the given transition index and old and new styles. Any number of animations may be returned, from zero (if the property did not animate) to one (for a single transition property) to arbitrarily many (for all).

fn from_transition_property(transition_property: TransitionProperty, transition_index: usize, old_style: &ServoComputedValues, new_style: &mut ServoComputedValues) -> Option<PropertyAnimation>

fn update(&self, style: &mut ServoComputedValues, time: f64)

fn does_not_animate(&self) -> bool