Struct style::applicable_declarations::ApplicableDeclarationBlock

pub struct ApplicableDeclarationBlock {
    pub source: StyleSource,
    pub specificity: u32,
    // some fields omitted
}

A property declaration together with its precedence among rules of equal specificity so that we can sort them.

This represents the declarations in a given declaration block for a given importance.

Fields
source: StyleSource

The style source, either a style rule, or a property declaration block.

specificity: u32

The specificity of the selector this block is represented by.

Methods

impl ApplicableDeclarationBlock

pub fn from_declarations(
declarations: Arc<Locked<PropertyDeclarationBlock>>,
level: CascadeLevel
) -> Self

Constructs an applicable declaration block from a given property declaration block and importance.

pub fn new(
source: StyleSource,
order: u32,
level: CascadeLevel,
specificity: u32
) -> Self

Constructs an applicable declaration block from the given components

pub fn source_order(&self) -> u32

Returns the source order of the block.

pub fn level(&self) -> CascadeLevel

Returns the cascade level of the block.

pub fn order_and_level(self) -> (StyleSource, CascadeLevel)

Convenience method to consume self and return the source alongside the level.