Enum style::attr::AttrValue

pub enum AttrValue {
    String(String),
    TokenList(String, Vec<Atom>),
    UInt(String, u32),
    Int(String, i32),
    Double(String, f64),
    Atom(Atom),
    Length(String, Option<Length>),
    Color(String, Option<RGBA>),
    Dimension(String, LengthOrPercentageOrAuto),
    Url(String, Option<ServoUrl>),
    Declaration(String, Arc<RwLock<PropertyDeclarationBlock>>),
}

Variants
String(String)TokenList(String, Vec<Atom>)UInt(String, u32)Int(String, i32)Double(String, f64)Atom(Atom)Length(String, Option<Length>)Color(String, Option<RGBA>)Dimension(String, LengthOrPercentageOrAuto)Url(String, Option<ServoUrl>)Declaration(String, Arc<RwLock<PropertyDeclarationBlock>>)

Note that this variant is only used transitively as a fast path to set the property declaration block relevant to the style of an element when set from the inline declaration of that element (that is, element.style).

This can, as of this writing, only correspond to the value of the style element, and is set from its relevant CSSInlineStyleDeclaration, and then converted to a string in Element::attribute_mutated.

Note that we don't necessarily need to do that (we could just clone the declaration block), but that avoids keeping a refcounted declarationblock for longer than needed.

Methods

impl AttrValue

fn from_serialized_tokenlist(tokens: String) -> AttrValue

fn from_comma_separated_tokenlist(tokens: String) -> AttrValue

fn from_u32(string: String, default: u32) -> AttrValue

fn from_i32(string: String, default: i32) -> AttrValue

fn from_double(string: String, default: f64) -> AttrValue

fn from_limited_i32(string: String, default: i32) -> AttrValue

fn from_limited_u32(string: String, default: u32) -> AttrValue

fn from_atomic(string: String) -> AttrValue

fn from_url(base: ServoUrl, url: String) -> AttrValue

fn from_legacy_color(string: String) -> AttrValue

fn from_dimension(string: String) -> AttrValue

fn from_nonzero_dimension(string: String) -> AttrValue

fn as_tokens(&self) -> &[Atom]

Assumes the AttrValue is a TokenList and returns its tokens

Panics

Panics if the AttrValue is not a TokenList

fn as_atom(&self) -> &Atom

Assumes the AttrValue is an Atom and returns its value

Panics

Panics if the AttrValue is not an Atom

fn as_color(&self) -> Option<&RGBA>

Assumes the AttrValue is a Color and returns its value

Panics

Panics if the AttrValue is not a Color

fn as_length(&self) -> Option<&Length>

Assumes the AttrValue is a Length and returns its value

Panics

Panics if the AttrValue is not a Length

fn as_dimension(&self) -> &LengthOrPercentageOrAuto

Assumes the AttrValue is a Dimension and returns its value

Panics

Panics if the AttrValue is not a Dimension

fn as_url(&self) -> Option<&ServoUrl>

Assumes the AttrValue is a Url and returns its value

Panics

Panics if the AttrValue is not a Url

fn as_uint(&self) -> u32

Return the AttrValue as its integer representation, if any. This corresponds to attribute values returned as AttrValue::UInt(_) by VirtualMethods::parse_plain_attribute().

Panics

Panics if the AttrValue is not a UInt