Struct style::cache::LRUCache

pub struct LRUCache<K, V> { /* fields omitted */ }

A LRU cache used to store a set of at most n elements at the same time.

Currently used for the style sharing candidate cache.

Methods

impl<K: PartialEq, V: Clone> LRUCache<K, V>

fn new(size: usize) -> LRUCache<K, V>

Create a new LRU cache with size elements at most.

fn touch(&mut self, pos: usize) -> &V

Touch a given position, and put it in the last item on the list.

fn iter(&self) -> Iter<(K, V)>

Iterate over the contents of this cache.

fn iter_mut(&mut self) -> IterMut<(K, V)>

Iterate mutably over the contents of this cache.

fn insert(&mut self, key: K, val: V)

Insert a given key and value in the cache.

fn find(&mut self, key: &K) -> Option<V>

Try to find a key in the cache.

fn find_or_create<F>(&mut self, key: K, blk: F) -> V where F: FnMut() -> V

Try to find a given key, or create a given item with that key executing blk.

fn evict_all(&mut self)

Evict all elements from the cache.