Struct style::cascade_info::CascadeInfo

pub struct CascadeInfo {
    pub saw_viewport_units: bool,
    // some fields omitted
}

A structure to collect information about the cascade.

This is useful to gather information about what an element is affected by, and can be used in the future to track optimisations like when a non-inherited property is explicitly inherited, in order to cut-off the traversal.

Fields
saw_viewport_units: bool

Whether we've seen viewport units so far.

Methods

impl CascadeInfo

fn new() -> Self

Construct a new CascadeInfo.

fn on_cascade_property<T>(&mut self,
_property_declaration: &PropertyDeclaration,
value: &DeclaredValue<T>) where T: HasViewportPercentage

Called when a property is cascaded.

NOTE: We can add a vast amount of information here.

fn finish<N: TNode>(self, node: &N)

Called when the cascade is finished, in order to use the information we've collected.

Currently used for styling to mark a node as needing restyling when the viewport size changes.