style::color

Enum ColorSpace

#[repr(u8)]
pub enum ColorSpace {

    Srgb = 0,
    Hsl = 1,
    Hwb = 2,
    Lab = 3,
    Lch = 4,
    Oklab = 5,
    Oklch = 6,
    SrgbLinear = 7,
    DisplayP3 = 8,
    DisplayP3Linear = 9,
    A98Rgb = 10,
    ProphotoRgb = 11,
    Rec2020 = 12,
    XyzD50 = 13,
    XyzD65 = 14,

}

A color space representation in the CSS specification.

https://drafts.csswg.org/css-color-4/#typedef-color-space

Variants

Srgb = 0

A color specified in the sRGB color space with either the rgb/rgba(..) functions or the newer color(srgb ..) function. If the color(..) function is used, the AS_COLOR_FUNCTION flag will be set. Examples: “color(srgb 0.691 0.139 0.259)”, “rgb(176, 35, 66)”

Hsl = 1

A color specified in the Hsl notation in the sRGB color space, e.g. “hsl(289.18 93.136% 65.531%)” https://drafts.csswg.org/css-color-4/#the-hsl-notation

Hwb = 2

A color specified in the Hwb notation in the sRGB color space, e.g. “hwb(740deg 20% 30%)” https://drafts.csswg.org/css-color-4/#the-hwb-notation

Lab = 3

A color specified in the Lab color format, e.g. “lab(29.2345% 39.3825 20.0664)”. https://w3c.github.io/csswg-drafts/css-color-4/#lab-colors

Lch = 4

A color specified in the Lch color format, e.g. “lch(29.2345% 44.2 27)”. https://w3c.github.io/csswg-drafts/css-color-4/#lch-colors

Oklab = 5

A color specified in the Oklab color format, e.g. “oklab(40.101% 0.1147 0.0453)”. https://w3c.github.io/csswg-drafts/css-color-4/#lab-colors

Oklch = 6

A color specified in the Oklch color format, e.g. “oklch(40.101% 0.12332 21.555)”. https://w3c.github.io/csswg-drafts/css-color-4/#lch-colors

SrgbLinear = 7

A color specified with the color(..) function and the “srgb-linear” color space, e.g. “color(srgb-linear 0.435 0.017 0.055)”.

DisplayP3 = 8

A color specified with the color(..) function and the “display-p3” color space, e.g. “color(display-p3 0.84 0.19 0.72)”.

DisplayP3Linear = 9

A color specified with the color(..) function and the “display-p3-linear” color space.

A98Rgb = 10

A color specified with the color(..) function and the “a98-rgb” color space, e.g. “color(a98-rgb 0.44091 0.49971 0.37408)”.

ProphotoRgb = 11

A color specified with the color(..) function and the “prophoto-rgb” color space, e.g. “color(prophoto-rgb 0.36589 0.41717 0.31333)”.

Rec2020 = 12

A color specified with the color(..) function and the “rec2020” color space, e.g. “color(rec2020 0.42210 0.47580 0.35605)”.

XyzD50 = 13

A color specified with the color(..) function and the “xyz-d50” color space, e.g. “color(xyz-d50 0.2005 0.14089 0.4472)”.

XyzD65 = 14

A color specified with the color(..) function and the “xyz-d65” or “xyz” color space, e.g. “color(xyz-d65 0.21661 0.14602 0.59452)”. NOTE: https://drafts.csswg.org/css-color-4/#resolving-color-function-values specifies that xyz is an alias for the xyz-d65 color space.

Implementations

impl ColorSpace

pub fn parse<'i, 't>(input: &mut Parser<'i, 't>) -> Result<Self, ParseError<'i>>

Parse this keyword.

pub fn from_ident(ident: &str) -> Result<Self, ()>

Parse this keyword from a string slice.

impl ColorSpace

pub fn is_rectangular(&self) -> bool

Returns whether this is a <rectangular-color-space>.

pub fn is_polar(&self) -> bool

Returns whether this is a <polar-color-space>.

pub fn is_rgb_or_xyz_like(&self) -> bool

Returns true if the color has RGB or XYZ components.

pub fn hue_index(&self) -> Option<usize>

Returns an index of the hue component in the color space, otherwise None.