style::color

Module mix

Color mixing/interpolation.

Structs

ColorInterpolationMethod

https://drafts.csswg.org/css-color-4/#color-interpolation-method

ColorMixItem

A color and its weight for use in a color mix.

InterpolatedAlpha 🔒

Enums

ComponentMixOutcome 🔒

What the outcome of each component should be in a mix result.

HueInterpolationMethod

A hue-interpolation-method as defined in 1.

Functions

adjust_hue 🔒

convert_for_mix 🔒

interpolate_alpha 🔒

interpolate_hue 🔒

interpolate_premultiplied 🔒

interpolate_premultiplied_component 🔒

mix_many

Mix N colors into one (left-to-right fold).

mix_with_weights 🔒

Mix two colors already in the interpolation color space.

normalize_hue 🔒