Struct style::computed_value_flags::ComputedValueFlags

#[repr(C)]
pub struct ComputedValueFlags {
    bits: u32,
}

Misc information about a given computed style.

All flags are currently inherited for text, pseudo elements, and anonymous boxes, see StyleBuilder::for_inheritance and its callsites. If we ever want to add some flags that shouldn’t inherit for them, we might want to add a function to handle this.

Fields
bits: u32
Implementations

impl ComputedValueFlags

pub const HAS_TEXT_DECORATION_LINES: Self = _

Whether the style or any of the ancestors has a text-decoration-line property that should get propagated to descendants.

text-decoration-line is a reset property, but gets propagated in the frame/box tree.

pub const SHOULD_SUPPRESS_LINEBREAK: Self = _

Whether line break inside should be suppressed.

If this flag is set, the line should not be broken inside, which means inlines act as if nowrap is set,
element is suppressed, and blocks are inlinized.

This bit is propagated to all children of line participants. It is currently used by ruby to make its content unbreakable.

pub const IS_TEXT_COMBINED: Self = _

A flag used to mark text that that has text-combine-upright.

This is used from Gecko’s layout engine.

pub const IS_RELEVANT_LINK_VISITED: Self = _

A flag used to mark styles under a relevant link that is also visited.

pub const IS_IN_PSEUDO_ELEMENT_SUBTREE: Self = _

A flag used to mark styles which are a pseudo-element or under one.

pub const SELF_OR_ANCESTOR_HAS_CONTAIN_STYLE: Self = _

A flag used to mark styles which have contain:style or under one.

pub const DISPLAY_DEPENDS_ON_INHERITED_STYLE: Self = _

Whether this style’s display property depends on our parent style.

This is important because it may affect our optimizations to avoid computing the style of pseudo-elements, given whether the pseudo-element is generated depends on the display value.

pub const CONTENT_DEPENDS_ON_INHERITED_STYLE: Self = _

Whether this style’s content depends on our parent style.

Important because of the same reason.

pub const INHERITS_RESET_STYLE: Self = _

Whether the child explicitly inherits any reset property.

pub const DEPENDS_ON_SELF_FONT_METRICS: Self = _

Whether any value on our style is font-metric-dependent on our primary font.

pub const DEPENDS_ON_INHERITED_FONT_METRICS: Self = _

Whether any value on our style is font-metric-dependent on the primary font of our parent.

pub const CAN_BE_FRAGMENTED: Self = _

Whether the style or any of the ancestors has a multicol style.

Only used in Servo.

pub const IS_ROOT_ELEMENT_STYLE: Self = _

Whether this style is the style of the document element.

pub const IS_IN_OPACITY_ZERO_SUBTREE: Self = _

Whether this element is inside an opacity: 0 subtree.

pub const HAS_AUTHOR_SPECIFIED_BORDER_BACKGROUND: Self = _

Whether there are author-specified rules for border-* properties (except border-image-*), background-color, or background-image.

TODO(emilio): Maybe do include border-image, see:

https://github.com/w3c/csswg-drafts/issues/4777#issuecomment-604424845

pub const HAS_AUTHOR_SPECIFIED_FONT_FAMILY: Self = _

Whether there are author-specified rules for font-family.

pub const HAS_AUTHOR_SPECIFIED_FONT_SYNTHESIS_WEIGHT: Self = _

Whether there are author-specified rules for font-synthesis-weight.

pub const HAS_AUTHOR_SPECIFIED_FONT_SYNTHESIS_STYLE: Self = _

Whether there are author-specified rules for font-synthesis-style.

pub const HAS_AUTHOR_SPECIFIED_LETTER_SPACING: Self = _

Whether there are author-specified rules for letter-spacing.

pub const HAS_AUTHOR_SPECIFIED_WORD_SPACING: Self = _

Whether there are author-specified rules for word-spacing.

pub const USES_VIEWPORT_UNITS: Self = _

Whether the style depends on viewport units.

pub const USES_VIEWPORT_UNITS_ON_CONTAINER_QUERIES: Self = _

Whether the style depends on viewport units on container queries.

This needs to be a separate flag from USES_VIEWPORT_UNITS because it causes us to re-match the style (rather than re-cascascading it, which is enough for other uses of viewport units).

pub const SELF_OR_ANCESTOR_HAS_SIZE_CONTAINER_TYPE: Self = _

A flag used to mark styles which have container-type of size or inline-size, or under one.

pub const CONSIDERED_RELATIVE_SELECTOR: Self = _

Whether the style evaluated any relative selector.

pub const ANCHORS_RELATIVE_SELECTOR: Self = _

Whether the style evaluated the matched element to be an anchor of a relative selector.

pub const USES_CONTAINER_UNITS: Self = _

Whether the style uses container query units, in which case the style depends on the container’s size and we can’t reuse it across cousins (without double-checking the container at least).

pub const HAS_AUTHOR_SPECIFIED_TEXT_COLOR: Self = _

Whether there are author-specific rules for text color.

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl ComputedValueFlags

fn inherited_flags() -> Self

Flags that are unconditionally propagated to descendants.

fn maybe_inherited_flags() -> Self

Flags that may be propagated to descendants.

fn cascade_input_flags() -> Self

Flags that are an input to the cascade.

pub fn inherited(self) -> Self

Returns the flags that are always propagated to descendants.

See StyleAdjuster::set_bits and StyleBuilder.

pub fn maybe_inherited(self) -> Self

Flags that are conditionally propagated to descendants, just to handle properly style invalidation.

pub fn for_cascade_inputs(self) -> Self

Flags that are an input to the cascade.