Enum style::context::SequentialTask

pub enum SequentialTask<E: TElement> {
    Unused(SendElement<E>),
}

A task to be run in sequential mode on the parent (non-worker) thread. This is used by the style system to queue up work which is not safe to do during the parallel traversal.

Variants
Unused(SendElement<E>)

Entry to avoid an unused type parameter error on servo.

Methods

impl<E: TElement> SequentialTask<E>

pub fn execute(self)

Executes this task.