Struct style::context::CurrentElementInfo

pub struct CurrentElementInfo {
    pub possibly_expired_animations: Vec<PropertyAnimation>,
    // some fields omitted
}

Information about the current element being processed. We group this together into a single struct within ThreadLocalStyleContext so that we can instantiate and destroy it easily at the beginning and end of element processing.

Fields
possibly_expired_animations: Vec<PropertyAnimation>

A Vec of possibly expired animations. Used only by Servo.