Struct style::context::SharedStyleContext

pub struct SharedStyleContext {
    pub stylist: Arc<Stylist>,
    pub running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub error_reporter: Box<ParseErrorReporter>,
    pub local_context_creation_data: Mutex<ThreadLocalStyleContextCreationInfo>,
    pub timer: Timer,
    pub quirks_mode: QuirksMode,
}

A shared style context.

There's exactly one of these during a given restyle traversal, and it's shared among the worker threads.

Fields
stylist: Arc<Stylist>

The CSS selector stylist.

running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

The animations that are currently running.

expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

The list of animations that have expired since the last style recalculation.

error_reporter: Box<ParseErrorReporter>

The CSS error reporter for all CSS loaded in this layout thread

local_context_creation_data: Mutex<ThreadLocalStyleContextCreationInfo>

Data needed to create the thread-local style context from the shared one.

timer: Timer

The current timer for transitions and animations. This is needed to test them.

quirks_mode: QuirksMode

The QuirksMode state which the document needs to be rendered with

Methods

impl SharedStyleContext

fn viewport_size(&self) -> Size2D<Au>

Return a suitable viewport size in order to be used for viewport units.