style

Module custom_properties

Support for custom properties for cascading variables.

Macros

lnf_int 🔒

lnf_int_variable 🔒

make_variable 🔒

Structs

AttrTaint

In CSS Values and Units, values produced by attr() are considered attr()-tainted, as are functions that contain an attr()-tainted value. Using an attr()-tainted value as or in a makes a declaration invalid at computed-value time. https://drafts.csswg.org/css-values-5/#attr-security

AttrTaintedRange

For a CSS string, the range, counted in bytes, that is attr()-tainted.

AttributeData

Data specific to an attr() call like type and namespace.

ComputedCustomProperties

A pair of separate CustomPropertiesMaps, split between custom properties that have the inherit flag set and those with the flag unset.

ComputedSubstitutionFunctions

A wrapper map that encapsulates both the custom properties and attributes for a given element.

CssEnvironment

The environment from which to get env function values.

EnvironmentVariable 🔒

NonCustomReferenceMap

A map from NonCustomReferenceMap to a T.

ReferenceFlags

Set of flags to references this custom property makes.

References

A struct holding information about the external references to that a custom property value may have.

Substitution 🔒

SubstitutionFunctionReference

A reference to a substitution function like env() / var() / attr().

SubstitutionResult

Result of var(), env(), and attr() substitution.

VariableFallback

The fallback of a particular value.

VariableValue

A value for a custom property is just a set of tokens.

Enums

AttributeType 🔒

SingleNonCustomReference

A non-custom reference that participates in cycle resolution. TODO(emilio): This should probably eventually become just PrioritaryPropertyId.

SubstitutionFunctionKind

Substitution function source: var, env, attr.

Statics

CHROME_ENVIRONMENT_VARIABLES 🔒

ENVIRONMENT_VARIABLES 🔒

Functions

compute_value 🔒

compute_variable_value 🔒

Given a potentially registered variable value turn it into a computed custom property value.

do_substitute_chunk 🔒

eval_gtk_csd_titlebar_radius 🔒

get_attr_value_for_cycle_resolution

Attribute values may reference other substitution functions we may need to process. See step 6: https://drafts.csswg.org/css-values-5/#attr-substitution

get_content_preferred_color_scheme 🔒

get_hairline 🔒

get_safearea_inset_bottom 🔒

get_safearea_inset_left 🔒

get_safearea_inset_right 🔒

get_safearea_inset_top 🔒

get_scrollbar_inline_size 🔒

handle_invalid_at_computed_value_time

See https://drafts.csswg.org/css-variables-2/#invalid-at-computed-value-time

parse_attr_type 🔒

Parse = type( ) | raw-string | number | . https://drafts.csswg.org/css-values-5/#attr-notation

parse_declaration_value 🔒

https://drafts.csswg.org/css-syntax-3/#typedef-declaration-value

parse_declaration_value_block 🔒

Like parse_declaration_value, but accept ! and ; since they are only invalid at the top level.

parse_name

Parse a custom property name.

quoted_css_string 🔒

reify_variable_value 🔒

reify_variable_value_range 🔒

Reify a slice of the CSS string into UnparsedSegment entries.

remove_and_insert_initial_value 🔒

Removes the named registered custom property and inserts its uncomputed initial value.

substitute

Replace var(), env(), and attr() functions, returning the resulting CSS string.

substitute_internal 🔒

Replace var(), env(), and attr() functions. Return Err(..) for invalid at computed time.

substitute_one_reference 🔒

substitute_references_if_needed_and_apply

Replace var(), env(), and attr() functions in a pre-existing variable value.

Type Aliases

ComputedValue

Both specified and computed values are VariableValues, the difference is whether var() functions are expanded.

EnvironmentEvaluator 🔒

Name

A custom property name is just an Atom.

SpecifiedValue

Both specified and computed values are VariableValues, the difference is whether var() functions are expanded.