style::custom_properties

Function substitute_all

fn substitute_all(
    custom_properties_map: &mut ComputedCustomProperties,
    deferred_properties_map: Option<&mut CustomPropertiesMap>,
    invalid_non_custom_properties: &mut LonghandIdSet,
    has_color_scheme: bool,
    seen: &PrecomputedHashSet<&Name>,
    references_from_non_custom_properties: &NonCustomReferenceMap<Vec<Name>>,
    stylist: &Stylist,
    computed_context: &Context<'_>,
)

Resolve all custom properties to either substituted, invalid, or unset (meaning we should use the inherited value).

It does cycle dependencies removal at the same time as substitution.