Function style::custom_properties::substitute_block

fn substitute_block<'i>(
    input: &mut Parser<'i, '_>, 
    position: &mut (SourcePosition, TokenSerializationType), 
    partial_computed_value: &mut ComputedValue, 
    custom_properties: &CustomPropertiesMap, 
    device: &Device
) -> Result<TokenSerializationType, ParseError<'i>>

Replace var() functions in an arbitrary bit of input.

If the variable has its initial value, the callback should return Err(()) and leave partial_computed_value unchanged.

Otherwise, it should push the value of the variable (with its own var() functions replaced) to partial_computed_value and return Ok(last_token_type of what was pushed)

Return Err(()) if input is invalid at computed-value time. or Ok(last_token_type that was pushed to partial_computed_value) otherwise.