style::custom_properties

Struct VariableValue

pub struct VariableValue {
    pub css: String,
    pub url_data: UrlExtraData,
    first_token_type: TokenSerializationType,
    last_token_type: TokenSerializationType,
    pub references: References,
}

A value for a custom property is just a set of tokens.

We preserve the original CSS for serialization, and also the variable references to other custom property names.

Fields
css: String

The raw CSS string.

url_data: UrlExtraData

The url data of the stylesheet where this value came from.

first_token_type: TokenSerializationTypelast_token_type: TokenSerializationTypereferences: References

var(), env(), attr() or non-custom property (e.g. through em) references.

Implementations

impl VariableValue

fn empty(url_data: &UrlExtraData) -> Self

pub fn new( css: String, url_data: &UrlExtraData, first_token_type: TokenSerializationType, last_token_type: TokenSerializationType, ) -> Self

Create a new custom property without parsing if the CSS is known to be valid and contain no references.

fn push<'i>( &mut self, css: &str, css_first_token_type: TokenSerializationType, css_last_token_type: TokenSerializationType, attr_taint: Option<&mut AttrTaint>, ) -> Result<(), ()>

pub fn parse<'i, 't>( input: &mut Parser<'i, 't>, namespaces: Option<&FxHashMap<Prefix, Namespace>>, url_data: &UrlExtraData, ) -> Result<Self, ParseError<'i>>

Parse a custom property value.

pub fn is_attr_tainted(&self) -> bool

Returns whether this value is tainted by attr().

fn integer(number: i32, url_data: &UrlExtraData) -> Self

Create VariableValue from an int.

fn ident(ident: &'static str, url_data: &UrlExtraData) -> Self

Create VariableValue from an int.

fn pixels(number: f32, url_data: &UrlExtraData) -> Self

Create VariableValue from a float amount of CSS pixels.

fn int_ms(number: i32, url_data: &UrlExtraData) -> Self

Create VariableValue from an integer amount of milliseconds.

fn int_pixels(number: i32, url_data: &UrlExtraData) -> Self

Create VariableValue from an integer amount of CSS pixels.

fn from_token(token: Token<'_>, url_data: &UrlExtraData) -> Self

pub fn css_text(&self) -> &str

Returns the raw CSS text from this VariableValue

pub fn has_references(&self) -> bool

Returns whether this variable value has any reference to the environment or other variables.

pub fn collect_attribute_references(&self, references: &mut AttrReferenceSet)

Returns the attribute references in this variable value, if any.