Type Alias style::custom_properties::CustomPropertiesMap

pub type CustomPropertiesMap = IndexMap<Name, Arc<VariableValue>, BuildHasherDefault<PrecomputedHasher>>;

A map from CSS variable names to CSS variable computed values, used for resolving.

A consistent ordering is required for CSSDeclaration objects in the DOM. CSSDeclarations expose property names as indexed properties, which need to be stable. So we keep an array of property names which order is determined on the order that they are added to the name-value map.

The variable values are guaranteed to not have references to other properties.

Aliased Type

struct CustomPropertiesMap {
    pub(crate) core: IndexMapCore<Atom<AtomStaticSet>, Arc<VariableValue>>,
    hash_builder: BuildHasherDefault<PrecomputedHasher>,
}

Fields
core: IndexMapCore<Atom<AtomStaticSet>, Arc<VariableValue>>hash_builder: BuildHasherDefault<PrecomputedHasher>