style

Module data

Per-node data used in style calculation.

Structs

EagerPseudoArray 🔒

EagerPseudoStyles

A lazily-allocated list of styles for eagerly-cascaded pseudo-elements.

ElementData

Style system data associated with an Element.

ElementDataFlags

Various flags stored on ElementData.

ElementDataMut

A read-only reference to ElementData.

ElementDataRef

A mutable reference to ElementData.

ElementDataWrapper

A struct that wraps ElementData, giving it the ability of doing thread-safety checks.

ElementStyles

The styles associated with a node, including the styles for any pseudo-elements.

Enums

RestyleKind

The kind of restyle that a single element should do.

ViewportUnitUsage

Information on how this element uses viewport units.

Constants

EMPTY_PSEUDO_ARRAY 🔒

Functions

needs_to_match_self 🔒

needs_to_recascade_self 🔒

Type Aliases

EagerPseudoArrayInner 🔒