Enum style::data::ElementData

pub enum ElementData {
    Initial(Option<ElementStyles>),
    Restyle(RestyleData),
    Persistent(ElementStyles),
}

Style system data associated with a node.

In Gecko, this hangs directly off a node, but is dropped when the frame takes ownership of the computed style data.

In Servo, this is embedded inside of layout data, which itself hangs directly off the node. Servo does not currently implement ownership transfer of the computed style data to the frame.

In both cases, it is wrapped inside an AtomicRefCell to ensure thread safety.

Variants
Initial(Option<ElementStyles>)Restyle(RestyleData)Persistent(ElementStyles)
Methods

impl ElementData

fn new(existing: Option<ElementStyles>) -> Self

fn is_initial(&self) -> bool

fn is_unstyled_initial(&self) -> bool

fn is_styled_initial(&self) -> bool

fn is_restyle(&self) -> bool

fn as_restyle(&self) -> Option<&RestyleData>

fn as_restyle_mut(&mut self) -> Option<&mut RestyleData>

fn is_persistent(&self) -> bool

fn restyle(&mut self) -> Option<&mut RestyleData>

Sets an element up for restyle, returning None for an unstyled element.

fn persist(&mut self)

Converts Initial and Restyle to Persistent. No-op for Persistent.

fn damage(&self) -> RestyleDamage

fn current_styles(&self) -> &ElementStyles

fn get_current_styles(&self) -> Option<&ElementStyles>

fn current_styles_mut(&mut self) -> &mut ElementStyles

fn previous_styles(&self) -> Option<&ElementStyles>

fn previous_styles_mut(&mut self) -> Option<&mut ElementStyles>

fn current_or_previous_styles(&self) -> &ElementStyles

fn finish_styling(&mut self, styles: ElementStyles, damage: RestyleDamage)