Enum style::data::ElementDataStyles

enum ElementDataStyles {
    Uninitialized,
    Previous(Option<ElementStyles>),
    Current(ElementStyles),
}

Variants
Uninitialized

The field has not been initialized.

Previous(Option<ElementStyles>)

The field holds the previous style of the node. If this is None, the node has not been previously styled.

This is the input to the styling algorithm. It would ideally be immutable, but for now we need to mutate it a bit before styling to handle animations.

Note that since ElementStyles contains an Arc, the null pointer optimization prevents the Option<> here from consuming an extra word.

Current(ElementStyles)

The field holds the current, up-to-date style.

This is the output of the styling algorithm.

Methods

impl ElementDataStyles

fn is_previous(&self) -> bool