Enum style::data::RestyleKind

pub enum RestyleKind {
    MatchAndCascade,
    CascadeWithReplacements(RestyleHint),
    CascadeOnly,
}

The kind of restyle that a single element should do.

Variants
MatchAndCascade

We need to run selector matching plus re-cascade, that is, a full restyle.

CascadeWithReplacements(RestyleHint)

We need to recascade with some replacement rule, such as the style attribute, or animation rules.

CascadeOnly

We only need to recascade, for example, because only inherited properties in the parent changed.