Struct style::data::RestyleData

pub struct RestyleData {
    pub hint: StoredRestyleHint,
    pub recascade: bool,
    pub damage: RestyleDamage,
    pub snapshot: SnapshotOption,
}

Transient data used by the restyle algorithm. This structure is instantiated either before or during restyle traversal, and is cleared at the end of node processing.

Fields
hint: StoredRestyleHint

The restyle hint, which indicates whether selectors need to be rematched for this element, its children, and its descendants.

recascade: bool

Whether we need to recascade. FIXME(bholley): This should eventually become more fine-grained.

damage: RestyleDamage

The restyle damage, indicating what kind of layout changes are required afte restyling.

snapshot: SnapshotOption

An optional snapshot of the original state and attributes of the element, from which we may compute additional restyle hints at traversal time.

Methods

impl RestyleData

fn expand_snapshot<E: TElement>(&mut self,
element: E,
stylist: &Stylist)
-> bool

Expands the snapshot (if any) into a restyle hint. Returns true if later siblings must be restyled.

fn has_invalidations(&self) -> bool

Returns true if this RestyleData might invalidate the current style.