Struct style::error_reporting::RustLogReporter

pub struct RustLogReporter;

An error reporter that uses the log crate at info level.

This logging is silent by default, and can be enabled with a RUST_LOG=style=info environment variable. (See env_logger.)