style

Module font_face

The @font-face at-rule.

Re-exports

pub use crate::properties::font_face::DescriptorId;

pub use crate::properties::font_face::DescriptorParser;

pub use crate::properties::font_face::Descriptors;

pub use crate::values::computed::font::FamilyName;

pub use crate::values::computed::font::FontStretch;

pub use crate::values::computed::font::FontStyle;

pub use crate::values::specified::font::AbsoluteFontWeight;

pub use crate::values::specified::font::FontFeatureSettings;

pub use crate::values::specified::font::FontLanguageOverride;

pub use crate::values::specified::font::FontStretch as SpecifiedFontStretch;

pub use crate::values::specified::font::FontVariationSettings;

pub use crate::values::specified::font::MetricsOverride;

pub use crate::values::specified::font::SpecifiedFontStyle;

Macros

impl_range 🔒

Structs

ComputedFontStretchRange

The computed representation of the above, so that Gecko and Servo can read them easily.

ComputedFontStyleRange

The computed representation of the above, with angles in degrees stored as signed 8.8 fixed-point values, so that Gecko and Servo can read them easily.

ComputedFontWeightRange

The computed representation of the above so Gecko and Servo can read them easily.

FontFaceRule

https://drafts.csswg.org/css-fonts/#font-face-rule

FontFaceSourceTechFlags

Flags for the @font-face tech() function, indicating font technologies required by the resource.

FontStretchRange

The font-stretch descriptor:

FontWeightRange

The font-weight descriptor:

SourceList

A list of sources for the font-face src descriptor.

UrlSource

A UrlSource represents a font-face source that has been specified with a url() function.

Enums

FontDisplay

A font-display value for a @font-face rule. The font-display descriptor determines how a font face is displayed based on whether and when it is downloaded and ready to use.

FontFaceSourceFormat

FontFaceSourceFormatKeyword

Keywords for the font-face src descriptor’s format() function. (‘None’ and ‘Unknown’ are for internal use in gfx, not exposed to CSS.)

FontStyleRange

The font-style descriptor:

Source

A source for a font-face rule.

Functions

parse_font_face_block

Parse the block inside a @font-face rule.

sort_range 🔒