Struct style::font_face::FontFaceRule

pub struct FontFaceRule {
    pub family: FamilyName,
    pub sources: Vec<Source>,
    pub style: T,
    pub weight: T,
    pub stretch: T,
    pub unicode_range: Vec<Range>,
}

A @font-face rule.

https://drafts.csswg.org/css-fonts/#font-face-rule

Fields
family: FamilyName

The name of this font face

sources: Vec<Source>

The alternative sources for this font face.

style: T

The style of this font face

weight: T

The weight of this font face

stretch: T

The stretch of this font face

unicode_range: Vec<Range>

The ranges of code points outside of which this font face should not be used.

Methods

impl FontFaceRule

fn effective_sources(&self) -> EffectiveSources

Returns the list of effective sources for that font-face, that is the sources which don't list any format hint, or the ones which list at least "truetype" or "opentype".