Struct style::font_face::FontFaceRuleData

pub struct FontFaceRuleData {

    pub family: Option<FamilyName>,
    pub sources: Option<SourceList>,
    pub style: Option<FontStyle>,
    pub weight: Option<FontWeightRange>,
    pub stretch: Option<FontStretchRange>,
    pub display: Option<FontDisplay>,
    pub unicode_range: Option<Vec<UnicodeRange>>,
    pub feature_settings: Option<FontFeatureSettings>,
    pub variation_settings: Option<FontVariationSettings>,
    pub language_override: Option<SpecifiedValue>,
    pub ascent_override: Option<MetricsOverride>,
    pub descent_override: Option<MetricsOverride>,
    pub line_gap_override: Option<MetricsOverride>,
    pub size_adjust: Option<NonNegativePercentage>,
    pub source_location: SourceLocation,

}

Data inside a @font-face rule.

https://drafts.csswg.org/css-fonts/#font-face-rule

Fields
family: Option<FamilyName>

The name of this font face

sources: Option<SourceList>

The alternative sources for this font face.

style: Option<FontStyle>

The style of this font face.

weight: Option<FontWeightRange>

The weight of this font face.

stretch: Option<FontStretchRange>

The stretch of this font face.

display: Option<FontDisplay>

The display of this font face.

unicode_range: Option<Vec<UnicodeRange>>

The ranges of code points outside of which this font face should not be used.

feature_settings: Option<FontFeatureSettings>

The feature settings of this font face.

variation_settings: Option<FontVariationSettings>

The variation settings of this font face.

language_override: Option<SpecifiedValue>

The language override of this font face.

ascent_override: Option<MetricsOverride>

The ascent override for this font face.

descent_override: Option<MetricsOverride>

The descent override for this font face.

line_gap_override: Option<MetricsOverride>

The line-gap override for this font face.

size_adjust: Option<NonNegativePercentage>

The size adjustment for this font face.

source_location: SourceLocation

Line and column of the @font-face rule source code.

Implementations

impl FontFaceRuleData

pub fn empty(location: SourceLocation) -> Self

Create an empty font-face rule

pub fn decl_to_css(&self, dest: &mut CssStringWriter) -> Result

Serialization of declarations in the FontFaceRule

impl FontFaceRuleData

pub fn font_face(&self) -> Option<FontFace<'_>>

Per https://github.com/w3c/csswg-drafts/issues/1133 an @font-face rule is valid as far as the CSS parser is concerned even if it doesn’t have a font-family or src declaration.

However both are required for the rule to represent an actual font face.