Module style::font_metrics

Access to font metrics from the style system.

Structs

FontMetrics
Represents the font metrics that style needs from a font to compute the value of certain CSS units like ex.

ServoMetricsProvider
Dummy metrics provider for Servo. Knows nothing about fonts and does not provide any metrics.

Enums

FontMetricsOrientation
Type of font metrics to retrieve.

Traits

FontMetricsProvider
A trait used to represent something capable of providing us font metrics.

Functions

get_metrics_provider_for_product
Construct a font metrics provider for the current product