Trait style::font_metrics::FontMetricsProvider

pub trait FontMetricsProvider {
    fn get_size(
        &self, 
        font_name: &Atom, 
        font_family: GenericFontFamily
    ) -> Length;

    fn create_from(context: &SharedStyleContext) -> Self
    where
        Self: Sized;

    fn query(
        &self, 
        _context: &Context, 
        _base_size: FontBaseSize, 
        _orientation: FontMetricsOrientation
    ) -> FontMetrics { ... }
}

A trait used to represent something capable of providing us font metrics.

Required methods

fn get_size(&self, font_name: &Atom, font_family: GenericFontFamily) -> Length

Get default size of a given language and generic family.

fn create_from(context: &SharedStyleContext) -> Self where
Self: Sized,

Construct from a shared style context

Loading content...
Provided methods

fn query(
&self,
_context: &Context,
_base_size: FontBaseSize,
_orientation: FontMetricsOrientation
) -> FontMetrics

Obtain the metrics for given font family.

Loading content...