Module style::gecko

Gecko-specific style-system bits.

Modules

conversions
This module contains conversion helpers between Servo and Gecko types Ideally, it would be in geckolib itself, but coherence forces us to keep the traits and implementations here

data
Data needed to style a Gecko document.

media_queries
Gecko's media-query device and expression representation.

restyle_damage
Gecko's restyle damage computation (aka change hints, aka nsChangeHint).

selector_parser
Gecko-specific bits for selector-parsing.

snapshot
A gecko snapshot, that stores the element attributes and state before they change in order to properly calculate restyle hints.

snapshot_helpers
Element an snapshot common logic.

traversal
Gecko-specific bits for the styling DOM traversal.

values
Different kind of helpers to interact with Gecko values.

wrapper
Wrapper definitions on top of Gecko types in order to be used in the style system.