Struct style::gecko::data::PerDocumentStyleDataImpl

pub struct PerDocumentStyleDataImpl {
    pub stylist: Arc<Stylist>,
    pub stylesheets: Vec<Arc<Stylesheet>>,
    pub stylesheets_changed: bool,
    pub new_animations_sender: Sender<Animation>,
    pub new_animations_receiver: Receiver<Animation>,
    pub running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub work_queue: Option<ThreadPool>,
    pub num_threads: usize,
}

The container for data that a Servo-backed Gecko document needs to style itself.

Fields
stylist: Arc<Stylist>

Rule processor.

stylesheets: Vec<Arc<Stylesheet>>

List of stylesheets, mirrored from Gecko.

stylesheets_changed: bool

Whether the stylesheets list above has changed since the last restyle.

new_animations_sender: Sender<Animation>

Unused. Will go away when we actually implement transitions and animations properly.

new_animations_receiver: Receiver<Animation>

Unused. Will go away when we actually implement transitions and animations properly.

running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

Unused. Will go away when we actually implement transitions and animations properly.

expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

Unused. Will go away when we actually implement transitions and animations properly.

work_queue: Option<ThreadPool>

The worker thread pool. FIXME(bholley): This shouldn't be per-document.

num_threads: usize

The number of threads of the work queue.

Methods

impl PerDocumentStyleDataImpl

fn reset_device(&mut self)

Reset the device state because it may have changed.

Implies also a stylesheet flush.

fn flush_stylesheets(&mut self)

Recreate the style data if the stylesheets have changed.

fn default_computed_values(&self) -> &Arc<ComputedValues>

Get the default computed values for this document.