Module style::gecko::media_queries

Gecko's media-query device and expression representation.

Structs

Device
The Device in Gecko wraps a pres context, has a default values computed, and contains all the viewport rule state.

Expression
A expression for gecko contains a reference to the media feature, the value the media query contained, and the range to evaluate.

Enums

MediaExpressionValue
A value found or expected in a media expression.

Resolution
A resolution.