Enum style::gecko::media_queries::MediaExpressionValue

pub enum MediaExpressionValue {
    Length(Length),
    Integer(u32),
    Float(CSSFloat),
    BoolInteger(bool),
    IntRatio(u32, u32),
    Resolution(Resolution),
    Enumerated(u32),
    Ident(String),
}

A value found or expected in a media expression.

Variants
Length(Length)

A length.

Integer(u32)

A (non-negative) integer.

Float(CSSFloat)

A floating point value.

BoolInteger(bool)

A boolean value, specified as an integer (i.e., either 0 or 1).

IntRatio(u32, u32)

Two integers separated by '/', with optional whitespace on either side of the '/'.

Resolution(Resolution)

A resolution.

Enumerated(u32)

An enumerated value, defined by the variant keyword table in the feature's mData member.

Ident(String)

An identifier.

TODO(emilio): Maybe atomize?