Struct style::gecko::media_queries::Device

pub struct Device { /* fields omitted */ }

The Device in Gecko wraps a pres context, has a default values computed, and contains all the viewport rule state.

Methods

impl Device

fn new(pres_context: RawGeckoPresContextOwned) -> Self

Trivially constructs a new Device.

fn account_for_viewport_rule(&mut self, constraints: &ViewportConstraints)

Tells the device that a new viewport rule has been found, and stores the relevant viewport constraints.

fn default_values(&self) -> &ComputedValues

Returns the default computed values as a reference, in order to match Servo.

fn default_values_arc(&self) -> &Arc<ComputedValues>

Returns the default computed values as an Arc, in order to avoid clones.

fn reset(&mut self)

Recreates all the temporary state that the Device stores.

This includes the viewport override from @viewport rules, and also the default computed values.

fn media_type(&self) -> MediaType

Returns the current media type of the device.

fn au_viewport_size(&self) -> Size2D<Au>

Returns the current viewport size in app units.