Struct style::gecko::restyle_damage::GeckoRestyleDamage

pub struct GeckoRestyleDamage(_);

The representation of Gecko's restyle damage is just a wrapper over nsChangeHint.

Methods

impl GeckoRestyleDamage

fn new(raw: nsChangeHint) -> Self

Trivially construct a new GeckoRestyleDamage.

fn as_change_hint(&self) -> nsChangeHint

Get the inner change hint for this damage.

fn empty() -> Self

Get an empty change hint, that is (nsChangeHint(0)).

fn is_empty(&self) -> bool

Returns whether this restyle damage represents the empty damage.

fn compute(source: &nsStyleContext, new_style: &Arc<ComputedValues>) -> Self

Computes a change hint given an old style (in the form of a nsStyleContext, and a new style (in the form of ComputedValues).

Note that we could in theory just get two ComputedValues here and diff them, but Gecko has an interesting optimization when they mark accessed structs, so they effectively only diff structs that have ever been accessed from layout.

fn rebuild_and_reflow() -> Self

Get a restyle damage that represents the maximum action to be taken (rebuild and reflow).