Struct style::gecko::selector_parser::SelectorImpl

pub struct SelectorImpl;

The dummy struct we use to implement our selector parsing.

Methods

impl SelectorImpl

fn pseudo_element_cascade_type(pseudo: &PseudoElement)
-> PseudoElementCascadeType

Returns the kind of cascade type that a given pseudo is going to use.

In Gecko we only compute ::before and ::after eagerly. We save the rules for anonymous boxes separately, so we resolve them as precomputed pseudos.

We resolve the others lazily, see Servo_ResolvePseudoStyle.

fn each_pseudo_element<F>(fun: F) where F: FnMut(PseudoElement)

Executes a function for each pseudo-element.

fn pseudo_is_before_or_after(pseudo: &PseudoElement) -> bool

Returns whether the given pseudo-element is ::before or ::after.

fn pseudo_class_state_flag(pc: &NonTSPseudoClass) -> ElementState

Returns the relevant state flag for a given non-tree-structural pseudo-class.

impl SelectorImpl

fn each_eagerly_cascaded_pseudo_element<F>(fun: F) where F: FnMut(PseudoElement)

A helper to traverse each eagerly cascaded pseudo-element, executing fun on it.

TODO(emilio): We can optimize this for Gecko using the pseudo-element macro, and we should consider doing that for Servo too.

fn each_precomputed_pseudo_element<F>(fun: F) where F: FnMut(PseudoElement)

A helper to traverse each precomputed pseudo-element, executing fun on it.

The optimization comment in each_eagerly_cascaded_pseudo_element also applies here.