Module style::gecko::snapshot_helpers

Element an snapshot common logic.

Functions

each_class
Given an item, a callback, and a getter, execute callback for each class this item has.

has_class
Given an item T, a class name, and a getter function, return whether that element has the class that name represents.

Type Definitions

ClassOrClassList
A function that, given an element of type T, allows you to get a single class or a class list.