Module style::gecko::values

Different kind of helpers to interact with Gecko values.

Traits

GeckoStyleCoordConvertible
A trait that defines an interface to convert from and to nsStyleCoords.

Functions

convert_nscolor_to_rgba
Convert a given nscolor to a Servo RGBA value.

convert_rgba_to_nscolor
Convert a given RGBA value to nscolor.

round_border_to_device_pixels
Round width down to the nearest device pixel, but any non-zero value that would round down to zero is clamped to 1 device pixel. Used for storing computed values of border-*-width and outline-width.