Trait style::gecko::values::GeckoStyleCoordConvertible

pub trait GeckoStyleCoordConvertible: Sized {
    fn to_gecko_style_coord<T: CoordDataMut>(&self, coord: &mut T);
    fn from_gecko_style_coord<T: CoordData>(coord: &T) -> Option<Self>;
}

A trait that defines an interface to convert from and to nsStyleCoords.

Required Methods

fn to_gecko_style_coord<T: CoordDataMut>(&self, coord: &mut T)

Convert this to a nsStyleCoord.

fn from_gecko_style_coord<T: CoordData>(coord: &T) -> Option<Self>

Given a nsStyleCoord, try to get a value of this type..