Module style::gecko::wrapper

Wrapper definitions on top of Gecko types in order to be used in the style system.

This really follows the Servo pattern in components/script/layout_wrapper.rs.

This theoretically should live in its own crate, but now it lives in the style system it's kind of pointless in the Stylo case, and only Servo forces the separation between the style system implementation and everything else.

Structs

DUMMY_BASE_URL
A dummy base url in order to get it where we don't have any available.

GeckoElement
A simple wrapper over a non-null Gecko Element pointer.

GeckoNode
A simple wrapper over a non-null Gecko node (nsINode) pointer.

Enums

GeckoChildrenIterator
A wrapper on top of two kind of iterators, depending on the parent being iterated.

Traits

AttrSelectorHelpers
A few helpers to help with attribute selectors and snapshotting.