Struct style::gecko::wrapper::DUMMY_BASE_URL

pub struct DUMMY_BASE_URL { /* fields omitted */ }

A dummy base url in order to get it where we don't have any available.

We need to get rid of this sooner than later.