Struct style::gecko::wrapper::GeckoElement

pub struct GeckoElement<'le>(pub &'le RawGeckoElement);

A simple wrapper over a non-null Gecko Element pointer.

Methods

impl<'le> GeckoElement<'le>

fn parse_style_attribute(value: &str) -> PropertyDeclarationBlock

Parse the style attribute of an element.

fn clear_data(&self)

Clear the element data for a given element.

unsafe fn ensure_data(&self) -> &AtomicRefCell<ElementData>

Ensures the element has data, returning the existing data or allocating it.

Only safe to call with exclusive access to the element, given otherwise it could race to allocate and leak.

fn create_snapshot(&self) -> Snapshot

Creates a blank snapshot for this element.