Struct style::gecko::wrapper::GeckoNode

pub struct GeckoNode<'ln>(pub &'ln RawGeckoNode);

A simple wrapper over a non-null Gecko node (nsINode) pointer.

Important: We don't currently refcount the DOM, because the wrapper lifetime magic guarantees that our LayoutFoo references won't outlive the root, and we don't mutate any of the references on the Gecko side during restyle.

We could implement refcounting if need be (at a potentially non-trivial performance cost) by implementing Drop and making LayoutFoo non-Copy.