Module style::gecko_bindings::bindings

Structs

RawServoImportRule

RawServoStyleRule

RawServoStyleSet

RawServoStyleSheet

ServoComputedValues

ServoCssRules

StyleChildrenIterator

nsACString
This type is the abstract type which is used for interacting with strings in rust. Each string type can derefence to an instance of this type, which provides the useful operations on strings.

nsAString
This type is the abstract type which is used for interacting with strings in rust. Each string type can derefence to an instance of this type, which provides the useful operations on strings.

Functions

Gecko_AddRefAtom

Gecko_AddRefCSSShadowArrayArbitraryThread

Gecko_AddRefCSSValueSharedListArbitraryThread

Gecko_AddRefCalcArbitraryThread

Gecko_AddRefPrincipalArbitraryThread

Gecko_AddRefQuoteValuesArbitraryThread

Gecko_AddRefURIArbitraryThread

Gecko_AnimationAppendKeyframe

Gecko_AtomAttrValue

Gecko_AtomEqualsUTF8

Gecko_AtomEqualsUTF8IgnoreCase

Gecko_Atomize

Gecko_AttrDashEquals

Gecko_AttrEquals

Gecko_AttrHasPrefix

Gecko_AttrHasSubstring

Gecko_AttrHasSuffix

Gecko_AttrIncludes

Gecko_CSSValue_Drop

Gecko_CSSValue_GetAbsoluteLength

Gecko_CSSValue_GetAngle

Gecko_CSSValue_GetArrayItem

Gecko_CSSValue_GetArrayItemConst

Gecko_CSSValue_GetCalc

Gecko_CSSValue_GetKeyword

Gecko_CSSValue_GetNumber

Gecko_CSSValue_GetPercentage

Gecko_CSSValue_SetAbsoluteLength

Gecko_CSSValue_SetAngle

Gecko_CSSValue_SetCalc

Gecko_CSSValue_SetFunction

Gecko_CSSValue_SetKeyword

Gecko_CSSValue_SetNumber

Gecko_CSSValue_SetPercentage

Gecko_CalcStyleDifference

Gecko_ChildrenCount

Gecko_ClassOrClassList

Gecko_ClearAndResizeStyleContents

Gecko_ClearPODTArray

Gecko_Construct_Default_nsStyleBackground

Gecko_Construct_Default_nsStyleBorder

Gecko_Construct_Default_nsStyleColor

Gecko_Construct_Default_nsStyleColumn

Gecko_Construct_Default_nsStyleContent

Gecko_Construct_Default_nsStyleDisplay

Gecko_Construct_Default_nsStyleEffects

Gecko_Construct_Default_nsStyleFont

Gecko_Construct_Default_nsStyleList

Gecko_Construct_Default_nsStyleMargin

Gecko_Construct_Default_nsStyleOutline

Gecko_Construct_Default_nsStylePadding

Gecko_Construct_Default_nsStylePosition

Gecko_Construct_Default_nsStyleSVG

Gecko_Construct_Default_nsStyleSVGReset

Gecko_Construct_Default_nsStyleTable

Gecko_Construct_Default_nsStyleTableBorder

Gecko_Construct_Default_nsStyleText

Gecko_Construct_Default_nsStyleTextReset

Gecko_Construct_Default_nsStyleUIReset

Gecko_Construct_Default_nsStyleUserInterface

Gecko_Construct_Default_nsStyleVariables

Gecko_Construct_Default_nsStyleVisibility

Gecko_Construct_Default_nsStyleXUL

Gecko_Construct_nsStyleVariables

Gecko_CopyClipPathValueFrom

Gecko_CopyConstruct_nsStyleBackground

Gecko_CopyConstruct_nsStyleBorder

Gecko_CopyConstruct_nsStyleColor

Gecko_CopyConstruct_nsStyleColumn

Gecko_CopyConstruct_nsStyleContent

Gecko_CopyConstruct_nsStyleDisplay

Gecko_CopyConstruct_nsStyleEffects

Gecko_CopyConstruct_nsStyleFont

Gecko_CopyConstruct_nsStyleList

Gecko_CopyConstruct_nsStyleMargin

Gecko_CopyConstruct_nsStyleOutline

Gecko_CopyConstruct_nsStylePadding

Gecko_CopyConstruct_nsStylePosition

Gecko_CopyConstruct_nsStyleSVG

Gecko_CopyConstruct_nsStyleSVGReset

Gecko_CopyConstruct_nsStyleTable

Gecko_CopyConstruct_nsStyleTableBorder

Gecko_CopyConstruct_nsStyleText

Gecko_CopyConstruct_nsStyleTextReset

Gecko_CopyConstruct_nsStyleUIReset

Gecko_CopyConstruct_nsStyleUserInterface

Gecko_CopyConstruct_nsStyleVariables

Gecko_CopyConstruct_nsStyleVisibility

Gecko_CopyConstruct_nsStyleXUL

Gecko_CopyCursorArrayFrom

Gecko_CopyFiltersFrom

Gecko_CopyFontFamilyFrom

Gecko_CopyImageValueFrom

Gecko_CopyListStyleImageFrom

Gecko_CopyListStyleTypeFrom

Gecko_CopyMozBindingFrom

Gecko_CopyStyleContentsFrom

Gecko_CreateElementSnapshot

Gecko_CreateGradient

Gecko_DestroyClipPath

Gecko_Destroy_nsStyleBackground

Gecko_Destroy_nsStyleBorder

Gecko_Destroy_nsStyleColor

Gecko_Destroy_nsStyleColumn

Gecko_Destroy_nsStyleContent

Gecko_Destroy_nsStyleDisplay

Gecko_Destroy_nsStyleEffects

Gecko_Destroy_nsStyleFont

Gecko_Destroy_nsStyleList

Gecko_Destroy_nsStyleMargin

Gecko_Destroy_nsStyleOutline

Gecko_Destroy_nsStylePadding

Gecko_Destroy_nsStylePosition

Gecko_Destroy_nsStyleSVG

Gecko_Destroy_nsStyleSVGReset

Gecko_Destroy_nsStyleTable

Gecko_Destroy_nsStyleTableBorder

Gecko_Destroy_nsStyleText

Gecko_Destroy_nsStyleTextReset

Gecko_Destroy_nsStyleUIReset

Gecko_Destroy_nsStyleUserInterface

Gecko_Destroy_nsStyleVariables

Gecko_Destroy_nsStyleVisibility

Gecko_Destroy_nsStyleXUL

Gecko_DropElementSnapshot

Gecko_DropStyleChildrenIterator

Gecko_ElementState

Gecko_EnsureImageLayersLength

Gecko_EnsureStyleAnimationArrayLength

Gecko_EnsureTArrayCapacity

Gecko_FillAllBackgroundLists

Gecko_FillAllMaskLists

Gecko_FlattenedTreeParentIsParent

Gecko_FontFamilyList_AppendGeneric

Gecko_FontFamilyList_AppendNamed

Gecko_FontFamilyList_Clear

Gecko_GetAnimationRule

Gecko_GetAtomAsUTF16

Gecko_GetDocumentElement

Gecko_GetElementId

Gecko_GetFirstChild

Gecko_GetFirstChildElement

Gecko_GetHTMLPresentationAttrDeclarationBlock

Gecko_GetLastChild

Gecko_GetLastChildElement

Gecko_GetMediaFeatures

Gecko_GetNextSibling

Gecko_GetNextSiblingElement

Gecko_GetNextStyleChild

Gecko_GetNodeFlags

Gecko_GetParentNode

Gecko_GetPrevSibling

Gecko_GetPrevSiblingElement

Gecko_GetStyleAttrDeclarationBlock

Gecko_GetStyleContext

Gecko_HasAttr

Gecko_IsInDocument

Gecko_IsLink

Gecko_IsRootElement

Gecko_IsTextNode

Gecko_IsUnvisitedLink

Gecko_IsVisitedLink

Gecko_LoadStyleSheet

Gecko_LocalName

Gecko_MatchesElement

Gecko_MaybeCreateStyleChildrenIterator

Gecko_Namespace

Gecko_NewBasicShape

Gecko_NewCSSShadowArray

Gecko_NewCSSValueSharedList

Gecko_NewStyleQuoteValues

Gecko_NodeIsElement

Gecko_PropertyId_IsPrefEnabled

Gecko_ReleaseAtom

Gecko_ReleaseCSSShadowArrayArbitraryThread

Gecko_ReleaseCSSValueSharedListArbitraryThread

Gecko_ReleaseCalcArbitraryThread

Gecko_ReleasePrincipalArbitraryThread

Gecko_ReleaseQuoteValuesArbitraryThread

Gecko_ReleaseURIArbitraryThread

Gecko_ResetFilters

Gecko_ResetStyleCoord

Gecko_SetCursorArrayLength

Gecko_SetCursorImage

Gecko_SetGradientImageValue

Gecko_SetListStyleImage

Gecko_SetListStyleImageNone

Gecko_SetListStyleType

Gecko_SetMozBinding

Gecko_SetNodeFlags

Gecko_SetNullImageValue

Gecko_SetOwnerDocumentNeedsStyleFlush

Gecko_SetStyleCoordCalcValue

Gecko_SetUrlImageValue

Gecko_SnapshotAtomAttrValue

Gecko_SnapshotAttrDashEquals

Gecko_SnapshotAttrEquals

Gecko_SnapshotAttrHasPrefix

Gecko_SnapshotAttrHasSubstring

Gecko_SnapshotAttrHasSuffix

Gecko_SnapshotAttrIncludes

Gecko_SnapshotClassOrClassList

Gecko_SnapshotHasAttr

Gecko_UnsetNodeFlags

Servo_AnimationValue_AddRef

Servo_AnimationValue_GetOpacity

Servo_AnimationValue_GetTransform

Servo_AnimationValue_Release

Servo_AnimationValue_Serialize

Servo_AnimationValues_Interpolate

Servo_AnimationValues_Populate

Servo_AnimationValues_Uncompute

Servo_AssertTreeIsClean

Servo_CSSSupports

Servo_CSSSupports2

Servo_ComputedValues_AddRef

Servo_ComputedValues_GetForAnonymousBox

Servo_ComputedValues_Inherit

Servo_ComputedValues_Release

Servo_CssRules_AddRef

Servo_CssRules_DeleteRule

Servo_CssRules_GetStyleRuleAt

Servo_CssRules_InsertRule

Servo_CssRules_ListTypes

Servo_CssRules_Release

Servo_DeclarationBlock_AddPresValue

Servo_DeclarationBlock_AddRef

Servo_DeclarationBlock_Clone

Servo_DeclarationBlock_Count

Servo_DeclarationBlock_CreateEmpty

Servo_DeclarationBlock_Equals

Servo_DeclarationBlock_GetCssText

Servo_DeclarationBlock_GetNthProperty

Servo_DeclarationBlock_GetPropertyIsImportant

Servo_DeclarationBlock_GetPropertyValue

Servo_DeclarationBlock_GetPropertyValueById

Servo_DeclarationBlock_Release

Servo_DeclarationBlock_RemoveProperty

Servo_DeclarationBlock_RemovePropertyById

Servo_DeclarationBlock_SerializeOneValue

Servo_DeclarationBlock_SetProperty

Servo_DeclarationBlock_SetPropertyById

Servo_Element_ClearData

Servo_Element_GetSnapshot

Servo_GetStyleBackground

Servo_GetStyleBorder

Servo_GetStyleColor

Servo_GetStyleColumn

Servo_GetStyleContent

Servo_GetStyleDisplay

Servo_GetStyleEffects

Servo_GetStyleFont

Servo_GetStyleList

Servo_GetStyleMargin

Servo_GetStyleOutline

Servo_GetStylePadding

Servo_GetStylePosition

Servo_GetStyleSVG

Servo_GetStyleSVGReset

Servo_GetStyleTable

Servo_GetStyleTableBorder

Servo_GetStyleText

Servo_GetStyleTextReset

Servo_GetStyleUIReset

Servo_GetStyleUserInterface

Servo_GetStyleVariables

Servo_GetStyleVisibility

Servo_GetStyleXUL

Servo_ImportRule_AddRef

Servo_ImportRule_GetSheet

Servo_ImportRule_Release

Servo_Initialize

Servo_NoteExplicitHints

Servo_ParseProperty

Servo_ParseStyleAttribute

Servo_ResolvePseudoStyle

Servo_ResolveStyle

Servo_ResolveStyleLazily

Servo_RestyleWithAddedDeclaration

Servo_Shutdown

Servo_StyleRule_AddRef

Servo_StyleRule_Debug

Servo_StyleRule_GetCssText

Servo_StyleRule_GetSelectorText

Servo_StyleRule_GetStyle

Servo_StyleRule_Release

Servo_StyleRule_SetStyle

Servo_StyleSet_AppendStyleSheet

Servo_StyleSet_Drop

Servo_StyleSet_FillKeyframesForName

Servo_StyleSet_FlushStyleSheets

Servo_StyleSet_Init

Servo_StyleSet_InsertStyleSheetBefore

Servo_StyleSet_NoteStyleSheetsChanged

Servo_StyleSet_PrependStyleSheet

Servo_StyleSet_RebuildData

Servo_StyleSet_RemoveStyleSheet

Servo_StyleSheet_AddRef

Servo_StyleSheet_ClearAndUpdate

Servo_StyleSheet_Empty

Servo_StyleSheet_FromUTF8Bytes

Servo_StyleSheet_GetRules

Servo_StyleSheet_HasRules

Servo_StyleSheet_Release

Servo_TakeChangeHint

Servo_TraverseSubtree

Type Definitions

RawGeckoAnimationValueListBorrowed

RawGeckoAnimationValueListBorrowedMut

RawGeckoAnimationValueListBorrowedMutOrNull

RawGeckoAnimationValueListBorrowedOrNull

RawGeckoDocumentBorrowed

RawGeckoDocumentBorrowedOrNull

RawGeckoElementBorrowed

RawGeckoElementBorrowedOrNull

RawGeckoKeyframeListBorrowed

RawGeckoKeyframeListBorrowedMut

RawGeckoKeyframeListBorrowedMutOrNull

RawGeckoKeyframeListBorrowedOrNull

RawGeckoNodeBorrowed

RawGeckoNodeBorrowedOrNull

RawGeckoPresContextBorrowed

RawGeckoPresContextBorrowedOrNull

RawServoAnimationValueBorrowed

RawServoAnimationValueBorrowedOrNull

RawServoAnimationValueStrong

RawServoDeclarationBlockBorrowed

RawServoDeclarationBlockBorrowedOrNull

RawServoDeclarationBlockStrong

RawServoDeclarationBlockStrongBorrowed

RawServoDeclarationBlockStrongBorrowedOrNull

RawServoImportRuleBorrowed

RawServoImportRuleBorrowedOrNull

RawServoImportRuleStrong

RawServoStyleRuleBorrowed

RawServoStyleRuleBorrowedOrNull

RawServoStyleRuleStrong

RawServoStyleSetBorrowed

RawServoStyleSetBorrowedMut

RawServoStyleSetBorrowedMutOrNull

RawServoStyleSetBorrowedOrNull

RawServoStyleSetOwned

RawServoStyleSetOwnedOrNull

RawServoStyleSheetBorrowed

RawServoStyleSheetBorrowedOrNull

RawServoStyleSheetStrong

ServoComputedValuesBorrowed

ServoComputedValuesBorrowedOrNull

ServoComputedValuesStrong

ServoCssRulesBorrowed

ServoCssRulesBorrowedOrNull

ServoCssRulesStrong

ServoElementSnapshotBorrowed

ServoElementSnapshotBorrowedMut

ServoElementSnapshotBorrowedMutOrNull

ServoElementSnapshotBorrowedOrNull

ServoElementSnapshotOwned

ServoElementSnapshotOwnedOrNull

StyleChildrenIteratorBorrowed

StyleChildrenIteratorBorrowedMut

StyleChildrenIteratorBorrowedMutOrNull

StyleChildrenIteratorBorrowedOrNull

StyleChildrenIteratorOwned

StyleChildrenIteratorOwnedOrNull

nsCSSValueBorrowed

nsCSSValueBorrowedMut

nsCSSValueBorrowedMutOrNull

nsCSSValueBorrowedOrNull

nsTArrayBorrowed_uintptr_t