Module style::gecko_bindings::structs::root::JS

Modules

dbg

Structs

DeletePolicy

FreePolicy

GCCellPtr
A GC pointer, tagged with the trace kind.

Handle
Reference to a T that has been rooted elsewhere. This is most useful as a parameter type, which guarantees that the T lvalue is properly rooted. See "Move GC Stack Rooting" above.

Heap
The Heap class is a heap-stored reference to a JS GC thing. All members of heap classes that refer to GC things should use Heap (or possibly TenuredHeap, described below).

MutableHandle
Similar to a handle, but the underlying storage can be changed. This is useful for outparams.

TenuredHeap
The TenuredHeap class is similar to the Heap class above in that it encapsulates the GC concerns of an on-heap reference to a JS object. However, it has two important differences:

Value
JS::Value is the interface for a single JavaScript Engine value. A few general notes on JS::Value:

Value_layout

Value_layout__bindgen_ty_1

Value_layout__bindgen_ty_2

Value_layout__bindgen_ty_2__bindgen_ty_1

Enums

Handle_CallerIdentity

Handle_Disambiguator

TenuredHeap__bindgen_ty_1

Constants

TenuredHeap_flagsMask

TenuredHeap_maskBits

Type Definitions

HandleValue

Handle_ElementType

Heap_ElementType

MutableHandleValue

MutableHandle_ElementType

TenuredHeap_ElementType

Value_PayloadType