Enum style::gecko_bindings::structs::root::nsIPresShell__bindgen_ty_3

#[repr(u32)]

pub enum nsIPresShell__bindgen_ty_3 {
    RENDER_IS_UNTRUSTED,
    RENDER_IGNORE_VIEWPORT_SCROLLING,
    RENDER_CARET,
    RENDER_USE_WIDGET_LAYERS,
    RENDER_ASYNC_DECODE_IMAGES,
    RENDER_DOCUMENT_RELATIVE,
    RENDER_DRAWWINDOW_NOT_FLUSHING,
}

Render the document into an arbitrary gfxContext Designed for getting a picture of a document or a piece of a document Note that callers will generally want to call FlushPendingNotifications to get an up-to-date view of the document @param aRect is the region to capture into the offscreen buffer, in the root frame's coordinate system (if aIgnoreViewportScrolling is false) or in the root scrolled frame's coordinate system (if aIgnoreViewportScrolling is true). The coordinates are in appunits. @param aFlags see below; set RENDER_IS_UNTRUSTED if the contents may be passed to malicious agents. E.g. we might choose not to paint the contents of sensitive widgets such as the file name in a file upload widget, and we might choose not to paint themes. set RENDER_IGNORE_VIEWPORT_SCROLLING to ignore clipping and scrollbar painting due to scrolling in the viewport set RENDER_CARET to draw the caret if one would be visible (by default the caret is never drawn) set RENDER_USE_LAYER_MANAGER to force rendering to go through the layer manager for the window. This may be unexpectedly slow (if the layer manager must read back data from the GPU) or low-quality (if the layer manager reads back pixel data and scales it instead of rendering using the appropriate scaling). It may also slow everything down if the area rendered does not correspond to the normal visible area of the window. set RENDER_ASYNC_DECODE_IMAGES to avoid having images synchronously decoded during rendering. (by default images decode synchronously with RenderDocument) set RENDER_DOCUMENT_RELATIVE to render the document as if there has been no scrolling and interpret |aRect| relative to the document instead of the CSS viewport. Only considered if RENDER_IGNORE_VIEWPORT_SCROLLING is set or the document is in ignore viewport scrolling mode (nsIPresShell::SetIgnoreViewportScrolling/IgnoringViewportScrolling). @param aBackgroundColor a background color to render onto @param aRenderedContext the gfxContext to render to. We render so that one CSS pixel in the source document is rendered to one unit in the current transform.

Variants
RENDER_IS_UNTRUSTEDRENDER_IGNORE_VIEWPORT_SCROLLINGRENDER_CARETRENDER_USE_WIDGET_LAYERSRENDER_ASYNC_DECODE_IMAGESRENDER_DOCUMENT_RELATIVERENDER_DRAWWINDOW_NOT_FLUSHING