Module style::gecko_bindings::structs::root::mozilla

Reexports

pub use self::super::super::root::mozilla::ServoElementSnapshotFlags as ServoElementSnapshot_Flags;

Modules

a11y

binding_danger

css

detail

dom

gfx

image

layers

net

widget

Structs

AbstractThread

AccessibleCaretEventHub

AlignedStorage2

AlignedStorage2_U

AnimationPerformanceWarning

AnimationValue

BlockingResourceBase
BlockingResourceBase Base class of resources that might block clients trying to acquire them. Does debugging and deadlock detection in DEBUG builds.

CSSVariableResolver

CSSVariableValues

CSSVariableValues_Variable

ChangesToFlush

ContainerLayerParameters

CounterStyle

CounterStyleManager

CounterStyle__bindgen_vtable

DeadlockDetector

DeclarationBlock

DefaultDelete
A default deletion policy using plain old operator delete.

DisplayItemClip
An DisplayItemClip represents the intersection of an optional rectangle with a list of rounded rectangles (which is often empty), all in appunits. It can represent everything CSS clipping can do to an element (except for SVG clip-path), including no clipping at all.

DisplayItemClip_RoundedRect

EffectCompositor

EffectCompositor_AnimationStyleRuleProcessor

EffectCompositor_cycleCollection

ErrorResult

EventChainPostVisitor

EventChainPreVisitor

EventChainVisitor

EventFlags

EventListenerManager

EventStateManager

EventStates
EventStates is the class used to represent the event states of nsIContent instances. These states are calculated by IntrinsicState() and ContentStatesChanged() has to be called when one of them changes thus informing the layout/style engine of the change. Event states are associated with pseudo-classes.

FontFamilyList
font family list, array of font families and a default font type. font family names are either named strings or generics. the default font type is used to preserve the variable font fallback behavior

FontFamilyName
font family name, a string for the name if not a generic and a font type indicated named family or which generic family

FramePropertyDescriptorUntyped

FramePropertyTable
The FramePropertyTable is optimized for storing 0 or 1 properties on a given frame. Storing very large numbers of properties on a single frame will not be efficient.

FramePropertyTable_Entry
Our hashtable entry. The key is an nsIFrame*, the value is a PropertyValue representing one or more property/value pairs.

FramePropertyTable_PropertyComparator
Used with an array of PropertyValues to allow lookups that compare only on the FramePropertyDescriptor.

FramePropertyTable_PropertyValue
Stores a property descriptor/value pair. It can also be used to store an nsTArray of PropertyValues.

FramePropertyTable_ReinterpretHelper

HandleRefPtr
A class for holding strong references to handle-managed objects.

Keyframe
A single keyframe.

LangGroupFontPrefs

LinkedList

LinkedListElement

LinkedList_Iterator

MallocAllocPolicy

Mutex
Mutex When possible, use MutexAutoLock/MutexAutoUnlock to lock/unlock this mutex within a scope, instead of calling Lock/Unlock directly.

NegativeType

OOMReporter

OffTheBooksMutex
OffTheBooksMutex is identical to Mutex, except that OffTheBooksMutex doesn't include leak checking. Sometimes you want to intentionally "leak" a mutex until shutdown; in these cases, OffTheBooksMutex is for you.

OriginAttributes

OwningNonNull

PadType

Pair
Pair is the logical concatenation of an instance of A with an instance B. Space is conserved when possible. Neither A nor B may be a final class.

PendingAnimationTracker

Position

PropertyStyleAnimationValuePair

PropertyValuePair
A property-value pair specified on a keyframe.

RestyleHintData
Additional data used in conjunction with an nsRestyleHint to control the restyle process.

RestyleManagerHandle
Smart pointer class that can hold a pointer to either a RestyleManager or a ServoRestyleManager.

RestyleManagerHandle_Ptr

RestyleTracker

ReverseIterator

Runnable

SVGAnimatedPreserveAspectRatio

SVGAttrAnimationRuleProcessor

SVGLengthList

SVGNumberList

SVGPathData

SVGPointList

SVGStringList

SVGTransformList

ScrollbarStyles

ServoAnimationRule

ServoAttrSnapshot
A structure representing a single attribute name and value.

ServoElementSnapshot
This class holds all non-tree-structural state of an element that might be used for selector matching eventually.

ServoStyleSheet

ShortcutKeyCandidate

StaticPresData
Some functionality that has historically lived on nsPresContext does not actually need to be per-document. This singleton class serves as a host for that functionality. We delegate to it from nsPresContext where appropriate, and use it standalone in some cases as well.

StaticRefPtr

StyleAnimation

StyleAnimationValue
Utility class to handle animated style values

StyleAnimationValue__bindgen_ty_1

StyleBasicShape

StyleComplexColor
This struct represents a combined color from a numeric color and the current foreground color (currentcolor keyword). Conceptually, the formula is "color * (1 - p) + currentcolor * p" where p is mForegroundRatio. See mozilla::LinearBlendColors for the actual algorithm.

StyleSetHandle
Smart pointer class that can hold a pointer to either an nsStyleSet or a ServoStyleSet.

StyleSetHandle_Ptr

StyleShapeSource

StyleShapeSource__bindgen_ty_1

StyleSheet
Superclass for data common to CSSStyleSheet and ServoStyleSheet.

StyleSheet_cycleCollection

StyleTransition

ThreadSafeAutoRefCnt

TimeStamp
Instances of this class represent moments in time, or a special "null" moment. We do not use the non-monotonic system clock or local time, since they can be reset, causing apparent backward travel in time, which can confuse algorithms. Instead we measure elapsed time according to the system. This time can never go backwards (i.e. it never wraps around, at least not in less than five million years of system elapsed time). It might not advance while the system is sleeping. If TimeStamp::SetNow() is not called at all for hours or days, we might not notice the passage of some of that time.

UniquePtr

WidgetEvent

WidgetGUIEvent

WidgetInputEvent

WritingMode

fallible_t

Enums

ArenaObjectID

BlockingResourceBase_BlockingResourceType

CORSMode

CSSEnabledState

CSSPseudoClassType

CSSPseudoElementType

Directionality

EffectCompositor_CascadeLevel

EffectCompositor_RestyleType

FlushType
This is the enum used by nsIDocument::FlushPendingNotifications to decide what to flush.

FontFamilyType
type of font family name, either a name (e.g. Helvetica) or a generic (e.g. serif, sans-serif), with the ability to distinguish between unquoted and quoted names for serializaiton

HalfCorner

LinkedListElement_NodeKind

LogicalSide

OriginAttributes__bindgen_ty_1

SelectionType

ServoElementSnapshotFlags
A bitflags enum class used to determine what data does a ServoElementSnapshot contains.

SheetType

Side

StyleAnimationValue_ColorConstructorType

StyleAnimationValue_CoordConstructorType

StyleAnimationValue_FloatConstructorType

StyleAnimationValue_IntegerConstructorType

StyleAnimationValue_PercentConstructorType

StyleAnimationValue_Unit
The types and values for the values that we extract and animate.

StyleBackendType
Enumeration that represents one of the two supported style system backends.

StyleBasicShapeType

StyleBoxAlign

StyleBoxDecorationBreak

StyleBoxDirection

StyleBoxOrient

StyleBoxPack

StyleBoxSizing

StyleClear

StyleDisplay

StyleFillRule

StyleFloat

StyleFloatEdge

StyleGeometryBox

StyleHyphens

StyleOrient

StyleShapeOutsideShapeBox

StyleShapeRadius

StyleShapeSourceType

StyleSheet_DocumentAssociationMode

StyleUserFocus

StyleUserInput

StyleUserModify

StyleUserSelect

StyleWindowDragging

TaskCategory

TraversalRootBehavior

UseCounter

Constants

CSSPseudoElementType_AnonBox

EffectCompositor_kCascadeLevelCount

FlushType_Frames

FontFamilyType_eFamily_generic_count

FontFamilyType_eFamily_generic_first

FontFamilyType_eFamily_generic_last

Keyframe_kComputedOffsetNotSet

LinkedListElement_NodeKind_Sentinel

OriginAttributes_STRIP_ADDON_ID

OriginAttributes_STRIP_FIRST_PARTY_DOMAIN

OriginAttributes_STRIP_USER_CONTEXT_ID

SERVO_PREF_ENABLED__moz_appearance

SERVO_PREF_ENABLED__moz_binding

SERVO_PREF_ENABLED__moz_border_bottom_colors

SERVO_PREF_ENABLED__moz_border_left_colors

SERVO_PREF_ENABLED__moz_border_right_colors

SERVO_PREF_ENABLED__moz_border_top_colors

SERVO_PREF_ENABLED__moz_box_align

SERVO_PREF_ENABLED__moz_box_direction

SERVO_PREF_ENABLED__moz_box_flex

SERVO_PREF_ENABLED__moz_box_ordinal_group

SERVO_PREF_ENABLED__moz_box_orient

SERVO_PREF_ENABLED__moz_box_pack

SERVO_PREF_ENABLED__moz_control_character_visibility

SERVO_PREF_ENABLED__moz_float_edge

SERVO_PREF_ENABLED__moz_force_broken_image_icon

SERVO_PREF_ENABLED__moz_image_region

SERVO_PREF_ENABLED__moz_math_display

SERVO_PREF_ENABLED__moz_math_variant

SERVO_PREF_ENABLED__moz_min_font_size_ratio

SERVO_PREF_ENABLED__moz_orient

SERVO_PREF_ENABLED__moz_osx_font_smoothing

SERVO_PREF_ENABLED__moz_outline_radius

SERVO_PREF_ENABLED__moz_outline_radius_bottomleft

SERVO_PREF_ENABLED__moz_outline_radius_bottomright

SERVO_PREF_ENABLED__moz_outline_radius_topleft

SERVO_PREF_ENABLED__moz_outline_radius_topright

SERVO_PREF_ENABLED__moz_script_level

SERVO_PREF_ENABLED__moz_script_min_size

SERVO_PREF_ENABLED__moz_script_size_multiplier

SERVO_PREF_ENABLED__moz_stack_sizing

SERVO_PREF_ENABLED__moz_tab_size

SERVO_PREF_ENABLED__moz_text_size_adjust

SERVO_PREF_ENABLED__moz_top_layer

SERVO_PREF_ENABLED__moz_transform

SERVO_PREF_ENABLED__moz_user_focus

SERVO_PREF_ENABLED__moz_user_input

SERVO_PREF_ENABLED__moz_user_modify

SERVO_PREF_ENABLED__moz_user_select

SERVO_PREF_ENABLED__moz_window_dragging

SERVO_PREF_ENABLED__moz_window_shadow

SERVO_PREF_ENABLED__webkit_text_fill_color

SERVO_PREF_ENABLED__webkit_text_stroke

SERVO_PREF_ENABLED__webkit_text_stroke_color

SERVO_PREF_ENABLED__webkit_text_stroke_width

SERVO_PREF_ENABLED__x_lang

SERVO_PREF_ENABLED__x_span

SERVO_PREF_ENABLED__x_system_font

SERVO_PREF_ENABLED__x_text_zoom

SERVO_PREF_ENABLED_align_content

SERVO_PREF_ENABLED_align_items

SERVO_PREF_ENABLED_align_self

SERVO_PREF_ENABLED_all

SERVO_PREF_ENABLED_animation

SERVO_PREF_ENABLED_animation_delay

SERVO_PREF_ENABLED_animation_direction

SERVO_PREF_ENABLED_animation_duration

SERVO_PREF_ENABLED_animation_fill_mode

SERVO_PREF_ENABLED_animation_iteration_count

SERVO_PREF_ENABLED_animation_name

SERVO_PREF_ENABLED_animation_play_state

SERVO_PREF_ENABLED_animation_timing_function

SERVO_PREF_ENABLED_backface_visibility

SERVO_PREF_ENABLED_background

SERVO_PREF_ENABLED_background_attachment

SERVO_PREF_ENABLED_background_blend_mode

SERVO_PREF_ENABLED_background_clip

SERVO_PREF_ENABLED_background_color

SERVO_PREF_ENABLED_background_image

SERVO_PREF_ENABLED_background_origin

SERVO_PREF_ENABLED_background_position

SERVO_PREF_ENABLED_background_position_x

SERVO_PREF_ENABLED_background_position_y

SERVO_PREF_ENABLED_background_repeat

SERVO_PREF_ENABLED_background_size

SERVO_PREF_ENABLED_block_size

SERVO_PREF_ENABLED_border

SERVO_PREF_ENABLED_border_block_end

SERVO_PREF_ENABLED_border_block_end_color

SERVO_PREF_ENABLED_border_block_end_style

SERVO_PREF_ENABLED_border_block_end_width

SERVO_PREF_ENABLED_border_block_start

SERVO_PREF_ENABLED_border_block_start_color

SERVO_PREF_ENABLED_border_block_start_style

SERVO_PREF_ENABLED_border_block_start_width

SERVO_PREF_ENABLED_border_bottom

SERVO_PREF_ENABLED_border_bottom_color

SERVO_PREF_ENABLED_border_bottom_left_radius

SERVO_PREF_ENABLED_border_bottom_right_radius

SERVO_PREF_ENABLED_border_bottom_style

SERVO_PREF_ENABLED_border_bottom_width

SERVO_PREF_ENABLED_border_collapse

SERVO_PREF_ENABLED_border_color

SERVO_PREF_ENABLED_border_image

SERVO_PREF_ENABLED_border_image_outset

SERVO_PREF_ENABLED_border_image_repeat

SERVO_PREF_ENABLED_border_image_slice

SERVO_PREF_ENABLED_border_image_source

SERVO_PREF_ENABLED_border_image_width

SERVO_PREF_ENABLED_border_inline_end

SERVO_PREF_ENABLED_border_inline_end_color

SERVO_PREF_ENABLED_border_inline_end_style

SERVO_PREF_ENABLED_border_inline_end_width

SERVO_PREF_ENABLED_border_inline_start

SERVO_PREF_ENABLED_border_inline_start_color

SERVO_PREF_ENABLED_border_inline_start_style

SERVO_PREF_ENABLED_border_inline_start_width

SERVO_PREF_ENABLED_border_left

SERVO_PREF_ENABLED_border_left_color

SERVO_PREF_ENABLED_border_left_style

SERVO_PREF_ENABLED_border_left_width

SERVO_PREF_ENABLED_border_radius

SERVO_PREF_ENABLED_border_right

SERVO_PREF_ENABLED_border_right_color

SERVO_PREF_ENABLED_border_right_style

SERVO_PREF_ENABLED_border_right_width

SERVO_PREF_ENABLED_border_spacing

SERVO_PREF_ENABLED_border_style

SERVO_PREF_ENABLED_border_top

SERVO_PREF_ENABLED_border_top_color

SERVO_PREF_ENABLED_border_top_left_radius

SERVO_PREF_ENABLED_border_top_right_radius

SERVO_PREF_ENABLED_border_top_style

SERVO_PREF_ENABLED_border_top_width

SERVO_PREF_ENABLED_border_width

SERVO_PREF_ENABLED_bottom

SERVO_PREF_ENABLED_box_decoration_break

SERVO_PREF_ENABLED_box_shadow

SERVO_PREF_ENABLED_box_sizing

SERVO_PREF_ENABLED_caption_side

SERVO_PREF_ENABLED_caret_color

SERVO_PREF_ENABLED_clear

SERVO_PREF_ENABLED_clip

SERVO_PREF_ENABLED_clip_path

SERVO_PREF_ENABLED_clip_rule

SERVO_PREF_ENABLED_color

SERVO_PREF_ENABLED_color_adjust

SERVO_PREF_ENABLED_color_interpolation

SERVO_PREF_ENABLED_color_interpolation_filters

SERVO_PREF_ENABLED_column_count

SERVO_PREF_ENABLED_column_fill

SERVO_PREF_ENABLED_column_gap

SERVO_PREF_ENABLED_column_rule

SERVO_PREF_ENABLED_column_rule_color

SERVO_PREF_ENABLED_column_rule_style

SERVO_PREF_ENABLED_column_rule_width

SERVO_PREF_ENABLED_column_width

SERVO_PREF_ENABLED_columns

SERVO_PREF_ENABLED_contain

SERVO_PREF_ENABLED_content

SERVO_PREF_ENABLED_counter_increment

SERVO_PREF_ENABLED_counter_reset

SERVO_PREF_ENABLED_cursor

SERVO_PREF_ENABLED_direction

SERVO_PREF_ENABLED_display

SERVO_PREF_ENABLED_dominant_baseline

SERVO_PREF_ENABLED_empty_cells

SERVO_PREF_ENABLED_fill

SERVO_PREF_ENABLED_fill_opacity

SERVO_PREF_ENABLED_fill_rule

SERVO_PREF_ENABLED_filter

SERVO_PREF_ENABLED_flex

SERVO_PREF_ENABLED_flex_basis

SERVO_PREF_ENABLED_flex_direction

SERVO_PREF_ENABLED_flex_flow

SERVO_PREF_ENABLED_flex_grow

SERVO_PREF_ENABLED_flex_shrink

SERVO_PREF_ENABLED_flex_wrap

SERVO_PREF_ENABLED_float_

SERVO_PREF_ENABLED_flood_color

SERVO_PREF_ENABLED_flood_opacity

SERVO_PREF_ENABLED_font

SERVO_PREF_ENABLED_font_family

SERVO_PREF_ENABLED_font_feature_settings

SERVO_PREF_ENABLED_font_kerning

SERVO_PREF_ENABLED_font_language_override

SERVO_PREF_ENABLED_font_size

SERVO_PREF_ENABLED_font_size_adjust

SERVO_PREF_ENABLED_font_stretch

SERVO_PREF_ENABLED_font_style

SERVO_PREF_ENABLED_font_synthesis

SERVO_PREF_ENABLED_font_variant

SERVO_PREF_ENABLED_font_variant_alternates

SERVO_PREF_ENABLED_font_variant_caps

SERVO_PREF_ENABLED_font_variant_east_asian

SERVO_PREF_ENABLED_font_variant_ligatures

SERVO_PREF_ENABLED_font_variant_numeric

SERVO_PREF_ENABLED_font_variant_position

SERVO_PREF_ENABLED_font_variation_settings

SERVO_PREF_ENABLED_font_weight

SERVO_PREF_ENABLED_grid

SERVO_PREF_ENABLED_grid_area

SERVO_PREF_ENABLED_grid_auto_columns

SERVO_PREF_ENABLED_grid_auto_flow

SERVO_PREF_ENABLED_grid_auto_rows

SERVO_PREF_ENABLED_grid_column

SERVO_PREF_ENABLED_grid_column_end

SERVO_PREF_ENABLED_grid_column_gap

SERVO_PREF_ENABLED_grid_column_start

SERVO_PREF_ENABLED_grid_gap

SERVO_PREF_ENABLED_grid_row

SERVO_PREF_ENABLED_grid_row_end

SERVO_PREF_ENABLED_grid_row_gap

SERVO_PREF_ENABLED_grid_row_start

SERVO_PREF_ENABLED_grid_template

SERVO_PREF_ENABLED_grid_template_areas

SERVO_PREF_ENABLED_grid_template_columns

SERVO_PREF_ENABLED_grid_template_rows

SERVO_PREF_ENABLED_height

SERVO_PREF_ENABLED_hyphens

SERVO_PREF_ENABLED_image_orientation

SERVO_PREF_ENABLED_image_rendering

SERVO_PREF_ENABLED_ime_mode

SERVO_PREF_ENABLED_initial_letter

SERVO_PREF_ENABLED_inline_size

SERVO_PREF_ENABLED_isolation

SERVO_PREF_ENABLED_justify_content

SERVO_PREF_ENABLED_justify_items

SERVO_PREF_ENABLED_justify_self

SERVO_PREF_ENABLED_left

SERVO_PREF_ENABLED_letter_spacing

SERVO_PREF_ENABLED_lighting_color

SERVO_PREF_ENABLED_line_height

SERVO_PREF_ENABLED_list_style

SERVO_PREF_ENABLED_list_style_image

SERVO_PREF_ENABLED_list_style_position

SERVO_PREF_ENABLED_list_style_type

SERVO_PREF_ENABLED_margin

SERVO_PREF_ENABLED_margin_block_end

SERVO_PREF_ENABLED_margin_block_start

SERVO_PREF_ENABLED_margin_bottom

SERVO_PREF_ENABLED_margin_inline_end

SERVO_PREF_ENABLED_margin_inline_start

SERVO_PREF_ENABLED_margin_left

SERVO_PREF_ENABLED_margin_right

SERVO_PREF_ENABLED_margin_top

SERVO_PREF_ENABLED_marker

SERVO_PREF_ENABLED_marker_end

SERVO_PREF_ENABLED_marker_mid

SERVO_PREF_ENABLED_marker_start

SERVO_PREF_ENABLED_mask

SERVO_PREF_ENABLED_mask_clip

SERVO_PREF_ENABLED_mask_composite

SERVO_PREF_ENABLED_mask_image

SERVO_PREF_ENABLED_mask_mode

SERVO_PREF_ENABLED_mask_origin

SERVO_PREF_ENABLED_mask_position

SERVO_PREF_ENABLED_mask_position_x

SERVO_PREF_ENABLED_mask_position_y

SERVO_PREF_ENABLED_mask_repeat

SERVO_PREF_ENABLED_mask_size

SERVO_PREF_ENABLED_mask_type

SERVO_PREF_ENABLED_max_block_size

SERVO_PREF_ENABLED_max_height

SERVO_PREF_ENABLED_max_inline_size

SERVO_PREF_ENABLED_max_width

SERVO_PREF_ENABLED_min_block_size

SERVO_PREF_ENABLED_min_height

SERVO_PREF_ENABLED_min_inline_size

SERVO_PREF_ENABLED_min_width

SERVO_PREF_ENABLED_mix_blend_mode

SERVO_PREF_ENABLED_object_fit

SERVO_PREF_ENABLED_object_position

SERVO_PREF_ENABLED_offset_block_end

SERVO_PREF_ENABLED_offset_block_start

SERVO_PREF_ENABLED_offset_inline_end

SERVO_PREF_ENABLED_offset_inline_start

SERVO_PREF_ENABLED_opacity

SERVO_PREF_ENABLED_order

SERVO_PREF_ENABLED_outline

SERVO_PREF_ENABLED_outline_color

SERVO_PREF_ENABLED_outline_offset

SERVO_PREF_ENABLED_outline_style

SERVO_PREF_ENABLED_outline_width

SERVO_PREF_ENABLED_overflow

SERVO_PREF_ENABLED_overflow_clip_box

SERVO_PREF_ENABLED_overflow_wrap

SERVO_PREF_ENABLED_overflow_x

SERVO_PREF_ENABLED_overflow_y

SERVO_PREF_ENABLED_padding

SERVO_PREF_ENABLED_padding_block_end

SERVO_PREF_ENABLED_padding_block_start

SERVO_PREF_ENABLED_padding_bottom

SERVO_PREF_ENABLED_padding_inline_end

SERVO_PREF_ENABLED_padding_inline_start

SERVO_PREF_ENABLED_padding_left

SERVO_PREF_ENABLED_padding_right

SERVO_PREF_ENABLED_padding_top

SERVO_PREF_ENABLED_page_break_after

SERVO_PREF_ENABLED_page_break_before

SERVO_PREF_ENABLED_page_break_inside

SERVO_PREF_ENABLED_paint_order

SERVO_PREF_ENABLED_perspective

SERVO_PREF_ENABLED_perspective_origin

SERVO_PREF_ENABLED_place_content

SERVO_PREF_ENABLED_place_items

SERVO_PREF_ENABLED_place_self

SERVO_PREF_ENABLED_pointer_events

SERVO_PREF_ENABLED_position

SERVO_PREF_ENABLED_quotes

SERVO_PREF_ENABLED_resize

SERVO_PREF_ENABLED_right

SERVO_PREF_ENABLED_ruby_align

SERVO_PREF_ENABLED_ruby_position

SERVO_PREF_ENABLED_scroll_behavior

SERVO_PREF_ENABLED_scroll_snap_coordinate

SERVO_PREF_ENABLED_scroll_snap_destination

SERVO_PREF_ENABLED_scroll_snap_points_x

SERVO_PREF_ENABLED_scroll_snap_points_y

SERVO_PREF_ENABLED_scroll_snap_type

SERVO_PREF_ENABLED_scroll_snap_type_x

SERVO_PREF_ENABLED_scroll_snap_type_y

SERVO_PREF_ENABLED_shape_outside

SERVO_PREF_ENABLED_shape_rendering

SERVO_PREF_ENABLED_stop_color

SERVO_PREF_ENABLED_stop_opacity

SERVO_PREF_ENABLED_stroke

SERVO_PREF_ENABLED_stroke_dasharray

SERVO_PREF_ENABLED_stroke_dashoffset

SERVO_PREF_ENABLED_stroke_linecap

SERVO_PREF_ENABLED_stroke_linejoin

SERVO_PREF_ENABLED_stroke_miterlimit

SERVO_PREF_ENABLED_stroke_opacity

SERVO_PREF_ENABLED_stroke_width

SERVO_PREF_ENABLED_table_layout

SERVO_PREF_ENABLED_text_align

SERVO_PREF_ENABLED_text_align_last

SERVO_PREF_ENABLED_text_anchor

SERVO_PREF_ENABLED_text_combine_upright

SERVO_PREF_ENABLED_text_decoration

SERVO_PREF_ENABLED_text_decoration_color

SERVO_PREF_ENABLED_text_decoration_line

SERVO_PREF_ENABLED_text_decoration_style

SERVO_PREF_ENABLED_text_emphasis

SERVO_PREF_ENABLED_text_emphasis_color

SERVO_PREF_ENABLED_text_emphasis_position

SERVO_PREF_ENABLED_text_emphasis_style

SERVO_PREF_ENABLED_text_indent

SERVO_PREF_ENABLED_text_orientation

SERVO_PREF_ENABLED_text_overflow

SERVO_PREF_ENABLED_text_rendering

SERVO_PREF_ENABLED_text_shadow

SERVO_PREF_ENABLED_text_transform

SERVO_PREF_ENABLED_top

SERVO_PREF_ENABLED_touch_action

SERVO_PREF_ENABLED_transform

SERVO_PREF_ENABLED_transform_box

SERVO_PREF_ENABLED_transform_origin

SERVO_PREF_ENABLED_transform_style

SERVO_PREF_ENABLED_transition

SERVO_PREF_ENABLED_transition_delay

SERVO_PREF_ENABLED_transition_duration

SERVO_PREF_ENABLED_transition_property

SERVO_PREF_ENABLED_transition_timing_function

SERVO_PREF_ENABLED_unicode_bidi

SERVO_PREF_ENABLED_vector_effect

SERVO_PREF_ENABLED_vertical_align

SERVO_PREF_ENABLED_visibility

SERVO_PREF_ENABLED_white_space

SERVO_PREF_ENABLED_width

SERVO_PREF_ENABLED_will_change

SERVO_PREF_ENABLED_word_break

SERVO_PREF_ENABLED_word_spacing

SERVO_PREF_ENABLED_writing_mode

SERVO_PREF_ENABLED_z_index

ThreadSafeAutoRefCnt_isThreadSafe

Statics

BlockingResourceBase_kResourceTypeName

BlockingResourceBase_sCallOnce

BlockingResourceBase_sDeadlockDetector

BlockingResourceBase_sResourceAcqnChainFrontTPI

EffectCompositor__cycleCollectorGlobal

StyleSheet__cycleCollectorGlobal

Type Definitions

BlockingResourceBase_AcquisitionState

BlockingResourceBase_DDT

CSSIntPoint

CSSPoint

CSSPseudoClassTypeBase

CSSPseudoElementTypeBase

CSSRect

CSSSize

CSSToLayoutDeviceScale

CSSToParentLayerScale2D

CounterStyleManager_HasThreadSafeRefCnt

CounterValue

DisplayItemClip_Color

DisplayItemClip_DrawTarget

DisplayItemClip_Path

EffectCompositor_AnimationStyleRuleProcessor_HasThreadSafeRefCnt

EffectCompositor_HasThreadSafeRefCnt

ErrorResult_BaseErrorResult

EventStates_InternalType

EventStates_ServoType

FalseType

FramePropertyDescriptorUntyped_UntypedDestructor
mDestructor will be called if it's non-null.

FramePropertyDescriptorUntyped_UntypedDestructorWithFrame
mDestructorWithFrame will be called if it's non-null and mDestructor is null. WARNING: The frame passed to mDestructorWithFrame may be a dangling frame pointer, if this is being called during presshell teardown. Do not use it except to compare against other frame pointers. No frame will have been allocated with the same address yet.

FramePropertyTable_UntypedDescriptor

LayoutDeviceIntPoint

LayoutDeviceIntRect

LayoutDeviceIntSize

LayoutDeviceToLayerScale2D

LinkedListElement_ClientType

LinkedListElement_ConstClientType

LinkedListElement_ConstRawType

LinkedListElement_RawType

LinkedListElement_Traits

LinkedList_ClientType

LinkedList_ConstClientType

LinkedList_ConstRawType

LinkedList_RawType

LinkedList_Traits

MallocSizeOf

Pair_Base

ParentLayerRect

Position_Coord

RawSelectionType

RestyleManagerHandle_RefPtr

Runnable_HasThreadSafeRefCnt

ScreenIntSize

ScreenMargin

ScreenToLayerScale2D

ServoElementSnapshot_BorrowedAttrInfo

ServoElementSnapshot_Element

ServoElementSnapshot_ServoStateType

StyleBasicShape_HasThreadSafeRefCnt

StyleClipPath

StyleShapeOutside

StyleSheet_HasThreadSafeRefCnt

TimeStampValue

TrueType
Convenient aliases.