Module style::gecko_bindings::structs::root::mozilla::binding_danger

Structs

AssertAndSuppressCleanupPolicy

JustAssertCleanupPolicy

TErrorResult
Templated implementation class for various ErrorResult-like things. The instantiations differ only in terms of their cleanup policies (used in the destructor), which they can specify via the template argument. Note that this means it's safe to reinterpret_cast between the instantiations unless you plan to invoke the destructor through such a cast pointer.

TErrorResult_DOMExceptionInfo

TErrorResult_Message

TErrorResult__bindgen_ty_1

Enums

TErrorResult_UnionState

Constants

AssertAndSuppressCleanupPolicy_assertHandled

AssertAndSuppressCleanupPolicy_suppress

JustAssertCleanupPolicy_assertHandled

JustAssertCleanupPolicy_suppress

TErrorResult_UnionState_HasDOMExceptionInfo

TErrorResult_UnionState_HasJSException

TErrorResult_UnionState_HasNothing