Module style::gecko_bindings::structs::root::mozilla::dom

Modules

binding_detail

constructors

namedpropertiesobjects

prototypes

workers

Structs

AccessibleNode

AddEventListenerOptionsOrBoolean

AllOwningUnionBase

Animation

AnimationFilter

AnonymousContent

Attr

Attr_cycleCollection

AudioContext

BorrowedAttrInfo
Struct that stores info on an attribute. The name and value must either both be null or both be non-null.

BoxObject

BoxQuadOptions

CDATASection

CSSRuleList

CallbackFunction

CallbackObject

CallbackObjectHolder

CallbackObjectHolderBase

CallbackObject_COMTypeInfo

CallbackObject_CallSetup

CallbackObject_FastCallbackConstructor

CallbackObject_JSObjectsDropper

CallbackObject_cycleCollection

Comment

ConvertCoordinateOptions

CustomElementData

CustomElementRegistry

DOMImplementation

DOMIntersectionObserver

DOMIntersectionObserverEntry

DOMIntersectionObserverEntry_cycleCollection

DOMIntersectionObserver_COMTypeInfo

DOMIntersectionObserver_cycleCollection

DOMPoint

DOMPointInit

DOMQuad

DOMRect

DOMRectList

DOMRectList_cycleCollection

DOMRectReadOnly

DOMRectReadOnly_cycleCollection

DOMString
A class for representing string return values. This can be either passed to callees that have an nsString or nsAString out param or passed to a callee that actually knows about this class and can work with it. Such a callee may call SetStringBuffer or SetEphemeralStringBuffer or SetOwnedString or SetOwnedAtom on this object. It's only OK to call SetStringBuffer/SetOwnedString/SetOwnedAtom if the caller of the method in question plans to keep holding a strong ref to the stringbuffer involved, whether it's a raw nsStringBuffer, or stored inside the string or atom being passed. In the string/atom cases that means the caller must own the string or atom, and not mutate it (in the string case) for the lifetime of the DOMString.

DOMStringList

DestinationInsertionPointList

DestinationInsertionPointList_cycleCollection

DictionaryBase

DictionaryBase_FastDictionaryInitializer

DispatcherTrait

DispatcherTrait__bindgen_vtable

DocGroup

DocumentFragment

DocumentTimeline

DocumentType

Element

ElementCreationOptions

ElementCreationOptionsOrString

ElementCreationOptionsOrString_Value

ElementRegistrationOptions

Element_COMTypeInfo

Element_MappedAttributeEntry

Element_StyleStateLocks
StyleStateLocks is used to specify which event states should be locked, and whether they should be locked to on or off.

Event

EventHandlerNonNull

EventListener

EventListenerOptionsOrBoolean

EventTarget

EventTarget_COMTypeInfo

FontFaceSet

FragmentOrElement

FragmentOrElement_cycleCollection

FragmentOrElement_nsDOMSlots
There are a set of DOM- and scripting-specific instance variables that may only be instantiated when a content object is accessed through the DOM. Rather than burn actual slots in the content objects for each of these instance variables, we put them off in a side structure that's only allocated when the content is accessed through the DOM.

FragmentOrElement_nsDOMSlots__bindgen_ty_1

FrameRequestCallback

FullscreenRequest

GetRootNodeOptions

GlobalObject

Grid

HTMLBodyElement

ImageTracker

ImportManager

IntersectionCallback

IntersectionObserverInit

Link

Location

MediaQueryList

NodeFilter

NodeInfo

NodeInfo_NodeInfoInner

NodeInfo_cycleCollection

NodeIterator

OriginAttributesDictionary

OwningDoubleOrDoubleSequence

OwningDoubleOrDoubleSequence_Value

OwningNodeOrString

OwningNodeOrString_Value

ParentObject

Performance

PrefSetting

ProcessingInstruction

Promise

SVGSVGElement

ScrollIntoViewOptions

ScrollOptions

ScrollToOptions

Selection

ServiceWorkerRegistration

ShadowRoot

StyleSheetList

Text

TextOrElementOrDocument

TimeoutManager

Touch

TouchList

TreeWalker

UnionMember

UnrestrictedDoubleOrKeyframeAnimationOptions

XPathEvaluator

XPathExpression

XPathNSResolver

XPathResult

Enums

CSSStyleSheetParsingMode

CallbackObject_ExceptionHandling

CallerType

DOMString_NullHandling

ElementCreationOptionsOrString_Type

FillMode

LargeAllocStatus

OrientationType

OwningDoubleOrDoubleSequence_Type

OwningNodeOrString_Type

PlaybackDirection

ScrollBehavior

ScrollLogicalPosition

VisibilityState

Constants

Element_kCallAfterSetAttr

Element_kDontCallAfterSetAttr

Element_kDontFireMutationEvent

Element_kDontNotifyDocumentObservers

Element_kFireMutationEvent

Element_kNotifyDocumentObservers

Statics

Attr__cycleCollectorGlobal

Attr_sInitialized

CallbackObject__cycleCollectorGlobal

DOMIntersectionObserverEntry__cycleCollectorGlobal

DOMIntersectionObserver__cycleCollectorGlobal

DOMRectList__cycleCollectorGlobal

DOMRectReadOnly__cycleCollectorGlobal

DestinationInsertionPointList__cycleCollectorGlobal

FragmentOrElement__cycleCollectorGlobal

NodeInfo__cycleCollectorGlobal

Type Definitions

Attr_HasThreadSafeRefCnt

CallbackObject_HasThreadSafeRefCnt

DOMIntersectionObserverEntry_HasThreadSafeRefCnt

DOMIntersectionObserver_HasThreadSafeRefCnt

DOMRectList_HasThreadSafeRefCnt

DOMRectReadOnly_HasThreadSafeRefCnt

DOMTokenListSupportedToken

DOMTokenListSupportedTokenArray

DestinationInsertionPointList_HasThreadSafeRefCnt

FragmentOrElement_HasThreadSafeRefCnt

NodeFilterHolder

NodeInfo_HasThreadSafeRefCnt