Struct style::gecko_bindings::structs::root::mozilla::dom::FragmentOrElement_nsDOMSlots

#[repr(C)]

pub struct FragmentOrElement_nsDOMSlots {
    pub _base: nsINode_nsSlots,
    pub mStyle: nsCOMPtr<nsICSSDeclaration>,
    pub mDataset: *mut nsDOMStringMap,
    pub mSMILOverrideStyle: nsCOMPtr<nsICSSDeclaration>,
    pub mSMILOverrideStyleDeclaration: RefPtr<DeclarationBlock>,
    pub mAttributeMap: RefPtr<nsDOMAttributeMap>,
    pub __bindgen_anon_1: FragmentOrElement_nsDOMSlots__bindgen_ty_1,
    pub mChildrenList: RefPtr<nsContentList>,
    pub mClassList: RefPtr<nsDOMTokenList>,
    pub mShadowRoot: RefPtr<ShadowRoot>,
    pub mContainingShadow: RefPtr<ShadowRoot>,
    pub mDestInsertionPoints: nsTArray<*mut nsIContent>,
    pub mXBLBinding: RefPtr<nsXBLBinding>,
    pub mXBLInsertionParent: nsCOMPtr<nsIContent>,
    pub mCustomElementData: RefPtr<CustomElementData>,
    pub mRegisteredIntersectionObservers: [u64; 6],
}

There are a set of DOM- and scripting-specific instance variables that may only be instantiated when a content object is accessed through the DOM. Rather than burn actual slots in the content objects for each of these instance variables, we put them off in a side structure that's only allocated when the content is accessed through the DOM.

Fields
_base: nsINode_nsSlots mStyle: nsCOMPtr<nsICSSDeclaration>

The .style attribute (an interface that forwards to the actual style rules) @see nsGenericHTMLElement::GetStyle

mDataset: *mut nsDOMStringMap

The .dataset attribute. @see nsGenericHTMLElement::GetDataset

mSMILOverrideStyle: nsCOMPtr<nsICSSDeclaration>

SMIL Overridde style rules (for SMIL animation of CSS properties) @see nsIContent::GetSMILOverrideStyle

mSMILOverrideStyleDeclaration: RefPtr<DeclarationBlock>

Holds any SMIL override style declaration for this element.

mAttributeMap: RefPtr<nsDOMAttributeMap>

An object implementing nsIDOMMozNamedAttrMap for this content (attributes) @see FragmentOrElement::GetAttributes

__bindgen_anon_1: FragmentOrElement_nsDOMSlots__bindgen_ty_1 mChildrenList: RefPtr<nsContentList>

An object implementing the .children property for this element.

mClassList: RefPtr<nsDOMTokenList>

An object implementing the .classList property for this element.

mShadowRoot: RefPtr<ShadowRoot>

ShadowRoot bound to the element.

mContainingShadow: RefPtr<ShadowRoot>

The root ShadowRoot of this element if it is in a shadow tree.

mDestInsertionPoints: nsTArray<*mut nsIContent>

An array of web component insertion points to which this element is distributed.

mXBLBinding: RefPtr<nsXBLBinding>

XBL binding installed on the element.

mXBLInsertionParent: nsCOMPtr<nsIContent>

XBL binding installed on the lement.

mCustomElementData: RefPtr<CustomElementData>

Web components custom element data.

mRegisteredIntersectionObservers: [u64; 6]

Registered Intersection Observers on the element.