Module style::gecko_bindings::structs::root::mozilla::layers

Structs

ContainerLayer

FrameMetrics
The viewport and displayport metrics for the painted frame at the time of a layer-tree transaction. These metrics are especially useful for shadow layers, because the metrics values are updated atomically with new pixels.

Layer

LayerManager

ScrollSnapInfo

Enums

FrameMetrics_ScrollOffsetUpdateType

Constants

FrameMetrics_START_SCROLL_ID

Statics

FrameMetrics_NULL_SCROLL_ID

Type Definitions

FrameMetrics_ViewID