Struct style::gecko_bindings::structs::root::mozilla::layers::FrameMetrics

#[repr(C)]

pub struct FrameMetrics {
    pub mScrollId: FrameMetrics_ViewID,
    pub mPresShellResolution: f32,
    pub mCompositionBounds: ParentLayerRect,
    pub mDisplayPort: CSSRect,
    pub mCriticalDisplayPort: CSSRect,
    pub mScrollableRect: CSSRect,
    pub mCumulativeResolution: LayoutDeviceToLayerScale2D,
    pub mDevPixelsPerCSSPixel: CSSToLayoutDeviceScale,
    pub mScrollOffset: CSSPoint,
    pub mZoom: CSSToParentLayerScale2D,
    pub mScrollGeneration: u32,
    pub mSmoothScrollOffset: CSSPoint,
    pub mRootCompositionSize: CSSSize,
    pub mDisplayPortMargins: ScreenMargin,
    pub mPresShellId: u32,
    pub mViewport: CSSRect,
    pub mExtraResolution: ScreenToLayerScale2D,
    pub mPaintRequestTime: TimeStamp,
    pub mScrollUpdateType: FrameMetrics_ScrollOffsetUpdateType,
    pub _bitfield_1: u8,
}

The viewport and displayport metrics for the painted frame at the time of a layer-tree transaction. These metrics are especially useful for shadow layers, because the metrics values are updated atomically with new pixels.

Fields
mScrollId: FrameMetrics_ViewID mPresShellResolution: f32 mCompositionBounds: ParentLayerRect mDisplayPort: CSSRect mCriticalDisplayPort: CSSRect mScrollableRect: CSSRect mCumulativeResolution: LayoutDeviceToLayerScale2D mDevPixelsPerCSSPixel: CSSToLayoutDeviceScale mScrollOffset: CSSPoint mZoom: CSSToParentLayerScale2D mScrollGeneration: u32 mSmoothScrollOffset: CSSPoint mRootCompositionSize: CSSSize mDisplayPortMargins: ScreenMargin mPresShellId: u32 mViewport: CSSRect mExtraResolution: ScreenToLayerScale2D mPaintRequestTime: TimeStamp mScrollUpdateType: FrameMetrics_ScrollOffsetUpdateType _bitfield_1: u8
Methods

impl FrameMetrics

fn mIsRootContent(&self) -> bool

fn set_mIsRootContent(&mut self, val: bool)

fn mDoSmoothScroll(&self) -> bool

fn set_mDoSmoothScroll(&mut self, val: bool)

fn mUseDisplayPortMargins(&self) -> bool

fn set_mUseDisplayPortMargins(&mut self, val: bool)

fn mIsScrollInfoLayer(&self) -> bool

fn set_mIsScrollInfoLayer(&mut self, val: bool)