Struct style::gecko_bindings::structs::root::mozilla::BlockingResourceBase

#[repr(C)]

pub struct BlockingResourceBase {
    pub mChainPrev: *mut BlockingResourceBase,
    pub mName: *const c_char,
    pub mType: BlockingResourceBase_BlockingResourceType,
    pub mAcquired: BlockingResourceBase_AcquisitionState,
}

BlockingResourceBase Base class of resources that might block clients trying to acquire them. Does debugging and deadlock detection in DEBUG builds.

Fields
mChainPrev: *mut BlockingResourceBase

mChainPrev A series of resource acquisitions creates a chain of orders. This chain is implemented as a linked list; |mChainPrev| points to the resource most recently Acquire()'d before this one.

mName: *const c_char

mName A descriptive name for this resource. Used in error messages etc.

mType: BlockingResourceBase_BlockingResourceType

mType The more specific type of this resource. Used to implement special semantics (e.g., reentrancy of monitors).

mAcquired: BlockingResourceBase_AcquisitionState

mAcquired Indicates if this resource is currently acquired.