Struct style::gecko_bindings::structs::root::mozilla::Pair

#[repr(C)]

pub struct Pair<A, B> {
    pub _address: u8,
    pub _phantom_0: PhantomData<A>,
    pub _phantom_1: PhantomData<B>,
}

Pair is the logical concatenation of an instance of A with an instance B. Space is conserved when possible. Neither A nor B may be a final class.

It's typically clearer to have individual A and B member fields. Except if you want the space-conserving qualities of Pair, you're probably better off not using this!

No guarantees are provided about the memory layout of A and B, the order of initialization or destruction of A and B, and so on. (This is approximately required to optimize space usage.) The first/second names are merely conceptual!

Fields
_address: u8 _phantom_0: PhantomData<A> _phantom_1: PhantomData<B>