Struct style::gecko_bindings::structs::root::nsIPresShell

#[repr(C)]

pub struct nsIPresShell {
    pub _base: nsISupports,
    pub mDocument: nsCOMPtr<nsIDocument>,
    pub mPresContext: RefPtr<nsPresContext>,
    pub mStyleSet: StyleSetHandle,
    pub mFrameConstructor: *mut nsCSSFrameConstructor,
    pub mViewManager: *mut nsViewManager,
    pub mFrameArena: nsPresArena,
    pub mSelection: RefPtr<nsFrameSelection>,
    pub mFrameManager: *mut nsFrameManagerBase,
    pub mForwardingContainer: u64,
    pub mHiddenInvalidationObserverRefreshDriver: *mut nsRefreshDriver,
    pub mDocAccessible: *mut DocAccessible,
    pub mReflowContinueTimer: nsCOMPtr<nsITimer>,
    pub mDrawEventTargetFrame: *mut nsIFrame,
    pub mAllocatedPointers: [u64; 6],
    pub mPaintCount: u64,
    pub mScrollPositionClampingScrollPortSize: nsSize,
    pub mWeakFrames: *mut nsWeakFrame,
    pub mCanvasBackgroundColor: nscolor,
    pub mResolution: [u64; 2],
    pub mSelectionFlags: i16,
    pub mRenderFlags: nsIPresShell_RenderFlags,
    pub _bitfield_1: u16,
    pub mPresShellId: u32,
    pub mChangedScopeStyleRoots: [u64; 3],
    pub mFontSizeInflationEmPerLine: u32,
    pub mFontSizeInflationMinTwips: u32,
    pub mFontSizeInflationLineThreshold: u32,
    pub mFontSizeInflationForceEnabled: bool,
    pub mFontSizeInflationDisabledInMasterProcess: bool,
    pub mFontSizeInflationEnabled: bool,
    pub mPaintingIsFrozen: bool,
    pub mFontSizeInflationEnabledIsDirty: bool,
    pub mIsNeverPainting: bool,
}

Presentation shell interface. Presentation shells are the controlling point for managing the presentation of a document. The presentation shell holds a live reference to the document, the presentation context, the style manager, the style set and the root frame.

When this object is Release'd, it will release the document, the presentation context, the style manager, the style set and the root frame.

Fields
_base: nsISupports mDocument: nsCOMPtr<nsIDocument> mPresContext: RefPtr<nsPresContext> mStyleSet: StyleSetHandle mFrameConstructor: *mut nsCSSFrameConstructor mViewManager: *mut nsViewManager mFrameArena: nsPresArena mSelection: RefPtr<nsFrameSelection> mFrameManager: *mut nsFrameManagerBase mForwardingContainer: u64 mHiddenInvalidationObserverRefreshDriver: *mut nsRefreshDriver mDocAccessible: *mut DocAccessible mReflowContinueTimer: nsCOMPtr<nsITimer> mDrawEventTargetFrame: *mut nsIFrame mAllocatedPointers: [u64; 6] mPaintCount: u64 mScrollPositionClampingScrollPortSize: nsSize mWeakFrames: *mut nsWeakFrame mCanvasBackgroundColor: nscolor mResolution: [u64; 2] mSelectionFlags: i16 mRenderFlags: nsIPresShell_RenderFlags _bitfield_1: u16 mPresShellId: u32 mChangedScopeStyleRoots: [u64; 3] mFontSizeInflationEmPerLine: u32 mFontSizeInflationMinTwips: u32 mFontSizeInflationLineThreshold: u32 mFontSizeInflationForceEnabled: bool mFontSizeInflationDisabledInMasterProcess: bool mFontSizeInflationEnabled: bool mPaintingIsFrozen: bool mFontSizeInflationEnabledIsDirty: bool mIsNeverPainting: bool
Methods

impl nsIPresShell

fn mStylesHaveChanged(&self) -> bool

fn set_mStylesHaveChanged(&mut self, val: bool)

fn mDidInitialize(&self) -> bool

fn set_mDidInitialize(&mut self, val: bool)

fn mIsDestroying(&self) -> bool

fn set_mIsDestroying(&mut self, val: bool)

fn mIsReflowing(&self) -> bool

fn set_mIsReflowing(&mut self, val: bool)

fn mPaintingSuppressed(&self) -> bool

fn set_mPaintingSuppressed(&mut self, val: bool)

fn mIsThemeSupportDisabled(&self) -> bool

fn set_mIsThemeSupportDisabled(&mut self, val: bool)

fn mIsActive(&self) -> bool

fn set_mIsActive(&mut self, val: bool)

fn mFrozen(&self) -> bool

fn set_mFrozen(&mut self, val: bool)

fn mIsFirstPaint(&self) -> bool

fn set_mIsFirstPaint(&mut self, val: bool)

fn mObservesMutationsForPrint(&self) -> bool

fn set_mObservesMutationsForPrint(&mut self, val: bool)

fn mReflowScheduled(&self) -> bool

fn set_mReflowScheduled(&mut self, val: bool)

fn mSuppressInterruptibleReflows(&self) -> bool

fn set_mSuppressInterruptibleReflows(&mut self, val: bool)

fn mScrollPositionClampingScrollPortSizeSet(&self) -> bool

fn set_mScrollPositionClampingScrollPortSizeSet(&mut self, val: bool)