Struct style::gecko_bindings::structs::root::nsMainThreadPtrHolder

#[repr(C)]

pub struct nsMainThreadPtrHolder<T> {
    pub mRefCnt: ThreadSafeAutoRefCnt,
    pub mRawPtr: *mut T,
    pub mStrict: bool,
}

Class to safely handle main-thread-only pointers off the main thread.

Classes like XPCWrappedJS are main-thread-only, which means that it is forbidden to call methods on instances of these classes off the main thread. For various reasons (see bug 771074), this restriction recently began to apply to AddRef/Release as well.

This presents a problem for consumers that wish to hold a callback alive on non-main-thread code. A common example of this is the proxy callback pattern, where non-main-thread code holds a strong-reference to the callback object, and dispatches new Runnables (also with a strong reference) to the main thread in order to execute the callback. This involves several AddRef and Release calls on the other thread, which is (now) verboten.

The basic idea of this class is to introduce a layer of indirection. nsMainThreadPtrHolder is a threadsafe reference-counted class that internally maintains one strong reference to the main-thread-only object. It must be instantiated on the main thread (so that the AddRef of the underlying object happens on the main thread), but consumers may subsequently pass references to the holder anywhere they please. These references are meant to be opaque when accessed off-main-thread (assertions enforce this).

The semantics of RefPtr > would be cumbersome, so we also introduce nsMainThreadPtrHandle, which is conceptually identical to the above (though it includes various convenience methods). The basic pattern is as follows.

// On the main thread: nsCOMPtr callback = ...; nsMainThreadPtrHandle callbackHandle = new nsMainThreadPtrHolder(callback); // Pass callbackHandle to structs/classes that might be accessed on other // threads.

All structs and classes that might be accessed on other threads should store an nsMainThreadPtrHandle rather than an nsCOMPtr.

Fields
mRefCnt: ThreadSafeAutoRefCnt mRawPtr: *mut T mStrict: bool