Struct style::gecko_bindings::structs::root::nsStyleAutoArray

#[repr(C)]

pub struct nsStyleAutoArray<T> {
    pub mFirstElement: T,
    pub mOtherElements: nsTArray<T>,
}

An array of objects, similar to AutoTArray but which is memmovable. It always has length >= 1.

Fields
mFirstElement: T mOtherElements: nsTArray<T>
Methods

impl<T> nsStyleAutoArray<T>

fn iter_mut(&mut self) -> Chain<Once<&mut T>, IterMut<T>>

Mutably iterate over the array elements.

fn iter(&self) -> Chain<Once<&T>, Iter<T>>

Iterate over the array elements.

fn len(&self) -> usize

Returns the length of the array.

Note that often structs containing autoarrays will have additional member fields that contain the length, which must be kept in sync.

fn ensure_len(&mut self, len: usize)

Ensures that the array has length at least the given length.