Struct style::gecko_bindings::structs::root::nsStyleCoord

#[repr(C)]

pub struct nsStyleCoord { /* fields omitted */ }

Class that hold a single size specification used by the style system. The size specification consists of two parts -- a number and a unit. The number is an integer, a floating point value, an nscoord, or undefined, and the unit is an nsStyleUnit. Checking the unit is a must before asking for the value in any particular form.

Methods

impl nsStyleCoord

fn set<T: GeckoStyleCoordConvertible>(&mut self, val: T)

Set this nsStyleCoord value to val.

impl nsStyleCoord

unsafe fn get_mUnit(&self) -> &nsStyleUnit

unsafe fn get_mUnit_mut(&mut self) -> &mut nsStyleUnit

unsafe fn get_mValue(&self) -> &nsStyleUnion

unsafe fn get_mValue_mut(&mut self) -> &mut nsStyleUnion

impl nsStyleCoord

fn null() -> Self

Get a null nsStyleCoord.