Struct style::gecko_bindings::structs::root::nsStyleSides

#[repr(C)]

pub struct nsStyleSides { /* fields omitted */ }

Class that represents a set of top/right/bottom/left nsStyleCoords. This is commonly used to hold the widths of the borders, margins, or paddings of a box.

Methods

impl nsStyleSides

unsafe fn get_mUnits(&self) -> &[nsStyleUnit; 4]

unsafe fn get_mUnits_mut(&mut self) -> &mut [nsStyleUnit; 4]

unsafe fn get_mValues(&self) -> &[nsStyleUnion; 4]

unsafe fn get_mValues_mut(&mut self) -> &mut [nsStyleUnion; 4]

impl nsStyleSides

fn data_at(&self, index: usize) -> SidesData

Immutably get the nsStyleCoord-like object representing the side at index index.

fn data_at_mut(&mut self, index: usize) -> SidesDataMut

Mutably get the nsStyleCoord-like object representing the side at index index.