Struct style::gecko_bindings::structs::root::nsWrapperCache

#[repr(C)]

pub struct nsWrapperCache {
    pub vtable_: *const nsWrapperCache__bindgen_vtable,
    pub mWrapper: *mut JSObject,
    pub mFlags: nsWrapperCache_FlagsType,
}

Class to store the wrapper for an object. This can only be used with objects that only have one non-security wrapper at a time (for an XPCWrappedNative this is usually ensured by setting an explicit parent in the PreCreate hook for the class).

An instance of nsWrapperCache can be gotten from an object that implements a wrapper cache by calling QueryInterface on it. Note that this breaks XPCOM rules a bit (this object doesn't derive from nsISupports).

The cache can store objects other than wrappers. We allow wrappers to use a separate JSObject to store their state (mostly expandos). If the wrapper is collected and we want to preserve this state we actually store the state object in the cache.

The cache can store 2 types of objects:

If WRAPPER_IS_NOT_DOM_BINDING is set (IsDOMBinding() returns false): - the JSObject of an XPCWrappedNative wrapper

If WRAPPER_IS_NOT_DOM_BINDING is not set (IsDOMBinding() returns true): - a DOM binding object (regular JS object or proxy)

The finalizer for the wrapper clears the cache.

A compacting GC can move the wrapper object. Pointers to moved objects are usually found and updated by tracing the heap, however non-preserved wrappers are weak references and are not traced, so another approach is necessary. Instead a class hook (objectMovedOp) is provided that is called when an object is moved and is responsible for ensuring pointers are updated. It does this by calling UpdateWrapper() on the wrapper cache. SetWrapper() asserts that the hook is implemented for any wrapper set.

A number of the methods are implemented in nsWrapperCacheInlines.h because we have to include some JS headers that don't play nicely with the rest of the codebase. Include nsWrapperCacheInlines.h if you need to call those methods.

Fields
vtable_: *const nsWrapperCache__bindgen_vtable mWrapper: *mut JSObject mFlags: nsWrapperCache_FlagsType