Trait style::gecko_bindings::sugar::ns_style_coord::CoordData

pub trait CoordData {
    fn unit(&self) -> nsStyleUnit;
    fn union(&self) -> nsStyleUnion;

    fn as_value(&self) -> CoordDataValue { ... }
    unsafe fn get_float(&self) -> f32 { ... }
    unsafe fn get_integer(&self) -> i32 { ... }
    unsafe fn get_calc_value(&self) -> nsStyleCoord_CalcValue { ... }
    unsafe fn as_calc(&self) -> &nsStyleCoord_Calc { ... }
}

A trait to abstract on top of a nsStyleCoord-like object.

Required Methods

fn unit(&self) -> nsStyleUnit

Get the unit of this object.

fn union(&self) -> nsStyleUnion

Get the nsStyleUnion for this object.

Provided Methods

fn as_value(&self) -> CoordDataValue

Get the appropriate value for this object.

unsafe fn get_float(&self) -> f32

Pretend inner value is a float; obtain it.

unsafe fn get_integer(&self) -> i32

Pretend inner value is an int; obtain it.

unsafe fn get_calc_value(&self) -> nsStyleCoord_CalcValue

Pretend inner value is a calc; obtain it. Ensure that the unit is Calc before calling this.

unsafe fn as_calc(&self) -> &nsStyleCoord_Calc

Pretend the inner value is a calc expression, and obtain it.