Trait style::gecko_bindings::sugar::ns_style_coord::CoordDataMut

pub trait CoordDataMut: CoordData {
    unsafe fn values_mut(&mut self) -> (&mut nsStyleUnit, &mut nsStyleUnion);

    fn reset(&mut self) { ... }
    fn copy_from<T: CoordData>(&mut self, other: &T) { ... }
    unsafe fn copy_from_unchecked<T: CoordData>(&mut self, other: &T) { ... }
    fn leaky_set_null(&mut self) { ... }
    fn set_value(&mut self, value: CoordDataValue) { ... }
    unsafe fn as_calc_mut(&mut self) -> &mut nsStyleCoord_Calc { ... }
    fn addref_if_calc(&mut self) { ... }
}

A trait to abstract on top of a mutable nsStyleCoord-like object.

Required Methods

unsafe fn values_mut(&mut self) -> (&mut nsStyleUnit, &mut nsStyleUnion)

Get mutably the unit and the union.

This is unsafe since it's possible to modify the unit without changing the union.

NB: This can't be two methods since we can't mutably borrow twice

Provided Methods

fn reset(&mut self)

Clean up any resources used by the union.

Currently, this only happens if the nsStyleUnit is a Calc.

fn copy_from<T: CoordData>(&mut self, other: &T)

Copies the unit and value from another CoordData type.

unsafe fn copy_from_unchecked<T: CoordData>(&mut self, other: &T)

Copies the unit and value from another CoordData type without checking the type of the value (so refcounted values like calc may leak).

fn leaky_set_null(&mut self)

Useful for initializing uninits, given that set_value may segfault on uninits.

fn set_value(&mut self, value: CoordDataValue)

Sets the inner value.

unsafe fn as_calc_mut(&mut self) -> &mut nsStyleCoord_Calc

Gets the Calc value mutably, asserts in debug builds if the unit is not Calc.

fn addref_if_calc(&mut self)

Does what it promises, if the unit is calc, it bumps the reference count of the calc expression.