Module style::gecko_bindings::sugar::ownership

Helpers for different FFI pointer kinds that Gecko's FFI layer uses.

Structs

Owned
Gecko-FFI-safe owned pointer.

OwnedOrNull
Gecko-FFI-safe owned pointer.

Strong
Gecko-FFI-safe Arc (T is an ArcInner).

Traits

FFIArcHelpers
A few helpers implemented on top of Arc to make it more comfortable to use and write safe code with.

HasArcFFI
Helper trait for conversions between FFI Strong/Borrowed types and Arcs

HasBoxFFI
Indicates that the given Servo type is passed over FFI as a Box

HasFFI
Indicates that a given Servo type has a corresponding Gecko FFI type.

HasSimpleFFI
Indicates that a given Servo type has the same layout as the corresponding HasFFI::FFIType type.