Struct style::gecko_bindings::sugar::ownership::OwnedOrNull

#[repr(C)]

pub struct OwnedOrNull<GeckoType> { /* fields omitted */ }

Gecko-FFI-safe owned pointer.

Can be null, and just as Owned leaks on Drop.

Methods

impl<GeckoType> OwnedOrNull<GeckoType>

fn is_null(&self) -> bool

Returns whether this pointer is null.

fn into_box_opt<ServoType>(self) -> Option<Box<ServoType>> where ServoType: HasBoxFFI<FFIType=GeckoType>

Returns an owned pointer if this is non-null, and None otherwise.

fn into_owned_opt(self) -> Option<Owned<GeckoType>>

Returns an Owned<GeckoType> if non-null, None otherwise.

fn borrow(&self) -> Option<&GeckoType>

Gets a immutable reference to the underlying Gecko type, or None if null.

fn borrow_mut(&self) -> Option<&mut GeckoType>

Gets a mutable reference to the underlying Gecko type, or None if null.