Trait style::gecko_bindings::sugar::ownership::FFIArcHelpers

pub unsafe trait FFIArcHelpers {
    type Inner: HasArcFFI;
    fn into_strong(self) -> Strong<Self::Inner::FFIType>;
    fn as_borrowed_opt(&self) -> Option<&Self::Inner::FFIType>;
}

A few helpers implemented on top of Arc to make it more comfortable to use and write safe code with.

Associated Types

type Inner: HasArcFFI

The Rust FFI type that we're implementing methods for.

Required Methods

fn into_strong(self) -> Strong<Self::Inner::FFIType>

Converts an Arc into a strong FFI reference.

Arc -> Strong

fn as_borrowed_opt(&self) -> Option<&Self::Inner::FFIType>

Produces a (nullable) borrowed FFI reference by borrowing an Arc.

&Arc -> Option<&GeckoType>

FIXME(emilio): What's the point of the nullability? Arc should be non-null, right?

Then the arc_as_borrowed method can go away.