Trait style::gecko_bindings::sugar::ownership::HasArcFFI

pub unsafe trait HasArcFFI: HasFFI {
    unsafe fn addref_opt(ptr: Option<&Self::FFIType>) { ... }
    unsafe fn release_opt(ptr: Option<&Self::FFIType>) { ... }
    unsafe fn addref(ptr: &Self::FFIType) { ... }
    unsafe fn release(ptr: &Self::FFIType) { ... }
    fn as_arc<'a>(ptr: &'a &Self::FFIType) -> &'a Arc<Self> { ... }
    fn arc_as_borrowed<'a>(arc: &'a Arc<Self>) -> &'a &Self::FFIType { ... }
    fn arc_from_borrowed<'a>(ptr: &'a Option<&Self::FFIType>)
                             -> Option<&'a Arc<Self>> { ... }
}

Helper trait for conversions between FFI Strong/Borrowed types and Arcs

Should be implemented by types which are passed over FFI as Arcs via Strong and Borrowed.

In this case, the FFIType is the rough equivalent of ArcInner.

Provided Methods

unsafe fn addref_opt(ptr: Option<&Self::FFIType>)

Artificially increments the refcount of a (possibly null) borrowed Arc over FFI.

unsafe fn release_opt(ptr: Option<&Self::FFIType>)

Given a (possibly null) borrowed FFI reference, decrements the refcount. Unsafe since it doesn't consume the backing Arc. Run it only when you know that a strong reference to the backing Arc is disappearing (usually on the C++ side) without running the Arc destructor.

unsafe fn addref(ptr: &Self::FFIType)

Artificially increments the refcount of a borrowed Arc over FFI.

unsafe fn release(ptr: &Self::FFIType)

Given a non-null borrowed FFI reference, decrements the refcount. Unsafe since it doesn't consume the backing Arc. Run it only when you know that a strong reference to the backing Arc is disappearing (usually on the C++ side) without running the Arc destructor.

fn as_arc<'a>(ptr: &'a &Self::FFIType) -> &'a Arc<Self>

Converts a borrowed FFI reference to a borrowed Arc.

&GeckoType -> &Arc

fn arc_as_borrowed<'a>(arc: &'a Arc<Self>) -> &'a &Self::FFIType

Converts a borrowed Arc to a borrowed FFI reference.

&Arc -> &GeckoType

fn arc_from_borrowed<'a>(ptr: &'a Option<&Self::FFIType>)
-> Option<&'a Arc<Self>>

Converts a borrowed nullable FFI reference to a borrowed Arc.

&GeckoType -> &Arc