Trait style::gecko_bindings::sugar::ownership::HasSimpleFFI

pub unsafe trait HasSimpleFFI: HasFFI {
    fn as_ffi(&self) -> &Self::FFIType { ... }
    fn as_ffi_mut(&mut self) -> &mut Self::FFIType { ... }
    fn from_ffi(ffi: &Self::FFIType) -> &Self { ... }
    fn from_ffi_mut(ffi: &mut Self::FFIType) -> &mut Self { ... }
}

Indicates that a given Servo type has the same layout as the corresponding HasFFI::FFIType type.

Provided Methods

fn as_ffi(&self) -> &Self::FFIType

Given a Servo-side reference, converts it to an FFI-safe reference which can be passed to Gecko.

&ServoType -> &GeckoType

fn as_ffi_mut(&mut self) -> &mut Self::FFIType

Given a Servo-side mutable reference, converts it to an FFI-safe mutable reference which can be passed to Gecko.

&mut ServoType -> &mut GeckoType

fn from_ffi(ffi: &Self::FFIType) -> &Self

Given an FFI-safe reference obtained from Gecko converts it to a Servo-side reference.

&GeckoType -> &ServoType

fn from_ffi_mut(ffi: &mut Self::FFIType) -> &mut Self

Given an FFI-safe mutable reference obtained from Gecko converts it to a Servo-side mutable reference.

&mut GeckoType -> &mut ServoType