Module style::gecko_bindings::sugar::refptr

A rust helper to ease the use of Gecko's refcounted types.

Structs

RefPtr
A custom RefPtr implementation to take into account Drop semantics and a bit less-painful memory management.

UniqueRefPtr
A RefPtr that we know is uniquely owned.

Traits

RefCounted
Trait for all objects that have Addref() and Release methods and can be placed inside RefPtr

ThreadSafeRefCounted
Trait for types which can be shared across threads in RefPtr.

Type Definitions

GeckoArcPrincipal
A Gecko ThreadSafePrincipalHolder wrapped in a safe refcounted pointer, to use during stylesheet parsing and style computation.

GeckoArcURI
A Gecko ThreadSafeURIHolder wrapped in a safe refcounted pointer, to use during stylesheet parsing and style computation.