Struct style::gecko_bindings::sugar::refptr::RefPtr

pub struct RefPtr<T: RefCounted> { /* fields omitted */ }

A custom RefPtr implementation to take into account Drop semantics and a bit less-painful memory management.

Methods

impl<T: RefCounted> RefPtr<T>

unsafe fn from_addrefed(ptr: *mut T) -> Self

Create a new RefPtr from an already addrefed pointer obtained from FFI.

The pointer must be valid, non-null and have been addrefed.

unsafe fn new(ptr: *mut T) -> Self

Create a new RefPtr from a pointer obtained from FFI.

The pointer must be valid and non null.

This method calls addref() internally

fn forget(self) -> RefPtr<T>

Produces an FFI-compatible RefPtr that can be stored in style structs.

structs::RefPtr does not have a destructor, so this may leak

fn get(&self) -> *mut T

Returns the raw inner pointer to be fed back into FFI.

fn addref(&self)

Addref the inner data, obviously leaky on its own.

unsafe fn release(&self)

Release the inner data.

Call only when the data actually needs releasing.