Struct style::gecko_properties::GeckoBorder

pub struct GeckoBorder { /* fields omitted */ }

Methods

impl GeckoBorder

fn set_border_top_style(&mut self, v: T)

fn copy_border_top_style_from(&mut self, other: &Self)

fn clone_border_top_style(&self) -> T

fn set_border_top_color(&mut self, v: T)

fn copy_border_top_color_from(&mut self, other: &Self)

fn clone_border_top_color(&self) -> T

fn set_border_top_width(&mut self, v: T)

fn copy_border_top_width_from(&mut self, other: &Self)

fn clone_border_top_width(&self) -> T

fn border_top_has_nonzero_width(&self) -> bool

fn set_border_right_style(&mut self, v: T)

fn copy_border_right_style_from(&mut self, other: &Self)

fn clone_border_right_style(&self) -> T

fn set_border_right_color(&mut self, v: T)

fn copy_border_right_color_from(&mut self, other: &Self)

fn clone_border_right_color(&self) -> T

fn set_border_right_width(&mut self, v: T)

fn copy_border_right_width_from(&mut self, other: &Self)

fn clone_border_right_width(&self) -> T

fn border_right_has_nonzero_width(&self) -> bool

fn set_border_bottom_style(&mut self, v: T)

fn copy_border_bottom_style_from(&mut self, other: &Self)

fn clone_border_bottom_style(&self) -> T

fn set_border_bottom_color(&mut self, v: T)

fn copy_border_bottom_color_from(&mut self, other: &Self)

fn clone_border_bottom_color(&self) -> T

fn set_border_bottom_width(&mut self, v: T)

fn copy_border_bottom_width_from(&mut self, other: &Self)

fn clone_border_bottom_width(&self) -> T

fn border_bottom_has_nonzero_width(&self) -> bool

fn set_border_left_style(&mut self, v: T)

fn copy_border_left_style_from(&mut self, other: &Self)

fn clone_border_left_style(&self) -> T

fn set_border_left_color(&mut self, v: T)

fn copy_border_left_color_from(&mut self, other: &Self)

fn clone_border_left_color(&self) -> T

fn set_border_left_width(&mut self, v: T)

fn copy_border_left_width_from(&mut self, other: &Self)

fn clone_border_left_width(&self) -> T

fn border_left_has_nonzero_width(&self) -> bool

fn set_border_top_left_radius(&mut self, v: T)

fn copy_border_top_left_radius_from(&mut self, other: &Self)

fn clone_border_top_left_radius(&self) -> T

fn set_border_top_right_radius(&mut self, v: T)

fn copy_border_top_right_radius_from(&mut self, other: &Self)

fn clone_border_top_right_radius(&self) -> T

fn set_border_bottom_right_radius(&mut self, v: T)

fn copy_border_bottom_right_radius_from(&mut self, other: &Self)

fn clone_border_bottom_right_radius(&self) -> T

fn set_border_bottom_left_radius(&mut self, v: T)

fn copy_border_bottom_left_radius_from(&mut self, other: &Self)

fn clone_border_bottom_left_radius(&self) -> T

fn set_border_image_source(&mut self, v: T)

fn copy_border_image_source_from(&mut self, other: &Self)

fn set_border_image_outset(&mut self, v: T)

fn copy_border_image_outset_from(&mut self, other: &Self)

fn set_border_image_repeat(&mut self, v: T)

fn copy_border_image_repeat_from(&mut self, other: &Self)

fn set_border_image_width(&mut self, v: T)

fn copy_border_image_width_from(&mut self, other: &Self)

fn set_border_image_slice(&mut self, v: T)

fn copy_border_image_slice_from(&mut self, other: &Self)

fn set_border_block_start_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-start-color given a writing mode.

fn copy_border_block_start_color_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-start-color given a writing mode.

fn set_border_block_end_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-end-color given a writing mode.

fn copy_border_block_end_color_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-end-color given a writing mode.

fn set_border_inline_start_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-start-color given a writing mode.

fn copy_border_inline_start_color_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-start-color given a writing mode.

fn set_border_inline_end_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-end-color given a writing mode.

fn copy_border_inline_end_color_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-end-color given a writing mode.

fn set_border_block_start_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-start-style given a writing mode.

fn copy_border_block_start_style_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-start-style given a writing mode.

fn clone_border_block_start_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-start-style given a writing mode.

fn set_border_block_end_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-end-style given a writing mode.

fn copy_border_block_end_style_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-end-style given a writing mode.

fn clone_border_block_end_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-end-style given a writing mode.

fn set_border_inline_start_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-start-style given a writing mode.

fn copy_border_inline_start_style_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-start-style given a writing mode.

fn clone_border_inline_start_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-start-style given a writing mode.

fn set_border_inline_end_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-end-style given a writing mode.

fn copy_border_inline_end_style_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-end-style given a writing mode.

fn clone_border_inline_end_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-end-style given a writing mode.

fn set_border_block_start_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-start-width given a writing mode.

fn copy_border_block_start_width_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-start-width given a writing mode.

fn set_border_block_end_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-end-width given a writing mode.

fn copy_border_block_end_width_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-end-width given a writing mode.

fn set_border_inline_start_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-start-width given a writing mode.

fn copy_border_inline_start_width_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-start-width given a writing mode.

fn set_border_inline_end_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-end-width given a writing mode.

fn copy_border_inline_end_width_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-end-width given a writing mode.

fn set_box_decoration_break(&mut self, v: T)

fn copy_box_decoration_break_from(&mut self, other: &Self)

fn set__moz_float_edge(&mut self, v: T)

fn copy__moz_float_edge_from(&mut self, other: &Self)

impl GeckoBorder

fn default(pres_context: RawGeckoPresContextBorrowed) -> Arc<Self>

fn get_gecko(&self) -> &nsStyleBorder