Struct style::gecko_properties::GeckoBox

pub struct GeckoBox { /* fields omitted */ }

Methods

impl Box

fn transition_duration_iter(&self) -> TransitionDurationIter

Iterate over the values of transition-duration.

fn transition_duration_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-duration.

fn transition_timing_function_iter(&self) -> TransitionTimingFunctionIter

Iterate over the values of transition-timing-function.

fn transition_timing_function_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-timing-function.

fn transition_property_iter(&self) -> TransitionPropertyIter

Iterate over the values of transition-property.

fn transition_property_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-property.

fn transition_delay_iter(&self) -> TransitionDelayIter

Iterate over the values of transition-delay.

fn transition_delay_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-delay.

fn animation_name_iter(&self) -> AnimationNameIter

Iterate over the values of animation-name.

fn animation_name_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-name.

fn animation_duration_iter(&self) -> AnimationDurationIter

Iterate over the values of animation-duration.

fn animation_duration_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-duration.

fn animation_timing_function_iter(&self) -> AnimationTimingFunctionIter

Iterate over the values of animation-timing-function.

fn animation_timing_function_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-timing-function.

fn animation_iteration_count_iter(&self) -> AnimationIterationCountIter

Iterate over the values of animation-iteration-count.

fn animation_iteration_count_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-iteration-count.

fn animation_direction_iter(&self) -> AnimationDirectionIter

Iterate over the values of animation-direction.

fn animation_direction_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-direction.

fn animation_play_state_iter(&self) -> AnimationPlayStateIter

Iterate over the values of animation-play-state.

fn animation_play_state_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-play-state.

fn animation_fill_mode_iter(&self) -> AnimationFillModeIter

Iterate over the values of animation-fill-mode.

fn animation_fill_mode_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-fill-mode.

fn animation_delay_iter(&self) -> AnimationDelayIter

Iterate over the values of animation-delay.

fn animation_delay_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-delay.

impl GeckoBox

fn set_display(&mut self, v: T)

fn copy_display_from(&mut self, other: &Self)

fn clone_display(&self) -> T

fn set_overflow_y(&mut self, v: T)

fn copy_overflow_y_from(&mut self, other: &Self)

fn clone_overflow_y(&self) -> T

fn set_vertical_align(&mut self, v: T)

fn clone_vertical_align(&self) -> T

fn copy_vertical_align_from(&mut self, other: &Self)

fn set__moz_binding(&mut self, v: T)

fn copy__moz_binding_from(&mut self, other: &Self)

fn set_page_break_before(&mut self, v: T)

fn copy_page_break_before_from(&mut self, other: &Self)

fn set_page_break_after(&mut self, v: T)

fn copy_page_break_after_from(&mut self, other: &Self)

fn set_scroll_snap_points_x(&mut self, v: T)

fn copy_scroll_snap_points_x_from(&mut self, other: &Self)

fn set_scroll_snap_points_y(&mut self, v: T)

fn copy_scroll_snap_points_y_from(&mut self, other: &Self)

fn set_scroll_snap_coordinate(&mut self, v: T)

fn copy_scroll_snap_coordinate_from(&mut self, other: &Self)

fn clone_scroll_snap_coordinate(&self) -> T

fn convert_transform(input: Vec<ComputedOperation>,
output: &mut RefPtr<nsCSSValueSharedList>)

fn set_transform(&mut self, other: T)

fn copy_transform_from(&mut self, other: &Self)

fn clone_transform(&self) -> T

fn set_animation_name(&mut self, v: T)

fn animation_name_at(&self, index: usize) -> SingleComputedValue

fn copy_animation_name_from(&mut self, other: &Self)

fn animation_name_count(&self) -> usize

fn set_animation_delay(&mut self, v: T)

fn animation_delay_at(&self, index: usize) -> SingleComputedValue

fn animation_delay_count(&self) -> usize

fn copy_animation_delay_from(&mut self, other: &Self)

fn set_animation_duration(&mut self, v: T)

fn animation_duration_at(&self, index: usize) -> SingleComputedValue

fn animation_duration_count(&self) -> usize

fn copy_animation_duration_from(&mut self, other: &Self)

fn set_animation_direction(&mut self, v: T)

fn animation_direction_at(&self, index: usize) -> SingleComputedValue

fn animation_direction_count(&self) -> usize

fn copy_animation_direction_from(&mut self, other: &Self)

fn set_animation_fill_mode(&mut self, v: T)

fn animation_fill_mode_at(&self, index: usize) -> SingleComputedValue

fn animation_fill_mode_count(&self) -> usize

fn copy_animation_fill_mode_from(&mut self, other: &Self)

fn set_animation_play_state(&mut self, v: T)

fn animation_play_state_at(&self, index: usize) -> SingleComputedValue

fn animation_play_state_count(&self) -> usize

fn copy_animation_play_state_from(&mut self, other: &Self)

fn set_animation_iteration_count(&mut self, v: T)

fn animation_iteration_count_at(&self, index: usize) -> SingleComputedValue

fn animation_iteration_count_count(&self) -> usize

fn copy_animation_iteration_count_from(&mut self, other: &Self)

fn set_animation_timing_function(&mut self, v: T)

fn animation_timing_function_count(&self) -> usize

fn copy_animation_timing_function_from(&mut self, other: &Self)

fn animation_timing_function_at(&self, index: usize) -> SingleComputedValue

fn set_scroll_snap_type_y(&mut self, v: T)

fn copy_scroll_snap_type_y_from(&mut self, other: &Self)

fn set_perspective_origin(&mut self, v: T)

fn copy_perspective_origin_from(&mut self, other: &Self)

fn set_transform_origin(&mut self, v: T)

fn copy_transform_origin_from(&mut self, other: &Self)

fn clone_transform_origin(&self) -> T

fn set__moz_top_layer(&mut self, v: T)

fn copy__moz_top_layer_from(&mut self, other: &Self)

fn clone__moz_top_layer(&self) -> T

fn set_position(&mut self, v: T)

fn copy_position_from(&mut self, other: &Self)

fn clone_position(&self) -> T

fn set_float(&mut self, v: T)

fn copy_float_from(&mut self, other: &Self)

fn clone_float(&self) -> T

fn set_clear(&mut self, v: T)

fn copy_clear_from(&mut self, other: &Self)

fn set_overflow_clip_box(&mut self, v: T)

fn copy_overflow_clip_box_from(&mut self, other: &Self)

fn set_overflow_x(&mut self, v: T)

fn copy_overflow_x_from(&mut self, other: &Self)

fn clone_overflow_x(&self) -> T

fn set_scroll_snap_destination(&mut self, v: T)

fn copy_scroll_snap_destination_from(&mut self, other: &Self)

fn clone_scroll_snap_destination(&self) -> T

fn set_scroll_behavior(&mut self, v: T)

fn copy_scroll_behavior_from(&mut self, other: &Self)

fn set_scroll_snap_type_x(&mut self, v: T)

fn copy_scroll_snap_type_x_from(&mut self, other: &Self)

fn set_isolation(&mut self, v: T)

fn copy_isolation_from(&mut self, other: &Self)

fn set_page_break_inside(&mut self, v: T)

fn copy_page_break_inside_from(&mut self, other: &Self)

fn set_resize(&mut self, v: T)

fn copy_resize_from(&mut self, other: &Self)

fn set_perspective(&mut self, v: T)

fn copy_perspective_from(&mut self, other: &Self)

fn clone_perspective(&self) -> T

fn set_backface_visibility(&mut self, v: T)

fn copy_backface_visibility_from(&mut self, other: &Self)

fn set_transform_box(&mut self, v: T)

fn copy_transform_box_from(&mut self, other: &Self)

fn set_transform_style(&mut self, v: T)

fn copy_transform_style_from(&mut self, other: &Self)

fn set__moz_appearance(&mut self, v: T)

fn copy__moz_appearance_from(&mut self, other: &Self)

fn set__moz_orient(&mut self, v: T)

fn copy__moz_orient_from(&mut self, other: &Self)

fn set_transition_duration(&mut self, _: T)

fn copy_transition_duration_from(&mut self, _: &Self)

fn transition_duration_count(&self) -> usize

fn transition_duration_at(&self, _index: usize) -> SingleComputedValue

fn set_transition_timing_function(&mut self, _: T)

fn copy_transition_timing_function_from(&mut self, _: &Self)

fn transition_timing_function_count(&self) -> usize

fn transition_timing_function_at(&self, _index: usize) -> SingleComputedValue

fn set_transition_property(&mut self, _: T)

fn copy_transition_property_from(&mut self, _: &Self)

fn transition_property_count(&self) -> usize

fn transition_property_at(&self, _index: usize) -> SingleComputedValue

fn set_transition_delay(&mut self, _: T)

fn copy_transition_delay_from(&mut self, _: &Self)

fn transition_delay_count(&self) -> usize

fn transition_delay_at(&self, _index: usize) -> SingleComputedValue

impl GeckoBox

fn default(pres_context: RawGeckoPresContextBorrowed) -> Arc<Self>

fn get_gecko(&self) -> &nsStyleDisplay