Struct style::gecko_properties::GeckoMargin

pub struct GeckoMargin { /* fields omitted */ }

Methods

impl GeckoMargin

fn set_margin_top(&mut self, v: T)

fn copy_margin_top_from(&mut self, other: &Self)

fn clone_margin_top(&self) -> T

fn set_margin_right(&mut self, v: T)

fn copy_margin_right_from(&mut self, other: &Self)

fn clone_margin_right(&self) -> T

fn set_margin_bottom(&mut self, v: T)

fn copy_margin_bottom_from(&mut self, other: &Self)

fn clone_margin_bottom(&self) -> T

fn set_margin_left(&mut self, v: T)

fn copy_margin_left_from(&mut self, other: &Self)

fn clone_margin_left(&self) -> T

fn set_margin_block_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-block-start given a writing mode.

fn copy_margin_block_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-block-start given a writing mode.

fn set_margin_block_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-block-end given a writing mode.

fn copy_margin_block_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-block-end given a writing mode.

fn set_margin_inline_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-inline-start given a writing mode.

fn copy_margin_inline_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-inline-start given a writing mode.

fn set_margin_inline_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-inline-end given a writing mode.

fn copy_margin_inline_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-inline-end given a writing mode.

impl GeckoMargin

fn default(pres_context: RawGeckoPresContextBorrowed) -> Arc<Self>

fn get_gecko(&self) -> &nsStyleMargin