Struct style::gecko_properties::GeckoPadding

pub struct GeckoPadding { /* fields omitted */ }

Methods

impl GeckoPadding

fn set_padding_top(&mut self, v: T)

fn copy_padding_top_from(&mut self, other: &Self)

fn clone_padding_top(&self) -> T

fn set_padding_right(&mut self, v: T)

fn copy_padding_right_from(&mut self, other: &Self)

fn clone_padding_right(&self) -> T

fn set_padding_bottom(&mut self, v: T)

fn copy_padding_bottom_from(&mut self, other: &Self)

fn clone_padding_bottom(&self) -> T

fn set_padding_left(&mut self, v: T)

fn copy_padding_left_from(&mut self, other: &Self)

fn clone_padding_left(&self) -> T

fn set_padding_block_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-block-start given a writing mode.

fn copy_padding_block_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-block-start given a writing mode.

fn set_padding_block_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-block-end given a writing mode.

fn copy_padding_block_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-block-end given a writing mode.

fn set_padding_inline_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-inline-start given a writing mode.

fn copy_padding_inline_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-inline-start given a writing mode.

fn set_padding_inline_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-inline-end given a writing mode.

fn copy_padding_inline_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-inline-end given a writing mode.

impl GeckoPadding

fn default(pres_context: RawGeckoPresContextBorrowed) -> Arc<Self>

fn get_gecko(&self) -> &nsStylePadding