Struct style::gecko_properties::GeckoPosition

pub struct GeckoPosition { /* fields omitted */ }

Methods

impl GeckoPosition

fn set_top(&mut self, v: T)

fn copy_top_from(&mut self, other: &Self)

fn clone_top(&self) -> T

fn set_right(&mut self, v: T)

fn copy_right_from(&mut self, other: &Self)

fn clone_right(&self) -> T

fn set_bottom(&mut self, v: T)

fn copy_bottom_from(&mut self, other: &Self)

fn clone_bottom(&self) -> T

fn set_left(&mut self, v: T)

fn copy_left_from(&mut self, other: &Self)

fn clone_left(&self) -> T

fn set_z_index(&mut self, v: T)

fn copy_z_index_from(&mut self, other: &Self)

fn clone_z_index(&self) -> T

fn set_box_sizing(&mut self, v: T)

fn copy_box_sizing_from(&mut self, other: &Self)

fn set_order(&mut self, v: T)

fn clone_order(&self) -> T

fn copy_order_from(&mut self, other: &Self)

fn set_grid_row_start(&mut self, v: T)

fn copy_grid_row_start_from(&mut self, other: &Self)

fn set_grid_row_end(&mut self, v: T)

fn copy_grid_row_end_from(&mut self, other: &Self)

fn set_grid_column_start(&mut self, v: T)

fn copy_grid_column_start_from(&mut self, other: &Self)

fn set_grid_column_end(&mut self, v: T)

fn copy_grid_column_end_from(&mut self, other: &Self)

fn set_offset_block_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-block-start given a writing mode.

fn copy_offset_block_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-block-start given a writing mode.

fn set_offset_block_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-block-end given a writing mode.

fn copy_offset_block_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-block-end given a writing mode.

fn set_offset_inline_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-inline-start given a writing mode.

fn copy_offset_inline_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-inline-start given a writing mode.

fn set_offset_inline_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-inline-end given a writing mode.

fn copy_offset_inline_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-inline-end given a writing mode.

fn set_flex_direction(&mut self, v: T)

fn copy_flex_direction_from(&mut self, other: &Self)

fn set_flex_wrap(&mut self, v: T)

fn copy_flex_wrap_from(&mut self, other: &Self)

fn set_align_items(&mut self, v: T)

fn copy_align_items_from(&mut self, other: &Self)

fn clone_align_items(&self) -> T

fn set_flex_grow(&mut self, v: T)

fn copy_flex_grow_from(&mut self, other: &Self)

fn clone_flex_grow(&self) -> T

fn set_flex_shrink(&mut self, v: T)

fn copy_flex_shrink_from(&mut self, other: &Self)

fn clone_flex_shrink(&self) -> T

fn set_align_self(&mut self, v: T)

fn copy_align_self_from(&mut self, other: &Self)

fn clone_align_self(&self) -> T

fn set_width(&mut self, v: T)

fn copy_width_from(&mut self, other: &Self)

fn clone_width(&self) -> T

fn set_min_width(&mut self, v: T)

fn copy_min_width_from(&mut self, other: &Self)

fn clone_min_width(&self) -> T

fn set_max_width(&mut self, v: T)

fn copy_max_width_from(&mut self, other: &Self)

fn clone_max_width(&self) -> T

fn set_height(&mut self, v: T)

fn copy_height_from(&mut self, other: &Self)

fn clone_height(&self) -> T

fn set_min_height(&mut self, v: T)

fn copy_min_height_from(&mut self, other: &Self)

fn clone_min_height(&self) -> T

fn set_max_height(&mut self, v: T)

fn copy_max_height_from(&mut self, other: &Self)

fn clone_max_height(&self) -> T

fn set_block_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for block-size given a writing mode.

fn copy_block_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for block-size given a writing mode.

fn set_min_block_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for min-block-size given a writing mode.

fn copy_min_block_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for min-block-size given a writing mode.

fn set_max_block_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for max-block-size given a writing mode.

fn copy_max_block_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for max-block-size given a writing mode.

fn set_inline_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for inline-size given a writing mode.

fn copy_inline_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for inline-size given a writing mode.

fn set_min_inline_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for min-inline-size given a writing mode.

fn copy_min_inline_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for min-inline-size given a writing mode.

fn set_max_inline_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for max-inline-size given a writing mode.

fn copy_max_inline_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for max-inline-size given a writing mode.

fn set_object_fit(&mut self, v: T)

fn copy_object_fit_from(&mut self, other: &Self)

fn set_object_position(&mut self, v: T)

fn copy_object_position_from(&mut self, other: &Self)

fn clone_object_position(&self) -> T

fn set_grid_row_gap(&mut self, v: T)

fn copy_grid_row_gap_from(&mut self, other: &Self)

fn clone_grid_row_gap(&self) -> T

fn set_grid_column_gap(&mut self, v: T)

fn copy_grid_column_gap_from(&mut self, other: &Self)

fn clone_grid_column_gap(&self) -> T

fn set_flex_basis(&mut self, _: T)

fn copy_flex_basis_from(&mut self, _: &Self)

impl GeckoPosition

fn default(pres_context: RawGeckoPresContextBorrowed) -> Arc<Self>

fn get_gecko(&self) -> &nsStylePosition