Struct style::gecko_string_cache::WeakAtom

pub struct WeakAtom(_);

An atom without a strong reference.

Only usable as &'a WeakAtom, where 'a is the lifetime of something that holds a strong reference to that atom.

Methods

impl WeakAtom

unsafe fn new<'a>(atom: *mut nsIAtom) -> &'a mut Self

Construct a WeakAtom from a raw nsIAtom.

fn clone(&self) -> Atom

Clone this atom, bumping the refcount if the atom is not static.

fn get_hash(&self) -> u32

Get the atom hash.

fn as_slice(&self) -> &[u16]

Get the atom as a slice of utf-16 chars.

fn with_str<F, Output>(&self, cb: F) -> Output where F: FnOnce(&str) -> Output

Execute cb with the string that this atom represents.

Find alternatives to this function when possible, please, since it's pretty slow.

fn to_string(&self) -> String

Convert this Atom into a string, decoding the UTF-16 bytes.

Find alternatives to this function when possible, please, since it's pretty slow.

fn is_static(&self) -> bool

Returns whether this atom is static.

fn len(&self) -> u32

Returns the length of the atom string.

fn as_ptr(&self) -> *mut nsIAtom

Returns the atom as a mutable pointer.