Enum style::hash::map::Entry

pub enum Entry<'a, K, V> where
    K: 'a,
    V: 'a,  {
    Occupied(OccupiedEntry<'a, K, V>),
    Vacant(VacantEntry<'a, K, V>),
}

A view into a single entry in a map, which may either be vacant or occupied.

This enum is constructed from the entry method on HashMap.

Variants
Occupied(OccupiedEntry<'a, K, V>)

An occupied entry.

Vacant(VacantEntry<'a, K, V>)

A vacant entry.

Methods

impl<'a, K, V> Entry<'a, K, V>

fn or_insert(self, default: V) -> &'a mut V

Ensures a value is in the entry by inserting the default if empty, and returns a mutable reference to the value in the entry.

Examples

use std::collections::HashMap;

let mut map: HashMap<&str, u32> = HashMap::new();
map.entry("poneyland").or_insert(12);

assert_eq!(map["poneyland"], 12);

*map.entry("poneyland").or_insert(12) += 10;
assert_eq!(map["poneyland"], 22);

fn or_insert_with<F>(self, default: F) -> &'a mut V where
F: FnOnce() -> V,

Ensures a value is in the entry by inserting the result of the default function if empty, and returns a mutable reference to the value in the entry.

Examples

use std::collections::HashMap;

let mut map: HashMap<&str, String> = HashMap::new();
let s = "hoho".to_string();

map.entry("poneyland").or_insert_with(|| s);

assert_eq!(map["poneyland"], "hoho".to_string());

fn key(&self) -> &K

Returns a reference to this entry's key.

Examples

use std::collections::HashMap;

let mut map: HashMap<&str, u32> = HashMap::new();
assert_eq!(map.entry("poneyland").key(), &"poneyland");

fn and_modify<F>(self, f: F) -> Entry<'a, K, V> where
F: FnMut(&mut V) -> (),

🔬 This is a nightly-only experimental API. (entry_and_modify)

Provides in-place mutable access to an occupied entry before any potential inserts into the map.

Examples

#![feature(entry_and_modify)]
use std::collections::HashMap;

let mut map: HashMap<&str, u32> = HashMap::new();

map.entry("poneyland")
   .and_modify(|e| { *e += 1 })
   .or_insert(42);
assert_eq!(map["poneyland"], 42);

map.entry("poneyland")
   .and_modify(|e| { *e += 1 })
   .or_insert(42);
assert_eq!(map["poneyland"], 43);

impl<'a, K, V> Entry<'a, K, V> where
V: Default,

fn or_default(self) -> &'a mut V

🔬 This is a nightly-only experimental API. (entry_or_default)

Ensures a value is in the entry by inserting the default value if empty, and returns a mutable reference to the value in the entry.

Examples

#![feature(entry_or_default)]
use std::collections::HashMap;

let mut map: HashMap<&str, Option<u32>> = HashMap::new();
map.entry("poneyland").or_default();

assert_eq!(map["poneyland"], None);