style::invalidation::element::element_wrapper

Struct ElementWrapper

pub struct ElementWrapper<'a, 'b, E>
where
    E: TElement,
{
    element: E,
    snapshots: &'a Snapshots<'b>,
}

A simple wrapper over an element and a snapshot, that allows us to selector-match against a past state of the element.

Fields
element: Esnapshots: &'a Snapshots<'b>
Implementations

impl<'a, 'b, E> ElementWrapper<'a, 'b, E>
where E: TElement,

pub fn new(el: E, snapshots: &'a Snapshots<'b>) -> Self

Trivially constructs an ElementWrapper.

pub fn snapshot(&self) -> Option<&'b Snapshot>

Gets the snapshot associated with this element, if any.

pub fn state_changes(&self) -> ElementState

Returns the states that have changed since the element was snapshotted.

fn get_lang(&self) -> Option<AttrValue>

Returns the value of the xml:lang="" (or, if appropriate, lang="") attribute from this element’s snapshot or the closest ancestor element snapshot with the attribute specified.