Struct style::invalidation::element::invalidation_map::InvalidationMap

pub struct InvalidationMap {
    pub class_to_selector: MaybeCaseInsensitiveHashMap<Atom, SmallVec<[Dependency; 1]>>,
    pub id_to_selector: MaybeCaseInsensitiveHashMap<Atom, SmallVec<[Dependency; 1]>>,
    pub state_affecting_selectors: SelectorMap<StateDependency>,
    pub document_state_selectors: Vec<DocumentStateDependency>,
    pub other_attribute_affecting_selectors: PrecomputedHashMap<LocalName, SmallVec<[Dependency; 1]>>,
}

A map where we store invalidations.

This is slightly different to a SelectorMap, in the sense of that the same selector may appear multiple times.

In particular, we want to lookup as few things as possible to get the fewer selectors the better, so this looks up by id, class, or looks at the list of state/other attribute affecting selectors.

Fields
class_to_selector: MaybeCaseInsensitiveHashMap<Atom, SmallVec<[Dependency; 1]>>

A map from a given class name to all the selectors with that class selector.

id_to_selector: MaybeCaseInsensitiveHashMap<Atom, SmallVec<[Dependency; 1]>>

A map from a given id to all the selectors with that ID in the stylesheets currently applying to the document.

state_affecting_selectors: SelectorMap<StateDependency>

A map of all the state dependencies.

document_state_selectors: Vec<DocumentStateDependency>

A list of document state dependencies in the rules we represent.

other_attribute_affecting_selectors: PrecomputedHashMap<LocalName, SmallVec<[Dependency; 1]>>

A map of other attribute affecting selectors.

Implementations

impl InvalidationMap

pub fn new() -> Self

Creates an empty InvalidationMap.

pub fn len(&self) -> usize

Returns the number of dependencies stored in the invalidation map.

pub fn clear(&mut self)

Clears this map, leaving it empty.

pub fn note_selector(
&mut self,
selector: &Selector<SelectorImpl>,
quirks_mode: QuirksMode
) -> Result<(), FailedAllocationError>

Adds a selector to this InvalidationMap. Returns Err(..) to signify OOM.