style::invalidation::element::invalidation_map

Struct RelativeSelectorDependencyCollector

struct RelativeSelectorDependencyCollector<'a, 'b> {
    map: &'a mut InvalidationMap,
    relative_selector_invalidation_map: &'a mut InvalidationMap,
    additional_relative_selector_invalidation_map: &'a mut AdditionalRelativeSelectorInvalidationMap,
    document_state: &'a mut DocumentState,
    selector: &'a RelativeSelector<SelectorImpl>,
    combinator_count: RelativeSelectorCombinatorCount,
    next_selectors: &'a mut SmallVec<[NextDependencyEntry; 5]>,
    quirks_mode: QuirksMode,
    compound_state: PerCompoundState,
    compound_state_attributes: RelativeSelectorCompoundStateAttributes,
    scope_dependencies: &'a mut ScopeSelectorCollectorState<'b>,
    alloc_error: &'a mut Option<AllocErr>,
}

A struct that collects invalidations for a given compound selector.

Fields
map: &'a mut InvalidationMaprelative_selector_invalidation_map: &'a mut InvalidationMapadditional_relative_selector_invalidation_map: &'a mut AdditionalRelativeSelectorInvalidationMapdocument_state: &'a mut DocumentState

The document this complex selector is affected by.

We don’t need to track state per compound selector, since it’s global state and it changes for everything.

selector: &'a RelativeSelector<SelectorImpl>

The current inner relative selector and offset we’re iterating.

combinator_count: RelativeSelectorCombinatorCount

Running combinator for this inner relative selector.

next_selectors: &'a mut SmallVec<[NextDependencyEntry; 5]>

The stack of next selectors that we have, and at which offset of the sequence.

This starts empty. It grows when we find nested :is and :where selector lists. The dependency field is cached and reference counted.

quirks_mode: QuirksMode

The quirks mode of the document where we’re inserting dependencies.

compound_state: PerCompoundState

State relevant to a given compound selector.

compound_state_attributes: RelativeSelectorCompoundStateAttributes

Attributes relevant to the relative compound selector state.

scope_dependencies: &'a mut ScopeSelectorCollectorState<'b>alloc_error: &'a mut Option<AllocErr>

The allocation error, if we OOM.

Implementations

impl<'a, 'b> RelativeSelectorDependencyCollector<'a, 'b>

fn visit_whole_selector(&mut self) -> bool