Struct style::invalidation::element::invalidation_map::RelativeSelectorInnerDependencyCollector

struct RelativeSelectorInnerDependencyCollector<'a, 'b> {
    map: &'a mut RelativeSelectorInvalidationMap,
    document_state: &'a mut DocumentState,
    parent_dependency: &'b Arc<Dependency>,
    selector: &'a Selector<SelectorImpl>,
    parent_selectors: &'a mut SmallVec<[ParentDependencyEntry; 5]>,
    quirks_mode: QuirksMode,
    compound_state: RelativeSelectorPerCompoundState,
    alloc_error: &'a mut Option<AllocErr>,
}

A struct that collects invalidations from a complex selector inside a relative selector. TODO(dshin): All of this duplication is not great Perhaps should be merged to the normal one, if possible? See bug 1855690.

Fields
map: &'a mut RelativeSelectorInvalidationMapdocument_state: &'a mut DocumentState

The document this complex selector is affected by.

We don’t need to track state per compound selector, since it’s global state and it changes for everything.

parent_dependency: &'b Arc<Dependency>

Parent relative selector dependency.

selector: &'a Selector<SelectorImpl>

The current inner relative selector and offset we’re iterating.

parent_selectors: &'a mut SmallVec<[ParentDependencyEntry; 5]>

The stack of parent selectors that we have, and at which offset of the sequence.

This starts empty. It grows when we find nested :is and :where selector lists. The dependency field is cached and reference counted.

quirks_mode: QuirksMode

The quirks mode of the document where we’re inserting dependencies.

compound_state: RelativeSelectorPerCompoundState

State relevant to a given compound selector.

alloc_error: &'a mut Option<AllocErr>

The allocation error, if we OOM.

Implementations

impl<'a, 'b> RelativeSelectorInnerDependencyCollector<'a, 'b>

fn visit_whole_selector(&mut self) -> bool