style::invalidation::element::invalidation_map

Struct SelectorDependencyCollector

struct SelectorDependencyCollector<'a, 'b, 'c> {
    map: &'a mut InvalidationMap,
    relative_selector_invalidation_map: &'a mut InvalidationMap,
    additional_relative_selector_invalidation_map: &'a mut AdditionalRelativeSelectorInvalidationMap,
    document_state: &'a mut DocumentState,
    selector: &'a Selector<SelectorImpl>,
    next_selectors: &'a mut SmallVec<[NextDependencyEntry; 5]>,
    quirks_mode: QuirksMode,
    compound_state: PerCompoundState,
    relative_inner_collector: Option<RelativeSelectorInnerCollectorState<'b>>,
    scope_dependencies: &'a mut ScopeSelectorCollectorState<'c>,
    alloc_error: &'a mut Option<AllocErr>,
}

A struct that collects invalidations for a given compound selector.

Fields
map: &'a mut InvalidationMaprelative_selector_invalidation_map: &'a mut InvalidationMapadditional_relative_selector_invalidation_map: &'a mut AdditionalRelativeSelectorInvalidationMapdocument_state: &'a mut DocumentState

The document this complex selector is affected by.

We don’t need to track state per compound selector, since it’s global state and it changes for everything.

selector: &'a Selector<SelectorImpl>

The current selector and offset we’re iterating.

next_selectors: &'a mut SmallVec<[NextDependencyEntry; 5]>

The stack of next selectors that we have, and at which offset of the sequence.

This starts empty. It grows when we find nested :is and :where selector lists. The dependency field is cached and reference counted.

quirks_mode: QuirksMode

The quirks mode of the document where we’re inserting dependencies.

compound_state: PerCompoundState

State relevant to a given compound selector.

relative_inner_collector: Option<RelativeSelectorInnerCollectorState<'b>>

Additional state to keep track of for collecting nested inner selectors of relative selectors Holds the next relative selector dependency and the state given to a relative selector.

scope_dependencies: &'a mut ScopeSelectorCollectorState<'c>alloc_error: &'a mut Option<AllocErr>

The allocation error, if we OOM.

Implementations

impl<'a, 'b, 'c> SelectorDependencyCollector<'a, 'b, 'c>

fn visit_whole_selector(&mut self) -> bool

fn visit_whole_selector_from( &mut self, iter: SelectorIter<'_, SelectorImpl>, index: usize, ) -> bool