Struct style::invalidation::element::state_and_attributes::Collector

struct Collector<'a, 'b: 'a, 'selectors: 'a, E> where
    E: TElement,  {
    element: E,
    wrapper: ElementWrapper<'b, E>,
    snapshot: &'a Snapshot,
    matching_context: &'a mut MatchingContext<'b, E::Impl>,
    lookup_element: E,
    removed_id: Option<&'a WeakAtom>,
    added_id: Option<&'a WeakAtom>,
    classes_removed: &'a SmallVec<[Atom; 8]>,
    classes_added: &'a SmallVec<[Atom; 8]>,
    state_changes: ElementState,
    descendant_invalidations: &'a mut DescendantInvalidationLists<'selectors>,
    sibling_invalidations: &'a mut InvalidationVector<'selectors>,
    invalidates_self: bool,
}

The collector implementation.

Fields
element: Ewrapper: ElementWrapper<'b, E>snapshot: &'a Snapshotmatching_context: &'a mut MatchingContext<'b, E::Impl>lookup_element: Eremoved_id: Option<&'a WeakAtom>added_id: Option<&'a WeakAtom>classes_removed: &'a SmallVec<[Atom; 8]>classes_added: &'a SmallVec<[Atom; 8]>state_changes: ElementStatedescendant_invalidations: &'a mut DescendantInvalidationLists<'selectors>sibling_invalidations: &'a mut InvalidationVector<'selectors>invalidates_self: bool
Implementations

impl<'a, 'b, 'selectors, E> Collector<'a, 'b, 'selectors, E> where
E: TElement,
'selectors: 'a,

fn collect_dependencies_in_invalidation_map(
&mut self,
map: &'selectors InvalidationMap
)

fn collect_state_dependencies(
&mut self,
map: &'selectors SelectorMap<StateDependency>,
state_changes: ElementState
)

fn check_dependency(&mut self, dependency: &Dependency) -> bool

Check whether a dependency should be taken into account.

fn scan_dependency(&mut self, dependency: &'selectors Dependency)

fn note_dependency(&mut self, dependency: &'selectors Dependency)

fn dependency_may_be_relevant(&self, dependency: &Dependency) -> bool

Returns whether dependency may cause us to invalidate the style of more elements than what we've already invalidated.