style::invalidation

Module stylesheets

A collection of invalidations due to changes in which stylesheets affect a document.

Structs

StylesheetInvalidationSet

A set of invalidations due to stylesheet changes.

Enums

Invalidation 🔒

A style sheet invalidation represents a kind of element or subtree that may need to be restyled. Whether it represents a whole subtree or just a single element is determined by the given InvalidationKind in StylesheetInvalidationSet’s maps.

InvalidationKind 🔒

Whether we should invalidate just the element, or the whole subtree within it.

RuleChangeKind

The kind of change that happened for a given rule.

Functions

invalidate_position_try

Invalidates for any absolutely positioned element that references the given @position-try fallback names.