Module style::keyframes

Structs

Keyframe
A keyframe.

KeyframeListParser
Parses a keyframes list, like: 0%, 50% { width: 50%; }

KeyframePercentage
A number from 1 to 100, indicating the percentage of the animation where this keyframe should run.

KeyframeSelector
A keyframes selector is a list of percentages or from/to symbols, which are converted at parse time to percentages.

KeyframesAnimation
This structure represents a list of animation steps computed from the list of keyframes, in order.

KeyframesStep
A single step from a keyframe animation.

Enums

KeyframesStepValue
A keyframes step value. This can be a synthetised keyframes animation, that is, one autogenerated from the current computed values, or a list of declarations to apply.

Void

Functions

get_animated_properties
Get all the animated properties in a keyframes animation. Note that it's not defined what happens when a property is not on a keyframe, so we only peek the props of the first one.

parse_keyframe_list