Module style::legacy

Legacy presentational attributes defined in the HTML5 specification: <td width>, <input size>, and so forth.

Enums

IntegerAttribute
Legacy presentational attributes that take an integer as defined in HTML5  2.4.4.2.

LengthAttribute
Legacy presentational attributes that take a length as defined in HTML5  2.4.4.4.

SimpleColorAttribute
Legacy presentational attributes that take a simple color as defined in HTML5  2.4.6.

UnsignedIntegerAttribute
Legacy presentational attributes that take a nonnegative integer as defined in HTML5  2.4.4.2.

Traits

PresentationalHintSynthesis
Extension methods for Stylist that cause rules to be synthesized for legacy attributes.

Functions

from_declaration
A convenience function to create a declaration block from a single declaration. This is primarily used in synthesize_rules_for_legacy_attributes.