Trait style::legacy::PresentationalHintSynthesis

pub trait PresentationalHintSynthesis {
    fn synthesize_presentational_hints_for_legacy_attributes<'a, N, V>(&self, node: &N, matching_rules_list: &mut V, shareable: &mut bool) where N: TNode<'a>, Element: TElementAttributes, V: VecLike<DeclarationBlock<Vec<PropertyDeclaration>>>;
    fn synthesize_presentational_hint_for_legacy_background_color_attribute<'a, E, V>(&self, element: E, matching_rules_list: &mut V, shareable: &mut bool) where E: TElement<'a> + TElementAttributes, V: VecLike<DeclarationBlock<Vec<PropertyDeclaration>>>;
    fn synthesize_presentational_hint_for_legacy_border_attribute<'a, E, V>(&self, element: E, matching_rules_list: &mut V, shareable: &mut bool) where E: TElement<'a> + TElementAttributes, V: VecLike<DeclarationBlock<Vec<PropertyDeclaration>>>;
}

Extension methods for Stylist that cause rules to be synthesized for legacy attributes.

Required Methods

fn synthesize_presentational_hints_for_legacy_attributes<'a, N, V>(&self, node: &N, matching_rules_list: &mut V, shareable: &mut bool) where N: TNode<'a>, Element: TElementAttributes, V: VecLike<DeclarationBlock<Vec<PropertyDeclaration>>>

Synthesizes rules from various HTML attributes (mostly legacy junk from HTML4) that confer presentational hints as defined in the HTML5 specification. This handles stuff like <body bgcolor>, <input size>, <td width>, and so forth.

NB: Beware! If you add an attribute to this list, be sure to add it to common_style_affecting_attributes or rare_style_affecting_attributes as appropriate. If you don't, you risk strange random nondeterministic failures due to false positives in style sharing.

fn synthesize_presentational_hint_for_legacy_background_color_attribute<'a, E, V>(&self, element: E, matching_rules_list: &mut V, shareable: &mut bool) where E: TElement<'a> + TElementAttributes, V: VecLike<DeclarationBlock<Vec<PropertyDeclaration>>>

Synthesizes rules for the legacy bgcolor attribute.

fn synthesize_presentational_hint_for_legacy_border_attribute<'a, E, V>(&self, element: E, matching_rules_list: &mut V, shareable: &mut bool) where E: TElement<'a> + TElementAttributes, V: VecLike<DeclarationBlock<Vec<PropertyDeclaration>>>

Synthesizes rules for the legacy border attribute.