style

Module logical_geometry

Geometry in flow-relative space.

Structs

DebugWritingMode 🔒

LogicalMargin

A “margin” in flow-relative dimensions Represents the four sides of the margins, borders, or padding of a CSS box, or a combination of those. A positive “margin” can be added to a rectangle to obtain a bigger rectangle.

LogicalPoint

A 2D point in flow-relative dimensions

LogicalRect

A rectangle in flow-relative dimensions

LogicalSize

A 2D size in flow-relative dimensions

WritingMode

Enums

BlockFlowDirection

Direction

InlineBaseDirection

LogicalAxis

LogicalCorner

LogicalSide

PhysicalAxis

PhysicalCorner

PhysicalSide

WritingModeProperty

The writing-mode property (different from the WritingMode enum). https://drafts.csswg.org/css-writing-modes/#block-flow Aliases come from https://drafts.csswg.org/css-writing-modes-4/#svg-writing-mode