Struct style::logical_geometry::WritingMode

#[repr(C)]
pub struct WritingMode(u8);

Tuple Fields
0: u8
Implementations

impl WritingMode

pub const VERTICAL: Self = _

A vertical writing mode; writing-mode is vertical-rl, vertical-lr, sideways-lr, or sideways-rl.

pub const INLINE_REVERSED: Self = _

The inline flow direction is reversed against the physical direction (i.e. right-to-left or bottom-to-top); writing-mode is sideways-lr or direction is rtl (but not both).

(This bit can be derived from the others, but we store it for convenience.)

pub const VERTICAL_LR: Self = _

A vertical writing mode whose block progression direction is left- to-right; writing-mode is vertical-lr or sideways-lr.

Never set without VERTICAL.

pub const LINE_INVERTED: Self = _

The line-over/line-under sides are inverted with respect to the block-start/block-end edge; writing-mode is vertical-lr.

Never set without VERTICAL and VERTICAL_LR.

pub const RTL: Self = _

direction is rtl.

pub const VERTICAL_SIDEWAYS: Self = _

All text within a vertical writing mode is displayed sideways and runs top-to-bottom or bottom-to-top; set in these cases:

writing-mode: sideways-rl;

writing-mode: sideways-lr;

Never set without VERTICAL.

pub const TEXT_SIDEWAYS: Self = _

Similar to VERTICAL_SIDEWAYS, but is set via text-orientation; set in these cases:

writing-mode: vertical-rl; text-orientation: sideways;

writing-mode: vertical-lr; text-orientation: sideways;

Never set without VERTICAL.

pub const UPRIGHT: Self = _

Horizontal text within a vertical writing mode is displayed with each glyph upright; set in these cases:

writing-mode: vertical-rl; text-orientation: upright;

writing-mode: vertical-lr: text-orientation: upright;

Never set without VERTICAL.

impl WritingMode

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u8

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u8) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u8) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u8) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl WritingMode

pub const fn iter(&self) -> Iter<WritingMode>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<WritingMode>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl WritingMode

pub fn new(inheritedbox_style: &InheritedBox) -> Self

Return a WritingMode bitflags from the relevant CSS properties.

pub fn horizontal_tb() -> Self

Returns the horizontal-tb value.

pub fn is_vertical(&self) -> bool

pub fn is_horizontal(&self) -> bool

pub fn is_vertical_lr(&self) -> bool

Assuming .is_vertical(), does the block direction go left to right?

pub fn is_inline_tb(&self) -> bool

Assuming .is_vertical(), does the inline direction go top to bottom?

pub fn is_bidi_ltr(&self) -> bool

pub fn is_sideways(&self) -> bool

pub fn is_upright(&self) -> bool

pub fn line_left_is_inline_start(&self) -> bool

https://drafts.csswg.org/css-writing-modes/#logical-to-physical

Return	line-left is…	line-right is…

true	inline-start	inline-end

false	inline-end	inline-start

pub fn inline_start_physical_side(&self) -> PhysicalSide

pub fn inline_end_physical_side(&self) -> PhysicalSide

pub fn block_start_physical_side(&self) -> PhysicalSide

pub fn block_end_physical_side(&self) -> PhysicalSide

pub fn start_start_physical_corner(&self) -> PhysicalCorner

pub fn start_end_physical_corner(&self) -> PhysicalCorner

pub fn end_start_physical_corner(&self) -> PhysicalCorner

pub fn end_end_physical_corner(&self) -> PhysicalCorner

pub fn block_flow_direction(&self) -> BlockFlowDirection

pub fn inline_base_direction(&self) -> InlineBaseDirection

pub fn to_bidi_level(&self) -> Level

The default bidirectional embedding level for this writing mode.

Returns bidi level 0 if the mode is LTR, or 1 otherwise.

pub fn is_text_vertical(&self) -> bool

Is the text layout vertical?