Macro style::longhand_properties_idents

macro_rules! longhand_properties_idents {
    ($macro_name: ident) => { ... };
}

Call the given macro with tokens like this for each longhand properties:

{ snake_case_ident, true }

… where the boolean indicates whether the property value type is wrapped in a Box<_> in the corresponding PropertyDeclaration variant.