Module style::macros

Various macro helpers.

Macros

local_name 🔒

ns 🔒

size_of_test 🔒

Asserts the size of a type at compile time.

trivial_to_computed_value 🔒

try_match_ident_ignore_ascii_case 🔒

A macro to parse an identifier, or return an UnexpectedIdent error otherwise.