Module style::matching

High-level interface to CSS selector matching.

Structs

CascadeBooleans

CommonStyleAffectingAttributeInfo

CommonStyleAffectingAttributes

MatchResults

StyleSharingCandidate
Information regarding a candidate.

StyleSharingCandidateCache
An LRU cache of the last few nodes seen, so that we can aggressively try to reuse their styles.

Enums

CacheMiss

CommonStyleAffectingAttributeMode

StyleSharingResult
The results of attempting to share a style.

Constants

ALIGN_CENTER_ATTRIBUTE

ALIGN_LEFT_ATTRIBUTE

ALIGN_RIGHT_ATTRIBUTE

HIDDEN_ATTRIBUTE

NO_WRAP_ATTRIBUTE

Statics

STYLE_SHARING_CANDIDATE_CACHE_SIZE

Traits

MatchMethods

PrivateMatchMethods

Functions

common_style_affecting_attributes

compute_rule_node

create_common_style_affecting_attributes_from_element

element_matches_candidate

have_same_class

have_same_common_style_affecting_attributes

have_same_presentational_hints

match_same_not_common_style_affecting_attributes_rules

match_same_sibling_affecting_rules

rare_style_affecting_attributes
Attributes that, if present, disable style sharing. All legacy HTML attributes must be in either this list or common_style_affecting_attributes. See the comment in synthesize_presentational_hints_for_legacy_attributes.

Type Definitions

PseudoRuleNodes