Enum style::matching::CacheMiss

pub enum CacheMiss {
    Parent,
    LocalName,
    Namespace,
    Link,
    UserAndAuthorRules,
    State,
    IdAttr,
    StyleAttr,
    Class,
    CommonStyleAffectingAttributes,
    PresHints,
    SiblingRules,
    NonCommonAttrRules,
}

A cache miss result.

Variants
Parent

The parents don't match.

LocalName

The local name of the element and the candidate don't match.

Namespace

The namespace of the element and the candidate don't match.

Link

One of the element or the candidate was a link, but the other one wasn't.

UserAndAuthorRules

The element and the candidate match different kind of rules. This can only happen in Gecko.

State

The element and the candidate are in a different state.

IdAttr

The element had an id attribute, which qualifies for a unique style.

StyleAttr

The element had a style attribute, which qualifies for a unique style.

Class

The element and the candidate class names didn't match.

CommonStyleAffectingAttributes

The element and the candidate common style affecting attributes didn't match.

PresHints

The presentation hints didn't match.

SiblingRules

The element and the candidate didn't match the same set of sibling-affecting rules.

NonCommonAttrRules

The element and the candidate didn't match the same set of non-common style affecting attribute selectors.