Struct style::matching::MatchResults

pub struct MatchResults {
    pub primary_relations: Option<StyleRelations>,
    pub rule_nodes_changed: bool,
}

The results returned from running selector matching on an element.

Fields
primary_relations: Option<StyleRelations>

A set of style relations (different hints about what rules matched or could have matched). This is necessary if the style will be shared. If None, the style will not be shared.

rule_nodes_changed: bool

Whether the rule nodes changed during selector matching.

Methods

impl MatchResults

fn primary_is_shareable(&self) -> bool

Returns true if the primary rule node is shareable with other nodes.