Struct style::matching::StyleSharingCandidate

struct StyleSharingCandidate<E: TElement> {
    element: SendElement<E>,
    common_style_affecting_attributes: Option<CommonStyleAffectingAttributes>,
    class_attributes: Option<Vec<Atom>>,
}

Information regarding a style sharing candidate.

Note that this information is stored in TLS and cleared after the traversal, and once here, the style information of the element is immutable, so it's safe to access.

TODO: We can stick a lot more info here.

Fields
element: SendElement<E>

The element. We use SendElement here so that the cache may live in ScopedTLS.

common_style_affecting_attributes: Option<CommonStyleAffectingAttributes>

The cached common style affecting attribute info.

class_attributes: Option<Vec<Atom>>

The cached class names.