Struct style::matching::StyleSharingCandidateCache

pub struct StyleSharingCandidateCache {
    cache: LRUCache<StyleSharingCandidate, ()>,
}

An LRU cache of the last few nodes seen, so that we can aggressively try to reuse their styles.

Note that this cache is flushed every time we steal work from the queue, so storing nodes here temporarily is safe.

NB: We store UnsafeNode's, but this is not unsafe. It's a shame being generic over elements is unfeasible (you can make compile style without much difficulty, but good luck with layout and all the types with assoc. lifetimes).

Fields
cache: LRUCache<StyleSharingCandidate, ()>
Methods

impl StyleSharingCandidateCache

fn new() -> Self

fn iter_mut(&mut self) -> IterMut<(StyleSharingCandidate, ())>

fn insert_if_possible<E: TElement>(&mut self,
element: &E,
style: &Arc<ComputedValues>,
relations: StyleRelations)

fn touch(&mut self, index: usize)

fn clear(&mut self)