Trait style::matching::ElementMatchMethods

pub trait ElementMatchMethods: TElement where Self::Impl: SelectorImplExt {
    fn match_element(&self, stylist: &Stylist<Self::Impl>, parent_bf: Option<&BloomFilter>, applicable_declarations: &mut ApplicableDeclarations<Self::Impl>) -> bool { ... }
    unsafe fn share_style_if_possible(&self, style_sharing_candidate_cache: &mut StyleSharingCandidateCache<Self::ConcreteNode::ConcreteComputedValues>, parent: Option<Self::ConcreteNode>) -> StyleSharingResult<Self::ConcreteNode::ConcreteRestyleDamage> { ... }
}

Provided Methods

fn match_element(&self, stylist: &Stylist<Self::Impl>, parent_bf: Option<&BloomFilter>, applicable_declarations: &mut ApplicableDeclarations<Self::Impl>) -> bool

unsafe fn share_style_if_possible(&self, style_sharing_candidate_cache: &mut StyleSharingCandidateCache<Self::ConcreteNode::ConcreteComputedValues>, parent: Option<Self::ConcreteNode>) -> StyleSharingResult<Self::ConcreteNode::ConcreteRestyleDamage>

Attempts to share a style with another node. This method is unsafe because it depends on the style_sharing_candidate_cache having only live nodes in it, and we have no way to guarantee that at the type system level yet.