style::matching

Trait MatchMethods

pub trait MatchMethods: TElement {
    // Provided methods
    fn layout_parent(&self) -> Self { ... }
    fn line_height_likely_changed(
        old_style: Option<&Arc<ComputedValues>>,
        new_style: &Arc<ComputedValues>,
    ) -> bool { ... }
    fn finish_restyle(
        &self,
        context: &mut StyleContext<'_, Self>,
        data: &mut ElementData,
        new_styles: ResolvedElementStyles,
        important_rules_changed: bool,
    ) -> RestyleHint { ... }
    fn replace_rules(
        &self,
        replacements: RestyleHint,
        context: &mut StyleContext<'_, Self>,
        cascade_inputs: &mut ElementCascadeInputs,
    ) -> bool { ... }
    fn compute_style_difference(
        &self,
        old_values: &ComputedValues,
        new_values: &ComputedValues,
        pseudo: Option<&PseudoElement>,
    ) -> StyleDifference { ... }
}

The public API that elements expose for selector matching.

Provided Methods

fn layout_parent(&self) -> Self

Returns the closest parent element that doesn’t have a display: contents style (and thus generates a box).

This is needed to correctly handle blockification of flex and grid items.

Returns itself if the element has no parent. In practice this doesn’t happen because the root element is blockified per spec, but it could happen if we decide to not blockify for roots of disconnected subtrees, which is a kind of dubious behavior.

fn line_height_likely_changed( old_style: Option<&Arc<ComputedValues>>, new_style: &Arc<ComputedValues>, ) -> bool

Rather than comparing the resolved line-height, which can be expensive to compute as it involves locking and font metrics access, we consider that line-height may have changed if the font-size or line-height property itself has changed, or if the value is ‘normal’ and one of the properties that affects font selection (family, style, weight, stretch) has changed.

fn finish_restyle( &self, context: &mut StyleContext<'_, Self>, data: &mut ElementData, new_styles: ResolvedElementStyles, important_rules_changed: bool, ) -> RestyleHint

Updates the styles with the new ones, diffs them, and stores the restyle damage.

fn replace_rules( &self, replacements: RestyleHint, context: &mut StyleContext<'_, Self>, cascade_inputs: &mut ElementCascadeInputs, ) -> bool

Updates the rule nodes without re-running selector matching, using just the rule tree.

Returns true if an !important rule was replaced.

fn compute_style_difference( &self, old_values: &ComputedValues, new_values: &ComputedValues, pseudo: Option<&PseudoElement>, ) -> StyleDifference

Given the old and new style of this element, and whether it’s a pseudo-element, compute the restyle damage used to determine which kind of layout or painting operations we’ll need.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.