Module style::media_queries::media_feature_expression

Parsing for media feature expressions, like (foo: bar) or (width >= 400px).

Structs

MediaFeatureExpression
A feature expression contains a reference to the media feature, the value the media query contained, and the range to evaluate.

Enums

MediaExpressionValue
A value found or expected in a media expression.

Operator
The operator that was specified in this media feature.

Range
The kind of matching that should be performed on a media feature value.

RangeOrOperator
Either a Range or an Operator.

Functions

consume_operation_or_colon
Consumes an operation or a colon, or returns an error.

disabled_by_pref