Struct style::media_queries::media_feature_expression::MediaFeatureExpression

pub struct MediaFeatureExpression {
    feature_index: usize,
    value: Option<MediaExpressionValue>,
    range_or_operator: Option<RangeOrOperator>,
}

A feature expression contains a reference to the media feature, the value the media query contained, and the range to evaluate.

Fields
feature_index: usizevalue: Option<MediaExpressionValue>range_or_operator: Option<RangeOrOperator>
Implementations

impl MediaFeatureExpression

fn new(
feature_index: usize,
value: Option<MediaExpressionValue>,
range_or_operator: Option<RangeOrOperator>
) -> Self

fn feature(&self) -> &'static MediaFeatureDescription

pub fn parse<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>
) -> Result<Self, ParseError<'i>>

Parse a media expression of the form:

(media-feature: media-value)

pub fn parse_in_parenthesis_block<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>
) -> Result<Self, ParseError<'i>>

Parse a media feature expression where we’ve already consumed the parenthesis.

pub fn matches(&self, device: &Device, quirks_mode: QuirksMode) -> bool

Returns whether this media query evaluates to true for the given device.