Function style::parallel::top_down_dom

fn top_down_dom<'a, 'scope, E, D>(
    nodes: &'a [SendNode<E::ConcreteNode>], 
    root: OpaqueNode, 
    traversal_data: PerLevelTraversalData, 
    scope: &'a ScopeFifo<'scope>, 
    pool: &'scope ThreadPool, 
    traversal: &'scope D, 
    tls: &'scope ScopedTLS<'scope, ThreadLocalStyleContext<E>>
) where
    E: TElement + 'scope,
    D: DomTraversal<E>, 

A parallel top-down DOM traversal.

This algorithm traverses the DOM in a breadth-first, top-down manner. The goals are:

Never process a child before its parent (since child style depends on parent style). If this were to happen, the styling algorithm would panic.

Prioritize discovering nodes as quickly as possible to maximize opportunities for parallelism. But this needs to be weighed against styling cousins on a single thread to improve sharing.

Style all the children of a given node (i.e. all sibling nodes) on a single thread (with an upper bound to handle nodes with an abnormally large number of children). This is important because we use a thread-local cache to share styles between siblings.