style

Module properties

Supported CSS properties and the cascade.

Re-exports

pub use self::cascade::*;

pub use self::declaration_block::*;

pub use self::generated::*;

Modules

cascade

The main cascading algorithm of the style system.

declaration_block

A property declaration block.

generated

The CSS properties supported by the style system. Generated from the properties.mako.rs template by build.rs

shorthands

Manual shorthand parsing and serialization

Structs

AllShorthandDeclarationIterator

An iterator over the all shorthand’s shorthand declarations.

CustomDeclaration

A custom property declaration with the property name and the declared value.

IdSet

A set of ids indexed in a bitfield. W is the number of u32 chunks needed to store Id::COUNT bits, and is filled in by the type aliases above.

IdSetIterator

An iterator over a set of ids.

NonCustomPropertyId

A longhand or shorthand property.

NonCustomPropertyIterator

An iterator over all the property ids that are enabled for a given shorthand, if that shorthand is enabled for all content too.

PropertyFlags

A set of flags for properties.

SourcePropertyDeclaration

A stack-allocated vector of PropertyDeclaration large enough to parse one CSS key: value declaration. (Shorthands expand to multiple PropertyDeclarations.)

SourcePropertyDeclarationDrain

Return type of SourcePropertyDeclaration::drain

TransitionPropertyIteration

A single iteration of the TransitionPropertyIterator.

TransitionPropertyIterator

An iterator over all the properties that transition on a given style.

UnparsedValue

An unparsed property value that contains var() functions.

VariableDeclaration

An unparsed declaration that contains var() functions.

WideKeywordDeclaration

A declaration using a CSS-wide keyword.

Enums

AllShorthand

A parsed all-shorthand value.

CSSWideKeyword

An enum to represent a CSS Wide keyword.

CustomDeclarationValue

A custom property declaration value is either an unparsed value or a CSS wide-keyword.

OwnedPropertyDeclarationId

A PropertyDeclarationId without references, for use as a hash map key.

PropertyDeclarationId

An identifier for a given property declaration, which can be either a longhand or a custom property.

PropertyId

Representation of a CSS property, that is, either a longhand, a shorthand, or a custom property.

Traits

IndexedId

A trait for property-id-like types that can be stored compactly in an IdSet bitfield.

Functions

enabled_arbitrary_substitution_functions

Return the names of arbitrary substitution functions that are enabled.

parse_non_custom_property_declaration_value_into 🔒

Type Aliases

LonghandIdSet

A set of longhand properties.

LonghandIdSetIterator

An iterator over longhand properties.

NonCustomPropertyIdSet

A set of non-custom properties.

NonCustomPropertyIdSetIterator

An iterator over non-custom properties.

PrioritaryPropertyIdSet

A set of prioritary properties.

PrioritaryPropertyIdSetIterator

An iterator over prioritary properties.

ShorthandsWithPropertyReferencesCache

A simple cache for properties that come from a shorthand and have variable references.

SubpropertiesVec

An ArrayVec of subproperties, contains space for the longest shorthand except all.