Module style::properties::animated_properties

A module with all the code related to animated properties.

This needs to be "included" by mako at least after all longhand modules, given they populate the global data.

Structs

InnerMatrix2D
A 2d matrix for interpolation.

MatrixDecomposed2D
A decomposed 2d matrix.

MatrixDecomposed3D
A decomposed 3d matrix.

Perspective
A 3d perspective transformation.

Quaternion
A quaternion used to represent a rotation.

Scale2D
A 2d scale function.

Scale3D
A 3d scale function.

Skew
A 3d skew function.

Translate2D
A 2d translation function.

Translate3D
A 3d translation.

Enums

AnimatedProperty
An animated property interpolation between two computed values for that property.

AnimationValue
An enum to represent a single computed value belonging to an animated property in order to be interpolated with another one. When interpolating, both values need to belong to the same property.

TransitionProperty
A given transition property, that is either All, or an animatable property.

Traits

Interpolate
A trait used to implement interpolation.

RepeatableListInterpolate
https://drafts.csswg.org/css-transitions/#animtype-repeatable-list

Functions

build_identity_transform_list
Build an equivalent 'identity transform function list' based on an existing transform list. http://dev.w3.org/csswg/css-transforms/#none-transform-animation

can_interpolate_list
Check if it's possible to do a direct numerical interpolation between these two transform lists. http://dev.w3.org/csswg/css-transforms/#transform-transform-animation

combine

cross

decompose_3d_matrix
Decompose a 3D matrix. https://drafts.csswg.org/css-transforms/#decomposing-a-3d-matrix

dot

interpolate_transform_list
Interpolate two transform lists. http://dev.w3.org/csswg/css-transforms/#interpolation-of-transforms

multiply

rotate_to_matrix
https://drafts.csswg.org/css-transforms/#Rotate3dDefined