Enum style::properties::animated_properties::AnimatedProperty

pub enum AnimatedProperty {
    AlignContent(T, T),
    AlignItems(T, T),
    AlignSelf(T, T),
    AspectRatio(T, T),
    BackfaceVisibility(T, T),
    BorderCollapse(T, T),
    BorderImageRepeat(T, T),
    BoxSizing(T, T),
    CaptionSide(T, T),
    Clear(T, T),
    ColumnCount(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Display(T, T),
    EmptyCells(T, T),
    FlexDirection(T, T),
    FlexWrap(T, T),
    Float(T, T),
    FontStretch(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    FontStyle(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    FontVariantCaps(T, T),
    FontWeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    ImageRendering(T, T),
    JustifyContent(T, T),
    ListStylePosition(T, T),
    ListStyleType(T, T),
    MixBlendMode(T, T),
    Opacity(T, T),
    Order(T, T),
    OutlineStyle(T, T),
    OverflowWrap(T, T),
    PointerEvents(T, T),
    Position(T, T),
    TableLayout(T, T),
    TextAlign(T, T),
    TextDecorationLine(T, T),
    TextJustify(T, T),
    TextRendering(T, T),
    TextTransform(T, T),
    TransformStyle(T, T),
    Visibility(T, T),
    WhiteSpace(T, T),
    WordBreak(T, T),
    ZIndex(T, T),
    FlexGrow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    FlexShrink(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    OverflowX(T, T),
    OverflowY(T, T),
    BorderBottomStyle(T, T),
    BorderLeftStyle(T, T),
    BorderRightStyle(T, T),
    BorderTopStyle(T, T),
    BackgroundAttachment(T, T),
    BackgroundClip(T, T),
    BackgroundImage(T, T),
    BackgroundOrigin(T, T),
    BackgroundPositionX(T, T),
    BackgroundPositionY(T, T),
    BackgroundRepeat(T, T),
    BackgroundSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderImageOutset(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderImageSlice(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderImageSource(T, T),
    BorderImageWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderSpacing(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BoxShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Clip(T, T),
    Color(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    ColumnGap(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    ColumnWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Content(T, T),
    CounterIncrement(T, T),
    CounterReset(T, T),
    Cursor(T, T),
    Filter(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    FlexBasis(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    FontFamily(T, T),
    FontSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    LetterSpacing(T, T),
    LineHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    ListStyleImage(T, T),
    OutlineOffset(T, T),
    Perspective(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PerspectiveOrigin(T, T),
    Quotes(T, T),
    Rotate(T, T),
    Scale(T, T),
    TextIndent(T, T),
    TextOverflow(T, T),
    TextShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Transform(T, T),
    TransformOrigin(T, T),
    Translate(T, T),
    VerticalAlign(T, T),
    WordSpacing(T, T),
    MaxHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    MaxWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PaddingBottom(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PaddingLeft(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PaddingRight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PaddingTop(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Height(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    MinHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    MinWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Width(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderLeftWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderRightWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    OutlineWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BackgroundColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderLeftColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderRightColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    OutlineColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Bottom(T, T),
    Left(T, T),
    MarginBottom(T, T),
    MarginLeft(T, T),
    MarginRight(T, T),
    MarginTop(T, T),
    Right(T, T),
    Top(T, T),
}

An animated property interpolation between two computed values for that property.

Variants

AlignContent(T, T)

align-content

AlignItems(T, T)

align-items

AlignSelf(T, T)

align-self

AspectRatio(T, T)

aspect-ratio

BackfaceVisibility(T, T)

backface-visibility

BorderCollapse(T, T)

border-collapse

BorderImageRepeat(T, T)

border-image-repeat

BoxSizing(T, T)

box-sizing

CaptionSide(T, T)

caption-side

Clear(T, T)

clear

ColumnCount(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

column-count

Display(T, T)

display

EmptyCells(T, T)

empty-cells

FlexDirection(T, T)

flex-direction

FlexWrap(T, T)

flex-wrap

Float(T, T)

float

FontStretch(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

font-stretch

FontStyle(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

font-style

FontVariantCaps(T, T)

font-variant-caps

FontWeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

font-weight

ImageRendering(T, T)

image-rendering

JustifyContent(T, T)

justify-content

ListStylePosition(T, T)

list-style-position

ListStyleType(T, T)

list-style-type

MixBlendMode(T, T)

mix-blend-mode

Opacity(T, T)

opacity

Order(T, T)

order

OutlineStyle(T, T)

outline-style

OverflowWrap(T, T)

overflow-wrap

PointerEvents(T, T)

pointer-events

Position(T, T)

position

TableLayout(T, T)

table-layout

TextAlign(T, T)

text-align

TextDecorationLine(T, T)

text-decoration-line

TextJustify(T, T)

text-justify

TextRendering(T, T)

text-rendering

TextTransform(T, T)

text-transform

TransformStyle(T, T)

transform-style

Visibility(T, T)

visibility

WhiteSpace(T, T)

white-space

WordBreak(T, T)

word-break

ZIndex(T, T)

z-index

FlexGrow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

flex-grow

FlexShrink(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

flex-shrink

OverflowX(T, T)

overflow-x

OverflowY(T, T)

overflow-y

BorderBottomStyle(T, T)

border-bottom-style

BorderLeftStyle(T, T)

border-left-style

BorderRightStyle(T, T)

border-right-style

BorderTopStyle(T, T)

border-top-style

BackgroundAttachment(T, T)

background-attachment

BackgroundClip(T, T)

background-clip

BackgroundImage(T, T)

background-image

BackgroundOrigin(T, T)

background-origin

BackgroundPositionX(T, T)

background-position-x

BackgroundPositionY(T, T)

background-position-y

BackgroundRepeat(T, T)

background-repeat

BackgroundSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

background-size

BorderImageOutset(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-image-outset

BorderImageSlice(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-image-slice

BorderImageSource(T, T)

border-image-source

BorderImageWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-image-width

BorderSpacing(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-spacing

BoxShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

box-shadow

Clip(T, T)

clip

Color(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

color

ColumnGap(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

column-gap

ColumnWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

column-width

Content(T, T)

content

CounterIncrement(T, T)

counter-increment

CounterReset(T, T)

counter-reset

Cursor(T, T)

cursor

Filter(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

filter

FlexBasis(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

flex-basis

FontFamily(T, T)

font-family

FontSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

font-size

LetterSpacing(T, T)

letter-spacing

LineHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

line-height

ListStyleImage(T, T)

list-style-image

OutlineOffset(T, T)

outline-offset

Perspective(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

perspective

PerspectiveOrigin(T, T)

perspective-origin

Quotes(T, T)

quotes

Rotate(T, T)

rotate

Scale(T, T)

scale

TextIndent(T, T)

text-indent

TextOverflow(T, T)

text-overflow

TextShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

text-shadow

Transform(T, T)

transform

TransformOrigin(T, T)

transform-origin

Translate(T, T)

translate

VerticalAlign(T, T)

vertical-align

WordSpacing(T, T)

word-spacing

MaxHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

max-height

MaxWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

max-width

BorderBottomLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-left-radius

BorderBottomRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-right-radius

BorderTopLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-left-radius

BorderTopRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-right-radius

PaddingBottom(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-bottom

PaddingLeft(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-left

PaddingRight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-right

PaddingTop(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-top

Height(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

height

MinHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

min-height

MinWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

min-width

Width(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

width

BorderBottomWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-width

BorderLeftWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-left-width

BorderRightWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-right-width

BorderTopWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-width

OutlineWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

outline-width

BackgroundColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

background-color

BorderBottomColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-color

BorderLeftColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-left-color

BorderRightColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-right-color

BorderTopColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-color

OutlineColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

outline-color

Bottom(T, T)

bottom

Left(T, T)

left

MarginBottom(T, T)

margin-bottom

MarginLeft(T, T)

margin-left

MarginRight(T, T)

margin-right

MarginTop(T, T)

margin-top

Right(T, T)

right

Top(T, T)

top

Methods

impl AnimatedProperty

pub fn id(&self) -> LonghandId

Get the id of the property we're animating.

pub fn name(&self) -> &'static str

Get the name of this property.

pub fn does_animate(&self) -> bool

Whether this interpolation does animate, that is, whether the start and end values are different.

pub fn has_the_same_end_value_as(&self, other: &Self) -> bool

Whether an animated property has the same end value as another.

pub fn update(&self, style: &mut ComputedValues, progress: f64)

Update style with the proper computed style corresponding to this animation at progress.

pub fn from_longhand(
property: LonghandId,
old_style: &ComputedValues,
new_style: &ComputedValues
) -> Option<AnimatedProperty>

Get an animatable value from a transition-property, an old style, and a new style.