Enum style::properties::animated_properties::AnimationValue

#[repr(u16)]pub enum AnimationValue {
    AlignContent(<SpecifiedValue as ToComputedValue>::ComputedValue),
    AlignItems(<SpecifiedValue as ToComputedValue>::ComputedValue),
    AlignSelf(<SpecifiedValue as ToComputedValue>::ComputedValue),
    AspectRatio(<AspectRatio as ToComputedValue>::ComputedValue),
    BackfaceVisibility(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BorderCollapse(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BorderImageRepeat(<BorderImageRepeat as ToComputedValue>::ComputedValue),
    BoxSizing(<SpecifiedValue as ToComputedValue>::ComputedValue),
    CaptionSide(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Clear(<Clear as ToComputedValue>::ComputedValue),
    ColumnCount(<<ColumnCount as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Direction(Void),
    Display(<Display as ToComputedValue>::ComputedValue),
    EmptyCells(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FlexDirection(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FlexWrap(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Float(<Float as ToComputedValue>::ComputedValue),
    FontStretch(<<FontStretch as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontStyle(<<FontStyle as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontVariantCaps(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FontWeight(<<FontWeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ImageRendering(<SpecifiedValue as ToComputedValue>::ComputedValue),
    JustifyContent(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ListStylePosition(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ListStyleType(<SpecifiedValue as ToComputedValue>::ComputedValue),
    MixBlendMode(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Opacity(<Opacity as ToComputedValue>::ComputedValue),
    Order(<Integer as ToComputedValue>::ComputedValue),
    OutlineStyle(<OutlineStyle as ToComputedValue>::ComputedValue),
    OverflowWrap(<OverflowWrap as ToComputedValue>::ComputedValue),
    PointerEvents(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Position(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ServoOverflowClipBox(Void),
    ServoTopLayer(Void),
    TableLayout(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextAlign(<TextAlign as ToComputedValue>::ComputedValue),
    TextDecorationLine(<TextDecorationLine as ToComputedValue>::ComputedValue),
    TextJustify(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextRendering(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextTransform(<TextTransform as ToComputedValue>::ComputedValue),
    TransformStyle(<TransformStyle as ToComputedValue>::ComputedValue),
    UnicodeBidi(Void),
    Visibility(<SpecifiedValue as ToComputedValue>::ComputedValue),
    WhiteSpace(<SpecifiedValue as ToComputedValue>::ComputedValue),
    WordBreak(<WordBreak as ToComputedValue>::ComputedValue),
    WritingMode(Void),
    ZIndex(<ZIndex as ToComputedValue>::ComputedValue),
    FlexGrow(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FlexShrink(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OverflowBlock(Void),
    OverflowInline(Void),
    OverflowX(<Overflow as ToComputedValue>::ComputedValue),
    OverflowY(<Overflow as ToComputedValue>::ComputedValue),
    BorderBlockEndStyle(Void),
    BorderBlockStartStyle(Void),
    BorderBottomStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderInlineEndStyle(Void),
    BorderInlineStartStyle(Void),
    BorderLeftStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderRightStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderTopStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    AnimationDelay(Void),
    AnimationDirection(Void),
    AnimationDuration(Void),
    AnimationFillMode(Void),
    AnimationIterationCount(Void),
    AnimationName(Void),
    AnimationPlayState(Void),
    AnimationTimingFunction(Void),
    BackgroundAttachment(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundClip(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundImage(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundOrigin(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundPositionX(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundPositionY(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundRepeat(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundSize(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageOutset(<<NonNegativeLengthOrNumberRect as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageSlice(<<BorderImageSlice as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageSource(<Image as ToComputedValue>::ComputedValue),
    BorderImageWidth(<<BorderImageWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderSpacing(<<BorderSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BoxShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Clip(<ClipRectOrAuto as ToComputedValue>::ComputedValue),
    Color(<<ColorPropertyValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ColumnGap(<<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ColumnWidth(<<NonNegativeLengthOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Content(<Content as ToComputedValue>::ComputedValue),
    CounterIncrement(<CounterIncrement as ToComputedValue>::ComputedValue),
    CounterReset(<CounterSetOrReset as ToComputedValue>::ComputedValue),
    Cursor(<Cursor as ToComputedValue>::ComputedValue),
    Filter(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FlexBasis(<<FlexBasis as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontFamily(<FontFamily as ToComputedValue>::ComputedValue),
    FontSize(<<FontSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    LetterSpacing(<LetterSpacing as ToComputedValue>::ComputedValue),
    LineHeight(<<LineHeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ListStyleImage(<ImageUrlOrNone as ToComputedValue>::ComputedValue),
    OutlineOffset(<Length as ToComputedValue>::ComputedValue),
    Perspective(<<Perspective as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PerspectiveOrigin(<Position as ToComputedValue>::ComputedValue),
    Quotes(<Quotes as ToComputedValue>::ComputedValue),
    Rotate(<Rotate as ToComputedValue>::ComputedValue),
    Scale(<Scale as ToComputedValue>::ComputedValue),
    TextIndent(<LengthPercentage as ToComputedValue>::ComputedValue),
    TextOverflow(<TextOverflow as ToComputedValue>::ComputedValue),
    TextShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Transform(<Transform as ToComputedValue>::ComputedValue),
    TransformOrigin(<TransformOrigin as ToComputedValue>::ComputedValue),
    TransitionDelay(Void),
    TransitionDuration(Void),
    TransitionProperty(Void),
    TransitionTimingFunction(Void),
    Translate(<Translate as ToComputedValue>::ComputedValue),
    VerticalAlign(<VerticalAlign as ToComputedValue>::ComputedValue),
    WordSpacing(<WordSpacing as ToComputedValue>::ComputedValue),
    MaxBlockSize(Void),
    MaxHeight(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MaxInlineSize(Void),
    MaxWidth(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBottomLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBottomRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderEndEndRadius(Void),
    BorderEndStartRadius(Void),
    BorderStartEndRadius(Void),
    BorderStartStartRadius(Void),
    BorderTopLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingBlockEnd(Void),
    PaddingBlockStart(Void),
    PaddingBottom(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingInlineEnd(Void),
    PaddingInlineStart(Void),
    PaddingLeft(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingRight(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingTop(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BlockSize(Void),
    Height(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    InlineSize(Void),
    MinBlockSize(Void),
    MinHeight(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MinInlineSize(Void),
    MinWidth(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Width(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBlockEndWidth(Void),
    BorderBlockStartWidth(Void),
    BorderBottomWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderInlineEndWidth(Void),
    BorderInlineStartWidth(Void),
    BorderLeftWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderRightWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BackgroundColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBlockEndColor(Void),
    BorderBlockStartColor(Void),
    BorderBottomColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderInlineEndColor(Void),
    BorderInlineStartColor(Void),
    BorderLeftColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderRightColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Bottom(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    InsetBlockEnd(Void),
    InsetBlockStart(Void),
    InsetInlineEnd(Void),
    InsetInlineStart(Void),
    Left(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginBlockEnd(Void),
    MarginBlockStart(Void),
    MarginBottom(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginInlineEnd(Void),
    MarginInlineStart(Void),
    MarginLeft(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginRight(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginTop(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    Right(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    Top(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
}

An enum to represent a single computed value belonging to an animated property in order to be interpolated with another one. When interpolating, both values need to belong to the same property.

FIXME: We need to add a path for custom properties, but that's trivial after this (is a similar path to that of PropertyDeclaration).

Variants

AlignContent(<SpecifiedValue as ToComputedValue>::ComputedValue)

align-content

AlignItems(<SpecifiedValue as ToComputedValue>::ComputedValue)

align-items

AlignSelf(<SpecifiedValue as ToComputedValue>::ComputedValue)

align-self

AspectRatio(<AspectRatio as ToComputedValue>::ComputedValue)

aspect-ratio

BackfaceVisibility(<SpecifiedValue as ToComputedValue>::ComputedValue)

backface-visibility

BorderCollapse(<SpecifiedValue as ToComputedValue>::ComputedValue)

border-collapse

BorderImageRepeat(<BorderImageRepeat as ToComputedValue>::ComputedValue)

border-image-repeat

BoxSizing(<SpecifiedValue as ToComputedValue>::ComputedValue)

box-sizing

CaptionSide(<SpecifiedValue as ToComputedValue>::ComputedValue)

caption-side

Clear(<Clear as ToComputedValue>::ComputedValue)

clear

ColumnCount(<<ColumnCount as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

column-count

Direction(Void)

direction

Display(<Display as ToComputedValue>::ComputedValue)

display

EmptyCells(<SpecifiedValue as ToComputedValue>::ComputedValue)

empty-cells

FlexDirection(<SpecifiedValue as ToComputedValue>::ComputedValue)

flex-direction

FlexWrap(<SpecifiedValue as ToComputedValue>::ComputedValue)

flex-wrap

Float(<Float as ToComputedValue>::ComputedValue)

float

FontStretch(<<FontStretch as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-stretch

FontStyle(<<FontStyle as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-style

FontVariantCaps(<SpecifiedValue as ToComputedValue>::ComputedValue)

font-variant-caps

FontWeight(<<FontWeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-weight

ImageRendering(<SpecifiedValue as ToComputedValue>::ComputedValue)

image-rendering

JustifyContent(<SpecifiedValue as ToComputedValue>::ComputedValue)

justify-content

ListStylePosition(<SpecifiedValue as ToComputedValue>::ComputedValue)

list-style-position

ListStyleType(<SpecifiedValue as ToComputedValue>::ComputedValue)

list-style-type

MixBlendMode(<SpecifiedValue as ToComputedValue>::ComputedValue)

mix-blend-mode

Opacity(<Opacity as ToComputedValue>::ComputedValue)

opacity

Order(<Integer as ToComputedValue>::ComputedValue)

order

OutlineStyle(<OutlineStyle as ToComputedValue>::ComputedValue)

outline-style

OverflowWrap(<OverflowWrap as ToComputedValue>::ComputedValue)

overflow-wrap

PointerEvents(<SpecifiedValue as ToComputedValue>::ComputedValue)

pointer-events

Position(<SpecifiedValue as ToComputedValue>::ComputedValue)

position

ServoOverflowClipBox(Void)

-servo-overflow-clip-box

ServoTopLayer(Void)

-servo-top-layer

TableLayout(<SpecifiedValue as ToComputedValue>::ComputedValue)

table-layout

TextAlign(<TextAlign as ToComputedValue>::ComputedValue)

text-align

TextDecorationLine(<TextDecorationLine as ToComputedValue>::ComputedValue)

text-decoration-line

TextJustify(<SpecifiedValue as ToComputedValue>::ComputedValue)

text-justify

TextRendering(<SpecifiedValue as ToComputedValue>::ComputedValue)

text-rendering

TextTransform(<TextTransform as ToComputedValue>::ComputedValue)

text-transform

TransformStyle(<TransformStyle as ToComputedValue>::ComputedValue)

transform-style

UnicodeBidi(Void)

unicode-bidi

Visibility(<SpecifiedValue as ToComputedValue>::ComputedValue)

visibility

WhiteSpace(<SpecifiedValue as ToComputedValue>::ComputedValue)

white-space

WordBreak(<WordBreak as ToComputedValue>::ComputedValue)

word-break

WritingMode(Void)

writing-mode

ZIndex(<ZIndex as ToComputedValue>::ComputedValue)

z-index

FlexGrow(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

flex-grow

FlexShrink(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

flex-shrink

OverflowBlock(Void)

overflow-block

OverflowInline(Void)

overflow-inline

OverflowX(<Overflow as ToComputedValue>::ComputedValue)

overflow-x

OverflowY(<Overflow as ToComputedValue>::ComputedValue)

overflow-y

BorderBlockEndStyle(Void)

border-block-end-style

BorderBlockStartStyle(Void)

border-block-start-style

BorderBottomStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-bottom-style

BorderInlineEndStyle(Void)

border-inline-end-style

BorderInlineStartStyle(Void)

border-inline-start-style

BorderLeftStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-left-style

BorderRightStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-right-style

BorderTopStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-top-style

AnimationDelay(Void)

animation-delay

AnimationDirection(Void)

animation-direction

AnimationDuration(Void)

animation-duration

AnimationFillMode(Void)

animation-fill-mode

AnimationIterationCount(Void)

animation-iteration-count

AnimationName(Void)

animation-name

AnimationPlayState(Void)

animation-play-state

AnimationTimingFunction(Void)

animation-timing-function

BackgroundAttachment(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-attachment

BackgroundClip(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-clip

BackgroundImage(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-image

BackgroundOrigin(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-origin

BackgroundPositionX(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-position-x

BackgroundPositionY(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-position-y

BackgroundRepeat(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-repeat

BackgroundSize(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

background-size

BorderImageOutset(<<NonNegativeLengthOrNumberRect as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-image-outset

BorderImageSlice(<<BorderImageSlice as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-image-slice

BorderImageSource(<Image as ToComputedValue>::ComputedValue)

border-image-source

BorderImageWidth(<<BorderImageWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-image-width

BorderSpacing(<<BorderSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-spacing

BoxShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

box-shadow

Clip(<ClipRectOrAuto as ToComputedValue>::ComputedValue)

clip

Color(<<ColorPropertyValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

color

ColumnGap(<<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

column-gap

ColumnWidth(<<NonNegativeLengthOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

column-width

Content(<Content as ToComputedValue>::ComputedValue)

content

CounterIncrement(<CounterIncrement as ToComputedValue>::ComputedValue)

counter-increment

CounterReset(<CounterSetOrReset as ToComputedValue>::ComputedValue)

counter-reset

Cursor(<Cursor as ToComputedValue>::ComputedValue)

cursor

Filter(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

filter

FlexBasis(<<FlexBasis as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

flex-basis

FontFamily(<FontFamily as ToComputedValue>::ComputedValue)

font-family

FontSize(<<FontSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-size

LetterSpacing(<LetterSpacing as ToComputedValue>::ComputedValue)

letter-spacing

LineHeight(<<LineHeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

line-height

ListStyleImage(<ImageUrlOrNone as ToComputedValue>::ComputedValue)

list-style-image

OutlineOffset(<Length as ToComputedValue>::ComputedValue)

outline-offset

Perspective(<<Perspective as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

perspective

PerspectiveOrigin(<Position as ToComputedValue>::ComputedValue)

perspective-origin

Quotes(<Quotes as ToComputedValue>::ComputedValue)

quotes

Rotate(<Rotate as ToComputedValue>::ComputedValue)

rotate

Scale(<Scale as ToComputedValue>::ComputedValue)

scale

TextIndent(<LengthPercentage as ToComputedValue>::ComputedValue)

text-indent

TextOverflow(<TextOverflow as ToComputedValue>::ComputedValue)

text-overflow

TextShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

text-shadow

Transform(<Transform as ToComputedValue>::ComputedValue)

transform

TransformOrigin(<TransformOrigin as ToComputedValue>::ComputedValue)

transform-origin

TransitionDelay(Void)

transition-delay

TransitionDuration(Void)

transition-duration

TransitionProperty(Void)

transition-property

TransitionTimingFunction(Void)

transition-timing-function

Translate(<Translate as ToComputedValue>::ComputedValue)

translate

VerticalAlign(<VerticalAlign as ToComputedValue>::ComputedValue)

vertical-align

WordSpacing(<WordSpacing as ToComputedValue>::ComputedValue)

word-spacing

MaxBlockSize(Void)

max-block-size

MaxHeight(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

max-height

MaxInlineSize(Void)

max-inline-size

MaxWidth(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

max-width

BorderBottomLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-left-radius

BorderBottomRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-right-radius

BorderEndEndRadius(Void)

border-end-end-radius

BorderEndStartRadius(Void)

border-end-start-radius

BorderStartEndRadius(Void)

border-start-end-radius

BorderStartStartRadius(Void)

border-start-start-radius

BorderTopLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-left-radius

BorderTopRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-right-radius

PaddingBlockEnd(Void)

padding-block-end

PaddingBlockStart(Void)

padding-block-start

PaddingBottom(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-bottom

PaddingInlineEnd(Void)

padding-inline-end

PaddingInlineStart(Void)

padding-inline-start

PaddingLeft(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-left

PaddingRight(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-right

PaddingTop(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-top

BlockSize(Void)

block-size

Height(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

height

InlineSize(Void)

inline-size

MinBlockSize(Void)

min-block-size

MinHeight(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

min-height

MinInlineSize(Void)

min-inline-size

MinWidth(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

min-width

Width(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

width

BorderBlockEndWidth(Void)

border-block-end-width

BorderBlockStartWidth(Void)

border-block-start-width

BorderBottomWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-width

BorderInlineEndWidth(Void)

border-inline-end-width

BorderInlineStartWidth(Void)

border-inline-start-width

BorderLeftWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-left-width

BorderRightWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-right-width

BorderTopWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-width

OutlineWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

outline-width

BackgroundColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

background-color

BorderBlockEndColor(Void)

border-block-end-color

BorderBlockStartColor(Void)

border-block-start-color

BorderBottomColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-color

BorderInlineEndColor(Void)

border-inline-end-color

BorderInlineStartColor(Void)

border-inline-start-color

BorderLeftColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-left-color

BorderRightColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-right-color

BorderTopColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-color

OutlineColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

outline-color

Bottom(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

bottom

InsetBlockEnd(Void)

inset-block-end

InsetBlockStart(Void)

inset-block-start

InsetInlineEnd(Void)

inset-inline-end

InsetInlineStart(Void)

inset-inline-start

Left(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

left

MarginBlockEnd(Void)

margin-block-end

MarginBlockStart(Void)

margin-block-start

MarginBottom(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

margin-bottom

MarginInlineEnd(Void)

margin-inline-end

MarginInlineStart(Void)

margin-inline-start

MarginLeft(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

margin-left

MarginRight(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

margin-right

MarginTop(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

margin-top

Right(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

right

Top(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue)

top

Implementations

impl AnimationValue

pub fn id(&self) -> LonghandId

Returns the longhand id this animated value corresponds to.

pub fn uncompute(&self) -> PropertyDeclaration

"Uncompute" this animation value in order to be used inside the CSS cascade.

pub fn from_declaration(
decl: &PropertyDeclaration,
context: &mut Context,
extra_custom_properties: Option<&Arc<CustomPropertiesMap>>,
initial: &ComputedValues
) -> Option<Self>

Construct an AnimationValue from a property declaration.

pub fn from_computed_values(
property: LonghandId,
style: &ComputedValues
) -> Option<Self>

Get an AnimationValue for an AnimatableLonghand from a given computed values.

pub fn set_in_style_for_servo(&self, style: &mut ComputedValues)

Update style with the value of this AnimationValue.

SERVO ONLY: This doesn't properly handle things like updating 'em' units when animated font-size.