Enum style::properties::animated_properties::TransitionProperty

pub enum TransitionProperty {
    All,
    BackgroundColor,
    BackgroundPositionX,
    BackgroundPositionY,
    BackgroundSize,
    BorderTopColor,
    BorderTopWidth,
    BorderLeftColor,
    BorderLeftWidth,
    BorderBottomColor,
    BorderBottomWidth,
    BorderRightColor,
    BorderRightWidth,
    BorderTopLeftRadius,
    BorderTopRightRadius,
    BorderBottomRightRadius,
    BorderBottomLeftRadius,
    VerticalAlign,
    Transform,
    Perspective,
    PerspectiveOrigin,
    TransformOrigin,
    Color,
    ColumnWidth,
    ColumnCount,
    ColumnGap,
    Opacity,
    BoxShadow,
    Clip,
    Filter,
    FontWeight,
    FontSize,
    FontStretch,
    Visibility,
    BorderSpacing,
    LineHeight,
    TextIndent,
    LetterSpacing,
    WordSpacing,
    TextShadow,
    MarginTop,
    MarginLeft,
    MarginBottom,
    MarginRight,
    OutlineColor,
    OutlineWidth,
    OutlineOffset,
    PaddingTop,
    PaddingLeft,
    PaddingBottom,
    PaddingRight,
    Top,
    Left,
    Bottom,
    Right,
    ZIndex,
    FlexGrow,
    FlexShrink,
    Order,
    Width,
    MinWidth,
    MaxWidth,
    Height,
    MinHeight,
    MaxHeight,
    Background,
    BackgroundPosition,
    BorderColor,
    BorderWidth,
    BorderTop,
    BorderLeft,
    BorderBottom,
    BorderRight,
    Border,
    BorderRadius,
    Columns,
    Font,
    Margin,
    Outline,
    Padding,
    Flex,
    Unsupported(CustomIdent),
}

A given transition property, that is either All, a transitionable longhand property, a shorthand with at least one transitionable longhand component, or an unsupported property.

Variants
All

All, any transitionable property changing should generate a transition.

BackgroundColor

background-color

BackgroundPositionX

background-position-x

BackgroundPositionY

background-position-y

BackgroundSize

background-size

BorderTopColor

border-top-color

BorderTopWidth

border-top-width

BorderLeftColor

border-left-color

BorderLeftWidth

border-left-width

BorderBottomColor

border-bottom-color

BorderBottomWidth

border-bottom-width

BorderRightColor

border-right-color

BorderRightWidth

border-right-width

BorderTopLeftRadius

border-top-left-radius

BorderTopRightRadius

border-top-right-radius

BorderBottomRightRadius

border-bottom-right-radius

BorderBottomLeftRadius

border-bottom-left-radius

VerticalAlign

vertical-align

Transform

transform

Perspective

perspective

PerspectiveOrigin

perspective-origin

TransformOrigin

transform-origin

Color

color

ColumnWidth

column-width

ColumnCount

column-count

ColumnGap

column-gap

Opacity

opacity

BoxShadow

box-shadow

Clip

clip

Filter

filter

FontWeight

font-weight

FontSize

font-size

FontStretch

font-stretch

Visibility

visibility

BorderSpacing

border-spacing

LineHeight

line-height

TextIndent

text-indent

LetterSpacing

letter-spacing

WordSpacing

word-spacing

TextShadow

text-shadow

MarginTop

margin-top

MarginLeft

margin-left

MarginBottom

margin-bottom

MarginRight

margin-right

OutlineColor

outline-color

OutlineWidth

outline-width

OutlineOffset

outline-offset

PaddingTop

padding-top

PaddingLeft

padding-left

PaddingBottom

padding-bottom

PaddingRight

padding-right

Top

top

Left

left

Bottom

bottom

Right

right

ZIndex

z-index

FlexGrow

flex-grow

FlexShrink

flex-shrink

Order

order

Width

width

MinWidth

min-width

MaxWidth

max-width

Height

height

MinHeight

min-height

MaxHeight

max-height

Background

background

BackgroundPosition

background-position

BorderColor

border-color

BorderWidth

border-width

BorderTop

border-top

BorderLeft

border-left

BorderBottom

border-bottom

BorderRight

border-right

Border

border

BorderRadius

border-radius

Columns

columns

Font

font

Margin

margin

Outline

outline

Padding

padding

Flex

flex

Unsupported(CustomIdent)

Unrecognized property which could be any non-transitionable, custom property, or unknown property.

Methods

impl TransitionProperty

pub fn each<F: FnMut(&TransitionProperty)>(cb: F)

Iterates over each longhand property.

pub fn any<F: FnMut(&TransitionProperty) -> bool>(cb: F) -> bool

Iterates over every longhand property that is not TransitionProperty::All, stopping and returning true when the provided callback returns true for the first time.

pub fn parse<'i, 't>(input: &mut Parser<'i, 't>) -> Result<Self, ParseError<'i>>

Parse a transition-property value.

pub fn longhands(&self) -> &'static [TransitionProperty]

Return transitionable longhands of this shorthand TransitionProperty, except for "all".

pub fn is_shorthand(&self) -> bool

Returns true if this TransitionProperty is a shorthand.