Trait style::properties::animated_properties::ListAnimation

pub trait ListAnimation<T>: Sized {
    fn animate_repeatable_list(
        &self, 
        other: &Self, 
        procedure: Procedure
    ) -> Result<Self, ()>
    where
        T: Animate;

    fn squared_distance_repeatable_list(
        &self, 
        other: &Self
    ) -> Result<SquaredDistance, ()>
    where
        T: ComputeSquaredDistance;

    fn animate_with_zero(
        &self, 
        other: &Self, 
        procedure: Procedure
    ) -> Result<Self, ()>
    where
        T: Animate + Clone + ToAnimatedZero;

    fn squared_distance_with_zero(
        &self, 
        other: &Self
    ) -> Result<SquaredDistance, ()>
    where
        T: ToAnimatedZero + ComputeSquaredDistance;
}

A trait to abstract away the different kind of animations over a list that there may be.

Required methods

fn animate_repeatable_list(
&self,
other: &Self,
procedure: Procedure
) -> Result<Self, ()> where
T: Animate,

https://drafts.csswg.org/css-transitions/#animtype-repeatable-list

fn squared_distance_repeatable_list(
&self,
other: &Self
) -> Result<SquaredDistance, ()> where
T: ComputeSquaredDistance,

https://drafts.csswg.org/css-transitions/#animtype-repeatable-list

fn animate_with_zero(
&self,
other: &Self,
procedure: Procedure
) -> Result<Self, ()> where
T: Animate + Clone + ToAnimatedZero,

This is the animation used for some of the types like shadows and filters, where the interpolation happens with the zero value if one of the sides is not present.

fn squared_distance_with_zero(
&self,
other: &Self
) -> Result<SquaredDistance, ()> where
T: ToAnimatedZero + ComputeSquaredDistance,

This is the animation used for some of the types like shadows and filters, where the interpolation happens with the zero value if one of the sides is not present.

Loading content...