style::properties

Module declaration_block

A property declaration block.

Structs

AnimationDeclarations

A set of property declarations including animations and transitions.

AnimationValueIterator

Iterator for AnimationValue to be generated from PropertyDeclarationBlock.

DeclarationImportanceIterator

Iterator over (PropertyDeclaration, Importance) pairs.

DeclarationParserState

The state needed to parse a declaration block.

PropertyDeclarationBlock

Overridden declarations are skipped.

PropertyDeclarationIdSet

A set of properties.

PropertyDeclarationIdSetIterator

An iterator over a set of longhand ids.

PropertyDeclarationParser 🔒

A struct to parse property declarations.

SourcePropertyDeclarationUpdate

A struct describes how a declaration block should be updated by a SourcePropertyDeclaration.

Enums

AppendableValue

A convenient enum to represent different kinds of stuff that can represent a value in the serialization of a property declaration.

DeclarationUpdate 🔒

An enum describes how a declaration should update the declaration block.

Importance

A declaration importance.

Functions

alias_of_known_property 🔒

append_declaration_value

Append a given kind of appendable value to a serialization.

append_serialization

Append a given property and value pair to a serialization.

handle_first_serialization 🔒

Potentially appends whitespace after the first (property: value;) pair.

is_non_mozilla_vendor_identifier 🔒

Based on NonMozillaVendorIdentifier from Gecko’s CSS parser.

parse_one_declaration_into

Parse a given property declaration. Can result in multiple PropertyDeclarations when expanding a shorthand, for example.

parse_property_declaration_list

Parse a list of property declarations and return a property declaration block.

parse_style_attribute

A helper to parse the style attribute of an element, in order for this to be shared between Servo and Gecko.

report_one_css_error 🔒

Type Aliases

SmallParseErrorVec 🔒