Enum style::properties::declaration_block::AppendableValue

pub enum AppendableValue<'a, I> where I: Iterator<Item=&'a PropertyDeclaration> {
    Declaration(&'a PropertyDeclaration),
    DeclarationsForShorthand(ShorthandId, I),
    Css(&'a str),
}

A convenient enum to represent different kinds of stuff that can represent a value in the serialization of a property declaration.

Variants
Declaration(&'a PropertyDeclaration)

A given declaration, of which we'll serialize just the value.

DeclarationsForShorthand(ShorthandId, I)

A set of declarations for a given shorthand.

FIXME: This needs more docs, where are the shorthands expanded? We print the property name before-hand, don't we?

Css(&'a str)

A raw CSS string, coming for example from a property with CSS variables.