Function style::properties::declaration_block::parse_style_attribute

pub fn parse_style_attribute<R>(
    input: &str, 
    url_data: &UrlExtraData, 
    error_reporter: &R, 
    quirks_mode: QuirksMode
) -> PropertyDeclarationBlock where
    R: ParseErrorReporter, 

A helper to parse the style attribute of an element, in order for this to be shared between Servo and Gecko.