style::properties::declaration_block

Struct DeclarationParserState

pub struct DeclarationParserState<'i> {
    output_block: PropertyDeclarationBlock,
    declarations: SourcePropertyDeclaration,
    importance: Importance,
    errors: SmallVec<[(ParseError<'i>, &'i str, Option<PropertyId>); 2]>,
    first_declaration_start: SourceLocation,
    last_parsed_property_id: Option<PropertyId>,
}

The state needed to parse a declaration block.

It stores declarations in output_block.

Fields
output_block: PropertyDeclarationBlock

The output block where results are stored.

declarations: SourcePropertyDeclaration

Declarations from the last declaration parsed. (note that a shorthand might expand to multiple declarations).

importance: Importance

The importance from the last declaration parsed.

errors: SmallVec<[(ParseError<'i>, &'i str, Option<PropertyId>); 2]>

A list of errors that have happened so far. Not all of them might be reported.

first_declaration_start: SourceLocation

The start of the first declaration

last_parsed_property_id: Option<PropertyId>

The last parsed property id, if any.

Implementations

impl<'i> DeclarationParserState<'i>

pub fn first_declaration_start(&self) -> SourceLocation

Getter for first_declaration_start.

pub fn has_parsed_declarations(&self) -> bool

Returns whether any parsed declarations have been parsed so far.

pub fn take_declarations(&mut self) -> PropertyDeclarationBlock

Takes the parsed declarations.

pub fn parse_value<'t>( &mut self, context: &ParserContext<'_>, name: CowRcStr<'i>, input: &mut Parser<'i, 't>, declaration_start: &ParserState, ) -> Result<(), ParseError<'i>>

Parse a single declaration value.

pub fn report_errors_if_needed( &mut self, context: &ParserContext<'_>, selectors: &[SelectorList<SelectorImpl>], )

Reports any CSS errors that have ocurred if needed.

fn do_report_css_errors( &mut self, context: &ParserContext<'_>, selectors: &[SelectorList<SelectorImpl>], )

pub fn did_error( &mut self, context: &ParserContext<'_>, error: ParseError<'i>, slice: &'i str, )

Resets the declaration parser state, and reports the error if needed.