Enum style::properties::ParsedDeclaration

pub enum ParsedDeclaration {
    Background(Longhands),
    BackgroundCSSWideKeyword(CSSWideKeyword),
    BackgroundWithVariables(Arc<UnparsedValue>),
    BackgroundPosition(Longhands),
    BackgroundPositionCSSWideKeyword(CSSWideKeyword),
    BackgroundPositionWithVariables(Arc<UnparsedValue>),
    BorderColor(Longhands),
    BorderColorCSSWideKeyword(CSSWideKeyword),
    BorderColorWithVariables(Arc<UnparsedValue>),
    BorderStyle(Longhands),
    BorderStyleCSSWideKeyword(CSSWideKeyword),
    BorderStyleWithVariables(Arc<UnparsedValue>),
    BorderWidth(Longhands),
    BorderWidthCSSWideKeyword(CSSWideKeyword),
    BorderWidthWithVariables(Arc<UnparsedValue>),
    BorderTop(Longhands),
    BorderTopCSSWideKeyword(CSSWideKeyword),
    BorderTopWithVariables(Arc<UnparsedValue>),
    BorderLeft(Longhands),
    BorderLeftCSSWideKeyword(CSSWideKeyword),
    BorderLeftWithVariables(Arc<UnparsedValue>),
    BorderBottom(Longhands),
    BorderBottomCSSWideKeyword(CSSWideKeyword),
    BorderBottomWithVariables(Arc<UnparsedValue>),
    BorderRight(Longhands),
    BorderRightCSSWideKeyword(CSSWideKeyword),
    BorderRightWithVariables(Arc<UnparsedValue>),
    BorderBlockStart(Longhands),
    BorderBlockStartCSSWideKeyword(CSSWideKeyword),
    BorderBlockStartWithVariables(Arc<UnparsedValue>),
    BorderBlockEnd(Longhands),
    BorderBlockEndCSSWideKeyword(CSSWideKeyword),
    BorderBlockEndWithVariables(Arc<UnparsedValue>),
    BorderInlineStart(Longhands),
    BorderInlineStartCSSWideKeyword(CSSWideKeyword),
    BorderInlineStartWithVariables(Arc<UnparsedValue>),
    BorderInlineEnd(Longhands),
    BorderInlineEndCSSWideKeyword(CSSWideKeyword),
    BorderInlineEndWithVariables(Arc<UnparsedValue>),
    Border(Longhands),
    BorderCSSWideKeyword(CSSWideKeyword),
    BorderWithVariables(Arc<UnparsedValue>),
    BorderRadius(Longhands),
    BorderRadiusCSSWideKeyword(CSSWideKeyword),
    BorderRadiusWithVariables(Arc<UnparsedValue>),
    BorderImage(Longhands),
    BorderImageCSSWideKeyword(CSSWideKeyword),
    BorderImageWithVariables(Arc<UnparsedValue>),
    Overflow(Longhands),
    OverflowCSSWideKeyword(CSSWideKeyword),
    OverflowWithVariables(Arc<UnparsedValue>),
    Transition(Longhands),
    TransitionCSSWideKeyword(CSSWideKeyword),
    TransitionWithVariables(Arc<UnparsedValue>),
    Animation(Longhands),
    AnimationCSSWideKeyword(CSSWideKeyword),
    AnimationWithVariables(Arc<UnparsedValue>),
    Columns(Longhands),
    ColumnsCSSWideKeyword(CSSWideKeyword),
    ColumnsWithVariables(Arc<UnparsedValue>),
    Font(Longhands),
    FontCSSWideKeyword(CSSWideKeyword),
    FontWithVariables(Arc<UnparsedValue>),
    ListStyle(Longhands),
    ListStyleCSSWideKeyword(CSSWideKeyword),
    ListStyleWithVariables(Arc<UnparsedValue>),
    Margin(Longhands),
    MarginCSSWideKeyword(CSSWideKeyword),
    MarginWithVariables(Arc<UnparsedValue>),
    Outline(Longhands),
    OutlineCSSWideKeyword(CSSWideKeyword),
    OutlineWithVariables(Arc<UnparsedValue>),
    Padding(Longhands),
    PaddingCSSWideKeyword(CSSWideKeyword),
    PaddingWithVariables(Arc<UnparsedValue>),
    FlexFlow(Longhands),
    FlexFlowCSSWideKeyword(CSSWideKeyword),
    FlexFlowWithVariables(Arc<UnparsedValue>),
    Flex(Longhands),
    FlexCSSWideKeyword(CSSWideKeyword),
    FlexWithVariables(Arc<UnparsedValue>),
    TextDecoration(Longhands),
    TextDecorationCSSWideKeyword(CSSWideKeyword),
    TextDecorationWithVariables(Arc<UnparsedValue>),
    LonghandOrCustom(PropertyDeclaration),
}

Includes shorthands before expansion

Variants
Background(Longhands)

background

BackgroundCSSWideKeyword(CSSWideKeyword)

background with a CSS-wide keyword

BackgroundWithVariables(Arc<UnparsedValue>)

background with var() functions

BackgroundPosition(Longhands)

background-position

BackgroundPositionCSSWideKeyword(CSSWideKeyword)

background-position with a CSS-wide keyword

BackgroundPositionWithVariables(Arc<UnparsedValue>)

background-position with var() functions

BorderColor(Longhands)

border-color

BorderColorCSSWideKeyword(CSSWideKeyword)

border-color with a CSS-wide keyword

BorderColorWithVariables(Arc<UnparsedValue>)

border-color with var() functions

BorderStyle(Longhands)

border-style

BorderStyleCSSWideKeyword(CSSWideKeyword)

border-style with a CSS-wide keyword

BorderStyleWithVariables(Arc<UnparsedValue>)

border-style with var() functions

BorderWidth(Longhands)

border-width

BorderWidthCSSWideKeyword(CSSWideKeyword)

border-width with a CSS-wide keyword

BorderWidthWithVariables(Arc<UnparsedValue>)

border-width with var() functions

BorderTop(Longhands)

border-top

BorderTopCSSWideKeyword(CSSWideKeyword)

border-top with a CSS-wide keyword

BorderTopWithVariables(Arc<UnparsedValue>)

border-top with var() functions

BorderLeft(Longhands)

border-left

BorderLeftCSSWideKeyword(CSSWideKeyword)

border-left with a CSS-wide keyword

BorderLeftWithVariables(Arc<UnparsedValue>)

border-left with var() functions

BorderBottom(Longhands)

border-bottom

BorderBottomCSSWideKeyword(CSSWideKeyword)

border-bottom with a CSS-wide keyword

BorderBottomWithVariables(Arc<UnparsedValue>)

border-bottom with var() functions

BorderRight(Longhands)

border-right

BorderRightCSSWideKeyword(CSSWideKeyword)

border-right with a CSS-wide keyword

BorderRightWithVariables(Arc<UnparsedValue>)

border-right with var() functions

BorderBlockStart(Longhands)

border-block-start

BorderBlockStartCSSWideKeyword(CSSWideKeyword)

border-block-start with a CSS-wide keyword

BorderBlockStartWithVariables(Arc<UnparsedValue>)

border-block-start with var() functions

BorderBlockEnd(Longhands)

border-block-end

BorderBlockEndCSSWideKeyword(CSSWideKeyword)

border-block-end with a CSS-wide keyword

BorderBlockEndWithVariables(Arc<UnparsedValue>)

border-block-end with var() functions

BorderInlineStart(Longhands)

border-inline-start

BorderInlineStartCSSWideKeyword(CSSWideKeyword)

border-inline-start with a CSS-wide keyword

BorderInlineStartWithVariables(Arc<UnparsedValue>)

border-inline-start with var() functions

BorderInlineEnd(Longhands)

border-inline-end

BorderInlineEndCSSWideKeyword(CSSWideKeyword)

border-inline-end with a CSS-wide keyword

BorderInlineEndWithVariables(Arc<UnparsedValue>)

border-inline-end with var() functions

Border(Longhands)

border

BorderCSSWideKeyword(CSSWideKeyword)

border with a CSS-wide keyword

BorderWithVariables(Arc<UnparsedValue>)

border with var() functions

BorderRadius(Longhands)

border-radius

BorderRadiusCSSWideKeyword(CSSWideKeyword)

border-radius with a CSS-wide keyword

BorderRadiusWithVariables(Arc<UnparsedValue>)

border-radius with var() functions

BorderImage(Longhands)

border-image

BorderImageCSSWideKeyword(CSSWideKeyword)

border-image with a CSS-wide keyword

BorderImageWithVariables(Arc<UnparsedValue>)

border-image with var() functions

Overflow(Longhands)

overflow

OverflowCSSWideKeyword(CSSWideKeyword)

overflow with a CSS-wide keyword

OverflowWithVariables(Arc<UnparsedValue>)

overflow with var() functions

Transition(Longhands)

transition

TransitionCSSWideKeyword(CSSWideKeyword)

transition with a CSS-wide keyword

TransitionWithVariables(Arc<UnparsedValue>)

transition with var() functions

Animation(Longhands)

animation

AnimationCSSWideKeyword(CSSWideKeyword)

animation with a CSS-wide keyword

AnimationWithVariables(Arc<UnparsedValue>)

animation with var() functions

Columns(Longhands)

columns

ColumnsCSSWideKeyword(CSSWideKeyword)

columns with a CSS-wide keyword

ColumnsWithVariables(Arc<UnparsedValue>)

columns with var() functions

Font(Longhands)

font

FontCSSWideKeyword(CSSWideKeyword)

font with a CSS-wide keyword

FontWithVariables(Arc<UnparsedValue>)

font with var() functions

ListStyle(Longhands)

list-style

ListStyleCSSWideKeyword(CSSWideKeyword)

list-style with a CSS-wide keyword

ListStyleWithVariables(Arc<UnparsedValue>)

list-style with var() functions

Margin(Longhands)

margin

MarginCSSWideKeyword(CSSWideKeyword)

margin with a CSS-wide keyword

MarginWithVariables(Arc<UnparsedValue>)

margin with var() functions

Outline(Longhands)

outline

OutlineCSSWideKeyword(CSSWideKeyword)

outline with a CSS-wide keyword

OutlineWithVariables(Arc<UnparsedValue>)

outline with var() functions

Padding(Longhands)

padding

PaddingCSSWideKeyword(CSSWideKeyword)

padding with a CSS-wide keyword

PaddingWithVariables(Arc<UnparsedValue>)

padding with var() functions

FlexFlow(Longhands)

flex-flow

FlexFlowCSSWideKeyword(CSSWideKeyword)

flex-flow with a CSS-wide keyword

FlexFlowWithVariables(Arc<UnparsedValue>)

flex-flow with var() functions

Flex(Longhands)

flex

FlexCSSWideKeyword(CSSWideKeyword)

flex with a CSS-wide keyword

FlexWithVariables(Arc<UnparsedValue>)

flex with var() functions

TextDecoration(Longhands)

text-decoration

TextDecorationCSSWideKeyword(CSSWideKeyword)

text-decoration with a CSS-wide keyword

TextDecorationWithVariables(Arc<UnparsedValue>)

text-decoration with var() functions

LonghandOrCustom(PropertyDeclaration)

Not a shorthand

Methods

impl ParsedDeclaration

fn expand<F>(self, push: F) where F: FnMut(PropertyDeclaration)

Transform this ParsedDeclaration into a sequence of PropertyDeclaration by expanding shorthand declarations into their corresponding longhands

fn parse(id: PropertyId,
context: &ParserContext,
input: &mut Parser,
in_keyframe_block: bool)
-> Result<ParsedDeclaration, PropertyDeclarationParseError>

The in_keyframe_block parameter controls this:

https://drafts.csswg.org/css-animations/#keyframes

The inside of accepts any CSS property except those defined in this specification, but does accept the animation-play-state property and interprets it specially.

This will not actually parse Importance values, and will always set things to Importance::Normal. Parsing Importance values is the job of PropertyDeclarationParser, we only set them here so that we don't have to reallocate