Enum style::properties::Shorthand

pub enum Shorthand {
    Background,
    BorderColor,
    BorderStyle,
    BorderWidth,
    BorderTop,
    BorderRight,
    BorderBottom,
    BorderLeft,
    Border,
    BorderRadius,
    Overflow,
    Transition,
    Columns,
    Font,
    WordWrap,
    ListStyle,
    Margin,
    Outline,
    Padding,
}

Variants
BackgroundBorderColorBorderStyleBorderWidthBorderTopBorderRightBorderBottomBorderLeftBorderBorderRadiusOverflowTransitionColumnsFontWordWrapListStyleMarginOutlinePadding
Methods

impl Shorthand

fn from_name(name: &str) -> Option<Shorthand>

fn name(&self) -> &'static str

fn longhands(&self) -> &'static [&'static str]

fn serialize_shorthand_value_to_string<'a, I>(self, declarations: I, is_important: bool) -> String where I: Iterator<Item=&'a PropertyDeclaration> + Clone

Serializes possible shorthand value to String.

fn serialize_shorthand_to_buffer<'a, W, I>(self, dest: &mut W, declarations: I, is_first_serialization: &mut bool) -> Result<bool, Error> where W: Write, I: Iterator<Item=&'a PropertyDeclaration> + Clone

Serializes possible shorthand name with value to input buffer given a list of longhand declarations. On success, returns true if shorthand value is written and false if no shorthand value is present.

fn get_shorthand_appendable_value<'a, I>(self, declarations: I) -> Option<AppendableValue<'a, I>> where I: Iterator<Item=&'a PropertyDeclaration> + Clone