Function style::properties::cascade

pub fn cascade(viewport_size: Size2D<Au>, applicable_declarations: &[DeclarationBlock<Vec<PropertyDeclaration>>], shareable: bool, parent_style: Option<&ComputedValues>, cached_style: Option<&ComputedValues>) -> (ComputedValues, bool)

Performs the CSS cascade, computing new styles for an element from its parent style and optionally a cached related style. The arguments are:

viewport_size: The size of the initial viewport.

applicable_declarations: The list of CSS rules that matched.

shareable: Whether the ComputedValues structure to be constructed should be considered shareable.

parent_style: The parent style, if applicable; if None, this is the root node.

cached_style: If present, cascading is short-circuited for everything but inherited values and these values are used instead. Obviously, you must be careful when supplying this that it is safe to only provide inherited declarations. If parent_style is None, this is ignored.

Returns the computed values and a boolean indicating whether the result is cacheable.