style::properties::generated

Enum CountedUnknownProperty

#[repr(u32)]
pub enum CountedUnknownProperty {

    WebkitTapHighlightColor = 0,
    Speak = 1,
    TextSizeAdjust = 2,
    WebkitUserDrag = 3,
    Orphans = 4,
    Widows = 5,
    WebkitUserModify = 6,
    WebkitMarginBefore = 7,
    WebkitMarginAfter = 8,
    WebkitMarginStart = 9,
    WebkitColumnBreakInside = 10,
    WebkitPaddingStart = 11,
    WebkitMarginEnd = 12,
    WebkitBoxReflect = 13,
    WebkitPrintColorAdjust = 14,
    WebkitMaskBoxImage = 15,
    WebkitLineBreak = 16,
    WebkitWritingMode = 17,
    WebkitHyphenateCharacter = 18,
    WebkitHighlight = 19,
    BackgroundRepeatX = 20,
    WebkitPaddingEnd = 21,
    BackgroundRepeatY = 22,
    WebkitTextEmphasisColor = 23,
    WebkitMarginTopCollapse = 24,
    WebkitRtlOrdering = 25,
    WebkitPaddingBefore = 26,
    WebkitTextDecorationsInEffect = 27,
    WebkitBorderVerticalSpacing = 28,
    WebkitLocale = 29,
    WebkitPaddingAfter = 30,
    WebkitBorderHorizontalSpacing = 31,
    ColorRendering = 32,
    WebkitColumnBreakBefore = 33,
    WebkitTransformOriginX = 34,
    WebkitTransformOriginY = 35,
    WebkitTextEmphasisPosition = 36,
    BufferedRendering = 37,
    WebkitTextOrientation = 38,
    WebkitTextCombine = 39,
    WebkitTextEmphasisStyle = 40,
    WebkitTextEmphasis = 41,
    WebkitMaskBoxImageWidth = 42,
    WebkitMaskBoxImageSource = 43,
    WebkitMaskBoxImageOutset = 44,
    WebkitMaskBoxImageSlice = 45,
    WebkitMaskBoxImageRepeat = 46,
    WebkitMarginAfterCollapse = 47,
    WebkitBorderBeforeColor = 48,
    WebkitBorderBeforeWidth = 49,
    WebkitPerspectiveOriginX = 50,
    WebkitPerspectiveOriginY = 51,
    WebkitMarginBeforeCollapse = 52,
    WebkitBorderBeforeStyle = 53,
    WebkitMarginBottomCollapse = 54,
    WebkitRubyPosition = 55,
    WebkitColumnBreakAfter = 56,
    WebkitMarginCollapse = 57,
    WebkitBorderBefore = 58,
    WebkitBorderEnd = 59,
    WebkitBorderAfter = 60,
    WebkitBorderStart = 61,
    WebkitMinLogicalWidth = 62,
    WebkitLogicalHeight = 63,
    WebkitTransformOriginZ = 64,
    WebkitFontSizeDelta = 65,
    WebkitLogicalWidth = 66,
    WebkitMaxLogicalWidth = 67,
    WebkitMinLogicalHeight = 68,
    WebkitMaxLogicalHeight = 69,
    WebkitBorderEndColor = 70,
    WebkitBorderEndWidth = 71,
    WebkitBorderStartColor = 72,
    WebkitBorderStartWidth = 73,
    WebkitBorderAfterColor = 74,
    WebkitBorderAfterWidth = 75,
    WebkitBorderEndStyle = 76,
    WebkitBorderAfterStyle = 77,
    WebkitBorderStartStyle = 78,
    WebkitMaskRepeatX = 79,
    WebkitMaskRepeatY = 80,
    UserZoom = 81,
    MinZoom = 82,
    WebkitBoxDecorationBreak = 83,
    Orientation = 84,
    MaxZoom = 85,
    WebkitAppRegion = 86,
    WebkitColumnRule = 87,
    WebkitColumnSpan = 88,
    WebkitColumnGap = 89,
    WebkitShapeOutside = 90,
    WebkitColumnRuleWidth = 91,
    WebkitColumnCount = 92,
    WebkitOpacity = 93,
    WebkitColumnWidth = 94,
    WebkitShapeImageThreshold = 95,
    WebkitColumnRuleStyle = 96,
    WebkitColumns = 97,
    WebkitColumnRuleColor = 98,
    WebkitShapeMargin = 99,

}

The counted unknown property list which is used for css use counters.

FIXME: This should be just #[repr(u8)], but can’t be because of ABI issues, see https://bugs.llvm.org/show_bug.cgi?id=44228.

Variants

WebkitTapHighlightColor = 0

-webkit-tap-highlight-color

Speak = 1

speak

TextSizeAdjust = 2

text-size-adjust

WebkitUserDrag = 3

-webkit-user-drag

Orphans = 4

orphans

Widows = 5

widows

WebkitUserModify = 6

-webkit-user-modify

WebkitMarginBefore = 7

-webkit-margin-before

WebkitMarginAfter = 8

-webkit-margin-after

WebkitMarginStart = 9

-webkit-margin-start

WebkitColumnBreakInside = 10

-webkit-column-break-inside

WebkitPaddingStart = 11

-webkit-padding-start

WebkitMarginEnd = 12

-webkit-margin-end

WebkitBoxReflect = 13

-webkit-box-reflect

WebkitPrintColorAdjust = 14

-webkit-print-color-adjust

WebkitMaskBoxImage = 15

-webkit-mask-box-image

WebkitLineBreak = 16

-webkit-line-break

WebkitWritingMode = 17

-webkit-writing-mode

WebkitHyphenateCharacter = 18

-webkit-hyphenate-character

WebkitHighlight = 19

-webkit-highlight

BackgroundRepeatX = 20

background-repeat-x

WebkitPaddingEnd = 21

-webkit-padding-end

BackgroundRepeatY = 22

background-repeat-y

WebkitTextEmphasisColor = 23

-webkit-text-emphasis-color

WebkitMarginTopCollapse = 24

-webkit-margin-top-collapse

WebkitRtlOrdering = 25

-webkit-rtl-ordering

WebkitPaddingBefore = 26

-webkit-padding-before

WebkitTextDecorationsInEffect = 27

-webkit-text-decorations-in-effect

WebkitBorderVerticalSpacing = 28

-webkit-border-vertical-spacing

WebkitLocale = 29

-webkit-locale

WebkitPaddingAfter = 30

-webkit-padding-after

WebkitBorderHorizontalSpacing = 31

-webkit-border-horizontal-spacing

ColorRendering = 32

color-rendering

WebkitColumnBreakBefore = 33

-webkit-column-break-before

WebkitTransformOriginX = 34

-webkit-transform-origin-x

WebkitTransformOriginY = 35

-webkit-transform-origin-y

WebkitTextEmphasisPosition = 36

-webkit-text-emphasis-position

BufferedRendering = 37

buffered-rendering

WebkitTextOrientation = 38

-webkit-text-orientation

WebkitTextCombine = 39

-webkit-text-combine

WebkitTextEmphasisStyle = 40

-webkit-text-emphasis-style

WebkitTextEmphasis = 41

-webkit-text-emphasis

WebkitMaskBoxImageWidth = 42

-webkit-mask-box-image-width

WebkitMaskBoxImageSource = 43

-webkit-mask-box-image-source

WebkitMaskBoxImageOutset = 44

-webkit-mask-box-image-outset

WebkitMaskBoxImageSlice = 45

-webkit-mask-box-image-slice

WebkitMaskBoxImageRepeat = 46

-webkit-mask-box-image-repeat

WebkitMarginAfterCollapse = 47

-webkit-margin-after-collapse

WebkitBorderBeforeColor = 48

-webkit-border-before-color

WebkitBorderBeforeWidth = 49

-webkit-border-before-width

WebkitPerspectiveOriginX = 50

-webkit-perspective-origin-x

WebkitPerspectiveOriginY = 51

-webkit-perspective-origin-y

WebkitMarginBeforeCollapse = 52

-webkit-margin-before-collapse

WebkitBorderBeforeStyle = 53

-webkit-border-before-style

WebkitMarginBottomCollapse = 54

-webkit-margin-bottom-collapse

WebkitRubyPosition = 55

-webkit-ruby-position

WebkitColumnBreakAfter = 56

-webkit-column-break-after

WebkitMarginCollapse = 57

-webkit-margin-collapse

WebkitBorderBefore = 58

-webkit-border-before

WebkitBorderEnd = 59

-webkit-border-end

WebkitBorderAfter = 60

-webkit-border-after

WebkitBorderStart = 61

-webkit-border-start

WebkitMinLogicalWidth = 62

-webkit-min-logical-width

WebkitLogicalHeight = 63

-webkit-logical-height

WebkitTransformOriginZ = 64

-webkit-transform-origin-z

WebkitFontSizeDelta = 65

-webkit-font-size-delta

WebkitLogicalWidth = 66

-webkit-logical-width

WebkitMaxLogicalWidth = 67

-webkit-max-logical-width

WebkitMinLogicalHeight = 68

-webkit-min-logical-height

WebkitMaxLogicalHeight = 69

-webkit-max-logical-height

WebkitBorderEndColor = 70

-webkit-border-end-color

WebkitBorderEndWidth = 71

-webkit-border-end-width

WebkitBorderStartColor = 72

-webkit-border-start-color

WebkitBorderStartWidth = 73

-webkit-border-start-width

WebkitBorderAfterColor = 74

-webkit-border-after-color

WebkitBorderAfterWidth = 75

-webkit-border-after-width

WebkitBorderEndStyle = 76

-webkit-border-end-style

WebkitBorderAfterStyle = 77

-webkit-border-after-style

WebkitBorderStartStyle = 78

-webkit-border-start-style

WebkitMaskRepeatX = 79

-webkit-mask-repeat-x

WebkitMaskRepeatY = 80

-webkit-mask-repeat-y

UserZoom = 81

user-zoom

MinZoom = 82

min-zoom

WebkitBoxDecorationBreak = 83

-webkit-box-decoration-break

Orientation = 84

orientation

MaxZoom = 85

max-zoom

WebkitAppRegion = 86

-webkit-app-region

WebkitColumnRule = 87

-webkit-column-rule

WebkitColumnSpan = 88

-webkit-column-span

WebkitColumnGap = 89

-webkit-column-gap

WebkitShapeOutside = 90

-webkit-shape-outside

WebkitColumnRuleWidth = 91

-webkit-column-rule-width

WebkitColumnCount = 92

-webkit-column-count

WebkitOpacity = 93

-webkit-opacity

WebkitColumnWidth = 94

-webkit-column-width

WebkitShapeImageThreshold = 95

-webkit-shape-image-threshold

WebkitColumnRuleStyle = 96

-webkit-column-rule-style

WebkitColumns = 97

-webkit-columns

WebkitColumnRuleColor = 98

-webkit-column-rule-color

WebkitShapeMargin = 99

-webkit-shape-margin

Implementations

impl CountedUnknownProperty

pub fn parse_for_testing(property_name: &str) -> Option<Self>

Parse the counted unknown property, for testing purposes only.

pub fn bit(self) -> usize

Returns the underlying index, used for use counter.