style::properties::generated

Enum LonghandId

#[repr(u16)]
pub enum LonghandId {

    AlignItems = 0,
    AlignmentBaseline = 1,
    BackfaceVisibility = 2,
    BaselineSource = 3,
    BorderCollapse = 4,
    BorderImageRepeat = 5,
    BoxDecorationBreak = 6,
    BoxSizing = 7,
    CaptionSide = 8,
    Clear = 9,
    ColumnSpan = 10,
    Contain = 11,
    ContainerType = 12,
    Direction = 13,
    Display = 14,
    EmptyCells = 15,
    FillRule = 16,
    FlexDirection = 17,
    FlexWrap = 18,
    Float = 19,
    FontKerning = 20,
    FontLanguageOverride = 21,
    FontOpticalSizing = 22,
    FontSynthesisWeight = 23,
    FontVariantCaps = 24,
    FontVariantEastAsian = 25,
    FontVariantLigatures = 26,
    FontVariantNumeric = 27,
    FontVariantPosition = 28,
    ForcedColorAdjust = 29,
    GridAutoFlow = 30,
    ImageRendering = 31,
    Isolation = 32,
    JustifyItems = 33,
    LineBreak = 34,
    ListStylePosition = 35,
    MaskType = 36,
    MathStyle = 37,
    MixBlendMode = 38,
    ObjectFit = 39,
    OutlineStyle = 40,
    OverflowWrap = 41,
    PointerEvents = 42,
    Position = 43,
    PositionArea = 44,
    ScrollBehavior = 45,
    ScrollbarWidth = 46,
    ServoTopLayer = 47,
    StrokeLinecap = 48,
    StrokeLinejoin = 49,
    TableLayout = 50,
    TextAlign = 51,
    TextAlignLast = 52,
    TextDecorationLine = 53,
    TextDecorationStyle = 54,
    TextJustify = 55,
    TextOrientation = 56,
    TextRendering = 57,
    TextTransform = 58,
    TextUnderlinePosition = 59,
    TextWrapMode = 60,
    TouchAction = 61,
    TransformStyle = 62,
    UnicodeBidi = 63,
    UserSelect = 64,
    Visibility = 65,
    WebkitTextSecurity = 66,
    WhiteSpaceCollapse = 67,
    WordBreak = 68,
    WritingMode = 69,
    XTextScale = 70,
    Appearance = 71,
    MozDefaultAppearance = 72,
    AlignContent = 73,
    JustifyContent = 74,
    AlignSelf = 75,
    JustifySelf = 76,
    OverflowBlock = 77,
    OverflowInline = 78,
    OverflowX = 79,
    OverflowY = 80,
    OverscrollBehaviorBlock = 81,
    OverscrollBehaviorInline = 82,
    OverscrollBehaviorX = 83,
    OverscrollBehaviorY = 84,
    BorderBlockEndStyle = 85,
    BorderBlockStartStyle = 86,
    BorderBottomStyle = 87,
    BorderInlineEndStyle = 88,
    BorderInlineStartStyle = 89,
    BorderLeftStyle = 90,
    BorderRightStyle = 91,
    BorderTopStyle = 92,
    AnimationComposition = 93,
    AnimationDelay = 94,
    AnimationDirection = 95,
    AnimationDuration = 96,
    AnimationFillMode = 97,
    AnimationIterationCount = 98,
    AnimationName = 99,
    AnimationPlayState = 100,
    AnimationRangeEnd = 101,
    AnimationRangeStart = 102,
    AnimationTimeline = 103,
    AnimationTimingFunction = 104,
    AspectRatio = 105,
    BackdropFilter = 106,
    BackgroundAttachment = 107,
    BackgroundBlendMode = 108,
    BackgroundClip = 109,
    BackgroundImage = 110,
    BackgroundOrigin = 111,
    BackgroundPositionX = 112,
    BackgroundPositionY = 113,
    BackgroundRepeat = 114,
    BackgroundSize = 115,
    BaselineShift = 116,
    BorderImageOutset = 117,
    BorderImageSlice = 118,
    BorderImageWidth = 119,
    BorderSpacing = 120,
    BoxShadow = 121,
    CaretColor = 122,
    Clip = 123,
    ClipPath = 124,
    Color = 125,
    ColorScheme = 126,
    ColumnCount = 127,
    ColumnWidth = 128,
    ContainerName = 129,
    Content = 130,
    CounterIncrement = 131,
    CounterReset = 132,
    Cursor = 133,
    D = 134,
    Filter = 135,
    FlexBasis = 136,
    FontFamily = 137,
    FontFeatureSettings = 138,
    FontSize = 139,
    FontSizeAdjust = 140,
    FontStretch = 141,
    FontStyle = 142,
    FontVariantAlternates = 143,
    FontVariationSettings = 144,
    FontWeight = 145,
    GridTemplateAreas = 146,
    LetterSpacing = 147,
    LineHeight = 148,
    ListStyleType = 149,
    MaskClip = 150,
    MaskComposite = 151,
    MaskImage = 152,
    MaskMode = 153,
    MaskOrigin = 154,
    MaskPositionX = 155,
    MaskPositionY = 156,
    MaskRepeat = 157,
    MaskSize = 158,
    MathDepth = 159,
    MozMinFontSizeRatio = 160,
    OffsetPath = 161,
    Opacity = 162,
    Order = 163,
    OutlineOffset = 164,
    OverflowClipMargin = 165,
    Perspective = 166,
    PositionTryFallbacks = 167,
    Quotes = 168,
    Rotate = 169,
    Scale = 170,
    ScrollbarColor = 171,
    StrokeDasharray = 172,
    StrokeDashoffset = 173,
    StrokeWidth = 174,
    TabSize = 175,
    TextDecorationInset = 176,
    TextDecorationThickness = 177,
    TextIndent = 178,
    TextOverflow = 179,
    TextShadow = 180,
    TextUnderlineOffset = 181,
    Transform = 182,
    TransformOrigin = 183,
    TransitionBehavior = 184,
    TransitionDelay = 185,
    TransitionDuration = 186,
    TransitionProperty = 187,
    TransitionTimingFunction = 188,
    Translate = 189,
    ViewTransitionClass = 190,
    ViewTransitionName = 191,
    WillChange = 192,
    WordSpacing = 193,
    XLang = 194,
    ZIndex = 195,
    Zoom = 196,
    ObjectPosition = 197,
    PerspectiveOrigin = 198,
    Fill = 199,
    Stroke = 200,
    GridTemplateColumns = 201,
    GridTemplateRows = 202,
    BorderImageSource = 203,
    ListStyleImage = 204,
    GridAutoColumns = 205,
    GridAutoRows = 206,
    Rx = 207,
    Ry = 208,
    FillOpacity = 209,
    StrokeOpacity = 210,
    ColumnGap = 211,
    RowGap = 212,
    FlexGrow = 213,
    FlexShrink = 214,
    StrokeMiterlimit = 215,
    GridColumnEnd = 216,
    GridColumnStart = 217,
    GridRowEnd = 218,
    GridRowStart = 219,
    Cx = 220,
    Cy = 221,
    X = 222,
    Y = 223,
    MaxBlockSize = 224,
    MaxHeight = 225,
    MaxInlineSize = 226,
    MaxWidth = 227,
    BorderBottomLeftRadius = 228,
    BorderBottomRightRadius = 229,
    BorderEndEndRadius = 230,
    BorderEndStartRadius = 231,
    BorderStartEndRadius = 232,
    BorderStartStartRadius = 233,
    BorderTopLeftRadius = 234,
    BorderTopRightRadius = 235,
    CornerBottomLeftShape = 236,
    CornerBottomRightShape = 237,
    CornerEndEndShape = 238,
    CornerEndStartShape = 239,
    CornerStartEndShape = 240,
    CornerStartStartShape = 241,
    CornerTopLeftShape = 242,
    CornerTopRightShape = 243,
    Bottom = 244,
    InsetBlockEnd = 245,
    InsetBlockStart = 246,
    InsetInlineEnd = 247,
    InsetInlineStart = 248,
    Left = 249,
    Right = 250,
    Top = 251,
    MarginBlockEnd = 252,
    MarginBlockStart = 253,
    MarginBottom = 254,
    MarginInlineEnd = 255,
    MarginInlineStart = 256,
    MarginLeft = 257,
    MarginRight = 258,
    MarginTop = 259,
    BlockSize = 260,
    Height = 261,
    InlineSize = 262,
    MinBlockSize = 263,
    MinHeight = 264,
    MinInlineSize = 265,
    MinWidth = 266,
    Width = 267,
    BorderBlockEndWidth = 268,
    BorderBlockStartWidth = 269,
    BorderBottomWidth = 270,
    BorderInlineEndWidth = 271,
    BorderInlineStartWidth = 272,
    BorderLeftWidth = 273,
    BorderRightWidth = 274,
    BorderTopWidth = 275,
    OutlineWidth = 276,
    PaddingBlockEnd = 277,
    PaddingBlockStart = 278,
    PaddingBottom = 279,
    PaddingInlineEnd = 280,
    PaddingInlineStart = 281,
    PaddingLeft = 282,
    PaddingRight = 283,
    PaddingTop = 284,
    R = 285,
    BackgroundColor = 286,
    BorderBlockEndColor = 287,
    BorderBlockStartColor = 288,
    BorderBottomColor = 289,
    BorderInlineEndColor = 290,
    BorderInlineStartColor = 291,
    BorderLeftColor = 292,
    BorderRightColor = 293,
    BorderTopColor = 294,
    OutlineColor = 295,
    TextDecorationColor = 296,

}

An identifier for a given longhand property.

Variants

AlignItems = 0

align-items

AlignmentBaseline = 1

alignment-baseline

BackfaceVisibility = 2

backface-visibility

BaselineSource = 3

baseline-source

BorderCollapse = 4

border-collapse

BorderImageRepeat = 5

border-image-repeat

BoxDecorationBreak = 6

box-decoration-break

BoxSizing = 7

box-sizing

CaptionSide = 8

caption-side

Clear = 9

clear

ColumnSpan = 10

column-span

Contain = 11

contain

ContainerType = 12

container-type

Direction = 13

direction

Display = 14

display

EmptyCells = 15

empty-cells

FillRule = 16

fill-rule

FlexDirection = 17

flex-direction

FlexWrap = 18

flex-wrap

Float = 19

float

FontKerning = 20

font-kerning

FontLanguageOverride = 21

font-language-override

FontOpticalSizing = 22

font-optical-sizing

FontSynthesisWeight = 23

font-synthesis-weight

FontVariantCaps = 24

font-variant-caps

FontVariantEastAsian = 25

font-variant-east-asian

FontVariantLigatures = 26

font-variant-ligatures

FontVariantNumeric = 27

font-variant-numeric

FontVariantPosition = 28

font-variant-position

ForcedColorAdjust = 29

forced-color-adjust

GridAutoFlow = 30

grid-auto-flow

ImageRendering = 31

image-rendering

Isolation = 32

isolation

JustifyItems = 33

justify-items

LineBreak = 34

line-break

ListStylePosition = 35

list-style-position

MaskType = 36

mask-type

MathStyle = 37

math-style

MixBlendMode = 38

mix-blend-mode

ObjectFit = 39

object-fit

OutlineStyle = 40

outline-style

OverflowWrap = 41

overflow-wrap

PointerEvents = 42

pointer-events

Position = 43

position

PositionArea = 44

position-area

ScrollBehavior = 45

scroll-behavior

ScrollbarWidth = 46

scrollbar-width

ServoTopLayer = 47

-servo-top-layer

StrokeLinecap = 48

stroke-linecap

StrokeLinejoin = 49

stroke-linejoin

TableLayout = 50

table-layout

TextAlign = 51

text-align

TextAlignLast = 52

text-align-last

TextDecorationLine = 53

text-decoration-line

TextDecorationStyle = 54

text-decoration-style

TextJustify = 55

text-justify

TextOrientation = 56

text-orientation

TextRendering = 57

text-rendering

TextTransform = 58

text-transform

TextUnderlinePosition = 59

text-underline-position

TextWrapMode = 60

text-wrap-mode

TouchAction = 61

touch-action

TransformStyle = 62

transform-style

UnicodeBidi = 63

unicode-bidi

UserSelect = 64

user-select

Visibility = 65

visibility

WebkitTextSecurity = 66

-webkit-text-security

WhiteSpaceCollapse = 67

white-space-collapse

WordBreak = 68

word-break

WritingMode = 69

writing-mode

XTextScale = 70

-x-text-scale

Appearance = 71

appearance

MozDefaultAppearance = 72

-moz-default-appearance

AlignContent = 73

align-content

JustifyContent = 74

justify-content

AlignSelf = 75

align-self

JustifySelf = 76

justify-self

OverflowBlock = 77

overflow-block

OverflowInline = 78

overflow-inline

OverflowX = 79

overflow-x

OverflowY = 80

overflow-y

OverscrollBehaviorBlock = 81

overscroll-behavior-block

OverscrollBehaviorInline = 82

overscroll-behavior-inline

OverscrollBehaviorX = 83

overscroll-behavior-x

OverscrollBehaviorY = 84

overscroll-behavior-y

BorderBlockEndStyle = 85

border-block-end-style

BorderBlockStartStyle = 86

border-block-start-style

BorderBottomStyle = 87

border-bottom-style

BorderInlineEndStyle = 88

border-inline-end-style

BorderInlineStartStyle = 89

border-inline-start-style

BorderLeftStyle = 90

border-left-style

BorderRightStyle = 91

border-right-style

BorderTopStyle = 92

border-top-style

AnimationComposition = 93

animation-composition

AnimationDelay = 94

animation-delay

AnimationDirection = 95

animation-direction

AnimationDuration = 96

animation-duration

AnimationFillMode = 97

animation-fill-mode

AnimationIterationCount = 98

animation-iteration-count

AnimationName = 99

animation-name

AnimationPlayState = 100

animation-play-state

AnimationRangeEnd = 101

animation-range-end

AnimationRangeStart = 102

animation-range-start

AnimationTimeline = 103

animation-timeline

AnimationTimingFunction = 104

animation-timing-function

AspectRatio = 105

aspect-ratio

BackdropFilter = 106

backdrop-filter

BackgroundAttachment = 107

background-attachment

BackgroundBlendMode = 108

background-blend-mode

BackgroundClip = 109

background-clip

BackgroundImage = 110

background-image

BackgroundOrigin = 111

background-origin

BackgroundPositionX = 112

background-position-x

BackgroundPositionY = 113

background-position-y

BackgroundRepeat = 114

background-repeat

BackgroundSize = 115

background-size

BaselineShift = 116

baseline-shift

BorderImageOutset = 117

border-image-outset

BorderImageSlice = 118

border-image-slice

BorderImageWidth = 119

border-image-width

BorderSpacing = 120

border-spacing

BoxShadow = 121

box-shadow

CaretColor = 122

caret-color

Clip = 123

clip

ClipPath = 124

clip-path

Color = 125

color

ColorScheme = 126

color-scheme

ColumnCount = 127

column-count

ColumnWidth = 128

column-width

ContainerName = 129

container-name

Content = 130

content

CounterIncrement = 131

counter-increment

CounterReset = 132

counter-reset

Cursor = 133

cursor

D = 134

d

Filter = 135

filter

FlexBasis = 136

flex-basis

FontFamily = 137

font-family

FontFeatureSettings = 138

font-feature-settings

FontSize = 139

font-size

FontSizeAdjust = 140

font-size-adjust

FontStretch = 141

font-stretch

FontStyle = 142

font-style

FontVariantAlternates = 143

font-variant-alternates

FontVariationSettings = 144

font-variation-settings

FontWeight = 145

font-weight

GridTemplateAreas = 146

grid-template-areas

LetterSpacing = 147

letter-spacing

LineHeight = 148

line-height

ListStyleType = 149

list-style-type

MaskClip = 150

mask-clip

MaskComposite = 151

mask-composite

MaskImage = 152

mask-image

MaskMode = 153

mask-mode

MaskOrigin = 154

mask-origin

MaskPositionX = 155

mask-position-x

MaskPositionY = 156

mask-position-y

MaskRepeat = 157

mask-repeat

MaskSize = 158

mask-size

MathDepth = 159

math-depth

MozMinFontSizeRatio = 160

-moz-min-font-size-ratio

OffsetPath = 161

offset-path

Opacity = 162

opacity

Order = 163

order

OutlineOffset = 164

outline-offset

OverflowClipMargin = 165

overflow-clip-margin

Perspective = 166

perspective

PositionTryFallbacks = 167

position-try-fallbacks

Quotes = 168

quotes

Rotate = 169

rotate

Scale = 170

scale

ScrollbarColor = 171

scrollbar-color

StrokeDasharray = 172

stroke-dasharray

StrokeDashoffset = 173

stroke-dashoffset

StrokeWidth = 174

stroke-width

TabSize = 175

tab-size

TextDecorationInset = 176

text-decoration-inset

TextDecorationThickness = 177

text-decoration-thickness

TextIndent = 178

text-indent

TextOverflow = 179

text-overflow

TextShadow = 180

text-shadow

TextUnderlineOffset = 181

text-underline-offset

Transform = 182

transform

TransformOrigin = 183

transform-origin

TransitionBehavior = 184

transition-behavior

TransitionDelay = 185

transition-delay

TransitionDuration = 186

transition-duration

TransitionProperty = 187

transition-property

TransitionTimingFunction = 188

transition-timing-function

Translate = 189

translate

ViewTransitionClass = 190

view-transition-class

ViewTransitionName = 191

view-transition-name

WillChange = 192

will-change

WordSpacing = 193

word-spacing

XLang = 194

-x-lang

ZIndex = 195

z-index

Zoom = 196

zoom

ObjectPosition = 197

object-position

PerspectiveOrigin = 198

perspective-origin

Fill = 199

fill

Stroke = 200

stroke

GridTemplateColumns = 201

grid-template-columns

GridTemplateRows = 202

grid-template-rows

BorderImageSource = 203

border-image-source

ListStyleImage = 204

list-style-image

GridAutoColumns = 205

grid-auto-columns

GridAutoRows = 206

grid-auto-rows

Rx = 207

rx

Ry = 208

ry

FillOpacity = 209

fill-opacity

StrokeOpacity = 210

stroke-opacity

ColumnGap = 211

column-gap

RowGap = 212

row-gap

FlexGrow = 213

flex-grow

FlexShrink = 214

flex-shrink

StrokeMiterlimit = 215

stroke-miterlimit

GridColumnEnd = 216

grid-column-end

GridColumnStart = 217

grid-column-start

GridRowEnd = 218

grid-row-end

GridRowStart = 219

grid-row-start

Cx = 220

cx

Cy = 221

cy

X = 222

x

Y = 223

y

MaxBlockSize = 224

max-block-size

MaxHeight = 225

max-height

MaxInlineSize = 226

max-inline-size

MaxWidth = 227

max-width

BorderBottomLeftRadius = 228

border-bottom-left-radius

BorderBottomRightRadius = 229

border-bottom-right-radius

BorderEndEndRadius = 230

border-end-end-radius

BorderEndStartRadius = 231

border-end-start-radius

BorderStartEndRadius = 232

border-start-end-radius

BorderStartStartRadius = 233

border-start-start-radius

BorderTopLeftRadius = 234

border-top-left-radius

BorderTopRightRadius = 235

border-top-right-radius

CornerBottomLeftShape = 236

corner-bottom-left-shape

CornerBottomRightShape = 237

corner-bottom-right-shape

CornerEndEndShape = 238

corner-end-end-shape

CornerEndStartShape = 239

corner-end-start-shape

CornerStartEndShape = 240

corner-start-end-shape

CornerStartStartShape = 241

corner-start-start-shape

CornerTopLeftShape = 242

corner-top-left-shape

CornerTopRightShape = 243

corner-top-right-shape

Bottom = 244

bottom

InsetBlockEnd = 245

inset-block-end

InsetBlockStart = 246

inset-block-start

InsetInlineEnd = 247

inset-inline-end

InsetInlineStart = 248

inset-inline-start

Left = 249

left

Right = 250

right

Top = 251

top

MarginBlockEnd = 252

margin-block-end

MarginBlockStart = 253

margin-block-start

MarginBottom = 254

margin-bottom

MarginInlineEnd = 255

margin-inline-end

MarginInlineStart = 256

margin-inline-start

MarginLeft = 257

margin-left

MarginRight = 258

margin-right

MarginTop = 259

margin-top

BlockSize = 260

block-size

Height = 261

height

InlineSize = 262

inline-size

MinBlockSize = 263

min-block-size

MinHeight = 264

min-height

MinInlineSize = 265

min-inline-size

MinWidth = 266

min-width

Width = 267

width

BorderBlockEndWidth = 268

border-block-end-width

BorderBlockStartWidth = 269

border-block-start-width

BorderBottomWidth = 270

border-bottom-width

BorderInlineEndWidth = 271

border-inline-end-width

BorderInlineStartWidth = 272

border-inline-start-width

BorderLeftWidth = 273

border-left-width

BorderRightWidth = 274

border-right-width

BorderTopWidth = 275

border-top-width

OutlineWidth = 276

outline-width

PaddingBlockEnd = 277

padding-block-end

PaddingBlockStart = 278

padding-block-start

PaddingBottom = 279

padding-bottom

PaddingInlineEnd = 280

padding-inline-end

PaddingInlineStart = 281

padding-inline-start

PaddingLeft = 282

padding-left

PaddingRight = 283

padding-right

PaddingTop = 284

padding-top

R = 285

r

BackgroundColor = 286

background-color

BorderBlockEndColor = 287

border-block-end-color

BorderBlockStartColor = 288

border-block-start-color

BorderBottomColor = 289

border-bottom-color

BorderInlineEndColor = 290

border-inline-end-color

BorderInlineStartColor = 291

border-inline-start-color

BorderLeftColor = 292

border-left-color

BorderRightColor = 293

border-right-color

BorderTopColor = 294

border-top-color

OutlineColor = 295

outline-color

TextDecorationColor = 296

text-decoration-color

Implementations

impl LonghandId

pub fn shorthands(self) -> NonCustomPropertyIterator<ShorthandId>

Returns an iterator over all the shorthands that include this longhand.

pub(super) fn parse_value<'i, 't>( self, context: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<PropertyDeclaration, ParseError<'i>>

fn logical_mapping_data(self) -> Option<&'static LogicalMappingData>

Return the relevant data to map a particular logical property into physical.

pub fn to_physical(self, wm: WritingMode) -> Self

If this is a logical property, return the corresponding physical one in the given writing mode. Otherwise, return unchanged.

pub fn logical_group(self) -> Option<LogicalGroupId>

Return the logical group of this longhand property.

pub fn flags(self) -> PropertyFlags

Returns PropertyFlags for given longhand property.

impl LonghandId

pub fn name(&self) -> &'static str

Get the name of this longhand property.

pub fn inherited(self) -> bool

Returns whether the longhand property is inherited by default.

pub fn zoom_dependent(self) -> bool

Returns whether the longhand property is zoom-dependent.

pub fn ignored_when_document_colors_disabled(self) -> bool

Returns true if the property is one that is ignored when document colors are disabled.

pub fn is_or_is_longhand_of(self, non_custom: NonCustomPropertyId) -> bool

Returns whether this longhand is non_custom or is a longhand of it.

pub fn is_longhand_of(self, shorthand: ShorthandId) -> bool

Returns whether this longhand is a longhand of shorthand.

pub fn is_animatable(self) -> bool

Returns whether this property is animatable.

pub fn is_discrete_animatable(self) -> bool

Returns whether this property is animatable in a discrete way.

pub fn is_logical(self) -> bool

Return whether this property is logical.