style::properties::generated

Enum PropertyDeclaration

#[repr(u16)]
pub enum PropertyDeclaration {

    AlignItems(ItemPlacement),
    AlignmentBaseline(AlignmentBaseline),
    BackfaceVisibility(SpecifiedValue),
    BaselineSource(BaselineSource),
    BorderCollapse(SpecifiedValue),
    BorderImageRepeat(BorderImageRepeat),
    BoxDecorationBreak(SpecifiedValue),
    BoxSizing(SpecifiedValue),
    CaptionSide(CaptionSide),
    Clear(Clear),
    ColumnSpan(SpecifiedValue),
    Contain(Contain),
    ContainerType(ContainerType),
    Direction(SpecifiedValue),
    Display(Display),
    EmptyCells(SpecifiedValue),
    FillRule(FillRule),
    FlexDirection(SpecifiedValue),
    FlexWrap(SpecifiedValue),
    Float(Float),
    FontKerning(SpecifiedValue),
    FontLanguageOverride(FontLanguageOverride),
    FontOpticalSizing(SpecifiedValue),
    FontSynthesisWeight(FontSynthesis),
    FontVariantCaps(SpecifiedValue),
    FontVariantEastAsian(FontVariantEastAsian),
    FontVariantLigatures(FontVariantLigatures),
    FontVariantNumeric(FontVariantNumeric),
    FontVariantPosition(SpecifiedValue),
    ForcedColorAdjust(ForcedColorAdjust),
    GridAutoFlow(GridAutoFlow),
    ImageRendering(ImageRendering),
    Isolation(SpecifiedValue),
    JustifyItems(JustifyItems),
    LineBreak(LineBreak),
    ListStylePosition(SpecifiedValue),
    MaskType(SpecifiedValue),
    MathStyle(SpecifiedValue),
    MixBlendMode(SpecifiedValue),
    ObjectFit(SpecifiedValue),
    OutlineStyle(OutlineStyle),
    OverflowWrap(OverflowWrap),
    PointerEvents(PointerEvents),
    Position(PositionProperty),
    PositionArea(PositionArea),
    ScrollBehavior(SpecifiedValue),
    ScrollbarWidth(SpecifiedValue),
    ServoTopLayer(SpecifiedValue),
    StrokeLinecap(SpecifiedValue),
    StrokeLinejoin(SpecifiedValue),
    TableLayout(SpecifiedValue),
    TextAlign(TextAlign),
    TextAlignLast(TextAlignLast),
    TextDecorationLine(TextDecorationLine),
    TextDecorationStyle(SpecifiedValue),
    TextJustify(TextJustify),
    TextOrientation(SpecifiedValue),
    TextRendering(SpecifiedValue),
    TextTransform(TextTransform),
    TextUnderlinePosition(TextUnderlinePosition),
    TextWrapMode(SpecifiedValue),
    TouchAction(TouchAction),
    TransformStyle(TransformStyle),
    UnicodeBidi(SpecifiedValue),
    UserSelect(UserSelect),
    Visibility(SpecifiedValue),
    WebkitTextSecurity(SpecifiedValue),
    WhiteSpaceCollapse(SpecifiedValue),
    WordBreak(WordBreak),
    WritingMode(WritingModeProperty),
    XTextScale(XTextScale),
    Appearance(Appearance),
    MozDefaultAppearance(Appearance),
    AlignContent(ContentDistribution),
    JustifyContent(ContentDistribution),
    AlignSelf(SelfAlignment),
    JustifySelf(SelfAlignment),
    OverflowBlock(Overflow),
    OverflowInline(Overflow),
    OverflowX(Overflow),
    OverflowY(Overflow),
    OverscrollBehaviorBlock(OverscrollBehavior),
    OverscrollBehaviorInline(OverscrollBehavior),
    OverscrollBehaviorX(OverscrollBehavior),
    OverscrollBehaviorY(OverscrollBehavior),
    BorderBlockEndStyle(BorderStyle),
    BorderBlockStartStyle(BorderStyle),
    BorderBottomStyle(BorderStyle),
    BorderInlineEndStyle(BorderStyle),
    BorderInlineStartStyle(BorderStyle),
    BorderLeftStyle(BorderStyle),
    BorderRightStyle(BorderStyle),
    BorderTopStyle(BorderStyle),
    AnimationComposition(SpecifiedValue),
    AnimationDelay(SpecifiedValue),
    AnimationDirection(SpecifiedValue),
    AnimationDuration(SpecifiedValue),
    AnimationFillMode(SpecifiedValue),
    AnimationIterationCount(SpecifiedValue),
    AnimationName(SpecifiedValue),
    AnimationPlayState(SpecifiedValue),
    AnimationRangeEnd(SpecifiedValue),
    AnimationRangeStart(SpecifiedValue),
    AnimationTimeline(SpecifiedValue),
    AnimationTimingFunction(SpecifiedValue),
    AspectRatio(Box<AspectRatio>),
    BackdropFilter(SpecifiedValue),
    BackgroundAttachment(SpecifiedValue),
    BackgroundBlendMode(SpecifiedValue),
    BackgroundClip(SpecifiedValue),
    BackgroundImage(SpecifiedValue),
    BackgroundOrigin(SpecifiedValue),
    BackgroundPositionX(SpecifiedValue),
    BackgroundPositionY(SpecifiedValue),
    BackgroundRepeat(SpecifiedValue),
    BackgroundSize(SpecifiedValue),
    BaselineShift(BaselineShift),
    BorderImageOutset(Box<NonNegativeLengthOrNumberRect>),
    BorderImageSlice(Box<BorderImageSlice>),
    BorderImageWidth(Box<BorderImageWidth>),
    BorderSpacing(BorderSpacing),
    BoxShadow(SpecifiedValue),
    CaretColor(CaretColor),
    Clip(Box<ClipRectOrAuto>),
    ClipPath(ClipPath),
    Color(ColorPropertyValue),
    ColorScheme(ColorScheme),
    ColumnCount(ColumnCount),
    ColumnWidth(NonNegativeLengthOrAuto),
    ContainerName(ContainerName),
    Content(Content),
    CounterIncrement(CounterIncrement),
    CounterReset(CounterReset),
    Cursor(Cursor),
    D(DProperty),
    Filter(SpecifiedValue),
    FlexBasis(Box<FlexBasis>),
    FontFamily(FontFamily),
    FontFeatureSettings(FontFeatureSettings),
    FontSize(FontSize),
    FontSizeAdjust(FontSizeAdjust),
    FontStretch(FontStretch),
    FontStyle(FontStyle),
    FontVariantAlternates(FontVariantAlternates),
    FontVariationSettings(FontVariationSettings),
    FontWeight(FontWeight),
    GridTemplateAreas(GridTemplateAreas),
    LetterSpacing(LetterSpacing),
    LineHeight(LineHeight),
    ListStyleType(ListStyleType),
    MaskClip(SpecifiedValue),
    MaskComposite(SpecifiedValue),
    MaskImage(SpecifiedValue),
    MaskMode(SpecifiedValue),
    MaskOrigin(SpecifiedValue),
    MaskPositionX(SpecifiedValue),
    MaskPositionY(SpecifiedValue),
    MaskRepeat(SpecifiedValue),
    MaskSize(SpecifiedValue),
    MathDepth(MathDepth),
    MozMinFontSizeRatio(Percentage),
    OffsetPath(OffsetPath),
    Opacity(Opacity),
    Order(Integer),
    OutlineOffset(BorderSideOffset),
    OverflowClipMargin(OverflowClipMargin),
    Perspective(Perspective),
    PositionTryFallbacks(PositionTryFallbacks),
    Quotes(Quotes),
    Rotate(Box<Rotate>),
    Scale(Box<Scale>),
    ScrollbarColor(Box<ScrollbarColor>),
    StrokeDasharray(SVGStrokeDashArray),
    StrokeDashoffset(SVGLength),
    StrokeWidth(SVGWidth),
    TabSize(NonNegativeLengthOrNumber),
    TextDecorationInset(Box<TextDecorationInset>),
    TextDecorationThickness(TextDecorationLength),
    TextIndent(TextIndent),
    TextOverflow(Box<TextOverflow>),
    TextShadow(SpecifiedValue),
    TextUnderlineOffset(LengthPercentageOrAuto),
    Transform(Transform),
    TransformOrigin(Box<TransformOrigin>),
    TransitionBehavior(SpecifiedValue),
    TransitionDelay(SpecifiedValue),
    TransitionDuration(SpecifiedValue),
    TransitionProperty(SpecifiedValue),
    TransitionTimingFunction(SpecifiedValue),
    Translate(Box<Translate>),
    ViewTransitionClass(ViewTransitionClass),
    ViewTransitionName(ViewTransitionName),
    WillChange(WillChange),
    WordSpacing(WordSpacing),
    XLang(XLang),
    ZIndex(ZIndex),
    Zoom(Zoom),
    ObjectPosition(Box<Position>),
    PerspectiveOrigin(Box<Position>),
    Fill(Box<SVGPaint>),
    Stroke(Box<SVGPaint>),
    GridTemplateColumns(GridTemplateComponent),
    GridTemplateRows(GridTemplateComponent),
    BorderImageSource(Image),
    ListStyleImage(Image),
    GridAutoColumns(ImplicitGridTracks),
    GridAutoRows(ImplicitGridTracks),
    Rx(NonNegativeLengthPercentageOrAuto),
    Ry(NonNegativeLengthPercentageOrAuto),
    FillOpacity(SVGOpacity),
    StrokeOpacity(SVGOpacity),
    ColumnGap(NonNegativeLengthPercentageOrNormal),
    RowGap(NonNegativeLengthPercentageOrNormal),
    FlexGrow(NonNegativeNumber),
    FlexShrink(NonNegativeNumber),
    StrokeMiterlimit(NonNegativeNumber),
    GridColumnEnd(GridLine),
    GridColumnStart(GridLine),
    GridRowEnd(GridLine),
    GridRowStart(GridLine),
    Cx(LengthPercentage),
    Cy(LengthPercentage),
    X(LengthPercentage),
    Y(LengthPercentage),
    MaxBlockSize(MaxSize),
    MaxHeight(MaxSize),
    MaxInlineSize(MaxSize),
    MaxWidth(MaxSize),
    BorderBottomLeftRadius(Box<BorderCornerRadius>),
    BorderBottomRightRadius(Box<BorderCornerRadius>),
    BorderEndEndRadius(Box<BorderCornerRadius>),
    BorderEndStartRadius(Box<BorderCornerRadius>),
    BorderStartEndRadius(Box<BorderCornerRadius>),
    BorderStartStartRadius(Box<BorderCornerRadius>),
    BorderTopLeftRadius(Box<BorderCornerRadius>),
    BorderTopRightRadius(Box<BorderCornerRadius>),
    CornerBottomLeftShape(CornerShape),
    CornerBottomRightShape(CornerShape),
    CornerEndEndShape(CornerShape),
    CornerEndStartShape(CornerShape),
    CornerStartEndShape(CornerShape),
    CornerStartStartShape(CornerShape),
    CornerTopLeftShape(CornerShape),
    CornerTopRightShape(CornerShape),
    Bottom(Inset),
    InsetBlockEnd(Inset),
    InsetBlockStart(Inset),
    InsetInlineEnd(Inset),
    InsetInlineStart(Inset),
    Left(Inset),
    Right(Inset),
    Top(Inset),
    MarginBlockEnd(Margin),
    MarginBlockStart(Margin),
    MarginBottom(Margin),
    MarginInlineEnd(Margin),
    MarginInlineStart(Margin),
    MarginLeft(Margin),
    MarginRight(Margin),
    MarginTop(Margin),
    BlockSize(Size),
    Height(Size),
    InlineSize(Size),
    MinBlockSize(Size),
    MinHeight(Size),
    MinInlineSize(Size),
    MinWidth(Size),
    Width(Size),
    BorderBlockEndWidth(BorderSideWidth),
    BorderBlockStartWidth(BorderSideWidth),
    BorderBottomWidth(BorderSideWidth),
    BorderInlineEndWidth(BorderSideWidth),
    BorderInlineStartWidth(BorderSideWidth),
    BorderLeftWidth(BorderSideWidth),
    BorderRightWidth(BorderSideWidth),
    BorderTopWidth(BorderSideWidth),
    OutlineWidth(BorderSideWidth),
    PaddingBlockEnd(NonNegativeLengthPercentage),
    PaddingBlockStart(NonNegativeLengthPercentage),
    PaddingBottom(NonNegativeLengthPercentage),
    PaddingInlineEnd(NonNegativeLengthPercentage),
    PaddingInlineStart(NonNegativeLengthPercentage),
    PaddingLeft(NonNegativeLengthPercentage),
    PaddingRight(NonNegativeLengthPercentage),
    PaddingTop(NonNegativeLengthPercentage),
    R(NonNegativeLengthPercentage),
    BackgroundColor(Color),
    BorderBlockEndColor(Color),
    BorderBlockStartColor(Color),
    BorderBottomColor(Color),
    BorderInlineEndColor(Color),
    BorderInlineStartColor(Color),
    BorderLeftColor(Color),
    BorderRightColor(Color),
    BorderTopColor(Color),
    OutlineColor(Color),
    TextDecorationColor(Color),
    CSSWideKeyword(WideKeywordDeclaration),
    WithVariables(VariableDeclaration),
    Custom(CustomDeclaration),

}

Servo’s representation for a property declaration.

Variants

AlignItems(ItemPlacement)

align-items

AlignmentBaseline(AlignmentBaseline)

alignment-baseline

BackfaceVisibility(SpecifiedValue)

backface-visibility

BaselineSource(BaselineSource)

baseline-source

BorderCollapse(SpecifiedValue)

border-collapse

BorderImageRepeat(BorderImageRepeat)

border-image-repeat

BoxDecorationBreak(SpecifiedValue)

box-decoration-break

BoxSizing(SpecifiedValue)

box-sizing

CaptionSide(CaptionSide)

caption-side

Clear(Clear)

clear

ColumnSpan(SpecifiedValue)

column-span

Contain(Contain)

contain

ContainerType(ContainerType)

container-type

Direction(SpecifiedValue)

direction

Display(Display)

display

EmptyCells(SpecifiedValue)

empty-cells

FillRule(FillRule)

fill-rule

FlexDirection(SpecifiedValue)

flex-direction

FlexWrap(SpecifiedValue)

flex-wrap

Float(Float)

float

FontKerning(SpecifiedValue)

font-kerning

FontLanguageOverride(FontLanguageOverride)

font-language-override

FontOpticalSizing(SpecifiedValue)

font-optical-sizing

FontSynthesisWeight(FontSynthesis)

font-synthesis-weight

FontVariantCaps(SpecifiedValue)

font-variant-caps

FontVariantEastAsian(FontVariantEastAsian)

font-variant-east-asian

FontVariantLigatures(FontVariantLigatures)

font-variant-ligatures

FontVariantNumeric(FontVariantNumeric)

font-variant-numeric

FontVariantPosition(SpecifiedValue)

font-variant-position

ForcedColorAdjust(ForcedColorAdjust)

forced-color-adjust

GridAutoFlow(GridAutoFlow)

grid-auto-flow

ImageRendering(ImageRendering)

image-rendering

Isolation(SpecifiedValue)

isolation

JustifyItems(JustifyItems)

justify-items

LineBreak(LineBreak)

line-break

ListStylePosition(SpecifiedValue)

list-style-position

MaskType(SpecifiedValue)

mask-type

MathStyle(SpecifiedValue)

math-style

MixBlendMode(SpecifiedValue)

mix-blend-mode

ObjectFit(SpecifiedValue)

object-fit

OutlineStyle(OutlineStyle)

outline-style

OverflowWrap(OverflowWrap)

overflow-wrap

PointerEvents(PointerEvents)

pointer-events

Position(PositionProperty)

position

PositionArea(PositionArea)

position-area

ScrollBehavior(SpecifiedValue)

scroll-behavior

ScrollbarWidth(SpecifiedValue)

scrollbar-width

ServoTopLayer(SpecifiedValue)

-servo-top-layer

StrokeLinecap(SpecifiedValue)

stroke-linecap

StrokeLinejoin(SpecifiedValue)

stroke-linejoin

TableLayout(SpecifiedValue)

table-layout

TextAlign(TextAlign)

text-align

TextAlignLast(TextAlignLast)

text-align-last

TextDecorationLine(TextDecorationLine)

text-decoration-line

TextDecorationStyle(SpecifiedValue)

text-decoration-style

TextJustify(TextJustify)

text-justify

TextOrientation(SpecifiedValue)

text-orientation

TextRendering(SpecifiedValue)

text-rendering

TextTransform(TextTransform)

text-transform

TextUnderlinePosition(TextUnderlinePosition)

text-underline-position

TextWrapMode(SpecifiedValue)

text-wrap-mode

TouchAction(TouchAction)

touch-action

TransformStyle(TransformStyle)

transform-style

UnicodeBidi(SpecifiedValue)

unicode-bidi

UserSelect(UserSelect)

user-select

Visibility(SpecifiedValue)

visibility

WebkitTextSecurity(SpecifiedValue)

-webkit-text-security

WhiteSpaceCollapse(SpecifiedValue)

white-space-collapse

WordBreak(WordBreak)

word-break

WritingMode(WritingModeProperty)

writing-mode

XTextScale(XTextScale)

-x-text-scale

Appearance(Appearance)

appearance

MozDefaultAppearance(Appearance)

-moz-default-appearance

AlignContent(ContentDistribution)

align-content

JustifyContent(ContentDistribution)

justify-content

AlignSelf(SelfAlignment)

align-self

JustifySelf(SelfAlignment)

justify-self

OverflowBlock(Overflow)

overflow-block

OverflowInline(Overflow)

overflow-inline

OverflowX(Overflow)

overflow-x

OverflowY(Overflow)

overflow-y

OverscrollBehaviorBlock(OverscrollBehavior)

overscroll-behavior-block

OverscrollBehaviorInline(OverscrollBehavior)

overscroll-behavior-inline

OverscrollBehaviorX(OverscrollBehavior)

overscroll-behavior-x

OverscrollBehaviorY(OverscrollBehavior)

overscroll-behavior-y

BorderBlockEndStyle(BorderStyle)

border-block-end-style

BorderBlockStartStyle(BorderStyle)

border-block-start-style

BorderBottomStyle(BorderStyle)

border-bottom-style

BorderInlineEndStyle(BorderStyle)

border-inline-end-style

BorderInlineStartStyle(BorderStyle)

border-inline-start-style

BorderLeftStyle(BorderStyle)

border-left-style

BorderRightStyle(BorderStyle)

border-right-style

BorderTopStyle(BorderStyle)

border-top-style

AnimationComposition(SpecifiedValue)

animation-composition

AnimationDelay(SpecifiedValue)

animation-delay

AnimationDirection(SpecifiedValue)

animation-direction

AnimationDuration(SpecifiedValue)

animation-duration

AnimationFillMode(SpecifiedValue)

animation-fill-mode

AnimationIterationCount(SpecifiedValue)

animation-iteration-count

AnimationName(SpecifiedValue)

animation-name

AnimationPlayState(SpecifiedValue)

animation-play-state

AnimationRangeEnd(SpecifiedValue)

animation-range-end

AnimationRangeStart(SpecifiedValue)

animation-range-start

AnimationTimeline(SpecifiedValue)

animation-timeline

AnimationTimingFunction(SpecifiedValue)

animation-timing-function

AspectRatio(Box<AspectRatio>)

aspect-ratio

BackdropFilter(SpecifiedValue)

backdrop-filter

BackgroundAttachment(SpecifiedValue)

background-attachment

BackgroundBlendMode(SpecifiedValue)

background-blend-mode

BackgroundClip(SpecifiedValue)

background-clip

BackgroundImage(SpecifiedValue)

background-image

BackgroundOrigin(SpecifiedValue)

background-origin

BackgroundPositionX(SpecifiedValue)

background-position-x

BackgroundPositionY(SpecifiedValue)

background-position-y

BackgroundRepeat(SpecifiedValue)

background-repeat

BackgroundSize(SpecifiedValue)

background-size

BaselineShift(BaselineShift)

baseline-shift

BorderImageOutset(Box<NonNegativeLengthOrNumberRect>)

border-image-outset

BorderImageSlice(Box<BorderImageSlice>)

border-image-slice

BorderImageWidth(Box<BorderImageWidth>)

border-image-width

BorderSpacing(BorderSpacing)

border-spacing

BoxShadow(SpecifiedValue)

box-shadow

CaretColor(CaretColor)

caret-color

Clip(Box<ClipRectOrAuto>)

clip

ClipPath(ClipPath)

clip-path

Color(ColorPropertyValue)

color

ColorScheme(ColorScheme)

color-scheme

ColumnCount(ColumnCount)

column-count

ColumnWidth(NonNegativeLengthOrAuto)

column-width

ContainerName(ContainerName)

container-name

Content(Content)

content

CounterIncrement(CounterIncrement)

counter-increment

CounterReset(CounterReset)

counter-reset

Cursor(Cursor)

cursor

D(DProperty)

d

Filter(SpecifiedValue)

filter

FlexBasis(Box<FlexBasis>)

flex-basis

FontFamily(FontFamily)

font-family

FontFeatureSettings(FontFeatureSettings)

font-feature-settings

FontSize(FontSize)

font-size

FontSizeAdjust(FontSizeAdjust)

font-size-adjust

FontStretch(FontStretch)

font-stretch

FontStyle(FontStyle)

font-style

FontVariantAlternates(FontVariantAlternates)

font-variant-alternates

FontVariationSettings(FontVariationSettings)

font-variation-settings

FontWeight(FontWeight)

font-weight

GridTemplateAreas(GridTemplateAreas)

grid-template-areas

LetterSpacing(LetterSpacing)

letter-spacing

LineHeight(LineHeight)

line-height

ListStyleType(ListStyleType)

list-style-type

MaskClip(SpecifiedValue)

mask-clip

MaskComposite(SpecifiedValue)

mask-composite

MaskImage(SpecifiedValue)

mask-image

MaskMode(SpecifiedValue)

mask-mode

MaskOrigin(SpecifiedValue)

mask-origin

MaskPositionX(SpecifiedValue)

mask-position-x

MaskPositionY(SpecifiedValue)

mask-position-y

MaskRepeat(SpecifiedValue)

mask-repeat

MaskSize(SpecifiedValue)

mask-size

MathDepth(MathDepth)

math-depth

MozMinFontSizeRatio(Percentage)

-moz-min-font-size-ratio

OffsetPath(OffsetPath)

offset-path

Opacity(Opacity)

opacity

Order(Integer)

order

OutlineOffset(BorderSideOffset)

outline-offset

OverflowClipMargin(OverflowClipMargin)

overflow-clip-margin

Perspective(Perspective)

perspective

PositionTryFallbacks(PositionTryFallbacks)

position-try-fallbacks

Quotes(Quotes)

quotes

Rotate(Box<Rotate>)

rotate

Scale(Box<Scale>)

scale

ScrollbarColor(Box<ScrollbarColor>)

scrollbar-color

StrokeDasharray(SVGStrokeDashArray)

stroke-dasharray

StrokeDashoffset(SVGLength)

stroke-dashoffset

StrokeWidth(SVGWidth)

stroke-width

TabSize(NonNegativeLengthOrNumber)

tab-size

TextDecorationInset(Box<TextDecorationInset>)

text-decoration-inset

TextDecorationThickness(TextDecorationLength)

text-decoration-thickness

TextIndent(TextIndent)

text-indent

TextOverflow(Box<TextOverflow>)

text-overflow

TextShadow(SpecifiedValue)

text-shadow

TextUnderlineOffset(LengthPercentageOrAuto)

text-underline-offset

Transform(Transform)

transform

TransformOrigin(Box<TransformOrigin>)

transform-origin

TransitionBehavior(SpecifiedValue)

transition-behavior

TransitionDelay(SpecifiedValue)

transition-delay

TransitionDuration(SpecifiedValue)

transition-duration

TransitionProperty(SpecifiedValue)

transition-property

TransitionTimingFunction(SpecifiedValue)

transition-timing-function

Translate(Box<Translate>)

translate

ViewTransitionClass(ViewTransitionClass)

view-transition-class

ViewTransitionName(ViewTransitionName)

view-transition-name

WillChange(WillChange)

will-change

WordSpacing(WordSpacing)

word-spacing

XLang(XLang)

-x-lang

ZIndex(ZIndex)

z-index

Zoom(Zoom)

zoom

ObjectPosition(Box<Position>)

object-position

PerspectiveOrigin(Box<Position>)

perspective-origin

Fill(Box<SVGPaint>)

fill

Stroke(Box<SVGPaint>)

stroke

GridTemplateColumns(GridTemplateComponent)

grid-template-columns

GridTemplateRows(GridTemplateComponent)

grid-template-rows

BorderImageSource(Image)

border-image-source

ListStyleImage(Image)

list-style-image

GridAutoColumns(ImplicitGridTracks)

grid-auto-columns

GridAutoRows(ImplicitGridTracks)

grid-auto-rows

Rx(NonNegativeLengthPercentageOrAuto)

rx

Ry(NonNegativeLengthPercentageOrAuto)

ry

FillOpacity(SVGOpacity)

fill-opacity

StrokeOpacity(SVGOpacity)

stroke-opacity

ColumnGap(NonNegativeLengthPercentageOrNormal)

column-gap

RowGap(NonNegativeLengthPercentageOrNormal)

row-gap

FlexGrow(NonNegativeNumber)

flex-grow

FlexShrink(NonNegativeNumber)

flex-shrink

StrokeMiterlimit(NonNegativeNumber)

stroke-miterlimit

GridColumnEnd(GridLine)

grid-column-end

GridColumnStart(GridLine)

grid-column-start

GridRowEnd(GridLine)

grid-row-end

GridRowStart(GridLine)

grid-row-start

Cx(LengthPercentage)

cx

Cy(LengthPercentage)

cy

X(LengthPercentage)

x

Y(LengthPercentage)

y

MaxBlockSize(MaxSize)

max-block-size

MaxHeight(MaxSize)

max-height

MaxInlineSize(MaxSize)

max-inline-size

MaxWidth(MaxSize)

max-width

BorderBottomLeftRadius(Box<BorderCornerRadius>)

border-bottom-left-radius

BorderBottomRightRadius(Box<BorderCornerRadius>)

border-bottom-right-radius

BorderEndEndRadius(Box<BorderCornerRadius>)

border-end-end-radius

BorderEndStartRadius(Box<BorderCornerRadius>)

border-end-start-radius

BorderStartEndRadius(Box<BorderCornerRadius>)

border-start-end-radius

BorderStartStartRadius(Box<BorderCornerRadius>)

border-start-start-radius

BorderTopLeftRadius(Box<BorderCornerRadius>)

border-top-left-radius

BorderTopRightRadius(Box<BorderCornerRadius>)

border-top-right-radius

CornerBottomLeftShape(CornerShape)

corner-bottom-left-shape

CornerBottomRightShape(CornerShape)

corner-bottom-right-shape

CornerEndEndShape(CornerShape)

corner-end-end-shape

CornerEndStartShape(CornerShape)

corner-end-start-shape

CornerStartEndShape(CornerShape)

corner-start-end-shape

CornerStartStartShape(CornerShape)

corner-start-start-shape

CornerTopLeftShape(CornerShape)

corner-top-left-shape

CornerTopRightShape(CornerShape)

corner-top-right-shape

Bottom(Inset)

bottom

InsetBlockEnd(Inset)

inset-block-end

InsetBlockStart(Inset)

inset-block-start

InsetInlineEnd(Inset)

inset-inline-end

InsetInlineStart(Inset)

inset-inline-start

Left(Inset)

left

Right(Inset)

right

Top(Inset)

top

MarginBlockEnd(Margin)

margin-block-end

MarginBlockStart(Margin)

margin-block-start

MarginBottom(Margin)

margin-bottom

MarginInlineEnd(Margin)

margin-inline-end

MarginInlineStart(Margin)

margin-inline-start

MarginLeft(Margin)

margin-left

MarginRight(Margin)

margin-right

MarginTop(Margin)

margin-top

BlockSize(Size)

block-size

Height(Size)

height

InlineSize(Size)

inline-size

MinBlockSize(Size)

min-block-size

MinHeight(Size)

min-height

MinInlineSize(Size)

min-inline-size

MinWidth(Size)

min-width

Width(Size)

width

BorderBlockEndWidth(BorderSideWidth)

border-block-end-width

BorderBlockStartWidth(BorderSideWidth)

border-block-start-width

BorderBottomWidth(BorderSideWidth)

border-bottom-width

BorderInlineEndWidth(BorderSideWidth)

border-inline-end-width

BorderInlineStartWidth(BorderSideWidth)

border-inline-start-width

BorderLeftWidth(BorderSideWidth)

border-left-width

BorderRightWidth(BorderSideWidth)

border-right-width

BorderTopWidth(BorderSideWidth)

border-top-width

OutlineWidth(BorderSideWidth)

outline-width

PaddingBlockEnd(NonNegativeLengthPercentage)

padding-block-end

PaddingBlockStart(NonNegativeLengthPercentage)

padding-block-start

PaddingBottom(NonNegativeLengthPercentage)

padding-bottom

PaddingInlineEnd(NonNegativeLengthPercentage)

padding-inline-end

PaddingInlineStart(NonNegativeLengthPercentage)

padding-inline-start

PaddingLeft(NonNegativeLengthPercentage)

padding-left

PaddingRight(NonNegativeLengthPercentage)

padding-right

PaddingTop(NonNegativeLengthPercentage)

padding-top

R(NonNegativeLengthPercentage)

r

BackgroundColor(Color)

background-color

BorderBlockEndColor(Color)

border-block-end-color

BorderBlockStartColor(Color)

border-block-start-color

BorderBottomColor(Color)

border-bottom-color

BorderInlineEndColor(Color)

border-inline-end-color

BorderInlineStartColor(Color)

border-inline-start-color

BorderLeftColor(Color)

border-left-color

BorderRightColor(Color)

border-right-color

BorderTopColor(Color)

border-top-color

OutlineColor(Color)

outline-color

TextDecorationColor(Color)

text-decoration-color

CSSWideKeyword(WideKeywordDeclaration)

A CSS-wide keyword.

WithVariables(VariableDeclaration)

An unparsed declaration.

Custom(CustomDeclaration)

A custom property declaration.

Implementations

impl PropertyDeclaration

pub(crate) unsafe fn unchecked_value_as<T>(&self) -> &T

Returns the given value for this declaration as a particular type. It’s the caller’s responsibility to guarantee that the longhand id has the right specified value representation.

pub(crate) fn debug_crash(&self, reason: &str)

Dumps the property declaration before crashing.

fn is_longhand_value(&self) -> bool

Returns whether this is a variant of the Longhand(Value) type, rather than one of the special variants in extra_variants.

pub fn to_css(&self, dest: &mut CssStringWriter) -> Result

Like the method on ToCss, but without the type parameter to avoid accidentally monomorphizing this large function multiple times for different writers.

pub fn to_typed_value_list(&self) -> Option<TypedValueList>

Like the method on ToTyped.

pub(super) fn color_value(&self) -> Option<&Color>

Returns the color value of a given property, for high-contrast-mode tweaks.

impl PropertyDeclaration

pub fn id(&self) -> PropertyDeclarationId<'_>

Given a property declaration, return the property declaration id.

pub fn to_physical(&self, wm: WritingMode) -> Self

Given a declaration, convert it into a declaration for a corresponding physical property.

pub fn get_system(&self) -> Option<SystemFont>

Returns whether or not the property is set by a system font

impl PropertyDeclaration

fn with_variables_from_shorthand(&self, shorthand: ShorthandId) -> Option<&str>

pub fn css_wide_keyword(id: LonghandId, keyword: CSSWideKeyword) -> Self

Returns a CSS-wide keyword declaration for a given property.

pub fn get_css_wide_keyword(&self) -> Option<CSSWideKeyword>

Returns a CSS-wide keyword if the declaration’s value is one.

pub fn may_serialize_as_part_of_shorthand(&self) -> bool

Returns whether the declaration may be serialized as part of a shorthand.

This method returns false if this declaration contains variable or has a CSS-wide keyword value, since these values cannot be serialized as part of a shorthand.

Caller should check with_variables_from_shorthand() and whether all needed declarations has the same CSS-wide keyword first.

Note that, serialization of a shorthand may still fail because of other property-specific requirement even when this method returns true for all the longhand declarations.

pub fn is_animatable(&self) -> bool

Returns true if this property declaration is for one of the animatable properties.

pub fn is_custom(&self) -> bool

Returns true if this property is a custom property, false otherwise.

pub fn parse_into<'i, 't>( declarations: &mut SourcePropertyDeclaration, id: PropertyId, context: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<(), ParseError<'i>>

The context parameter controls this:

https://drafts.csswg.org/css-animations/#keyframes

The inside of accepts any CSS property except those defined in this specification, but does accept the animation-play-state property and interprets it specially.

This will not actually parse Importance values, and will always set things to Importance::Normal. Parsing Importance values is the job of PropertyDeclarationParser, we only set them here so that we don’t have to reallocate