style::properties::generated

Enum ShorthandId

#[repr(u16)]
pub enum ShorthandId {

    Background = 0,
    BackgroundPosition = 1,
    BorderColor = 2,
    BorderStyle = 3,
    BorderWidth = 4,
    BorderTop = 5,
    BorderRight = 6,
    BorderBottom = 7,
    BorderLeft = 8,
    BorderBlockStart = 9,
    BorderBlockEnd = 10,
    BorderInlineStart = 11,
    BorderInlineEnd = 12,
    Border = 13,
    BorderRadius = 14,
    BorderImage = 15,
    BorderBlockWidth = 16,
    BorderBlockStyle = 17,
    BorderBlockColor = 18,
    BorderInlineWidth = 19,
    BorderInlineStyle = 20,
    BorderInlineColor = 21,
    BorderBlock = 22,
    BorderInline = 23,
    Overflow = 24,
    WebkitPerspective = 25,
    WebkitTransform = 26,
    Columns = 27,
    Font = 28,
    FontVariant = 29,
    WhiteSpace = 30,
    ListStyle = 31,
    Margin = 32,
    MarginBlock = 33,
    MarginInline = 34,
    Outline = 35,
    Padding = 36,
    PaddingBlock = 37,
    PaddingInline = 38,
    FlexFlow = 39,
    Flex = 40,
    Gap = 41,
    GridRow = 42,
    GridColumn = 43,
    GridArea = 44,
    GridTemplate = 45,
    Grid = 46,
    PlaceContent = 47,
    PlaceSelf = 48,
    PlaceItems = 49,
    Inset = 50,
    InsetBlock = 51,
    InsetInline = 52,
    TextDecoration = 53,
    Transition = 54,
    Animation = 55,
    All = 56,

}

An identifier for a given shorthand property.

Variants

Background = 0

background

BackgroundPosition = 1

background-position

BorderColor = 2

border-color

BorderStyle = 3

border-style

BorderWidth = 4

border-width

BorderTop = 5

border-top

BorderRight = 6

border-right

BorderBottom = 7

border-bottom

BorderLeft = 8

border-left

BorderBlockStart = 9

border-block-start

BorderBlockEnd = 10

border-block-end

BorderInlineStart = 11

border-inline-start

BorderInlineEnd = 12

border-inline-end

Border = 13

border

BorderRadius = 14

border-radius

BorderImage = 15

border-image

BorderBlockWidth = 16

border-block-width

BorderBlockStyle = 17

border-block-style

BorderBlockColor = 18

border-block-color

BorderInlineWidth = 19

border-inline-width

BorderInlineStyle = 20

border-inline-style

BorderInlineColor = 21

border-inline-color

BorderBlock = 22

border-block

BorderInline = 23

border-inline

Overflow = 24

overflow

WebkitPerspective = 25

-webkit-perspective

WebkitTransform = 26

-webkit-transform

Columns = 27

columns

Font = 28

font

FontVariant = 29

font-variant

WhiteSpace = 30

white-space

ListStyle = 31

list-style

Margin = 32

margin

MarginBlock = 33

margin-block

MarginInline = 34

margin-inline

Outline = 35

outline

Padding = 36

padding

PaddingBlock = 37

padding-block

PaddingInline = 38

padding-inline

FlexFlow = 39

flex-flow

Flex = 40

flex

Gap = 41

gap

GridRow = 42

grid-row

GridColumn = 43

grid-column

GridArea = 44

grid-area

GridTemplate = 45

grid-template

Grid = 46

grid

PlaceContent = 47

place-content

PlaceSelf = 48

place-self

PlaceItems = 49

place-items

Inset = 50

inset

InsetBlock = 51

inset-block

InsetInline = 52

inset-inline

TextDecoration = 53

text-decoration

Transition = 54

transition

Animation = 55

animation

All = 56

all

Implementations

impl ShorthandId

pub fn longhands(self) -> NonCustomPropertyIterator<LonghandId>

Get the longhand ids that form this shorthand.

pub fn longhands_to_css( self, declarations: &[&PropertyDeclaration], dest: &mut CssStringWriter, ) -> Result

Try to serialize the given declarations as this shorthand.

Returns an error if writing to the stream fails, or if the declarations do not map to a shorthand.

pub fn flags(self) -> PropertyFlags

Returns PropertyFlags for the given shorthand property.

pub fn idl_name_sort_order(self) -> u32

Returns the order in which this property appears relative to other shorthands in idl-name-sorting order.

pub(super) fn parse_into<'i, 't>( self, declarations: &mut SourcePropertyDeclaration, context: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<(), ParseError<'i>>

impl ShorthandId

pub fn name(&self) -> &'static str

Get the name for this shorthand property.

pub fn get_shorthand_appendable_value<'a, 'b: 'a>( self, declarations: &'a [&'b PropertyDeclaration], ) -> Option<AppendableValue<'a, 'b>>

Finds and returns an appendable value for the given declarations.

Returns the optional appendable value.

pub fn is_legacy_shorthand(self) -> bool

Returns whether this property is a legacy shorthand.