Module style::properties::generated::shorthands

A module with code for all the shorthand css properties, and a few serialization helpers.

Modules

animation

https://drafts.csswg.org/css-animations/#propdef-animation

background

https://drafts.csswg.org/css-backgrounds/#the-background

background_position

https://drafts.csswg.org/css-backgrounds-4/#the-background-position

border

https://drafts.csswg.org/css-backgrounds/#border

border_block

https://drafts.csswg.org/css-logical/#propdef-border-block

border_block_color

https://drafts.csswg.org/css-logical/#propdef-border-block-color

border_block_end

https://drafts.csswg.org/css-logical-props/#propdef-border-block-end

border_block_start

https://drafts.csswg.org/css-logical-props/#propdef-border-block-start

border_block_style

https://drafts.csswg.org/css-logical/#propdef-border-block-style

border_block_width

https://drafts.csswg.org/css-logical/#propdef-border-block-width

border_bottom

https://drafts.csswg.org/css-backgrounds/#border-bottom

border_color

https://drafts.csswg.org/css-backgrounds/#border-color

border_image

https://drafts.csswg.org/css-backgrounds-3/#border-image

border_inline

https://drafts.csswg.org/css-logical/#propdef-border-inline

border_inline_color

https://drafts.csswg.org/css-logical/#propdef-border-inline-color

border_inline_end

https://drafts.csswg.org/css-logical-props/#propdef-border-inline-end

border_inline_start

https://drafts.csswg.org/css-logical-props/#propdef-border-inline-start

border_inline_style

https://drafts.csswg.org/css-logical/#propdef-border-inline-style

border_inline_width

https://drafts.csswg.org/css-logical/#propdef-border-inline-width

border_left

https://drafts.csswg.org/css-backgrounds/#border-left

border_radius

https://drafts.csswg.org/css-backgrounds/#border-radius

border_right

https://drafts.csswg.org/css-backgrounds/#border-right

border_style

https://drafts.csswg.org/css-backgrounds/#border-style

border_top

https://drafts.csswg.org/css-backgrounds/#border-top

border_width

https://drafts.csswg.org/css-backgrounds/#border-width

columns

https://drafts.csswg.org/css-multicol/#propdef-columns

flex

https://drafts.csswg.org/css-flexbox/#flex-property

flex_flow

https://drafts.csswg.org/css-flexbox/#flex-flow-property

font

https://drafts.csswg.org/css-fonts-3/#propdef-font

font_variant

https://drafts.csswg.org/css-fonts-3/#propdef-font-variant

inset

https://drafts.csswg.org/css-logical/#propdef-inset

inset_block

https://drafts.csswg.org/css-logical/#propdef-inset-block

inset_inline

https://drafts.csswg.org/css-logical/#propdef-inset-inline

list_style

https://drafts.csswg.org/css-lists/#propdef-list-style

margin

https://drafts.csswg.org/css-box/#propdef-margin

margin_block

https://drafts.csswg.org/css-logical/#propdef-margin-block

margin_inline

https://drafts.csswg.org/css-logical/#propdef-margin-inline

outline

https://drafts.csswg.org/css-ui/#propdef-outline

overflow

https://drafts.csswg.org/css-overflow/#propdef-overflow

padding

https://drafts.csswg.org/css-box-3/#propdef-padding

padding_block

https://drafts.csswg.org/css-logical/#propdef-padding-block

padding_inline

https://drafts.csswg.org/css-logical/#propdef-padding-inline

text_decoration

https://drafts.csswg.org/css-text-decor/#propdef-text-decoration

transition

https://drafts.csswg.org/css-transitions/#propdef-transition

Macros

try_parse_one 🔒

Constants

ALL_SHORTHAND_MAX_LEN

The max amount of longhands that the all shorthand will ever contain.

Functions

parse_border