Struct style::properties::generated::shorthands::border_block_color::LonghandsToSerialize

pub struct LonghandsToSerialize<'a> {
    pub border_block_start_color: &'a SpecifiedValue,
    pub border_block_end_color: &'a SpecifiedValue,
}

Represents a serializable set of all of the longhand properties that correspond to a shorthand.

Fields
border_block_start_color: &'a SpecifiedValueborder_block_end_color: &'a SpecifiedValue
Implementations

impl<'a> LonghandsToSerialize<'a>

pub fn from_iter( iter: impl Iterator<Item = &'a PropertyDeclaration> ) -> Result<Self, ()>

Tries to get a serializable set of longhands given a set of property declarations.