style::properties::generated::shorthands

Module border_width

https://drafts.csswg.org/css-backgrounds/#border-width

Structs

Longhands

LonghandsToSerialize

Represents a serializable set of all of the longhand properties that correspond to a shorthand.

Functions

parse_into

Parse the given shorthand and fill the result into the declarations vector.

parse_value 🔒

to_css

Try to serialize a given shorthand to a string.

Type Aliases

Single 🔒