style::properties::generated::shorthands

Module flex_flow

https://drafts.csswg.org/css-flexbox/#flex-flow-property

Structs

Longhands

LonghandsToSerialize

Represents a serializable set of all of the longhand properties that correspond to a shorthand.

Functions

parse_into

Parse the given shorthand and fill the result into the declarations vector.

to_css

Try to serialize a given shorthand to a string.