style::properties::generated

Struct ComputedValues

pub struct ComputedValues {
    inner: ComputedValuesInner,
    pseudo: Option<PseudoElement>,
}

The struct that Servo uses to represent computed values.

This struct contains an immutable atomically-reference-counted pointer to every kind of style struct.

When needed, the structs may be copied in order to get mutated.

Fields
inner: ComputedValuesInner

The actual computed values

In Gecko the outer ComputedValues is actually a ComputedStyle, whereas ComputedValuesInner is the core set of computed values.

We maintain this distinction in servo to reduce the amount of special casing.

pseudo: Option<PseudoElement>

The pseudo-element that we’re using.

Implementations

impl ComputedValues

pub fn pseudo(&self) -> Option<PseudoElement>

Returns the pseudo-element that this style represents.

pub fn is_pseudo_style(&self) -> bool

Returns true if this is the style for a pseudo-element.

pub fn is_display_contents(&self) -> bool

Returns whether this style’s display value is equal to contents.

pub fn rules(&self) -> &StrongRuleNode

Gets a reference to the rule node. Panic if no rule node exists.

pub fn visited_rules(&self) -> Option<&StrongRuleNode>

Returns the visited rules, if applicable.

pub fn custom_properties(&self) -> &ComputedCustomProperties

Gets a reference to the custom properties map (if one exists).

pub fn clone_align_items(&self) -> T

Gets the computed value of a given property.

pub fn align_items_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_alignment_baseline(&self) -> T

Gets the computed value of a given property.

pub fn alignment_baseline_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_backface_visibility(&self) -> T

Gets the computed value of a given property.

pub fn backface_visibility_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_baseline_source(&self) -> T

Gets the computed value of a given property.

pub fn baseline_source_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_collapse(&self) -> T

Gets the computed value of a given property.

pub fn border_collapse_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_image_repeat(&self) -> T

Gets the computed value of a given property.

pub fn border_image_repeat_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_box_decoration_break(&self) -> T

Gets the computed value of a given property.

pub fn box_decoration_break_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_box_sizing(&self) -> T

Gets the computed value of a given property.

pub fn box_sizing_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_caption_side(&self) -> T

Gets the computed value of a given property.

pub fn caption_side_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_clear(&self) -> T

Gets the computed value of a given property.

pub fn clear_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_column_span(&self) -> T

Gets the computed value of a given property.

pub fn column_span_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_contain(&self) -> T

Gets the computed value of a given property.

pub fn contain_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_container_type(&self) -> T

Gets the computed value of a given property.

pub fn container_type_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_direction(&self) -> T

Gets the computed value of a given property.

pub fn direction_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_display(&self) -> T

Gets the computed value of a given property.

pub fn display_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_empty_cells(&self) -> T

Gets the computed value of a given property.

pub fn empty_cells_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_fill_rule(&self) -> T

Gets the computed value of a given property.

pub fn fill_rule_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_flex_direction(&self) -> T

Gets the computed value of a given property.

pub fn flex_direction_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_flex_wrap(&self) -> T

Gets the computed value of a given property.

pub fn flex_wrap_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_float(&self) -> T

Gets the computed value of a given property.

pub fn float_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_kerning(&self) -> T

Gets the computed value of a given property.

pub fn font_kerning_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_language_override(&self) -> T

Gets the computed value of a given property.

pub fn font_language_override_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_optical_sizing(&self) -> T

Gets the computed value of a given property.

pub fn font_optical_sizing_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_synthesis_weight(&self) -> T

Gets the computed value of a given property.

pub fn font_synthesis_weight_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_variant_caps(&self) -> T

Gets the computed value of a given property.

pub fn font_variant_caps_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_variant_east_asian(&self) -> T

Gets the computed value of a given property.

pub fn font_variant_east_asian_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_variant_ligatures(&self) -> T

Gets the computed value of a given property.

pub fn font_variant_ligatures_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_variant_numeric(&self) -> T

Gets the computed value of a given property.

pub fn font_variant_numeric_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_variant_position(&self) -> T

Gets the computed value of a given property.

pub fn font_variant_position_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_forced_color_adjust(&self) -> T

Gets the computed value of a given property.

pub fn forced_color_adjust_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_auto_flow(&self) -> T

Gets the computed value of a given property.

pub fn grid_auto_flow_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_image_rendering(&self) -> T

Gets the computed value of a given property.

pub fn image_rendering_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_isolation(&self) -> T

Gets the computed value of a given property.

pub fn isolation_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_justify_items(&self) -> T

Gets the computed value of a given property.

pub fn justify_items_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_line_break(&self) -> T

Gets the computed value of a given property.

pub fn line_break_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_list_style_position(&self) -> T

Gets the computed value of a given property.

pub fn list_style_position_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_type(&self) -> T

Gets the computed value of a given property.

pub fn mask_type_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_math_style(&self) -> T

Gets the computed value of a given property.

pub fn math_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mix_blend_mode(&self) -> T

Gets the computed value of a given property.

pub fn mix_blend_mode_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_object_fit(&self) -> T

Gets the computed value of a given property.

pub fn object_fit_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_outline_style(&self) -> T

Gets the computed value of a given property.

pub fn outline_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_overflow_wrap(&self) -> T

Gets the computed value of a given property.

pub fn overflow_wrap_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_pointer_events(&self) -> T

Gets the computed value of a given property.

pub fn pointer_events_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_position(&self) -> T

Gets the computed value of a given property.

pub fn position_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_position_area(&self) -> T

Gets the computed value of a given property.

pub fn position_area_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_scroll_behavior(&self) -> T

Gets the computed value of a given property.

pub fn scroll_behavior_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_scrollbar_width(&self) -> T

Gets the computed value of a given property.

pub fn scrollbar_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone__servo_top_layer(&self) -> T

Gets the computed value of a given property.

pub fn _servo_top_layer_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke_linecap(&self) -> T

Gets the computed value of a given property.

pub fn stroke_linecap_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke_linejoin(&self) -> T

Gets the computed value of a given property.

pub fn stroke_linejoin_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_table_layout(&self) -> T

Gets the computed value of a given property.

pub fn table_layout_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_align(&self) -> T

Gets the computed value of a given property.

pub fn text_align_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_align_last(&self) -> T

Gets the computed value of a given property.

pub fn text_align_last_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_decoration_line(&self) -> T

Gets the computed value of a given property.

pub fn text_decoration_line_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_decoration_style(&self) -> T

Gets the computed value of a given property.

pub fn text_decoration_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_justify(&self) -> T

Gets the computed value of a given property.

pub fn text_justify_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_orientation(&self) -> T

Gets the computed value of a given property.

pub fn text_orientation_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_rendering(&self) -> T

Gets the computed value of a given property.

pub fn text_rendering_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_transform(&self) -> T

Gets the computed value of a given property.

pub fn text_transform_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_underline_position(&self) -> T

Gets the computed value of a given property.

pub fn text_underline_position_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_wrap_mode(&self) -> T

Gets the computed value of a given property.

pub fn text_wrap_mode_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_touch_action(&self) -> T

Gets the computed value of a given property.

pub fn touch_action_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transform_style(&self) -> T

Gets the computed value of a given property.

pub fn transform_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_unicode_bidi(&self) -> T

Gets the computed value of a given property.

pub fn unicode_bidi_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_user_select(&self) -> T

Gets the computed value of a given property.

pub fn user_select_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_visibility(&self) -> T

Gets the computed value of a given property.

pub fn visibility_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone__webkit_text_security(&self) -> T

Gets the computed value of a given property.

pub fn _webkit_text_security_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_white_space_collapse(&self) -> T

Gets the computed value of a given property.

pub fn white_space_collapse_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_word_break(&self) -> T

Gets the computed value of a given property.

pub fn word_break_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_writing_mode(&self) -> T

Gets the computed value of a given property.

pub fn writing_mode_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone__x_text_scale(&self) -> T

Gets the computed value of a given property.

pub fn _x_text_scale_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_appearance(&self) -> T

Gets the computed value of a given property.

pub fn appearance_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone__moz_default_appearance(&self) -> T

Gets the computed value of a given property.

pub fn _moz_default_appearance_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_align_content(&self) -> T

Gets the computed value of a given property.

pub fn align_content_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_justify_content(&self) -> T

Gets the computed value of a given property.

pub fn justify_content_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_align_self(&self) -> T

Gets the computed value of a given property.

pub fn align_self_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_justify_self(&self) -> T

Gets the computed value of a given property.

pub fn justify_self_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_overflow_x(&self) -> T

Gets the computed value of a given property.

pub fn overflow_x_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_overflow_y(&self) -> T

Gets the computed value of a given property.

pub fn overflow_y_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_overscroll_behavior_x(&self) -> T

Gets the computed value of a given property.

pub fn overscroll_behavior_x_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_overscroll_behavior_y(&self) -> T

Gets the computed value of a given property.

pub fn overscroll_behavior_y_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_bottom_style(&self) -> T

Gets the computed value of a given property.

pub fn border_bottom_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_left_style(&self) -> T

Gets the computed value of a given property.

pub fn border_left_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_right_style(&self) -> T

Gets the computed value of a given property.

pub fn border_right_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_top_style(&self) -> T

Gets the computed value of a given property.

pub fn border_top_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_composition(&self) -> T

Gets the computed value of a given property.

pub fn animation_composition_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_delay(&self) -> T

Gets the computed value of a given property.

pub fn animation_delay_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_direction(&self) -> T

Gets the computed value of a given property.

pub fn animation_direction_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_duration(&self) -> T

Gets the computed value of a given property.

pub fn animation_duration_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_fill_mode(&self) -> T

Gets the computed value of a given property.

pub fn animation_fill_mode_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_iteration_count(&self) -> T

Gets the computed value of a given property.

pub fn animation_iteration_count_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_name(&self) -> T

Gets the computed value of a given property.

pub fn animation_name_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_play_state(&self) -> T

Gets the computed value of a given property.

pub fn animation_play_state_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_range_end(&self) -> T

Gets the computed value of a given property.

pub fn animation_range_end_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_range_start(&self) -> T

Gets the computed value of a given property.

pub fn animation_range_start_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_timeline(&self) -> T

Gets the computed value of a given property.

pub fn animation_timeline_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_animation_timing_function(&self) -> T

Gets the computed value of a given property.

pub fn animation_timing_function_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_aspect_ratio(&self) -> T

Gets the computed value of a given property.

pub fn aspect_ratio_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_backdrop_filter(&self) -> T

Gets the computed value of a given property.

pub fn backdrop_filter_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_attachment(&self) -> T

Gets the computed value of a given property.

pub fn background_attachment_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_blend_mode(&self) -> T

Gets the computed value of a given property.

pub fn background_blend_mode_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_clip(&self) -> T

Gets the computed value of a given property.

pub fn background_clip_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_image(&self) -> T

Gets the computed value of a given property.

pub fn background_image_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_origin(&self) -> T

Gets the computed value of a given property.

pub fn background_origin_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_position_x(&self) -> T

Gets the computed value of a given property.

pub fn background_position_x_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_position_y(&self) -> T

Gets the computed value of a given property.

pub fn background_position_y_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_repeat(&self) -> T

Gets the computed value of a given property.

pub fn background_repeat_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_size(&self) -> T

Gets the computed value of a given property.

pub fn background_size_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_baseline_shift(&self) -> T

Gets the computed value of a given property.

pub fn baseline_shift_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_image_outset(&self) -> T

Gets the computed value of a given property.

pub fn border_image_outset_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_image_slice(&self) -> T

Gets the computed value of a given property.

pub fn border_image_slice_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_image_width(&self) -> T

Gets the computed value of a given property.

pub fn border_image_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_spacing(&self) -> T

Gets the computed value of a given property.

pub fn border_spacing_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_box_shadow(&self) -> T

Gets the computed value of a given property.

pub fn box_shadow_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_caret_color(&self) -> T

Gets the computed value of a given property.

pub fn caret_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_clip(&self) -> T

Gets the computed value of a given property.

pub fn clip_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_clip_path(&self) -> T

Gets the computed value of a given property.

pub fn clip_path_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_color(&self) -> T

Gets the computed value of a given property.

pub fn color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_color_scheme(&self) -> T

Gets the computed value of a given property.

pub fn color_scheme_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_column_count(&self) -> T

Gets the computed value of a given property.

pub fn column_count_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_column_width(&self) -> T

Gets the computed value of a given property.

pub fn column_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_container_name(&self) -> T

Gets the computed value of a given property.

pub fn container_name_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_content(&self) -> T

Gets the computed value of a given property.

pub fn content_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_counter_increment(&self) -> T

Gets the computed value of a given property.

pub fn counter_increment_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_counter_reset(&self) -> T

Gets the computed value of a given property.

pub fn counter_reset_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_cursor(&self) -> T

Gets the computed value of a given property.

pub fn cursor_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_d(&self) -> T

Gets the computed value of a given property.

pub fn d_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_filter(&self) -> T

Gets the computed value of a given property.

pub fn filter_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_flex_basis(&self) -> T

Gets the computed value of a given property.

pub fn flex_basis_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_family(&self) -> T

Gets the computed value of a given property.

pub fn font_family_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_feature_settings(&self) -> T

Gets the computed value of a given property.

pub fn font_feature_settings_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_size(&self) -> T

Gets the computed value of a given property.

pub fn font_size_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_size_adjust(&self) -> T

Gets the computed value of a given property.

pub fn font_size_adjust_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_stretch(&self) -> T

Gets the computed value of a given property.

pub fn font_stretch_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_style(&self) -> T

Gets the computed value of a given property.

pub fn font_style_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_variant_alternates(&self) -> T

Gets the computed value of a given property.

pub fn font_variant_alternates_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_variation_settings(&self) -> T

Gets the computed value of a given property.

pub fn font_variation_settings_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_font_weight(&self) -> T

Gets the computed value of a given property.

pub fn font_weight_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_template_areas(&self) -> T

Gets the computed value of a given property.

pub fn grid_template_areas_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_letter_spacing(&self) -> T

Gets the computed value of a given property.

pub fn letter_spacing_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_line_height(&self) -> T

Gets the computed value of a given property.

pub fn line_height_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_list_style_type(&self) -> T

Gets the computed value of a given property.

pub fn list_style_type_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_clip(&self) -> T

Gets the computed value of a given property.

pub fn mask_clip_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_composite(&self) -> T

Gets the computed value of a given property.

pub fn mask_composite_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_image(&self) -> T

Gets the computed value of a given property.

pub fn mask_image_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_mode(&self) -> T

Gets the computed value of a given property.

pub fn mask_mode_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_origin(&self) -> T

Gets the computed value of a given property.

pub fn mask_origin_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_position_x(&self) -> T

Gets the computed value of a given property.

pub fn mask_position_x_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_position_y(&self) -> T

Gets the computed value of a given property.

pub fn mask_position_y_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_repeat(&self) -> T

Gets the computed value of a given property.

pub fn mask_repeat_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_mask_size(&self) -> T

Gets the computed value of a given property.

pub fn mask_size_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_math_depth(&self) -> T

Gets the computed value of a given property.

pub fn math_depth_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone__moz_min_font_size_ratio(&self) -> T

Gets the computed value of a given property.

pub fn _moz_min_font_size_ratio_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_offset_path(&self) -> T

Gets the computed value of a given property.

pub fn offset_path_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_opacity(&self) -> T

Gets the computed value of a given property.

pub fn opacity_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_order(&self) -> T

Gets the computed value of a given property.

pub fn order_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_outline_offset(&self) -> T

Gets the computed value of a given property.

pub fn outline_offset_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_overflow_clip_margin(&self) -> T

Gets the computed value of a given property.

pub fn overflow_clip_margin_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_perspective(&self) -> T

Gets the computed value of a given property.

pub fn perspective_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_position_try_fallbacks(&self) -> T

Gets the computed value of a given property.

pub fn position_try_fallbacks_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_quotes(&self) -> T

Gets the computed value of a given property.

pub fn quotes_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_rotate(&self) -> T

Gets the computed value of a given property.

pub fn rotate_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_scale(&self) -> T

Gets the computed value of a given property.

pub fn scale_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_scrollbar_color(&self) -> T

Gets the computed value of a given property.

pub fn scrollbar_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke_dasharray(&self) -> T

Gets the computed value of a given property.

pub fn stroke_dasharray_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke_dashoffset(&self) -> T

Gets the computed value of a given property.

pub fn stroke_dashoffset_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke_width(&self) -> T

Gets the computed value of a given property.

pub fn stroke_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_tab_size(&self) -> T

Gets the computed value of a given property.

pub fn tab_size_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_decoration_inset(&self) -> T

Gets the computed value of a given property.

pub fn text_decoration_inset_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_decoration_thickness(&self) -> T

Gets the computed value of a given property.

pub fn text_decoration_thickness_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_indent(&self) -> T

Gets the computed value of a given property.

pub fn text_indent_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_overflow(&self) -> T

Gets the computed value of a given property.

pub fn text_overflow_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_shadow(&self) -> T

Gets the computed value of a given property.

pub fn text_shadow_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_underline_offset(&self) -> T

Gets the computed value of a given property.

pub fn text_underline_offset_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transform(&self) -> T

Gets the computed value of a given property.

pub fn transform_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transform_origin(&self) -> T

Gets the computed value of a given property.

pub fn transform_origin_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transition_behavior(&self) -> T

Gets the computed value of a given property.

pub fn transition_behavior_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transition_delay(&self) -> T

Gets the computed value of a given property.

pub fn transition_delay_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transition_duration(&self) -> T

Gets the computed value of a given property.

pub fn transition_duration_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transition_property(&self) -> T

Gets the computed value of a given property.

pub fn transition_property_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_transition_timing_function(&self) -> T

Gets the computed value of a given property.

pub fn transition_timing_function_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_translate(&self) -> T

Gets the computed value of a given property.

pub fn translate_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_view_transition_class(&self) -> T

Gets the computed value of a given property.

pub fn view_transition_class_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_view_transition_name(&self) -> T

Gets the computed value of a given property.

pub fn view_transition_name_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_will_change(&self) -> T

Gets the computed value of a given property.

pub fn will_change_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_word_spacing(&self) -> T

Gets the computed value of a given property.

pub fn word_spacing_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone__x_lang(&self) -> T

Gets the computed value of a given property.

pub fn _x_lang_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_z_index(&self) -> T

Gets the computed value of a given property.

pub fn z_index_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_zoom(&self) -> T

Gets the computed value of a given property.

pub fn zoom_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_object_position(&self) -> T

Gets the computed value of a given property.

pub fn object_position_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_perspective_origin(&self) -> T

Gets the computed value of a given property.

pub fn perspective_origin_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_fill(&self) -> T

Gets the computed value of a given property.

pub fn fill_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke(&self) -> T

Gets the computed value of a given property.

pub fn stroke_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_template_columns(&self) -> T

Gets the computed value of a given property.

pub fn grid_template_columns_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_template_rows(&self) -> T

Gets the computed value of a given property.

pub fn grid_template_rows_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_image_source(&self) -> T

Gets the computed value of a given property.

pub fn border_image_source_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_list_style_image(&self) -> T

Gets the computed value of a given property.

pub fn list_style_image_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_auto_columns(&self) -> T

Gets the computed value of a given property.

pub fn grid_auto_columns_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_auto_rows(&self) -> T

Gets the computed value of a given property.

pub fn grid_auto_rows_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_rx(&self) -> T

Gets the computed value of a given property.

pub fn rx_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_ry(&self) -> T

Gets the computed value of a given property.

pub fn ry_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_fill_opacity(&self) -> T

Gets the computed value of a given property.

pub fn fill_opacity_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke_opacity(&self) -> T

Gets the computed value of a given property.

pub fn stroke_opacity_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_column_gap(&self) -> T

Gets the computed value of a given property.

pub fn column_gap_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_row_gap(&self) -> T

Gets the computed value of a given property.

pub fn row_gap_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_flex_grow(&self) -> T

Gets the computed value of a given property.

pub fn flex_grow_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_flex_shrink(&self) -> T

Gets the computed value of a given property.

pub fn flex_shrink_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_stroke_miterlimit(&self) -> T

Gets the computed value of a given property.

pub fn stroke_miterlimit_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_column_end(&self) -> T

Gets the computed value of a given property.

pub fn grid_column_end_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_column_start(&self) -> T

Gets the computed value of a given property.

pub fn grid_column_start_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_row_end(&self) -> T

Gets the computed value of a given property.

pub fn grid_row_end_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_grid_row_start(&self) -> T

Gets the computed value of a given property.

pub fn grid_row_start_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_cx(&self) -> T

Gets the computed value of a given property.

pub fn cx_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_cy(&self) -> T

Gets the computed value of a given property.

pub fn cy_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_x(&self) -> T

Gets the computed value of a given property.

pub fn x_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_y(&self) -> T

Gets the computed value of a given property.

pub fn y_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_max_height(&self) -> T

Gets the computed value of a given property.

pub fn max_height_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_max_width(&self) -> T

Gets the computed value of a given property.

pub fn max_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_bottom_left_radius(&self) -> T

Gets the computed value of a given property.

pub fn border_bottom_left_radius_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_bottom_right_radius(&self) -> T

Gets the computed value of a given property.

pub fn border_bottom_right_radius_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_top_left_radius(&self) -> T

Gets the computed value of a given property.

pub fn border_top_left_radius_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_top_right_radius(&self) -> T

Gets the computed value of a given property.

pub fn border_top_right_radius_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_corner_bottom_left_shape(&self) -> T

Gets the computed value of a given property.

pub fn corner_bottom_left_shape_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_corner_bottom_right_shape(&self) -> T

Gets the computed value of a given property.

pub fn corner_bottom_right_shape_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_corner_top_left_shape(&self) -> T

Gets the computed value of a given property.

pub fn corner_top_left_shape_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_corner_top_right_shape(&self) -> T

Gets the computed value of a given property.

pub fn corner_top_right_shape_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_bottom(&self) -> T

Gets the computed value of a given property.

pub fn bottom_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_left(&self) -> T

Gets the computed value of a given property.

pub fn left_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_right(&self) -> T

Gets the computed value of a given property.

pub fn right_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_top(&self) -> T

Gets the computed value of a given property.

pub fn top_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_margin_bottom(&self) -> T

Gets the computed value of a given property.

pub fn margin_bottom_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_margin_left(&self) -> T

Gets the computed value of a given property.

pub fn margin_left_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_margin_right(&self) -> T

Gets the computed value of a given property.

pub fn margin_right_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_margin_top(&self) -> T

Gets the computed value of a given property.

pub fn margin_top_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_height(&self) -> T

Gets the computed value of a given property.

pub fn height_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_min_height(&self) -> T

Gets the computed value of a given property.

pub fn min_height_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_min_width(&self) -> T

Gets the computed value of a given property.

pub fn min_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_width(&self) -> T

Gets the computed value of a given property.

pub fn width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_bottom_width(&self) -> T

Gets the computed value of a given property.

pub fn border_bottom_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_left_width(&self) -> T

Gets the computed value of a given property.

pub fn border_left_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_right_width(&self) -> T

Gets the computed value of a given property.

pub fn border_right_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_top_width(&self) -> T

Gets the computed value of a given property.

pub fn border_top_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_outline_width(&self) -> T

Gets the computed value of a given property.

pub fn outline_width_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_padding_bottom(&self) -> T

Gets the computed value of a given property.

pub fn padding_bottom_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_padding_left(&self) -> T

Gets the computed value of a given property.

pub fn padding_left_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_padding_right(&self) -> T

Gets the computed value of a given property.

pub fn padding_right_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_padding_top(&self) -> T

Gets the computed value of a given property.

pub fn padding_top_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_r(&self) -> T

Gets the computed value of a given property.

pub fn r_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_background_color(&self) -> T

Gets the computed value of a given property.

pub fn background_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_bottom_color(&self) -> T

Gets the computed value of a given property.

pub fn border_bottom_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_left_color(&self) -> T

Gets the computed value of a given property.

pub fn border_left_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_right_color(&self) -> T

Gets the computed value of a given property.

pub fn border_right_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_border_top_color(&self) -> T

Gets the computed value of a given property.

pub fn border_top_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_outline_color(&self) -> T

Gets the computed value of a given property.

pub fn outline_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn clone_text_decoration_color(&self) -> T

Gets the computed value of a given property.

pub fn text_decoration_color_equals(&self, other: &Self) -> bool

Gets the computed value of a given property.

pub fn computed_or_resolved_value( &self, property_id: LonghandId, context: Option<&mut Context<'_>>, dest: &mut CssStringWriter, ) -> Result

Writes the (resolved or computed) value of the given longhand as a string in dest.

TODO(emilio): We should move all the special resolution from nsComputedDOMStyle to ToResolvedValue instead.

pub fn property_value_to_typed_value_list( &self, property_id: LonghandId, ) -> Option<TypedValueList>

Returns the computed value of the given longhand as a TypedValueList, if supported.

pub fn computed_or_resolved_declaration( &self, property_id: LonghandId, context: Option<&mut Context<'_>>, ) -> PropertyDeclaration

Returns the given longhand’s resolved value as a property declaration.

pub fn resolve_color(&self, color: &Color) -> AbsoluteColor

Resolves the currentColor keyword.

Any color value from computed values (except for the ‘color’ property itself) should go through this method.

Usage example: let top_color = style.resolve_color(&style.get_border().clone_border_top_color());

pub fn transition_properties<'a>(&'a self) -> TransitionPropertyIterator<'a>

Create a TransitionPropertyIterator for this styles transition properties.

impl ComputedValues

pub fn new( pseudo: Option<&PseudoElement>, custom_properties: ComputedCustomProperties, attribute_references: AttributeReferences, writing_mode: WritingMode, effective_zoom: Zoom, flags: ComputedValueFlags, rules: Option<StrongRuleNode>, visited_style: Option<Arc<ComputedValues>>, background: Arc<Background>, border: Arc<Border>, box_: Arc<Box>, column: Arc<Column>, counters: Arc<Counters>, effects: Arc<Effects>, font: Arc<Font>, inherited_box: Arc<InheritedBox>, inherited_svg: Arc<InheritedSVG>, inherited_table: Arc<InheritedTable>, inherited_text: Arc<InheritedText>, inherited_ui: Arc<InheritedUI>, list: Arc<List>, margin: Arc<Margin>, outline: Arc<Outline>, padding: Arc<Padding>, position: Arc<Position>, svg: Arc<SVG>, table: Arc<Table>, text: Arc<Text>, ui: Arc<UI>, ) -> Arc<Self>

Create a new refcounted ComputedValues

pub fn initial_values_with_font_override(default_font: Font) -> Arc<Self>

Get the initial computed values.

pub fn to_arc(&self) -> Arc<Self>

Converts the computed values to an Arc<> from a reference.

pub fn computed_value_to_string( &self, property: PropertyDeclarationId<'_>, ) -> String

Serializes the computed value of this property as a string.

pub fn each_cached_lazy_pseudo<F>(&self, _f: F)
where F: FnMut(&Self),

Calls the given function for each cached lazy pseudo-element style.

impl ComputedValues

pub fn guarantees_stacking_context(&self) -> bool

Some properties establish a stacking context when they are set to a non-initial value. In that case, the damage is only set to ServoRestyleDamage::REPAINT because we don’t need to rebuild stacking contexts when the style changes between different non-initial values. This function checks whether any of these properties is set to a value that guarantees a stacking context, so that we only do the work when this changes. Note that it’s still possible to establish a stacking context when this returns false.