style::properties::generated::style_structs

Struct Background

pub struct Background {
    pub background_color: T,
    pub background_image: T,
    pub background_position_x: T,
    pub background_position_y: T,
    pub background_repeat: T,
    pub background_size: T,
    pub background_attachment: T,
    pub background_clip: T,
    pub background_origin: T,
    pub background_blend_mode: T,
}

The Background style struct.

Fields
background_color: T

The background-color computed value.

background_image: T

The background-image computed value.

background_position_x: T

The background-position-x computed value.

background_position_y: T

The background-position-y computed value.

background_repeat: T

The background-repeat computed value.

background_size: T

The background-size computed value.

background_attachment: T

The background-attachment computed value.

background_clip: T

The background-clip computed value.

background_origin: T

The background-origin computed value.

background_blend_mode: T

The background-blend-mode computed value.

Implementations

impl Background

pub fn set_background_color(&mut self, v: T)

Set background-color.

pub fn copy_background_color_from(&mut self, other: &Self)

Set background-color from other struct.

pub fn reset_background_color(&mut self, other: &Self)

Reset background-color from the initial struct.

pub fn clone_background_color(&self) -> T

Get the computed value for background-color.

pub fn background_color_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-color.

pub fn set_background_image(&mut self, v: T)

Set background-image.

pub fn copy_background_image_from(&mut self, other: &Self)

Set background-image from other struct.

pub fn reset_background_image(&mut self, other: &Self)

Reset background-image from the initial struct.

pub fn clone_background_image(&self) -> T

Get the computed value for background-image.

pub fn background_image_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-image.

pub fn set_background_position_x(&mut self, v: T)

Set background-position-x.

pub fn copy_background_position_x_from(&mut self, other: &Self)

Set background-position-x from other struct.

pub fn reset_background_position_x(&mut self, other: &Self)

Reset background-position-x from the initial struct.

pub fn clone_background_position_x(&self) -> T

Get the computed value for background-position-x.

pub fn background_position_x_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-position-x.

pub fn set_background_position_y(&mut self, v: T)

Set background-position-y.

pub fn copy_background_position_y_from(&mut self, other: &Self)

Set background-position-y from other struct.

pub fn reset_background_position_y(&mut self, other: &Self)

Reset background-position-y from the initial struct.

pub fn clone_background_position_y(&self) -> T

Get the computed value for background-position-y.

pub fn background_position_y_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-position-y.

pub fn set_background_repeat(&mut self, v: T)

Set background-repeat.

pub fn copy_background_repeat_from(&mut self, other: &Self)

Set background-repeat from other struct.

pub fn reset_background_repeat(&mut self, other: &Self)

Reset background-repeat from the initial struct.

pub fn clone_background_repeat(&self) -> T

Get the computed value for background-repeat.

pub fn background_repeat_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-repeat.

pub fn set_background_size(&mut self, v: T)

Set background-size.

pub fn copy_background_size_from(&mut self, other: &Self)

Set background-size from other struct.

pub fn reset_background_size(&mut self, other: &Self)

Reset background-size from the initial struct.

pub fn clone_background_size(&self) -> T

Get the computed value for background-size.

pub fn background_size_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-size.

pub fn set_background_attachment(&mut self, v: T)

Set background-attachment.

pub fn copy_background_attachment_from(&mut self, other: &Self)

Set background-attachment from other struct.

pub fn reset_background_attachment(&mut self, other: &Self)

Reset background-attachment from the initial struct.

pub fn clone_background_attachment(&self) -> T

Get the computed value for background-attachment.

pub fn background_attachment_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-attachment.

pub fn set_background_clip(&mut self, v: T)

Set background-clip.

pub fn copy_background_clip_from(&mut self, other: &Self)

Set background-clip from other struct.

pub fn reset_background_clip(&mut self, other: &Self)

Reset background-clip from the initial struct.

pub fn clone_background_clip(&self) -> T

Get the computed value for background-clip.

pub fn background_clip_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-clip.

pub fn set_background_origin(&mut self, v: T)

Set background-origin.

pub fn copy_background_origin_from(&mut self, other: &Self)

Set background-origin from other struct.

pub fn reset_background_origin(&mut self, other: &Self)

Reset background-origin from the initial struct.

pub fn clone_background_origin(&self) -> T

Get the computed value for background-origin.

pub fn background_origin_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-origin.

pub fn set_background_blend_mode(&mut self, v: T)

Set background-blend-mode.

pub fn copy_background_blend_mode_from(&mut self, other: &Self)

Set background-blend-mode from other struct.

pub fn reset_background_blend_mode(&mut self, other: &Self)

Reset background-blend-mode from the initial struct.

pub fn clone_background_blend_mode(&self) -> T

Get the computed value for background-blend-mode.

pub fn background_blend_mode_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for background-blend-mode.