style::properties::generated::style_structs

Struct InheritedText

pub struct InheritedText {

    pub color: T,
    pub letter_spacing: T,
    pub overflow_wrap: T,
    pub text_align: T,
    pub text_align_last: T,
    pub text_indent: T,
    pub text_justify: T,
    pub text_shadow: T,
    pub text_transform: T,
    pub word_break: T,
    pub word_spacing: T,
    pub forced_color_adjust: T,
    pub line_break: T,
    pub tab_size: T,
    pub text_underline_offset: T,
    pub text_underline_position: T,
    pub text_rendering: T,
    pub _webkit_text_security: T,
    pub text_wrap_mode: T,
    pub white_space_collapse: T,

}

The InheritedText style struct.

Fields
color: T

The color computed value.

letter_spacing: T

The letter-spacing computed value.

overflow_wrap: T

The overflow-wrap computed value.

text_align: T

The text-align computed value.

text_align_last: T

The text-align-last computed value.

text_indent: T

The text-indent computed value.

text_justify: T

The text-justify computed value.

text_shadow: T

The text-shadow computed value.

text_transform: T

The text-transform computed value.

word_break: T

The word-break computed value.

word_spacing: T

The word-spacing computed value.

forced_color_adjust: T

The forced-color-adjust computed value.

line_break: T

The line-break computed value.

tab_size: T

The tab-size computed value.

text_underline_offset: T

The text-underline-offset computed value.

text_underline_position: T

The text-underline-position computed value.

text_rendering: T

The text-rendering computed value.

_webkit_text_security: T

The -webkit-text-security computed value.

text_wrap_mode: T

The text-wrap-mode computed value.

white_space_collapse: T

The white-space-collapse computed value.

Implementations

impl InheritedText

pub fn set_color(&mut self, v: T)

Set color.

pub fn copy_color_from(&mut self, other: &Self)

Set color from other struct.

pub fn reset_color(&mut self, other: &Self)

Reset color from the initial struct.

pub fn clone_color(&self) -> T

Get the computed value for color.

pub fn color_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for color.

pub fn set_letter_spacing(&mut self, v: T)

Set letter-spacing.

pub fn copy_letter_spacing_from(&mut self, other: &Self)

Set letter-spacing from other struct.

pub fn reset_letter_spacing(&mut self, other: &Self)

Reset letter-spacing from the initial struct.

pub fn clone_letter_spacing(&self) -> T

Get the computed value for letter-spacing.

pub fn letter_spacing_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for letter-spacing.

pub fn set_overflow_wrap(&mut self, v: T)

Set overflow-wrap.

pub fn copy_overflow_wrap_from(&mut self, other: &Self)

Set overflow-wrap from other struct.

pub fn reset_overflow_wrap(&mut self, other: &Self)

Reset overflow-wrap from the initial struct.

pub fn clone_overflow_wrap(&self) -> T

Get the computed value for overflow-wrap.

pub fn overflow_wrap_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for overflow-wrap.

pub fn set_text_align(&mut self, v: T)

Set text-align.

pub fn copy_text_align_from(&mut self, other: &Self)

Set text-align from other struct.

pub fn reset_text_align(&mut self, other: &Self)

Reset text-align from the initial struct.

pub fn clone_text_align(&self) -> T

Get the computed value for text-align.

pub fn text_align_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-align.

pub fn set_text_align_last(&mut self, v: T)

Set text-align-last.

pub fn copy_text_align_last_from(&mut self, other: &Self)

Set text-align-last from other struct.

pub fn reset_text_align_last(&mut self, other: &Self)

Reset text-align-last from the initial struct.

pub fn clone_text_align_last(&self) -> T

Get the computed value for text-align-last.

pub fn text_align_last_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-align-last.

pub fn set_text_indent(&mut self, v: T)

Set text-indent.

pub fn copy_text_indent_from(&mut self, other: &Self)

Set text-indent from other struct.

pub fn reset_text_indent(&mut self, other: &Self)

Reset text-indent from the initial struct.

pub fn clone_text_indent(&self) -> T

Get the computed value for text-indent.

pub fn text_indent_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-indent.

pub fn set_text_justify(&mut self, v: T)

Set text-justify.

pub fn copy_text_justify_from(&mut self, other: &Self)

Set text-justify from other struct.

pub fn reset_text_justify(&mut self, other: &Self)

Reset text-justify from the initial struct.

pub fn clone_text_justify(&self) -> T

Get the computed value for text-justify.

pub fn text_justify_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-justify.

pub fn set_text_shadow(&mut self, v: T)

Set text-shadow.

pub fn copy_text_shadow_from(&mut self, other: &Self)

Set text-shadow from other struct.

pub fn reset_text_shadow(&mut self, other: &Self)

Reset text-shadow from the initial struct.

pub fn clone_text_shadow(&self) -> T

Get the computed value for text-shadow.

pub fn text_shadow_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-shadow.

pub fn set_text_transform(&mut self, v: T)

Set text-transform.

pub fn copy_text_transform_from(&mut self, other: &Self)

Set text-transform from other struct.

pub fn reset_text_transform(&mut self, other: &Self)

Reset text-transform from the initial struct.

pub fn clone_text_transform(&self) -> T

Get the computed value for text-transform.

pub fn text_transform_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-transform.

pub fn set_word_break(&mut self, v: T)

Set word-break.

pub fn copy_word_break_from(&mut self, other: &Self)

Set word-break from other struct.

pub fn reset_word_break(&mut self, other: &Self)

Reset word-break from the initial struct.

pub fn clone_word_break(&self) -> T

Get the computed value for word-break.

pub fn word_break_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for word-break.

pub fn set_word_spacing(&mut self, v: T)

Set word-spacing.

pub fn copy_word_spacing_from(&mut self, other: &Self)

Set word-spacing from other struct.

pub fn reset_word_spacing(&mut self, other: &Self)

Reset word-spacing from the initial struct.

pub fn clone_word_spacing(&self) -> T

Get the computed value for word-spacing.

pub fn word_spacing_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for word-spacing.

pub fn set_forced_color_adjust(&mut self, v: T)

Set forced-color-adjust.

pub fn copy_forced_color_adjust_from(&mut self, other: &Self)

Set forced-color-adjust from other struct.

pub fn reset_forced_color_adjust(&mut self, other: &Self)

Reset forced-color-adjust from the initial struct.

pub fn clone_forced_color_adjust(&self) -> T

Get the computed value for forced-color-adjust.

pub fn forced_color_adjust_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for forced-color-adjust.

pub fn set_line_break(&mut self, v: T)

Set line-break.

pub fn copy_line_break_from(&mut self, other: &Self)

Set line-break from other struct.

pub fn reset_line_break(&mut self, other: &Self)

Reset line-break from the initial struct.

pub fn clone_line_break(&self) -> T

Get the computed value for line-break.

pub fn line_break_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for line-break.

pub fn set_tab_size(&mut self, v: T)

Set tab-size.

pub fn copy_tab_size_from(&mut self, other: &Self)

Set tab-size from other struct.

pub fn reset_tab_size(&mut self, other: &Self)

Reset tab-size from the initial struct.

pub fn clone_tab_size(&self) -> T

Get the computed value for tab-size.

pub fn tab_size_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for tab-size.

pub fn set_text_underline_offset(&mut self, v: T)

Set text-underline-offset.

pub fn copy_text_underline_offset_from(&mut self, other: &Self)

Set text-underline-offset from other struct.

pub fn reset_text_underline_offset(&mut self, other: &Self)

Reset text-underline-offset from the initial struct.

pub fn clone_text_underline_offset(&self) -> T

Get the computed value for text-underline-offset.

pub fn text_underline_offset_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-underline-offset.

pub fn set_text_underline_position(&mut self, v: T)

Set text-underline-position.

pub fn copy_text_underline_position_from(&mut self, other: &Self)

Set text-underline-position from other struct.

pub fn reset_text_underline_position(&mut self, other: &Self)

Reset text-underline-position from the initial struct.

pub fn clone_text_underline_position(&self) -> T

Get the computed value for text-underline-position.

pub fn text_underline_position_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-underline-position.

pub fn set_text_rendering(&mut self, v: T)

Set text-rendering.

pub fn copy_text_rendering_from(&mut self, other: &Self)

Set text-rendering from other struct.

pub fn reset_text_rendering(&mut self, other: &Self)

Reset text-rendering from the initial struct.

pub fn clone_text_rendering(&self) -> T

Get the computed value for text-rendering.

pub fn text_rendering_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-rendering.

pub fn set__webkit_text_security(&mut self, v: T)

Set -webkit-text-security.

pub fn copy__webkit_text_security_from(&mut self, other: &Self)

Set -webkit-text-security from other struct.

pub fn reset__webkit_text_security(&mut self, other: &Self)

Reset -webkit-text-security from the initial struct.

pub fn clone__webkit_text_security(&self) -> T

Get the computed value for -webkit-text-security.

pub fn _webkit_text_security_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for -webkit-text-security.

pub fn set_text_wrap_mode(&mut self, v: T)

Set text-wrap-mode.

pub fn copy_text_wrap_mode_from(&mut self, other: &Self)

Set text-wrap-mode from other struct.

pub fn reset_text_wrap_mode(&mut self, other: &Self)

Reset text-wrap-mode from the initial struct.

pub fn clone_text_wrap_mode(&self) -> T

Get the computed value for text-wrap-mode.

pub fn text_wrap_mode_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-wrap-mode.

pub fn set_white_space_collapse(&mut self, v: T)

Set white-space-collapse.

pub fn copy_white_space_collapse_from(&mut self, other: &Self)

Set white-space-collapse from other struct.

pub fn reset_white_space_collapse(&mut self, other: &Self)

Reset white-space-collapse from the initial struct.

pub fn clone_white_space_collapse(&self) -> T

Get the computed value for white-space-collapse.

pub fn white_space_collapse_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for white-space-collapse.