style::properties::generated::style_structs

Struct InheritedUI

pub struct InheritedUI {
    pub caret_color: T,
    pub color_scheme: T,
    pub cursor: T,
    pub pointer_events: T,
    pub scrollbar_color: T,
}

The InheritedUI style struct.

Fields
caret_color: T

The caret-color computed value.

color_scheme: T

The color-scheme computed value.

cursor: T

The cursor computed value.

pointer_events: T

The pointer-events computed value.

scrollbar_color: T

The scrollbar-color computed value.

Implementations

impl InheritedUI

pub fn set_caret_color(&mut self, v: T)

Set caret-color.

pub fn copy_caret_color_from(&mut self, other: &Self)

Set caret-color from other struct.

pub fn reset_caret_color(&mut self, other: &Self)

Reset caret-color from the initial struct.

pub fn clone_caret_color(&self) -> T

Get the computed value for caret-color.

pub fn caret_color_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for caret-color.

pub fn set_color_scheme(&mut self, v: T)

Set color-scheme.

pub fn copy_color_scheme_from(&mut self, other: &Self)

Set color-scheme from other struct.

pub fn reset_color_scheme(&mut self, other: &Self)

Reset color-scheme from the initial struct.

pub fn clone_color_scheme(&self) -> T

Get the computed value for color-scheme.

pub fn color_scheme_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for color-scheme.

pub fn set_cursor(&mut self, v: T)

Set cursor.

pub fn copy_cursor_from(&mut self, other: &Self)

Set cursor from other struct.

pub fn reset_cursor(&mut self, other: &Self)

Reset cursor from the initial struct.

pub fn clone_cursor(&self) -> T

Get the computed value for cursor.

pub fn cursor_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for cursor.

pub fn set_pointer_events(&mut self, v: T)

Set pointer-events.

pub fn copy_pointer_events_from(&mut self, other: &Self)

Set pointer-events from other struct.

pub fn reset_pointer_events(&mut self, other: &Self)

Reset pointer-events from the initial struct.

pub fn clone_pointer_events(&self) -> T

Get the computed value for pointer-events.

pub fn pointer_events_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for pointer-events.

pub fn set_scrollbar_color(&mut self, v: T)

Set scrollbar-color.

pub fn copy_scrollbar_color_from(&mut self, other: &Self)

Set scrollbar-color from other struct.

pub fn reset_scrollbar_color(&mut self, other: &Self)

Reset scrollbar-color from the initial struct.

pub fn clone_scrollbar_color(&self) -> T

Get the computed value for scrollbar-color.

pub fn scrollbar_color_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for scrollbar-color.

pub fn color_scheme_bits(&self) -> ColorSchemeFlags

Returns the ColorSchemeFlags corresponding to the value of color-scheme.