style::properties::generated::style_structs

Struct Outline

pub struct Outline {
    pub outline_color: T,
    pub outline_offset: T,
    pub outline_style: T,
    pub outline_width: T,
}

The Outline style struct.

Fields
outline_color: T

The outline-color computed value.

outline_offset: T

The outline-offset computed value.

outline_style: T

The outline-style computed value.

outline_width: T

The outline-width computed value.

Implementations

impl Outline

pub fn set_outline_color(&mut self, v: T)

Set outline-color.

pub fn copy_outline_color_from(&mut self, other: &Self)

Set outline-color from other struct.

pub fn reset_outline_color(&mut self, other: &Self)

Reset outline-color from the initial struct.

pub fn clone_outline_color(&self) -> T

Get the computed value for outline-color.

pub fn outline_color_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for outline-color.

pub fn set_outline_offset(&mut self, v: T)

Set outline-offset.

pub fn copy_outline_offset_from(&mut self, other: &Self)

Set outline-offset from other struct.

pub fn reset_outline_offset(&mut self, other: &Self)

Reset outline-offset from the initial struct.

pub fn clone_outline_offset(&self) -> T

Get the computed value for outline-offset.

pub fn outline_offset_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for outline-offset.

pub fn set_outline_style(&mut self, v: T)

Set outline-style.

pub fn copy_outline_style_from(&mut self, other: &Self)

Set outline-style from other struct.

pub fn reset_outline_style(&mut self, other: &Self)

Reset outline-style from the initial struct.

pub fn clone_outline_style(&self) -> T

Get the computed value for outline-style.

pub fn outline_style_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for outline-style.

pub fn set_outline_width(&mut self, v: T)

Set outline-width.

pub fn copy_outline_width_from(&mut self, other: &Self)

Set outline-width from other struct.

pub fn reset_outline_width(&mut self, other: &Self)

Reset outline-width from the initial struct.

pub fn clone_outline_width(&self) -> T

Get the computed value for outline-width.

pub fn outline_width_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for outline-width.