style::properties::generated::style_structs

Struct SVG

pub struct SVG {

    pub clip_path: T,
    pub d: T,
    pub cx: T,
    pub cy: T,
    pub mask_image: T,
    pub mask_position_x: T,
    pub mask_position_y: T,
    pub mask_repeat: T,
    pub mask_size: T,
    pub r: T,
    pub rx: T,
    pub ry: T,
    pub x: T,
    pub y: T,
    pub mask_type: T,
    pub mask_mode: T,
    pub mask_clip: T,
    pub mask_origin: T,
    pub mask_composite: T,

}

The SVG style struct.

Fields
clip_path: T

The clip-path computed value.

d: T

The d computed value.

cx: T

The cx computed value.

cy: T

The cy computed value.

mask_image: T

The mask-image computed value.

mask_position_x: T

The mask-position-x computed value.

mask_position_y: T

The mask-position-y computed value.

mask_repeat: T

The mask-repeat computed value.

mask_size: T

The mask-size computed value.

r: T

The r computed value.

rx: T

The rx computed value.

ry: T

The ry computed value.

x: T

The x computed value.

y: T

The y computed value.

mask_type: T

The mask-type computed value.

mask_mode: T

The mask-mode computed value.

mask_clip: T

The mask-clip computed value.

mask_origin: T

The mask-origin computed value.

mask_composite: T

The mask-composite computed value.

Implementations

impl SVG

pub fn set_clip_path(&mut self, v: T)

Set clip-path.

pub fn copy_clip_path_from(&mut self, other: &Self)

Set clip-path from other struct.

pub fn reset_clip_path(&mut self, other: &Self)

Reset clip-path from the initial struct.

pub fn clone_clip_path(&self) -> T

Get the computed value for clip-path.

pub fn clip_path_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for clip-path.

pub fn set_d(&mut self, v: T)

Set d.

pub fn copy_d_from(&mut self, other: &Self)

Set d from other struct.

pub fn reset_d(&mut self, other: &Self)

Reset d from the initial struct.

pub fn clone_d(&self) -> T

Get the computed value for d.

pub fn d_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for d.

pub fn set_cx(&mut self, v: T)

Set cx.

pub fn copy_cx_from(&mut self, other: &Self)

Set cx from other struct.

pub fn reset_cx(&mut self, other: &Self)

Reset cx from the initial struct.

pub fn clone_cx(&self) -> T

Get the computed value for cx.

pub fn cx_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for cx.

pub fn set_cy(&mut self, v: T)

Set cy.

pub fn copy_cy_from(&mut self, other: &Self)

Set cy from other struct.

pub fn reset_cy(&mut self, other: &Self)

Reset cy from the initial struct.

pub fn clone_cy(&self) -> T

Get the computed value for cy.

pub fn cy_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for cy.

pub fn set_mask_image(&mut self, v: T)

Set mask-image.

pub fn copy_mask_image_from(&mut self, other: &Self)

Set mask-image from other struct.

pub fn reset_mask_image(&mut self, other: &Self)

Reset mask-image from the initial struct.

pub fn clone_mask_image(&self) -> T

Get the computed value for mask-image.

pub fn mask_image_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-image.

pub fn set_mask_position_x(&mut self, v: T)

Set mask-position-x.

pub fn copy_mask_position_x_from(&mut self, other: &Self)

Set mask-position-x from other struct.

pub fn reset_mask_position_x(&mut self, other: &Self)

Reset mask-position-x from the initial struct.

pub fn clone_mask_position_x(&self) -> T

Get the computed value for mask-position-x.

pub fn mask_position_x_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-position-x.

pub fn set_mask_position_y(&mut self, v: T)

Set mask-position-y.

pub fn copy_mask_position_y_from(&mut self, other: &Self)

Set mask-position-y from other struct.

pub fn reset_mask_position_y(&mut self, other: &Self)

Reset mask-position-y from the initial struct.

pub fn clone_mask_position_y(&self) -> T

Get the computed value for mask-position-y.

pub fn mask_position_y_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-position-y.

pub fn set_mask_repeat(&mut self, v: T)

Set mask-repeat.

pub fn copy_mask_repeat_from(&mut self, other: &Self)

Set mask-repeat from other struct.

pub fn reset_mask_repeat(&mut self, other: &Self)

Reset mask-repeat from the initial struct.

pub fn clone_mask_repeat(&self) -> T

Get the computed value for mask-repeat.

pub fn mask_repeat_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-repeat.

pub fn set_mask_size(&mut self, v: T)

Set mask-size.

pub fn copy_mask_size_from(&mut self, other: &Self)

Set mask-size from other struct.

pub fn reset_mask_size(&mut self, other: &Self)

Reset mask-size from the initial struct.

pub fn clone_mask_size(&self) -> T

Get the computed value for mask-size.

pub fn mask_size_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-size.

pub fn set_r(&mut self, v: T)

Set r.

pub fn copy_r_from(&mut self, other: &Self)

Set r from other struct.

pub fn reset_r(&mut self, other: &Self)

Reset r from the initial struct.

pub fn clone_r(&self) -> T

Get the computed value for r.

pub fn r_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for r.

pub fn set_rx(&mut self, v: T)

Set rx.

pub fn copy_rx_from(&mut self, other: &Self)

Set rx from other struct.

pub fn reset_rx(&mut self, other: &Self)

Reset rx from the initial struct.

pub fn clone_rx(&self) -> T

Get the computed value for rx.

pub fn rx_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for rx.

pub fn set_ry(&mut self, v: T)

Set ry.

pub fn copy_ry_from(&mut self, other: &Self)

Set ry from other struct.

pub fn reset_ry(&mut self, other: &Self)

Reset ry from the initial struct.

pub fn clone_ry(&self) -> T

Get the computed value for ry.

pub fn ry_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for ry.

pub fn set_x(&mut self, v: T)

Set x.

pub fn copy_x_from(&mut self, other: &Self)

Set x from other struct.

pub fn reset_x(&mut self, other: &Self)

Reset x from the initial struct.

pub fn clone_x(&self) -> T

Get the computed value for x.

pub fn x_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for x.

pub fn set_y(&mut self, v: T)

Set y.

pub fn copy_y_from(&mut self, other: &Self)

Set y from other struct.

pub fn reset_y(&mut self, other: &Self)

Reset y from the initial struct.

pub fn clone_y(&self) -> T

Get the computed value for y.

pub fn y_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for y.

pub fn set_mask_type(&mut self, v: T)

Set mask-type.

pub fn copy_mask_type_from(&mut self, other: &Self)

Set mask-type from other struct.

pub fn reset_mask_type(&mut self, other: &Self)

Reset mask-type from the initial struct.

pub fn clone_mask_type(&self) -> T

Get the computed value for mask-type.

pub fn mask_type_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-type.

pub fn set_mask_mode(&mut self, v: T)

Set mask-mode.

pub fn copy_mask_mode_from(&mut self, other: &Self)

Set mask-mode from other struct.

pub fn reset_mask_mode(&mut self, other: &Self)

Reset mask-mode from the initial struct.

pub fn clone_mask_mode(&self) -> T

Get the computed value for mask-mode.

pub fn mask_mode_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-mode.

pub fn set_mask_clip(&mut self, v: T)

Set mask-clip.

pub fn copy_mask_clip_from(&mut self, other: &Self)

Set mask-clip from other struct.

pub fn reset_mask_clip(&mut self, other: &Self)

Reset mask-clip from the initial struct.

pub fn clone_mask_clip(&self) -> T

Get the computed value for mask-clip.

pub fn mask_clip_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-clip.

pub fn set_mask_origin(&mut self, v: T)

Set mask-origin.

pub fn copy_mask_origin_from(&mut self, other: &Self)

Set mask-origin from other struct.

pub fn reset_mask_origin(&mut self, other: &Self)

Reset mask-origin from the initial struct.

pub fn clone_mask_origin(&self) -> T

Get the computed value for mask-origin.

pub fn mask_origin_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-origin.

pub fn set_mask_composite(&mut self, v: T)

Set mask-composite.

pub fn copy_mask_composite_from(&mut self, other: &Self)

Set mask-composite from other struct.

pub fn reset_mask_composite(&mut self, other: &Self)

Reset mask-composite from the initial struct.

pub fn clone_mask_composite(&self) -> T

Get the computed value for mask-composite.

pub fn mask_composite_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for mask-composite.