style::properties::generated::style_structs

Struct Text

pub struct Text {
    pub text_overflow: T,
    pub text_decoration_line: T,
    pub text_decoration_color: T,
    pub text_decoration_inset: T,
    pub text_decoration_thickness: T,
    pub unicode_bidi: T,
    pub text_decoration_style: T,
}

The Text style struct.

Fields
text_overflow: T

The text-overflow computed value.

text_decoration_line: T

The text-decoration-line computed value.

text_decoration_color: T

The text-decoration-color computed value.

text_decoration_inset: T

The text-decoration-inset computed value.

text_decoration_thickness: T

The text-decoration-thickness computed value.

unicode_bidi: T

The unicode-bidi computed value.

text_decoration_style: T

The text-decoration-style computed value.

Implementations

impl Text

pub fn set_text_overflow(&mut self, v: T)

Set text-overflow.

pub fn copy_text_overflow_from(&mut self, other: &Self)

Set text-overflow from other struct.

pub fn reset_text_overflow(&mut self, other: &Self)

Reset text-overflow from the initial struct.

pub fn clone_text_overflow(&self) -> T

Get the computed value for text-overflow.

pub fn text_overflow_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-overflow.

pub fn set_text_decoration_line(&mut self, v: T)

Set text-decoration-line.

pub fn copy_text_decoration_line_from(&mut self, other: &Self)

Set text-decoration-line from other struct.

pub fn reset_text_decoration_line(&mut self, other: &Self)

Reset text-decoration-line from the initial struct.

pub fn clone_text_decoration_line(&self) -> T

Get the computed value for text-decoration-line.

pub fn text_decoration_line_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-decoration-line.

pub fn set_text_decoration_color(&mut self, v: T)

Set text-decoration-color.

pub fn copy_text_decoration_color_from(&mut self, other: &Self)

Set text-decoration-color from other struct.

pub fn reset_text_decoration_color(&mut self, other: &Self)

Reset text-decoration-color from the initial struct.

pub fn clone_text_decoration_color(&self) -> T

Get the computed value for text-decoration-color.

pub fn text_decoration_color_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-decoration-color.

pub fn set_text_decoration_inset(&mut self, v: T)

Set text-decoration-inset.

pub fn copy_text_decoration_inset_from(&mut self, other: &Self)

Set text-decoration-inset from other struct.

pub fn reset_text_decoration_inset(&mut self, other: &Self)

Reset text-decoration-inset from the initial struct.

pub fn clone_text_decoration_inset(&self) -> T

Get the computed value for text-decoration-inset.

pub fn text_decoration_inset_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-decoration-inset.

pub fn set_text_decoration_thickness(&mut self, v: T)

Set text-decoration-thickness.

pub fn copy_text_decoration_thickness_from(&mut self, other: &Self)

Set text-decoration-thickness from other struct.

pub fn reset_text_decoration_thickness(&mut self, other: &Self)

Reset text-decoration-thickness from the initial struct.

pub fn clone_text_decoration_thickness(&self) -> T

Get the computed value for text-decoration-thickness.

pub fn text_decoration_thickness_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-decoration-thickness.

pub fn set_unicode_bidi(&mut self, v: T)

Set unicode-bidi.

pub fn copy_unicode_bidi_from(&mut self, other: &Self)

Set unicode-bidi from other struct.

pub fn reset_unicode_bidi(&mut self, other: &Self)

Reset unicode-bidi from the initial struct.

pub fn clone_unicode_bidi(&self) -> T

Get the computed value for unicode-bidi.

pub fn unicode_bidi_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for unicode-bidi.

pub fn set_text_decoration_style(&mut self, v: T)

Set text-decoration-style.

pub fn copy_text_decoration_style_from(&mut self, other: &Self)

Set text-decoration-style from other struct.

pub fn reset_text_decoration_style(&mut self, other: &Self)

Reset text-decoration-style from the initial struct.

pub fn clone_text_decoration_style(&self) -> T

Get the computed value for text-decoration-style.

pub fn text_decoration_style_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for text-decoration-style.