Struct style::properties::longhands::animation_play_state::computed_value::OwnedList

pub struct OwnedList<T>(pub UnderlyingOwnedList<T>);

The generic type defining the animated and resolved values for this property.

Making this type generic allows the compiler to figure out the animated value for us, instead of having to implement it manually for every type we care about.

Tuple Fields
0: UnderlyingOwnedList<T>