Struct style::properties::longhands::transform::computed_value::T

pub struct T(pub Option<Vec<ComputedOperation>>);

Methods

impl TransformList

pub fn to_transform_3d_matrix(
&self,
reference_box: Option<&Rect<Au>>
) -> Option<Transform3D<CSSFloat>>

Return the equivalent 3d matrix of this transform list. If |reference_box| is None, we will drop the percent part from translate because we can resolve it without the layout info.

pub fn create_perspective_matrix(d: CSSFloat) -> Transform3D<f32>

Return the transform matrix from a perspective length.

pub fn get_normalized_vector_and_angle(
x: f32,
y: f32,
z: f32,
angle: Angle
) -> (f32, f32, f32, Angle)

Return the normalized direction vector and its angle for Rotate3D.