Enum style::properties::longhands::transform::SpecifiedOperation

pub enum SpecifiedOperation {
    Matrix(Matrix<Number>),
    PrefixedMatrix(Matrix<Number, LoPoNumber>),
    Matrix3D {
        m11: Number,
        m12: Number,
        m13: Number,
        m14: Number,
        m21: Number,
        m22: Number,
        m23: Number,
        m24: Number,
        m31: Number,
        m32: Number,
        m33: Number,
        m34: Number,
        m41: Number,
        m42: Number,
        m43: Number,
        m44: Number,
    },
    PrefixedMatrix3D {
        m11: Number,
        m12: Number,
        m13: Number,
        m14: Number,
        m21: Number,
        m22: Number,
        m23: Number,
        m24: Number,
        m31: Number,
        m32: Number,
        m33: Number,
        m34: Number,
        m41: LoPoNumber,
        m42: LoPoNumber,
        m43: LengthOrNumber,
        m44: Number,
    },
    Skew(Angle, Option<Angle>),
    SkewX(Angle),
    SkewY(Angle),
    Translate(LengthOrPercentage, Option<LengthOrPercentage>),
    TranslateX(LengthOrPercentage),
    TranslateY(LengthOrPercentage),
    TranslateZ(Length),
    Translate3D(LengthOrPercentage, LengthOrPercentage, Length),
    Scale(Number, Option<Number>),
    ScaleX(Number),
    ScaleY(Number),
    ScaleZ(Number),
    Scale3D(Number, Number, Number),
    Rotate(Angle),
    RotateX(Angle),
    RotateY(Angle),
    RotateZ(Angle),
    Rotate3D(Number, Number, Number, Angle),
    Perspective(Length),
    InterpolateMatrix {
        from_list: SpecifiedValue,
        to_list: SpecifiedValue,
        progress: Percentage,
    },
    AccumulateMatrix {
        from_list: SpecifiedValue,
        to_list: SpecifiedValue,
        count: Integer,
    },
}

Describes a single parsed Transform Function.

Multiple transform functions compose a transformation.

Some transformations can be expressed by other more general functions.

Variants
Matrix(Matrix<Number>)

Represents a 2D 2x3 matrix.

PrefixedMatrix(Matrix<Number, LoPoNumber>)

Represents a 3D 4x4 matrix with percentage and length values. For moz-transform.

Matrix3D

Represents a 3D 4x4 matrix.

Fields of Matrix3D

m11: Number

m12: Number

m13: Number

m14: Number

m21: Number

m22: Number

m23: Number

m24: Number

m31: Number

m32: Number

m33: Number

m34: Number

m41: Number

m42: Number

m43: Number

m44: Number

PrefixedMatrix3D

Represents a 3D 4x4 matrix with percentage and length values. For moz-transform.

Fields of PrefixedMatrix3D

m11: Number

m12: Number

m13: Number

m14: Number

m21: Number

m22: Number

m23: Number

m24: Number

m31: Number

m32: Number

m33: Number

m34: Number

m41: LoPoNumber

m42: LoPoNumber

m43: LengthOrNumber

m44: Number

Skew(Angle, Option<Angle>)

A 2D skew.

If the second angle is not provided it is assumed zero.

SkewX(Angle)SkewY(Angle)Translate(LengthOrPercentage, Option<LengthOrPercentage>)TranslateX(LengthOrPercentage)TranslateY(LengthOrPercentage)TranslateZ(Length)Translate3D(LengthOrPercentage, LengthOrPercentage, Length)Scale(Number, Option<Number>)

A 2D scaling factor.

scale(2) is parsed as Scale(Number::new(2.0), None) and is equivalent to writing scale(2, 2) (Scale(Number::new(2.0), Some(Number::new(2.0)))).

Negative values are allowed and flip the element.

ScaleX(Number)ScaleY(Number)ScaleZ(Number)Scale3D(Number, Number, Number)Rotate(Angle)

Describes a 2D Rotation.

In a 3D scene rotate(angle) is equivalent to rotateZ(angle).

RotateX(Angle)

Rotation in 3D space around the x-axis.

RotateY(Angle)

Rotation in 3D space around the y-axis.

RotateZ(Angle)

Rotation in 3D space around the z-axis.

Rotate3D(Number, Number, Number, Angle)

Rotation in 3D space.

Generalization of rotateX, rotateY and rotateZ.

Perspective(Length)

Specifies a perspective projection matrix.

Part of CSS Transform Module Level 2 and defined at  13.1. 3D Transform Function.

The value must be greater than or equal to zero.

InterpolateMatrix

A intermediate type for interpolation of mismatched transform lists.

Fields of InterpolateMatrix

from_list: SpecifiedValue

to_list: SpecifiedValue

progress: Percentage

AccumulateMatrix

A intermediate type for accumulation of mismatched transform lists.

Fields of AccumulateMatrix

from_list: SpecifiedValue

to_list: SpecifiedValue

count: Integer