Struct style::properties::shorthands::animation::LonghandsToSerialize

pub struct LonghandsToSerialize<'a> {
    pub animation_name: &'a SpecifiedValue,
    pub animation_duration: &'a SpecifiedValue,
    pub animation_timing_function: &'a SpecifiedValue,
    pub animation_delay: &'a SpecifiedValue,
    pub animation_iteration_count: &'a SpecifiedValue,
    pub animation_direction: &'a SpecifiedValue,
    pub animation_fill_mode: &'a SpecifiedValue,
    pub animation_play_state: &'a SpecifiedValue,
}

Represents a serializable set of all of the longhand properties that correspond to a shorthand.

Fields
animation_name: &'a SpecifiedValue animation_duration: &'a SpecifiedValue animation_timing_function: &'a SpecifiedValue animation_delay: &'a SpecifiedValue animation_iteration_count: &'a SpecifiedValue animation_direction: &'a SpecifiedValue animation_fill_mode: &'a SpecifiedValue animation_play_state: &'a SpecifiedValue
Methods

impl<'a> LonghandsToSerialize<'a>

pub fn from_iter<I>(iter: I) -> Result<Self, ()> where
I: Iterator<Item = &'a PropertyDeclaration>,

Tries to get a serializable set of longhands given a set of property declarations.