Struct style::properties::shorthands::border_block_end::LonghandsToSerialize

pub struct LonghandsToSerialize<'a> {
    pub border_block_end_color: &'a SpecifiedValue,
    pub border_block_end_style: &'a SpecifiedValue,
    pub border_block_end_width: &'a SpecifiedValue,
}

Represents a serializable set of all of the longhand properties that correspond to a shorthand.

Fields
border_block_end_color: &'a SpecifiedValue border_block_end_style: &'a SpecifiedValue border_block_end_width: &'a SpecifiedValue
Methods

impl<'a> LonghandsToSerialize<'a>

pub fn from_iter<I>(iter: I) -> Result<Self, ()> where
I: Iterator<Item = &'a PropertyDeclaration>,

Tries to get a serializable set of longhands given a set of property declarations.