Struct style::properties::shorthands::outline::LonghandsToSerialize

pub struct LonghandsToSerialize<'a> {
    pub outline_color: &'a DeclaredValue<Box<SpecifiedValue>>,
    pub outline_style: &'a DeclaredValue<SpecifiedValue>,
    pub outline_width: &'a DeclaredValue<SpecifiedValue>,
}

Represents a serializable set of all of the longhand properties that correspond to a shorthand.

Fields
outline_color: &'a DeclaredValue<Box<SpecifiedValue>> outline_style: &'a DeclaredValue<SpecifiedValue> outline_width: &'a DeclaredValue<SpecifiedValue>
Methods

impl<'a> LonghandsToSerialize<'a>

fn from_iter<I>(iter: I) -> Result<Self, ()> where I: Iterator<Item=&'a PropertyDeclaration>

Tries to get a serializable set of longhands given a set of property declarations.