Struct style::properties::ComputedValuesInner

pub struct ComputedValuesInner {
    pub writing_mode: WritingMode,
    pub font_computation_data: FontComputationData,
    pub flags: ComputedValueFlags,
    pub rules: Option<StrongRuleNode>,
    // some fields omitted
}

Actual data of ComputedValues, to match up with Gecko

Fields
writing_mode: WritingMode

The writing mode of this computed values struct.

font_computation_data: FontComputationData

The keyword behind the current font-size property, if any

flags: ComputedValueFlags

A set of flags we use to store misc information regarding this style.

rules: Option<StrongRuleNode>

The rule node representing the ordered list of rules matched for this node. Can be None for default values and text nodes. This is essentially an optimization to avoid referencing the root rule node.

Methods

impl ComputedValuesInner

pub fn new(
custom_properties: Option<Arc<CustomPropertiesMap>>,
writing_mode: WritingMode,
font_size_keyword: FontComputationData,
flags: ComputedValueFlags,
rules: Option<StrongRuleNode>,
visited_style: Option<Arc<ComputedValues>>,
background: Arc<Background>,
border: Arc<Border>,
box_: Arc<Box>,
color: Arc<Color>,
column: Arc<Column>,
counters: Arc<Counters>,
effects: Arc<Effects>,
font: Arc<Font>,
inheritedbox: Arc<InheritedBox>,
inheritedtable: Arc<InheritedTable>,
inheritedtext: Arc<InheritedText>,
list: Arc<List>,
margin: Arc<Margin>,
outline: Arc<Outline>,
padding: Arc<Padding>,
pointing: Arc<Pointing>,
position: Arc<Position>,
table: Arc<Table>,
text: Arc<Text>
) -> Self

Construct a ComputedValuesInner instance.

impl ComputedValuesInner

pub fn clone_background(&self) -> Arc<Background>

Clone the Background struct.

pub fn get_background(&self) -> &Background

Get a immutable reference to the Background struct.

pub fn background_arc(&self) -> &Arc<Background>

Gets an immutable reference to the refcounted value that wraps Background.

pub fn mutate_background(&mut self) -> &mut Background

Get a mutable reference to the Background struct.

pub fn clone_border(&self) -> Arc<Border>

Clone the Border struct.

pub fn get_border(&self) -> &Border

Get a immutable reference to the Border struct.

pub fn border_arc(&self) -> &Arc<Border>

Gets an immutable reference to the refcounted value that wraps Border.

pub fn mutate_border(&mut self) -> &mut Border

Get a mutable reference to the Border struct.

pub fn clone_box(&self) -> Arc<Box>

Clone the Box struct.

pub fn get_box(&self) -> &Box

Get a immutable reference to the Box struct.

pub fn box_arc(&self) -> &Arc<Box>

Gets an immutable reference to the refcounted value that wraps Box.

pub fn mutate_box(&mut self) -> &mut Box

Get a mutable reference to the Box struct.

pub fn clone_color(&self) -> Arc<Color>

Clone the Color struct.

pub fn get_color(&self) -> &Color

Get a immutable reference to the Color struct.

pub fn color_arc(&self) -> &Arc<Color>

Gets an immutable reference to the refcounted value that wraps Color.

pub fn mutate_color(&mut self) -> &mut Color

Get a mutable reference to the Color struct.

pub fn clone_column(&self) -> Arc<Column>

Clone the Column struct.

pub fn get_column(&self) -> &Column

Get a immutable reference to the Column struct.

pub fn column_arc(&self) -> &Arc<Column>

Gets an immutable reference to the refcounted value that wraps Column.

pub fn mutate_column(&mut self) -> &mut Column

Get a mutable reference to the Column struct.

pub fn clone_counters(&self) -> Arc<Counters>

Clone the Counters struct.

pub fn get_counters(&self) -> &Counters

Get a immutable reference to the Counters struct.

pub fn counters_arc(&self) -> &Arc<Counters>

Gets an immutable reference to the refcounted value that wraps Counters.

pub fn mutate_counters(&mut self) -> &mut Counters

Get a mutable reference to the Counters struct.

pub fn clone_effects(&self) -> Arc<Effects>

Clone the Effects struct.

pub fn get_effects(&self) -> &Effects

Get a immutable reference to the Effects struct.

pub fn effects_arc(&self) -> &Arc<Effects>

Gets an immutable reference to the refcounted value that wraps Effects.

pub fn mutate_effects(&mut self) -> &mut Effects

Get a mutable reference to the Effects struct.

pub fn clone_font(&self) -> Arc<Font>

Clone the Font struct.

pub fn get_font(&self) -> &Font

Get a immutable reference to the Font struct.

pub fn font_arc(&self) -> &Arc<Font>

Gets an immutable reference to the refcounted value that wraps Font.

pub fn mutate_font(&mut self) -> &mut Font

Get a mutable reference to the Font struct.

pub fn clone_inheritedbox(&self) -> Arc<InheritedBox>

Clone the InheritedBox struct.

pub fn get_inheritedbox(&self) -> &InheritedBox

Get a immutable reference to the InheritedBox struct.

pub fn inheritedbox_arc(&self) -> &Arc<InheritedBox>

Gets an immutable reference to the refcounted value that wraps InheritedBox.

pub fn mutate_inheritedbox(&mut self) -> &mut InheritedBox

Get a mutable reference to the InheritedBox struct.

pub fn clone_inheritedtable(&self) -> Arc<InheritedTable>

Clone the InheritedTable struct.

pub fn get_inheritedtable(&self) -> &InheritedTable

Get a immutable reference to the InheritedTable struct.

pub fn inheritedtable_arc(&self) -> &Arc<InheritedTable>

Gets an immutable reference to the refcounted value that wraps InheritedTable.

pub fn mutate_inheritedtable(&mut self) -> &mut InheritedTable

Get a mutable reference to the InheritedTable struct.

pub fn clone_inheritedtext(&self) -> Arc<InheritedText>

Clone the InheritedText struct.

pub fn get_inheritedtext(&self) -> &InheritedText

Get a immutable reference to the InheritedText struct.

pub fn inheritedtext_arc(&self) -> &Arc<InheritedText>

Gets an immutable reference to the refcounted value that wraps InheritedText.

pub fn mutate_inheritedtext(&mut self) -> &mut InheritedText

Get a mutable reference to the InheritedText struct.

pub fn clone_list(&self) -> Arc<List>

Clone the List struct.

pub fn get_list(&self) -> &List

Get a immutable reference to the List struct.

pub fn list_arc(&self) -> &Arc<List>

Gets an immutable reference to the refcounted value that wraps List.

pub fn mutate_list(&mut self) -> &mut List

Get a mutable reference to the List struct.

pub fn clone_margin(&self) -> Arc<Margin>

Clone the Margin struct.

pub fn get_margin(&self) -> &Margin

Get a immutable reference to the Margin struct.

pub fn margin_arc(&self) -> &Arc<Margin>

Gets an immutable reference to the refcounted value that wraps Margin.

pub fn mutate_margin(&mut self) -> &mut Margin

Get a mutable reference to the Margin struct.

pub fn clone_outline(&self) -> Arc<Outline>

Clone the Outline struct.

pub fn get_outline(&self) -> &Outline

Get a immutable reference to the Outline struct.

pub fn outline_arc(&self) -> &Arc<Outline>

Gets an immutable reference to the refcounted value that wraps Outline.

pub fn mutate_outline(&mut self) -> &mut Outline

Get a mutable reference to the Outline struct.

pub fn clone_padding(&self) -> Arc<Padding>

Clone the Padding struct.

pub fn get_padding(&self) -> &Padding

Get a immutable reference to the Padding struct.

pub fn padding_arc(&self) -> &Arc<Padding>

Gets an immutable reference to the refcounted value that wraps Padding.

pub fn mutate_padding(&mut self) -> &mut Padding

Get a mutable reference to the Padding struct.

pub fn clone_pointing(&self) -> Arc<Pointing>

Clone the Pointing struct.

pub fn get_pointing(&self) -> &Pointing

Get a immutable reference to the Pointing struct.

pub fn pointing_arc(&self) -> &Arc<Pointing>

Gets an immutable reference to the refcounted value that wraps Pointing.

pub fn mutate_pointing(&mut self) -> &mut Pointing

Get a mutable reference to the Pointing struct.

pub fn clone_position(&self) -> Arc<Position>

Clone the Position struct.

pub fn get_position(&self) -> &Position

Get a immutable reference to the Position struct.

pub fn position_arc(&self) -> &Arc<Position>

Gets an immutable reference to the refcounted value that wraps Position.

pub fn mutate_position(&mut self) -> &mut Position

Get a mutable reference to the Position struct.

pub fn clone_table(&self) -> Arc<Table>

Clone the Table struct.

pub fn get_table(&self) -> &Table

Get a immutable reference to the Table struct.

pub fn table_arc(&self) -> &Arc<Table>

Gets an immutable reference to the refcounted value that wraps Table.

pub fn mutate_table(&mut self) -> &mut Table

Get a mutable reference to the Table struct.

pub fn clone_text(&self) -> Arc<Text>

Clone the Text struct.

pub fn get_text(&self) -> &Text

Get a immutable reference to the Text struct.

pub fn text_arc(&self) -> &Arc<Text>

Gets an immutable reference to the refcounted value that wraps Text.

pub fn mutate_text(&mut self) -> &mut Text

Get a mutable reference to the Text struct.

pub fn rules(&self) -> &StrongRuleNode

Gets a reference to the rule node. Panic if no rule node exists.

pub fn has_visited_style(&self) -> bool

Whether there is a visited style.

pub fn get_visited_style(&self) -> Option<&ComputedValues>

Gets a reference to the visited style, if any.

pub fn visited_style(&self) -> &ComputedValues

Gets a reference to the visited style. Panic if no visited style exists.

pub fn clone_visited_style(&self) -> Option<Arc<ComputedValues>>

Clone the visited style. Used for inheriting parent styles in StyleBuilder::for_inheritance.

pub fn get_custom_properties(&self) -> Option<&CustomPropertiesMap>

Gets a reference to the custom properties map (if one exists).

pub fn custom_properties(&self) -> Option<Arc<CustomPropertiesMap>>

Get the custom properties map if necessary.

Cloning the Arc here is fine because it only happens in the case where we have custom properties, and those are both rare and expensive.

pub fn has_moz_binding(&self) -> bool

Whether this style has a -moz-binding value. This is always false for Servo for obvious reasons.

pub fn is_style_if_visited(&self) -> bool

Whether we're a visited style.

pub fn is_display_contents(&self) -> bool

Returns whether this style's display value is equal to contents.

Since this isn't supported in Servo, this is always false for Servo.

pub fn ineffective_content_property(&self) -> bool

Returns whether the "content" property for the given style is completely ineffective, and would yield an empty ::before or ::after pseudo-element.

pub fn is_multicol(&self) -> bool

Whether the current style is multicolumn.

pub fn resolve_color(&self, color: Color) -> RGBA

Resolves the currentColor keyword.

Any color value from computed values (except for the 'color' property itself) should go through this method.

Usage example: let top_color = style.resolve_color(style.Border.border_top_color);

pub fn content_inline_size(&self) -> LengthOrPercentageOrAuto

Get the logical computed inline size.

pub fn content_block_size(&self) -> LengthOrPercentageOrAuto

Get the logical computed block size.

pub fn min_inline_size(&self) -> LengthOrPercentage

Get the logical computed min inline size.

pub fn min_block_size(&self) -> LengthOrPercentage

Get the logical computed min block size.

pub fn max_inline_size(&self) -> LengthOrPercentageOrNone

Get the logical computed max inline size.

pub fn max_block_size(&self) -> LengthOrPercentageOrNone

Get the logical computed max block size.

pub fn logical_padding(&self) -> LogicalMargin<LengthOrPercentage>

Get the logical computed padding for this writing mode.

pub fn border_width_for_writing_mode(
&self,
writing_mode: WritingMode
) -> LogicalMargin<Au>

Get the logical border width

pub fn logical_border_width(&self) -> LogicalMargin<Au>

Gets the logical computed border widths for this style.

pub fn logical_margin(&self) -> LogicalMargin<LengthOrPercentageOrAuto>

Gets the logical computed margin from this style.

pub fn logical_position(&self) -> LogicalMargin<LengthOrPercentageOrAuto>

Gets the logical position from this style.

pub fn overrides_transform_style(&self) -> bool

Return true if the effects force the transform style to be Flat

pub fn get_used_transform_style(&self) -> T

https://drafts.csswg.org/css-transforms/#grouping-property-values

pub fn transform_requires_layer(&self) -> bool

Whether given this transform value, the compositor would require a layer.

pub fn computed_value_to_string(
&self,
property: PropertyDeclarationId
) -> String

Serializes the computed value of this property as a string.