Struct style::properties::LonghandIdSet

pub struct LonghandIdSet {
    storage: [u32; 6],
}

A set of longhand properties

Fields
storage: [u32; 6]
Implementations

impl LonghandIdSet

fn reset() -> &'static Self

fn animatable() -> &'static Self

fn discrete_animatable() -> &'static Self

fn logical() -> &'static Self

pub fn ignored_when_colors_disabled() -> &'static Self

Returns the set of longhands that are ignored when document colors are disabled.

pub fn padding_properties() -> &'static Self

Returns the set of padding properties for the purpose of disabling native appearance.

pub fn border_background_properties() -> &'static Self

Returns the set of border properties for the purpose of disabling native appearance.

pub fn iter(&self) -> LonghandIdSetIterator<'_>

Iterate over the current longhand id set.

pub fn contains_all(&self, other: &Self) -> bool

Returns whether this set contains at least every longhand that other also contains.

pub fn contains_any(&self, other: &Self) -> bool

Returns whether this set contains any longhand that other also contains.

pub fn remove_all(&mut self, other: &Self)

Remove all the given properties from the set.

pub fn new() -> LonghandIdSet

Create an empty set

pub fn contains(&self, id: LonghandId) -> bool

Return whether the given property is in the set

pub fn contains_any_reset(&self) -> bool

Return whether this set contains any reset longhand.

pub fn insert(&mut self, id: LonghandId)

Add the given property to the set

pub fn remove(&mut self, id: LonghandId)

Remove the given property from the set

pub fn clear(&mut self)

Clear all bits

pub fn is_empty(&self) -> bool

Returns whether the set is empty.