Struct style::properties::style_structs::Padding

pub struct Padding {
    pub padding_top: T,
    pub padding_left: T,
    pub padding_bottom: T,
    pub padding_right: T,
    pub padding_block_start: T,
    pub padding_block_end: T,
    pub padding_inline_start: T,
    pub padding_inline_end: T,
}

The Padding style struct.

Fields
padding_top: T

The padding-top computed value.

padding_left: T

The padding-left computed value.

padding_bottom: T

The padding-bottom computed value.

padding_right: T

The padding-right computed value.

padding_block_start: T

The padding-block-start computed value.

padding_block_end: T

The padding-block-end computed value.

padding_inline_start: T

The padding-inline-start computed value.

padding_inline_end: T

The padding-inline-end computed value.

Methods

impl Padding

pub fn set_padding_top(&mut self, v: T)

Set padding-top.

pub fn copy_padding_top_from(&mut self, other: &Self)

Set padding-top from other struct.

pub fn reset_padding_top(&mut self, other: &Self)

Reset padding-top from the initial struct.

pub fn clone_padding_top(&self) -> T

Get the computed value for padding-top.

pub fn set_padding_left(&mut self, v: T)

Set padding-left.

pub fn copy_padding_left_from(&mut self, other: &Self)

Set padding-left from other struct.

pub fn reset_padding_left(&mut self, other: &Self)

Reset padding-left from the initial struct.

pub fn clone_padding_left(&self) -> T

Get the computed value for padding-left.

pub fn set_padding_bottom(&mut self, v: T)

Set padding-bottom.

pub fn copy_padding_bottom_from(&mut self, other: &Self)

Set padding-bottom from other struct.

pub fn reset_padding_bottom(&mut self, other: &Self)

Reset padding-bottom from the initial struct.

pub fn clone_padding_bottom(&self) -> T

Get the computed value for padding-bottom.

pub fn set_padding_right(&mut self, v: T)

Set padding-right.

pub fn copy_padding_right_from(&mut self, other: &Self)

Set padding-right from other struct.

pub fn reset_padding_right(&mut self, other: &Self)

Reset padding-right from the initial struct.

pub fn clone_padding_right(&self) -> T

Get the computed value for padding-right.

pub fn set_padding_block_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-block-start given a writing mode.

pub fn copy_padding_block_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-block-start given a writing mode.

pub fn reset_padding_block_start(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-block-start given a writing mode.

pub fn clone_padding_block_start(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for padding-block-start given a writing mode.

pub fn set_padding_block_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-block-end given a writing mode.

pub fn copy_padding_block_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-block-end given a writing mode.

pub fn reset_padding_block_end(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-block-end given a writing mode.

pub fn clone_padding_block_end(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for padding-block-end given a writing mode.

pub fn set_padding_inline_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-inline-start given a writing mode.

pub fn copy_padding_inline_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-inline-start given a writing mode.

pub fn reset_padding_inline_start(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-inline-start given a writing mode.

pub fn clone_padding_inline_start(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for padding-inline-start given a writing mode.

pub fn set_padding_inline_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for padding-inline-end given a writing mode.

pub fn copy_padding_inline_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-inline-end given a writing mode.

pub fn reset_padding_inline_end(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for padding-inline-end given a writing mode.

pub fn clone_padding_inline_end(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for padding-inline-end given a writing mode.