Struct style::properties::style_structs::Position

pub struct Position {
    pub top: T,
    pub left: T,
    pub bottom: T,
    pub right: T,
    pub offset_block_start: T,
    pub offset_block_end: T,
    pub offset_inline_start: T,
    pub offset_inline_end: T,
    pub z_index: T,
    pub flex_direction: T,
    pub flex_wrap: T,
    pub justify_content: T,
    pub align_content: T,
    pub align_items: T,
    pub flex_grow: T,
    pub flex_shrink: T,
    pub align_self: T,
    pub order: T,
    pub flex_basis: T,
    pub width: T,
    pub min_width: T,
    pub max_width: T,
    pub height: T,
    pub min_height: T,
    pub max_height: T,
    pub block_size: T,
    pub min_block_size: T,
    pub max_block_size: T,
    pub inline_size: T,
    pub min_inline_size: T,
    pub max_inline_size: T,
    pub box_sizing: T,
}

The Position style struct.

Fields
top: T

The top computed value.

left: T

The left computed value.

bottom: T

The bottom computed value.

right: T

The right computed value.

offset_block_start: T

The offset-block-start computed value.

offset_block_end: T

The offset-block-end computed value.

offset_inline_start: T

The offset-inline-start computed value.

offset_inline_end: T

The offset-inline-end computed value.

z_index: T

The z-index computed value.

flex_direction: T

The flex-direction computed value.

flex_wrap: T

The flex-wrap computed value.

justify_content: T

The justify-content computed value.

align_content: T

The align-content computed value.

align_items: T

The align-items computed value.

flex_grow: T

The flex-grow computed value.

flex_shrink: T

The flex-shrink computed value.

align_self: T

The align-self computed value.

order: T

The order computed value.

flex_basis: T

The flex-basis computed value.

width: T

The width computed value.

min_width: T

The min-width computed value.

max_width: T

The max-width computed value.

height: T

The height computed value.

min_height: T

The min-height computed value.

max_height: T

The max-height computed value.

block_size: T

The block-size computed value.

min_block_size: T

The min-block-size computed value.

max_block_size: T

The max-block-size computed value.

inline_size: T

The inline-size computed value.

min_inline_size: T

The min-inline-size computed value.

max_inline_size: T

The max-inline-size computed value.

box_sizing: T

The box-sizing computed value.

Methods

impl Position

pub fn set_top(&mut self, v: T)

Set top.

pub fn copy_top_from(&mut self, other: &Self)

Set top from other struct.

pub fn reset_top(&mut self, other: &Self)

Reset top from the initial struct.

pub fn clone_top(&self) -> T

Get the computed value for top.

pub fn set_left(&mut self, v: T)

Set left.

pub fn copy_left_from(&mut self, other: &Self)

Set left from other struct.

pub fn reset_left(&mut self, other: &Self)

Reset left from the initial struct.

pub fn clone_left(&self) -> T

Get the computed value for left.

pub fn set_bottom(&mut self, v: T)

Set bottom.

pub fn copy_bottom_from(&mut self, other: &Self)

Set bottom from other struct.

pub fn reset_bottom(&mut self, other: &Self)

Reset bottom from the initial struct.

pub fn clone_bottom(&self) -> T

Get the computed value for bottom.

pub fn set_right(&mut self, v: T)

Set right.

pub fn copy_right_from(&mut self, other: &Self)

Set right from other struct.

pub fn reset_right(&mut self, other: &Self)

Reset right from the initial struct.

pub fn clone_right(&self) -> T

Get the computed value for right.

pub fn set_offset_block_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-block-start given a writing mode.

pub fn copy_offset_block_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-block-start given a writing mode.

pub fn reset_offset_block_start(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-block-start given a writing mode.

pub fn clone_offset_block_start(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for offset-block-start given a writing mode.

pub fn set_offset_block_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-block-end given a writing mode.

pub fn copy_offset_block_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-block-end given a writing mode.

pub fn reset_offset_block_end(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-block-end given a writing mode.

pub fn clone_offset_block_end(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for offset-block-end given a writing mode.

pub fn set_offset_inline_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-inline-start given a writing mode.

pub fn copy_offset_inline_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-inline-start given a writing mode.

pub fn reset_offset_inline_start(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-inline-start given a writing mode.

pub fn clone_offset_inline_start(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for offset-inline-start given a writing mode.

pub fn set_offset_inline_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for offset-inline-end given a writing mode.

pub fn copy_offset_inline_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-inline-end given a writing mode.

pub fn reset_offset_inline_end(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for offset-inline-end given a writing mode.

pub fn clone_offset_inline_end(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for offset-inline-end given a writing mode.

pub fn set_z_index(&mut self, v: T)

Set z-index.

pub fn copy_z_index_from(&mut self, other: &Self)

Set z-index from other struct.

pub fn reset_z_index(&mut self, other: &Self)

Reset z-index from the initial struct.

pub fn clone_z_index(&self) -> T

Get the computed value for z-index.

pub fn set_flex_direction(&mut self, v: T)

Set flex-direction.

pub fn copy_flex_direction_from(&mut self, other: &Self)

Set flex-direction from other struct.

pub fn reset_flex_direction(&mut self, other: &Self)

Reset flex-direction from the initial struct.

pub fn clone_flex_direction(&self) -> T

Get the computed value for flex-direction.

pub fn set_flex_wrap(&mut self, v: T)

Set flex-wrap.

pub fn copy_flex_wrap_from(&mut self, other: &Self)

Set flex-wrap from other struct.

pub fn reset_flex_wrap(&mut self, other: &Self)

Reset flex-wrap from the initial struct.

pub fn clone_flex_wrap(&self) -> T

Get the computed value for flex-wrap.

pub fn set_justify_content(&mut self, v: T)

Set justify-content.

pub fn copy_justify_content_from(&mut self, other: &Self)

Set justify-content from other struct.

pub fn reset_justify_content(&mut self, other: &Self)

Reset justify-content from the initial struct.

pub fn clone_justify_content(&self) -> T

Get the computed value for justify-content.

pub fn set_align_content(&mut self, v: T)

Set align-content.

pub fn copy_align_content_from(&mut self, other: &Self)

Set align-content from other struct.

pub fn reset_align_content(&mut self, other: &Self)

Reset align-content from the initial struct.

pub fn clone_align_content(&self) -> T

Get the computed value for align-content.

pub fn set_align_items(&mut self, v: T)

Set align-items.

pub fn copy_align_items_from(&mut self, other: &Self)

Set align-items from other struct.

pub fn reset_align_items(&mut self, other: &Self)

Reset align-items from the initial struct.

pub fn clone_align_items(&self) -> T

Get the computed value for align-items.

pub fn set_flex_grow(&mut self, v: T)

Set flex-grow.

pub fn copy_flex_grow_from(&mut self, other: &Self)

Set flex-grow from other struct.

pub fn reset_flex_grow(&mut self, other: &Self)

Reset flex-grow from the initial struct.

pub fn clone_flex_grow(&self) -> T

Get the computed value for flex-grow.

pub fn set_flex_shrink(&mut self, v: T)

Set flex-shrink.

pub fn copy_flex_shrink_from(&mut self, other: &Self)

Set flex-shrink from other struct.

pub fn reset_flex_shrink(&mut self, other: &Self)

Reset flex-shrink from the initial struct.

pub fn clone_flex_shrink(&self) -> T

Get the computed value for flex-shrink.

pub fn set_align_self(&mut self, v: T)

Set align-self.

pub fn copy_align_self_from(&mut self, other: &Self)

Set align-self from other struct.

pub fn reset_align_self(&mut self, other: &Self)

Reset align-self from the initial struct.

pub fn clone_align_self(&self) -> T

Get the computed value for align-self.

pub fn set_order(&mut self, v: T)

Set order.

pub fn copy_order_from(&mut self, other: &Self)

Set order from other struct.

pub fn reset_order(&mut self, other: &Self)

Reset order from the initial struct.

pub fn clone_order(&self) -> T

Get the computed value for order.

pub fn set_flex_basis(&mut self, v: T)

Set flex-basis.

pub fn copy_flex_basis_from(&mut self, other: &Self)

Set flex-basis from other struct.

pub fn reset_flex_basis(&mut self, other: &Self)

Reset flex-basis from the initial struct.

pub fn clone_flex_basis(&self) -> T

Get the computed value for flex-basis.

pub fn set_width(&mut self, v: T)

Set width.

pub fn copy_width_from(&mut self, other: &Self)

Set width from other struct.

pub fn reset_width(&mut self, other: &Self)

Reset width from the initial struct.

pub fn clone_width(&self) -> T

Get the computed value for width.

pub fn set_min_width(&mut self, v: T)

Set min-width.

pub fn copy_min_width_from(&mut self, other: &Self)

Set min-width from other struct.

pub fn reset_min_width(&mut self, other: &Self)

Reset min-width from the initial struct.

pub fn clone_min_width(&self) -> T

Get the computed value for min-width.

pub fn set_max_width(&mut self, v: T)

Set max-width.

pub fn copy_max_width_from(&mut self, other: &Self)

Set max-width from other struct.

pub fn reset_max_width(&mut self, other: &Self)

Reset max-width from the initial struct.

pub fn clone_max_width(&self) -> T

Get the computed value for max-width.

pub fn set_height(&mut self, v: T)

Set height.

pub fn copy_height_from(&mut self, other: &Self)

Set height from other struct.

pub fn reset_height(&mut self, other: &Self)

Reset height from the initial struct.

pub fn clone_height(&self) -> T

Get the computed value for height.

pub fn set_min_height(&mut self, v: T)

Set min-height.

pub fn copy_min_height_from(&mut self, other: &Self)

Set min-height from other struct.

pub fn reset_min_height(&mut self, other: &Self)

Reset min-height from the initial struct.

pub fn clone_min_height(&self) -> T

Get the computed value for min-height.

pub fn set_max_height(&mut self, v: T)

Set max-height.

pub fn copy_max_height_from(&mut self, other: &Self)

Set max-height from other struct.

pub fn reset_max_height(&mut self, other: &Self)

Reset max-height from the initial struct.

pub fn clone_max_height(&self) -> T

Get the computed value for max-height.

pub fn set_block_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for block-size given a writing mode.

pub fn copy_block_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for block-size given a writing mode.

pub fn reset_block_size(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for block-size given a writing mode.

pub fn clone_block_size(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for block-size given a writing mode.

pub fn set_min_block_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for min-block-size given a writing mode.

pub fn copy_min_block_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for min-block-size given a writing mode.

pub fn reset_min_block_size(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for min-block-size given a writing mode.

pub fn clone_min_block_size(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for min-block-size given a writing mode.

pub fn set_max_block_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for max-block-size given a writing mode.

pub fn copy_max_block_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for max-block-size given a writing mode.

pub fn reset_max_block_size(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for max-block-size given a writing mode.

pub fn clone_max_block_size(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for max-block-size given a writing mode.

pub fn set_inline_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for inline-size given a writing mode.

pub fn copy_inline_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for inline-size given a writing mode.

pub fn reset_inline_size(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for inline-size given a writing mode.

pub fn clone_inline_size(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for inline-size given a writing mode.

pub fn set_min_inline_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for min-inline-size given a writing mode.

pub fn copy_min_inline_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for min-inline-size given a writing mode.

pub fn reset_min_inline_size(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for min-inline-size given a writing mode.

pub fn clone_min_inline_size(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for min-inline-size given a writing mode.

pub fn set_max_inline_size(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for max-inline-size given a writing mode.

pub fn copy_max_inline_size_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for max-inline-size given a writing mode.

pub fn reset_max_inline_size(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for max-inline-size given a writing mode.

pub fn clone_max_inline_size(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for max-inline-size given a writing mode.

pub fn set_box_sizing(&mut self, v: T)

Set box-sizing.

pub fn copy_box_sizing_from(&mut self, other: &Self)

Set box-sizing from other struct.

pub fn reset_box_sizing(&mut self, other: &Self)

Reset box-sizing from the initial struct.

pub fn clone_box_sizing(&self) -> T

Get the computed value for box-sizing.