Type Definition style::properties::FontComputationData

type FontComputationData = Option<(KeywordSize, f32, NonNegativeLength)>;

This is where we store extra font data while while computing font sizes.

font-size keyword values (and font-size-relative values applied to keyword values) need to preserve their identity as originating from keywords and relative font sizes. We store this information out of band in the ComputedValues. When None, the font size on the current struct was computed from a value that was not a keyword or a chain of font-size-relative values applying to successive parents terminated by a keyword. When Some, this means the font-size was derived from a keyword value or a keyword value on some ancestor with only font-size-relative keywords and regular inheritance in between. The integer stores the final ratio of the chain of font size relative values. and is 1 when there was just a keyword and no relative values.

When this is Some, we compute font sizes by computing the keyword against the generic font, and then multiplying it by the ratio (as well as adding any absolute offset from calcs)