style::properties

Type Alias LonghandIdSet

pub type LonghandIdSet = IdSet<LonghandId, { _ }>;

A set of longhand properties.

Aliased Type

pub struct LonghandIdSet {
    storage: [u32; 10],
    _phantom: PhantomData<LonghandId>,
}

Fields
storage: [u32; 10]_phantom: PhantomData<LonghandId>
Implementations

impl LonghandIdSet

pub(super) fn reset() -> &'static Self

The set of non-inherited longhands.

pub(super) fn discrete_animatable() -> &'static Self

pub(super) fn logical() -> &'static Self

pub(super) fn ignored_when_colors_disabled() -> &'static Self

Returns the set of longhands that are ignored when document colors are disabled.

pub(super) fn visited_dependent() -> &'static Self

Only a few properties are allowed to depend on the visited state of links. When cascading visited styles, we can save time by only processing these properties.

pub(super) fn prioritary_properties() -> &'static Self

pub(super) fn late_group_only_inherited() -> &'static Self

pub(super) fn late_group() -> &'static Self

pub fn margin_properties() -> &'static Self

Returns the set of margin properties, for the purposes of
use counters / warnings.

pub fn border_background_properties() -> &'static Self

Returns the set of border properties for the purpose of disabling native appearance.

pub fn zoom_dependent() -> &'static Self

Returns properties that are zoom dependent (basically, that contain lengths).

pub fn zoom_dependent_inherited_properties() -> &'static Self

Note that it’s different from zoom_dependent(), as this only includes inherited, physical properties.

impl LonghandIdSet

pub fn contains_any_reset(&self) -> bool

Return whether this set contains any reset longhand.