style::properties_and_values::rule

Type Alias InitialValue

pub type InitialValue = Arc<SpecifiedValue>;

Specifies the initial value of the custom property registration represented by the @property rule, controlling the property’s initial value.

The SpecifiedValue is wrapped in an Arc to avoid copying when using it.

Aliased Type

#[repr(C)]
pub struct InitialValue { /* private fields */ }