style::properties_and_values

Module value

Parsing for registered custom properties.

Structs

ComponentList

A list of component values, including the list’s multiplier.

CustomAnimatedValue

An animated value for custom property.

Parser 🔒

Value

A struct for a single specified registered custom property value that includes its original URL data so the value can be uncomputed later.

Enums

AllowComputationallyDependent

Whether the computed value parsing should allow computationaly dependent values like 3em or var(-foo).

GenericValueComponent

A generic enum used for both specified value components and computed value components.

ValueInner

A specified registered custom property value.

Type Aliases

ComputedValue

Computed custom property value.

ComputedValueComponent

A single component of the computed value.

SmallComponentVec 🔒

SpecifiedValue

Specified custom property value.

SpecifiedValueComponent

A single component of the specified value.