Struct style::properties_and_values::value::Value

pub struct Value<Component> {
    pub(crate) v: ValueInner<Component>,
    url_data: UrlExtraData,
}

A struct for a single specified registered custom property value that includes its original URL

Fields
v: ValueInner<Component>

The registered custom property value.

url_data: UrlExtraData

The URL data of the registered custom property from before it was computed. This is necessary to uncompute registered custom properties.

Implementations

impl<Component> Value<Component>

pub fn new(v: ValueInner<Component>, url_data: UrlExtraData) -> Self

Creates a new registered custom property value.

pub fn universal(var: Arc<ComputedPropertyValue>) -> Self

Creates a new registered custom property value presumed to have universal syntax.

impl Value<GenericValueComponent<Length, Number, Percentage, LengthPercentage, Color, GenericImage<GenericGradient<LineDirection, LengthPercentage, NonNegative<Length>, NonNegative<LengthPercentage>, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, Angle, AngleOrPercentage, Color>, CssUrl, Color, Percentage, Resolution>, CssUrl, Integer, Angle, Time, Resolution, GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>>>

pub fn compute<'i, 't>( input: &mut CSSParser<'i, 't>, registration: &PropertyRegistrationData, url_data: &UrlExtraData, context: &Context<'_>, allow_computationally_dependent: AllowComputationallyDependent, ) -> Result<ComputedValue, ()>

Convert a registered custom property to a Computed custom property value, given input and a property registration.

pub fn parse<'i, 't>( input: &mut CSSParser<'i, 't>, syntax: &Descriptor, url_data: &UrlExtraData, allow_computationally_dependent: AllowComputationallyDependent, ) -> Result<Self, StyleParseError<'i>>

Parse and validate a registered custom property value according to its syntax descriptor, and check for computational independence.

impl Value<GenericValueComponent<CSSPixelLength, f32, Percentage, LengthPercentage, GenericColor<Percentage>, GenericImage<GenericGradient<LineDirection, LengthPercentage, NonNegative<CSSPixelLength>, NonNegative<LengthPercentage>, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>, ComputedUrl, i32, Angle, Time, Resolution, GenericTransform<GenericTransformOperation<Angle, f32, CSSPixelLength, i32, LengthPercentage>>>>

fn serialization_types( &self, ) -> (TokenSerializationType, TokenSerializationType)

fn to_declared_value(&self) -> CustomDeclarationValue

pub fn as_universal(&self) -> Option<&Arc<ComputedPropertyValue>>

Returns the contained variable value if it exists, otherwise None.

pub fn is_parsed(&self, registration: &PropertyRegistrationData) -> bool

Returns whether the the property is computed.

pub fn to_variable_value(&self) -> ComputedPropertyValue

Convert to an untyped variable value.