style::properties_and_values::value

Type Alias ComputedValue

pub type ComputedValue = Value<ComputedValueComponent>;

Computed custom property value.

Aliased Type

pub struct ComputedValue {
    pub(crate) v: ValueInner<GenericValueComponent<CSSPixelLength, f32, Percentage, LengthPercentage, GenericColor<Percentage>, GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>, ComputedUrl, i32, Angle, Time, Resolution, GenericTransform<GenericTransformOperation<Angle, f32, CSSPixelLength, i32, LengthPercentage>>>>,
    url_data: UrlExtraData,
    pub attr_tainted: bool,
}

Fields
v: ValueInner<GenericValueComponent<CSSPixelLength, f32, Percentage, LengthPercentage, GenericColor<Percentage>, GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>, ComputedUrl, i32, Angle, Time, Resolution, GenericTransform<GenericTransformOperation<Angle, f32, CSSPixelLength, i32, LengthPercentage>>>>

The registered custom property value.

url_data: UrlExtraData

The URL data of the registered custom property from before it was computed. This is necessary to uncompute registered custom properties.

attr_tainted: bool

Flag indicating whether this value is tainted by an attr().

Implementations

impl ComputedValue

pub fn to_declared_value(&self) -> CustomDeclarationValue

Uncomputes the value so that it can go back into the cascade.

pub fn as_universal(&self) -> Option<&Arc<ComputedPropertyValue>>

Returns the contained variable value if it exists, otherwise None.