style::properties_and_values::value

Type Alias SpecifiedValue

pub type SpecifiedValue = Value<SpecifiedValueComponent>;

Specified custom property value.

Aliased Type

pub struct SpecifiedValue {
    pub(crate) v: ValueInner<GenericValueComponent<Length, Number, Percentage, LengthPercentage, Color, GenericImage<GenericGradient<LineDirection, Length, LengthPercentage, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, Angle, AngleOrPercentage, Color>, CssUrl, Color, Percentage, Resolution>, CssUrl, Integer, Angle, Time, Resolution, GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>>>,
    url_data: UrlExtraData,
    pub attr_tainted: bool,
}

Fields
v: ValueInner<GenericValueComponent<Length, Number, Percentage, LengthPercentage, Color, GenericImage<GenericGradient<LineDirection, Length, LengthPercentage, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, Angle, AngleOrPercentage, Color>, CssUrl, Color, Percentage, Resolution>, CssUrl, Integer, Angle, Time, Resolution, GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>>>

The registered custom property value.

url_data: UrlExtraData

The URL data of the registered custom property from before it was computed. This is necessary to uncompute registered custom properties.

attr_tainted: bool

Flag indicating whether this value is tainted by an attr().

Implementations

impl SpecifiedValue

pub fn compute<'i, 't>( input: &mut CSSParser<'i, 't>, registration: &PropertyDescriptors, namespaces: Option<&FxHashMap<Prefix, Namespace>>, url_data: &UrlExtraData, context: &Context<'_>, allow_computationally_dependent: AllowComputationallyDependent, attr_taint: AttrTaint, ) -> Result<ComputedValue, ()>

Convert a registered custom property to a Computed custom property value, given input and a property registration.

pub fn parse<'i, 't>( input: &mut CSSParser<'i, 't>, syntax: &Descriptor, url_data: &UrlExtraData, namespaces: Option<&FxHashMap<Prefix, Namespace>>, allow_computationally_dependent: AllowComputationallyDependent, attr_taint: AttrTaint, ) -> Result<Self, StyleParseError<'i>>

Parse and validate a registered custom property value according to its syntax descriptor, and check for computational independence.