style::queries

Module condition

A query condition:

https://drafts.csswg.org/mediaqueries-4/#typedef-media-condition https://drafts.csswg.org/css-contain-3/#typedef-container-condition

Structs

MozPrefFeature

A custom -moz-pref(, ) query feature.

StyleFeaturePlain

A style feature consisting of a custom property name and (optionally) value.

Enums

AllowOr 🔒

Whether to allow an or condition or not during parsing.

BoolValue

A boolean value for a pref query.

MozPrefFeatureValue

Simple values we support for -moz-pref(). We don’t want to deal with calc() and other shenanigans for now.

Operator

A binary and or or operator.

QueryCondition

Represents a condition.

StyleFeature

A style query feature: https://drafts.csswg.org/css-conditional-5/#typedef-style-feature

StyleFeatureValue 🔒

StyleQuery

https://drafts.csswg.org/css-conditional-5/#typedef-style-query

Traits

OperationParser 🔒

Trait for query elements that parse a series of conditions separated by AND or OR operators, or prefixed with NOT.

Functions

consume_any_value 🔒

https://drafts.csswg.org/css-syntax-3/#typedef-any-value

Type Aliases

ComputedMozPrefFeatureValue

The computed -moz-pref() value.

SpecifiedMozPrefFeatureValue 🔒