Module style::refcell

A fork of std::cell::RefCell that makes as_unsafe_cell usable on stable Rust.

FIXME(https://github.com/rust-lang/rust/issues/27708): Remove this (revert commit f7f81e0ed0b62402db291e28a9bb16f7194ebf78 / PR #11835) when std::cell::RefCell::as_unsafe_cell is in Rust’s stable channel.

Structs

BorrowRef

BorrowRefMut

Ref
Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell
A fork of std::cell::RefCell that makes as_unsafe_cell usable on stable Rust.

RefMut
A wrapper type for a mutably borrowed value from a RefCell<T>.

Enums

BorrowState
An enumeration of values returned from the state method on a RefCell<T>.

Constants

UNUSED

WRITING

Type Definitions

BorrowFlag