Module style::restyle_hints

Structs

Dependency

DependencySet

ElementSnapshot
In order to compute restyle hints, we perform a selector match against a list of partial selectors whose rightmost simple selector may be sensitive to the thing being changed. We do this matching twice, once for the element as it exists now and once for the element as it existed at the time of the last restyle. If the results of the selector match differ, that means that the given partial selector is sensitive to the change, and we compute a restyle hint based on its combinator.

ElementWrapper

RestyleHint

Sensitivities

Constants

RESTYLE_DESCENDANTS
Rerun selector matching on all of the element's descendants.

RESTYLE_LATER_SIBLINGS
Rerun selector matching on all later siblings of the element and all of their descendants.

RESTYLE_SELF
Rerun selector matching on the element.

Statics

EMPTY_SNAPSHOT

Functions

combinator_to_restyle_hint

is_attr_selector

selector_to_state