Struct style::restyle_hints::Dependency

struct Dependency {
    selector: Arc<ComplexSelector<TheSelectorImpl>>,
    hint: RestyleHint,
    sensitivities: Sensitivities,
}

Mapping between (partial) CompoundSelectors (and the combinator to their right) and the states and attributes they depend on.

In general, for all selectors in all applicable stylesheets of the form:

|a _ b _ c _ d _ e|

Where: * |b| and |d| are simple selectors that depend on state (like :hover) or attributes (like [attr...], .foo, or #foo). * |a|, |c|, and |e| are arbitrary simple selectors that do not depend on state or attributes.

We generate a Dependency for both |a _ b:X _| and |a _ b:X _ c _ d:Y _|, even though those selectors may not appear on their own in any stylesheet. This allows us to quickly scan through the dependency sites of all style rules and determine the maximum effect that a given state or attribute change may have on the style of elements in the document.

Fields
selector: Arc<ComplexSelector<TheSelectorImpl>> hint: RestyleHint sensitivities: Sensitivities