style

Module rule_tree

The rule tree.

Re-exports

pub use self::level::CascadeLevel;

pub use self::level::CascadeOrigin;

pub use self::level::ShadowCascadeOrder;

Modules

core 🔒

level

The cascade level and shadow cascade order for tracking shadow tree rules.

map 🔒

source 🔒

unsafe_box 🔒

Structs

RuleCascadeFlags

Flags that are part of the cascade priority, and that we use to track information about where the rule came from.

RuleTree

The rule tree, the structure servo uses to preserve the results of selector matching.

SelfAndAncestors

An iterator over a rule node and its ancestors.

StrongRuleNode

A strong reference to a rule node.

StyleSource

A style source for the rule node. It is a declaration block that may come from either a style rule or a standalone block like animations / transitions / smil / preshints / style attr…