Enum style::rule_tree::CascadeLevel

#[repr(u8)]
pub enum CascadeLevel {
    UANormal,
    UserNormal,
    PresHints,
    AuthorNormal {
        shadow_cascade_order: ShadowCascadeOrder,
    },
    SMILOverride,
    Animations,
    AuthorImportant {
        shadow_cascade_order: ShadowCascadeOrder,
    },
    UserImportant,
    UAImportant,
    Transitions,
}

The cascade level these rules are relevant at, as per13.

Presentational hints for SVG and HTML are in the “author-level zero-specificity” level, that is, right after user rules, and before author rules.

The order of variants declared here is significant, and must be in ascending order of precedence.

See also 4 for the Shadow DOM bits. We rely on the invariant that rules from outside the tree the element is in can’t affect the element.

The opposite is not true (i.e., :host and ::slotted) from an “inner” shadow tree may affect an element connected to the document or an “outer” shadow tree.

Variants

UANormal

Normal User-Agent rules.

UserNormal

User normal rules.

PresHints

Presentational hints.

AuthorNormal

Fields

shadow_cascade_order: ShadowCascadeOrder

The order in the shadow tree hierarchy. This number is relative to the tree of the element, and thus the only invariants that need to be preserved is:

Zero is the same tree as the element that matched the rule. This is important so that we can optimize style attribute insertions.

The levels are ordered in accordance with https://drafts.csswg.org/css-scoping/#shadow-cascading

Shadow DOM styles from author styles.

SMILOverride

SVG SMIL animations.

Animations

CSS animations and script-generated animations.

AuthorImportant

Fields

shadow_cascade_order: ShadowCascadeOrder

The order in the shadow tree hierarchy, inverted, so that PartialOrd does the right thing.

Author-supplied important rules.

UserImportant

User important rules.

UAImportant

User-agent important rules.

Transitions

Transitions

Implementations

impl CascadeLevel

pub fn to_byte_lossy(&self) -> u8

Pack this cascade level in a single byte.

We have 10 levels, which we can represent with 4 bits, and then a cascade order optionally, which we can clamp to three bits max, and represent with a fourth bit for the sign.

So this creates: SOOODDDD

Where S is the sign of the order (one if negative, 0 otherwise), O is the absolute value of the order, and Ds are the discriminant.

pub fn from_byte(b: u8) -> Self

Convert back from the single-byte representation of the cascade level explained above.

pub fn guard<'a>(
&self,
guards: &'a StylesheetGuards<'a>
) -> &'a SharedRwLockReadGuard<'a>

Select a lock guard for this level

pub fn same_tree_author_important() -> Self

Returns the cascade level for author important declarations from the same tree as the element.

pub fn same_tree_author_normal() -> Self

Returns the cascade level for author normal declarations from the same tree as the element.

pub fn is_important(&self) -> bool

Returns whether this cascade level represents important rules of some sort.

pub fn importance(&self) -> Importance

Returns the importance relevant for this rule. Pretty similar to is_important.

pub fn origin(&self) -> Origin

Returns the cascade origin of the rule.

pub fn is_animation(&self) -> bool

Returns whether this cascade level represents an animation rules.