style::rule_tree::level

Enum CascadeOrigin

#[repr(u8)]
pub enum CascadeOrigin {
    UA = 0,
    User = 1,
    PresHints = 2,
    Author = 3,
    PositionFallback = 4,
    SMILOverride = 5,
    Animations = 6,
    Transitions = 7,
}

The cascade origin of rules.

Presentational hints for SVG and HTML are in the “author-level zero-specificity” level, that is, right after user rules, and before author rules.

The order of variants declared here is significant, and must be in ascending order of precedence.

Variants

UA = 0

Normal User-Agent rules.

User = 1

User normal rules.

PresHints = 2

Presentational hints.

Author = 3

Styles from author styles.

PositionFallback = 4

https://drafts.csswg.org/css-anchor-position-1/#position-fallback-origin

SMILOverride = 5

SVG SMIL animations.

Animations = 6

CSS animations and script-generated animations.

Transitions = 7

CSS Transitions

Implementations

impl CascadeOrigin

pub fn origin(self) -> Origin

Returns the “simplified” origin.

pub fn is_author_origin(self) -> bool

Returns whether this is an “author” origin (in the “simplified” sense of the word).

pub fn guard<'a>( &self, guards: &'a StylesheetGuards<'a>, ) -> &'a SharedRwLockReadGuard<'a>

Select a lock guard for this origin.