Struct style::rule_tree::RuleNode

struct RuleNode {
    root: Option<WeakRuleNode>,
    parent: Option<StrongRuleNode>,
    source: Option<StyleSource>,
    importance: Importance,
    refcount: AtomicUsize,
    first_child: AtomicPtr<RuleNode>,
    next_sibling: AtomicPtr<RuleNode>,
    prev_sibling: AtomicPtr<RuleNode>,
    next_free: AtomicPtr<RuleNode>,
    free_count: AtomicUsize,
}

Fields
root: Option<WeakRuleNode>

The root node. Only the root has no root pointer, for obvious reasons.

parent: Option<StrongRuleNode>

The parent rule node. Only the root has no parent.

source: Option<StyleSource>

The actual style source, either coming from a selector in a StyleRule, or a raw property declaration block (like the style attribute).

importance: Importance

The importance of the declarations relevant in the style rule, meaningless in the root node.

refcount: AtomicUsize first_child: AtomicPtr<RuleNode> next_sibling: AtomicPtr<RuleNode> prev_sibling: AtomicPtr<RuleNode> next_free: AtomicPtr<RuleNode>

The next item in the rule tree free list, that starts on the root node.

free_count: AtomicUsize

Number of RuleNodes we have added to the free list since the last GC. (We don't update this if we rescue a RuleNode from the free list. It's just used as a heuristic to decide when to run GC.)

Only used on the root RuleNode. (We could probably re-use one of the sibling pointers to save space.)

Methods

impl RuleNode

fn new(root: WeakRuleNode,
parent: StrongRuleNode,
source: StyleSource,
importance: Importance)
-> Self

fn root() -> Self

fn is_root(&self) -> bool

unsafe fn remove_from_child_list(&self)

Remove this rule node from the child list.

This method doesn't use proper synchronization, and it's expected to be called in a single-threaded fashion, thus the unsafety.

This is expected to be called before freeing the node from the free list.

fn dump<W: Write>(&self, writer: &mut W, indent: usize)

fn iter_children(&self) -> RuleChildrenListIter