Struct style::rule_tree::StrongRuleNode

pub struct StrongRuleNode {
    p: NonNull<RuleNode>,
}

A strong reference to a rule node.

Fields
p: NonNull<RuleNode>
Methods

impl StrongRuleNode

fn new(n: Box<RuleNode>) -> Self

fn from_ptr(p: NonNull<RuleNode>) -> Self

fn downgrade(&self) -> WeakRuleNode

pub fn parent(&self) -> Option<&StrongRuleNode>

Get the parent rule node of this rule node.

fn ensure_child(
&self,
root: WeakRuleNode,
source: StyleSource,
level: CascadeLevel
) -> StrongRuleNode

pub fn ptr(&self) -> *mut RuleNode

Raw pointer to the RuleNode

fn get(&self) -> &RuleNode

pub fn style_source(&self) -> Option<&StyleSource>

Get the style source corresponding to this rule node. May return None if it's the root node, which means that the node hasn't matched any rules.

pub fn cascade_level(&self) -> CascadeLevel

The cascade level for this node

pub fn importance(&self) -> Importance

Get the importance that this rule node represents.

Important traits for SelfAndAncestors<'a>
impl<'a> Iterator for SelfAndAncestors<'a> type Item = &'a StrongRuleNode;

pub fn self_and_ancestors(&self) -> SelfAndAncestors

Get an iterator for this rule node and its ancestors.

pub unsafe fn has_children_for_testing(&self) -> bool

Returns whether this node has any child, only intended for testing purposes, and called on a single-threaded fashion only.

unsafe fn pop_from_free_list(&self) -> Option<WeakRuleNode>

unsafe fn assert_free_list_has_no_duplicates_or_null(&self)

unsafe fn gc(&self)

unsafe fn maybe_gc(&self)

pub fn has_animation_or_transition_rules(&self) -> bool

Returns true if there is either animation or transition level rule.

pub fn get_properties_overriding_animations(
&self,
guards: &StylesheetGuards
) -> (LonghandIdSet, bool)

Get a set of properties whose CascadeLevel are higher than Animations but not equal to Transitions.

If there are any custom properties, we set the boolean value of the returned tuple to true.