Struct style::scoped_tls::ScopedTLS

pub struct ScopedTLS<'scope, T: Send> {
    pool: &'scope ThreadPool,
    slots: Box<[RefCell<Option<T>>]>,
}

A scoped TLS set, that is alive during the 'scope lifetime.

We use this on Servo to construct thread-local contexts, but clear them once we're done with restyling.

Fields
pool: &'scope ThreadPool slots: Box<[RefCell<Option<T>>]>
Methods

impl<'scope, T: Send> ScopedTLS<'scope, T>

fn new(p: &'scope ThreadPool) -> Self

Create a new scoped TLS that will last as long as this rayon threadpool reference.

fn borrow(&self) -> Ref<Option<T>>

Return an immutable reference to the Option<T> that this thread owns.

fn borrow_mut(&self) -> RefMut<Option<T>>

Return a mutable reference to the Option<T> that this thread owns.

fn ensure<F: FnOnce() -> T>(&self, f: F) -> RefMut<T>

Ensure that the current data this thread owns is initialized, or initialize it using f.

unsafe fn unsafe_get(&self) -> &[RefCell<Option<T>>]

Unsafe access to the slots. This can be used to access the TLS when the caller knows that the pool does not have access to the TLS.