Enum style::selector_impl::PseudoElementCascadeType

pub enum PseudoElementCascadeType {
    Eager,
    Lazy,
    Precomputed,
}

This function determines if a pseudo-element is eagerly cascaded or not.

Eagerly cascaded pseudo-elements are "normal" pseudo-elements (i.e. ::before and ::after). They inherit styles normally as another selector would do, and they're part of the cascade.

Lazy pseudo-elements are affected by selector matching, but they're only computed when needed, and not before. They're useful for general pseudo-elements that are not very common.

Note that in Servo lazy pseudo-elements are restricted to a subset of selectors, so you can't use it for public pseudo-elements. This is not the case with Gecko though.

Precomputed ones skip the cascade process entirely, mostly as an optimisation since they are private pseudo-elements (like ::-servo-details-content).

This pseudo-elements are resolved on the fly using only global rules (rules of the form *|*), and applying them to the parent style.

If you're implementing a public selector that the end-user might customize, then you probably need to make it eager.

Variants
EagerLazyPrecomputed
Methods

impl PseudoElementCascadeType

fn is_eager(&self) -> bool

fn is_lazy(&self) -> bool

fn is_precomputed(&self) -> bool