Struct style::selector_matching::PerOriginSelectorMap

struct PerOriginSelectorMap<Impl: SelectorImpl> {
    normal: SelectorMap<Vec<PropertyDeclaration>, Impl>,
    important: SelectorMap<Vec<PropertyDeclaration>, Impl>,
}

Map that contains the CSS rules for a given origin.

Fields

normal

Rules that contains at least one property declararion with normal importance.

important

Rules that contains at least one property declararion with !important.

Methods

impl<Impl: SelectorImpl> PerOriginSelectorMap<Impl>

fn new() -> PerOriginSelectorMap<Impl>