Struct style::selector_matching::Stylist

pub struct Stylist {
    pub device: Device,
    viewport_constraints: Option<ViewportConstraints>,
    quirks_mode: bool,
    is_device_dirty: bool,
    element_map: PerPseudoElementSelectorMap,
    pseudos_map: FnvHashMap<PseudoElement, PerPseudoElementSelectorMap>,
    animations: FnvHashMap<Atom, KeyframesAnimation>,
    precomputed_pseudo_element_decls: FnvHashMap<PseudoElement, Vec<ApplicableDeclarationBlock>>,
    rules_source_order: usize,
    state_deps: DependencySet,
    sibling_affecting_selectors: Vec<Selector<TheSelectorImpl>>,
    non_common_style_affecting_attributes_selectors: Vec<Selector<TheSelectorImpl>>,
}

This structure holds all the selectors and device characteristics for a given document. The selectors are converted into Rules (defined in rust-selectors), and introduced in a SelectorMap depending on the pseudo-element (see PerPseudoElementSelectorMap), and stylesheet origin (see the fields of PerPseudoElementSelectorMap).

This structure is effectively created once per pipeline, in the LayoutThread corresponding to that pipeline.

Fields
device: Device

Device that the stylist is currently evaluating against.

viewport_constraints: Option<ViewportConstraints>

Viewport constraints based on the current device.

quirks_mode: bool

If true, the quirks-mode stylesheet is applied.

is_device_dirty: bool

If true, the device has changed, and the stylist needs to be updated.

element_map: PerPseudoElementSelectorMap

The current selector maps, after evaluating media rules against the current device.

pseudos_map: FnvHashMap<PseudoElement, PerPseudoElementSelectorMap>

The selector maps corresponding to a given pseudo-element (depending on the implementation)

animations: FnvHashMap<Atom, KeyframesAnimation>

A map with all the animations indexed by name.

precomputed_pseudo_element_decls: FnvHashMap<PseudoElement, Vec<ApplicableDeclarationBlock>>

Applicable declarations for a given non-eagerly cascaded pseudo-element. These are eagerly computed once, and then used to resolve the new computed values on the fly on layout.

rules_source_order: usize state_deps: DependencySet

Selector dependencies used to compute restyle hints.

sibling_affecting_selectors: Vec<Selector<TheSelectorImpl>>

Selectors in the page affecting siblings

non_common_style_affecting_attributes_selectors: Vec<Selector<TheSelectorImpl>>

Selectors in the page matching elements with non-common style-affecting attributes.

Methods

impl Stylist

fn new(device: Device) -> Self

fn update(&mut self,
doc_stylesheets: &[Arc<Stylesheet>],
ua_stylesheets: Option<&UserAgentStylesheets>,
stylesheets_changed: bool)
-> bool

fn add_stylesheet(&mut self, stylesheet: &Stylesheet)

fn precomputed_values_for_pseudo(&self,
pseudo: &PseudoElement,
parent: Option<&Arc<ComputedValues>>,
inherit_all: bool)
-> Option<Arc<ComputedValues>>

Computes the style for a given "precomputed" pseudo-element, taking the universal rules and applying them.

If inherit_all is true, then all properties are inherited from the parent; otherwise, non-inherited properties are reset to their initial values. The flow constructor uses this flag when constructing anonymous flows.

fn style_for_anonymous_box(&self,
pseudo: &PseudoElement,
parent_style: &Arc<ComputedValues>)
-> Arc<ComputedValues>

Returns the style for an anonymous box of the given type.

fn lazily_compute_pseudo_element_style<E>(&self,
element: &E,
pseudo: &PseudoElement,
parent: &Arc<ComputedValues>)
-> Option<Arc<ComputedValues>> where E: Element<Impl=TheSelectorImpl> + Debug + PresentationalHintsSynthetizer

fn set_device(&mut self, device: Device, stylesheets: &[Arc<Stylesheet>])

fn viewport_constraints(&self) -> &Option<ViewportConstraints>

fn set_quirks_mode(&mut self, enabled: bool)

fn push_applicable_declarations<E, V>(&self,
element: &E,
parent_bf: Option<&BloomFilter>,
style_attribute: Option<&Arc<RwLock<PropertyDeclarationBlock>>>,
pseudo_element: Option<&PseudoElement>,
applicable_declarations: &mut V,
reason: MatchingReason)
-> StyleRelations where E: Element<Impl=TheSelectorImpl> + Debug + PresentationalHintsSynthetizer, V: Push<ApplicableDeclarationBlock> + VecLike<ApplicableDeclarationBlock>

Returns the applicable CSS declarations for the given element. This corresponds to ElementRuleCollector in WebKit.

The returned boolean indicates whether the style is shareable; that is, whether the matched selectors are simple enough to allow the matching logic to be reduced to the logic in css::matching::PrivateMatchMethods::candidate_element_allows_for_style_sharing.

fn is_device_dirty(&self) -> bool

fn animations(&self) -> &FnvHashMap<Atom, KeyframesAnimation>

fn match_same_not_common_style_affecting_attributes_rules<E>(&self,
element: &E,
candidate: &E)
-> bool where E: ElementExt

fn match_same_sibling_affecting_rules<E>(&self,
element: &E,
candidate: &E)
-> bool where E: ElementExt

fn compute_restyle_hint<E>(&self,
element: &E,
snapshot: &E::Snapshot,
current_state: ElementState)
-> RestyleHint where E: ElementExt + Clone