style

Module servo

Servo-specific bits of the style system.

These get compiled out on a Gecko build.

Modules

animation

CSS transitions and animations.

attr

Parsed representations of DOM attributes.

encoding_support 🔒

Parsing stylesheets from bytes (not &str).

media_features

Servo’s media feature list and evaluator.

restyle_damage

The restyle damage is a hint that tells layout which kind of operations may be needed in presence of incremental style changes.

selector_parser

Servo’s selector parser.

shadow_parts 🔒