style::servo

Module animation

CSS transitions and animations.

Structs

Animation

A CSS Animation

AnimationSetKey

A key that is used to identify nodes in the DocumentAnimationSet.

ComputedKeyframe 🔒

A single computed keyframe for a CSS Animation.

DocumentAnimationSet

A set of animations for a document.

ElementAnimationSet

Holds the animation state for a particular element.

IntermediateComputedKeyframe 🔒

A temporary data structure used when calculating ComputedKeyframes for an animation. This data structure is used to collapse information for steps which may be spread across multiple keyframe declarations into a single instance per start_percentage.

KeyframeDataForProperty 🔒

KeyframeOffsetCacheForProperty 🔒

Caches the indices of keyframes that declare a specific property.

PropertyAnimation

Represents an animation for a given property.

PropertyDeclarationOffsets 🔒

Transition

A CSS Transition

Enums

AnimationState

This structure represents the state of an animation.

AnimationValueOrReference 🔒

Direction 🔒

IgnoreTransitions 🔒

KeyframesIterationState

This structure represents a keyframes animation current iteration state.

Functions

composite_animation_value 🔒

Composite a keyframe value with the underlying value according to the given composite operation.

maybe_start_animations

Triggers animations for a given node looking at the animation property values.

start_transitions_if_applicable

Kick off any new transitions for this node and return all of the properties that are transitioning. This is at the end of calculating style for a single node.