Module style::servo::media_queries

Servo's media-query device and expression representation.

Structs

Device
A device is a structure that represents the current media a given document is displayed in.

MEDIA_FEATURES
A list with all the media features that Servo supports.

Enums

Scan

Constants

_IMPL_NUM_FromPrimitive_FOR_Scan

Functions

eval_scan
https://drafts.csswg.org/mediaqueries-4/#scan

eval_width
https://drafts.csswg.org/mediaqueries-4/#width