Struct style::servo::media_queries::Device

pub struct Device {
    media_type: MediaType,
    viewport_size: Size2D<f32, CSSPixel>,
    device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>,
    quirks_mode: QuirksMode,
    root_font_size: AtomicU32,
    used_root_font_size: AtomicBool,
    used_viewport_units: AtomicBool,
    environment: CssEnvironment,
}

A device is a structure that represents the current media a given document is displayed in.

This is the struct against which media queries are evaluated.

Fields
media_type: MediaType

The current media type used by de device.

viewport_size: Size2D<f32, CSSPixel>

The current viewport size, in CSS pixels.

device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>

The current device pixel ratio, from CSS pixels to device pixels.

quirks_mode: QuirksMode

The current quirks mode.

root_font_size: AtomicU32

The font size of the root element This is set when computing the style of the root element, and used for rem units in other elements

When computing the style of the root element, there can’t be any other style being computed at the same time, given we need the style of the parent to compute everything else. So it is correct to just use a relaxed atomic here.

used_root_font_size: AtomicBool

Whether any styles computed in the document relied on the root font-size by using rem units.

used_viewport_units: AtomicBool

Whether any styles computed in the document relied on the viewport size.

environment: CssEnvironment

The CssEnvironment object responsible of getting CSS environment variables.

Implementations

impl Device

pub fn new( media_type: MediaType, quirks_mode: QuirksMode, viewport_size: Size2D<f32, CSSPixel>, device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel> ) -> Device

Trivially construct a new Device.

pub fn environment(&self) -> &CssEnvironment

Get the relevant environment to resolve env() functions.

pub fn default_computed_values(&self) -> &ComputedValues

Return the default computed values for this device.

pub fn root_font_size(&self) -> CSSPixelLength

Get the font size of the root element (for rem)

pub fn set_root_font_size(&self, size: CSSPixelLength)

Set the font size of the root element (for rem)

pub fn quirks_mode(&self) -> QuirksMode

Get the quirks mode of the current device.

pub fn set_body_text_color(&self, _color: AbsoluteColor)

Sets the body text color for the “inherit color from body” quirk.

https://quirks.spec.whatwg.org/#the-tables-inherit-color-from-body-quirk

pub fn animation_name_may_be_referenced(&self, _: &KeyframesName) -> bool

Whether a given animation name may be referenced from style.

pub fn used_root_font_size(&self) -> bool

Returns whether we ever looked up the root font size of the Device.

pub fn au_viewport_size(&self) -> UntypedSize2D<Au>

Returns the viewport size of the current device in app units, needed, among other things, to resolve viewport units.

pub fn au_viewport_size_for_viewport_unit_resolution( &self, _: ViewportVariant ) -> UntypedSize2D<Au>

Like the above, but records that we’ve used viewport units.

pub fn used_viewport_units(&self) -> bool

Whether viewport units were used since the last device change.

pub fn app_units_per_device_pixel(&self) -> i32

Returns the number of app units per device pixel we’re using currently.

pub fn device_pixel_ratio(&self) -> Scale<f32, CSSPixel, DevicePixel>

Returns the device pixel ratio.

pub fn scrollbar_inline_size(&self) -> CSSPixelLength

Gets the size of the scrollbar in CSS pixels.

pub fn query_font_metrics( &self, _vertical: bool, _font: &Font, _base_size: CSSPixelLength, _in_media_query: bool, _retrieve_math_scales: bool ) -> FontMetrics

Queries dummy font metrics for Servo. Knows nothing about fonts and does not provide any metrics. TODO: Servo’s font metrics provider will probably not live in this crate, so this will have to be replaced with something else (perhaps a trait method on TElement) when we get there

pub fn media_type(&self) -> MediaType

Return the media type of the current device.

pub fn use_document_colors(&self) -> bool

Returns whether document colors are enabled.

pub fn default_background_color(&self) -> AbsoluteColor

Returns the default background color.

pub fn default_color(&self) -> AbsoluteColor

Returns the default foreground color.

pub fn safe_area_insets(&self) -> SideOffsets2D<f32, CSSPixel>

Returns safe area insets

pub fn is_supported_mime_type(&self, mime_type: &str) -> bool

Returns true if the given MIME type is supported

pub fn chrome_rules_enabled_for_document(&self) -> bool

Return whether the document is a chrome document.