style::servo::media_queries

Trait FontMetricsProvider

pub trait FontMetricsProvider: Debug + Sync {
    // Required methods
    fn query_font_metrics(
        &self,
        vertical: bool,
        font: &Font,
        base_size: CSSPixelLength,
        flags: QueryFontMetricsFlags,
    ) -> FontMetrics;
    fn base_size_for_generic(&self, generic: GenericFontFamily) -> Length;
}

A trait used to query font metrics in clients of Stylo. This is used by Device to query font metrics in a way that is specific to the client using Stylo.

Required Methods

fn query_font_metrics( &self, vertical: bool, font: &Font, base_size: CSSPixelLength, flags: QueryFontMetricsFlags, ) -> FontMetrics

Query the font metrics for the given font and the given base font size.

fn base_size_for_generic(&self, generic: GenericFontFamily) -> Length

Gets the base size given a generic font family.