Enum style::servo::selector_parser::NonTSPseudoClass

pub enum NonTSPseudoClass {
    Active,
    AnyLink,
    Checked,
    Defined,
    Disabled,
    Enabled,
    Focus,
    Fullscreen,
    Hover,
    Indeterminate,
    Lang(Lang),
    Link,
    PlaceholderShown,
    ReadWrite,
    ReadOnly,
    ServoNonZeroBorder,
    Target,
    Visited,
}

A non tree-structural pseudo-class. See https://drafts.csswg.org/selectors-4/#structural-pseudos

Variants

Active

AnyLink

Checked

Defined

Disabled

Enabled

Focus

Fullscreen

Hover

Indeterminate

Lang(Lang)

Link

PlaceholderShown

ReadWrite

ReadOnly

ServoNonZeroBorder

Target

Visited

Implementations

impl NonTSPseudoClass

pub fn state_flag(&self) -> ElementState

Gets a given state flag for this pseudo-class. This is used to do selector matching, and it's set from the DOM.

pub fn document_state_flag(&self) -> DocumentState

Get the document state flag associated with a pseudo-class, if any.

pub fn needs_cache_revalidation(&self) -> bool

Returns true if the given pseudoclass should trigger style sharing cache revalidation.