style::servo::selector_parser

Enum PseudoElement

#[repr(u8)]
pub enum PseudoElement {

    After = 0,
    Before = 1,
    Selection = 2,
    FirstLetter = 3,
    Backdrop = 4,
    DetailsContent = 5,
    Marker = 6,
    ColorSwatch = 7,
    FileSelectorButton = 8,
    Placeholder = 9,
    SliderFill = 10,
    SliderThumb = 11,
    SliderTrack = 12,
    ServoTextControlInnerContainer = 13,
    ServoTextControlInnerEditor = 14,
    ServoAnonymousBox = 15,
    ServoAnonymousTable = 16,
    ServoAnonymousTableCell = 17,
    ServoAnonymousTableRow = 18,
    ServoTableGrid = 19,
    ServoTableWrapper = 20,

}

A pseudo-element, both public and private.

NB: If you add to this list, be sure to update each_simple_pseudo_element too.

Variants

After = 0

Before = 1

Selection = 2

FirstLetter = 3

Backdrop = 4

DetailsContent = 5

Marker = 6

ColorSwatch = 7

FileSelectorButton = 8

Placeholder = 9

SliderFill = 10

SliderThumb = 11

SliderTrack = 12

ServoTextControlInnerContainer = 13

ServoTextControlInnerEditor = 14

ServoAnonymousBox = 15

ServoAnonymousTable = 16

ServoAnonymousTableCell = 17

ServoAnonymousTableRow = 18

ServoTableGrid = 19

ServoTableWrapper = 20

Implementations

impl PseudoElement

pub fn eager_index(&self) -> usize

Gets the canonical index of this eagerly-cascaded pseudo-element.

pub fn index(&self) -> usize

An index for this pseudo-element to be indexed in an enumerated array.

pub fn pseudo_none_array<T>() -> [Option<T>; 21]

An array of None, one per pseudo-element.

pub fn from_eager_index(i: usize) -> Self

Creates a pseudo-element from an eager index.

pub fn is_before_or_after(&self) -> bool

Whether the current pseudo element is ::before or ::after.

pub fn is_unknown_webkit_pseudo_element(&self) -> bool

Whether this is an unknown ::-webkit- pseudo-element.

pub fn is_marker(&self) -> bool

Whether this pseudo-element is the ::marker pseudo.

pub fn is_selection(&self) -> bool

Whether this pseudo-element is the ::selection pseudo.

pub fn is_before(&self) -> bool

Whether this pseudo-element is the ::before pseudo.

pub fn is_after(&self) -> bool

Whether this pseudo-element is the ::after pseudo.

pub fn is_first_letter(&self) -> bool

Whether the current pseudo element is :first-letter

pub fn is_first_line(&self) -> bool

Whether the current pseudo element is :first-line

pub fn is_color_swatch(&self) -> bool

Whether this pseudo-element is representing the color swatch inside an <input> element.

pub fn is_eager(&self) -> bool

Whether this pseudo-element is eagerly-cascaded.

pub fn is_lazy(&self) -> bool

Whether this pseudo-element is lazily-cascaded.

pub fn is_anon_box(&self) -> bool

Whether this pseudo-element is for an anonymous box.

pub fn skip_item_display_fixup(&self) -> bool

Whether this pseudo-element skips flex/grid container display-based fixup.

pub fn is_precomputed(&self) -> bool

Whether this pseudo-element is precomputed.

pub fn cascade_type(&self) -> PseudoElementCascadeType

Returns which kind of cascade type has this pseudo.

See the documentation for PseudoElementCascadeType for how we choose which cascade type to use.

Note: Keep eager pseudos in sync with EAGER_PSEUDO_COUNT and EMPTY_PSEUDO_ARRAY in style/data.rs

pub fn canonical(&self) -> PseudoElement

Covert non-canonical pseudo-element to canonical one, and keep a canonical one as it is.

pub fn pseudo_info(&self)

Stub, only Gecko needs this

pub fn property_restriction(&self) -> Option<PropertyFlags>

Property flag that properties must have to apply to this pseudo-element.

pub fn should_exist(&self, style: &ComputedValues) -> bool

Whether this pseudo-element should actually exist if it has the given styles.

pub fn is_highlight(&self) -> bool

Whether this pseudo-element is the ::highlight pseudo.

pub fn is_target_text(&self) -> bool

Whether this pseudo-element is the ::target-text pseudo.

pub fn is_lazy_painted_highlight_pseudo(&self) -> bool

Whether this is a highlight pseudo-element that is styled lazily during painting rather than during the restyle traversal. These pseudos need explicit repaint triggering when their styles change.

pub fn is_element_backed(&self) -> bool

Whether this pseudo-element is “element-backed”, which means that it inherits from its regular flat tree parent, which might not be the originating element.