style::servo::selector_parser

Struct ServoElementSnapshot

pub struct ServoElementSnapshot {
    pub state: Option<ElementState>,
    pub attrs: Option<Vec<(AttrIdentifier, AttrValue)>>,
    pub changed_attrs: Vec<LocalName>,
    pub class_changed: bool,
    pub id_changed: bool,
    pub other_attributes_changed: bool,
}

Servo’s version of an element snapshot.

Fields
state: Option<ElementState>

The stored state of the element.

attrs: Option<Vec<(AttrIdentifier, AttrValue)>>

The set of stored attributes and its values.

changed_attrs: Vec<LocalName>

The set of changed attributes and its values.

class_changed: bool

Whether the class attribute changed or not.

id_changed: bool

Whether the id attribute changed or not.

other_attributes_changed: bool

Whether other attributes other than id or class changed or not.

Implementations

impl ServoElementSnapshot

pub fn new() -> Self

Create an empty element snapshot.

pub fn id_changed(&self) -> bool

Returns whether the id attribute changed or not.

pub fn class_changed(&self) -> bool

Returns whether the class attribute changed or not.

pub fn other_attr_changed(&self) -> bool

Returns whether other attributes other than id or class changed or not.

fn get_attr( &self, namespace: &Namespace, name: &LocalName, ) -> Option<&AttrValue>

pub fn each_attr_changed<F>(&self, callback: F)
where F: FnMut(&LocalName),

Executes the callback once for each attribute that changed.

fn any_attr_ignore_ns<F>(&self, name: &LocalName, f: F) -> bool
where F: FnMut(&AttrValue) -> bool,

impl ServoElementSnapshot

pub fn attr_matches( &self, ns: &NamespaceConstraint<&Namespace>, local_name: &LocalName, operation: &AttrSelectorOperation<&AtomString>, ) -> bool

selectors::Element::attr_matches