Type Alias style::servo::url::SpecifiedUrl

pub type SpecifiedUrl = CssUrl;

A specified url() value for servo.

Aliased Type

struct SpecifiedUrl {
    original: Option<Arc<String>>,
    resolved: Option<Arc<Url>>,
}

Fields
original: Option<Arc<String>>

The original URI. This might be optional since we may insert computed values of images into the cascade directly, and we don’t bother to convert their serialization.

Refcounted since cloning this should be cheap and data: uris can be really large.

resolved: Option<Arc<Url>>

The resolved value for the url, if valid.

Implementations

impl CssUrl

pub fn parse_from_string( url: String, context: &ParserContext<'_>, _: CorsMode ) -> Self

Try to parse a URL from a string value that is a valid CSS token for a URL.

FIXME(emilio): Should honor CorsMode.

pub fn is_invalid(&self) -> bool

Returns true if the URL is definitely invalid. For Servo URLs, we can use its |resolved| status.

pub fn is_fragment(&self) -> bool

Returns true if this URL looks like a fragment. See https://drafts.csswg.org/css-values/#local-urls

Since Servo currently stores resolved URLs, this is hard to implement. We either need to change servo to lazily resolve (like Gecko), or note this information in the tokenizer.

pub fn url(&self) -> Option<&Arc<Url>>

Returns the resolved url if it was valid.

pub fn as_str(&self) -> &str

Return the resolved url as string, or the empty string if it’s invalid.

TODO(emilio): Should we return the original one if needed?

pub fn for_cascade(url: Arc<Url>) -> Self

Creates an already specified url value from an already resolved URL for insertion in the cascade.

pub fn new_for_testing(url: &str) -> Self

Gets a new url from a string for unit tests.

pub fn parse_with_cors_mode<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, cors_mode: CorsMode ) -> Result<Self, ParseError<'i>>

Parses a URL request and records that the corresponding request needs to be CORS-enabled.

This is only for shape images and masks in Gecko, thus unimplemented for now so somebody notices when trying to do so.