Module style::shared_lock

Different objects protected by the same lock

Modules

compile_time_assert

Structs

Locked
Data protect by a shared lock.

SharedRwLock
A shared read/write lock that can protect multiple objects.

SharedRwLockReadGuard
Proof that a shared lock was obtained for reading.

SharedRwLockWriteGuard
Proof that a shared lock was obtained for writing.

StylesheetGuards
Guards for a document

Traits

ToCssWithGuard
Like ToCss, but with a lock guard given by the caller.