Struct style::shared_lock::SharedRwLock

pub struct SharedRwLock { /* fields omitted */ }

A shared read/write lock that can protect multiple objects.

In Gecko builds, we don't need the blocking behavior, just the safety. As such we implement this with an AtomicRefCell instead in Gecko builds, which is ~2x as fast, and panics (rather than deadlocking) when things go wrong (which is much easier to debug on CI).

Servo needs the blocking behavior for its unsynchronized animation setup, but that may not be web-compatible and may need to be changed (at which point Servo could use AtomicRefCell too).

Methods

impl SharedRwLock

pub fn new() -> Self

Create a new shared lock (servo).

pub fn wrap<T>(&self, data: T) -> Locked<T>

Wrap the given data to make its access protected by this lock.

pub fn read(&self) -> SharedRwLockReadGuard

Obtain the lock for reading (servo).

pub fn write(&self) -> SharedRwLockWriteGuard

Obtain the lock for writing (servo).