style::sharing

Module checks

Different checks done during the style sharing process in order to determine quickly whether it’s worth to share style, and whether two different elements can indeed share the same style.

Functions

have_same_class

Whether a given element has the same class attribute as a given candidate.

have_same_parts

Whether a given element has the same part attribute as a given candidate.

have_same_presentational_hints

Whether two elements have the same same presentational attributes.

have_same_referenced_attrs

Whether the given element and a candidate have the same values for the the attributes used in an attr() function.

have_same_style_attribute

Whether two elements have the same style attribute.

have_shareable_tree_counting_functions

Whether two elements have compatible tree-counting functions.

may_match_different_id_rules

Checks whether we might have rules for either of the two ids.

parents_allow_sharing

Determines whether a target and a candidate have compatible parents for sharing.

revalidate

Whether a given element and a candidate match the same set of “revalidation” selectors.

revalidate_scope

Whether a given element and a candidate share a set of scope activations for revalidation.

shadow_root_style_data_equals

Returns whether the cascade data of the given shadow roots is the same.