style::sharing

Struct StyleSharingCandidate

pub struct StyleSharingCandidate<E: TElement> {
    element: E,
    validation_data: ValidationData,
    considered_nontrivial_scoped_style: bool,
}

Information regarding a style sharing candidate, that is, an entry in the style sharing cache.

Note that this information is stored in TLS and cleared after the traversal, and once here, the style information of the element is immutable, so itโs safe to access.

Important: If you change the members/layout here, You need to do the same for FakeCandidate below.

Fields
element: E

The element.

validation_data: ValidationDataconsidered_nontrivial_scoped_style: bool
Implementations

impl<E: TElement> StyleSharingCandidate<E>

fn class_list(&mut self) -> &[AtomIdent]

Get the classlist of this candidate.

fn part_list(&mut self) -> &[AtomIdent]

Get the part list of this candidate.

fn pres_hints(&mut self) -> &[ApplicableDeclarationBlock]

Get the pres hints of this candidate.

fn parent_style_identity(&mut self) -> OpaqueComputedValues

Get the parent style identity.

fn revalidation_match_results( &mut self, stylist: &Stylist, bloom: &StyleBloom<E>, selector_caches: &mut SelectorCaches, ) -> &RevalidationResult

Compute the bit vector of revalidation selector match results for this candidate.

fn scope_revalidation_results( &mut self, stylist: &Stylist, selector_caches: &mut SelectorCaches, ) -> ScopeRevalidationResult