Struct style::sharing::StyleSharingTarget

pub struct StyleSharingTarget<E: TElement> {
    element: E,
    validation_data: ValidationData,
}

An element we want to test against the style sharing cache.

Fields
element: Evalidation_data: ValidationData
Implementations

impl<E: TElement> StyleSharingTarget<E>

pub fn new(element: E) -> Self

Trivially construct a new StyleSharingTarget to test against the cache.

fn class_list(&mut self) -> &[AtomIdent]

fn part_list(&mut self) -> &[AtomIdent]

fn pres_hints(&mut self) -> &[ApplicableDeclarationBlock]

Get the pres hints of this candidate.

fn parent_style_identity(&mut self) -> OpaqueComputedValues

Get the parent style identity.

fn revalidation_match_results(
&mut self,
stylist: &Stylist,
bloom: &StyleBloom<E>,
nth_index_cache: &mut NthIndexCache,
selector_flags_map: &mut SelectorFlagsMap<E>
) -> &SmallBitVec

pub fn share_style_if_possible(
&mut self,
context: &mut StyleContext<'_, E>
) -> Option<ResolvedElementStyles>

Attempts to share a style with another node.

pub fn take_validation_data(&mut self) -> ValidationData

Gets the validation data used to match against this target, if any.