Struct style::OwnedSlice

#[repr(C)]pub struct OwnedSlice<T> {
    ptr: NonNull<T>,
    len: usize,
    _phantom: PhantomData<T>,
}

A struct that basically replaces a Box<[T]>, but which cbindgen can understand.

We could rely on the struct layout of Box<[T]> per:

https://github.com/rust-lang/unsafe-code-guidelines/blob/master/reference/src/layout/pointers.md

But handling fat pointers with cbindgen both in structs and argument positions more generally is a bit tricky.

cbindgen:derive-eq=false cbindgen:derive-neq=false

Fields
ptr: NonNull<T>len: usize_phantom: PhantomData<T>
Implementations

impl<T> OwnedSlice<T>

pub fn into_box(self) -> Box<[T]>

Convert the OwnedSlice into a boxed slice.

pub fn into_vec(self) -> Vec<T>

Convert the OwnedSlice into a Vec.

pub fn into_iter(self) -> impl ExactSizeIterator + Iterator<Item = T>

Iterate over all the elements in the slice taking ownership of them.

pub fn from_slice(s: &[T]) -> OwnedSlice<T> where
T: Clone,

Convert the regular slice into an owned slice.