Module style::style_resolver

Style resolution for a given element or pseudo-element.

Structs

MatchingResults

PrimaryStyle
The primary style of an element or an element-backed pseudo-element.

ResolvedElementStyles
A set of style returned from the resolver machinery.

ResolvedStyle
A style returned from the resolver machinery.

StyleResolverForElement
A struct that takes care of resolving the style of a given element.

Enums

PseudoElementResolution
Whether pseudo-elements should be resolved or not.

Functions

eager_pseudo_is_definitely_not_generated

layout_parent_style_for_pseudo

with_default_parent_styles