Struct style::style_resolver::PrimaryStyle

pub struct PrimaryStyle {
    pub style: ResolvedStyle,
    pub reused_via_rule_node: bool,
}

The primary style of an element or an element-backed pseudo-element.

Fields
style: ResolvedStyle

The style itself.

reused_via_rule_node: bool

Whether the style was reused from another element via the rule node (see StyleSharingCache::lookup_by_rules).

Methods

impl PrimaryStyle

pub fn style(&self) -> &ComputedValues

Convenience accessor for the style.