Enum style::stylesheet_set::OriginValidity

pub enum OriginValidity {
    Valid,
    CascadeInvalid,
    FullyInvalid,
}

The validity of the data in a given cascade origin.

Variants
Valid

The origin is clean, all the data already there is valid, though we may have new sheets at the end.

CascadeInvalid

The cascade data is invalid, but not the invalidation data (which is order-independent), and thus only the cascade data should be inserted.

FullyInvalid

Everything needs to be rebuilt.