Enum style::stylesheet_set::SheetRebuildKind

pub enum SheetRebuildKind {
    Full,
    CascadeOnly,
}

The type of rebuild that we need to do for a given stylesheet.

Variants
Full

A full rebuild, of both cascade data and invalidation data.

CascadeOnly

A partial rebuild, of only the cascade data.

Methods

impl SheetRebuildKind

pub fn should_rebuild_invalidation(&self) -> bool

Whether the stylesheet invalidation data should be rebuilt.