style::stylesheet_set

Struct AuthorStylesheetSet

pub struct AuthorStylesheetSet<S>
where
    S: StylesheetInDocument + PartialEq + 'static,
{
    collection: SheetCollection<S>,
    invalidations: StylesheetInvalidationSet,
}

The set of stylesheets effective for a given Shadow Root.

Fields
collection: SheetCollection<S>

The actual style sheets.

invalidations: StylesheetInvalidationSet

The set of invalidations scheduled for this collection.

Implementations

impl<S> AuthorStylesheetSet<S>
where S: StylesheetInDocument + PartialEq + 'static,

pub fn new() -> Self

Create a new empty AuthorStylesheetSet.

pub fn dirty(&self) -> bool

Whether anything has changed since the last time this was flushed.

pub fn is_empty(&self) -> bool

Whether the collection is empty.

pub fn get(&self, index: usize) -> Option<&S>

Returns the indexth stylesheet in the collection of author styles if present.

pub fn len(&self) -> usize

Returns the number of author stylesheets.

fn collection_for( &mut self, _: &S, _: &SharedRwLockReadGuard<'_>, ) -> &mut SheetCollection<S>

fn collect_invalidations_for( &mut self, device: Option<&Device>, custom_media: &CustomMediaMap, sheet: &S, guard: &SharedRwLockReadGuard<'_>, )

pub fn append_stylesheet( &mut self, device: Option<&Device>, custom_media: &CustomMediaMap, sheet: S, guard: &SharedRwLockReadGuard<'_>, )

Appends a new stylesheet to the current set.

No device implies not computing invalidations.

pub fn insert_stylesheet_before( &mut self, device: Option<&Device>, custom_media: &CustomMediaMap, sheet: S, before_sheet: S, guard: &SharedRwLockReadGuard<'_>, )

Insert a given stylesheet before another stylesheet in the document.

pub fn remove_stylesheet( &mut self, device: Option<&Device>, custom_media: &CustomMediaMap, sheet: S, guard: &SharedRwLockReadGuard<'_>, )

Remove a given stylesheet from the set.

pub fn rule_changed( &mut self, device: Option<&Device>, custom_media: &CustomMediaMap, sheet: &S, rule: &CssRule, guard: &SharedRwLockReadGuard<'_>, change_kind: RuleChangeKind, ancestors: &[CssRuleRef<'_>], )

Notify the set that a rule from a given stylesheet has changed somehow.

pub fn iter(&self) -> impl Iterator<Item = &S>

Iterate over the list of stylesheets.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut S>

Returns a mutable iterator over the current list of stylesheets.

pub fn force_dirty(&mut self)

Mark the sheet set dirty, as appropriate.

pub fn flush( &mut self, ) -> (AuthorStylesheetFlusher<'_, S>, StylesheetInvalidationSet)

Flush the stylesheets for this author set.