Struct style::stylesheet_set::DocumentStylesheetSet

pub struct DocumentStylesheetSet<S> where
    S: StylesheetInDocument + PartialEq + 'static,  {
    collections: PerOrigin<SheetCollection<S>>,
    invalidations: StylesheetInvalidationSet,
}

The set of stylesheets effective for a given document.

Fields
collections: PerOrigin<SheetCollection<S>>

The collections of sheets per each origin.

invalidations: StylesheetInvalidationSet

The invalidations for stylesheets added or removed from this document.

Methods

impl<S> DocumentStylesheetSet<S> where
S: StylesheetInDocument + PartialEq + 'static,

pub fn new() -> Self

Create a new empty DocumentStylesheetSet.

fn collection_for(
&mut self,
sheet: &S,
guard: &SharedRwLockReadGuard
) -> &mut SheetCollection<S>

fn collect_invalidations_for(
&mut self,
device: Option<&Device>,
sheet: &S,
guard: &SharedRwLockReadGuard
)

pub fn append_stylesheet(
&mut self,
device: Option<&Device>,
sheet: S,
guard: &SharedRwLockReadGuard
)

Appends a new stylesheet to the current set.

No device implies not computing invalidations.

pub fn insert_stylesheet_before(
&mut self,
device: Option<&Device>,
sheet: S,
before_sheet: S,
guard: &SharedRwLockReadGuard
)

Insert a given stylesheet before another stylesheet in the document.

pub fn remove_stylesheet(
&mut self,
device: Option<&Device>,
sheet: S,
guard: &SharedRwLockReadGuard
)

Remove a given stylesheet from the set.

pub fn len(&self) -> usize

Returns the number of stylesheets in the set.

pub fn sheet_count(&self, origin: Origin) -> usize

Returns the count of stylesheets for a given origin.

pub fn get(&self, origin: Origin, index: usize) -> Option<&S>

Returns the indexth stylesheet in the set for the given origin.

pub fn has_changed(&self) -> bool

Returns whether the given set has changed from the last flush.

pub fn flush<E>(
&mut self,
document_element: Option<E>,
snapshots: Option<&SnapshotMap>
) -> DocumentStylesheetFlusher<S> where
E: TElement,

Flush the current set, unmarking it as dirty, and returns a DocumentStylesheetFlusher in order to rebuild the stylist.

pub fn flush_without_invalidation(&mut self) -> OriginSet

Flush stylesheets, but without running any of the invalidation passes.

Important traits for StylesheetIterator<'a, S>
impl<'a, S> Iterator for StylesheetIterator<'a, S> where
S: StylesheetInDocument + PartialEq + 'static, type Item = (&'a S, Origin);

pub fn iter(&self) -> StylesheetIterator<S>

Return an iterator over the flattened view of all the stylesheets.

pub fn force_dirty(&mut self, origins: OriginSet)

Mark the stylesheets for the specified origin as dirty, because something external may have invalidated it.