Struct style::stylesheet_set::StylesheetFlusher

pub struct StylesheetFlusher<'a, S> where
    S: StylesheetInDocument + PartialEq + 'static,  { /* fields omitted */ }

A struct to iterate over the different stylesheets to be flushed.

Methods

impl<'a, S> StylesheetFlusher<'a, S> where
S: StylesheetInDocument + PartialEq + 'static,

pub fn origin_validity(&self, origin: Origin) -> OriginValidity

The data validity for a given origin.

pub fn origin_dirty(&self, origin: Origin) -> bool

Whether the origin data is dirty in any way.

pub fn manual_origin_sheets<'b>(
&'b mut self,
origin: Origin
) -> StylesheetCollectionIterator<'b, S> where
'a: 'b,

Returns an iterator over the stylesheets of a given origin, assuming all of them will be flushed.

pub fn origin_sheets<'b>(
&'b mut self,
origin: Origin
) -> PerOriginFlusher<'b, S> where
'a: 'b,

Returns a flusher for the dirty origin origin.

pub fn nothing_to_do(&self) -> bool

Returns whether running the whole flushing process would be a no-op.

pub fn had_invalidations(&self) -> bool

Returns whether any DOM invalidations were processed as a result of the stylesheet flush.