Struct style::stylesheet_set::StylesheetSet

pub struct StylesheetSet<S> where
    S: StylesheetInDocument + PartialEq + 'static,  { /* fields omitted */ }

The set of stylesheets effective for a given document.

Methods

impl<S> StylesheetSet<S> where
S: StylesheetInDocument + PartialEq + 'static,

pub fn new() -> Self

Create a new empty StylesheetSet.

pub fn len(&self) -> usize

Returns the number of stylesheets in the set.

pub fn get(&self, origin: Origin, index: usize) -> Option<&S>

Returns the indexth stylesheet in the set for the given origin.

pub fn author_style_disabled(&self) -> bool

Returns whether author styles have been disabled for the current stylesheet set.

pub fn append_stylesheet(
&mut self,
device: Option<&Device>,
sheet: S,
guard: &SharedRwLockReadGuard
)

Appends a new stylesheet to the current set.

No device implies not computing invalidations.

pub fn prepend_stylesheet(
&mut self,
device: Option<&Device>,
sheet: S,
guard: &SharedRwLockReadGuard
)

Prepend a new stylesheet to the current set.

pub fn insert_stylesheet_before(
&mut self,
device: Option<&Device>,
sheet: S,
before_sheet: S,
guard: &SharedRwLockReadGuard
)

Insert a given stylesheet before another stylesheet in the document.

pub fn remove_stylesheet(
&mut self,
device: Option<&Device>,
sheet: S,
guard: &SharedRwLockReadGuard
)

Remove a given stylesheet from the set.

pub fn set_author_style_disabled(&mut self, disabled: bool)

Notes that the author style has been disabled for this document.

pub fn has_changed(&self) -> bool

Returns whether the given set has changed from the last flush.

pub fn flush<'a, E>(
&'a mut self,
document_element: Option<E>
) -> StylesheetFlusher<'a, S> where
E: TElement,

Flush the current set, unmarking it as dirty, and returns a StylesheetFlusher in order to rebuild the stylist.

pub fn flush_without_invalidation(&mut self) -> OriginSet

Flush stylesheets, but without running any of the invalidation passes.

pub fn iter(&self) -> StylesheetIterator<S>

Return an iterator over the flattened view of all the stylesheets.

pub fn force_dirty(&mut self, origins: OriginSet)

Mark the stylesheets for the specified origin as dirty, because something external may have invalidated it.