style::stylesheets

Module container_rule

A @container rule.

Structs

ContainerCondition

A container condition and filter, combined.

ContainerConditions

Contains all container conditions for a container rule.

ContainerInfo

Information needed to evaluate an individual container query.

ContainerLookupResult

The result of a successful container query lookup.

ContainerRule

A container rule.

ContainerSizeQueryResult

Result of a container size query, signifying the hypothetical containment boundary in terms of physical axes. Defined by up to two size containers. Queries on logical axes are resolved with respect to the querying element’s writing mode.

Enums

ContainerAttributeDependencyKind

Whether this container uses attr() references and if so whether or not this is a named or unnamed container query. Used for determining which kind of restyle hint we need when traversing invalidation.

ContainerSizeQuery

Unevaluated lazy container size query.

TraversalResult 🔒

Statics

CONTAINER_FEATURES

https://drafts.csswg.org/css-contain-3/#container-features

Functions

container_type_axes 🔒

eval_aspect_ratio 🔒

eval_block_size 🔒

eval_height 🔒

eval_inline_size 🔒

eval_orientation 🔒

eval_width 🔒

traverse_container 🔒

Type Aliases

AttrReferenceSet

A set of attribute names used for invalidation of attr() inside style container queries.