style::stylesheets

Module keyframes_rule

Keyframes: https://drafts.csswg.org/css-animations/#keyframes

Structs

Keyframe

A keyframe.

KeyframeListParser 🔒

Parses a keyframes list, like: 0%, 50% { width: 50%; }

KeyframePercentage

A number from 0 to 1, indicating the percentage of the animation when this keyframe should run.

KeyframeSelector

A single <keyframe-selector>: <keyframe-selector> = from | to | <percentage [0,100]> | <timeline-range-name> <percentage> It could be a percentage, from/to, or a timeline range name together with a percentage. https://drafts.csswg.org/scroll-animations-1/#named-range-keyframes

KeyframeSelectors

A list of <keyframe-selector>s.

KeyframesAnimation

This structure represents a list of animation steps computed from the list of keyframes, in order.

KeyframesRule

A @keyframes rule.

KeyframesStep

A single step from a keyframe animation.

Enums

KeyframeOffset

The Keyframe offset for Web Animations. Since we support double value from JS, so we need to include a number for it as well.

KeyframesStepValue

A keyframes step value. This can be a synthetised keyframes animation, that is, one autogenerated from the current computed values, or a list of declarations to apply.

Functions

get_animated_properties 🔒

Get all the animated properties in a keyframes animation.

parse_keyframe_list

Parses a keyframe list from CSS input.