Enum style::stylesheets::keyframes_rule::KeyframesStepValue

pub enum KeyframesStepValue {
    Declarations {
        block: Arc<Locked<PropertyDeclarationBlock>>,
    },
    ComputedValues,
}

A keyframes step value. This can be a synthetised keyframes animation, that is, one autogenerated from the current computed values, or a list of declarations to apply.

TODO: Find a better name for this?

Variants
Declarations

A step formed by a declaration block specified by the CSS.

Fields of Declarations

block: Arc<Locked<PropertyDeclarationBlock>>

The declaration block per se.

ComputedValues

A synthetic step computed from the current computed values at the time of the animation.