Struct style::stylesheets::keyframes_rule::Keyframe

pub struct Keyframe {
    pub selector: KeyframeSelector,
    pub block: Arc<Locked<PropertyDeclarationBlock>>,
    pub source_location: SourceLocation,
}

A keyframe.

Fields
selector: KeyframeSelector

The selector this keyframe was specified from.

block: Arc<Locked<PropertyDeclarationBlock>>

The declaration block that was declared inside this keyframe.

Note that !important rules in keyframes don't apply, but we keep this Arc just for convenience.

source_location: SourceLocation

The line and column of the rule's source code.

Methods

impl Keyframe

pub fn parse<'i>(
css: &'i str,
parent_stylesheet_contents: &StylesheetContents,
lock: &SharedRwLock
) -> Result<Arc<Locked<Self>>, ParseError<'i>>

Parse a CSS keyframe.