Struct style::stylesheets::keyframes_rule::KeyframesAnimation

pub struct KeyframesAnimation {
    pub steps: Vec<KeyframesStep>,
    pub properties_changed: LonghandIdSet,
    pub vendor_prefix: Option<VendorPrefix>,
}

This structure represents a list of animation steps computed from the list of keyframes, in order.

It only takes into account animable properties.

Fields
steps: Vec<KeyframesStep>

The difference steps of the animation.

properties_changed: LonghandIdSet

The properties that change in this animation.

vendor_prefix: Option<VendorPrefix>

Vendor prefix type the @keyframes has.

Methods

impl KeyframesAnimation

pub fn from_keyframes(
keyframes: &[Arc<Locked<Keyframe>>],
vendor_prefix: Option<VendorPrefix>,
guard: &SharedRwLockReadGuard
) -> Self

Create a keyframes animation from a given list of keyframes.

This will return a keyframe animation with empty steps and properties_changed if the list of keyframes is empty, or there are no animated properties obtained from the keyframes.

Otherwise, this will compute and sort the steps used for the animation, and return the animation object.