Struct style::stylesheets::keyframes_rule::KeyframesStep

pub struct KeyframesStep {
    pub start_percentage: KeyframePercentage,
    pub value: KeyframesStepValue,
    pub declared_timing_function: bool,
}

A single step from a keyframe animation.

Fields
start_percentage: KeyframePercentage

The percentage of the animation duration when this step starts.

value: KeyframesStepValue

Declarations that will determine the final style during the step, or ComputedValues if this is an autogenerated step.

declared_timing_function: bool

Wether a animation-timing-function declaration exists in the list of declarations.

This is used to know when to override the keyframe animation style.

Methods

impl KeyframesStep

pub fn get_animation_timing_function(
&self,
guard: &SharedRwLockReadGuard
) -> Option<SpecifiedTimingFunction>

Return specified TransitionTimingFunction if this KeyframesSteps has 'animation-timing-function'.