Module style::stylesheets::layer_rule

A @layer rule.

Structs

LayerBlockRule

A block @layer <name>? { ... } https://drafts.csswg.org/css-cascade-5/#layer-block

LayerName

A <layer-name>: https://drafts.csswg.org/css-cascade-5/#typedef-layer-name

LayerOrder

The order of a given layer. We use 16 bits so that we can pack LayerOrder and CascadeLevel in a single 32-bit struct. If we need more bits we can go back to packing CascadeLevel in a single byte as we did before.

LayerStatementRule

A statement @layer <name>, <name>, <name>;