Module style::stylesheets::origin

CSS cascade origins.

Structs

OriginSet
A set of origins. This is equivalent to Gecko's OriginFlags.

OriginSetIterator
Iterates over the origins present in an OriginSet, in order from highest priority (author) to lower (user agent).

PerOrigin
An object that stores a T for each origin of the CSS cascade.

PerOriginIter
Iterator over PerOrigin<T>, from highest level (author) to lowest (user agent).

PerOriginIterMut
Like PerOriginIter<T>, but iterates over mutable references to the per-origin data.

Enums

Origin
Each style rule has an origin, which determines where it enters the cascade.