style::stylesheets

Module page_rule

A @page rule.

Macros

page_pseudo_classes 🔒

Structs

PagePseudoClassFlags

Bit-flags for pseudo-class. This should only be used for querying if a page-rule applies.

PageRule

A @page rule.

PageSelector

Type of a single [@page][page selector]

PageSelectors

A list of [@page][page selectors]

Enums

PagePseudoClass

[@page][page] rule pseudo-classes.

Functions

parse_page_name 🔒

selector_specificity 🔒

Computes the specificity given the g, h, and f values as in the spec.

Type Aliases

PagePseudoClasses 🔒