style::stylesheets::rule_parser

Enum AtRulePrelude

pub enum AtRulePrelude {

    FontFace,
    FontFeatureValues(Vec<FamilyName>),
    FontPaletteValues(DashedIdent),
    CounterStyle(CustomIdent),
    Media(Arc<Locked<MediaList>>),
    Container(ArcSlice<ContainerCondition>),
    Supports(SupportsCondition),
    Keyframes(KeyframesName, Option<VendorPrefix>),
    Page(PageSelectors),
    Property(PropertyRuleName),
    Document(DocumentCondition),
    Import(CssUrl, Arc<Locked<MediaList>>, Option<ImportSupportsCondition>, ImportLayer),
    Margin(MarginRuleType),
    Namespace(Option<Prefix>, Namespace),
    Layer(Vec<LayerName>),
    Scope(ScopeBounds),
    StartingStyle,
    AppearanceBase,
    PositionTry(DashedIdent),
    CustomMedia(DashedIdent, CustomMediaCondition),
    ViewTransition,

}

A rule prelude for at-rule with block.

Variants

FontFace

A @font-face rule prelude.

FontFeatureValues(Vec<FamilyName>)

A @font-feature-values rule prelude, with its FamilyName list.

FontPaletteValues(DashedIdent)

A @font-palette-values rule prelude, with its identifier.

CounterStyle(CustomIdent)

A @counter-style rule prelude, with its counter style name.

Media(Arc<Locked<MediaList>>)

A @media rule prelude, with its media queries.

Container(ArcSlice<ContainerCondition>)

A @container rule prelude.

Supports(SupportsCondition)

An @supports rule, with its conditional

Keyframes(KeyframesName, Option<VendorPrefix>)

A @keyframes rule, with its animation name and vendor prefix if exists.

Page(PageSelectors)

A @page rule prelude, with its page name if it exists.

Property(PropertyRuleName)

A @property rule prelude.

Document(DocumentCondition)

A @document rule, with its conditional.

Import(CssUrl, Arc<Locked<MediaList>>, Option<ImportSupportsCondition>, ImportLayer)

A @import rule prelude.

Margin(MarginRuleType)

A @margin rule prelude.

Namespace(Option<Prefix>, Namespace)

A @namespace rule prelude.

Layer(Vec<LayerName>)

A @layer rule prelude.

Scope(ScopeBounds)

A @scope rule prelude.

StartingStyle

A @starting-style prelude.

AppearanceBase

A @appearance-base prelude (UA sheets only).

PositionTry(DashedIdent)

A @position-try prelude for Anchor Positioning.

CustomMedia(DashedIdent, CustomMediaCondition)

A @custom-media prelude.

ViewTransition

A @view-transition prelude.

Implementations

impl AtRulePrelude

fn name(&self) -> &'static str