style::stylesheets::rules_iterator

Type Alias EffectiveRulesIterator

pub type EffectiveRulesIterator<'a, 'b, CMM> = RulesIterator<'a, 'b, EffectiveRules, CMM>;

An iterator over all the effective rules of a stylesheet.

NOTE: This iterator recurses into @import rules.

Aliased Type

pub struct EffectiveRulesIterator<'a, 'b, CMM> {
    device: &'a Device,
    quirks_mode: QuirksMode,
    custom_media: CMM,
    guard: &'a SharedRwLockReadGuard<'b>,
    stack: SmallVec<[Iter<'a, CssRule>; 3]>,
    last_rule_had_children: bool,
    _phantom: PhantomData<EffectiveRules>,
}

Fields
device: &'a Devicequirks_mode: QuirksModecustom_media: CMMguard: &'a SharedRwLockReadGuard<'b>stack: SmallVec<[Iter<'a, CssRule>; 3]>last_rule_had_children: bool_phantom: PhantomData<EffectiveRules>
Implementations

impl<'a, 'b, CMM> EffectiveRulesIterator<'a, 'b, CMM>
where CMM: Deref<Target = CustomMediaMap>,

pub fn effective_children( device: &'a Device, quirks_mode: QuirksMode, custom_media_map: CMM, guard: &'a SharedRwLockReadGuard<'b>, rule: &'a CssRule, ) -> Self

Returns an iterator over the effective children of a rule, even if rule itself is not effective.