style::stylesheets

Module scope_rule

A @scope rule.

Structs

ScopeBounds

Bounds of the scope.

ScopeRootCandidate

A scope root candidate.

ScopeRule

A scoped rule.

ScopeSubjectMap

A map containing simple selectors in subjects of scope selectors. This allows fast-rejecting scopes before running the full match.

Enums

ImplicitScopeRoot

Types of implicit scope root.

ImplicitScopeTarget

Target of this implicit scope.

ScopeTarget

Target of this scope.

Functions

collect_scope_roots

Collect potential scope roots for a given element and its scope target. The check may not pass the ceiling, if specified.

element_is_outside_of_scope

Given the scope-end selector, check if the element is outside of the scope. That is, check if any ancestor to the root matches the scope-end selector.

parse_scope 🔒

scope_selector_list_is_trivial

Determine if this selector list, when used as a scope bound selector, is considered trivial.