style::stylesheets::scope_rule

Function collect_scope_roots

pub fn collect_scope_roots<E>(
    element: E,
    ceiling: Option<OpaqueElement>,
    context: &mut MatchingContext<'_, E::Impl>,
    target: &ScopeTarget<'_>,
    matches_shadow_host: bool,
    scope_subject_map: &ScopeSubjectMap,
) -> Vec<ScopeRootCandidate>
where
    E: TElement,

Collect potential scope roots for a given element and its scope target. The check may not pass the ceiling, if specified.