Struct style::stylesheets::Stylesheet

pub struct Stylesheet {
    pub rules: Vec<CSSRule>,
    pub media: Option<MediaQueryList>,
    pub origin: Origin,
    pub dirty_on_viewport_size_change: bool,
}

Fields
rules: Vec<CSSRule>

List of rules in the order they were found (important for cascading order)

media: Option<MediaQueryList>

List of media associated with the Stylesheet, if any.

origin: Origin dirty_on_viewport_size_change: bool
Methods

impl Stylesheet

fn from_bytes_iter<I: Iterator<Item=Vec<u8>>>(input: I,
base_url: Url,
protocol_encoding_label: Option<&str>,
environment_encoding: Option<EncodingRef>,
origin: Origin,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)
-> Stylesheet

fn from_bytes(bytes: &[u8],
base_url: Url,
protocol_encoding_label: Option<&str>,
environment_encoding: Option<EncodingRef>,
origin: Origin,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)
-> Stylesheet

fn from_str(css: &str,
base_url: Url,
origin: Origin,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)
-> Stylesheet

fn set_media(&mut self, media: Option<MediaQueryList>)

Set the MediaQueryList associated with the style-sheet.

fn is_effective_for_device(&self, device: &Device) -> bool

Returns whether the style-sheet applies for the current device depending on the associated MediaQueryList.

Always true if no associated MediaQueryList exists.

fn effective_rules<F>(&self, device: &Device, f: F) where F: FnMut(&CSSRule)

Return an iterator over the effective rules within the style-sheet, as according to the supplied Device.

If a condition does not hold, its associated conditional group rule and nested rules will be skipped. Use rules if all rules need to be examined.

impl Stylesheet

fn effective_style_rules<F>(&self, device: &Device, f: F) where F: FnMut(&StyleRule)

fn effective_media_rules<F>(&self, device: &Device, f: F) where F: FnMut(&MediaRule)

fn effective_font_face_rules<F>(&self, device: &Device, f: F) where F: FnMut(&FontFaceRule)

fn effective_viewport_rules<F>(&self, device: &Device, f: F) where F: FnMut(&ViewportRule)

fn effective_keyframes_rules<F>(&self, device: &Device, f: F) where F: FnMut(&KeyframesRule)