style::stylesheets::stylesheet

Trait StylesheetInDocument

pub trait StylesheetInDocument: Debug {
    // Required methods
    fn enabled(&self) -> bool;
    fn media<'a>(
        &'a self,
        guard: &'a SharedRwLockReadGuard<'_>,
    ) -> Option<&'a MediaList>;
    fn contents<'a>(
        &'a self,
        guard: &'a SharedRwLockReadGuard<'_>,
    ) -> &'a StylesheetContents;
    fn implicit_scope_root(&self) -> Option<ImplicitScopeRoot>;

    // Provided method
    fn is_effective_for_device(
        &self,
        device: &Device,
        custom_media: &CustomMediaMap,
        guard: &SharedRwLockReadGuard<'_>,
    ) -> bool { ... }
}

A trait to represent a given stylesheet in a document.

Required Methods

fn enabled(&self) -> bool

Get whether this stylesheet is enabled.

fn media<'a>( &'a self, guard: &'a SharedRwLockReadGuard<'_>, ) -> Option<&'a MediaList>

Get the media associated with this stylesheet.

fn contents<'a>( &'a self, guard: &'a SharedRwLockReadGuard<'_>, ) -> &'a StylesheetContents

Returns a reference to the contents of the stylesheet.

fn implicit_scope_root(&self) -> Option<ImplicitScopeRoot>

Return the implicit scope root for this stylesheet, if one exists.

Provided Methods

fn is_effective_for_device( &self, device: &Device, custom_media: &CustomMediaMap, guard: &SharedRwLockReadGuard<'_>, ) -> bool

Returns whether the style-sheet applies for the current device.