Type Definition style::stylesheets::EffectiveRulesIterator

pub type EffectiveRulesIterator<'a, 'b> = RulesIterator<'a, 'b, EffectiveRules>;

An iterator over all the effective rules of a stylesheet.

NOTE: This iterator recurses into @import rules.

Implementations

impl<'a, 'b> EffectiveRulesIterator<'a, 'b>

pub fn effective_children(
device: &'a Device,
quirks_mode: QuirksMode,
guard: &'a SharedRwLockReadGuard<'b>,
rule: &'a CssRule
) -> Self

Returns an iterator over the effective children of a rule, even if rule itself is not effective.