style::stylist

Struct DocumentCascadeData

pub struct DocumentCascadeData {
    user_agent: Arc<UserAgentCascadeData>,
    user: CascadeData,
    author: CascadeData,
    per_origin: PerOrigin<()>,
}

All the computed information for all the stylesheets that apply to the document.

Fields
user_agent: Arc<UserAgentCascadeData>user: CascadeDataauthor: CascadeDataper_origin: PerOrigin<()>
Implementations

impl DocumentCascadeData

pub fn borrow_for_origin(&self, origin: Origin) -> &CascadeData

Borrows the cascade data for a given origin.

fn iter_origins(&self) -> DocumentCascadeDataIter<'_>

fn iter_origins_rev(&self) -> DocumentCascadeDataIter<'_>

fn custom_media_for_sheet( &self, s: &StylistSheet, guard: &SharedRwLockReadGuard<'_>, ) -> &CustomMediaMap

fn rebuild<'a, S>( &mut self, device: &Device, quirks_mode: QuirksMode, flusher: DocumentStylesheetFlusher<'a, S>, guards: &StylesheetGuards<'_>, difference: &mut CascadeDataDifference, ) -> Result<(), AllocErr>
where S: StylesheetInDocument + PartialEq + 'static,

Rebuild the cascade data for the given document stylesheets, and optionally with a set of user agent stylesheets. Returns Err(..) to signify OOM.