Struct style::stylist::GenericElementAndPseudoRules

struct GenericElementAndPseudoRules<Map> {
    element_map: Map,
    pseudos_map: PerPseudoElementMap<Box<Map>>,
}

A set of rules for element and pseudo-elements.

Fields
element_map: Map

Rules from stylesheets at this CascadeData's origin.

pseudos_map: PerPseudoElementMap<Box<Map>>

Rules from stylesheets at this CascadeData's origin that correspond to a given pseudo-element.

FIXME(emilio): There are a bunch of wasted entries here in practice. Figure out a good way to do a PerNonAnonBox and PerAnonBox (for precomputed_values_for_pseudo) without duplicating a lot of code.

Implementations

impl<Map: Default + MallocSizeOf> GenericElementAndPseudoRules<Map>

fn for_insertion(&mut self, pseudo_element: Option<&PseudoElement>) -> &mut Map

fn rules(&self, pseudo: Option<&PseudoElement>) -> Option<&Map>

impl GenericElementAndPseudoRules<SelectorMap<Rule>>

fn clear(&mut self)

impl GenericElementAndPseudoRules<PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>

fn clear(&mut self)