Struct style::stylist::Stylist

pub struct Stylist {
    pub device: Arc<Device>,
    viewport_constraints: Option<ViewportConstraints>,
    quirks_mode: bool,
    is_device_dirty: bool,
    element_map: PerPseudoElementSelectorMap,
    pub rule_tree: RuleTree,
    pseudos_map: FnvHashMap<PseudoElement, PerPseudoElementSelectorMap>,
    animations: FnvHashMap<Atom, KeyframesAnimation>,
    precomputed_pseudo_element_decls: FnvHashMap<PseudoElement, Vec<ApplicableDeclarationBlock>>,
    rules_source_order: usize,
    state_deps: DependencySet,
    sibling_affecting_selectors: Vec<Selector<SelectorImpl>>,
    non_common_style_affecting_attributes_selectors: Vec<Selector<SelectorImpl>>,
}

This structure holds all the selectors and device characteristics for a given document. The selectors are converted into Rules (defined in rust-selectors), and introduced in a SelectorMap depending on the pseudo-element (see PerPseudoElementSelectorMap), and stylesheet origin (see the fields of PerPseudoElementSelectorMap).

This structure is effectively created once per pipeline, in the LayoutThread corresponding to that pipeline.

Fields
device: Arc<Device>

Device that the stylist is currently evaluating against.

This field deserves a bigger comment due to the different use that Gecko and Servo give to it (that we should eventually unify).

With Gecko, the device is never changed. Gecko manually tracks whether the device data should be reconstructed, and "resets" the state of the device.

On Servo, on the other hand, the device is a really cheap representation that is recreated each time some constraint changes and calling set_device.

In both cases, the device is actually owned by the Stylist, and it's only an Arc so we can implement add_stylesheet more idiomatically.

viewport_constraints: Option<ViewportConstraints>

Viewport constraints based on the current device.

quirks_mode: bool

If true, the quirks-mode stylesheet is applied.

is_device_dirty: bool

If true, the device has changed, and the stylist needs to be updated.

element_map: PerPseudoElementSelectorMap

The current selector maps, after evaluating media rules against the current device.

rule_tree: RuleTree

The rule tree, that stores the results of selector matching.

FIXME(emilio): Not pub!

pseudos_map: FnvHashMap<PseudoElement, PerPseudoElementSelectorMap>

The selector maps corresponding to a given pseudo-element (depending on the implementation)

animations: FnvHashMap<Atom, KeyframesAnimation>

A map with all the animations indexed by name.

precomputed_pseudo_element_decls: FnvHashMap<PseudoElement, Vec<ApplicableDeclarationBlock>>

Applicable declarations for a given non-eagerly cascaded pseudo-element. These are eagerly computed once, and then used to resolve the new computed values on the fly on layout.

FIXME(emilio): Use the rule tree!

rules_source_order: usize

A monotonically increasing counter to represent the order on which a style rule appears in a stylesheet, needed to sort them by source order.

state_deps: DependencySet

Selector dependencies used to compute restyle hints.

sibling_affecting_selectors: Vec<Selector<SelectorImpl>>

Selectors in the page affecting siblings

non_common_style_affecting_attributes_selectors: Vec<Selector<SelectorImpl>>

Selectors in the page matching elements with non-common style-affecting attributes.

Methods

impl Stylist

fn new(device: Device) -> Self

Construct a new Stylist, using a given Device.

fn update(&mut self,
doc_stylesheets: &[Arc<Stylesheet>],
ua_stylesheets: Option<&UserAgentStylesheets>,
stylesheets_changed: bool)
-> bool

Update the stylist for the given document stylesheets, and optionally with a set of user agent stylesheets.

This method resets all the style data each time the stylesheets change (which is indicated by the stylesheets_changed parameter), or the device is dirty, which means we need to re-evaluate media queries.

fn add_stylesheet(&mut self, stylesheet: &Stylesheet)

fn precomputed_values_for_pseudo(&self,
pseudo: &PseudoElement,
parent: Option<&Arc<ComputedValues>>,
default: &Arc<ComputedValues>,
inherit_all: bool)
-> ComputedStyle

Computes the style for a given "precomputed" pseudo-element, taking the universal rules and applying them.

If inherit_all is true, then all properties are inherited from the parent; otherwise, non-inherited properties are reset to their initial values. The flow constructor uses this flag when constructing anonymous flows.

fn style_for_anonymous_box(&self,
pseudo: &PseudoElement,
parent_style: &Arc<ComputedValues>,
default_style: &Arc<ComputedValues>)
-> Arc<ComputedValues>

Returns the style for an anonymous box of the given type.

fn lazily_compute_pseudo_element_style<E>(&self,
element: &E,
pseudo: &PseudoElement,
parent: &Arc<ComputedValues>,
default: &Arc<ComputedValues>)
-> Option<ComputedStyle> where E: TElement + Debug + PresentationalHintsSynthetizer

Computes a pseudo-element style lazily during layout.

This can only be done for a certain set of pseudo-elements, like :selection.

Check the documentation on lazy pseudo-elements in docs/components/style.md

fn set_device(&mut self, device: Device, stylesheets: &[Arc<Stylesheet>])

Set a given device, which may change the styles that apply to the document.

This means that we may need to rebuild style data even if the stylesheets haven't changed.

Also, the device that arrives here may need to take the viewport rules into account.

TODO(emilio): Probably should be unified with update, right now I don't think we take into account dynamic updates to viewport rules.

Probably worth to make the stylist own a single Device, and have a update_device function?

feature = "servo" because gecko only has one device, and manually tracks when the device is dirty.

FIXME(emilio): The semantics of the device for Servo and Gecko are different enough we may want to unify them.

fn viewport_constraints(&self) -> Option<&ViewportConstraints>

Returns the viewport constraints that apply to this document because of a @viewport rule.

fn set_quirks_mode(&mut self, enabled: bool)

Sets the quirks mode of the document.

fn push_applicable_declarations<E, V>(&self,
element: &E,
parent_bf: Option<&BloomFilter>,
style_attribute: Option<&Arc<RwLock<PropertyDeclarationBlock>>>,
animation_rules: AnimationRules,
pseudo_element: Option<&PseudoElement>,
applicable_declarations: &mut V,
flags: &mut ElementSelectorFlags)
-> StyleRelations where E: TElement + Debug + PresentationalHintsSynthetizer,
V: Push<ApplicableDeclarationBlock> + VecLike<ApplicableDeclarationBlock>

Returns the applicable CSS declarations for the given element.

This corresponds to ElementRuleCollector in WebKit.

The returned StyleRelations indicate hints about which kind of rules have matched.

fn is_device_dirty(&self) -> bool

Return whether the device is dirty, that is, whether the screen size or media type have changed (for now).

fn animations(&self) -> &FnvHashMap<Atom, KeyframesAnimation>

Returns the map of registered @keyframes animations.

fn match_same_not_common_style_affecting_attributes_rules<E>(&self,
element: &E,
candidate: &E)
-> bool where E: TElement

Whether two elements match the same not-common style-affecting attribute rules.

This is used to test elements and candidates in the style-sharing candidate cache.

fn rule_tree_root(&self) -> StrongRuleNode

Returns the rule root node.

fn match_same_sibling_affecting_rules<E>(&self,
element: &E,
candidate: &E)
-> bool where E: TElement

Returns whether two elements match the same sibling-affecting rules.

This is also for the style sharing candidate cache.

fn compute_restyle_hint<E>(&self,
element: &E,
snapshot: &Snapshot)
-> RestyleHint where E: TElement

Given an element, and a snapshot that represents a previous state of the element, compute the appropriate restyle hint, that is, the kind of restyle we need to do.