Struct style::stylist::StylistSelectorVisitor

struct StylistSelectorVisitor<'a> {
    passed_rightmost_selector: bool,
    needs_revalidation: &'a mut bool,
    in_selector_list_of: SelectorListKind,
    mapped_ids: &'a mut PrecomputedHashSet<Atom>,
    nth_of_mapped_ids: &'a mut PrecomputedHashSet<Atom>,
    attribute_dependencies: &'a mut PrecomputedHashSet<LocalName>,
    nth_of_class_dependencies: &'a mut PrecomputedHashSet<Atom>,
    nth_of_attribute_dependencies: &'a mut PrecomputedHashSet<LocalName>,
    state_dependencies: &'a mut ElementState,
    nth_of_state_dependencies: &'a mut ElementState,
    document_state_dependencies: &'a mut DocumentState,
}

A selector visitor implementation that collects all the state the Stylist cares about a selector.

Fields
passed_rightmost_selector: bool

Whether we’ve past the rightmost compound selector, not counting pseudo-elements.

needs_revalidation: &'a mut bool

Whether the selector needs revalidation for the style sharing cache.

in_selector_list_of: SelectorListKind

Flags for which selector list-containing components the visitor is inside of, if any

mapped_ids: &'a mut PrecomputedHashSet<Atom>

The filter with all the id’s getting referenced from rightmost selectors.

nth_of_mapped_ids: &'a mut PrecomputedHashSet<Atom>

The filter with the IDs getting referenced from the selector list of :nth-child(… of ) selectors.

attribute_dependencies: &'a mut PrecomputedHashSet<LocalName>

The filter with the local names of attributes there are selectors for.

nth_of_class_dependencies: &'a mut PrecomputedHashSet<Atom>

The filter with the classes getting referenced from the selector list of :nth-child(… of ) selectors.

nth_of_attribute_dependencies: &'a mut PrecomputedHashSet<LocalName>

The filter with the local names of attributes there are selectors for within the selector list of :nth-child(… of ) selectors.

state_dependencies: &'a mut ElementState

All the states selectors in the page reference.

nth_of_state_dependencies: &'a mut ElementState

All the state selectors in the page reference within the selector list of :nth-child(… of ) selectors.

document_state_dependencies: &'a mut DocumentState

All the document states selectors in the page reference.