Type Alias style::stylist::ElementAndPseudoRules

type ElementAndPseudoRules = GenericElementAndPseudoRules<SelectorMap<Rule>>;

Aliased Type

struct ElementAndPseudoRules {
    element_map: SelectorMap<Rule>,
    pseudos_map: PerPseudoElementMap<Box<SelectorMap<Rule>, Global>>,
}

Fields
element_map: SelectorMap<Rule>

Rules from stylesheets at this CascadeData’s origin.

pseudos_map: PerPseudoElementMap<Box<SelectorMap<Rule>, Global>>

Rules from stylesheets at this CascadeData’s origin that correspond to a given pseudo-element.

FIXME(emilio): There are a bunch of wasted entries here in practice. Figure out a good way to do a PerNonAnonBox and PerAnonBox (for precomputed_values_for_pseudo) without duplicating a lot of code.

Implementations

impl<Map: Default + MallocSizeOf> GenericElementAndPseudoRules<Map>

fn for_insertion(&mut self, pseudo_element: Option<&PseudoElement>) -> &mut Map

fn rules(&self, pseudo: Option<&PseudoElement>) -> Option<&Map>

impl GenericElementAndPseudoRules<SelectorMap<Rule>>

fn clear(&mut self)

fn shrink_if_needed(&mut self)