style::stylist

Type Alias PartElementAndPseudoRules

type PartElementAndPseudoRules = GenericElementAndPseudoRules<PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>;

Aliased Type

struct PartElementAndPseudoRules {
    element_map: HashMap<Atom<AtomStaticSet>, SmallVec<[Rule; 1]>, BuildHasherDefault<PrecomputedHasher>>,
    pseudos_map: PerPseudoElementMap<GenericElementAndPseudoRules<HashMap<Atom<AtomStaticSet>, SmallVec<[Rule; 1]>, BuildHasherDefault<PrecomputedHasher>>>>,
}

Fields
element_map: HashMap<Atom<AtomStaticSet>, SmallVec<[Rule; 1]>, BuildHasherDefault<PrecomputedHasher>>

Rules from stylesheets at this CascadeData’s origin.

pseudos_map: PerPseudoElementMap<GenericElementAndPseudoRules<HashMap<Atom<AtomStaticSet>, SmallVec<[Rule; 1]>, BuildHasherDefault<PrecomputedHasher>>>>

Rules from stylesheets at this CascadeData’s origin that correspond to a given pseudo-element.

FIXME(emilio): There are a bunch of wasted entries here in practice. Figure out a good way to do a PerNonAnonBox and PerAnonBox (for precomputed_values_for_pseudo) without duplicating a lot of code.