style

Module thread_state

Supports dynamic assertions about what sort of thread is running and what state it’s in.

Structs

ThreadState

A thread state flag, used for multiple assertions.

Constants

STATE 🔒

Functions

enter

Enters into a given temporary state. Panics if re-entering.

exit

Exits a given temporary state.

get

Gets the current thread state.

initialize

Initializes the current thread state.

initialize_layout_worker_thread

Initializes the current thread as a layout worker thread.