Struct style::timer::Timer

pub struct Timer { /* fields omitted */ }

A Timer struct that takes care of giving the current time for animations.

This is needed to be allowed to hook the time in the animations' test-mode.

Methods

impl Timer

pub fn new() -> Self

Creates a new "normal" timer, i.e., a "Current" mode timer.

pub fn test_mode() -> Self

Creates a new "test mode" timer, with initial time 0.

pub fn seconds(&self) -> f64

Returns the current time, at least from the caller's perspective. In test mode returns whatever the value is.

pub fn increment(&mut self, by: f64)

Increments the current clock. Panics if the clock is not on test mode.