Struct style::traversal::LazyRestyleData

struct LazyRestyleData<'b, E: TElement + 'b, D: DomTraversalContext<E::ConcreteNode>> {
    data: Option<AtomicRefMut<'b, ElementData>>,
    element: &'b E,
    phantom: PhantomData<D>,
}

Various steps in the child preparation algorithm above may cause us to lazily instantiate the ElementData on the child. Encapsulate that logic into a convenient abstraction.

Fields
data: Option<AtomicRefMut<'b, ElementData>> element: &'b E phantom: PhantomData<D>
Methods

impl<'b, E: TElement, D: DomTraversalContext<E::ConcreteNode>> LazyRestyleData<'b, E, D>

unsafe fn new(element: &'b E) -> Self

This may lazily instantiate ElementData, and is therefore only safe to call on an element for which we have exclusive access.

fn ensure(&mut self) -> Option<&mut RestyleData>

fn has_snapshot(&self) -> bool

Checks for the existence of an element snapshot without lazily instantiating anything. This allows the traversal to cheaply pass through already-styled nodes when they don't need a restyle.