Trait style::traversal::DomTraversalContext

pub trait DomTraversalContext<N: TNode> {
    type SharedContext: Sync + 'static;
    fn new<'a>(&'a Self::SharedContext, OpaqueNode) -> Self;
    fn process_preorder(&self, node: N, data: &mut PerLevelTraversalData);
    fn process_postorder(&self, node: N);
    fn should_traverse_child(child: N) -> bool;
    unsafe fn ensure_element_data(element: &N::ConcreteElement)
                              -> &AtomicRefCell<ElementData>;
    unsafe fn clear_element_data(element: &N::ConcreteElement);
    fn local_context(&self) -> &LocalStyleContext;

    fn needs_postorder_traversal(&self) -> bool { ... }
    fn traverse_children<F: FnMut(N)>(parent: N::ConcreteElement, f: F) { ... }
}

Associated Types

type SharedContext: Sync + 'static

Required Methods

fn new<'a>(&'a Self::SharedContext, OpaqueNode) -> Self

fn process_preorder(&self, node: N, data: &mut PerLevelTraversalData)

Process node on the way down, before its children have been processed.

fn process_postorder(&self, node: N)

Process node on the way up, after its children have been processed.

This is only executed if needs_postorder_traversal returns true.

fn should_traverse_child(child: N) -> bool

Returns true if traversal should visit the given child.

unsafe fn ensure_element_data(element: &N::ConcreteElement)
-> &AtomicRefCell<ElementData>

Ensures the existence of the ElementData, and returns it. This can't live on TNode because of the trait-based separation between Servo's script and layout crates.

This is only safe to call in top-down traversal before processing the children of |element|.

unsafe fn clear_element_data(element: &N::ConcreteElement)

Clears the ElementData attached to this element, if any.

This is only safe to call in top-down traversal before processing the children of |element|.

fn local_context(&self) -> &LocalStyleContext

Provided Methods

fn needs_postorder_traversal(&self) -> bool

Boolean that specifies whether a bottom up traversal should be performed.

If it's false, then process_postorder has no effect at all.

fn traverse_children<F: FnMut(N)>(parent: N::ConcreteElement, f: F)

Helper for the traversal implementations to select the children that should be enqueued for processing.