Module style::traversal_flags

Flags that control the traversal process.

We CamelCase rather than UPPER_CASING so that we can grep for the same strings across gecko and servo.

Structs

TraversalFlags
Flags that control the traversal process.

Constants

AnimationOnly
Traverse only elements for animation restyles.

ClearAnimationOnlyDirtyDescendants
Clears the animation-only dirty descendants bit in the subtree.

ClearDirtyBits
Clears all the dirty bits on the elements traversed.

FlushThrottledAnimations
Flush throttled animations. By default, we only update throttled animations when we have other non-throttled work to do. With this flag, we unconditionally tick and process them.

ForCSSRuleChanges
Traverse and update all elements with CSS animations since @keyframes rules may have changed. Triggered by CSS rule changes.

Forgetful
A forgetful traversal ignores the previous state of the frame tree, and thus does not compute damage or maintain other state describing the styles pre-traversal. A forgetful traversal is usually the right thing if you aren't going to do a post-traversal.

ParallelTraversal
Allows the traversal to run in parallel if there are sufficient cores on the machine.

UnstyledOnly
Styles unstyled elements, but does not handle invalidations on already-styled elements.