style

Module typed_om

Typed OM.

https://drafts.css-houdini.org/css-typed-om-1/

Re-exports

pub use numeric_type::NumericType;

Modules

numeric

Typed OM Numeric Value.

numeric_declaration

Typed OM Numeric Declaration.

numeric_type

Typed OM Numeric Type.

sum_value

Typed OM Sum Value.

Macros

impl_to_typed_for_predefined_type 🔒

Structs

KeywordValue

A keyword value used by the Typed OM.

MathClamp

A clamp expression over numeric values.

MathInvert

An inverted numeric value.

MathMax

A maximum expression over numeric values.

MathMin

A minimum expression over numeric values.

MathNegate

A negated numeric value.

MathProduct

A product of numeric values.

MathSum

A sum of numeric values.

MatrixComponent

A matrix transform component used by the Typed OM.

PerspectiveComponent

A perspective transform component used by the Typed OM.

RotateComponent

A rotate transform component used by the Typed OM.

ScaleComponent

A scale transform component used by the Typed OM.

SkewComponent

A skew transform component used by the Typed OM.

TranslateComponent

A translate transform component used by the Typed OM.

TypedValueList

A list of property-agnostic values used by the Typed OM.

UnitValue

A single numeric value with an associated unit.

VariableReferenceValue

A variable reference inside an unparsed Typed OM value.

Enums

ImageValue

An image value used by the Typed OM.

MathValue

A math expression used by the Typed OM.

NumericValue

A numeric value used by the Typed OM.

PerspectiveValue

A perspective value used by a perspective component.

TransformComponent

A single transform component used by the Typed OM.

TypedValue

A property-agnostic representation of a value, used by Typed OM.

UnparsedSegment

A single segment of an unparsed Typed OM value.

Traits

ToTyped

Reifies a value into its Typed OM representation.

Type Aliases

SkewXComponent

A skewX transform component used by the Typed OM.

SkewYComponent

A skewY transform component used by the Typed OM.

TransformValue

A transform value used by the Typed OM.

UnparsedValue

An unparsed value used by the Typed OM.