style::typed_om::numeric

Enum NoCalcNumeric

#[repr(u8)]
pub enum NoCalcNumeric {
    Length(NoCalcLength),
    Angle(NoCalcAngle),
    Time(NoCalcTime),
    Number(NoCalcNumber),
    Percentage(NoCalcPercentage),
}

A numeric value without a calc expression.

Variants

Length(NoCalcLength)

A <length> value.

https://drafts.csswg.org/css-values/#lengths

Angle(NoCalcAngle)

An <angle> value.

https://drafts.csswg.org/css-values/#angles

Time(NoCalcTime)

A <time> value.

https://drafts.csswg.org/css-values/#time

Number(NoCalcNumber)

A <number> value.

https://drafts.csswg.org/css-values/#number-value

Percentage(NoCalcPercentage)

A <percentage> value.

https://drafts.csswg.org/css-values/#percentages

Implementations

impl NoCalcNumeric

pub fn unitless_value(&self) -> CSSFloat

Return the unitless, raw value.

pub fn unit(&self) -> &'static str

Return the unit, as a string.

TODO: Investigate returning SortKey or adding a new variant for returning the unit as SortKey. Tracked in https://bugzilla.mozilla.org/show_bug.cgi?id=2015863

pub fn canonical_unit(&self) -> Option<&'static str>

Return the canonical unit for this value, if one exists.

TODO: Investigate returning SortKey. Tracked in https://bugzilla.mozilla.org/show_bug.cgi?id=2015863

pub fn to(&self, unit: &str) -> Result<Self, ()>

Convert this value to the specified unit, if possible.

TODO: Investigate using SortKey. Tracked in https://bugzilla.mozilla.org/show_bug.cgi?id=2015863

pub fn parse_unit_value(value: CSSFloat, unit: &str) -> Result<Self, ()>

Parse a given unit value.