style::values::animated

Module lists

Lists have various ways of being animated, this module implements them.

See https://drafts.csswg.org/web-animations-1/#animating-properties

Modules

by_computed_value

https://drafts.csswg.org/web-animations-1/#by-computed-value

repeatable_list

https://drafts.csswg.org/web-animations-1/#repeatable-list

with_zero

This is the animation used for some of the types like shadows and filters, where the interpolation happens with the zero value if one of the sides is not present.