Enum style::values::computed::angle::Angle

pub enum Angle {
    Deg(CSSFloat),
    Grad(CSSFloat),
    Rad(CSSFloat),
    Turn(CSSFloat),
}

A computed angle.

Variants
Deg(CSSFloat)

An angle with degree unit.

Grad(CSSFloat)

An angle with gradian unit.

Rad(CSSFloat)

An angle with radian unit.

Turn(CSSFloat)

An angle with turn unit.

Methods

impl Angle

pub fn from_radians(radians: CSSFloat) -> Self

Creates a computed Angle value from a radian amount.

pub fn radians(&self) -> CSSFloat

Returns the amount of radians this angle represents.

pub fn radians64(&self) -> f64

Returns the amount of radians this angle represents as a f64.

Gecko stores angles as singles, but does this computation using doubles. See nsCSSValue::GetAngleValueInRadians. This is significant enough to mess up rounding to the nearest quarter-turn for 225 degrees, for example.