Module style::values::computed::box_

Computed types for box properties.

Re-exports

pub use crate::values::specified::box_::Appearance;

pub use crate::values::specified::box_::BaselineSource;

pub use crate::values::specified::box_::BreakBetween;

pub use crate::values::specified::box_::BreakWithin;

pub use crate::values::specified::box_::Clear as SpecifiedClear;

pub use crate::values::specified::box_::Contain;

pub use crate::values::specified::box_::ContainerName;

pub use crate::values::specified::box_::ContainerType;

pub use crate::values::specified::box_::ContentVisibility;

pub use crate::values::specified::box_::Display;

pub use crate::values::specified::box_::Float as SpecifiedFloat;

pub use crate::values::specified::box_::Overflow;

pub use crate::values::specified::box_::OverflowAnchor;

pub use crate::values::specified::box_::OverflowClipBox;

pub use crate::values::specified::box_::OverscrollBehavior;

pub use crate::values::specified::box_::ScrollSnapAlign;

pub use crate::values::specified::box_::ScrollSnapAxis;

pub use crate::values::specified::box_::ScrollSnapStop;

pub use crate::values::specified::box_::ScrollSnapStrictness;

pub use crate::values::specified::box_::ScrollSnapType;

pub use crate::values::specified::box_::ScrollbarGutter;

pub use crate::values::specified::box_::TouchAction;

pub use crate::values::specified::box_::WillChange;

Structs

Zoom

The computed zoom property value. We store it as a 16-bit fixed point because we need to store it efficiently in the ComputedStyle representation. The assumption being that zooms over 1000 aren’t quite useful.

Enums

Clear

A computed value for the clear property.

Float

A computed value for the float property.

Resize

A computed value for the resize property.

Constants

ZOOM_FRACTION_BITS

We use an unsigned 10.6 fixed-point value (range 0.0 - 1023.984375).

Type Aliases

ContainIntrinsicSize

A computed value for the contain-intrinsic-size property.

LineClamp

A computed value for the line-clamp property.

Perspective

A computed value for the perspective property.

VerticalAlign

A computed value for the vertical-align property.

ZoomFixedPoint

This is an alias which is useful mostly as a cbindgen / C++ inference workaround.