Struct style::values::computed::color::Color

pub struct Color {
    pub color: RGBA,
    pub foreground_ratio: u8,
}

This struct represents a combined color from a numeric color and the current foreground color (currentcolor keyword). Conceptually, the formula is "color * (1 - p) + currentcolor * p" where p is foreground_ratio.

Fields
color: RGBA

RGBA color.

foreground_ratio: u8

The ratio of currentcolor in complex color.

Methods

impl Color

pub fn rgba(rgba: RGBA) -> Color

Returns a numeric color representing the given RGBA value.

pub fn transparent() -> Color

Returns a complex color value representing transparent.

pub fn currentcolor() -> Color

Returns a complex color value representing currentcolor.

pub fn is_numeric(&self) -> bool

Whether it is a numeric color (no currentcolor component).

pub fn is_currentcolor(&self) -> bool

Whether it is a currentcolor value (no numeric color component).

pub fn to_rgba(&self, fg_color: RGBA) -> RGBA

Combine this complex color with the given foreground color into a numeric RGBA color. It currently uses linear blending.