Type Alias style::values::computed::color::ColorPropertyValue

pub type ColorPropertyValue = AbsoluteColor;

The computed value of the color property.

Aliased Type

struct ColorPropertyValue {
    pub components: ColorComponents,
    pub alpha: f32,
    pub color_space: ColorSpace,
    pub flags: ColorFlags,
}

Fields
components: ColorComponents

The 3 components that make up colors in any color space.

alpha: f32

The alpha component of the color.

color_space: ColorSpace

The current color space that the components represent.

flags: ColorFlags

Extra flags used durring serialization of this color.

Implementations

impl AbsoluteColor

pub const TRANSPARENT: Self = _

A fully transparent color in the legacy syntax.

pub const BLACK: Self = _

An opaque black color in the legacy syntax.

pub const WHITE: Self = _

An opaque white color in the legacy syntax.

pub fn new( color_space: ColorSpace, c1: impl Into<ComponentDetails>, c2: impl Into<ComponentDetails>, c3: impl Into<ComponentDetails>, alpha: impl Into<ComponentDetails> ) -> Self

Create a new AbsoluteColor with the given ColorSpace and components.

pub fn into_srgb_legacy(self) -> Self

Convert this color into the sRGB color space and set it to the legacy syntax.

pub fn srgb_legacy(red: u8, green: u8, blue: u8, alpha: f32) -> Self

Create a new AbsoluteColor from rgba legacy syntax values in the sRGB color space.

pub fn raw_components(&self) -> &[f32; 4]

Return all the components of the color in an array. (Includes alpha)

pub fn is_legacy_syntax(&self) -> bool

Returns true if this color is in the legacy color syntax.

pub fn alpha(&self) -> f32

Return the alpha component.

pub fn to_color_space(&self, color_space: ColorSpace) -> Self

Convert this color to the specified color space.