Struct style::values::computed::font::KeywordInfo

pub struct KeywordInfo {
    pub kw: KeywordSize,
    pub factor: f32,
    pub offset: NonNegativeLength,
}

Additional information for keyword-derived font sizes.

Fields
kw: KeywordSize

The keyword used

factor: f32

A factor to be multiplied by the computed size of the keyword

offset: NonNegativeLength

An additional Au offset to add to the kw*factor in the case of calcs

Methods

impl KeywordInfo

pub fn to_computed_value(&self, context: &Context) -> NonNegativeLength

Computes the final size for this font-size keyword, accounting for text-zoom.

pub fn compose(self, factor: f32, offset: NonNegativeLength) -> Self

Given a parent keyword info (self), apply an additional factor/offset to it

pub fn medium() -> Self

KeywordInfo value for font-size: medium