style::values::computed::image

Type Alias Image

pub type Image = GenericImage<Gradient, ComputedUrl, Color, Percentage, Resolution>;

Computed values for an image according to CSS-IMAGES. https://drafts.csswg.org/css-images/#image-values

Aliased Type

#[repr(C, u8)]
pub enum Image {
    None,
    Url(ComputedUrl),
    Gradient(Box<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>>),
    PaintWorklet(Box<PaintWorklet>),
    CrossFade(Box<GenericCrossFade<GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>, GenericColor<Percentage>, Percentage>>),
    ImageSet(Box<GenericImageSet<GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>, Resolution>>),
    LightDark(Box<GenericLightDark<GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>>>),
    Image(Box<GenericColor<Percentage>>),
}

Variants

None

none variant.

Url(ComputedUrl)

A <url()> image.

Gradient(Box<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>>)

A <gradient> image. Gradients are rather large, and not nearly as common as urls, so we box them here to keep the size of this enum sane.

PaintWorklet(Box<PaintWorklet>)

A paint worklet image. https://drafts.css-houdini.org/css-paint-api/

CrossFade(Box<GenericCrossFade<GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>, GenericColor<Percentage>, Percentage>>)

A <cross-fade()> image. Storing this directly inside of GenericImage increases the size by 8 bytes so we box it here and store images directly inside of cross-fade instead of boxing them there.

ImageSet(Box<GenericImageSet<GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>, Resolution>>)

An image-set() function.

LightDark(Box<GenericLightDark<GenericImage<GenericGradient<LineDirection, CSSPixelLength, LengthPercentage, GenericPosition<LengthPercentage, LengthPercentage>, Angle, AngleOrPercentage, GenericColor<Percentage>>, ComputedUrl, GenericColor<Percentage>, Percentage, Resolution>>>)

A light-dark() function.

Image(Box<GenericColor<Percentage>>)

An image(<color>) function.