Type Alias style::values::computed::length::Length

pub type Length = CSSPixelLength;

An alias of computed <length> value.

Aliased Type

struct Length(f32);

Fields
0: f32
Implementations

impl CSSPixelLength

pub fn new(px: CSSFloat) -> Self

Return a new CSSPixelLength.

pub fn normalized(self) -> Self

Returns a normalized (NaN turned to zero) version of this length.

pub fn finite(self) -> Self

Returns a finite (normalized and clamped to float min and max) version of this length.

pub fn scale_by(self, scale: CSSFloat) -> Self

Scale the length by a given amount.

pub fn px(self) -> CSSFloat

Return the containing pixel value.

pub fn to_i32_au(self) -> i32

Return the length with app_unit i32 type.

pub fn abs(self) -> Self

Return the absolute value of this length.

pub fn clamp_to_non_negative(self) -> Self

Return the clamped value of this length.

pub fn min(self, other: Self) -> Self

Returns the minimum between self and other.

pub fn max(self, other: Self) -> Self

Returns the maximum between self and other.

pub fn max_assign(&mut self, other: Self)

Sets self to the maximum between self and other.

pub fn clamp_between_extremums( self, min_size: Self, max_size: Option<Self> ) -> Self

Clamp the value to a lower bound and an optional upper bound.

Can be used for example with min-width and max-width.

pub fn clamp_below_max(self, max_size: Option<Self>) -> Self

Clamp the value to an optional upper bound.

Can be used for example with max-width.