Type Definition style::values::computed::length::NonNegativeLength

type NonNegativeLength = NonNegative<Length>;

A wrapper of Length, whose value must be >= 0.

Methods

impl NonNegativeLength

pub fn new(px: CSSFloat) -> Self

Create a NonNegativeLength.

pub fn zero() -> Self

Return a zero value.

pub fn px(&self) -> CSSFloat

Return the pixel value of |NonNegativeLength|.

pub fn clamp(self) -> Self

Ensures it is non negative

pub fn scale_by(&self, factor: f32) -> Self

Scale this NonNegativeLength. We scale NonNegativeLength by zero if the factor is negative because it doesn't make sense to scale a negative factor on a non-negative length.