style::values::computed::length_percentage

Type Alias CalcNode

pub type CalcNode = GenericCalcNode<ComputedLeaf>;

The computed version of a calc() node for <length-percentage> values.

Aliased Type

#[repr(u8)]
pub enum CalcNode {

    Leaf(ComputedLeaf),
    Negate(Box<GenericCalcNode<ComputedLeaf>>),
    Invert(Box<GenericCalcNode<ComputedLeaf>>),
    Sum(OwnedSlice<GenericCalcNode<ComputedLeaf>>),
    Product(OwnedSlice<GenericCalcNode<ComputedLeaf>>),
    MinMax(OwnedSlice<GenericCalcNode<ComputedLeaf>>, MinMaxOp),
    Clamp {
        min: Box<GenericCalcNode<ComputedLeaf>>,
        center: Box<GenericCalcNode<ComputedLeaf>>,
        max: Box<GenericCalcNode<ComputedLeaf>>,
    },
    Round {
        strategy: RoundingStrategy,
        value: Box<GenericCalcNode<ComputedLeaf>>,
        step: Box<GenericCalcNode<ComputedLeaf>>,
    },
    ModRem {
        dividend: Box<GenericCalcNode<ComputedLeaf>>,
        divisor: Box<GenericCalcNode<ComputedLeaf>>,
        op: ModRemOp,
    },
    Sin(Box<GenericCalcNode<ComputedLeaf>>),
    Cos(Box<GenericCalcNode<ComputedLeaf>>),
    Tan(Box<GenericCalcNode<ComputedLeaf>>),
    Asin(Box<GenericCalcNode<ComputedLeaf>>),
    Acos(Box<GenericCalcNode<ComputedLeaf>>),
    Atan(Box<GenericCalcNode<ComputedLeaf>>),
    Atan2(Box<GenericCalcNode<ComputedLeaf>>, Box<GenericCalcNode<ComputedLeaf>>),
    Pow(Box<GenericCalcNode<ComputedLeaf>>, Box<GenericCalcNode<ComputedLeaf>>),
    Sqrt(Box<GenericCalcNode<ComputedLeaf>>),
    Hypot(OwnedSlice<GenericCalcNode<ComputedLeaf>>),
    Log(Box<GenericCalcNode<ComputedLeaf>>, Optional<Box<GenericCalcNode<ComputedLeaf>>>),
    Exp(Box<GenericCalcNode<ComputedLeaf>>),
    Abs(Box<GenericCalcNode<ComputedLeaf>>),
    Sign(Box<GenericCalcNode<ComputedLeaf>>),
    Progress {
        clamping_mode: ProgressClampingMode,
        value: Box<GenericCalcNode<ComputedLeaf>>,
        start: Box<GenericCalcNode<ComputedLeaf>>,
        end: Box<GenericCalcNode<ComputedLeaf>>,
    },
    Anchor(Box<GenericAnchorFunction<Box<GenericCalcNode<ComputedLeaf>>, Box<GenericAnchorFunctionFallback<ComputedLeaf>>>>),
    AnchorSize(Box<GenericAnchorSizeFunction<Box<GenericAnchorFunctionFallback<ComputedLeaf>>>>),

}

Variants

Leaf(ComputedLeaf)

A leaf node.

Negate(Box<GenericCalcNode<ComputedLeaf>>)

A node that negates its child, e.g. Negate(1) == -1.

Invert(Box<GenericCalcNode<ComputedLeaf>>)

A node that inverts its child, e.g. Invert(10) == 1 / 10 == 0.1. The child must always resolve to a number unit.

Sum(OwnedSlice<GenericCalcNode<ComputedLeaf>>)

A sum node, representing a + b + c where a, b, and c are the arguments.

Product(OwnedSlice<GenericCalcNode<ComputedLeaf>>)

A product node, representing a * b * c where a, b, and c are the arguments.

MinMax(OwnedSlice<GenericCalcNode<ComputedLeaf>>, MinMaxOp)

A min or max function.

Clamp

A clamp() function.

Fields

min: Box<GenericCalcNode<ComputedLeaf>>

The minimum value.

center: Box<GenericCalcNode<ComputedLeaf>>

The central value.

max: Box<GenericCalcNode<ComputedLeaf>>

The maximum value.

Round

A round() function.

Fields

strategy: RoundingStrategy

The rounding strategy.

value: Box<GenericCalcNode<ComputedLeaf>>

The value to round.

step: Box<GenericCalcNode<ComputedLeaf>>

The step value.

ModRem

A mod() or rem() function.

Fields

dividend: Box<GenericCalcNode<ComputedLeaf>>

The dividend calculation.

divisor: Box<GenericCalcNode<ComputedLeaf>>

The divisor calculation.

op: ModRemOp

Is the function mod or rem?

Sin(Box<GenericCalcNode<ComputedLeaf>>)

A sin() function.

Cos(Box<GenericCalcNode<ComputedLeaf>>)

A cos() function.

Tan(Box<GenericCalcNode<ComputedLeaf>>)

A tan() function.

Asin(Box<GenericCalcNode<ComputedLeaf>>)

An asin() function.

Acos(Box<GenericCalcNode<ComputedLeaf>>)

An acos() function.

Atan(Box<GenericCalcNode<ComputedLeaf>>)

An atan() function.

Atan2(Box<GenericCalcNode<ComputedLeaf>>, Box<GenericCalcNode<ComputedLeaf>>)

An atan2() function.

Pow(Box<GenericCalcNode<ComputedLeaf>>, Box<GenericCalcNode<ComputedLeaf>>)

A pow() function.

Sqrt(Box<GenericCalcNode<ComputedLeaf>>)

A sqrt() function.

Hypot(OwnedSlice<GenericCalcNode<ComputedLeaf>>)

A hypot() function

Log(Box<GenericCalcNode<ComputedLeaf>>, Optional<Box<GenericCalcNode<ComputedLeaf>>>)

A log() function.

Exp(Box<GenericCalcNode<ComputedLeaf>>)

An exp() function.

Abs(Box<GenericCalcNode<ComputedLeaf>>)

An abs() function.

Sign(Box<GenericCalcNode<ComputedLeaf>>)

A sign() function.

Progress

A progress() function.

Fields

clamping_mode: ProgressClampingMode

Clamping mode for the result.

value: Box<GenericCalcNode<ComputedLeaf>>

The progress value calculation.

start: Box<GenericCalcNode<ComputedLeaf>>

The progress start calculation.

end: Box<GenericCalcNode<ComputedLeaf>>

The progress end calculation.

Anchor(Box<GenericAnchorFunction<Box<GenericCalcNode<ComputedLeaf>>, Box<GenericAnchorFunctionFallback<ComputedLeaf>>>>)

An anchor() function.

AnchorSize(Box<GenericAnchorSizeFunction<Box<GenericAnchorFunctionFallback<ComputedLeaf>>>>)

An anchor-size() function.