Type Alias style::values::computed::motion::OffsetPosition

pub type OffsetPosition = GenericOffsetPosition<LengthPercentage, LengthPercentage>;

The computed value of offset-position.

Aliased Type

enum OffsetPosition {
    Normal,
    Auto,
    Position(GenericPosition<LengthPercentage, LengthPercentage>),
}

Variants

Normal

The element does not have an offset starting position.

Auto

The offset starting position is the top-left corner of the box.

Position(GenericPosition<LengthPercentage, LengthPercentage>)

The offset starting position is the result of using the to position a 0x0 object area within the box’s containing block.