style::values::computed::position

Trait TryTacticAdjustment

pub(crate) trait TryTacticAdjustment {
    // Required method
    fn try_tactic_adjustment(
        &mut self,
        old_side: PhysicalSide,
        new_side: PhysicalSide,
    );
}

Perform the adjustment of a given value for a given try tactic, as per: https://drafts.csswg.org/css-anchor-position-1/#swap-due-to-a-try-tactic

Required Methods

fn try_tactic_adjustment( &mut self, old_side: PhysicalSide, new_side: PhysicalSide, )

Performs the adjustments necessary given an old side we’re relative to, and a new side we’re relative to.