style::values::computed::position

Type Alias AnchorFunction

pub type AnchorFunction = GenericAnchorFunction<Percentage, Inset>;

The computed value of an anchor() function.

Aliased Type

#[repr(C)]
pub struct AnchorFunction {
    pub target_element: TreeScoped<DashedIdent>,
    pub side: GenericAnchorSide<Percentage>,
    pub fallback: Optional<GenericInset<Percentage, LengthPercentage>>,
}

Fields
target_element: TreeScoped<DashedIdent>

Anchor name of the element to anchor to. If omitted, selects the implicit anchor element. The shadow cascade order of the tree-scoped anchor name associates the name with the host of the originating stylesheet.

side: GenericAnchorSide<Percentage>

Where relative to the target anchor element to position the anchored element to.

fallback: Optional<GenericInset<Percentage, LengthPercentage>>

Value to use in case the anchor function is invalid.