style::values::computed

Struct Context

pub struct Context<'a> {

    pub builder: StyleBuilder<'a>,
    pub cached_system_font: Option<()>,
    pub in_media_query: bool,
    pub in_container_query: bool,
    pub quirks_mode: QuirksMode,
    pub for_animation: bool,
    pub container_info: Option<ContainerInfo>,
    pub for_non_inherited_property: bool,
    pub rule_cache_conditions: RefCell<&'a mut RuleCacheConditions>,
    pub scope: CascadeLevel,
    pub included_cascade_flags: RuleCascadeFlags,
    container_size_query: RefCell<ContainerSizeQuery<'a>>,
    element_context: &'a dyn ElementContext,
    pub tree_counting_caches: RefCell<&'a mut TreeCountingCaches>,

}

A Context is all the data a specified value could ever need to compute itself and be transformed to a computed value.

Fields
builder: StyleBuilder<'a>

Values accessed through this need to be in the properties “computed early”: color, text-decoration, font-size, display, position, float, border-*-style, outline-style, font-family, writing-mode…

cached_system_font: Option<()>

A dummy option for servo so initializing a computed::Context isn’t painful.

TODO(emilio): Make constructors for Context, and drop this.

in_media_query: bool

Whether or not we are computing the media list in a media query.

in_container_query: bool

Whether or not we are computing the container query condition.

quirks_mode: QuirksMode

The quirks mode of this context.

for_animation: bool

Whether this computation is being done for animation.

Allows opacity to interpolate out-of-range values

container_info: Option<ContainerInfo>

Returns the container information to evaluate a given container query.

for_non_inherited_property: bool

Whether we’re computing a value for a non-inherited property. False if we are computed a value for an inherited property or not computing for a property at all (e.g. in a media query evaluation).

rule_cache_conditions: RefCell<&'a mut RuleCacheConditions>

The conditions to cache a rule node on the rule cache.

FIXME(emilio): Drop the refcell.

scope: CascadeLevel

The cascade level in the shadow tree hierarchy.

included_cascade_flags: RuleCascadeFlags

The set of RuleCascadeFlags whose rules should be included during the cascade. STARTING_STYLE is set from the caller for re-cascade. APPEARANCE_BASE is added dynamically after the appearance property is resolved to a non-None value.

container_size_query: RefCell<ContainerSizeQuery<'a>>

Container size query for this context.

element_context: &'a dyn ElementContext

Element context for the element being styled, used to resolve tree-counting functions (sibling-index(), sibling-count()) and attribute values (attr()).

tree_counting_caches: RefCell<&'a mut TreeCountingCaches>

Caches for evaluation of tree-counting functions.

Implementations

impl<'a> Context<'a>

pub fn get_container_size_query(&self) -> ContainerSizeQueryResult

Lazily evaluate the container size query, returning the result.

pub fn for_media_query_evaluation<F, R>( device: &Device, quirks_mode: QuirksMode, f: F, ) -> R
where F: FnOnce(&Context<'_>) -> R,

Creates a suitable context for media query evaluation, in which font-relative units compute against the system_font, and executes f with it.

pub fn for_container_query_evaluation<F, R>( device: &Device, stylist: Option<&Stylist>, container_info_and_style: Option<(ContainerInfo, Arc<ComputedValues>)>, container_size_query: ContainerSizeQuery<'_>, element_context: &dyn ElementContext, f: F, ) -> R
where F: FnOnce(&Context<'_>) -> R,

Creates a suitable context for container query evaluation for the style specified.

pub fn new( builder: StyleBuilder<'a>, quirks_mode: QuirksMode, rule_cache_conditions: &'a mut RuleCacheConditions, container_size_query: ContainerSizeQuery<'a>, included_cascade_flags: RuleCascadeFlags, element_context: &'a dyn ElementContext, tree_counting_caches: &'a mut TreeCountingCaches, ) -> Self

Creates a context suitable for more general cases.

pub fn new_for_animation( builder: StyleBuilder<'a>, quirks_mode: QuirksMode, rule_cache_conditions: &'a mut RuleCacheConditions, container_size_query: ContainerSizeQuery<'a>, element_context: &'a dyn ElementContext, tree_counting_caches: &'a mut TreeCountingCaches, ) -> Self

Creates a context suitable for computing animations.

pub fn new_for_initial_at_property_value( stylist: &'a Stylist, rule_cache_conditions: &'a mut RuleCacheConditions, tree_counting_caches: &'a mut TreeCountingCaches, ) -> Self

Creates a context suitable for computing the initial value of @property.

pub fn device(&self) -> &Device

The current device.

pub fn inherited_custom_properties(&self) -> &ComputedCustomProperties

Get the inherited custom properties map.

pub fn is_root_element(&self) -> bool

Whether the style is for the root element.

pub fn query_font_metrics( &self, base_size: FontBaseSize, orientation: FontMetricsOrientation, flags: QueryFontMetricsFlags, ) -> FontMetrics

Queries font metrics.

pub fn viewport_size_for_viewport_unit_resolution( &self, variant: ViewportVariant, ) -> Size2D<Au>

The current viewport size, used to resolve viewport units.

fn resolve_tree_counting_result(&self) -> TreeCountingResult

pub fn query_sibling_count(&self) -> u32

Returns the number of siblings of the element that matches the declaration (including the element itself). Also marks the style as depending on sibling-count().

pub fn query_sibling_index(&self) -> u32

Returns the 1-based index of the element that matches the declaration amongst its siblings. Also marks the style as depending on sibling-index().

pub fn in_media_or_container_query(&self) -> bool

Whether we’re in a media or container query.

pub fn default_style(&self) -> &ComputedValues

The default computed style we’re getting our reset style from.

pub fn style(&self) -> &StyleBuilder<'a>

The current style.

pub fn current_scope(&self) -> CascadeLevel

The current tree scope.

pub fn maybe_zoom_text(&self, size: CSSPixelLength) -> CSSPixelLength

Apply text-zoom if enabled.