Struct style::values::computed::text::TextOverflow

pub struct TextOverflow {
    pub first: TextOverflowSide,
    pub second: TextOverflowSide,
    pub sides_are_logical: bool,
}

text-overflow. When the specified value only has one side, that's the "second" side, and the sides are logical, so "second" means "end". The start side is Clip in that case.

When the specified value has two sides, those are our "first" and "second" sides, and they are physical sides ("left" and "right").

Fields
first: TextOverflowSide

First side

second: TextOverflowSide

Second side

sides_are_logical: bool

True if the specified value only has one side.

Methods

impl TextOverflow

pub fn get_initial_value() -> TextOverflow

Returns the initial text-overflow value