style::values::computed::text

Type Alias HyphenateLimitChars

pub type HyphenateLimitChars = GenericHyphenateLimitChars<CSSInteger>;

A computed value for the hyphenate-character property.

Aliased Type

#[repr(C)]
pub struct HyphenateLimitChars {
    pub total_word_length: NumberOrAuto<i32>,
    pub pre_hyphen_length: NumberOrAuto<i32>,
    pub post_hyphen_length: NumberOrAuto<i32>,
}

Fields
total_word_length: NumberOrAuto<i32>

Required minimum number of characters in a hyphenated word.

pre_hyphen_length: NumberOrAuto<i32>

Required minumum number of characters before the hyphen.

post_hyphen_length: NumberOrAuto<i32>

Required minumum number of characters after the hyphen.

Implementations

impl HyphenateLimitChars

pub fn auto() -> Self

Return the auto value, which has all three component values as auto.