Type Alias style::values::computed::text::WordSpacing

pub type WordSpacing = LengthPercentage;

A computed value for the word-spacing property.

Aliased Type

struct WordSpacing(LengthPercentageUnion);

Fields
0: LengthPercentageUnion
Implementations

impl WordSpacing

pub fn normal() -> Self

Return the normal computed value, which is just zero.

impl LengthPercentage

pub fn one() -> Self

1px length value for SVG defaults

pub fn zero_percent() -> Self

0%

fn to_calc_node(&self) -> Cow<'_, CalcNode>

pub fn new_length(length: Length) -> Self

Constructs a length value.

pub fn new_percent(percentage: Percentage) -> Self

Constructs a percentage value.

pub fn hundred_percent_minus(v: Self, clamping_mode: AllowedNumericType) -> Self

Given a LengthPercentage value v, construct the value representing calc(100% - v).

pub fn hundred_percent_minus_list( list: &[&Self], clamping_mode: AllowedNumericType ) -> Self

Given a list of LengthPercentage values, construct the value representing calc(100% - the sum of the list).

pub fn new_calc(node: CalcNode, clamping_mode: AllowedNumericType) -> Self

Constructs a calc() value.

fn new_calc_unchecked(calc: Box<CalcLengthPercentage>) -> Self

Private version of new_calc() that constructs a calc() variant without checking.

fn tag(&self) -> Tag

fn unpack_mut<'a>(&'a mut self) -> UnpackedMut<'a>

fn unpack<'a>(&'a self) -> Unpacked<'a>

unsafe fn calc_ptr(&self) -> *mut CalcLengthPercentage

fn to_serializable(&self) -> Serializable

fn from_serializable(s: Serializable) -> Self

pub fn is_definitely_zero(&self) -> bool

Returns true if the computed value is absolute 0 or 0%.

pub fn resolve(&self, basis: Length) -> Length

Resolves the percentage.

pub fn percentage_relative_to(&self, basis: Length) -> Length

Resolves the percentage. Just an alias of resolve().

pub fn has_percentage(&self) -> bool

Return whether there’s any percentage in this value.

pub fn to_length(&self) -> Option<Length>

Converts to a <length> if possible.

pub fn to_percentage(&self) -> Option<Percentage>

Converts to a <percentage> if possible.

pub fn to_used_value(&self, containing_length: Au) -> Au

Returns the used value.

pub fn to_pixel_length(&self, containing_length: Au) -> Length

Returns the used value as CSSPixelLength.

pub fn maybe_to_used_value(&self, container_len: Option<Length>) -> Option<Au>

Convert the computed value into used value.

pub fn maybe_percentage_relative_to( &self, container_len: Option<Length> ) -> Option<Length>

If there are special rules for computing percentages in a value (e.g. the height property), they apply whenever a calc() expression contains percentages.

pub fn clamp_to_non_negative(self) -> Self

Returns the clamped non-negative values.