style::values::computed

Type Alias TrackSize

pub type TrackSize = GenericTrackSize<LengthPercentage>;

The computed value of a grid <track-size>

Aliased Type

#[repr(C, u8)]
pub enum TrackSize {
    Breadth(GenericTrackBreadth<LengthPercentage>),
    Minmax(GenericTrackBreadth<LengthPercentage>, GenericTrackBreadth<LengthPercentage>),
    FitContent(GenericTrackBreadth<LengthPercentage>),
}

Variants

Breadth(GenericTrackBreadth<LengthPercentage>)

A flexible <track-breadth>

Minmax(GenericTrackBreadth<LengthPercentage>, GenericTrackBreadth<LengthPercentage>)

A minmax function for a range over an inflexible <track-breadth> and a flexible <track-breadth>

https://drafts.csswg.org/css-grid/#valdef-grid-template-columns-minmax

FitContent(GenericTrackBreadth<LengthPercentage>)

A fit-content function.

This stores a TrackBreadth for convenience, but it can only be a LengthPercentage.

https://drafts.csswg.org/css-grid/#valdef-grid-template-columns-fit-content