style::values::generics

Module box_

Generic types for box properties.

Re-exports

pub use self::GenericBaselineShift as BaselineShift;

pub use self::GenericContainIntrinsicSize as ContainIntrinsicSize;

pub use self::GenericMaxLines as MaxLines;

pub use self::GenericLineClamp as LineClamp;

pub use self::GenericPerspective as Perspective;

pub use self::GenericOverflowClipMargin as OverflowClipMargin;

Structs

GenericLineClamp

A generic value for the line-clamp property.

GenericMaxLines

A generic value for the maximum lines used by line-clamp. https://drafts.csswg.org/css-overflow-4/#propdef-max-lines

GenericOverflowClipMargin

https://drafts.csswg.org/css-overflow-4/#overflow-clip-margin

Enums

BaselineShiftKeyword

BlockEllipsis

A block ellipsis for line clamping.

GenericBaselineShift

A generic value for the baseline-shift property. https://drafts.csswg.org/css-inline-3/#baseline-shift

GenericContainIntrinsicSize

https://drafts.csswg.org/css-sizing-4/#intrinsic-size-override

GenericPerspective

A generic value for the perspective property.

MaxLinesKeyword

A keyword for the maximum lines used by line-clamp.

OverflowClipMarginBox

https://drafts.csswg.org/css-overflow-4/#overflow-clip-margin’s . Note that the spec has special behavior for the omitted keyword, but that’s rather odd, see: https://github.com/w3c/csswg-drafts/issues/13185

PositionProperty