style::values::generics::calc

Enum GenericCalcNode

#[repr(u8)]
pub enum GenericCalcNode<L> {

    Leaf(L),
    Negate(Box<Self>),
    Invert(Box<Self>),
    Sum(OwnedSlice<Self>),
    Product(OwnedSlice<Self>),
    MinMax(OwnedSlice<Self>, MinMaxOp),
    Clamp {
        min: Box<Self>,
        center: Box<Self>,
        max: Box<Self>,
    },
    Round {
        strategy: RoundingStrategy,
        value: Box<Self>,
        step: Box<Self>,
    },
    ModRem {
        dividend: Box<Self>,
        divisor: Box<Self>,
        op: ModRemOp,
    },
    Sin(Box<Self>),
    Cos(Box<Self>),
    Tan(Box<Self>),
    Asin(Box<Self>),
    Acos(Box<Self>),
    Atan(Box<Self>),
    Atan2(Box<Self>, Box<Self>),
    Pow(Box<Self>, Box<Self>),
    Sqrt(Box<Self>),
    Hypot(OwnedSlice<Self>),
    Log(Box<Self>, Optional<Box<Self>>),
    Exp(Box<Self>),
    Abs(Box<Self>),
    Sign(Box<Self>),
    Progress {
        clamping_mode: ProgressClampingMode,
        value: Box<Self>,
        start: Box<Self>,
        end: Box<Self>,
    },
    Anchor(Box<GenericCalcAnchorFunction<L>>),
    AnchorSize(Box<GenericCalcAnchorSizeFunction<L>>),

}

A generic node in a calc expression.

FIXME: This would be much more elegant if we used Self in the types below, but we can’t because of https://github.com/serde-rs/serde/issues/1565.

FIXME: The following annotations are to workaround an LLVM inlining bug, see bug 1631929.

cbindgen:destructor-attributes=MOZ_NEVER_INLINE cbindgen:copy-constructor-attributes=MOZ_NEVER_INLINE cbindgen:eq-attributes=MOZ_NEVER_INLINE

Variants

Leaf(L)

A leaf node.

Negate(Box<Self>)

A node that negates its child, e.g. Negate(1) == -1.

Invert(Box<Self>)

A node that inverts its child, e.g. Invert(10) == 1 / 10 == 0.1. The child must always resolve to a number unit.

Sum(OwnedSlice<Self>)

A sum node, representing a + b + c where a, b, and c are the arguments.

Product(OwnedSlice<Self>)

A product node, representing a * b * c where a, b, and c are the arguments.

MinMax(OwnedSlice<Self>, MinMaxOp)

A min or max function.

Clamp

A clamp() function.

Fields

min: Box<Self>

The minimum value.

center: Box<Self>

The central value.

max: Box<Self>

The maximum value.

Round

A round() function.

Fields

strategy: RoundingStrategy

The rounding strategy.

value: Box<Self>

The value to round.

step: Box<Self>

The step value.

ModRem

A mod() or rem() function.

Fields

dividend: Box<Self>

The dividend calculation.

divisor: Box<Self>

The divisor calculation.

op: ModRemOp

Is the function mod or rem?

Sin(Box<Self>)

A sin() function.

Cos(Box<Self>)

A cos() function.

Tan(Box<Self>)

A tan() function.

Asin(Box<Self>)

An asin() function.

Acos(Box<Self>)

An acos() function.

Atan(Box<Self>)

An atan() function.

Atan2(Box<Self>, Box<Self>)

An atan2() function.

Pow(Box<Self>, Box<Self>)

A pow() function.

Sqrt(Box<Self>)

A sqrt() function.

Hypot(OwnedSlice<Self>)

A hypot() function

Log(Box<Self>, Optional<Box<Self>>)

A log() function.

Exp(Box<Self>)

An exp() function.

Abs(Box<Self>)

An abs() function.

Sign(Box<Self>)

A sign() function.

Progress

A progress() function.

Fields

clamping_mode: ProgressClampingMode

Clamping mode for the result.

value: Box<Self>

The progress value calculation.

start: Box<Self>

The progress start calculation.

end: Box<Self>

The progress end calculation.

Anchor(Box<GenericCalcAnchorFunction<L>>)

An anchor() function.

AnchorSize(Box<GenericCalcAnchorSizeFunction<L>>)

An anchor-size() function.

Implementations

impl<L: CalcNodeLeaf> CalcNode<L>

fn dummy() -> Self

Create a dummy CalcNode that can be used to do replacements of other nodes.

fn coerce_to_value(&mut self, value: f32) -> Result<(), ()>

Change all the leaf nodes to have the given value. This is useful when you have calc(1px * nan) and you want to replace the product node with calc(nan), in which case the unit will be retained.

pub fn is_product_distributive(&self) -> bool

Return true if a product is distributive over this node. Is distributive: (2 + 3) * 4 = 8 + 12 Not distributive: sign(2 + 3) * 4 != sign(8 + 12)

pub fn unit(&self) -> Result<CalcUnits, ()>

If the node has a valid unit outcome, then return it, otherwise fail.

pub fn negate(&mut self)

Negate the node inline. If the node is distributive, it is replaced by the result, otherwise the node is wrapped in a [Negate] node.

fn sort_key(&self) -> SortKey

pub fn as_leaf(&self) -> Option<&L>

Returns the leaf if we can (if simplification has allowed it).

pub fn try_sum_in_place(&mut self, other: &Self) -> Result<(), ()>

Tries to merge one node into another using the sum, that is, perform x + y.

pub fn try_product_in_place(&mut self, other: &mut Self) -> bool

Tries to merge one node into another using the product, that is, perform x * y.

fn try_op<O>(&self, other: &Self, op: O) -> Result<Self, ()>
where O: Fn(f32, f32) -> f32,

Tries to apply a generic arithmetic operator

pub fn map(&mut self, op: impl FnMut(f32) -> f32) -> Result<(), ()>

Map the value of this node with the given operation.

pub fn map_leaves<O, F>(&self, map: F) -> CalcNode<O>
where O: CalcNodeLeaf, F: FnMut(&L) -> O,

Convert this CalcNode into a CalcNode with a different leaf kind.

fn map_leaves_internal<O, F>(&self, map: &mut F) -> CalcNode<O>
where O: CalcNodeLeaf, F: FnMut(&L) -> O,

pub fn resolve(&self) -> Result<L, ()>

Resolve this node into a value.

pub fn resolve_map<F>(&self, leaf_to_output_fn: F) -> Result<L, ()>
where F: FnMut(&L) -> Result<L, ()>,

Resolve this node into a value, given a function that maps the leaf values.

fn resolve_internal<F>(&self, leaf_to_output_fn: &mut F) -> Result<L, ()>
where F: FnMut(&L) -> Result<L, ()>,

pub fn map_node<F>(&mut self, mapping_fn: F) -> Result<(), ()>
where F: FnMut(&CalcNode<L>) -> Result<Option<CalcNode<L>>, ()>,

Mutate nodes within this calc node tree using given the mapping function.

fn map_node_internal<F>(&mut self, mapping_fn: &mut F) -> Result<(), ()>
where F: FnMut(&CalcNode<L>) -> Result<Option<CalcNode<L>>, ()>,

fn is_negative_leaf(&self) -> Result<bool, ()>

fn is_zero_leaf(&self) -> Result<bool, ()>

fn is_infinite_leaf(&self) -> Result<bool, ()>

fn is_nan_leaf(&self) -> Result<bool, ()>

pub fn visit_depth_first(&mut self, f: impl FnMut(&mut Self))

Visits all the nodes in this calculation tree recursively, starting by the leaves and bubbling all the way up.

This is useful for simplification, but can also be used for validation and such.

fn visit_depth_first_internal(&mut self, f: &mut impl FnMut(&mut Self))

pub fn simplify_and_sort_direct_children(&mut self) -> SimplificationResult

This function simplifies and sorts the calculation of the specified node. It simplifies directly nested nodes while assuming that all nodes below it have already been simplified. It is recommended to use this function in combination with visit_depth_first().

This function is necessary only if the node needs to be preserved after parsing, specifically for <length-percentage> cases where the calculation contains percentages or relative units. Otherwise, the node can be evaluated using resolve(), which will automatically provide a simplified value.

https://drafts.csswg.org/css-values-4/#calc-simplification

pub fn simplify_and_sort(&mut self) -> SimplificationResult

Simplifies and sorts the kids in the whole calculation subtree.

fn to_css_impl<W>( &self, dest: &mut CssWriter<'_, W>, level: ArgumentLevel, ) -> Result
where W: Write,

fn to_typed_impl( &self, dest: &mut ThinVec<TypedValue>, level: ArgumentLevel, ) -> Result<(), ()>

fn compare( &self, other: &Self, basis_positive: PositivePercentageBasis, ) -> Option<Ordering>

fn gt(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool

Return whether a leaf is greater than another.

fn lt(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool

Return whether a leaf is less than another.

fn lte(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool

Return whether a leaf is smaller or equal than another.

impl GenericCalcNode<Leaf>

fn parse_one<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, flags: CalcParseFlags, ) -> Result<Self, ParseError<'i>>

Tries to parse a single element in the expression, that is, a <length>, <angle>, <time>, <percentage>, <resolution>, etc.

May return a “complex” CalcNode, in the presence of a parenthesized expression, for example.

pub fn parse<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, function: MathFunction, flags: CalcParseFlags, ) -> Result<Self, ParseError<'i>>

Parse a top-level calc expression, with all nested sub-expressions.

This is in charge of parsing, for example, 2 + 3 * 100%.

fn parse_argument<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, flags: CalcParseFlags, ) -> Result<Self, ParseError<'i>>

fn parse_product<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, flags: CalcParseFlags, ) -> Result<Self, ParseError<'i>>

Parse a top-level calc expression, and all the products that may follow, and stop as soon as a non-product expression is found.

This should parse correctly:

2

2 * 2

2 * 2 + 2 (but will leave the + 2 unparsed).

pub fn to_computed_value( &self, context: Option<&Context<'_>>, origin_color: Option<&AbsoluteColor>, ) -> Self

Computes this calc tree against the given context (if any), resolving context-dependent leaves (e.g. lengths) and substituting color channel references against origin_color when provided.

pub fn into_length_or_percentage( self, clamping_mode: AllowedNumericType, ) -> Result<CalcLengthPercentage, ()>

Tries to simplify this expression into a <length> or <percentage> value.

fn into_time(self, clamping_mode: AllowedNumericType) -> Result<CalcNumeric, ()>

Tries to simplify this expression into a <time> value.

fn into_resolution(self) -> Result<CalcNumeric, ()>

Tries to simplify this expression into a <resolution> value.

fn into_angle( self, clamping_mode: AllowedNumericType, ) -> Result<CalcNumeric, ()>

Tries to simplify this expression into a CalcNumeric value.

fn into_number( self, clamping_mode: AllowedNumericType, ) -> Result<CalcNumeric, ()>

Tries to convert this expression into a CalcNumeric, keeping the AST for later evaluation at computed-value time.

fn into_percentage( self, clamping_mode: AllowedNumericType, ) -> Result<CalcNumeric, ()>

Tries to convert this expression into a CalcNumeric, keeping the AST for later evaluation at computed-value time.

pub fn math_function<'i>( _: &ParserContext<'_>, name: &CowRcStr<'i>, location: SourceLocation, ) -> Result<MathFunction, ParseError<'i>>

Given a function name, and the location from where the token came from, return a mathematical function corresponding to that name or an error.

pub fn parse_length_or_percentage<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, clamping_mode: AllowedNumericType, function: MathFunction, allow_anchor: AllowAnchorPositioningFunctions, ) -> Result<CalcLengthPercentage, ParseError<'i>>

Convenience parsing function for <length> | <percentage>, and, optionally, anchor().

pub fn parse_percentage<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, clamping_mode: AllowedNumericType, function: MathFunction, ) -> Result<CalcNumeric, ParseError<'i>>

Convenience parsing function for percentages.

pub fn parse_length<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, clamping_mode: AllowedNumericType, function: MathFunction, ) -> Result<CalcLengthPercentage, ParseError<'i>>

Convenience parsing function for <length>.

pub fn parse_number<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, clamping_mode: AllowedNumericType, function: MathFunction, ) -> Result<CalcNumeric, ParseError<'i>>

Convenience parsing function for <number>.

pub fn parse_angle<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, function: MathFunction, ) -> Result<CalcNumeric, ParseError<'i>>

Convenience parsing function for <angle>.

pub fn parse_time<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, clamping_mode: AllowedNumericType, function: MathFunction, ) -> Result<CalcNumeric, ParseError<'i>>

Convenience parsing function for <time>.

pub fn parse_resolution<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, function: MathFunction, ) -> Result<CalcNumeric, ParseError<'i>>

Convenience parsing function for <resolution>.