style::values::generics::calc

Trait CalcNodeLeaf

pub trait CalcNodeLeaf:
    Clone
    + Sized
    + PartialEq
    + ToCss
    + ToTyped {

    // Required methods
    fn unit(&self) -> CalcUnits;
    fn unitless_value(&self) -> Option<f32>;
    fn as_angle_radians(&self) -> Option<f32>;
    fn new_angle_from_radians(radians: f32) -> Self;
    fn compare(
        &self,
        other: &Self,
        base_is_positive: PositivePercentageBasis,
    ) -> Option<Ordering>;
    fn new_number(value: f32) -> Self;
    fn as_number(&self) -> Option<f32>;
    fn try_sum_in_place(&mut self, other: &Self) -> Result<(), ()>;
    fn try_product_in_place(&mut self, other: &mut Self) -> bool;
    fn try_op<O>(&self, other: &Self, op: O) -> Result<Self, ()>
       where O: Fn(f32, f32) -> f32;
    fn map(&mut self, op: impl FnMut(f32) -> f32) -> Result<(), ()>;
    fn simplify(&mut self) -> SimplificationResult;
    fn sort_key(&self) -> SortKey;

    // Provided methods
    fn is_same_unit_as(&self, other: &Self) -> bool { ... }
    fn gt(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool { ... }
    fn lt(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool { ... }
    fn lte(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool { ... }
    fn as_number_or_angle_radians(&self) -> Option<f32> { ... }
    fn is_negative(&self) -> Result<bool, ()> { ... }
    fn is_infinite(&self) -> Result<bool, ()> { ... }
    fn is_zero(&self) -> Result<bool, ()> { ... }
    fn is_nan(&self) -> Result<bool, ()> { ... }
    fn sign_from(leaf: &impl CalcNodeLeaf) -> Result<Self, ()> { ... }
    fn should_serialize_with_root_calc_wrapper(&self) -> bool { ... }

}

A trait that represents all the stuff a valid leaf of a calc expression.

Required Methods

fn unit(&self) -> CalcUnits

Returns the unit of the leaf.

fn unitless_value(&self) -> Option<f32>

Returns the unitless value of this leaf if one is available.

fn as_angle_radians(&self) -> Option<f32>

Returns the angle value in radians if this leaf is an angle.

fn new_angle_from_radians(radians: f32) -> Self

Creates a new angle leaf from a value in radians.

fn compare( &self, other: &Self, base_is_positive: PositivePercentageBasis, ) -> Option<Ordering>

Do a partial comparison of these values.

fn new_number(value: f32) -> Self

Create a new leaf with a number value.

fn as_number(&self) -> Option<f32>

Returns a float value if the leaf is a number.

fn try_sum_in_place(&mut self, other: &Self) -> Result<(), ()>

Tries to merge one leaf into another using the sum, that is, perform x + y.

fn try_product_in_place(&mut self, other: &mut Self) -> bool

Try to merge the right leaf into the left by using a multiplication. Return true if the merge was successful, otherwise false.

fn try_op<O>(&self, other: &Self, op: O) -> Result<Self, ()>
where O: Fn(f32, f32) -> f32,

Tries a generic arithmetic operation.

fn map(&mut self, op: impl FnMut(f32) -> f32) -> Result<(), ()>

Map the value of this node with the given operation.

fn simplify(&mut self) -> SimplificationResult

Canonicalizes the expression if necessary.

fn sort_key(&self) -> SortKey

Returns the sort key for simplification.

Provided Methods

fn is_same_unit_as(&self, other: &Self) -> bool

Return true if the units of both leaves are equal. (NOTE: Does not take the values into account)

fn gt(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool

Return whether a leaf is greater than another.

fn lt(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool

Return whether a leaf is less than another.

fn lte(&self, other: &Self, basis_positive: PositivePercentageBasis) -> bool

Return whether a leaf is smaller or equal than another.

fn as_number_or_angle_radians(&self) -> Option<f32>

Returns a number or angle radians if the leaf is a number or angle.

fn is_negative(&self) -> Result<bool, ()>

Whether this value is known-negative.

fn is_infinite(&self) -> Result<bool, ()>

Whether this value is infinite.

fn is_zero(&self) -> Result<bool, ()>

Whether this value is zero.

fn is_nan(&self) -> Result<bool, ()>

Whether this value is NaN.

fn sign_from(leaf: &impl CalcNodeLeaf) -> Result<Self, ()>

Create a new leaf containing the sign() result of the given leaf.

fn should_serialize_with_root_calc_wrapper(&self) -> bool

Whether this leaf node should serialize with a calc() wrapper if this node is the root of the calculation tree.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.