style::values::generics::calc

Type Alias GenericCalcAnchorFunction

pub type GenericCalcAnchorFunction<L> = GenericAnchorFunction<Box<GenericCalcNode<L>>, Box<GenericAnchorFunctionFallback<L>>>;

anchor() function used in math functions.

Aliased Type

#[repr(C)]
pub struct GenericCalcAnchorFunction<L> {
    pub target_element: TreeScoped<DashedIdent>,
    pub side: GenericAnchorSide<Box<GenericCalcNode<L>>>,
    pub fallback: Optional<Box<GenericAnchorFunctionFallback<L>>>,
}

Fields
target_element: TreeScoped<DashedIdent>

Anchor name of the element to anchor to. If omitted, selects the implicit anchor element. The shadow cascade order of the tree-scoped anchor name associates the name with the host of the originating stylesheet.

side: GenericAnchorSide<Box<GenericCalcNode<L>>>

Where relative to the target anchor element to position the anchored element to.

fallback: Optional<Box<GenericAnchorFunctionFallback<L>>>

Value to use in case the anchor function is invalid.