style::values::generics

Module color

Generic types for color properties.

Re-exports

pub use self::GenericColorMix as ColorMix;

pub use self::GenericColor as Color;

pub use self::GenericColorOrAuto as ColorOrAuto;

pub use self::GenericCaretColor as CaretColor;

Structs

ColorMixFlags

Flags used to modify the calculation of a color mix result.

GenericCaretColor

Caret color is effectively a ColorOrAuto, but resolves auto to currentColor.

GenericColorMix

A restricted version of the css color-mix() function, which only supports percentages.

GenericColorMixItem

One (color, percentage) component of a color-mix() expression.

GenericLightDark

A light-dark(, ) function.

Enums

GenericColor

This struct represents a combined color from a numeric color and the current foreground color (currentcolor keyword).

GenericColorOrAuto

Either <color> or auto.