style::values::generics::color

Enum GenericColor

#[repr(C)]
pub enum GenericColor<Percentage> {
    Absolute(AbsoluteColor),
    ColorFunction(Box<ColorFunction<Self>>),
    CurrentColor,
    ColorMix(Box<GenericColorMix<Self, Percentage>>),
    ContrastColor(Box<Self>),
}

This struct represents a combined color from a numeric color and the current foreground color (currentcolor keyword).

Variants

Absolute(AbsoluteColor)

The actual numeric color.

ColorFunction(Box<ColorFunction<Self>>)

A unresolvable color.

CurrentColor

The CurrentColor keyword.

ColorMix(Box<GenericColorMix<Self, Percentage>>)

The color-mix() function.

ContrastColor(Box<Self>)

The contrast-color() function.

Implementations

impl GenericColor<Percentage>

pub const TRANSPARENT_BLACK: Self

A fully transparent color.

pub const BLACK: Self

An opaque black color.

pub const WHITE: Self

An opaque white color.

pub fn from_color_mix(color_mix: ColorMix) -> Self

Create a new computed Color from a given color-mix, simplifying it to an absolute color if possible.

pub fn resolve_to_absolute( &self, current_color: &AbsoluteColor, ) -> AbsoluteColor

Combine this complex color with the given foreground color into an absolute color.

pub fn resolve_contrast_color(bg_color: &AbsoluteColor) -> AbsoluteColor

Performs the resolution of contrast-color given a background color.

fn contrast_ratio(a: &AbsoluteColor, b: &AbsoluteColor) -> f32

impl<Percentage> Color<Percentage>

pub fn as_absolute(&self) -> Option<&AbsoluteColor>

If this color is absolute return it’s value, otherwise return None.

pub fn currentcolor() -> Self

Returns a color value representing currentcolor.

pub fn is_currentcolor(&self) -> bool

Whether it is a currentcolor value (no numeric color component).

pub fn is_absolute(&self) -> bool

Whether this color is an absolute color.