Enum style::values::generics::grid::TrackListType

pub enum TrackListType {
    Auto(u16),
    Normal,
    Explicit,
}

The type of a <track-list> as determined during parsing.

https://drafts.csswg.org/css-grid/#typedef-track-list

Variants
Auto(u16)

<auto-track-list>

If this type exists, then the value at the index in line_names field in TrackList has the <line-names>? list that comes before <auto-repeat>. If it's a specified value, then the repeat() function (that follows the line names list) is also at the given index in values field. On the contrary, if it's a computed value, then the repeat() function is in the auto_repeat field.

Normal

<track-list>

Explicit

<explicit-track-list>

Note that this is a subset of the normal <track-list>, and so it could be used in place of the latter.