Struct style::values::generics::grid::GenericGridLine

#[repr(C)]
pub struct GenericGridLine<Integer> {
    pub ident: CustomIdent,
    pub line_num: Integer,
    pub is_span: bool,
}

A <grid-line> type.

https://drafts.csswg.org/css-grid/#typedef-grid-row-start-grid-line

Fields
ident: CustomIdent

A custom identifier for named lines, or the empty atom otherwise.

https://drafts.csswg.org/css-grid/#grid-placement-slot

line_num: Integer

Denotes the nth grid line from grid item’s placement.

This is clamped by MIN_GRID_LINE and MAX_GRID_LINE.

NOTE(emilio): If we ever allow animating these we need to either do something more complicated for the clamping, or do this clamping at used-value time.

is_span: bool

Flag to check whether it’s a span keyword.

Implementations

impl<Integer> GridLine<Integer>
where Integer: PartialEq + Zero,

pub fn auto() -> Self

The auto value.

pub fn is_auto(&self) -> bool

Check whether this <grid-line> represents an auto value.

pub fn is_ident_only(&self) -> bool

Check whether this <grid-line> represents a <custom-ident> value.

pub fn can_omit(&self, other: &Self) -> bool

Check if self makes other omittable according to the rules at: https://drafts.csswg.org/css-grid/#propdef-grid-column https://drafts.csswg.org/css-grid/#propdef-grid-area