Struct style::values::generics::grid::TrackList

pub struct TrackList<LengthOrPercentage, Integer> {
    pub list_type: TrackListType,
    pub values: Vec<TrackListValue<LengthOrPercentage, Integer>>,
    pub line_names: Box<[Box<[CustomIdent]>]>,
    pub auto_repeat: Option<TrackRepeat<LengthOrPercentage, Integer>>,
}

A grid <track-list> type.

https://drafts.csswg.org/css-grid/#typedef-track-list

Fields
list_type: TrackListType

The type of this <track-list> (auto, explicit or general).

In order to avoid parsing the same value multiple times, this does a single traversal and arrives at the type of value it has parsed (or bails out gracefully with an error).

values: Vec<TrackListValue<LengthOrPercentage, Integer>>

A vector of <track-size> | <track-repeat> values.

line_names: Box<[Box<[CustomIdent]>]>

<line-names> accompanying <track-size> | <track-repeat> values.

If there's no <line-names>, then it's represented by an empty vector. For N values, there will be N+1 <line-names>, and so this vector's length is always one value more than that of the <track-size>.

auto_repeat: Option<TrackRepeat<LengthOrPercentage, Integer>>

<auto-repeat> value. There can only be one <auto-repeat> in a TrackList.