Struct style::values::generics::grid::TrackRepeat

pub struct TrackRepeat<L, I> {
    pub count: RepeatCount<I>,
    pub line_names: Box<[Box<[CustomIdent]>]>,
    pub track_sizes: Vec<TrackSize<L>>,
}

The structure containing <line-names> and <track-size> values.

It can also hold repeat() function parameters, which expands into the respective values in its computed form.

Fields
count: RepeatCount<I>

The number of times for the value to be repeated (could also be auto-fit or auto-fill)

line_names: Box<[Box<[CustomIdent]>]>

<line-names> accompanying <track_size> values.

If there's no <line-names>, then it's represented by an empty vector. For N <track-size> values, there will be N+1 <line-names>, and so this vector's length is always one value more than that of the <track-size>.

track_sizes: Vec<TrackSize<L>>

<track-size> values.

Methods

impl<L: Clone> TrackRepeat<L, Integer>

pub fn expand(&self) -> Self

If the repeat count is numeric, then expand the values and merge accordingly.