style::values::generics

Module length

Generic types for CSS values related to length.

Re-exports

pub use self::GenericLengthPercentageOrAuto as LengthPercentageOrAuto;

pub use self::GenericSize as Size;

pub use self::GenericMaxSize as MaxSize;

pub use self::GenericLengthOrNumber as LengthOrNumber;

pub use self::GenericLengthPercentageOrNormal as LengthPercentageOrNormal;

Structs

GenericAnchorSizeFunction

Anchor size function used by sizing, margin and inset properties. This resolves to the size of the anchor at computed time.

Enums

AnchorResolutionResult

Result of resolving an anchor function.

AnchorSizeKeyword

Keyword values for the anchor size function.

GenericLengthOrNumber

A generic <length> | <number> value for the tab-size property.

GenericLengthPercentageOrAuto

A <length-percentage> | auto value.

GenericLengthPercentageOrNormal

A generic <length-percentage> | normal` value.

GenericMargin

Specified type for margin properties, which allows the use of the anchor-size() function.

GenericMaxSize

A generic value for the max-width or max-height property.

GenericSize

A generic value for the width, height, min-width, or min-height property.