style::values::generics::position

Trait IsTreeScoped

pub trait IsTreeScoped {
    // Provided method
    fn is_tree_scoped(&self) -> bool { ... }
}

Trait to check if the value of a potentially-tree-scoped type T is actually tree-scoped. e.g. none value of anchor-scope should not be tree-scoped.

Provided Methods

fn is_tree_scoped(&self) -> bool

Returns true if the current value should be considered tree-scoped. Default implementation assumes that the value is always tree-scoped.