Struct style::values::generics::NonNegative

#[repr(transparent)]pub struct NonNegative<T>(pub T);

A wrapper of Non-negative values.

Implementations

impl NonNegative<CSSPixelLength>

pub fn new(px: CSSFloat) -> Self

Create a NonNegativeLength.

pub fn px(&self) -> CSSFloat

Return the pixel value of |NonNegativeLength|.

pub fn clamp(self) -> Self

Ensures it is non negative

impl NonNegative<LengthPercentage>

pub fn is_definitely_zero(&self) -> bool

Returns true if the computed value is absolute 0 or 0%.

pub fn to_used_value(&self, containing_length: Au) -> Au

Returns the used value.

pub fn maybe_to_used_value(&self, containing_length: Option<Au>) -> Option<Au>

Convert the computed value into used value.

impl NonNegative<Percentage>

pub fn hundred() -> Self

100%

impl NonNegative<NumberOrPercentage>

pub fn hundred_percent() -> Self

Returns the 100% value.

impl NonNegative<Length>

pub fn from_px(px_value: CSSFloat) -> Self

Get an absolute length from a px value.

pub fn parse_quirky<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
allow_quirks: AllowQuirks
) -> Result<Self, ParseError<'i>>

Parses a non-negative length, optionally with quirks.

impl NonNegative<LengthPercentage>

pub fn zero_percent() -> Self

Returns a 0% value.

pub fn parse_quirky<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
allow_quirks: AllowQuirks
) -> Result<Self, ParseError<'i>>

Parses a length or a percentage, allowing the unitless length quirk. https://quirks.spec.whatwg.org/#the-unitless-length-quirk

impl NonNegative<Number>

pub fn new(val: CSSFloat) -> Self

Returns a new non-negative number with the value val.

pub fn get(&self) -> f32

Returns the numeric value.

impl NonNegative<NumberOrPercentage>

pub fn hundred_percent() -> Self

Returns the 100% value.