style::values::generics

Module transform

Generic types for CSS values that are related to transformations.

Re-exports

pub use self::GenericMatrix as Matrix;

pub use self::GenericMatrix3D as Matrix3D;

pub use self::GenericTransformOrigin as TransformOrigin;

pub use self::GenericPerspectiveFunction as PerspectiveFunction;

pub use self::GenericTransformOperation as TransformOperation;

pub use self::GenericTransform as Transform;

pub use self::GenericRotate as Rotate;

pub use self::GenericScale as Scale;

pub use self::GenericTranslate as Translate;

Structs

GenericMatrix

A generic 2D transformation matrix.

GenericMatrix3D

GenericTransform

A value of the transform property

GenericTransformOrigin

A generic transform origin.

Enums

GenericPerspectiveFunction

A value for the perspective() transform function, which is either a non-negative <length> or none.

GenericRotate

A value of the Rotate property

GenericScale

A value of the Scale property

GenericTransformOperation

A single operation in the list of a transform value

GenericTranslate

A value of the translate property

TransformStyle

Traits

IsParallelTo

A trait to check if the current 3D vector is parallel to the DirectionVector. This is especially for serialization on Rotate.

ToAbsoluteLength

Convert a length type into the absolute lengths.

ToFloat

Convert a number type into a float.

ToMatrix

Support the conversion to a 3d matrix.

ToRadians

A little helper to deal with both specified and computed angles.

ToTransformComponent

Converts a transform operation into a transform component.

Functions

create_perspective_matrix

Return the transform matrix from a perspective length.

get_normalized_vector_and_angle

Return the normalized direction vector and its angle for Rotate3D.

is_same 🔒

y_axis_and_z_axis_are_zero 🔒