Enum style::values::generics::transform::TransformOperation

pub enum TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber> {
    Matrix(Matrix<Number>),
    PrefixedMatrix(Matrix<Number, LoPoNumber>),
    Matrix3D(Matrix3D<Number>),
    PrefixedMatrix3D(Matrix3D<Number, LoPoNumber, LengthOrNumber>),
    Skew(Angle, Option<Angle>),
    SkewX(Angle),
    SkewY(Angle),
    Translate(LengthOrPercentage, Option<LengthOrPercentage>),
    TranslateX(LengthOrPercentage),
    TranslateY(LengthOrPercentage),
    TranslateZ(Length),
    Translate3D(LengthOrPercentage, LengthOrPercentage, Length),
    Scale(Number, Option<Number>),
    ScaleX(Number),
    ScaleY(Number),
    ScaleZ(Number),
    Scale3D(Number, Number, Number),
    Rotate(Angle),
    RotateX(Angle),
    RotateY(Angle),
    RotateZ(Angle),
    Rotate3D(Number, Number, Number, Angle),
    Perspective(Length),
    InterpolateMatrix {
        from_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>,
        to_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>,
        progress: Percentage,
    },
    AccumulateMatrix {
        from_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>,
        to_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>,
        count: Integer,
    },
}

A single operation in the list of a transform value

Variants
Matrix(Matrix<Number>)

Represents a 2D 2x3 matrix.

PrefixedMatrix(Matrix<Number, LoPoNumber>)

Represents a 3D 4x4 matrix with percentage and length values. For moz-transform.

Matrix3D(Matrix3D<Number>)

Represents a 3D 4x4 matrix.

PrefixedMatrix3D(Matrix3D<Number, LoPoNumber, LengthOrNumber>)

Represents a 3D 4x4 matrix with percentage and length values. For moz-transform.

Skew(Angle, Option<Angle>)

A 2D skew.

If the second angle is not provided it is assumed zero.

Syntax can be skew(angle) or skew(angle, angle)

SkewX(Angle)

skewX(angle)

SkewY(Angle)

skewY(angle)

Translate(LengthOrPercentage, Option<LengthOrPercentage>)

translate(x, y) or translate(x)

TranslateX(LengthOrPercentage)

translateX(x)

TranslateY(LengthOrPercentage)

translateY(y)

TranslateZ(Length)

translateZ(z)

Translate3D(LengthOrPercentage, LengthOrPercentage, Length)

translate3d(x, y, z)

Scale(Number, Option<Number>)

A 2D scaling factor.

scale(2) is parsed as Scale(Number::new(2.0), None) and is equivalent to writing scale(2, 2) (Scale(Number::new(2.0), Some(Number::new(2.0)))).

Negative values are allowed and flip the element.

Syntax can be scale(factor) or scale(factor, factor)

ScaleX(Number)

scaleX(factor)

ScaleY(Number)

scaleY(factor)

ScaleZ(Number)

scaleZ(factor)

Scale3D(Number, Number, Number)

scale3D(factorX, factorY, factorZ)

Rotate(Angle)

Describes a 2D Rotation.

In a 3D scene rotate(angle) is equivalent to rotateZ(angle).

RotateX(Angle)

Rotation in 3D space around the x-axis.

RotateY(Angle)

Rotation in 3D space around the y-axis.

RotateZ(Angle)

Rotation in 3D space around the z-axis.

Rotate3D(Number, Number, Number, Angle)

Rotation in 3D space.

Generalization of rotateX, rotateY and rotateZ.

Perspective(Length)

Specifies a perspective projection matrix.

Part of CSS Transform Module Level 2 and defined at  13.1. 3D Transform Function.

The value must be greater than or equal to zero.

InterpolateMatrix

A intermediate type for interpolation of mismatched transform lists.

Fields of InterpolateMatrix

from_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>

to_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>

progress: Percentage

AccumulateMatrix

A intermediate type for accumulation of mismatched transform lists.

Fields of AccumulateMatrix

from_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>

to_list: Transform<TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>>

count: Integer

Methods

impl<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber> TransformOperation<Angle, Number, Length, Integer, LengthOrNumber, LengthOrPercentage, LoPoNumber>

pub fn is_translate(&self) -> bool

Check if it is any translate function

pub fn is_scale(&self) -> bool

Check if it is any scale function