style::values

Module resolved

Resolved values. These are almost always computed values, but in some cases there are used values.

Modules

animation 🔒

Resolved animation values.

color 🔒

Resolved color values.

counters 🔒

Resolved values for counter properties

Macros

trivial_to_resolved_value 🔒

Structs

Context

Information needed to resolve a given value.

Traits

ToResolvedValue

A trait to represent the conversion between resolved and resolved values.