Trait style::values::resolved::ToResolvedValue

pub trait ToResolvedValue {
    type ResolvedValue;

    // Required methods
    fn to_resolved_value(self, context: &Context<'_>) -> Self::ResolvedValue;
    fn from_resolved_value(resolved: Self::ResolvedValue) -> Self;
}

A trait to represent the conversion between resolved and resolved values.

This trait is derivable with #[derive(ToResolvedValue)].

The deriving code assumes that if the type isn’t generic, then the trait can be implemented as simple move. This means that a manual implementation with ResolvedValue = Self is bogus if it returns anything else than a clone.

Required Associated Types

type ResolvedValue

The resolved value type we’re going to be converted to.

Required Methods

fn to_resolved_value(self, context: &Context<'_>) -> Self::ResolvedValue

Convert a resolved value to a resolved value.

fn from_resolved_value(resolved: Self::ResolvedValue) -> Self

Convert a resolved value to resolved value form.