style::values::specified::align

Struct SelfAlignment

#[repr(C)]
pub struct SelfAlignment(pub AlignFlags);

The specified value of the {align,justify}-self properties.

https://drafts.csswg.org/css-align/#self-alignment https://drafts.csswg.org/css-align/#propdef-align-self

Tuple Fields
0: AlignFlags
Implementations

impl SelfAlignment

pub fn auto() -> Self

The initial value ‘auto’

pub fn is_valid_on_both_axes(&self) -> bool

Returns whether this value is valid for both axis directions.

pub fn parse_block<'i, 't>( _: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<Self, ParseError<'i>>

Parse self-alignment on the block axis (for align-self)

pub fn parse_inline<'i, 't>( _: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<Self, ParseError<'i>>

Parse self-alignment on the block axis (for align-self)

fn parse<'i, 't>( input: &mut Parser<'i, 't>, axis: AxisDirection, ) -> Result<Self, ParseError<'i>>

Parse a self-alignment value on one of the axes.

fn list_keywords(f: KeywordsCollectFn<'_>, axis: AxisDirection)

pub fn flip_position(self) -> Self

Performs a flip of the position, that is, for self-start we return self-end, for left we return right, etc.

pub fn with_value(self, value: AlignFlags) -> Self

Returns a fixed-up alignment value.