Struct style::values::specified::angle::Angle

pub struct Angle { /* fields omitted */ }

A specified angle.

Computed angles are essentially same as specified ones except for calc() value serialization. Therefore we are storing a computed angle inside to hold the actual value and its unit.

Methods

impl Angle

pub fn from_degrees(value: CSSFloat, was_calc: bool) -> Self

Creates an angle with the given value in degrees.

pub fn from_gradians(value: CSSFloat, was_calc: bool) -> Self

Creates an angle with the given value in gradians.

pub fn from_turns(value: CSSFloat, was_calc: bool) -> Self

Creates an angle with the given value in turns.

pub fn from_radians(value: CSSFloat, was_calc: bool) -> Self

Creates an angle with the given value in radians.

pub fn radians(self) -> f32

Returns the amount of radians this angle represents.

pub fn degrees(self) -> f32

Returns the amount of degrees this angle represents.

pub fn zero() -> Self

Returns 0deg.

pub fn from_calc(radians: CSSFloat) -> Self

Returns an Angle parsed from a calc() expression.

impl Angle

pub fn parse_dimension(
value: CSSFloat,
unit: &str,
from_calc: bool
) -> Result<Angle, ()>

Parse an <angle> value given a value and an unit.

pub fn parse_with_unitless<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>
) -> Result<Self, ParseError<'i>>

Parse an <angle> allowing unitless zero to represent a zero angle.

See the comment in AllowUnitlessZeroAngle for why.