style::values::specified

Module animation

Specified types for properties related to animations and transitions.

Structs

AnimationName

A value for the animation-name property.

ScrollFunction

The scroll() notation. https://drafts.csswg.org/scroll-animations-1/#scroll-notation

TimelineIdent

The identifier for a timeline name.

ViewTransitionClassList

The view-transition-class: none | <custom-ident>+.

ViewTransitionNameKeyword

The view-transition-name: none | <custom-ident> | match-element.

Enums

AnimationComposition

https://drafts.csswg.org/css-animations-2/#animation-composition

AnimationDirection

https://drafts.csswg.org/css-animations/#propdef-animation-direction

AnimationFillMode

https://drafts.csswg.org/css-animations/#propdef-animation-fill-mode

AnimationIterationCount

https://drafts.csswg.org/css-animations/#animation-iteration-count

AnimationPlayState

https://drafts.csswg.org/css-animations/#animation-play-state

ScrollAxis

A value for the used in scroll(), or a value for {scroll|view}-timeline-axis.

Scroller

A value for the used in scroll().

TimelineRangeName

The value type, which indicates a CSS identifier representing one of the predefined named timeline ranges. https://drafts.csswg.org/scroll-animations-1/#named-ranges

TransitionBehavior

A specified value for .

TransitionProperty

A given transition property, that is either All, a longhand or shorthand property, or an unsupported or custom property.

Functions

parse_animation_range 🔒

Type Aliases

AnimationDuration

A specified value for the animation-duration property.

AnimationRangeEnd

A specified value for the animation-range-end.

AnimationRangeStart

A specified value for the animation-range-start.

AnimationRangeValue

The internal value for animation-range-start and animation-range-end.

AnimationTimeline

A specified value for the animation-timeline property.

TimelineName

The typedef of scroll-timeline-name or view-timeline-name.

ViewTimelineInset

A specified value for the view-timeline-inset property.

ViewTransitionClass

https://drafts.csswg.org/css-view-transitions-2/#view-transition-class-prop

ViewTransitionName

https://drafts.csswg.org/css-view-transitions-1/#view-transition-name-prop