style::values::specified

Module basic_shape

CSS handling for the specified value of basic-shapes

Re-exports

pub use crate::values::generics::basic_shape::FillRule;

Structs

AllowedBasicShapes

The flags to represent which basic shapes we would like to support.

ShapeRectFunction

Defines a rectangle via insets from the top and left edges of the reference box.

Xywh

The specified value of xywh(). Defines a rectangle via offsets from the top and left edge of the reference box, and a specified width and height.

Enums

BasicShapeRect

The specified value of . = <inset()> | <rect()> | <xywh()>

ByTo 🔒

This determines whether the command is absolutely or relatively positioned. https://drafts.csswg.org/css-shapes-1/#typedef-shape-command-end-point

ShapeType

For filled shapes, we use fill-rule, and store it for path() and polygon(). For outline shapes, we should ignore fill-rule.

Functions

parse_at_position 🔒

parse_fill_rule 🔒

parse_round 🔒

parse_shape_or_box 🔒

A helper for both clip-path and shape-outside parsing of shapes.

round_to_css 🔒

Type Aliases

BasicShape

A specified basic shape.

Circle

A specified circle.

ClipPath

A specified clip-path value.

Ellipse

A specified ellipse.

InsetRect

The specified value of inset().

PathOrShapeFunction

The specified value of PathOrShapeFunction.

Polygon

The specified value of Polygon.

ShapeCommand

The specified value of ShapeCommand.

ShapeOutside

A specified shape-outside value.

ShapeRadius

The specified value of ShapeRadius.