style::values::specified::border

Type Alias BorderImageSideWidth

pub type BorderImageSideWidth = GenericBorderImageSideWidth<NonNegativeLengthPercentage, NonNegativeNumber>;

A specified value for a single side of a border-image-width property.

Aliased Type

#[repr(C, u8)]
pub enum BorderImageSideWidth {
    Number(NonNegative<Number>),
    LengthPercentage(NonNegative<LengthPercentage>),
    Auto,
}

Variants

Number(NonNegative<Number>)

<number>

NOTE: Numbers need to be before length-percentagess, in order to parse them first, since 0 should be a number, not the 0px length.

LengthPercentage(NonNegative<LengthPercentage>)

<length-or-percentage>

Auto

auto

Implementations

impl BorderImageSideWidth

pub fn one() -> Self

Returns 1.