style::values::specified

Module box_

Specified types for box properties.

Re-exports

pub use crate::logical_geometry::WritingModeProperty;

pub use crate::typed_om::KeywordValue;

pub use crate::typed_om::ToTyped;

pub use crate::typed_om::TypedValue;

pub use crate::values::generics::box_::PositionProperty;

Structs

Contain

Constants for contain: https://drafts.csswg.org/css-contain/#contain-property

ContainerName

https://drafts.csswg.org/css-contain-3/#container-name

ContainerType

Specified keyword values for the container-type property. Spec: normal | [ [ size | inline-size ] || scroll-state ]

Display

MarginTrim

Specified value of the margin-trim property, which trims the margins of a container’s children where they meet the container’s edges. https://drafts.csswg.org/css-box-4/#propdef-margin-trim Pending https://github.com/w3c/csswg-drafts/issues/13731, we’re only targetting block trimming for now and skipping inline-start and inline-end

ScrollSnapAlign

https://drafts.csswg.org/css-scroll-snap-1/#scroll-snap-align

ScrollSnapType

https://drafts.csswg.org/css-scroll-snap-1/#scroll-snap-type

ScrollbarGutter

Values for scrollbar-gutter: https://drafts.csswg.org/css-overflow-3/#scrollbar-gutter-property

TouchAction

Values for the touch-action property.

WillChange

Provides a rendering hint to the user agent, stating what kinds of changes the author expects to perform on the element.

WillChangeBits

The change bits that we care about.

Enums

AlignmentBaseline

A specified value for the alignment-baseline property. https://drafts.csswg.org/css-inline-3/#alignment-baseline

Appearance

The value for the appearance property.

BaselineSource

A specified value for the baseline-source property. https://drafts.csswg.org/css-inline-3/#baseline-source

BreakBetween

A kind of break between two boxes.

BreakWithin

A kind of break within a box.

Clear

https://drafts.csswg.org/css2/#propdef-clear

ContentVisibility

https://drafts.csswg.org/css-contain-2/#content-visibility

DisplayInside

DisplayKeyword 🔒

DisplayOutside

Defines an element’s display type, which consists of the two basic qualities of how an element generates boxes https://drafts.csswg.org/css-display/#propdef-display

DominantBaseline

A specified value for the dominant-baseline property. https://drafts.csswg.org/css-inline-3/#dominant-baseline

Float

https://drafts.csswg.org/css-box/#propdef-float

Overflow

The value for the overflow-x / overflow-y properties.

OverflowAnchor

OverscrollBehavior

Resize

https://drafts.csswg.org/css-ui/#propdef-resize

ScrollSnapAlignKeyword

Specified value of scroll-snap-align keyword value.

ScrollSnapAxis

https://drafts.csswg.org/css-scroll-snap-1/#snap-axis

ScrollSnapStop

ScrollSnapStrictness

https://drafts.csswg.org/css-scroll-snap-1/#snap-strictness

Zoom

A specified value for the zoom property.

Functions

appearance_base_enabled 🔒

appearance_base_select_enabled 🔒

change_bits_for_longhand 🔒

change_bits_for_maybe_property 🔒

grid_enabled 🔒

Type Aliases

BaselineShift

A specified value for the baseline-shift property.

ContainIntrinsicSize

A specified value for the contain-intrinsic-size property.

LineClamp

A specified value for the line-clamp property.

OverflowClipMargin

The specified value of overflow-clip-margin.

Perspective

A specified value for the perspective property.