style::values::specified

Module calc

Calc expressions.

Structs

AdditionalFunctions 🔒

Additional functions within math functions that are permitted to be parsed depending on the context of parsing (e.g. Parsing inset allows use of anchor() within calc()).

AllowParse

What is allowed to be parsed for math functions within in this context?

CalcLengthPercentage

A struct to hold a simplified <length> or <percentage> expression.

Enums

AllowAnchorPositioningFunctions

Should parsing anchor-positioning functions in calc() be allowed?

Leaf

A leaf node inside a Calc expression’s AST.

MathFunction

The name of the mathematical function that we’re parsing.

Type Aliases

CalcAnchorFunction

Specified anchor() function in math functions.

CalcAnchorSizeFunction

Specified anchor-size() function in math functions.

CalcNode

A calc node representation for specified values.