Enum style::values::specified::calc::CalcUnit

pub enum CalcUnit {
    Number,
    Integer,
    Length,
    Percentage,
    LengthOrPercentage,
    Angle,
    Time,
}

An expected unit we intend to parse within a calc() expression.

This is used as a hint for the parser to fast-reject invalid expressions.

Variants
Number

<number>

Integer

<integer>

Length

<length>

Percentage

<percentage>

LengthOrPercentage

<length> | <percentage>

Angle

<angle>

Time

<time>