style::values::specified::calc

Enum Leaf

#[repr(u8)]
pub enum Leaf {
    Length(NoCalcLength),
    Angle(NoCalcAngle),
    Time(NoCalcTime),
    Resolution(NoCalcResolution),
    ColorComponent(ChannelKeyword),
    Percentage(NoCalcPercentage),
    Number(NoCalcNumber),
    TreeCountingFunction(TreeCountingFunction),
}

A leaf node inside a Calc expression’s AST.

Variants

Length(NoCalcLength)

<length>

Angle(NoCalcAngle)

<angle>

Time(NoCalcTime)

<time>

Resolution(NoCalcResolution)

<resolution>

ColorComponent(ChannelKeyword)

A component of a color.

Percentage(NoCalcPercentage)

<percentage>

Number(NoCalcNumber)

<number>

TreeCountingFunction(TreeCountingFunction)

A tree-counting function.

Implementations

impl Leaf

pub fn to_computed_value( &self, context: Option<&Context<'_>>, origin_color: Option<&AbsoluteColor>, ) -> Self

Computes this leaf against the given context (if any), substituting color channel references with the matching channel of origin_color when it is provided. If no origin color is available, channel references are kept symbolic so they can be resolved later.