style::values::specified

Module color

Specified color values.

Structs

Absolute

Container holding an absolute color and the text specified by an author.

ColorPropertyValue

Specified value for the “color” property, which resolves the currentcolor keyword to the parent color instead of self’s color.

ColorScheme

https://drafts.csswg.org/css-color-adjust/#color-scheme-prop

ColorSchemeFlags

Various flags to represent the color-scheme property in an efficient way.

Enums

Color

Specified color value

ForcedColorAdjust

https://drafts.csswg.org/css-color-adjust-1/#forced-color-adjust-prop

ForcedColors

Possible values for the forced-colors media query. https://drafts.csswg.org/mediaqueries-5/#forced-colors

PreserveAuthored 🔒

Whether to preserve authored colors during parsing. That’s useful only if we plan to serialize the color back.

PrintColorAdjust

https://drafts.csswg.org/css-color-adjust/#print-color-adjust

Type Aliases

CaretColor

caret-color

ColorMix

A specified color-mix().

ColorOrAuto

auto |