Enum style::values::specified::color::Color

pub enum Color {
    CurrentColor,
    Absolute(Box<Absolute>),
    ColorMix(Box<ColorMix>),
    LightDark(Box<LightDark>),
}

Specified color value

Variants

CurrentColor

The ‘currentColor’ keyword

Absolute(Box<Absolute>)

An absolute color. https://w3c.github.io/csswg-drafts/css-color-4/#typedef-absolute-color-function

ColorMix(Box<ColorMix>)

A color mix.

LightDark(Box<LightDark>)

A light-dark() color.

Implementations

impl Color

fn parse_internal<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, preserve_authored: PreserveAuthored ) -> Result<Self, ParseError<'i>>

pub fn is_valid(context: &ParserContext<'_>, input: &mut Parser<'_, '_>) -> bool

Returns whether a given color is valid for authors.

pub fn parse_and_compute( context: &ParserContext<'_>, input: &mut Parser<'_, '_>, device: Option<&Device> ) -> Option<ComputedColor>

Tries to parse a color and compute it with a given device.

impl Color

pub fn honored_in_forced_colors_mode(&self, allow_transparent: bool) -> bool

Returns whether this color is allowed in forced-colors mode.

pub fn currentcolor() -> Self

Returns currentcolor value.

pub fn transparent() -> Self

Returns transparent value.

pub fn from_absolute_color(color: AbsoluteColor) -> Self

Create a color from an AbsoluteColor.

pub fn parse_quirky<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, allow_quirks: AllowQuirks ) -> Result<Self, ParseError<'i>>

Parse a color, with quirks.

https://quirks.spec.whatwg.org/#the-hashless-hex-color-quirk

fn parse_quirky_color<'i, 't>( input: &mut Parser<'i, 't> ) -> Result<Self, ParseError<'i>>

Parse a value.

https://quirks.spec.whatwg.org/#the-hashless-hex-color-quirk

impl Color

pub fn to_computed_color( &self, context: Option<&Context<'_>> ) -> Option<ComputedColor>

Converts this Color into a ComputedColor.

If context is None, and the specified color requires data from the context to resolve, then None is returned.

impl Color

fn has_modern_syntax(&self) -> bool