Enum style::values::specified::font::MozScriptLevel

pub enum MozScriptLevel {
    Relative(i32),
    MozAbsolute(i32),
    Auto,
}

Changes the scriptlevel in effect for the children. Ref: https://wiki.mozilla.org/MathML:mstyle

The main effect of scriptlevel is to control the font size. https://www.w3.org/TR/MathML3/chapter3.html#presm.scriptlevel

Variants
Relative(i32)

Change font-size relatively.

MozAbsolute(i32)

Change font-size absolutely.

Should only be serialized by presentation attributes, so even though serialization for this would look the same as for the Relative variant, it is unexposed, so no big deal.

Auto

Change font-size automatically.