Struct style::values::specified::font::FontVariantEastAsian

pub struct FontVariantEastAsian {
    bits: u16,
}

Vairants for east asian variant

Fields
bits: u16
Implementations

impl FontVariantEastAsian

pub const NORMAL: Self = _

None of the features

pub const JIS78: Self = _

Enables rendering of JIS78 forms (OpenType feature: jp78)

pub const JIS83: Self = _

Enables rendering of JIS83 forms (OpenType feature: jp83).

pub const JIS90: Self = _

Enables rendering of JIS90 forms (OpenType feature: jp90).

pub const JIS04: Self = _

Enables rendering of JIS2004 forms (OpenType feature: jp04).

pub const SIMPLIFIED: Self = _

Enables rendering of simplified forms (OpenType feature: smpl).

pub const TRADITIONAL: Self = _

Enables rendering of traditional forms (OpenType feature: trad).

pub const FULL_WIDTH: Self = _

Enables rendering of full-width variants (OpenType feature: fwid).

pub const PROPORTIONAL_WIDTH: Self = _

Enables rendering of proportionally-spaced variants (OpenType feature: pwid).

pub const RUBY: Self = _

Enables display of ruby variant glyphs (OpenType feature: ruby).

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u16

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u16) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u16) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u16) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.