Module style::values::specified::image

CSS handling for the specified value of images

Structs

ColorStop
Specified values for one color stop in a linear gradient. https://drafts.csswg.org/css-images/#typedef-color-stop-list

Gradient
Specified values for a CSS gradient. https://drafts.csswg.org/css-images/#gradients

Enums

AngleOrCorner
Specified values for an angle or a corner in a linear gradient.

EndingShape
Determines whether the gradient's ending shape is a circle or an ellipse. If is omitted, the ending shape defaults to a circle if the is a single , and to an ellipse otherwise. https://drafts.csswg.org/css-images/#valdef-radial-gradient-ending-shape

GradientKind
Specified values for CSS linear or radial gradients. https://drafts.csswg.org/css-images/#gradients

HorizontalDirection

Image
Specified values for an image according to CSS-IMAGES. https://drafts.csswg.org/css-images/#image-values

LengthOrKeyword
https://drafts.csswg.org/css-images/#valdef-radial-gradient-size

LengthOrPercentageOrKeyword
https://drafts.csswg.org/css-images/#valdef-radial-gradient-size

SizeKeyword

VerticalDirection