Type Alias style::values::specified::image::CrossFadeElement

pub type CrossFadeElement = CrossFadeElement<Image, Color, Percentage>;

CrossFadeElement = percent? CrossFadeImage

Aliased Type

struct CrossFadeElement {
    pub percent: Optional<Percentage>,
    pub image: GenericCrossFadeImage<GenericImage<GenericGradient<LineDirection, LengthPercentage, NonNegative<Length>, NonNegative<LengthPercentage>, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, Angle, AngleOrPercentage, Color>, MozImageRect, CssUrl, Color, Percentage, Resolution>, Color>,
}

Fields
percent: Optional<Percentage>

The percent of the final image that image will be.

image: GenericCrossFadeImage<GenericImage<GenericGradient<LineDirection, LengthPercentage, NonNegative<Length>, NonNegative<LengthPercentage>, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, Angle, AngleOrPercentage, Color>, MozImageRect, CssUrl, Color, Percentage, Resolution>, Color>

A color or image that will be blended when cross-fade is evaluated.

Implementations

impl CrossFadeElement

fn parse_percentage<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't> ) -> Option<Percentage>

fn parse<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, cors_mode: CorsMode, flags: ParseImageFlags ) -> Result<Self, ParseError<'i>>

= ? && [ | ]