Enum style::values::specified::length::Length

pub enum Length {
    NoCalc(NoCalcLength),
    Calc(Box<CalcLengthOrPercentage>, AllowedNumericType),
}

An extension to NoCalcLength to parse calc expressions. This is commonly used for the <length> values.

https://drafts.csswg.org/css-values/#lengths

Variants
NoCalc(NoCalcLength)

The internal length type that cannot parse calc

Calc(Box<CalcLengthOrPercentage>, AllowedNumericType)

A calc expression.

https://drafts.csswg.org/css-values/#calc-notation

TODO(emilio): We have more Calc variants around, we should only use one.

Methods

impl Length

fn zero() -> Length

Returns a zero length.

fn from_str(s: &str) -> Option<Length>

https://drafts.csswg.org/css-fonts-3/#font-size-prop

fn parse_dimension(value: CSSFloat, unit: &str) -> Result<Length, ()>

Parse a given absolute or relative dimension.

fn from_font_size_int(i: u8) -> Self

https://drafts.csswg.org/css-fonts-3/#font-size-prop

fn parse_non_negative(input: &mut Parser) -> Result<Length, ()>

Parse a non-negative length

fn from_px(px_value: CSSFloat) -> Length

Get an absolute length from a px value.

fn take(&mut self) -> Self

Extract inner length without a clone, replacing it with a 0 Au

Use when you need to move out of a length array without cloning