Enum style::values::specified::length::NoCalcLength

pub enum NoCalcLength {
    Absolute(AbsoluteLength),
    FontRelative(FontRelativeLength),
    ViewportPercentage(ViewportPercentageLength),
    ServoCharacterWidth(CharacterWidth),
}

A <length> without taking calc expressions into account

https://drafts.csswg.org/css-values/#lengths

Variants

Absolute(AbsoluteLength)

An absolute length

https://drafts.csswg.org/css-values/#absolute-length

FontRelative(FontRelativeLength)

A font-relative length:

https://drafts.csswg.org/css-values/#font-relative-lengths

ViewportPercentage(ViewportPercentageLength)

A viewport-relative length.

https://drafts.csswg.org/css-values/#viewport-relative-lengths

ServoCharacterWidth(CharacterWidth)

HTML5 "character width", as defined in HTML5  14.5.4.

This cannot be specified by the user directly and is only generated by Stylist::synthesize_rules_for_legacy_attributes().

Methods

impl NoCalcLength

pub fn parse_dimension(
context: &ParserContext,
value: CSSFloat,
unit: &str
) -> Result<Self, ()>

Parse a given absolute or relative dimension.

pub(crate) fn try_sum(&self, other: &Self) -> Result<Self, ()>

Try to sume two lengths if compatible into the left hand side.

pub fn to_computed_pixel_length_without_context(&self) -> Result<CSSFloat, ()>

Get a px value without context.

pub fn from_px(px_value: CSSFloat) -> NoCalcLength

Get an absolute length from a px value.