style::values::specified::length

Struct NoCalcLength

#[repr(C)]
pub struct NoCalcLength {
    unit: LengthUnit,
    value: CSSFloat,
}

A <length> without taking calc expressions into account

https://drafts.csswg.org/css-values/#lengths

Fields
unit: LengthUnitvalue: CSSFloat
Implementations

impl NoCalcLength

pub const EM: &'static str = "em"

Unit identifier for em.

pub const EX: &'static str = "ex"

Unit identifier for ex.

pub const REX: &'static str = "rex"

Unit identifier for rex.

pub const CH: &'static str = "ch"

Unit identifier for ch.

pub const RCH: &'static str = "rch"

Unit identifier for rch.

pub const CAP: &'static str = "cap"

Unit identifier for cap.

pub const RCAP: &'static str = "rcap"

Unit identifier for rcap.

pub const IC: &'static str = "ic"

Unit identifier for ic.

pub const RIC: &'static str = "ric"

Unit identifier for ric.

pub const REM: &'static str = "rem"

Unit identifier for rem.

pub const LH: &'static str = "lh"

Unit identifier for lh.

pub const RLH: &'static str = "rlh"

Unit identifier for rlh.

pub fn new(unit: LengthUnit, value: CSSFloat) -> Self

Creates a length with the given unit and value.

pub fn length_unit(&self) -> LengthUnit

Returns the unit of this length.

pub fn unitless_value(&self) -> CSSFloat

Return the unitless, raw value.

pub fn unit(&self) -> &'static str

Return the unit, as a string.

pub fn canonical_unit(&self) -> Option<&'static str>

Return the canonical unit for this value, if one exists.

pub fn to(&self, unit: &str) -> Result<Self, ()>

Convert this value to the specified unit, if possible.

pub fn is_negative(&self) -> bool

Returns whether the value of this length without unit is less than zero.

pub fn is_zero(&self) -> bool

Returns whether the value of this length without unit is equal to zero.

pub fn is_infinite(&self) -> bool

Returns whether the value of this length without unit is infinite.

pub fn is_nan(&self) -> bool

Returns whether the value of this length without unit is NaN.

pub fn should_zoom_text(&self) -> bool

Whether text-only zoom should be applied to this length.

Generally, font-dependent/relative units don’t get text-only-zoomed, because the font they’re relative to should be zoomed already.

pub(crate) fn sort_key(&self) -> SortKey

Returns the SortKey for this length. Must not be called on the internal ServoCharacterWidth unit.

pub fn parse_dimension_with_flags( parsing_mode: ParsingMode, in_page_rule: bool, value: CSSFloat, unit: &str, ) -> Result<Self, ()>

Parse a given absolute or relative dimension.

pub fn parse_dimension_with_context( context: &ParserContext<'_>, value: CSSFloat, unit: &str, ) -> Result<Self, ()>

Parse a given absolute or relative dimension.

pub(crate) fn try_op<O>(&self, other: &Self, op: O) -> Result<Self, ()>
where O: Fn(f32, f32) -> f32,

pub(crate) fn map(&self, op: impl FnMut(f32) -> f32) -> Self

pub fn to_computed_pixel_length_without_context(&self) -> Result<CSSFloat, ()>

Get a px value without context (so only absolute units can be handled).

pub fn from_px(px_value: CSSFloat) -> NoCalcLength

Get an absolute length from a px value.

pub fn to_px_if_absolute(&self) -> Option<CSSFloat>

Returns the value as a px-canonical length (absolute lengths only).

pub fn from_em(value: CSSFloat) -> Self

Construct a font-relative em value.

pub fn from_servo_character_width(value: i32) -> Self

Construct an internal ServoCharacterWidth length from an i32 column count.

fn font_relative_to_computed_value( &self, context: &Context<'_>, base_size: FontBaseSize, line_height_base: LineHeightBase, ) -> Length

Compute a font-relative length against the given base sizes. Must only be called on a font-relative unit.

fn reference_font_size_and_length( &self, context: &Context<'_>, base_size: FontBaseSize, line_height_base: LineHeightBase, ) -> (Length, CSSFloat)

fn viewport_percentage_to_computed_value( &self, context: &Context<'_>, ) -> CSSPixelLength

Compute the viewport-percentage length. Must only be called on a viewport-relative unit.

fn container_relative_to_computed_value( &self, context: &Context<'_>, ) -> CSSPixelLength

Compute the container-relative length. Must only be called on a container-relative unit.

fn servo_character_width_to_computed_value( &self, reference_font_size: Length, ) -> Length

Computes a ServoCharacterWidth length against a reference font size.

pub fn to_computed_value_with_base_size( &self, context: &Context<'_>, base_size: FontBaseSize, line_height_base: LineHeightBase, ) -> CSSPixelLength

Computes a length with a given font-relative base size.