Type Definition style::values::specified::length::LengthOrAuto

type LengthOrAuto = LengthPercentageOrAuto<Length>;

Either a <length> or the auto keyword.

Note that we use LengthPercentage just for convenience, since it pretty much is everything we care about, but we could just add a similar LengthOrAuto instead if we think getting rid of this weirdness is worth it.

Implementations

impl LengthOrAuto

pub fn parse_quirky<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
allow_quirks: AllowQuirks
) -> Result<Self, ParseError<'i>>

Parses a length, allowing the unitless length quirk. https://quirks.spec.whatwg.org/#the-unitless-length-quirk