style::values::specified::length

Type Alias NonNegativeLengthPercentage

pub type NonNegativeLengthPercentage = NonNegative<LengthPercentage>;

A wrapper of LengthPercentage, whose value must be >= 0.

Aliased Type

#[repr(transparent)]
pub struct NonNegativeLengthPercentage(pub LengthPercentage);

Tuple Fields
0: LengthPercentage
Implementations

impl NonNegativeLengthPercentage

pub fn zero_percent() -> Self

Returns a 0% value.

pub fn parse_quirky<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, allow_quirks: AllowQuirks, ) -> Result<Self, ParseError<'i>>

Parses a length or a percentage, allowing the unitless length quirk. https://quirks.spec.whatwg.org/#the-unitless-length-quirk

pub fn parse_non_negative_with_anchor_size<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, allow_quirks: AllowQuirks, ) -> Result<Self, ParseError<'i>>

Parses a length or a percentage, allowing the unitless length quirk, as well as allowing anchor-size().