style::values::specified::motion

Type Alias OffsetPath

pub type OffsetPath = GenericOffsetPath<OffsetPathFunction>;

The specified value of offset-path.

Aliased Type

#[repr(C, u8)]
pub enum OffsetPath {
    OffsetPath {
        path: Box<GenericOffsetPathFunction<GenericBasicShape<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, LengthPercentage, BasicShapeRect>, GenericRayFunction<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>>, CssUrl>>,
        coord_box: CoordBox,
    },
    CoordBox(CoordBox),
    None,
}

Variants

OffsetPath

|| .

Fields

path: Box<GenericOffsetPathFunction<GenericBasicShape<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, LengthPercentage, BasicShapeRect>, GenericRayFunction<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>>, CssUrl>>

part.

coord_box: CoordBox

part.

CoordBox(CoordBox)

Only . This represents that is omitted, so we use the default value, inset(0 round X), where X is the value of border-radius on the element that establishes the containing block for this element.

None

None value.