style::values::specified::motion

Type Alias OffsetPathFunction

pub type OffsetPathFunction = GenericOffsetPathFunction<BasicShape, RayFunction, SpecifiedUrl>;

The specified value of .

Aliased Type

#[repr(C, u8)]
pub enum OffsetPathFunction {
    Ray(GenericRayFunction<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>>),
    Url(CssUrl),
    Shape(GenericBasicShape<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, LengthPercentage, BasicShapeRect>),
}

Variants

Ray(GenericRayFunction<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>>)

ray() function, which defines a path in the polar coordinate system. Use Box<> to make sure the size of offset-path is not too large.

Url(CssUrl)

A URL reference to an SVG shape element. If the URL does not reference a shape element, this behaves as path(“m 0 0”) instead.

Shape(GenericBasicShape<Angle, GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>, LengthPercentage, BasicShapeRect>)

The value.