style::values::specified::motion

Type Alias OffsetPosition

pub type OffsetPosition = GenericOffsetPosition<HorizontalPosition, VerticalPosition>;

The specified value of offset-position.

Aliased Type

#[repr(C, u8)]
pub enum OffsetPosition {
    Normal,
    Auto,
    Position(GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>),
}

Variants

Normal

The element does not have an offset starting position.

Auto

The offset starting position is the top-left corner of the box.

Position(GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>)

The offset starting position is the result of using the to position a 0x0 object area within the box’s containing block.