style::values::specified::motion

Type Alias RayFunction

pub type RayFunction = GenericRayFunction<Angle, Position>;

The specified value of ray() function.

Aliased Type

#[repr(C)]
pub struct RayFunction {
    pub angle: Angle,
    pub size: RaySize,
    pub contain: bool,
    pub position: GenericPositionOrAuto<GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>>,
}

Fields
angle: Angle

The bearing angle with 0deg pointing up and positive angles representing clockwise rotation.

size: RaySize

Decide the path length used when offset-distance is expressed as a percentage.

contain: bool

Clamp offset-distance so that the box is entirely contained within the path.

position: GenericPositionOrAuto<GenericPosition<PositionComponent<HorizontalPositionKeyword>, PositionComponent<VerticalPositionKeyword>>>

The “at ” part. If omitted, we use auto to represent it.

Implementations

impl RayFunction

fn parse_function_arguments<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<Self, ParseError<'i>>

Parse the inner arguments of a ray function.