style::values::specified::percentage

Trait ToPercentage

pub trait ToPercentage {
    // Required method
    fn to_percentage(&self) -> Option<CSSFloat>;

    // Provided method
    fn is_calc(&self) -> bool { ... }
}

Turns the percentage into a plain float.

Required Methods

fn to_percentage(&self) -> Option<CSSFloat>

Returns the percentage as a plain float, or None for calc expressions that require computed context. Will always return Some if is_calc is false.

Provided Methods

fn is_calc(&self) -> bool

Returns whether this percentage used to be a calc().