style::values::specified::percentage

Type Alias NonNegativePercentage

pub type NonNegativePercentage = NonNegative<Percentage>;

A wrapper of Percentage, whose value must be >= 0.

Aliased Type

#[repr(transparent)]
pub struct NonNegativePercentage(pub Percentage);

Tuple Fields
0: Percentage
Implementations

impl NonNegativePercentage

pub fn compute(&self) -> Option<ComputedPercentage>

Convert to ComputedPercentage, for FontFaceRule size-adjust getter. Returns None if the value is a calc expression that cannot be resolved at parse time.