style::values::specified

Module position

CSS handling for the specified value of positions

Structs

AnchorNameIdent

https://drafts.csswg.org/css-anchor-position-1/#propdef-anchor-name

DashedIdentAndOrTryTactic

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-try-fallbacks ||

GridAutoFlow

Controls how the auto-placement algorithm works specifying exactly how auto-placed items get flowed into the grid: [ row | column ] || dense https://drafts.csswg.org/css-grid-2/#grid-auto-flow-property

MasonryAutoFlow

Controls how the Masonry layout algorithm works specifying exactly how auto-placed items get flowed in the masonry axis.

NamedArea

Not associated with any particular grid item, but can be referenced from the grid-placement properties.

PositionArea

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-area

PositionTryFallbacksList

https://drafts.csswg.org/css-anchor-position-1/#position-try-fallbacks

PositionTryFallbacksTryTactic

Changes for the automatically-generated position option. Note that this is order-dependent - e.g. flip-start flip-inline != flip-inline flip-start.

PositionVisibility

Specified keyword values for the position-visibility property.

ScopedNameList

List of scoped names, or none.

TemplateAreas

https://drafts.csswg.org/css-grid/#named-grid-area

TemplateAreasArc

Arc type for Arc<TemplateAreas>

TemplateAreasParser

Parser for grid template areas.

TemplateAreasTokenizer 🔒

Tokenize the string into a list of the tokens, using longest-match semantics

UnsignedRange

A range of rows or columns. Using this instead of std::ops::Range for FFI purposes.

Enums

GridTemplateAreas

This property specifies named grid areas.

HorizontalPositionKeyword

A keyword for the X direction.

MasonryItemOrder

Masonry auto-placement algorithm item sorting option.

MasonryPlacement

Masonry auto-placement algorithm packing.

PositionAnchorKeyword

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-anchor

PositionAreaAxis

A three-bit value that represents the axis in which position-area operates on. Represented as 4 bits: axis type (physical or logical), direction type (physical or logical), axis value.

PositionAreaKeyword

Possible values for the position-area property’s keywords. Represented by [0z xxx yyy], where z means “self wm resolution”, xxxx is the axis (as in PositionAreaAxis) and yyy is the PositionAreaTrack https://drafts.csswg.org/css-anchor-position-1/#propdef-position-area

PositionAreaTrack

Specifies which tracks(s) on the axis that the position-area span occupies. Represented as 3 bits: start, center, end track.

PositionAreaType

A value indicating which high level group in the formal grammar a PositionAreaKeyword or PositionArea belongs to.

PositionComponent

The specified value of a component of a CSS <position>.

PositionTryFallbacksItem

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-try-fallbacks [ [ || ] | <‘position-area’> ]

PositionTryFallbacksTryTacticKeyword

How to swap values for the automatically-generated position tactic.

PositionTryOrder

https://drafts.csswg.org/css-anchor-position-1/#position-try-order-property

VerticalPositionKeyword

A keyword for the Y direction.

Constants

AXIS_MASK

The mask used to extract the axis.

AXIS_SHIFT

The shift to the left needed to set the axis.

SELF_WM

The self-wm bit.

TRACK_MASK

The mask used to extract the track.

Traits

Side

Represents a side, either horizontal or vertical, of a CSS position.

Functions

is_item_order_definite_first 🔒

is_name_code_point 🔒

is_pack_with_non_default_order 🔒

Type Aliases

AnchorFunction

A specified value for anchor() function.

AnchorName

https://drafts.csswg.org/css-anchor-position-1/#propdef-anchor-name

AspectRatio

A specified value for the aspect-ratio property.

HorizontalPosition

The specified value of a horizontal position.

Inset

A specified value for inset types.

Position

The specified value of a CSS <position>

PositionAnchor

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-anchor

PositionOrAuto

The specified value of an auto | <position>.

PositionTryFallbacks

https://drafts.csswg.org/css-anchor-position-1/#position-try-fallbacks

ScopedName

A scoped name type, such as:

VerticalPosition

The specified value of a vertical position.

ZIndex

A specified value for the z-index property.