Type Definition style::values::specified::position::LegacyPosition

type LegacyPosition = GenericPosition<LegacyHPosition, LegacyVPosition>;

The specified value of a legacy CSS <position> Modern position syntax supports 3 and 4-value syntax. That means: If three or four values are given, then each or represents an offset and must be preceded by a keyword, which specifies from which edge the offset is given. For example, bottom 10px right 20px represents a 10px vertical offset up from the bottom edge and a 20px horizontal offset leftward from the right edge. If three values are given, the missing offset is assumed to be zero. But for some historical reasons we need to keep CSS Level 2 syntax which only supports up to 2-value. This type represents this 2-value syntax.

Methods

impl LegacyPosition

pub fn parse_quirky<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
allow_quirks: AllowQuirks
) -> Result<Self, ParseError<'i>>

Parses a <position>, with quirks.

pub fn center() -> Self

center center