style::values::specified

Module random

Specified types for the <random-key> of the CSS random() function. https://drafts.csswg.org/css-values-5/#randomness

Structs

RandomCacheKey

A that specifies parts of the random cache name.

RandomUaIdent

The part of a . This is a starting with ua- that is typically automatically constructed by the user agent.

Enums

RandomKey

A that specifies how to compute the random base value for a random function. Note that the auto value is not modeled as it is immediately converted to the element-scoped property-index-scoped cache key format during parsing.