Struct style::values::specified::LengthAndPercentage

pub struct LengthAndPercentage {
    pub lengths: Vec<Length>,
    pub percentage: CSSFloat,
}

The sum of a series of lengths and a percentage. This is used in calc() and other things that effectively work like it (e.g. transforms).

Fields

lengths

The length components.

percentage

The percentage component.

Methods

impl LengthAndPercentage

fn zero() -> LengthAndPercentage

fn from_length_or_percentage(length_or_percentage: &LengthOrPercentage) -> LengthAndPercentage

fn parse(input: &mut Parser) -> Result<LengthAndPercentage, ()>

fn from_length(length: Length) -> LengthAndPercentage

fn from_percentage(percentage: CSSFloat) -> LengthAndPercentage