style::values::specified

Module text

Specified types for text properties.

Structs

LetterSpacing

A specified value for the letter-spacing property.

TextAutospace

Specified value for the text-autospace property which takes the grammar: normal | | auto where: = no-autospace | [ ideograph-alpha || ideograph-numeric || punctuation ] || [ insert | replace ]

TextBoxTrim

Specified value for the text-box-trim property.

TextDecorationLine

Specified keyword values for the text-decoration-line property.

TextEdge

The typedef, used by the line-fit-edge and text-box-edge properties.

TextEmphasisPosition

Values for text-emphasis-position: https://drafts.csswg.org/css-text-decor/#text-emphasis-position-property

TextOverflow

text-overflow. When the specified value only has one side, that’s the “second” side, and the sides are logical, so “second” means “end”. The start side is Clip in that case.

TextTransform

Specified value for the text-transform property. (The spec grammar gives none | math-auto | [capitalize | uppercase | lowercase] || full-width || full-size-kana.) https://drafts.csswg.org/css-text-4/#text-transform-property

TextUnderlinePosition

Specified keyword values for the text-underline-position property. (Non-exclusive, but not all combinations are allowed: the spec grammar gives auto | [ from-font | under ] || [ left | right ].) https://drafts.csswg.org/css-text-decor-4/#text-underline-position-property

WordSpacing

A specified value for the word-spacing property.

Enums

HyphenateCharacter

A value for the hyphenate-character property.

LineBreak

Values for the line-break property.

MozControlCharacterVisibility

Values for the -moz-control-character-visibility CSS property.

OverflowWrap

Values for the overflow-wrap property.

RubyPosition

Values for ruby-position property

Spacing

A spacing value used by either the letter-spacing or word-spacing properties.

TextAlign

Specified value of text-align property.

TextAlignKeyword

Specified value of text-align keyword value.

TextAlignLast

Specified and computed value of text-align-last.

TextBoxEdge

Specified value for the text-box-edge property.

TextDecorationSkipInk

Implements text-decoration-skip-ink which takes the keywords auto | none | all

TextEdgeKeyword

Identifies specific font metrics for use in the typedef.

TextEmphasisFillMode

Fill mode for the text-emphasis-style property

TextEmphasisShapeKeyword

Shape keyword for the text-emphasis-style property

TextEmphasisStyle

Specified value of text-emphasis-style property.

TextJustify

Values for the text-justify CSS property.

TextOverflowSide

A generic value for the text-overflow property.

TextTransformCase

Specified keyword values for case transforms in the text-transform property. (These are exclusive.)

WordBreak

Values for the word-break property.

Functions

fill_mode_is_default_and_shape_exists 🔒

parse_inset_endpoint 🔒

Type Aliases

HyphenateLimitChars

A value for the hyphenate-limit-chars property.

InitialLetter

A specified type for the initial-letter property.

TextDecorationInset

Implements type for text-decoration-inset property

TextDecorationLength

Implements type for text-decoration-thickness property

TextIndent

A specified value for the text-indent property which takes the grammar of [] && hanging? && each-line?