style::values::specified::text

Type Alias TextIndent

pub type TextIndent = GenericTextIndent<LengthPercentage>;

A specified value for the text-indent property which takes the grammar of [] && hanging? && each-line?

https://drafts.csswg.org/css-text/#propdef-text-indent

Aliased Type

#[repr(C)]
pub struct TextIndent {
    pub length: LengthPercentage,
    pub hanging: bool,
    pub each_line: bool,
}

Fields
length: LengthPercentage

The amount of indent to be applied to the inline-start of the first line.

hanging: bool

Apply indent to non-first lines instead of first.

each_line: bool

Apply to each line after a hard break, not only first in block.