Type Alias style::values::specified::transform::Transform

pub type Transform = Transform<TransformOperation>;

A specified CSS transform

Aliased Type

struct Transform(pub OwnedSlice<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>);

Fields
0: OwnedSlice<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>
Implementations

impl Transform

fn parse_internal<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<Self, ParseError<'i>>

Internal parse function for deciding if we wish to accept prefixed values or not

transform allows unitless zero angles as an exception, see: https://github.com/w3c/csswg-drafts/issues/1162