style::values::specified

Type Alias GridLine

pub type GridLine = GridLine<Integer>;

The specified value of a <grid-line>.

Aliased Type

#[repr(C)]
pub struct GridLine {
    pub ident: CustomIdent,
    pub line_num: Integer,
    pub is_span: bool,
}

Fields
ident: CustomIdent

A custom identifier for named lines, or the empty atom otherwise.

https://drafts.csswg.org/css-grid/#grid-placement-slot

line_num: Integer

Denotes the nth grid line from grid item’s placement.

This is clamped by MIN_GRID_LINE and MAX_GRID_LINE.

NOTE(emilio): If we ever allow animating these we need to either do something more complicated for the clamping, or do this clamping at used-value time.

is_span: bool

Flag to check whether it’s a span keyword.