Struct style::values::specified::url::SpecifiedUrl

pub struct SpecifiedUrl {
    original: Option<Arc<String>>,
    resolved: Option<ServoUrl>,
    extra_data: UrlExtraData,
}

A specified url() value.

Fields
original: Option<Arc<String>>

The original URI. This might be optional since we may insert computed values of images into the cascade directly, and we don't bother to convert their serialization.

Refcounted since cloning this should be cheap and data: uris can be really large.

resolved: Option<ServoUrl>

The resolved value for the url, if valid.

extra_data: UrlExtraData

Extra data used for Stylo.

Methods

impl SpecifiedUrl

fn extra_data(&self) -> &UrlExtraData

fn url(&self) -> Option<&ServoUrl>

fn as_str(&self) -> &str

fn as_slice_components(&self) -> (*const u8, usize)

Little helper for Gecko's ffi.

fn for_cascade(url: Option<ServoUrl>, extra_data: UrlExtraData) -> Self

Creates an already specified url value from an already resolved URL for insertion in the cascade.

fn new_for_testing(url: &str) -> Self