Struct style::values::RGBA

pub struct RGBA {
    pub red: u8,
    pub green: u8,
    pub blue: u8,
    pub alpha: u8,
}

A color with red, green, blue, and alpha components, in a byte each.

Fields
red: u8

The red component.

green: u8

The green component.

blue: u8

The blue component.

alpha: u8

The alpha component.

Methods

impl RGBA

fn from_floats(red: f32, green: f32, blue: f32, alpha: f32) -> RGBA

Constructs a new RGBA value from float components. It expects the red, green, blue and alpha channels in that order, and all values will be clamped to the 0.0 ... 1.0 range.

fn transparent() -> RGBA

Returns a transparent color.

fn new(red: u8, green: u8, blue: u8, alpha: u8) -> RGBA

Same thing, but with u8 values instead of floats in the 0 to 1 range.

fn red_f32(&self) -> f32

Returns the red channel in a floating point number form, from 0 to 1.

fn green_f32(&self) -> f32

Returns the green channel in a floating point number form, from 0 to 1.

fn blue_f32(&self) -> f32

Returns the blue channel in a floating point number form, from 0 to 1.

fn alpha_f32(&self) -> f32

Returns the alpha channel in a floating point number form, from 0 to 1.