Crate style

Calculate specified and computed values from a tree of DOM nodes and a set of stylesheets.

In particular, this crate contains the definitions of supported properties, the code to parse them into specified values and calculate the computed values based on the specified values, as well as the code to serialize both specified and computed values.

The main entry point is recalc_style_at.

Major dependencies are the cssparser and selectors crates.

Modules

animation
CSS transitions and animations.

attr
Parsed representations of DOM attributes.

bezier
Parametric Bézier curves.

computed_values
Types for computed values.

context
The context within which style is calculated.

custom_properties
Support for custom properties for cascading variables.

data
Per-node data used in style calculation.

dom
Types and traits used to access the DOM from style calculation.

element_state
States elements can be in.

error_reporting
Types used to report parsing errors.

font_face
The @font-face at-rule.

logical_geometry
Geometry in flow-relative space.

matching
High-level interface to CSS selector matching.

media_queries
Media queries.

parallel
Implements parallel traversal over the DOM tree.

parser
The context within which CSS code is parsed.

properties
The CSS properties supported by the style system.

refcell
A fork of std::cell::RefCell that makes as_unsafe_cell usable on stable Rust.

restyle_hints
Restyle hints: an optimization to avoid unnecessarily matching selectors.

selector_impl
The pseudo-classes and pseudo-elements supported by the style system.

selector_matching
Selector matching.

sequential
Implements sequential traversal over the DOM tree.

servo
Concrete types for servo Style implementation

stylesheets
Style sheets and their CSS rules.

traversal
Traversing the DOM tree; the bloom filter.

values
Common values used in CSS.

viewport
The @viewport at-rule and meta element.

Macros

css_properties_accessors!