Module style::animation

Structs

Animation
State relating to an animation.

PropertyAnimation

Enums

AnimatedProperty

Traits

GetMod
Accesses an element of an array, "wrapping around" using modular arithmetic. This is needed to handle values of differing lengths according to CSS-TRANSITIONS  2.

Interpolate

Functions

build_identity_transform_list
Build an equivalent 'identity transform function list' based on an existing transform list. http://dev.w3.org/csswg/css-transforms/#none-transform-animation

can_interpolate_list
Check if it's possible to do a direct numerical interpolation between these two transform lists. http://dev.w3.org/csswg/css-transforms/#transform-transform-animation

interpolate_transform_list
Interpolate two transform lists. http://dev.w3.org/csswg/css-transforms/#interpolation-of-transforms

start_transitions_if_applicable
Inserts transitions into the queue of running animations as applicable for the given style difference. This is called from the layout worker threads. Returns true if any animations were kicked off and false otherwise.

update_style_for_animation
Updates a single animation and associated style based on the current time. If damage is provided, inserts the appropriate restyle damage.