Module style::animation

CSS transitions and animations.

Structs

AnimationFrame
A single animation frame of a single property.

KeyframesAnimationState
This structure represents the current keyframe animation state, i.e., the duration, the current and maximum iteration count, and the state (either playing or paused).

PropertyAnimation
Represents an animation for a given property.

Enums

Animation
State relating to an animation.

AnimationUpdate
Returns the kind of animation update that happened.

KeyframesIterationState
This structure represents a keyframes animation current iteration state.

KeyframesRunningState
This structure represents wether an animation is actually running.

Functions

complete_expired_transitions
Update the style in the node when it finishes.

compute_style_for_animation_step

maybe_start_animations
Triggers animations for a given node looking at the animation property values.

start_transitions_if_applicable
Inserts transitions into the queue of running animations as applicable for the given style difference. This is called from the layout worker threads. Returns true if any animations were kicked off and false otherwise.

update_style_for_animation
Updates a single animation and associated style based on the current time.

update_style_for_animation_frame
Updates a given computed style for a given animation frame. Returns a bool representing if the style was indeed updated.