Struct style::animation::Animation

pub struct Animation {
    pub node: OpaqueNode,
    pub property_animation: PropertyAnimation,
    pub start_time: f64,
    pub end_time: f64,
}

State relating to an animation.

Fields
node: OpaqueNode

An opaque reference to the DOM node participating in the animation.

property_animation: PropertyAnimation

A description of the property animation that is occurring.

start_time: f64

The start time of the animation, as returned by time::precise_time_s().

end_time: f64

The end time of the animation, as returned by time::precise_time_s().

Methods

impl Animation

fn duration(&self) -> f64

Returns the duration of this animation in seconds.