Struct style::animation::Animation

pub struct Animation {

    pub name: Atom,
    properties_changed: LonghandIdSet,
    computed_steps: Vec<ComputedKeyframe>,
    pub started_at: f64,
    pub duration: f64,
    pub delay: f64,
    pub fill_mode: AnimationFillMode,
    pub iteration_state: KeyframesIterationState,
    pub state: AnimationState,
    pub direction: AnimationDirection,
    pub current_direction: AnimationDirection,
    pub cascade_style: Arc<ComputedValues>,
    pub is_new: bool,

}

A CSS Animation

Fields
name: Atom

The name of this animation as defined by the style.

properties_changed: LonghandIdSet

The properties that change in this animation.

computed_steps: Vec<ComputedKeyframe>

The computed style for each keyframe of this animation.

started_at: f64

The time this animation started at, which is the current value of the animation timeline when this animation was created plus any animation delay.

duration: f64

The duration of this animation.

delay: f64

The delay of the animation.

fill_mode: AnimationFillMode

The animation-fill-mode property of this animation.

iteration_state: KeyframesIterationState

The current iteration state for the animation.

state: AnimationState

Whether this animation is paused.

direction: AnimationDirection

The declared animation direction of this animation.

current_direction: AnimationDirection

The current animation direction. This can only be normal or reverse.

cascade_style: Arc<ComputedValues>

The original cascade style, needed to compute the generated keyframes of the animation.

is_new: bool

Whether or not this animation is new and or has already been tracked by the script thread.

Implementations

impl Animation

fn is_cancelled_in_new_style(&self, new_style: &Arc<ComputedValues>) -> bool

Whether or not this animation is cancelled by changes from a new style.

pub fn iterate_if_necessary(&mut self, time: f64) -> bool

Given the current time, advances this animation to the next iteration, updates times, and then toggles the direction if appropriate. Otherwise does nothing. Returns true if this animation has iterated.

fn iterate(&mut self)

pub fn current_iteration_end_progress(&self) -> f64

A number (> 0 and <= 1) which represents the fraction of a full iteration that the current iteration of the animation lasts. This will be less than 1 if the current iteration is the fractional remainder of a non-integral iteration count.

pub fn current_iteration_duration(&self) -> f64

The duration of the current iteration of this animation which may be less than the animation duration if it has a non-integral iteration count.

fn iteration_over(&self, time: f64) -> bool

Whether or not the current iteration is over. Note that this method assumes that the animation is still running.

fn on_last_iteration(&self) -> bool

Assuming this animation is running, whether or not it is on the last iteration.

pub fn has_ended(&self, time: f64) -> bool

Whether or not this animation has finished at the provided time. This does not take into account canceling i.e. when an animation or transition is canceled due to changes in the style.

pub fn update_from_other(&mut self, other: &Self, now: f64)

Updates the appropiate state from other animation.

This happens when an animation is re-submitted to layout, presumably because of an state change.

There are some bits of state we can’t just replace, over all taking in account times, so here’s that logic.

fn get_property_declaration_at_time(
&self,
now: f64,
map: &mut AnimationValueMap
)

Fill in an AnimationValueMap with values calculated from this animation at the given time value.