Struct style::animation::PropertyAnimation

pub struct PropertyAnimation { /* fields omitted */ }

Represents an animation for a given property.

Methods

impl PropertyAnimation

pub fn property_name(&self) -> &'static str

Returns the given property name.

pub fn from_transition(
transition_index: usize,
old_style: &ComputedValues,
new_style: &mut ComputedValues
) -> Vec<PropertyAnimation>

Creates a new property animation for the given transition index and old and new styles. Any number of animations may be returned, from zero (if the property did not animate) to one (for a single transition property) to arbitrarily many (for all).

pub fn update(&self, style: &mut ComputedValues, time: f64)

Update the given animation at a given point of progress.

pub fn has_the_same_end_value_as(&self, other: &Self) -> bool

Whether this animation has the same end value as another one.