Module style::context

The context within which style is calculated.

Structs

CascadeInputs
The structure holds various intermediate inputs that are eventually used by by the cascade.

EagerPseudoCascadeInputs
A list of cascade inputs for eagerly-cascaded pseudo-elements. The list is stored inline.

ElementCascadeInputs
The cascade inputs associated with a node, including those for any pseudo-elements.

SelectorFlagsMap
Map from Elements to ElementSelectorFlags. Used to defer applying selector flags until after the traversal.

SequentialTaskList
A list of SequentialTasks that get executed on Drop.

SharedStyleContext
A shared style context.

StackLimitChecker
A helper type for stack limit checking. This assumes that stacks grow down, which is true for all non-ancient CPU architectures.

StyleContext
A StyleContext is just a simple container for a immutable reference to a shared style context, and a mutable reference to a local one.

StyleSystemOptions
A global options structure for the style system. We use this instead of opts to abstract across Gecko and Servo.

ThreadLocalStyleContext
A thread-local style context.

ThreadLocalStyleContextCreationInfo
This structure is used to create a local style context from a shared one.

TraversalStatistics
Statistics gathered during the traversal. We gather statistics on each thread and then combine them after the threads join via the Add implementation below.

Enums

QuirksMode
Which quirks mode is this document in.

SequentialTask
A task to be run in sequential mode on the parent (non-worker) thread. This is used by the style system to queue up work which is not safe to do during the parallel traversal.

Traits

RegisteredSpeculativePainter
A registered painter

RegisteredSpeculativePainters
A set of registered painters