Enum style::context::SequentialTask

pub enum SequentialTask<E: TElement> {
    SetSelectorFlags(SendElement<E>, ElementSelectorFlags),
}

A task to be run in sequential mode on the parent (non-worker) thread. This is used by the style system to queue up work which is not safe to do during the parallel traversal.

Variants
SetSelectorFlags(SendElement<E>, ElementSelectorFlags)

Sets selector flags. This is used when we need to set flags on an element that we don't have exclusive access to (i.e. the parent).

Methods

impl<E: TElement> SequentialTask<E>

fn execute(self)

Executes this task.

fn set_selector_flags(el: E, flags: ElementSelectorFlags) -> Self

Creates a task to set the selector flags on an element.