Struct style::context::CurrentElementInfo

struct CurrentElementInfo {
    element: OpaqueNode,
    is_initial_style: bool,
}

Information about the current element being processed. We group this together into a single struct within ThreadLocalStyleContext so that we can instantiate and destroy it easily at the beginning and end of element processing.

Fields
element: OpaqueNode

The element being processed. Currently we use an OpaqueNode since we only use this for identity checks, but we could use SendElement if there were a good reason to.

is_initial_style: bool

Whether the element is being styled for the first time.