Struct style::context::SharedStyleContext

pub struct SharedStyleContext<Impl: SelectorImplExt> {
    pub viewport_size: Size2D<Au>,
    pub screen_size_changed: bool,
    pub stylist: Arc<Stylist<Impl>>,
    pub generation: u32,
    pub new_animations_sender: Mutex<Sender<Animation>>,
    pub goal: ReflowGoal,
    pub running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub error_reporter: Box<ParseErrorReporter + Sync>,
}

Fields

viewport_size

The current viewport size.

screen_size_changed

Screen sized changed?

stylist

The CSS selector stylist.

generation

Starts at zero, and increased by one every time a layout completes. This can be used to easily check for invalid stale data.

new_animations_sender

A channel on which new animations that have been triggered by style recalculation can be sent.

goal

Why is this reflow occurring

running_animations

The animations that are currently running.

expired_animations

The list of animations that have expired since the last style recalculation.

error_reporter

The CSS error reporter for all CSS loaded in this layout thread