Struct style::context::SharedStyleContext

pub struct SharedStyleContext {
    pub viewport_size: Size2D<Au>,
    pub screen_size_changed: bool,
    pub stylist: Arc<Stylist>,
    pub goal: ReflowGoal,
    pub running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    pub error_reporter: Box<ParseErrorReporter + Sync>,
    pub local_context_creation_data: Mutex<ThreadLocalStyleContextCreationInfo>,
    pub timer: Timer,
    pub quirks_mode: QuirksMode,
    pub default_computed_values: Arc<ComputedValues>,
}

A shared style context.

There's exactly one of these during a given restyle traversal, and it's shared among the worker threads.

Fields
viewport_size: Size2D<Au>

The current viewport size.

screen_size_changed: bool

Screen sized changed?

stylist: Arc<Stylist>

The CSS selector stylist.

goal: ReflowGoal

Why is this reflow occurring

running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

The animations that are currently running.

expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

The list of animations that have expired since the last style recalculation.

error_reporter: Box<ParseErrorReporter + Sync>

The CSS error reporter for all CSS loaded in this layout thread

local_context_creation_data: Mutex<ThreadLocalStyleContextCreationInfo>

Data needed to create the thread-local style context from the shared one.

timer: Timer

The current timer for transitions and animations. This is needed to test them.

quirks_mode: QuirksMode

The QuirksMode state which the document needs to be rendered with

default_computed_values: Arc<ComputedValues>

The default computed values to use for elements with no rules applying to them.