Struct style::context::ThreadLocalStyleContext

pub struct ThreadLocalStyleContext<E: TElement> {
    pub style_sharing_candidate_cache: StyleSharingCandidateCache<E>,
    pub bloom_filter: StyleBloom<E>,
    pub new_animations_sender: Sender<Animation>,
    pub tasks: Vec<SequentialTask<E>>,
    pub statistics: TraversalStatistics,
    // some fields omitted
}

A thread-local style context.

This context contains data that needs to be used during restyling, but is not required to be unique among worker threads, so we create one per worker thread in order to be able to mutate it without locking.

Fields
style_sharing_candidate_cache: StyleSharingCandidateCache<E>

A cache to share style among siblings.

bloom_filter: StyleBloom<E>

The bloom filter used to fast-reject selector-matching.

new_animations_sender: Sender<Animation>

A channel on which new animations that have been triggered by style recalculation can be sent.

tasks: Vec<SequentialTask<E>>

A set of tasks to be run (on the parent thread) in sequential mode after the rest of the styling is complete. This is useful for infrequently-needed non-threadsafe operations.

statistics: TraversalStatistics

Statistics about the traversal.

Methods

impl<E: TElement> ThreadLocalStyleContext<E>

fn new(shared: &SharedStyleContext) -> Self

Creates a new ThreadLocalStyleContext from a shared one.

fn begin_element(&mut self, element: E, data: &ElementData)

Notes when the style system starts traversing an element.

fn end_element(&mut self, element: E)

Notes when the style system finishes traversing an element.

fn is_initial_style(&self) -> bool

Returns true if the current element being traversed is being styled for the first time.

Panics if called while no element is being traversed.