Struct style::counter_style::CounterStyleRuleData

pub struct CounterStyleRuleData {

    name: CustomIdent,
    generation: Wrapping<u32>,
    system: Option<System>,
    negative: Option<Negative>,
    prefix: Option<Symbol>,
    suffix: Option<Symbol>,
    range: Option<CounterRanges>,
    pad: Option<Pad>,
    fallback: Option<Fallback>,
    symbols: Option<Symbols>,
    additive_symbols: Option<AdditiveSymbols>,
    speak_as: Option<SpeakAs>,
    pub source_location: SourceLocation,

}

An @counter-style rule

Fields
name: CustomIdentgeneration: Wrapping<u32>system: Option<System>

https://drafts.csswg.org/css-counter-styles/#counter-style-system

negative: Option<Negative>

https://drafts.csswg.org/css-counter-styles/#counter-style-negative

prefix: Option<Symbol>

https://drafts.csswg.org/css-counter-styles/#counter-style-prefix

suffix: Option<Symbol>

https://drafts.csswg.org/css-counter-styles/#counter-style-suffix

range: Option<CounterRanges>

https://drafts.csswg.org/css-counter-styles/#counter-style-range

pad: Option<Pad>

https://drafts.csswg.org/css-counter-styles/#counter-style-pad

fallback: Option<Fallback>

https://drafts.csswg.org/css-counter-styles/#counter-style-fallback

symbols: Option<Symbols>

https://drafts.csswg.org/css-counter-styles/#descdef-counter-style-symbols

additive_symbols: Option<AdditiveSymbols>

https://drafts.csswg.org/css-counter-styles/#descdef-counter-style-additive-symbols

speak_as: Option<SpeakAs>

https://drafts.csswg.org/css-counter-styles/#counter-style-speak-as

source_location: SourceLocation

Line and column of the @counter-style rule source code.

Implementations

impl CounterStyleRuleData

fn empty(name: CustomIdent, source_location: SourceLocation) -> Self

pub fn system(&self) -> Option<&System>

https://drafts.csswg.org/css-counter-styles/#counter-style-system

pub fn negative(&self) -> Option<&Negative>

https://drafts.csswg.org/css-counter-styles/#counter-style-negative

pub fn prefix(&self) -> Option<&Symbol>

https://drafts.csswg.org/css-counter-styles/#counter-style-prefix

pub fn suffix(&self) -> Option<&Symbol>

https://drafts.csswg.org/css-counter-styles/#counter-style-suffix

pub fn range(&self) -> Option<&CounterRanges>

https://drafts.csswg.org/css-counter-styles/#counter-style-range

pub fn pad(&self) -> Option<&Pad>

https://drafts.csswg.org/css-counter-styles/#counter-style-pad

pub fn fallback(&self) -> Option<&Fallback>

https://drafts.csswg.org/css-counter-styles/#counter-style-fallback

pub fn symbols(&self) -> Option<&Symbols>

https://drafts.csswg.org/css-counter-styles/#descdef-counter-style-symbols

pub fn additive_symbols(&self) -> Option<&AdditiveSymbols>

https://drafts.csswg.org/css-counter-styles/#descdef-counter-style-additive-symbols

pub fn speak_as(&self) -> Option<&SpeakAs>

https://drafts.csswg.org/css-counter-styles/#counter-style-speak-as

pub fn set_system(&mut self, value: System) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-system

pub fn set_negative(&mut self, value: Negative) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-negative

pub fn set_prefix(&mut self, value: Symbol) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-prefix

pub fn set_suffix(&mut self, value: Symbol) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-suffix

pub fn set_range(&mut self, value: CounterRanges) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-range

pub fn set_pad(&mut self, value: Pad) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-pad

pub fn set_fallback(&mut self, value: Fallback) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-fallback

pub fn set_symbols(&mut self, value: Symbols) -> bool

https://drafts.csswg.org/css-counter-styles/#descdef-counter-style-symbols

pub fn set_additive_symbols(&mut self, value: AdditiveSymbols) -> bool

https://drafts.csswg.org/css-counter-styles/#descdef-counter-style-additive-symbols

pub fn set_speak_as(&mut self, value: SpeakAs) -> bool

https://drafts.csswg.org/css-counter-styles/#counter-style-speak-as

impl CounterStyleRuleData

fn check_system(&self, value: &System) -> bool

Check that the system is effectively not changed. Only params of system descriptor is changeable.

fn check_symbols(&self, value: &Symbols) -> bool

fn check_additive_symbols(&self, _value: &AdditiveSymbols) -> bool

impl CounterStyleRuleData

pub fn name(&self) -> &CustomIdent

Get the name of the counter style rule.

pub fn set_name(&mut self, name: CustomIdent)

Set the name of the counter style rule. Caller must ensure that the name is valid.

pub fn generation(&self) -> u32

Get the current generation of the counter style rule.

pub fn resolved_system(&self) -> &System

Get the system of this counter style rule, default to symbolic if not specified.