Module style::custom_properties

Support for custom properties for cascading variables.

Structs

BorrowedSpecifiedValue

ComputedValue

SpecifiedValue

Functions

cascade
Add one custom property declaration to a map, unless another with the same name was already there.

finish_cascade

parse

parse_declaration_value
https://drafts.csswg.org/css-syntax-3/#typedef-declaration-value

parse_declaration_value_block
Like parse_declaration_value, but accept ! and ; since they are only invalid at the top level

parse_name

parse_non_custom_with_var
Parse the value of a non-custom property that contains var() references.

parse_self_contained_declaration_value

parse_var_function

remove_cycles
https://drafts.csswg.org/css-variables/#cycles The initial value of a custom property is represented by this property not being in the map.

substitute
Replace var() functions for a non-custom property. Return Err(()) for invalid at computed time.

substitute_all
Replace var() functions for all custom properties.

substitute_block
Replace var() functions in an arbitrary bit of input.

substitute_one
Replace var() functions for one custom property. Also recursively record results for other custom properties referenced by var() functions. Return Err(()) for invalid at computed time. or Ok(last_token_type that was pushed to partial_computed_value) otherwise.

Type Definitions

ComputedValuesMap

Name