style

Module custom_properties

Support for custom properties for cascading variables.

Macros

lnf_int ğŸ”’

lnf_int_variable ğŸ”’

make_variable ğŸ”’

Structs

ComputedCustomProperties

A pair of separate CustomPropertiesMaps, split between custom properties that have the inherit flag set and those with the flag unset.

CssEnvironment

The environment from which to get env function values.

CustomPropertiesBuilder

A struct that takes care of encapsulating the cascade process for custom properties.

EnvironmentVariable ğŸ”’

NonCustomReferenceMap ğŸ”’

NonCustomReferences ğŸ”’

Set of flags to non-custom references this custom property makes.

References ğŸ”’

A struct holding information about the external references to that a custom property value may have.

Substitution ğŸ”’

SubstitutionFunctionReference ğŸ”’

VariableFallback ğŸ”’

VariableValue

A value for a custom property is just a set of tokens.

Enums

DeferFontRelativeCustomPropertyResolution

Whether to defer resolving custom properties referencing font relative units.

SingleNonCustomReference ğŸ”’

SubstitutionFunctionKind ğŸ”’

Statics

CHROME_ENVIRONMENT_VARIABLES ğŸ”’

ENVIRONMENT_VARIABLES ğŸ”’

Functions

compute_value ğŸ”’

compute_variable_value

Given a potentially registered variable value turn it into a computed custom property value.

do_substitute_chunk ğŸ”’

eval_gtk_csd_titlebar_radius ğŸ”’

find_non_custom_references ğŸ”’

get_content_preferred_color_scheme ğŸ”’

get_hairline ğŸ”’

get_safearea_inset_bottom ğŸ”’

get_safearea_inset_left ğŸ”’

get_safearea_inset_right ğŸ”’

get_safearea_inset_top ğŸ”’

get_scrollbar_inline_size ğŸ”’

handle_invalid_at_computed_value_time ğŸ”’

parse_declaration_value ğŸ”’

https://drafts.csswg.org/css-syntax-3/#typedef-declaration-value

parse_declaration_value_block ğŸ”’

Like parse_declaration_value, but accept ! and ; since they are only invalid at the top level.

parse_name

Parse a custom property name.

remove_and_insert_initial_value ğŸ”’

Removes the named registered custom property and inserts its uncomputed initial value.

substitute

Replace var() and env() functions, returning the resulting CSS string.

substitute_all ğŸ”’

Resolve all custom properties to either substituted, invalid, or unset (meaning we should use the inherited value).

substitute_internal ğŸ”’

Replace var() and env() functions. Return Err(..) for invalid at computed time.

substitute_one_reference ğŸ”’

substitute_references_if_needed_and_apply ğŸ”’

Replace var() and env() functions in a pre-existing variable value.

Type Aliases

ComputedValue

Both specified and computed values are VariableValues, the difference is whether var() functions are expanded.

EnvironmentEvaluator ğŸ”’

Name

A custom property name is just an Atom.

SpecifiedValue

Both specified and computed values are VariableValues, the difference is whether var() functions are expanded.