Module style::custom_properties

Support for custom properties for cascading variables.

Structs

CustomPropertiesBuilder
A struct that takes care of encapsulating the cascade process for custom properties.

OrderedMap
A map that preserves order for the keys, and that is easily indexable.

OrderedMapIterator
An iterator for OrderedMap.

VariableValue
A value for a custom property is just a set of tokens.

Functions

parse_name
Parse a custom property name.

parse_non_custom_with_var
Parse the value of a non-custom property that contains var() references.

substitute
Replace var() functions for a non-custom property. Return Err(()) for invalid at computed time.

Type Definitions

ComputedValue
Both specified and computed values are VariableValues, the difference is whether var() functions are expanded.

CustomPropertiesMap
A map from CSS variable names to CSS variable computed values, used for resolving.

Name
A custom property name is just an Atom.

SpecifiedValue
Both specified and computed values are VariableValues, the difference is whether var() functions are expanded.