Function style::custom_properties::substitute_block

fn substitute_block<F>(input: &mut Parser, position: &mut (SourcePosition, TokenSerializationType), partial_computed_value: &mut ComputedValue, substitute_one: &mut F) -> Result<TokenSerializationType, ()> where F: FnMut(&Name, &mut ComputedValue) -> Result<TokenSerializationType, ()>

Replace var() functions in an arbitrary bit of input.

The substitute_one callback is called for each var() function in input. If the variable has its initial value, the callback should return Err(()) and leave partial_computed_value unchanged. Otherwise, it should push the value of the variable (with its own var() functions replaced) to partial_computed_value and return Ok(last_token_type of what was pushed)

Return Err(()) if input is invalid at computed-value time. or Ok(last_token_type that was pushed to partial_computed_value) otherwise.