style

Module data

Per-node data used in style calculation.

Structs

EagerPseudoArray 🔒

EagerPseudoStyles

A lazily-allocated list of styles for eagerly-cascaded pseudo-elements.

ElementData

Style system data associated with an Element.

ElementDataFlags

Various flags stored on ElementData.

ElementStyles

The styles associated with a node, including the styles for any pseudo-elements.

Enums

RestyleKind

The kind of restyle that a single element should do.

ViewportUnitUsage

Information on how this element uses viewport units.

Constants

EMPTY_PSEUDO_ARRAY 🔒

Type Aliases

EagerPseudoArrayInner 🔒