Module style::data

Per-node data used in style calculation.

Structs

ComputedStyle
The structure that represents the result of style computation. This is effectively a tuple of rules and computed values, that is, the rule node, and the result of computing that rule node's rules, the ComputedValues.

ElementData
Style system data associated with an Element.

ElementStyles
The styles associated with a node, including the styles for any pseudo-elements.

PseudoStyles
A set of styles for a given element's pseudo-elements.

RestyleData
Transient data used by the restyle algorithm. This structure is instantiated either before or during restyle traversal, and is cleared at the end of node processing.

SnapshotOption
We really want to store an Option here, but we can't drop Gecko Snapshots off-main-thread. So we make a convenient little wrapper to provide the semantics of Option, while deferring the actual drop.

StoredRestyleHint
Restyle hint for storing on ElementData. We use a separate representation to provide more type safety while propagating restyle hints down the tree.

Enums

DescendantRestyleHint
Enum to describe the different requirements that a restyle hint may impose on its descendants.

RestyleKind
The kind of restyle that a single element should do.