Struct style::data::ElementData

pub struct ElementData {
    styles: ElementDataStyles,
    pub restyle_data: Option<RestyleData>,
}

Style system data associated with a node.

In Gecko, this hangs directly off a node, but is dropped when the frame takes ownership of the computed style data.

In Servo, this is embedded inside of layout data, which itself hangs directly off the node. Servo does not currently implement ownership transfer of the computed style data to the frame.

In both cases, it is wrapped inside an AtomicRefCell to ensure thread safety.

Fields
styles: ElementDataStyles restyle_data: Option<RestyleData>
Methods

impl ElementData

fn new() -> Self

fn has_current_styles(&self) -> bool

fn get_current_styles(&self) -> Option<&ElementStyles>

fn current_styles(&self) -> &ElementStyles

fn current_pseudos_mut(&mut self) -> &mut PseudoStyles

fn previous_styles(&self) -> Option<&ElementStyles>

fn previous_styles_mut(&mut self) -> Option<&mut ElementStyles>

fn gather_previous_styles<F>(&mut self, f: F) where F: FnOnce() -> Option<ElementStyles>

fn ensure_restyle_data(&mut self)

fn finish_styling(&mut self, styles: ElementStyles)