Struct style::data::ElementData

pub struct ElementData { /* fields omitted */ }

Style system data associated with an Element.

In Gecko, this hangs directly off the Element. Servo, this is embedded inside of layout data, which itself hangs directly off the Element. In both cases, it is wrapped inside an AtomicRefCell to ensure thread safety.

Methods

impl ElementData

fn new(existing: Option<ElementStyles>) -> Self

Trivially construct an ElementData.

fn has_styles(&self) -> bool

Returns true if this element has a computed styled.

fn has_current_styles(&self) -> bool

Returns true if this element's style is up-to-date and has no potential invalidation.

fn restyle_kind(&self) -> RestyleKind

Returns the kind of restyling that we're going to need to do on this element, based of the stored restyle hint.

fn get_styles(&self) -> Option<&ElementStyles>

Gets the element styles, if any.

fn styles(&self) -> &ElementStyles

Gets the element styles. Panic if the element has never been styled.

fn get_styles_mut(&mut self) -> Option<&mut ElementStyles>

Gets a mutable reference to the element styles, if any.

fn styles_mut(&mut self) -> &mut ElementStyles

Gets a mutable reference to the element styles. Panic if the element has never been styled.

fn styles_and_restyle_mut(&mut self)
-> (&mut ElementStyles, Option<&mut RestyleData>)

Borrows both styles and restyle mutably at the same time.

fn set_styles(&mut self, styles: ElementStyles)

Sets the computed element styles.

fn has_restyle(&self) -> bool

Returns true if the Element has a RestyleData.

fn clear_restyle(&mut self)

Drops any RestyleData.

fn ensure_restyle(&mut self) -> &mut RestyleData

Creates a RestyleData if one doesn't exist.

Asserts that the Element has been styled.

fn get_restyle(&self) -> Option<&RestyleData>

Gets a reference to the restyle data, if any.

fn restyle(&self) -> &RestyleData

Gets a reference to the restyle data. Panic if the element does not have restyle data.

fn get_restyle_mut(&mut self) -> Option<&mut RestyleData>

Gets a mutable reference to the restyle data, if any.

fn restyle_mut(&mut self) -> &mut RestyleData

Gets a mutable reference to the restyle data. Panic if the element does not have restyle data.