Trait style::dom::NodeInfo

pub trait NodeInfo {
    fn is_element(&self) -> bool;

    fn is_text_node(&self) -> bool;
}

Simple trait to provide basic information about the type of an element.

We avoid exposing the full type id, since computing it in the general case would be difficult for Gecko nodes.

Required methods

fn is_element(&self) -> bool

Whether this node is an element.

fn is_text_node(&self) -> bool

Whether this node is a text node.

Loading content...