Trait style::dom::TDocument

pub trait TDocument: Sized + Copy + Clone {
    type ConcreteNode: TNode<ConcreteDocument = Self>;
    fn as_node(&self) -> Self::ConcreteNode;

    fn is_html_document(&self) -> bool;

    fn quirks_mode(&self) -> QuirksMode;

    fn shared_lock(&self) -> &SharedRwLock;

    fn elements_with_id<'a>(
        &self, 
        _id: &Atom
    ) -> Result<&'a [<Self::ConcreteNode as TNode>::ConcreteElement], ()>
    where
        Self: 'a,
    { ... }
}

The TDocument trait, to represent a document node.

Associated Types

type ConcreteNode: TNode<ConcreteDocument = Self>

The concrete TNode type.

Loading content...
Required methods

fn as_node(&self) -> Self::ConcreteNode

Get this document as a TNode.

fn is_html_document(&self) -> bool

Returns whether this document is an HTML document.

fn quirks_mode(&self) -> QuirksMode

Returns the quirks mode of this document.

fn shared_lock(&self) -> &SharedRwLock

This document's shared lock.

Loading content...
Provided methods

fn elements_with_id<'a>(
&self,
_id: &Atom
) -> Result<&'a [<Self::ConcreteNode as TNode>::ConcreteElement], ()> where
Self: 'a,

Get a list of elements with a given ID in this document, sorted by tree position.

Can return an error to signal that this list is not available, or also return an empty slice.

Loading content...