Trait style::dom::TElement

pub trait TElement: PartialEq + Debug + Sized + Copy + Clone + ElementExt + PresentationalHintsSynthetizer {
    type ConcreteNode: TNode<ConcreteElement=Self, ConcreteDocument=Self::ConcreteDocument>;
    type ConcreteDocument: TDocument<ConcreteNode=Self::ConcreteNode, ConcreteElement=Self>;
    fn as_node(&self) -> Self::ConcreteNode;
    fn style_attribute(&self) -> Option<&Arc<RwLock<PropertyDeclarationBlock>>>;
    fn get_state(&self) -> ElementState;
    fn has_attr(&self, namespace: &Namespace, attr: &LocalName) -> bool;
    fn attr_equals(&self,
               namespace: &Namespace,
               attr: &LocalName,
               value: &Atom)
               -> bool;
    fn set_restyle_damage(self, damage: RestyleDamage);
    fn existing_style_for_restyle_damage<'a>(&'a self,
                                         current_computed_values: Option<&'a Arc<ComputedValues>>,
                                         pseudo: Option<&PseudoElement>)
                                         -> Option<&'a RestyleDamage::PreExistingComputedValues>;
    fn deprecated_dirty_bit_is_set(&self) -> bool;
    fn has_dirty_descendants(&self) -> bool;
    unsafe fn set_dirty_descendants(&self);
    fn store_children_to_process(&self, n: isize);
    fn did_process_child(&self) -> isize;
    fn borrow_data(&self) -> Option<AtomicRef<ElementData>>;
    fn get_data(&self) -> Option<&AtomicRefCell<ElementData>>;

    fn is_display_none(&self) -> bool { ... }
    fn frame_has_style(&self) -> bool { ... }
    fn get_styles_from_frame(&self) -> Option<ElementStyles> { ... }
    fn styling_mode(&self) -> StylingMode { ... }
    fn note_restyle_hint<C: DomTraversalContext<Self::ConcreteNode>>(&self,
                                                                 hint: RestyleHint) { ... }
}

Associated Types

type ConcreteNode: TNode<ConcreteElement=Self, ConcreteDocument=Self::ConcreteDocument>

type ConcreteDocument: TDocument<ConcreteNode=Self::ConcreteNode, ConcreteElement=Self>

Required Methods

fn as_node(&self) -> Self::ConcreteNode

fn style_attribute(&self) -> Option<&Arc<RwLock<PropertyDeclarationBlock>>>

fn get_state(&self) -> ElementState

fn has_attr(&self, namespace: &Namespace, attr: &LocalName) -> bool

fn attr_equals(&self,
namespace: &Namespace,
attr: &LocalName,
value: &Atom)
-> bool

fn set_restyle_damage(self, damage: RestyleDamage)

Set the restyle damage field.

fn existing_style_for_restyle_damage<'a>(&'a self,
current_computed_values: Option<&'a Arc<ComputedValues>>,
pseudo: Option<&PseudoElement>)
-> Option<&'a RestyleDamage::PreExistingComputedValues>

XXX: It's a bit unfortunate we need to pass the current computed values as an argument here, but otherwise Servo would crash due to double borrows to return it.

fn deprecated_dirty_bit_is_set(&self) -> bool

The concept of a dirty bit doesn't exist in our new restyle algorithm. Instead, we associate restyle and change hints with nodes. However, we continue to allow the dirty bit to trigger unconditional restyles while we transition both Servo and Stylo to the new architecture.

fn has_dirty_descendants(&self) -> bool

unsafe fn set_dirty_descendants(&self)

fn store_children_to_process(&self, n: isize)

Atomically stores the number of children of this node that we will need to process during bottom-up traversal.

fn did_process_child(&self) -> isize

Atomically notes that a child has been processed during bottom-up traversal. Returns the number of children left to process.

fn borrow_data(&self) -> Option<AtomicRef<ElementData>>

Immutable borrows the ElementData.

fn get_data(&self) -> Option<&AtomicRefCell<ElementData>>

Gets a reference to the ElementData container.

Provided Methods

fn is_display_none(&self) -> bool

Returns true if this element's current style is display:none. Only valid to call after styling.

fn frame_has_style(&self) -> bool

Returns true if this node has a styled layout frame that owns the style.

fn get_styles_from_frame(&self) -> Option<ElementStyles>

Returns the styles from the layout frame that owns them, if any.

FIXME(bholley): Once we start dropping ElementData from nodes when creating frames, we'll want to teach this method to actually get style data from the frame.

fn styling_mode(&self) -> StylingMode

Returns the styling mode for this node. This is only valid to call before and during restyling, before finish_styling is invoked.

See the comments around StylingMode.

fn note_restyle_hint<C: DomTraversalContext<Self::ConcreteNode>>(&self,
hint: RestyleHint)

Properly marks nodes as dirty in response to restyle hints.