Trait style::dom::TElement

pub trait TElement: PartialEq + Debug + Sized + Copy + Clone + ElementExt + PresentationalHintsSynthetizer {
    type ConcreteNode: TNode<ConcreteElement=Self>;
    fn as_node(&self) -> Self::ConcreteNode;
    fn style_attribute(&self) -> Option<&Arc<RwLock<PropertyDeclarationBlock>>>;
    fn get_state(&self) -> ElementState;
    fn has_attr(&self, namespace: &Namespace, attr: &LocalName) -> bool;
    fn attr_equals(&self,
                   namespace: &Namespace,
                   attr: &LocalName,
                   value: &Atom)
                   -> bool;
    fn existing_style_for_restyle_damage<'a>(&'a self,
                                             current_computed_values: &'a Arc<ComputedValues>,
                                             pseudo: Option<&PseudoElement>)
                                             -> Option<&'a PreExistingComputedValues>;
    fn has_dirty_descendants(&self) -> bool;
    unsafe fn set_dirty_descendants(&self);
    unsafe fn unset_dirty_descendants(&self);
    fn store_children_to_process(&self, n: isize);
    fn did_process_child(&self) -> isize;
    fn get_data(&self) -> Option<&AtomicRefCell<ElementData>>;
    fn skip_root_and_item_based_display_fixup(&self) -> bool;
    unsafe fn set_selector_flags(&self, flags: ElementSelectorFlags);
    fn has_selector_flags(&self, flags: ElementSelectorFlags) -> bool;

    fn layout_parent_element(self, reflow_root: OpaqueNode) -> Option<Self> { ... }
    fn get_animation_rules(&self,
                           _pseudo: Option<&PseudoElement>)
                           -> AnimationRules { ... }
    fn borrow_data(&self) -> Option<AtomicRef<ElementData>> { ... }
    fn mutate_data(&self) -> Option<AtomicRefMut<ElementData>> { ... }
}

The element trait, the main abstraction the style crate acts over.

Associated Types

type ConcreteNode: TNode<ConcreteElement=Self>

The concrete node type.

Required Methods

fn as_node(&self) -> Self::ConcreteNode

Get this element as a node.

fn style_attribute(&self) -> Option<&Arc<RwLock<PropertyDeclarationBlock>>>

Get this element's style attribute.

fn get_state(&self) -> ElementState

Get this element's state, for non-tree-structural pseudos.

fn has_attr(&self, namespace: &Namespace, attr: &LocalName) -> bool

Whether this element has an attribute with a given namespace.

fn attr_equals(&self,
namespace: &Namespace,
attr: &LocalName,
value: &Atom)
-> bool

Whether an attribute value equals value.

fn existing_style_for_restyle_damage<'a>(&'a self,
current_computed_values: &'a Arc<ComputedValues>,
pseudo: Option<&PseudoElement>)
-> Option<&'a PreExistingComputedValues>

Get the pre-existing style to calculate restyle damage (change hints).

This needs to be generic since it varies between Servo and Gecko.

XXX(emilio): It's a bit unfortunate we need to pass the current computed values as an argument here, but otherwise Servo would crash due to double borrows to return it.

fn has_dirty_descendants(&self) -> bool

Returns true if this element may have a descendant needing style processing.

Note that we cannot guarantee the existence of such an element, because it may have been removed from the DOM between marking it for restyle and the actual restyle traversal.

unsafe fn set_dirty_descendants(&self)

Flag that this element has a descendant for style processing.

Only safe to call with exclusive access to the element.

unsafe fn unset_dirty_descendants(&self)

Flag that this element has no descendant for style processing.

Only safe to call with exclusive access to the element.

fn store_children_to_process(&self, n: isize)

Atomically stores the number of children of this node that we will need to process during bottom-up traversal.

fn did_process_child(&self) -> isize

Atomically notes that a child has been processed during bottom-up traversal. Returns the number of children left to process.

fn get_data(&self) -> Option<&AtomicRefCell<ElementData>>

Gets a reference to the ElementData container.

fn skip_root_and_item_based_display_fixup(&self) -> bool

Whether we should skip any root- or item-based display property blockification on this element. (This function exists so that Gecko native anonymous content can opt out of this style fixup.)

unsafe fn set_selector_flags(&self, flags: ElementSelectorFlags)

Sets selector flags, which indicate what kinds of selectors may have matched on this element and therefore what kind of work may need to be performed when DOM state changes.

This is unsafe, like all the flag-setting methods, because it's only safe to call with exclusive access to the element. When setting flags on the parent during parallel traversal, we use SequentialTask to queue up the set to run after the threads join.

fn has_selector_flags(&self, flags: ElementSelectorFlags) -> bool

Returns true if the element has all the specified selector flags.

Provided Methods

fn layout_parent_element(self, reflow_root: OpaqueNode) -> Option<Self>

While doing a reflow, the element at the root has no parent, as far as we're concerned. This method returns None at the reflow root.

fn get_animation_rules(&self, _pseudo: Option<&PseudoElement>) -> AnimationRules

Get this element's animation rules.

fn borrow_data(&self) -> Option<AtomicRef<ElementData>>

Immutably borrows the ElementData.

fn mutate_data(&self) -> Option<AtomicRefMut<ElementData>>

Mutably borrows the ElementData.