Trait style::dom::TNode

pub trait TNode: Sized + Copy + Clone + NodeInfo {
    type ConcreteElement: TElement<ConcreteNode=Self, ConcreteDocument=Self::ConcreteDocument>;
    type ConcreteDocument: TDocument<ConcreteNode=Self, ConcreteElement=Self::ConcreteElement>;
    type ConcreteChildrenIterator: Iterator<Item=Self>;
    fn to_unsafe(&self) -> UnsafeNode;
    unsafe fn from_unsafe(n: &UnsafeNode) -> Self;
    fn dump(self);
    fn dump_style(self);
    fn children(self) -> LayoutIterator<Self::ConcreteChildrenIterator>;
    fn opaque(&self) -> OpaqueNode;
    fn layout_parent_node(self, reflow_root: OpaqueNode) -> Option<Self>;
    fn debug_id(self) -> usize;
    fn as_element(&self) -> Option<Self::ConcreteElement>;
    fn as_document(&self) -> Option<Self::ConcreteDocument>;
    fn deprecated_dirty_bit_is_set(&self) -> bool;
    fn has_dirty_descendants(&self) -> bool;
    unsafe fn set_dirty_descendants(&self);
    fn needs_dirty_on_viewport_size_changed(&self) -> bool;
    unsafe fn set_dirty_on_viewport_size_changed(&self);
    fn can_be_fragmented(&self) -> bool;
    unsafe fn set_can_be_fragmented(&self, value: bool);
    fn store_children_to_process(&self, n: isize);
    fn did_process_child(&self) -> isize;
    fn begin_styling(&self) -> AtomicRefMut<NodeData>;
    fn style_text_node(&self, style: Arc<ComputedValues>);
    fn borrow_data(&self) -> Option<AtomicRef<NodeData>>;
    fn parent_node(&self) -> Option<Self>;
    fn first_child(&self) -> Option<Self>;
    fn last_child(&self) -> Option<Self>;
    fn prev_sibling(&self) -> Option<Self>;
    fn next_sibling(&self) -> Option<Self>;

    fn frame_has_style(&self) -> bool { ... }
    fn get_styles_from_frame(&self) -> Option<NodeStyles> { ... }
    fn styling_mode(&self) -> StylingMode { ... }
}

Associated Types

type ConcreteElement: TElement<ConcreteNode=Self, ConcreteDocument=Self::ConcreteDocument>

type ConcreteDocument: TDocument<ConcreteNode=Self, ConcreteElement=Self::ConcreteElement>

type ConcreteChildrenIterator: Iterator<Item=Self>

Required Methods

fn to_unsafe(&self) -> UnsafeNode

unsafe fn from_unsafe(n: &UnsafeNode) -> Self

fn dump(self)

fn dump_style(self)

fn children(self) -> LayoutIterator<Self::ConcreteChildrenIterator>

Returns an iterator over this node's children.

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn layout_parent_node(self, reflow_root: OpaqueNode) -> Option<Self>

While doing a reflow, the node at the root has no parent, as far as we're concerned. This method returns None at the reflow root.

fn debug_id(self) -> usize

fn as_element(&self) -> Option<Self::ConcreteElement>

fn as_document(&self) -> Option<Self::ConcreteDocument>

fn deprecated_dirty_bit_is_set(&self) -> bool

The concept of a dirty bit doesn't exist in our new restyle algorithm. Instead, we associate restyle and change hints with nodes. However, we continue to allow the dirty bit to trigger unconditional restyles while we transition both Servo and Stylo to the new architecture.

fn has_dirty_descendants(&self) -> bool

unsafe fn set_dirty_descendants(&self)

fn needs_dirty_on_viewport_size_changed(&self) -> bool

unsafe fn set_dirty_on_viewport_size_changed(&self)

fn can_be_fragmented(&self) -> bool

unsafe fn set_can_be_fragmented(&self, value: bool)

fn store_children_to_process(&self, n: isize)

Atomically stores the number of children of this node that we will need to process during bottom-up traversal.

fn did_process_child(&self) -> isize

Atomically notes that a child has been processed during bottom-up traversal. Returns the number of children left to process.

fn begin_styling(&self) -> AtomicRefMut<NodeData>

Sets up the appropriate data structures to style a node, returing a mutable handle to the node data upon which further style calculations can be performed.

fn style_text_node(&self, style: Arc<ComputedValues>)

Set the style directly for a text node. This skips various unnecessary steps from begin_styling like computing the previous style.

fn borrow_data(&self) -> Option<AtomicRef<NodeData>>

Immutable borrows the NodeData.

fn parent_node(&self) -> Option<Self>

fn first_child(&self) -> Option<Self>

fn last_child(&self) -> Option<Self>

fn prev_sibling(&self) -> Option<Self>

fn next_sibling(&self) -> Option<Self>

Provided Methods

fn frame_has_style(&self) -> bool

Returns true if this node has a styled layout frame that owns the style.

fn get_styles_from_frame(&self) -> Option<NodeStyles>

Returns the styles from the layout frame that owns them, if any.

FIXME(bholley): Once we start dropping NodeData from nodes when creating frames, we'll want to teach this method to actually get style data from the frame.

fn styling_mode(&self) -> StylingMode

Returns the styling mode for this node. This is only valid to call before and during restyling, before finish_styling is invoked.

See the comments around StylingMode.