Trait style::dom::TNode

pub trait TNode: Sized + Copy + Clone {
    type ConcreteElement: TElement<ConcreteNode=Self, ConcreteDocument=Self::ConcreteDocument>;
    type ConcreteDocument: TDocument<ConcreteNode=Self, ConcreteElement=Self::ConcreteElement>;
    type ConcreteRestyleDamage: TRestyleDamage<ConcreteComputedValues=Self::ConcreteComputedValues>;
    type ConcreteComputedValues: ComputedValues;
    fn to_unsafe(&self) -> UnsafeNode;
    unsafe fn from_unsafe(n: &UnsafeNode) -> Self;
    fn is_text_node(&self) -> bool;
    fn is_element(&self) -> bool;
    fn dump(self);
    fn opaque(&self) -> OpaqueNode;
    fn initialize_data(self);
    fn layout_parent_node(self, reflow_root: OpaqueNode) -> Option<Self>;
    fn debug_id(self) -> usize;
    fn as_element(&self) -> Option<Self::ConcreteElement>;
    fn as_document(&self) -> Option<Self::ConcreteDocument>;
    fn children_count(&self) -> u32;
    fn has_changed(&self) -> bool;
    unsafe fn set_changed(&self, value: bool);
    fn is_dirty(&self) -> bool;
    unsafe fn set_dirty(&self, value: bool);
    fn has_dirty_descendants(&self) -> bool;
    unsafe fn set_dirty_descendants(&self, value: bool);
    fn can_be_fragmented(&self) -> bool;
    unsafe fn set_can_be_fragmented(&self, value: bool);
    unsafe fn borrow_data_unchecked(&self) -> Option<*const PrivateStyleData<Self::ConcreteElement::Impl, Self::ConcreteComputedValues>>;
    fn borrow_data(&self) -> Option<Ref<PrivateStyleData<Self::ConcreteElement::Impl, Self::ConcreteComputedValues>>>;
    fn mutate_data(&self) -> Option<RefMut<PrivateStyleData<Self::ConcreteElement::Impl, Self::ConcreteComputedValues>>>;
    fn restyle_damage(self) -> Self::ConcreteRestyleDamage;
    fn set_restyle_damage(self, damage: Self::ConcreteRestyleDamage);
    fn parent_node(&self) -> Option<Self>;
    fn first_child(&self) -> Option<Self>;
    fn last_child(&self) -> Option<Self>;
    fn prev_sibling(&self) -> Option<Self>;
    fn next_sibling(&self) -> Option<Self>;

    fn traverse_preorder(self) -> TreeIterator<Self> { ... }
    fn children(self) -> ChildrenIterator<Self> { ... }
    fn rev_children(self) -> ReverseChildrenIterator<Self> { ... }
    fn dirty_self(&self) { ... }
    fn dirty_descendants(&self) { ... }
    fn style(&self, _context: &SharedStyleContext<Self::ConcreteElement::Impl>) -> Ref<Arc<Self::ConcreteComputedValues>> where Self::ConcreteElement::Impl: SelectorImplExt<ComputedValues=Self::ConcreteComputedValues> { ... }
    fn unstyle(self) { ... }
}

Associated Types

type ConcreteElement: TElement<ConcreteNode=Self, ConcreteDocument=Self::ConcreteDocument>

type ConcreteDocument: TDocument<ConcreteNode=Self, ConcreteElement=Self::ConcreteElement>

type ConcreteRestyleDamage: TRestyleDamage<ConcreteComputedValues=Self::ConcreteComputedValues>

type ConcreteComputedValues: ComputedValues

Required Methods

fn to_unsafe(&self) -> UnsafeNode

unsafe fn from_unsafe(n: &UnsafeNode) -> Self

fn is_text_node(&self) -> bool

Returns whether this is a text node. It turns out that this is all the style system cares about, and thus obviates the need to compute the full type id, which would be expensive in Gecko.

fn is_element(&self) -> bool

fn dump(self)

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn initialize_data(self)

Initializes style and layout data for the node. No-op if the data is already initialized.

FIXME(pcwalton): Do this as part of fragment building instead of in a traversal.

fn layout_parent_node(self, reflow_root: OpaqueNode) -> Option<Self>

While doing a reflow, the node at the root has no parent, as far as we're concerned. This method returns None at the reflow root.

fn debug_id(self) -> usize

fn as_element(&self) -> Option<Self::ConcreteElement>

fn as_document(&self) -> Option<Self::ConcreteDocument>

fn children_count(&self) -> u32

fn has_changed(&self) -> bool

unsafe fn set_changed(&self, value: bool)

fn is_dirty(&self) -> bool

unsafe fn set_dirty(&self, value: bool)

fn has_dirty_descendants(&self) -> bool

unsafe fn set_dirty_descendants(&self, value: bool)

fn can_be_fragmented(&self) -> bool

unsafe fn set_can_be_fragmented(&self, value: bool)

unsafe fn borrow_data_unchecked(&self) -> Option<*const PrivateStyleData<Self::ConcreteElement::Impl, Self::ConcreteComputedValues>>

Borrows the PrivateStyleData without checks.

fn borrow_data(&self) -> Option<Ref<PrivateStyleData<Self::ConcreteElement::Impl, Self::ConcreteComputedValues>>>

Borrows the PrivateStyleData immutably. Fails on a conflicting borrow.

fn mutate_data(&self) -> Option<RefMut<PrivateStyleData<Self::ConcreteElement::Impl, Self::ConcreteComputedValues>>>

Borrows the PrivateStyleData mutably. Fails on a conflicting borrow.

fn restyle_damage(self) -> Self::ConcreteRestyleDamage

Get the description of how to account for recent style changes.

fn set_restyle_damage(self, damage: Self::ConcreteRestyleDamage)

Set the restyle damage field.

fn parent_node(&self) -> Option<Self>

fn first_child(&self) -> Option<Self>

fn last_child(&self) -> Option<Self>

fn prev_sibling(&self) -> Option<Self>

fn next_sibling(&self) -> Option<Self>

Provided Methods

fn traverse_preorder(self) -> TreeIterator<Self>

fn children(self) -> ChildrenIterator<Self>

Returns an iterator over this node's children.

fn rev_children(self) -> ReverseChildrenIterator<Self>

fn dirty_self(&self)

fn dirty_descendants(&self)

fn style(&self, _context: &SharedStyleContext<Self::ConcreteElement::Impl>) -> Ref<Arc<Self::ConcreteComputedValues>> where Self::ConcreteElement::Impl: SelectorImplExt<ComputedValues=Self::ConcreteComputedValues>

Returns the style results for the given node. If CSS selector matching has not yet been performed, fails.

fn unstyle(self)

Removes the style from this node.