style::dom

Trait TNode

pub trait TNode:
    Sized
    + Copy
    + Clone
    + Debug
    + NodeInfo
    + PartialEq {
    type ConcreteElement: TElement<ConcreteNode = Self>;
    type ConcreteDocument: TDocument<ConcreteNode = Self>;
    type ConcreteShadowRoot: TShadowRoot<ConcreteNode = Self>;

    // Required methods
    fn parent_node(&self) -> Option<Self>;
    fn first_child(&self) -> Option<Self>;
    fn last_child(&self) -> Option<Self>;
    fn prev_sibling(&self) -> Option<Self>;
    fn next_sibling(&self) -> Option<Self>;
    fn owner_doc(&self) -> Self::ConcreteDocument;
    fn is_in_document(&self) -> bool;
    fn traversal_parent(&self) -> Option<Self::ConcreteElement>;
    fn opaque(&self) -> OpaqueNode;
    fn debug_id(self) -> usize;
    fn as_element(&self) -> Option<Self::ConcreteElement>;
    fn as_document(&self) -> Option<Self::ConcreteDocument>;
    fn as_shadow_root(&self) -> Option<Self::ConcreteShadowRoot>;

    // Provided methods
    fn dom_children(&self) -> DomChildren<Self>  { ... }
    fn dom_descendants(&self) -> DomDescendants<Self>  { ... }
    fn next_in_preorder(&self, scoped_to: Self) -> Option<Self> { ... }
    fn depth(&self) -> usize { ... }
    fn parent_element(&self) -> Option<Self::ConcreteElement> { ... }
    fn parent_element_or_host(&self) -> Option<Self::ConcreteElement> { ... }

}

The TNode trait. This is the main generic trait over which the style system can be implemented.

Required Associated Types

type ConcreteElement: TElement<ConcreteNode = Self>

The concrete TElement type.

type ConcreteDocument: TDocument<ConcreteNode = Self>

The concrete TDocument type.

type ConcreteShadowRoot: TShadowRoot<ConcreteNode = Self>

The concrete TShadowRoot type.

Required Methods

fn parent_node(&self) -> Option<Self>

Get this node’s parent node.

fn first_child(&self) -> Option<Self>

Get this node’s first child.

fn last_child(&self) -> Option<Self>

Get this node’s last child.

fn prev_sibling(&self) -> Option<Self>

Get this node’s previous sibling.

fn next_sibling(&self) -> Option<Self>

Get this node’s next sibling.

fn owner_doc(&self) -> Self::ConcreteDocument

Get the owner document of this node.

fn is_in_document(&self) -> bool

Returns whether the node is attached to a document.

fn traversal_parent(&self) -> Option<Self::ConcreteElement>

Get this node’s parent element from the perspective of a restyle traversal.

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn debug_id(self) -> usize

A debug id, only useful, mm… for debugging.

fn as_element(&self) -> Option<Self::ConcreteElement>

Get this node as an element, if it’s one.

fn as_document(&self) -> Option<Self::ConcreteDocument>

Get this node as a document, if it’s one.

fn as_shadow_root(&self) -> Option<Self::ConcreteShadowRoot>

Get this node as a ShadowRoot, if it’s one.

Provided Methods

fn dom_children(&self) -> DomChildren<Self>

Iterate over the DOM children of a node.

fn dom_descendants(&self) -> DomDescendants<Self>

Iterate over the DOM children of a node, in preorder.

fn next_in_preorder(&self, scoped_to: Self) -> Option<Self>

Returns the next node after this one, in a pre-order tree-traversal of the subtree rooted at scoped_to.

fn depth(&self) -> usize

Returns the depth of this node in the DOM.

fn parent_element(&self) -> Option<Self::ConcreteElement>

Get this node’s parent element if present.

fn parent_element_or_host(&self) -> Option<Self::ConcreteElement>

Get this node’s parent element, or shadow host if it’s a shadow root.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.