Trait style::dom::TNode

pub trait TNode: Sized + Copy + Clone + Debug + NodeInfo + PartialEq {
    type ConcreteElement: TElement<ConcreteNode = Self>;
    type ConcreteDocument: TDocument<ConcreteNode = Self>;
    type ConcreteShadowRoot: TShadowRoot<ConcreteNode = Self>;
    fn parent_node(&self) -> Option<Self>;

    fn first_child(&self) -> Option<Self>;

    fn last_child(&self) -> Option<Self>;

    fn prev_sibling(&self) -> Option<Self>;

    fn next_sibling(&self) -> Option<Self>;

    fn owner_doc(&self) -> Self::ConcreteDocument;

    fn is_in_document(&self) -> bool;

    fn traversal_parent(&self) -> Option<Self::ConcreteElement>;

    fn opaque(&self) -> OpaqueNode;

    fn debug_id(self) -> usize;

    fn as_element(&self) -> Option<Self::ConcreteElement>;

    fn as_document(&self) -> Option<Self::ConcreteDocument>;

    fn as_shadow_root(&self) -> Option<Self::ConcreteShadowRoot>;

    fn dom_children(&self) -> DomChildren<Self>ⓘ

Important traits for DomChildren<N>
impl<N> Iterator for DomChildren<N> where
    N: TNode,     type Item = N;
 { ... }

    fn dom_descendants(&self) -> DomDescendants<Self>ⓘ

Important traits for DomDescendants<N>
impl<N> Iterator for DomDescendants<N> where
    N: TNode,     type Item = N;
 { ... }

    fn next_in_preorder(&self, scoped_to: Option<Self>) -> Option<Self> { ... }

    fn parent_element(&self) -> Option<Self::ConcreteElement> { ... }
}

The TNode trait. This is the main generic trait over which the style system can be implemented.

Associated Types

type ConcreteElement: TElement<ConcreteNode = Self>

The concrete TElement type.

type ConcreteDocument: TDocument<ConcreteNode = Self>

The concrete TDocument type.

type ConcreteShadowRoot: TShadowRoot<ConcreteNode = Self>

The concrete TShadowRoot type.

Loading content...
Required methods

fn parent_node(&self) -> Option<Self>

Get this node's parent node.

fn first_child(&self) -> Option<Self>

Get this node's first child.

fn last_child(&self) -> Option<Self>

Get this node's first child.

fn prev_sibling(&self) -> Option<Self>

Get this node's previous sibling.

fn next_sibling(&self) -> Option<Self>

Get this node's next sibling.

fn owner_doc(&self) -> Self::ConcreteDocument

Get the owner document of this node.

fn is_in_document(&self) -> bool

Returns whether the node is attached to a document.

fn traversal_parent(&self) -> Option<Self::ConcreteElement>

Get this node's parent element from the perspective of a restyle traversal.

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn debug_id(self) -> usize

A debug id, only useful, mm... for debugging.

fn as_element(&self) -> Option<Self::ConcreteElement>

Get this node as an element, if it's one.

fn as_document(&self) -> Option<Self::ConcreteDocument>

Get this node as a document, if it's one.

fn as_shadow_root(&self) -> Option<Self::ConcreteShadowRoot>

Get this node as a ShadowRoot, if it's one.

Loading content...
Provided methods

fn dom_children(&self) -> DomChildren<Self>ⓘ

Important traits for DomChildren<N>
impl<N> Iterator for DomChildren<N> where
N: TNode, type Item = N;

Iterate over the DOM children of a node.

fn dom_descendants(&self) -> DomDescendants<Self>ⓘ

Important traits for DomDescendants<N>
impl<N> Iterator for DomDescendants<N> where
N: TNode, type Item = N;

Iterate over the DOM children of a node, in preorder.

fn next_in_preorder(&self, scoped_to: Option<Self>) -> Option<Self>

Returns the next children in pre-order, optionally scoped to a subtree root.

fn parent_element(&self) -> Option<Self::ConcreteElement>

Get this node's parent element if present.

Loading content...