Trait style::dom::TNode

pub trait TNode: Sized + Copy + Clone + Debug + NodeInfo {
    type ConcreteElement: TElement<ConcreteNode = Self>;
    type ConcreteChildrenIterator: Iterator<Item = Self>;
    fn to_unsafe(&self) -> UnsafeNode;

    unsafe fn from_unsafe(n: &UnsafeNode) -> Self;

    fn parent_node(&self) -> Option<Self>;

    fn children(&self) -> LayoutIterator<Self::ConcreteChildrenIterator>;

    fn traversal_parent(&self) -> Option<Self::ConcreteElement>;

    fn traversal_children(
        &self
    ) -> LayoutIterator<Self::ConcreteChildrenIterator>;

    fn opaque(&self) -> OpaqueNode;

    fn debug_id(self) -> usize;

    fn as_element(&self) -> Option<Self::ConcreteElement>;

    fn can_be_fragmented(&self) -> bool;

    unsafe fn set_can_be_fragmented(&self, value: bool);

    fn is_in_doc(&self) -> bool;

    fn parent_element(&self) -> Option<Self::ConcreteElement> { ... }
}

The TNode trait. This is the main generic trait over which the style system can be implemented.

Associated Types

type ConcreteElement: TElement<ConcreteNode = Self>

The concrete TElement type.

type ConcreteChildrenIterator: Iterator<Item = Self>

A concrete children iterator type in order to iterate over the Nodes.

TODO(emilio): We should eventually replace this with the impl Trait syntax.

Required Methods

fn to_unsafe(&self) -> UnsafeNode

Convert this node in an UnsafeNode.

unsafe fn from_unsafe(n: &UnsafeNode) -> Self

Get a node back from an UnsafeNode.

fn parent_node(&self) -> Option<Self>

Get this node's parent node.

fn children(&self) -> LayoutIterator<Self::ConcreteChildrenIterator>

Returns an iterator over this node's children.

fn traversal_parent(&self) -> Option<Self::ConcreteElement>

Get this node's parent element from the perspective of a restyle traversal.

fn traversal_children(&self) -> LayoutIterator<Self::ConcreteChildrenIterator>

Get this node's children from the perspective of a restyle traversal.

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn debug_id(self) -> usize

A debug id, only useful, mm... for debugging.

fn as_element(&self) -> Option<Self::ConcreteElement>

Get this node as an element, if it's one.

fn can_be_fragmented(&self) -> bool

Whether this node can be fragmented. This is used for multicol, and only for Servo.

unsafe fn set_can_be_fragmented(&self, value: bool)

Set whether this node can be fragmented.

fn is_in_doc(&self) -> bool

Whether this node is in the document right now needed to clear the restyle data appropriately on some forced restyles.

Provided Methods

fn parent_element(&self) -> Option<Self::ConcreteElement>

Get this node's parent element if present.