Trait style::dom::TShadowRoot

pub trait TShadowRoot: Sized + Copy + Clone + Debug + PartialEq {
    type ConcreteNode: TNode<ConcreteShadowRoot = Self>;
    fn as_node(&self) -> Self::ConcreteNode;

    fn host(&self) -> <Self::ConcreteNode as TNode>::ConcreteElement;

    fn style_data<'a>(&self) -> Option<&'a CascadeData>
    where
        Self: 'a;

    fn parts<'a>(&self) -> &[<Self::ConcreteNode as TNode>::ConcreteElement]ⓘ

Important traits for &'_ [u8]
impl<'_> Read for &'_ [u8]impl<'_> Write for &'_ mut [u8]

    where
        Self: 'a,
    { ... }

    fn elements_with_id<'a>(
        &self, 
        _id: &Atom
    ) -> Result<&'a [<Self::ConcreteNode as TNode>::ConcreteElement], ()>
    where
        Self: 'a,
    { ... }
}

The ShadowRoot trait.

Associated Types

type ConcreteNode: TNode<ConcreteShadowRoot = Self>

The concrete node type.

Loading content...
Required methods

fn as_node(&self) -> Self::ConcreteNode

Get this ShadowRoot as a node.

fn host(&self) -> <Self::ConcreteNode as TNode>::ConcreteElement

Get the shadow host that hosts this ShadowRoot.

fn style_data<'a>(&self) -> Option<&'a CascadeData> where
Self: 'a,

Get the style data for this ShadowRoot.

Loading content...
Provided methods

fn parts<'a>(&self) -> &[<Self::ConcreteNode as TNode>::ConcreteElement]ⓘ

Important traits for &'_ [u8]
impl<'_> Read for &'_ [u8]impl<'_> Write for &'_ mut [u8]
where
Self: 'a,

Get the list of shadow parts for this shadow root.

fn elements_with_id<'a>(
&self,
_id: &Atom
) -> Result<&'a [<Self::ConcreteNode as TNode>::ConcreteElement], ()> where
Self: 'a,

Get a list of elements with a given ID in this shadow root, sorted by tree position.

Can return an error to signal that this list is not available, or also return an empty slice.

Loading content...