Module style::dom_apis

Generic implementations of some DOM APIs so they can be shared between Servo and Gecko.

Structs

QueryAll

A query for all the elements in a subtree.

QueryFirst

A query for the first in-tree match of all the elements in a subtree.

QuerySelectorProcessor

Enums

MayUseInvalidation

Whether the invalidation machinery should be used for this query.

SimpleFilter

Traits

SelectorQuery

A selector query abstraction, in order to be generic over QuerySelector and QuerySelectorAll.

Functions

collect_all_elements

collect_elements_with_id

Collects elements with a given id under root, that pass filter.

connected_element_is_descendant_of

Returns whether a given element connected to root is descendant of root.

element_closest

https://dom.spec.whatwg.org/#dom-element-closest

element_matches

https://dom.spec.whatwg.org/#dom-element-matches

fast_connected_elements_with_id

Fast path for iterating over every element with a given id in the document or shadow root that root is connected to.

local_name_matches

query_selector

https://dom.spec.whatwg.org/#dom-parentnode-queryselector

query_selector_fast

Fast paths for a given selector query.

query_selector_single_query

Fast paths for querySelector with a single simple selector.

query_selector_slow

Type Definitions

QuerySelectorAllResult

The result of a querySelectorAll call.