Module style::dom_apis

Generic implementations of some DOM APIs so they can be shared between Servo and Gecko.

Structs

QueryAll
A query for all the elements in a subtree.

QueryFirst
A query for the first in-tree match of all the elements in a subtree.

Enums

MayUseInvalidation
Whether the invalidation machinery should be used for this query.

Traits

SelectorQuery
A selector query abstraction, in order to be generic over QuerySelector and QuerySelectorAll.

Functions

element_closest
https://dom.spec.whatwg.org/#dom-element-closest

element_matches
https://dom.spec.whatwg.org/#dom-element-matches

query_selector
https://dom.spec.whatwg.org/#dom-parentnode-queryselector

Type Definitions

QuerySelectorAllResult
The result of a querySelectorAll call.