Module style::font_face

The @font-face at-rule.

Structs

ComputedFontStretchRange
The computed representation of the above, so that Gecko can read them easily.

ComputedFontWeightRange
The computed representation of the above so Gecko can read them easily.

EffectiveSources
A list of effective sources that we send over through IPC to the font cache.

FontFace
A @font-face rule that is known to have font-family and src declarations.

FontFaceRuleData
Data inside a @font-face rule.

FontFaceRuleParser

FontStretchRange
The font-stretch descriptor:

FontWeightRange
The font-weight descriptor:

UrlSource
A UrlSource represents a font-face source that has been specified with a url() function.

Enums

ComputedFontStyleDescriptor
The computed representation of the above, with angles in degrees, so that Gecko can read them easily.

FontDisplay
A font-display value for a @font-face rule. The font-display descriptor determines how a font face is displayed based on whether and when it is downloaded and ready to use.

FontStyle
The font-style descriptor:

Source
A source for a font-face rule.

Functions

parse_font_face_block
Parse the block inside a @font-face rule.

sort_range