Module style::font_face

The @font-face at-rule.

Macros

font_face_descriptors 🔒

font_face_descriptors_common 🔒

impl_range 🔒

is_descriptor_enabled 🔒

Structs

ComputedFontStretchRange

The computed representation of the above, so that Gecko can read them easily.

ComputedFontWeightRange

The computed representation of the above so Gecko can read them easily.

FontFace

A @font-face rule that is known to have font-family and src declarations.

FontFaceRuleData

Data inside a @font-face rule.

FontFaceRuleParser 🔒

FontFaceSourceTechFlags

Flags for the @font-face tech() function, indicating font technologies required by the resource.

FontStretchRange

The font-stretch descriptor:

FontWeightRange

The font-weight descriptor:

SourceList

A list of sources for the font-face src descriptor.

UrlSource

A UrlSource represents a font-face source that has been specified with a url() function.

Enums

ComputedFontStyleDescriptor

The computed representation of the above, with angles in degrees, so that Gecko can read them easily.

FontDisplay

A font-display value for a @font-face rule. The font-display descriptor determines how a font face is displayed based on whether and when it is downloaded and ready to use.

FontFaceSourceFormat

FontFaceSourceFormatKeyword

Keywords for the font-face src descriptor’s format() function. (‘None’ and ‘Unknown’ are for internal use in gfx, not exposed to CSS.)

FontStyle

The font-style descriptor:

Source

A source for a font-face rule.

Functions

parse_font_face_block

Parse the block inside a @font-face rule.

sort_range 🔒