Struct style::font_face::FontFaceRuleData

pub struct FontFaceRuleData {
    pub family: Option<FamilyName>,
    pub sources: Option<Vec<Source>>,
    pub source_location: SourceLocation,
}

Data inside a @font-face rule.

https://drafts.csswg.org/css-fonts/#font-face-rule

Fields
family: Option<FamilyName>

The name of this font face

sources: Option<Vec<Source>>

The alternative sources for this font face.

source_location: SourceLocation

Line and column of the @font-face rule source code.

Methods

impl FontFaceRuleData

pub fn font_face(&self) -> Option<FontFace>

Per https://github.com/w3c/csswg-drafts/issues/1133 an @font-face rule is valid as far as the CSS parser is concerned even if it doesn’t have a font-family or src declaration.

However both are required for the rule to represent an actual font face.

impl FontFaceRule

pub fn clone_conditionally_gecko_or_servo(&self) -> FontFaceRule