style

Module global_style_data

Global style data

Structs

GLOBAL_STYLE_DATA

Global style data

GlobalStyleData

Global style data

STYLE_THREAD_JOIN_HANDLES 🔒

JoinHandles for spawned style threads. These will be joined during StyleThreadPool::shutdown() after exiting the thread pool.

STYLE_THREAD_POOL

Global thread pool

StyleThreadPool

Global thread pool.

Constants

STYLO_MAX_THREADS 🔒

The performance benefit of additional threads seems to level off at around six, so we cap it there on many-core machines (see bug 1431285 comment 14).

Functions

stylo_threads_pref 🔒

thread_name 🔒

thread_shutdown 🔒

thread_spawn 🔒

thread_startup 🔒

Type Aliases

PlatformThreadHandle

Platform-specific handle to a thread.