Module style::invalidation::element

Invalidation of element styles due to attribute or style changes.

Modules

document_state
An invalidation processor for style changes due to document state changes.

element_wrapper
A wrapper over an element and a snapshot, that allows us to selector-match against a past state of the element.

invalidation_map
Code for invalidations due to state or attribute changes.

invalidator
The struct that takes care of encapsulating all the logic on where and how element styles need to be invalidated.

restyle_hints
Restyle hints: an optimization to avoid unnecessarily matching selectors.

state_and_attributes
An invalidation processor for style changes due to state and attribute changes.