Struct style::invalidation::element::element_wrapper::ElementWrapper

pub struct ElementWrapper<'a, E> where
    E: TElement,  { /* fields omitted */ }

A simple wrapper over an element and a snapshot, that allows us to selector-match against a past state of the element.

Methods

impl<'a, E> ElementWrapper<'a, E> where
E: TElement,

pub fn new(el: E, snapshot_map: &'a SnapshotMap) -> Self

Trivially constructs an ElementWrapper.

pub fn snapshot(&self) -> Option<&'a Snapshot>

Gets the snapshot associated with this element, if any.

pub fn state_changes(&self) -> ElementState

Returns the states that have changed since the element was snapshotted.