style::invalidation::element

Module invalidation_map

Code for invalidations due to state or attribute changes.

Structs

AdditionalRelativeSelectorInvalidationMap

A map to store invalidation dependencies specific to relative selectors. This keeps a lot more data than the usual map, because any change can generate upward traversals that need to be handled separately.

Dependency

Mapping between (partial) CompoundSelectors (and the combinator to their right) and the states and attributes they depend on.

DocumentStateDependency

The same, but for document state selectors.

InvalidationMap

A map where we store invalidations.

NextDependencyEntry 🔒

PerCompoundState 🔒

RelativeSelectorCompoundStateAttributes 🔒

RelativeSelectorDependencyCollector 🔒

A struct that collects invalidations for a given compound selector.

RelativeSelectorInnerCollectorState 🔒

ScopeSelectorCollectorState 🔒

SelectorDependencyCollector 🔒

A struct that collects invalidations for a given compound selector.

StateDependency

The same, but for state selectors, which can track more exactly what state do they track.

TSStateDependency

Dependency for tree-structural pseudo-classes.

TSStateForInvalidation

Tree-structural pseudoclasses that we care about for (Relative selector) invalidation. Specifically, we need to store information on ones that don’t generate the inner selector. Given the nature of these selectors:

Enums

ComponentVisitResult 🔒

DependencyInvalidationKind

Invalidation kind merging normal and relative dependencies.

GeneratedInvalidation

The type of invalidation a non-relative selector can generate.

NormalDependencyInvalidationKind

The kind of elements down the tree this dependency may affect.

RelativeDependencyInvalidationKind

The kind of elements up the tree this relative selector dependency may affect. Because this travels upwards, it’s not viable for parallel subtree traversal, and is handled separately.

ScopeDependencyInvalidationKind

The kind of invalidation the subject of this dependency triggers.

Traits

Collector 🔒

Functions

add_attr_dependency 🔒

add_custom_state_dependency 🔒

add_local_name 🔒

add_non_unique_info 🔒

add_pseudo_class_dependency 🔒

add_ts_pseudo_class_dependency 🔒

get_non_relative_invalidation_kind 🔒

Return the type of normal invalidation given a selector & an offset.

next_dependency 🔒

note_selector_for_invalidation

Adds a selector to the given InvalidationMap. Returns Err(..) to signify OOM.

on_attribute 🔒

on_id_or_class 🔒

on_pseudo_class 🔒

on_scope 🔒

on_simple_selector 🔒

visit_all_in_iter_compound 🔒

Visit all the simple selectors in the iter compound and return the number of simple selectors visited. We need to return a tuple because we need to keep track of two things:

Type Aliases

AnyDependencyMap

Dependency mapping for * selectors.

CustomStateDependencyMap

Dependency mapping for customstates

IdOrClassDependencyMap

Dependency mapping for classes or IDs.

LocalNameDependencyMap

Dependency mapping for local names.

NextSelectors 🔒

StateDependencyMap

Dependency mapping for non-tree-strctural pseudo-class states.

TSStateDependencyMap

Dependency mapping for tree-structural pseudo-class states.