Module style::invalidation::element::invalidation_map

Code for invalidations due to state or attribute changes.

Structs

Dependency
Mapping between (partial) CompoundSelectors (and the combinator to their right) and the states and attributes they depend on.

DocumentStateDependency
The same, but for document state selectors.

InvalidationMap
A map where we store invalidations.

PerCompoundState

SelectorDependencyCollector
A struct that collects invalidations for a given compound selector.

StateDependency
The same, but for state selectors, which can track more exactly what state do they track.

Enums

DependencyInvalidationKind
The kind of elements down the tree this dependency may affect.