Module style::invalidation::element::invalidation_map

Code for invalidations due to state or attribute changes.

Structs

Dependency

Mapping between (partial) CompoundSelectors (and the combinator to their right) and the states and attributes they depend on.

DocumentStateDependency

The same, but for document state selectors.

InvalidationMap

A map where we store invalidations.

ParentDependencyEntry 🔒

PerCompoundState 🔒

RelativeSelectorDependencyCollector 🔒

A struct that collects invalidations for a given compound selector.

RelativeSelectorInnerDependencyCollector 🔒

A struct that collects invalidations from a complex selector inside a relative selector. TODO(dshin): All of this duplication is not great Perhaps should be merged to the normal one, if possible? See bug 1855690.

RelativeSelectorInvalidationMap

A map to store all relative selector invalidations. This keeps a lot more data than the usual map, because any change can generate upward traversals that need to be handled separately.

RelativeSelectorPerCompoundState 🔒

SelectorDependencyCollector 🔒

A struct that collects invalidations for a given compound selector.

StateDependency

The same, but for state selectors, which can track more exactly what state do they track.

TSStateDependency

Dependency for tree-structural pseudo-classes.

TSStateForInvalidation

Tree-structural pseudoclasses that we care about for (Relative selector) invalidation. Specifically, we need to store information on ones that don’t generate the inner selector.

Enums

ComponentVisitResult 🔒

DependencyInvalidationKind

Invalidation kind merging normal and relative dependencies.

NormalDependencyInvalidationKind

The kind of elements down the tree this dependency may affect.

RelativeDependencyInvalidationKind

The kind of elements up the tree this relative selector dependency may affect. Because this travels upwards, it’s not viable for parallel subtree traversal, and is handled separately.

Traits

Collector 🔒

Functions

add_attr_dependency 🔒

add_custom_state_dependency 🔒

add_local_name 🔒

add_non_unique_info 🔒

add_pseudo_class_dependency 🔒

add_ts_pseudo_class_dependency 🔒

note_selector_for_invalidation

Adds a selector to the given InvalidationMap. Returns Err(..) to signify OOM.

on_attribute 🔒

on_id_or_class 🔒

on_pseudo_class 🔒

on_simple_selector 🔒

parent_dependency 🔒

Type Aliases

AnyDependencyMap

Dependency mapping for * selectors.

CustomStateDependencyMap

Dependency mapping for customstates

IdOrClassDependencyMap

Dependency mapping for classes or IDs.

LocalNameDependencyMap

Dependency mapping for local names.

ParentSelectors 🔒

StateDependencyMap

Dependency mapping for non-tree-strctural pseudo-class states.

TSStateDependencyMap

Dependency mapping for tree-structural pseudo-class states.