style::invalidation::element::invalidation_map

Type Alias TSStateDependencyMap

pub type TSStateDependencyMap = SelectorMap<TSStateDependency>;

Dependency mapping for tree-structural pseudo-class states.

Aliased Type

pub struct TSStateDependencyMap {
    pub root: SmallVec<[TSStateDependency; 1]>,
    pub id_hash: MaybeCaseInsensitiveHashMap<Atom<AtomStaticSet>, SmallVec<[TSStateDependency; 1]>>,
    pub class_hash: MaybeCaseInsensitiveHashMap<Atom<AtomStaticSet>, SmallVec<[TSStateDependency; 1]>>,
    pub local_name_hash: HashMap<GenericAtomIdent<LocalNameStaticSet>, SmallVec<[TSStateDependency; 1]>, BuildHasherDefault<PrecomputedHasher>>,
    pub attribute_hash: HashMap<GenericAtomIdent<LocalNameStaticSet>, SmallVec<[TSStateDependency; 1]>, BuildHasherDefault<PrecomputedHasher>>,
    pub namespace_hash: HashMap<GenericAtomIdent<NamespaceStaticSet>, SmallVec<[TSStateDependency; 1]>, BuildHasherDefault<PrecomputedHasher>>,
    pub rare_pseudo_classes: SmallVec<[TSStateDependency; 1]>,
    pub other: SmallVec<[TSStateDependency; 1]>,
    pub count: usize,
}

Fields
root: SmallVec<[TSStateDependency; 1]>

Rules that have :root selectors.

id_hash: MaybeCaseInsensitiveHashMap<Atom<AtomStaticSet>, SmallVec<[TSStateDependency; 1]>>

A hash from an ID to rules which contain that ID selector.

class_hash: MaybeCaseInsensitiveHashMap<Atom<AtomStaticSet>, SmallVec<[TSStateDependency; 1]>>

A hash from a class name to rules which contain that class selector.

local_name_hash: HashMap<GenericAtomIdent<LocalNameStaticSet>, SmallVec<[TSStateDependency; 1]>, BuildHasherDefault<PrecomputedHasher>>

A hash from local name to rules which contain that local name selector.

attribute_hash: HashMap<GenericAtomIdent<LocalNameStaticSet>, SmallVec<[TSStateDependency; 1]>, BuildHasherDefault<PrecomputedHasher>>

A hash from attributes to rules which contain that attribute selector.

namespace_hash: HashMap<GenericAtomIdent<NamespaceStaticSet>, SmallVec<[TSStateDependency; 1]>, BuildHasherDefault<PrecomputedHasher>>

A hash from namespace to rules which contain that namespace selector.

rare_pseudo_classes: SmallVec<[TSStateDependency; 1]>

Rules for pseudo-states that are rare but have global selectors.

other: SmallVec<[TSStateDependency; 1]>

All other rules.

count: usize

The number of entries in this map.