Module style::invalidation::element::invalidator

The struct that takes care of encapsulating all the logic on where and how element styles need to be invalidated.

Structs

DescendantInvalidationLists

Different invalidation lists for descendants.

Invalidation

An Invalidation is a complex selector that describes which elements, relative to a current element we are processing, must be restyled.

InvalidationResult

The result of a whole invalidation process for a given element.

SingleInvalidationResult 🔒

The result of processing a single invalidation for a given element.

TreeStyleInvalidator

The struct that takes care of encapsulating all the logic on where and how element styles need to be invalidated.

Enums

DescendantInvalidationKind 🔒

The kind of descendant invalidation we’re processing.

InvalidationKind 🔒

The kind of invalidation we’re processing.

Traits

InvalidationProcessor

A trait to abstract the collection of invalidations for a given pass.

Type Aliases

InvalidationVector

A vector of invalidations, optimized for small invalidation sets.