style::invalidation::element::invalidator

Function note_scope_dependency_force_at_subject

pub fn note_scope_dependency_force_at_subject<'selectors>(
    dependency: &'selectors Dependency,
    current_host: Option<OpaqueElement>,
    scope: Option<OpaqueElement>,
    traversed_non_subject: bool,
) -> Vec<Invalidation<'selectors>>

Note the child dependencies of a scope end selector This is necessary because the scope end selector is not bound to :scope

e.g @scope to (.b) {:scope .a .c {…}} in the case of the following:

If we toggle class “b” in x, we would have to go up to find .a if we wanted to invalidate correctly. However, this is costly. Instead we just invalidate to the subject of the selector .c