Struct style::invalidation::element::invalidator::TreeStyleInvalidator

pub struct TreeStyleInvalidator<'a, 'b, E, P: 'a> where
    'b: 'a,
    E: TElement,
    P: InvalidationProcessor<'b, E>,  { /* fields omitted */ }

The struct that takes care of encapsulating all the logic on where and how element styles need to be invalidated.

Methods

impl<'a, 'b, E, P: 'a> TreeStyleInvalidator<'a, 'b, E, P> where
'b: 'a,
E: TElement,
P: InvalidationProcessor<'b, E>,

pub fn new(
element: E,
stack_limit_checker: Option<&'a StackLimitChecker>,
processor: &'a mut P
) -> Self

Trivially constructs a new TreeStyleInvalidator.

pub fn invalidate(self) -> InvalidationResult

Perform the invalidation pass.