Module style::invalidation::element::state_and_attributes

An invalidation processor for style changes due to state and attribute changes.

Structs

Collector 🔒

The collector implementation.

StateAndAttrInvalidationProcessor

An invalidation processor for style changes due to state and attribute changes.

Functions

check_dependency

Checks a dependency against a given element and wrapper, to see if something changed.

invalidated_descendants

Propagates the bits after invalidating a descendant child, if needed.

invalidated_self

Sets the appropriate restyle hint after invalidating the style of a given element.

invalidated_sibling

Sets the appropriate hint after invalidating the style of a sibling.

propagate_dirty_bit_up_to

Propagates the bits after invalidating a descendant child.

should_process_descendants

Whether we should process the descendants of a given element for style invalidation.