Module style::invalidation::media_queries

Code related to the invalidation of media-query-affected rules.

Structs

EffectiveMediaQueryResults
A struct that holds the result of a media query evaluation pass for the media queries that evaluated successfully.

MediaListKey
A key for a given media query result.

PotentiallyEffectiveMediaRules
A filter that filters over effective rules, but allowing all potentially effective @media rules.

Traits

ToMediaListKey
A trait to get a given MediaListKey for a given item that can hold a MediaList.