Struct style::invalidation::stylesheets::StylesheetInvalidationSet

pub struct StylesheetInvalidationSet { /* fields omitted */ }

A set of invalidations due to stylesheet additions.

TODO(emilio): We might be able to do the same analysis for media query changes too (or even selector changes?).

Methods

impl StylesheetInvalidationSet

pub fn new() -> Self

Create an empty StylesheetInvalidationSet.

pub fn invalidate_fully(&mut self)

Mark the DOM tree styles' as fully invalid.

pub fn collect_invalidations_for<S>(
&mut self,
device: &Device,
stylesheet: &S,
guard: &SharedRwLockReadGuard
) where
S: StylesheetInDocument,

Analyze the given stylesheet, and collect invalidations from their rules, in order to avoid doing a full restyle when we style the document next time.

pub fn flush<E>(&mut self, document_element: Option<E>) -> bool where
E: TElement,

Clears the invalidation set, invalidating elements as needed if document_element is provided.

Returns true if any invalidations ocurred.

pub fn clear(&mut self)

Clears the invalidation set without processing.