Struct style::logical_geometry::LogicalMargin

pub struct LogicalMargin<T> {
    pub block_start: T,
    pub inline_end: T,
    pub block_end: T,
    pub inline_start: T,
    debug_writing_mode: DebugWritingMode,
}

A "margin" in flow-relative dimensions Represents the four sides of the margins, borders, or padding of a CSS box, or a combination of those. A positive "margin" can be added to a rectangle to obtain a bigger rectangle.

Fields
block_start: Tinline_end: Tblock_end: Tinline_start: Tdebug_writing_mode: DebugWritingMode
Implementations

impl<T: Zero> LogicalMargin<T>

pub fn zero(mode: WritingMode) -> LogicalMargin<T>

impl<T> LogicalMargin<T>

pub fn new(
mode: WritingMode,
block_start: T,
inline_end: T,
block_end: T,
inline_start: T
) -> LogicalMargin<T>

pub fn from_physical(
mode: WritingMode,
offsets: SideOffsets2D<T>
) -> LogicalMargin<T>

impl<T: Copy> LogicalMargin<T>

pub fn new_all_same(mode: WritingMode, value: T) -> LogicalMargin<T>

pub fn top(&self, mode: WritingMode) -> T

pub fn set_top(&mut self, mode: WritingMode, top: T)

pub fn right(&self, mode: WritingMode) -> T

pub fn set_right(&mut self, mode: WritingMode, right: T)

pub fn bottom(&self, mode: WritingMode) -> T

pub fn set_bottom(&mut self, mode: WritingMode, bottom: T)

pub fn left(&self, mode: WritingMode) -> T

pub fn set_left(&mut self, mode: WritingMode, left: T)

pub fn to_physical(&self, mode: WritingMode) -> SideOffsets2D<T>

pub fn convert(
&self,
mode_from: WritingMode,
mode_to: WritingMode
) -> LogicalMargin<T>

impl<T: PartialEq + Zero> LogicalMargin<T>

pub fn is_zero(&self) -> bool

impl<T: Copy + Add<T, Output = T>> LogicalMargin<T>

pub fn inline_start_end(&self) -> T

pub fn block_start_end(&self) -> T

pub fn start_end(&self, direction: Direction) -> T

pub fn top_bottom(&self, mode: WritingMode) -> T

pub fn left_right(&self, mode: WritingMode) -> T