Struct style::logical_geometry::LogicalMargin

pub struct LogicalMargin<T> {
    pub block_start: T,
    pub inline_end: T,
    pub block_end: T,
    pub inline_start: T,
    // some fields omitted
}

A "margin" in flow-relative dimensions Represents the four sides of the margins, borders, or padding of a CSS box, or a combination of those. A positive "margin" can be added to a rectangle to obtain a bigger rectangle.

Fields
block_start: T inline_end: T block_end: T inline_start: T
Methods

impl<T: Zero> LogicalMargin<T>

fn zero(mode: WritingMode) -> LogicalMargin<T>

impl<T: Copy> LogicalMargin<T>

fn new(mode: WritingMode,
block_start: T,
inline_end: T,
block_end: T,
inline_start: T)
-> LogicalMargin<T>

fn new_all_same(mode: WritingMode, value: T) -> LogicalMargin<T>

fn from_physical(mode: WritingMode,
offsets: SideOffsets2D<T>)
-> LogicalMargin<T>

fn top(&self, mode: WritingMode) -> T

fn set_top(&mut self, mode: WritingMode, top: T)

fn right(&self, mode: WritingMode) -> T

fn set_right(&mut self, mode: WritingMode, right: T)

fn bottom(&self, mode: WritingMode) -> T

fn set_bottom(&mut self, mode: WritingMode, bottom: T)

fn left(&self, mode: WritingMode) -> T

fn set_left(&mut self, mode: WritingMode, left: T)

fn to_physical(&self, mode: WritingMode) -> SideOffsets2D<T>

fn convert(&self,
mode_from: WritingMode,
mode_to: WritingMode)
-> LogicalMargin<T>

impl<T: PartialEq + Zero> LogicalMargin<T>

fn is_zero(&self) -> bool

impl<T: Copy + Add<T, Output=T>> LogicalMargin<T>

fn inline_start_end(&self) -> T

fn block_start_end(&self) -> T

fn start_end(&self, direction: Direction) -> T

fn top_bottom(&self, mode: WritingMode) -> T

fn left_right(&self, mode: WritingMode) -> T