Module style::matching

Structs

ApplicableDeclarations

ApplicableDeclarationsCache

ApplicableDeclarationsCacheEntry

ApplicableDeclarationsCacheQuery

StyleSharingCandidate

StyleSharingCandidateCache
An LRU cache of the last few nodes seen, so that we can aggressively try to reuse their styles.

Enums

StyleSharingResult
The results of attempting to share a style.

Statics

APPLICABLE_DECLARATIONS_CACHE_SIZE

STYLE_SHARING_CANDIDATE_CACHE_SIZE

Traits

ElementMatchMethods

MatchMethods

PrivateElementMatchMethods

PrivateMatchMethods

Functions

create_common_style_affecting_attributes_from_element
High-level interface to CSS selector matching.