Module style::matching

High-level interface to CSS selector matching.

Structs

CommonStyleAffectingAttributeInfo
The information of how to match a given common-style affecting attribute.

CommonStyleAffectingAttributes
A set of common style-affecting attributes we check separately to optimize the style sharing cache.

MatchResults
The results returned from running selector matching on an element.

StyleSharingCandidateCache
An LRU cache of the last few nodes seen, so that we can aggressively try to reuse their styles.

Enums

CacheMiss
A cache miss result.

CommonStyleAffectingAttributeMode
How should we match a given common style-affecting attribute?

StyleSharingResult
The results of attempting to share a style.

Constants

ALIGN_CENTER_ATTRIBUTE
The align="center" attribute.

ALIGN_LEFT_ATTRIBUTE
The align="left" attribute.

ALIGN_RIGHT_ATTRIBUTE
The align="right" attribute.

HIDDEN_ATTRIBUTE
The hidden attribute.

NO_WRAP_ATTRIBUTE
The nowrap attribute.

Traits

MatchMethods
The public API that elements expose for selector matching.

Functions

common_style_affecting_attributes
The common style affecting attribute array.

rare_style_affecting_attributes
Attributes that, if present, disable style sharing. All legacy HTML attributes must be in either this list or common_style_affecting_attributes. See the comment in synthesize_presentational_hints_for_legacy_attributes.