Module style::matching

High-level interface to CSS selector matching.

Structs

StyleDifference

Represents the result of comparing an element’s old and new style.

Enums

CascadeVisitedMode 🔒

Determines which styles are being cascaded currently.

ChildRestyleRequirement

Whether or not newly computed values for an element need to be cascaded to children (or children might need to be re-matched, e.g., for container queries).

StyleChange

Represents whether or not the style of an element has changed.

Traits

MatchMethods

The public API that elements expose for selector matching.

PrivateMatchMethods 🔒