Struct style::matching::StyleSharingCandidate

pub struct StyleSharingCandidate<C: ComputedValues> {
    pub style: Arc<C>,
    pub parent_style: Arc<C>,
    pub local_name: Atom,
    pub class: Option<String>,
    pub namespace: Namespace,
    pub common_style_affecting_attributes: CommonStyleAffectingAttributes,
    pub link: bool,
}

Fields

style

parent_style

local_name

class

namespace

common_style_affecting_attributes

link

Methods

impl<C: ComputedValues> StyleSharingCandidate<C>

fn new<N: TNode<ConcreteComputedValues=C>>(element: &N::ConcreteElement) -> Option<Self>

Attempts to create a style sharing candidate from this node. Returns the style sharing candidate or None if this node is ineligible for style sharing.

fn can_share_style_with<E: TElement>(&self, element: &E) -> bool