Struct style::matching::StyleSharingCandidateCache

pub struct StyleSharingCandidateCache<C: ComputedValues> {
    cache: LRUCache<StyleSharingCandidate<C>, ()>,
}

An LRU cache of the last few nodes seen, so that we can aggressively try to reuse their styles.

Fields
cache: LRUCache<StyleSharingCandidate<C>, ()>
Methods

impl<C: ComputedValues> StyleSharingCandidateCache<C>

fn new() -> Self

fn iter(&self) -> Iter<(StyleSharingCandidate<C>, ())>

fn insert_if_possible<N: TNode<ConcreteComputedValues=C>>(&mut self, element: &N::ConcreteElement)

fn touch(&mut self, index: usize)