Struct style::matching::StyleSharingCandidateCache

pub struct StyleSharingCandidateCache<E: TElement> { /* fields omitted */ }

An LRU cache of the last few nodes seen, so that we can aggressively try to reuse their styles.

Note that this cache is flushed every time we steal work from the queue, so storing nodes here temporarily is safe.

Methods

impl<E: TElement> StyleSharingCandidateCache<E>

fn new() -> Self

Create a new style sharing candidate cache.

fn insert_if_possible(&mut self,
element: &E,
style: &Arc<ComputedValues>,
relations: StyleRelations)

Tries to insert an element in the style sharing cache.

Fails if we know it should never be in the cache.

fn touch(&mut self, index: usize)

Touch a given index in the style sharing candidate cache.

fn clear(&mut self)

Clear the style sharing candidate cache.