Trait style::matching::MatchMethods

pub trait MatchMethods: TNode {
    fn insert_into_bloom_filter(&self, bf: &mut BloomFilter) { ... }
    fn remove_from_bloom_filter(&self, bf: &mut BloomFilter) { ... }
    unsafe fn cascade_node(&self, context: &SharedStyleContext<Self::ConcreteElement::Impl>, parent: Option<Self>, applicable_declarations: &ApplicableDeclarations<Self::ConcreteElement::Impl>, applicable_declarations_cache: &mut ApplicableDeclarationsCache<Self::ConcreteComputedValues>, new_animations_sender: &Mutex<Sender<Animation>>) where Self::ConcreteElement::Impl: SelectorImplExt { ... }
}

Provided Methods

fn insert_into_bloom_filter(&self, bf: &mut BloomFilter)

Inserts and removes the matching Descendant selectors from a bloom filter. This is used to speed up CSS selector matching to remove unnecessary tree climbs for Descendant queries.

A bloom filter of the local names, namespaces, IDs, and classes is kept. Therefore, each node must have its matching selectors inserted after its own selector matching and before its children start.

fn remove_from_bloom_filter(&self, bf: &mut BloomFilter)

After all the children are done css selector matching, this must be called to reset the bloom filter after an insert.

unsafe fn cascade_node(&self, context: &SharedStyleContext<Self::ConcreteElement::Impl>, parent: Option<Self>, applicable_declarations: &ApplicableDeclarations<Self::ConcreteElement::Impl>, applicable_declarations_cache: &mut ApplicableDeclarationsCache<Self::ConcreteComputedValues>, new_animations_sender: &Mutex<Sender<Animation>>) where Self::ConcreteElement::Impl: SelectorImplExt