Trait style::matching::MatchMethods

pub trait MatchMethods: TElement {
    fn match_element(&self,
                     context: &mut StyleContext<Self>,
                     data: &mut ElementData)
                     -> MatchResults { ... }
    fn cascade_with_replacements(&self,
                                 hint: RestyleHint,
                                 context: &StyleContext<Self>,
                                 data: &mut AtomicRefMut<ElementData>)
                                 -> bool { ... }
    unsafe fn share_style_if_possible(&self,
                                      style_sharing_candidate_cache: &mut StyleSharingCandidateCache<Self>,
                                      shared_context: &SharedStyleContext,
                                      data: &mut AtomicRefMut<ElementData>)
                                      -> StyleSharingResult { ... }
    fn insert_into_bloom_filter(&self, bf: &mut BloomFilter) { ... }
    fn remove_from_bloom_filter(&self, bf: &mut BloomFilter) { ... }
    fn compute_restyle_damage(&self,
                              old_values: &Arc<ComputedValues>,
                              new_values: &Arc<ComputedValues>,
                              pseudo: Option<&PseudoElement>)
                              -> RestyleDamage { ... }
    fn cascade_element(&self,
                       context: &StyleContext<Self>,
                       data: &mut AtomicRefMut<ElementData>,
                       primary_is_shareable: bool) { ... }
}

The public API that elements expose for selector matching.

Provided Methods

fn match_element(&self,
context: &mut StyleContext<Self>,
data: &mut ElementData)
-> MatchResults

Runs selector matching to (re)compute rule nodes for this element.

fn cascade_with_replacements(&self,
hint: RestyleHint,
context: &StyleContext<Self>,
data: &mut AtomicRefMut<ElementData>)
-> bool

Updates the rule nodes without re-running selector matching, using just the rule tree. Returns true if the rule nodes changed.

unsafe fn share_style_if_possible(&self,
style_sharing_candidate_cache: &mut StyleSharingCandidateCache<Self>,
shared_context: &SharedStyleContext,
data: &mut AtomicRefMut<ElementData>)
-> StyleSharingResult

Attempts to share a style with another node. This method is unsafe because it depends on the style_sharing_candidate_cache having only live nodes in it, and we have no way to guarantee that at the type system level yet.

fn insert_into_bloom_filter(&self, bf: &mut BloomFilter)

Inserts and removes the matching Descendant selectors from a bloom filter. This is used to speed up CSS selector matching to remove unnecessary tree climbs for Descendant queries.

A bloom filter of the local names, namespaces, IDs, and classes is kept. Therefore, each node must have its matching selectors inserted after its own selector matching and before its children start.

fn remove_from_bloom_filter(&self, bf: &mut BloomFilter)

After all the children are done css selector matching, this must be called to reset the bloom filter after an insert.

fn compute_restyle_damage(&self,
old_values: &Arc<ComputedValues>,
new_values: &Arc<ComputedValues>,
pseudo: Option<&PseudoElement>)
-> RestyleDamage

Given the old and new style of this element, and whether it's a pseudo-element, compute the restyle damage used to determine which kind of layout or painting operations we'll need.

fn cascade_element(&self,
context: &StyleContext<Self>,
data: &mut AtomicRefMut<ElementData>,
primary_is_shareable: bool)

Run the CSS cascade and compute values for the element, potentially starting any new transitions or animations.