Trait style::matching::PrivateMatchMethods

trait PrivateMatchMethods: TElement {
    fn cascade_node_pseudo_element<'a, Ctx>(&self,
                                            context: &Ctx,
                                            parent_style: Option<&Arc<ComputedValues>>,
                                            old_style: Option<&Arc<ComputedValues>>,
                                            rule_node: &StrongRuleNode,
                                            possibly_expired_animations: &[PropertyAnimation],
                                            booleans: CascadeBooleans)
                                            -> Arc<ComputedValues> where Ctx: StyleContext<'a> { ... }
    fn update_animations_for_cascade(&self,
                                     context: &SharedStyleContext,
                                     style: &mut Arc<ComputedValues>,
                                     possibly_expired_animations: &mut Vec<PropertyAnimation>) { ... }
    fn share_style_with_candidate_if_possible(&self,
                                              shared_context: &SharedStyleContext,
                                              candidate: &mut StyleSharingCandidate)
                                              -> Result<ComputedStyle, CacheMiss> { ... }
}

Provided Methods

fn cascade_node_pseudo_element<'a, Ctx>(&self,
context: &Ctx,
parent_style: Option<&Arc<ComputedValues>>,
old_style: Option<&Arc<ComputedValues>>,
rule_node: &StrongRuleNode,
possibly_expired_animations: &[PropertyAnimation],
booleans: CascadeBooleans)
-> Arc<ComputedValues> where Ctx: StyleContext<'a>

Actually cascades style for a node or a pseudo-element of a node.

Note that animations only apply to nodes or ::before or ::after pseudo-elements.

fn update_animations_for_cascade(&self,
context: &SharedStyleContext,
style: &mut Arc<ComputedValues>,
possibly_expired_animations: &mut Vec<PropertyAnimation>)

fn share_style_with_candidate_if_possible(&self,
shared_context: &SharedStyleContext,
candidate: &mut StyleSharingCandidate)
-> Result<ComputedStyle, CacheMiss>