Trait style::matching::PrivateMatchMethods

trait PrivateMatchMethods: TElement {
    fn layout_parent(&self) -> Self { ... }
    fn cascade_internal(&self,
                        context: &StyleContext<Self>,
                        primary_style: &ComputedStyle,
                        pseudo_style: &mut Option<(&PseudoElement, &mut ComputedStyle)>,
                        booleans: &CascadeBooleans)
                        -> Arc<ComputedValues> { ... }
    fn cascade_primary_or_pseudo<'a>(&self,
                                     context: &mut StyleContext<Self>,
                                     data: &mut ElementData,
                                     pseudo: Option<&PseudoElement>,
                                     possibly_expired_animations: &mut Vec<PropertyAnimation>,
                                     booleans: CascadeBooleans) { ... }
    fn process_animations(&self,
                          context: &mut StyleContext<Self>,
                          old_values: &mut Option<Arc<ComputedValues>>,
                          new_values: &mut Arc<ComputedValues>,
                          _pseudo: Option<&PseudoElement>,
                          possibly_expired_animations: &mut Vec<PropertyAnimation>) { ... }
    fn accumulate_damage(&self,
                         restyle: &mut RestyleData,
                         old_values: &Arc<ComputedValues>,
                         new_values: &Arc<ComputedValues>,
                         pseudo: Option<&PseudoElement>) { ... }
    fn update_animations_for_cascade(&self,
                                     context: &SharedStyleContext,
                                     style: &mut Arc<ComputedValues>,
                                     possibly_expired_animations: &mut Vec<PropertyAnimation>) { ... }
    fn share_style_with_candidate_if_possible(&self,
                                              shared_context: &SharedStyleContext,
                                              candidate: &mut StyleSharingCandidate<Self>)
                                              -> Result<ComputedStyle, CacheMiss> { ... }
}

Provided Methods

fn layout_parent(&self) -> Self

Returns the closest parent element that doesn't have a display: contents style (and thus generates a box).

This is needed to correctly handle blockification of flex and grid items.

Returns itself if the element has no parent. In practice this doesn't happen because the root element is blockified per spec, but it could happen if we decide to not blockify for roots of disconnected subtrees, which is a kind of dubious beahavior.

fn cascade_internal(&self,
context: &StyleContext<Self>,
primary_style: &ComputedStyle,
pseudo_style: &mut Option<(&PseudoElement, &mut ComputedStyle)>,
booleans: &CascadeBooleans)
-> Arc<ComputedValues>

fn cascade_primary_or_pseudo<'a>(&self,
context: &mut StyleContext<Self>,
data: &mut ElementData,
pseudo: Option<&PseudoElement>,
possibly_expired_animations: &mut Vec<PropertyAnimation>,
booleans: CascadeBooleans)

Computes values and damage for the primary or pseudo style of an element, setting them on the ElementData.

fn process_animations(&self,
context: &mut StyleContext<Self>,
old_values: &mut Option<Arc<ComputedValues>>,
new_values: &mut Arc<ComputedValues>,
_pseudo: Option<&PseudoElement>,
possibly_expired_animations: &mut Vec<PropertyAnimation>)

fn accumulate_damage(&self,
restyle: &mut RestyleData,
old_values: &Arc<ComputedValues>,
new_values: &Arc<ComputedValues>,
pseudo: Option<&PseudoElement>)

Computes and applies restyle damage unless we've already maxed it out.

fn update_animations_for_cascade(&self,
context: &SharedStyleContext,
style: &mut Arc<ComputedValues>,
possibly_expired_animations: &mut Vec<PropertyAnimation>)

fn share_style_with_candidate_if_possible(&self,
shared_context: &SharedStyleContext,
candidate: &mut StyleSharingCandidate<Self>)
-> Result<ComputedStyle, CacheMiss>