style::parser

Struct ParserContext

pub struct ParserContext<'a> {
    pub stylesheet_origin: Origin,
    pub url_data: &'a UrlExtraData,
    pub parsing_mode: ParsingMode,
    pub quirks_mode: QuirksMode,
    error_reporter: Option<&'a dyn ParseErrorReporter>,
    pub namespaces: Cow<'a, Namespaces>,
    pub use_counters: Option<&'a UseCounters>,
    pub nesting_context: NestingContext,
}

The data that the parser needs from outside in order to parse a stylesheet.

Fields
stylesheet_origin: Origin

The Origin of the stylesheet, whether it’s a user, author or user-agent stylesheet.

url_data: &'a UrlExtraData

The extra data we need for resolving url values.

parsing_mode: ParsingMode

The mode to use when parsing.

quirks_mode: QuirksMode

The quirks mode of this stylesheet.

error_reporter: Option<&'a dyn ParseErrorReporter>

The active error reporter, or none if error reporting is disabled.

namespaces: Cow<'a, Namespaces>

The currently active namespaces.

use_counters: Option<&'a UseCounters>

The use counters we want to record while parsing style rules, if any.

nesting_context: NestingContext

Current nesting context.

Implementations

impl<'a> ParserContext<'a>

pub fn new( stylesheet_origin: Origin, url_data: &'a UrlExtraData, rule_type: Option<CssRuleType>, parsing_mode: ParsingMode, quirks_mode: QuirksMode, namespaces: Cow<'a, Namespaces>, error_reporter: Option<&'a dyn ParseErrorReporter>, use_counters: Option<&'a UseCounters>, ) -> Self

Create a parser context.

pub fn nest_for_rule<R>( &mut self, rule_type: CssRuleType, cb: impl FnOnce(&mut Self) -> R, ) -> R

Temporarily sets the rule_type and executes the callback function, returning its result.

pub fn in_page_rule(&self) -> bool

Whether we’re in a @page rule.

pub fn allows_important_declarations(&self) -> bool

Returns whether !important declarations are forbidden.

pub fn rule_types(&self) -> CssRuleTypes

Get the rule type, which assumes that one is available.

pub fn error_reporting_enabled(&self) -> bool

Returns whether CSS error reporting is enabled.

pub fn log_css_error( &self, location: SourceLocation, error: ContextualParseError<'_>, )

Record a CSS parse error with this context’s error reporting.

pub fn in_ua_sheet(&self) -> bool

Whether we’re in a user-agent stylesheet.

pub fn chrome_rules_enabled(&self) -> bool

Returns whether chrome-only rules should be parsed.

pub fn allows_computational_dependence(&self) -> bool

Whether the parsing mode allows units or functions that are not computationally independent.