Struct style::parser::ParserContext

pub struct ParserContext<'a> {
    pub stylesheet_origin: Origin,
    pub base_url: &'a ServoUrl,
    pub error_reporter: Box<ParseErrorReporter + Send>,
    pub extra_data: ParserContextExtraData,
}

The data that the parser needs from outside in order to parse a stylesheet.

Fields
stylesheet_origin: Origin

The Origin of the stylesheet, whether it's a user, author or user-agent stylesheet.

base_url: &'a ServoUrl

The base url we're parsing this stylesheet as.

error_reporter: Box<ParseErrorReporter + Send>

An error reporter to report syntax errors.

extra_data: ParserContextExtraData

Implementation-dependent extra data.

Methods

impl<'a> ParserContext<'a>

fn new_with_extra_data(stylesheet_origin: Origin,
base_url: &'a ServoUrl,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)
-> ParserContext<'a>

Create a ParserContext with extra data.

fn new(stylesheet_origin: Origin,
base_url: &'a ServoUrl,
error_reporter: Box<ParseErrorReporter + Send>)
-> ParserContext<'a>

Create a parser context with the default extra data.

fn new_for_cssom(base_url: &'a ServoUrl) -> ParserContext<'a>

Create a parser context for on-the-fly parsing in CSSOM