Trait style::parser::Parse

pub trait Parse: Sized {
    fn parse<'i, 't>(
        context: &ParserContext, 
        input: &mut Parser<'i, 't>
    ) -> Result<Self, ParseError<'i>>;
}

A trait to abstract parsing of a specified value given a ParserContext and CSS input.

This can be derived on keywords with #[derive(Parse)].

The derive code understands the following attributes on each of the variants:

#[parse(aliases = "foo,bar")] can be used to alias a value with another at parse-time.

#[parse(condition = "function")] can be used to make the parsing of the value conditional on function, which needs to fulfill fn(&ParserContext) -> bool.

Required methods

fn parse<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>
) -> Result<Self, ParseError<'i>>

Parse a value of this type.

Returns an error on failure.

Loading content...