Module style::properties

The CSS properties supported by the style system. Generated from the properties.mako.rs template by build.rs

Reexports

pub use self::declaration_block::*;

Modules

animated_properties
A module with all the code related to animated properties.

declaration_block
A property declaration block.

lazy_static_module

longhands
A module with all the code for longhand properties.

shorthands
A module with code for all the shorthand css properties, and a few serialization helpers.

style_structs
The module where all the style structs are defined.

Structs

AnimationDelayIter
An iterator over the values of the animation-delay properties.

AnimationDirectionIter
An iterator over the values of the animation-direction properties.

AnimationDurationIter
An iterator over the values of the animation-duration properties.

AnimationFillModeIter
An iterator over the values of the animation-fill-mode properties.

AnimationIterationCountIter
An iterator over the values of the animation-iteration-count properties.

AnimationNameIter
An iterator over the values of the animation-name properties.

AnimationPlayStateIter
An iterator over the values of the animation-play-state properties.

AnimationTimingFunctionIter
An iterator over the values of the animation-timing-function properties.

CascadeFlags
A set of flags to tweak the behavior of the cascade function.

ComputedValues
The struct that Servo uses to represent computed values.

FontFamilyIter
An iterator over the values of the font-family properties.

INITIAL_SERVO_VALUES

LonghandIdSet
A set of longhand properties

PropertyDeclarationIdSet
A specialized set of PropertyDeclarationId

TransitionDelayIter
An iterator over the values of the transition-delay properties.

TransitionDurationIter
An iterator over the values of the transition-duration properties.

TransitionPropertyIter
An iterator over the values of the transition-property properties.

TransitionTimingFunctionIter
An iterator over the values of the transition-timing-function properties.

UnparsedValue
An unparsed property value that contains var() functions.

Enums

CSSWideKeyword
An enum to represent a CSS Wide keyword.

DeclaredValue
Servo's representation of a declared value for a given T, which is the declared value for that property.

LonghandId
An identifier for a given longhand property.

PropertyDeclaration
Servo's representation for a property declaration.

PropertyDeclarationId
An identifier for a given property declaration, which can be either a longhand or a custom property.

PropertyDeclarationParseResult
The result of parsing a property declaration.

PropertyId
Servo's representation of a CSS property, that is, either a longhand, a shorthand, or a custom property.

ShorthandId
An identifier for a given shorthand property.

Constants

INHERIT_ALL
Whether to inherit all styles from the parent. If this flag is not present, non-inherited styles are reset to their initial values.

SHAREABLE
Whether the ComputedValues structure to be constructed should be considered shareable.

SKIP_ROOT_AND_ITEM_BASED_DISPLAY_FIXUP
Whether to skip any root element and flex/grid item display style fixup.

Statics

CASCADE_PROPERTY
A per-longhand array of functions to perform the CSS cascade on each of them, effectively doing virtual dispatch.

Functions

apply_declarations
NOTE: This function expects the declaration with more priority to appear first.

cascade
Performs the CSS cascade, computing new styles for an element from its parent style.

get_writing_mode
Return a WritingMode bitflags from the relevant CSS properties.

modify_border_style_for_inline_sides
Adjusts borders as appropriate to account for a fragment's status as the first or last fragment within the range of an element.

modify_style_for_anonymous_flow
Modifies the style for an anonymous flow so it resets all its non-inherited style structs, and set their borders and outlines to zero.

modify_style_for_inline_absolute_hypothetical_fragment
Adjusts the clip property so that an inline absolute hypothetical fragment doesn't clip its children.

modify_style_for_outer_inline_block_fragment
Adjusts the position property as necessary for the outer fragment wrapper of an inline-block.

modify_style_for_replaced_content
Alters the given style to accommodate replaced content. This is called in flow construction. It handles cases like:

modify_style_for_text
Adjusts the position and padding properties as necessary to account for text.

substitute_variables__servo_overflow_clip_box
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables__servo_overflow_clip_box_slow

substitute_variables_align_content
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_align_content_slow

substitute_variables_align_items
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_align_items_slow

substitute_variables_align_self
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_align_self_slow

substitute_variables_animation_delay
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_delay_slow

substitute_variables_animation_direction
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_direction_slow

substitute_variables_animation_duration
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_duration_slow

substitute_variables_animation_fill_mode
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_fill_mode_slow

substitute_variables_animation_iteration_count
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_iteration_count_slow

substitute_variables_animation_name
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_name_slow

substitute_variables_animation_play_state
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_play_state_slow

substitute_variables_animation_timing_function
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_animation_timing_function_slow

substitute_variables_backface_visibility
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_backface_visibility_slow

substitute_variables_background_attachment
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_attachment_slow

substitute_variables_background_clip
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_clip_slow

substitute_variables_background_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_color_slow

substitute_variables_background_image
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_image_slow

substitute_variables_background_origin
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_origin_slow

substitute_variables_background_position_x
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_position_x_slow

substitute_variables_background_position_y
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_position_y_slow

substitute_variables_background_repeat
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_repeat_slow

substitute_variables_background_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_background_size_slow

substitute_variables_block_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_block_size_slow

substitute_variables_border_block_end_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_block_end_color_slow

substitute_variables_border_block_end_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_block_end_style_slow

substitute_variables_border_block_end_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_block_end_width_slow

substitute_variables_border_block_start_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_block_start_color_slow

substitute_variables_border_block_start_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_block_start_style_slow

substitute_variables_border_block_start_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_block_start_width_slow

substitute_variables_border_bottom_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_bottom_color_slow

substitute_variables_border_bottom_left_radius
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_bottom_left_radius_slow

substitute_variables_border_bottom_right_radius
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_bottom_right_radius_slow

substitute_variables_border_bottom_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_bottom_style_slow

substitute_variables_border_bottom_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_bottom_width_slow

substitute_variables_border_collapse
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_collapse_slow

substitute_variables_border_image_outset
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_image_outset_slow

substitute_variables_border_image_repeat
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_image_repeat_slow

substitute_variables_border_image_slice
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_image_slice_slow

substitute_variables_border_image_source
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_image_source_slow

substitute_variables_border_image_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_image_width_slow

substitute_variables_border_inline_end_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_inline_end_color_slow

substitute_variables_border_inline_end_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_inline_end_style_slow

substitute_variables_border_inline_end_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_inline_end_width_slow

substitute_variables_border_inline_start_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_inline_start_color_slow

substitute_variables_border_inline_start_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_inline_start_style_slow

substitute_variables_border_inline_start_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_inline_start_width_slow

substitute_variables_border_left_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_left_color_slow

substitute_variables_border_left_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_left_style_slow

substitute_variables_border_left_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_left_width_slow

substitute_variables_border_right_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_right_color_slow

substitute_variables_border_right_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_right_style_slow

substitute_variables_border_right_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_right_width_slow

substitute_variables_border_spacing
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_spacing_slow

substitute_variables_border_top_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_top_color_slow

substitute_variables_border_top_left_radius
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_top_left_radius_slow

substitute_variables_border_top_right_radius
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_top_right_radius_slow

substitute_variables_border_top_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_top_style_slow

substitute_variables_border_top_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_border_top_width_slow

substitute_variables_bottom
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_bottom_slow

substitute_variables_box_shadow
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_box_shadow_slow

substitute_variables_box_sizing
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_box_sizing_slow

substitute_variables_caption_side
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_caption_side_slow

substitute_variables_clear
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_clear_slow

substitute_variables_clip
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_clip_slow

substitute_variables_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_color_slow

substitute_variables_column_count
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_column_count_slow

substitute_variables_column_gap
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_column_gap_slow

substitute_variables_column_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_column_width_slow

substitute_variables_content
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_content_slow

substitute_variables_counter_increment
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_counter_increment_slow

substitute_variables_counter_reset
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_counter_reset_slow

substitute_variables_cursor
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_cursor_slow

substitute_variables_direction
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_direction_slow

substitute_variables_display
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_display_slow

substitute_variables_empty_cells
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_empty_cells_slow

substitute_variables_filter
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_filter_slow

substitute_variables_flex_basis
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_flex_basis_slow

substitute_variables_flex_direction
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_flex_direction_slow

substitute_variables_flex_grow
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_flex_grow_slow

substitute_variables_flex_shrink
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_flex_shrink_slow

substitute_variables_flex_wrap
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_flex_wrap_slow

substitute_variables_float
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_float_slow

substitute_variables_font_family
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_font_family_slow

substitute_variables_font_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_font_size_slow

substitute_variables_font_stretch
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_font_stretch_slow

substitute_variables_font_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_font_style_slow

substitute_variables_font_variant
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_font_variant_slow

substitute_variables_font_weight
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_font_weight_slow

substitute_variables_height
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_height_slow

substitute_variables_image_rendering
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_image_rendering_slow

substitute_variables_inline_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_inline_size_slow

substitute_variables_justify_content
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_justify_content_slow

substitute_variables_left
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_left_slow

substitute_variables_letter_spacing
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_letter_spacing_slow

substitute_variables_line_height
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_line_height_slow

substitute_variables_list_style_image
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_list_style_image_slow

substitute_variables_list_style_position
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_list_style_position_slow

substitute_variables_list_style_type
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_list_style_type_slow

substitute_variables_margin_block_end
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_block_end_slow

substitute_variables_margin_block_start
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_block_start_slow

substitute_variables_margin_bottom
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_bottom_slow

substitute_variables_margin_inline_end
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_inline_end_slow

substitute_variables_margin_inline_start
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_inline_start_slow

substitute_variables_margin_left
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_left_slow

substitute_variables_margin_right
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_right_slow

substitute_variables_margin_top
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_margin_top_slow

substitute_variables_max_block_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_max_block_size_slow

substitute_variables_max_height
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_max_height_slow

substitute_variables_max_inline_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_max_inline_size_slow

substitute_variables_max_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_max_width_slow

substitute_variables_min_block_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_min_block_size_slow

substitute_variables_min_height
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_min_height_slow

substitute_variables_min_inline_size
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_min_inline_size_slow

substitute_variables_min_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_min_width_slow

substitute_variables_mix_blend_mode
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_mix_blend_mode_slow

substitute_variables_offset_block_end
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_offset_block_end_slow

substitute_variables_offset_block_start
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_offset_block_start_slow

substitute_variables_offset_inline_end
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_offset_inline_end_slow

substitute_variables_offset_inline_start
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_offset_inline_start_slow

substitute_variables_opacity
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_opacity_slow

substitute_variables_order
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_order_slow

substitute_variables_outline_color
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_outline_color_slow

substitute_variables_outline_offset
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_outline_offset_slow

substitute_variables_outline_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_outline_style_slow

substitute_variables_outline_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_outline_width_slow

substitute_variables_overflow_wrap
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_overflow_wrap_slow

substitute_variables_overflow_x
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_overflow_x_slow

substitute_variables_overflow_y
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_overflow_y_slow

substitute_variables_padding_block_end
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_block_end_slow

substitute_variables_padding_block_start
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_block_start_slow

substitute_variables_padding_bottom
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_bottom_slow

substitute_variables_padding_inline_end
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_inline_end_slow

substitute_variables_padding_inline_start
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_inline_start_slow

substitute_variables_padding_left
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_left_slow

substitute_variables_padding_right
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_right_slow

substitute_variables_padding_top
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_padding_top_slow

substitute_variables_perspective
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_perspective_origin
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_perspective_origin_slow

substitute_variables_perspective_slow

substitute_variables_pointer_events
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_pointer_events_slow

substitute_variables_position
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_position_slow

substitute_variables_quotes
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_quotes_slow

substitute_variables_right
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_right_slow

substitute_variables_table_layout
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_table_layout_slow

substitute_variables_text_align
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_align_slow

substitute_variables_text_decoration
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_decoration_slow

substitute_variables_text_indent
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_indent_slow

substitute_variables_text_justify
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_justify_slow

substitute_variables_text_overflow
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_overflow_slow

substitute_variables_text_rendering
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_rendering_slow

substitute_variables_text_shadow
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_shadow_slow

substitute_variables_text_transform
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_text_transform_slow

substitute_variables_top
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_top_slow

substitute_variables_transform
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_transform_origin
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_transform_origin_slow

substitute_variables_transform_slow

substitute_variables_transform_style
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_transform_style_slow

substitute_variables_transition_delay
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_transition_delay_slow

substitute_variables_transition_duration
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_transition_duration_slow

substitute_variables_transition_property
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_transition_property_slow

substitute_variables_transition_timing_function
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_transition_timing_function_slow

substitute_variables_unicode_bidi
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_unicode_bidi_slow

substitute_variables_vertical_align
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_vertical_align_slow

substitute_variables_visibility
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_visibility_slow

substitute_variables_white_space
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_white_space_slow

substitute_variables_width
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_width_slow

substitute_variables_word_break
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_word_break_slow

substitute_variables_word_spacing
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_word_spacing_slow

substitute_variables_writing_mode
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_writing_mode_slow

substitute_variables_z_index
Perform CSS variable substitution if needed, and execute f with the resulting declared value.

substitute_variables_z_index_slow

Type Definitions

CascadePropertyFn
A per-longhand function that performs the CSS cascade for that longhand.

ServoComputedValues
A legacy alias for a servo-version of ComputedValues. Should go away soon.