Module style::properties

Modules

longhands

shorthands

style_structs

Structs

ComputedValues

PropertyDeclarationBlock
Declarations are stored in reverse order. Overridden declarations are skipped.

Enums

CSSWideKeyword

DeclaredValue

PropertyDeclaration

PropertyDeclarationParseResult

Functions

cascade
Performs the CSS cascade, computing new styles for an element from its parent style and optionally a cached related style. The arguments are:

cascade_anonymous
Equivalent to cascade() with an empty applicable_declarations Performs the CSS cascade for an anonymous box.

is_supported_property

longhands_from_shorthand

make_border
Sets border_left_width to the passed in values. If border_left_width == 0 also sets border_left_style = none.

make_inline
Sets display to inline and position to static.

parse_property_declaration_list

parse_style_attribute