Module style::properties

The CSS properties supported by the style system. Generated from the properties.mako.rs template by build.rs

Reexports

pub use self::declaration_block::*;

Modules

animated_properties
A module with all the code related to animated properties.

computed_value_flags
Misc information about a given computed style.

declaration_block
A property declaration block.

longhands
A module with all the code for longhand properties.

shorthands
A module with code for all the shorthand css properties, and a few serialization helpers.

style_structs
The module where all the style structs are defined.

Structs

AnimationDelayIter
An iterator over the values of the animation-delay properties.

AnimationDirectionIter
An iterator over the values of the animation-direction properties.

AnimationDurationIter
An iterator over the values of the animation-duration properties.

AnimationFillModeIter
An iterator over the values of the animation-fill-mode properties.

AnimationIterationCountIter
An iterator over the values of the animation-iteration-count properties.

AnimationNameIter
An iterator over the values of the animation-name properties.

AnimationPlayStateIter
An iterator over the values of the animation-play-state properties.

AnimationTimingFunctionIter
An iterator over the values of the animation-timing-function properties.

BuilderArc

CascadeFlags
A set of flags to tweak the behavior of the cascade function.

ComputedValues
The struct that Servo uses to represent computed values.

ComputedValuesInner
Actual data of ComputedValues, to match up with Gecko

INITIAL_SERVO_VALUES

LonghandIdSet
A set of longhand properties

LonghandIdSetIterator
An iterator over a set of longhand ids.

NonCustomPropertyId
A longhand or shorthand porperty

NonCustomPropertyIdSet
A set of all properties

PropertyDeclarationIdSet
A specialized set of PropertyDeclarationId

PropertyFlags
A set of flags for properties.

SourcePropertyDeclaration
A stack-allocated vector of PropertyDeclaration large enough to parse one CSS key: value declaration. (Shorthands expand to multiple PropertyDeclarations.)

SourcePropertyDeclarationDrain
Return type of SourcePropertyDeclaration::drain

StyleBuilder
A type used to compute a struct with minimal overhead.

TransitionDelayIter
An iterator over the values of the transition-delay properties.

TransitionDurationIter
An iterator over the values of the transition-duration properties.

TransitionPropertyIter
An iterator over the values of the transition-property properties.

TransitionTimingFunctionIter
An iterator over the values of the transition-timing-function properties.

UnparsedValue
An unparsed property value that contains var() functions.

Enums

AliasId
An identifier for a given alias property.

CSSWideKeyword
An enum to represent a CSS Wide keyword.

DeclaredValue
Servo's representation of a declared value for a given T, which is the declared value for that property.

DeclaredValueOwned
A variant of DeclaredValue that owns its data. This separation exists so that PropertyDeclaration can avoid embedding a DeclaredValue (and its extra discriminant word) and synthesize dependent DeclaredValues for PropertyDeclaration instances as needed.

LonghandId
An identifier for a given longhand property.

PropertyDeclaration
Servo's representation for a property declaration.

PropertyDeclarationId
An identifier for a given property declaration, which can be either a longhand or a custom property.

PropertyId
Servo's representation of a CSS property, that is, either a longhand, a shorthand, or a custom property.

ShorthandId
An identifier for a given shorthand property.

StyleStructRef
A reference to a style struct of the parent, or our own style struct.

Traits

MaybeBoxed
Conversion with fewer impls than From/Into

Functions

adjust_border_width
See StyleAdjuster::adjust_for_border_width.

apply_declarations
NOTE: This function expects the declaration with more priority to appear first.

cascade
Performs the CSS cascade, computing new styles for an element from its parent style.

Type Definitions

CascadePropertyFn
A per-longhand function that performs the CSS cascade for that longhand.