Module style::properties::animated_properties

A module with all the code related to animated properties.

This needs to be “included” by mako at least after all longhand modules, given they populate the global data.

Macros

animated_list_impl

Structs

AnimationValueVariantRepr

TransitionPropertyIteration

A single iteration of the TransitionPropertyIterator.

TransitionPropertyIterator

An iterator over all the properties that transition on a given style.

Enums

AnimationValue

An enum to represent a single computed value belonging to an animated property in order to be interpolated with another one. When interpolating, both values need to belong to the same property.

Traits

ListAnimation

A trait to abstract away the different kind of animations over a list that there may be.

Functions

animate_discrete

Type Definitions

AnimationValueMap

A collection of AnimationValue that were composed on an element. This HashMap stores the values that are the last AnimationValue to be composed for each TransitionProperty.