Enum style::properties::animated_properties::AnimatedProperty

pub enum AnimatedProperty {
    BackgroundColor(T, T),
    BackgroundPositionX(T, T),
    BackgroundPositionY(T, T),
    BackgroundSize(T, T),
    BorderTopColor(T, T),
    BorderLeftColor(T, T),
    BorderBottomColor(T, T),
    BorderRightColor(T, T),
    BorderTopWidth(T, T),
    BorderLeftWidth(T, T),
    BorderBottomWidth(T, T),
    BorderRightWidth(T, T),
    BorderTopLeftRadius(T, T),
    BorderTopRightRadius(T, T),
    BorderBottomRightRadius(T, T),
    BorderBottomLeftRadius(T, T),
    VerticalAlign(T, T),
    ScrollSnapDestination(T, T),
    ScrollSnapCoordinate(T, T),
    Transform(T, T),
    Perspective(T, T),
    TransformOrigin(T, T),
    Color(T, T),
    ColumnRuleWidth(T, T),
    ColumnRuleColor(T, T),
    Opacity(T, T),
    BoxShadow(T, T),
    FontWeight(T, T),
    FontSize(T, T),
    FontSizeAdjust(T, T),
    Visibility(T, T),
    LineHeight(T, T),
    TextIndent(T, T),
    TextShadow(T, T),
    TextEmphasisColor(T, T),
    WebkitTextFillColor(T, T),
    WebkitTextStrokeColor(T, T),
    MarginTop(T, T),
    MarginLeft(T, T),
    MarginBottom(T, T),
    MarginRight(T, T),
    OutlineColor(T, T),
    OutlineWidth(T, T),
    OutlineOffset(T, T),
    PaddingTop(T, T),
    PaddingLeft(T, T),
    PaddingBottom(T, T),
    PaddingRight(T, T),
    Top(T, T),
    Left(T, T),
    Bottom(T, T),
    Right(T, T),
    ZIndex(T, T),
    FlexGrow(T, T),
    FlexShrink(T, T),
    Order(T, T),
    Width(T, T),
    MinWidth(T, T),
    MaxWidth(T, T),
    Height(T, T),
    MinHeight(T, T),
    MaxHeight(T, T),
    ObjectPosition(T, T),
    GridRowGap(T, T),
    GridColumnGap(T, T),
    TextDecorationColor(T, T),
    MaskPositionX(T, T),
    MaskPositionY(T, T),
    MaskSize(T, T),
}

An animated property interpolation between two computed values for that property.

Variants
BackgroundColor(T, T)

background-color

BackgroundPositionX(T, T)

background-position-x

BackgroundPositionY(T, T)

background-position-y

BackgroundSize(T, T)

background-size

BorderTopColor(T, T)

border-top-color

BorderLeftColor(T, T)

border-left-color

BorderBottomColor(T, T)

border-bottom-color

BorderRightColor(T, T)

border-right-color

BorderTopWidth(T, T)

border-top-width

BorderLeftWidth(T, T)

border-left-width

BorderBottomWidth(T, T)

border-bottom-width

BorderRightWidth(T, T)

border-right-width

BorderTopLeftRadius(T, T)

border-top-left-radius

BorderTopRightRadius(T, T)

border-top-right-radius

BorderBottomRightRadius(T, T)

border-bottom-right-radius

BorderBottomLeftRadius(T, T)

border-bottom-left-radius

VerticalAlign(T, T)

vertical-align

ScrollSnapDestination(T, T)

scroll-snap-destination

ScrollSnapCoordinate(T, T)

scroll-snap-coordinate

Transform(T, T)

transform

Perspective(T, T)

perspective

TransformOrigin(T, T)

transform-origin

Color(T, T)

color

ColumnRuleWidth(T, T)

column-rule-width

ColumnRuleColor(T, T)

column-rule-color

Opacity(T, T)

opacity

BoxShadow(T, T)

box-shadow

FontWeight(T, T)

font-weight

FontSize(T, T)

font-size

FontSizeAdjust(T, T)

font-size-adjust

Visibility(T, T)

visibility

LineHeight(T, T)

line-height

TextIndent(T, T)

text-indent

TextShadow(T, T)

text-shadow

TextEmphasisColor(T, T)

text-emphasis-color

WebkitTextFillColor(T, T)

-webkit-text-fill-color

WebkitTextStrokeColor(T, T)

-webkit-text-stroke-color

MarginTop(T, T)

margin-top

MarginLeft(T, T)

margin-left

MarginBottom(T, T)

margin-bottom

MarginRight(T, T)

margin-right

OutlineColor(T, T)

outline-color

OutlineWidth(T, T)

outline-width

OutlineOffset(T, T)

outline-offset

PaddingTop(T, T)

padding-top

PaddingLeft(T, T)

padding-left

PaddingBottom(T, T)

padding-bottom

PaddingRight(T, T)

padding-right

Top(T, T)

top

Left(T, T)

left

Bottom(T, T)

bottom

Right(T, T)

right

ZIndex(T, T)

z-index

FlexGrow(T, T)

flex-grow

FlexShrink(T, T)

flex-shrink

Order(T, T)

order

Width(T, T)

width

MinWidth(T, T)

min-width

MaxWidth(T, T)

max-width

Height(T, T)

height

MinHeight(T, T)

min-height

MaxHeight(T, T)

max-height

ObjectPosition(T, T)

object-position

GridRowGap(T, T)

grid-row-gap

GridColumnGap(T, T)

grid-column-gap

TextDecorationColor(T, T)

text-decoration-color

MaskPositionX(T, T)

mask-position-x

MaskPositionY(T, T)

mask-position-y

MaskSize(T, T)

mask-size

Methods

impl AnimatedProperty

fn name(&self) -> &'static str

Get the name of this property.

fn does_animate(&self) -> bool

Whether this interpolation does animate, that is, whether the start and end values are different.

fn has_the_same_end_value_as(&self, other: &Self) -> bool

Whether an animated property has the same end value as another.

fn update(&self, style: &mut ComputedValues, progress: f64)

Update style with the proper computed style corresponding to this animation at progress.

fn from_transition_property(transition_property: &TransitionProperty,
old_style: &ComputedValues,
new_style: &ComputedValues)
-> AnimatedProperty

Get an animatable value from a transition-property, an old style, and a new style.