Enum style::properties::animated_properties::AnimatedProperty

pub enum AnimatedProperty {
    BackgroundColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BackgroundImage(T, T),
    BackgroundPositionX(T, T),
    BackgroundPositionY(T, T),
    BackgroundRepeat(T, T),
    BackgroundAttachment(T, T),
    BackgroundClip(T, T),
    BackgroundOrigin(T, T),
    BackgroundSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopStyle(T, T),
    BorderTopWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderLeftColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderLeftStyle(T, T),
    BorderLeftWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomStyle(T, T),
    BorderBottomWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderRightColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderRightStyle(T, T),
    BorderRightWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderTopRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderBottomLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    BorderImageSource(T, T),
    BorderImageOutset(T, T),
    BorderImageRepeat(T, T),
    BorderImageWidth(T, T),
    BorderImageSlice(T, T),
    Display(T, T),
    Position(T, T),
    Float(T, T),
    Clear(T, T),
    VerticalAlign(T, T),
    OverflowX(T, T),
    OverflowY(T, T),
    Transform(T, T),
    Perspective(T, T),
    PerspectiveOrigin(T, T),
    BackfaceVisibility(T, T),
    TransformStyle(T, T),
    TransformOrigin(T, T),
    Color(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    ColumnWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    ColumnCount(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    ColumnGap(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Content(T, T),
    CounterIncrement(T, T),
    CounterReset(T, T),
    Opacity(T, T),
    BoxShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Clip(T, T),
    Filter(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    MixBlendMode(T, T),
    FontFamily(T, T),
    FontStyle(T, T),
    FontVariantCaps(T, T),
    FontWeight(T, T),
    FontSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    FontStretch(T, T),
    Visibility(T, T),
    WritingMode(T, T),
    Direction(T, T),
    ImageRendering(T, T),
    BorderCollapse(T, T),
    EmptyCells(T, T),
    CaptionSide(T, T),
    BorderSpacing(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    LineHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    TextTransform(T, T),
    TextIndent(T, T),
    OverflowWrap(T, T),
    WordBreak(T, T),
    TextJustify(T, T),
    TextAlign(T, T),
    LetterSpacing(T, T),
    WordSpacing(T, T),
    WhiteSpace(T, T),
    TextShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    TextRendering(T, T),
    ListStylePosition(T, T),
    ListStyleType(T, T),
    ListStyleImage(T, T),
    Quotes(T, T),
    MarginTop(T, T),
    MarginLeft(T, T),
    MarginBottom(T, T),
    MarginRight(T, T),
    OutlineColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    OutlineStyle(T, T),
    OutlineWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    OutlineOffset(T, T),
    PaddingTop(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PaddingLeft(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PaddingBottom(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    PaddingRight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    Cursor(T, T),
    PointerEvents(T, T),
    Top(T, T),
    Left(T, T),
    Bottom(T, T),
    Right(T, T),
    ZIndex(T, T),
    FlexDirection(T, T),
    FlexWrap(T, T),
    JustifyContent(T, T),
    AlignContent(T, T),
    AlignItems(T, T),
    FlexGrow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    FlexShrink(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue),
    AlignSelf(T, T),
    Order(T, T),
    Width(T, T),
    MinWidth(T, T),
    MaxWidth(T, T),
    Height(T, T),
    MinHeight(T, T),
    MaxHeight(T, T),
    BoxSizing(T, T),
    TableLayout(T, T),
    TextOverflow(T, T),
    UnicodeBidi(T, T),
    TextDecorationLine(T, T),
}

An animated property interpolation between two computed values for that property.

Variants
BackgroundColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

background-color

BackgroundImage(T, T)

background-image

BackgroundPositionX(T, T)

background-position-x

BackgroundPositionY(T, T)

background-position-y

BackgroundRepeat(T, T)

background-repeat

BackgroundAttachment(T, T)

background-attachment

BackgroundClip(T, T)

background-clip

BackgroundOrigin(T, T)

background-origin

BackgroundSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

background-size

BorderTopColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-color

BorderTopStyle(T, T)

border-top-style

BorderTopWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-width

BorderLeftColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-left-color

BorderLeftStyle(T, T)

border-left-style

BorderLeftWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-left-width

BorderBottomColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-color

BorderBottomStyle(T, T)

border-bottom-style

BorderBottomWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-width

BorderRightColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-right-color

BorderRightStyle(T, T)

border-right-style

BorderRightWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-right-width

BorderTopLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-left-radius

BorderTopRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-top-right-radius

BorderBottomRightRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-right-radius

BorderBottomLeftRadius(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-bottom-left-radius

BorderImageSource(T, T)

border-image-source

BorderImageOutset(T, T)

border-image-outset

BorderImageRepeat(T, T)

border-image-repeat

BorderImageWidth(T, T)

border-image-width

BorderImageSlice(T, T)

border-image-slice

Display(T, T)

display

Position(T, T)

position

Float(T, T)

float

Clear(T, T)

clear

VerticalAlign(T, T)

vertical-align

OverflowX(T, T)

overflow-x

OverflowY(T, T)

overflow-y

Transform(T, T)

transform

Perspective(T, T)

perspective

PerspectiveOrigin(T, T)

perspective-origin

BackfaceVisibility(T, T)

backface-visibility

TransformStyle(T, T)

transform-style

TransformOrigin(T, T)

transform-origin

Color(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

color

ColumnWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

column-width

ColumnCount(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

column-count

ColumnGap(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

column-gap

Content(T, T)

content

CounterIncrement(T, T)

counter-increment

CounterReset(T, T)

counter-reset

Opacity(T, T)

opacity

BoxShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

box-shadow

Clip(T, T)

clip

Filter(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

filter

MixBlendMode(T, T)

mix-blend-mode

FontFamily(T, T)

font-family

FontStyle(T, T)

font-style

FontVariantCaps(T, T)

font-variant-caps

FontWeight(T, T)

font-weight

FontSize(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

font-size

FontStretch(T, T)

font-stretch

Visibility(T, T)

visibility

WritingMode(T, T)

writing-mode

Direction(T, T)

direction

ImageRendering(T, T)

image-rendering

BorderCollapse(T, T)

border-collapse

EmptyCells(T, T)

empty-cells

CaptionSide(T, T)

caption-side

BorderSpacing(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

border-spacing

LineHeight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

line-height

TextTransform(T, T)

text-transform

TextIndent(T, T)

text-indent

OverflowWrap(T, T)

overflow-wrap

WordBreak(T, T)

word-break

TextJustify(T, T)

text-justify

TextAlign(T, T)

text-align

LetterSpacing(T, T)

letter-spacing

WordSpacing(T, T)

word-spacing

WhiteSpace(T, T)

white-space

TextShadow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

text-shadow

TextRendering(T, T)

text-rendering

ListStylePosition(T, T)

list-style-position

ListStyleType(T, T)

list-style-type

ListStyleImage(T, T)

list-style-image

Quotes(T, T)

quotes

MarginTop(T, T)

margin-top

MarginLeft(T, T)

margin-left

MarginBottom(T, T)

margin-bottom

MarginRight(T, T)

margin-right

OutlineColor(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

outline-color

OutlineStyle(T, T)

outline-style

OutlineWidth(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

outline-width

OutlineOffset(T, T)

outline-offset

PaddingTop(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-top

PaddingLeft(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-left

PaddingBottom(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-bottom

PaddingRight(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

padding-right

Cursor(T, T)

cursor

PointerEvents(T, T)

pointer-events

Top(T, T)

top

Left(T, T)

left

Bottom(T, T)

bottom

Right(T, T)

right

ZIndex(T, T)

z-index

FlexDirection(T, T)

flex-direction

FlexWrap(T, T)

flex-wrap

JustifyContent(T, T)

justify-content

AlignContent(T, T)

align-content

AlignItems(T, T)

align-items

FlexGrow(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

flex-grow

FlexShrink(<T as ToAnimatedValue>::AnimatedValue, <T as ToAnimatedValue>::AnimatedValue)

flex-shrink

AlignSelf(T, T)

align-self

Order(T, T)

order

Width(T, T)

width

MinWidth(T, T)

min-width

MaxWidth(T, T)

max-width

Height(T, T)

height

MinHeight(T, T)

min-height

MaxHeight(T, T)

max-height

BoxSizing(T, T)

box-sizing

TableLayout(T, T)

table-layout

TextOverflow(T, T)

text-overflow

UnicodeBidi(T, T)

unicode-bidi

TextDecorationLine(T, T)

text-decoration-line

Methods

impl AnimatedProperty

pub fn name(&self) -> &'static str

Get the name of this property.

pub fn does_animate(&self) -> bool

Whether this interpolation does animate, that is, whether the start and end values are different.

pub fn has_the_same_end_value_as(&self, other: &Self) -> bool

Whether an animated property has the same end value as another.

pub fn update(&self, style: &mut ComputedValues, progress: f64)

Update style with the proper computed style corresponding to this animation at progress.

pub fn from_longhand(
property: &LonghandId,
old_style: &ComputedValues,
new_style: &ComputedValues
) -> Option<AnimatedProperty>

Get an animatable value from a transition-property, an old style, and a new style.