Enum style::properties::animated_properties::AnimationValue

#[repr(u16)]
pub enum AnimationValue {

    AlignContent(<SpecifiedValue as ToComputedValue>::ComputedValue),
    AlignItems(<SpecifiedValue as ToComputedValue>::ComputedValue),
    AlignSelf(<SpecifiedValue as ToComputedValue>::ComputedValue),
    AspectRatio(<AspectRatio as ToComputedValue>::ComputedValue),
    BackfaceVisibility(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BorderCollapse(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BorderImageRepeat(<BorderImageRepeat as ToComputedValue>::ComputedValue),
    BoxSizing(<SpecifiedValue as ToComputedValue>::ComputedValue),
    CaptionSide(<CaptionSide as ToComputedValue>::ComputedValue),
    Clear(<Clear as ToComputedValue>::ComputedValue),
    ColumnCount(<<ColumnCount as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Direction(Void),
    Display(<Display as ToComputedValue>::ComputedValue),
    EmptyCells(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FlexDirection(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FlexWrap(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Float(<Float as ToComputedValue>::ComputedValue),
    FontStretch(<<FontStretch as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontStyle(<<FontStyle as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontVariantCaps(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FontWeight(<<FontWeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ImageRendering(<SpecifiedValue as ToComputedValue>::ComputedValue),
    JustifyContent(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ListStylePosition(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ListStyleType(<SpecifiedValue as ToComputedValue>::ComputedValue),
    MixBlendMode(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Opacity(<Opacity as ToComputedValue>::ComputedValue),
    Order(<Integer as ToComputedValue>::ComputedValue),
    OutlineStyle(<OutlineStyle as ToComputedValue>::ComputedValue),
    OverflowWrap(<OverflowWrap as ToComputedValue>::ComputedValue),
    PointerEvents(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Position(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ServoOverflowClipBox(Void),
    ServoTopLayer(Void),
    TableLayout(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextAlign(<TextAlign as ToComputedValue>::ComputedValue),
    TextDecorationLine(<TextDecorationLine as ToComputedValue>::ComputedValue),
    TextJustify(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextRendering(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextTransform(<TextTransform as ToComputedValue>::ComputedValue),
    TransformStyle(<TransformStyle as ToComputedValue>::ComputedValue),
    UnicodeBidi(Void),
    Visibility(<SpecifiedValue as ToComputedValue>::ComputedValue),
    WhiteSpace(<SpecifiedValue as ToComputedValue>::ComputedValue),
    WordBreak(<WordBreak as ToComputedValue>::ComputedValue),
    WritingMode(Void),
    ZIndex(<ZIndex as ToComputedValue>::ComputedValue),
    FlexGrow(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FlexShrink(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OverflowBlock(Void),
    OverflowInline(Void),
    OverflowX(<Overflow as ToComputedValue>::ComputedValue),
    OverflowY(<Overflow as ToComputedValue>::ComputedValue),
    BorderBlockEndStyle(Void),
    BorderBlockStartStyle(Void),
    BorderBottomStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderInlineEndStyle(Void),
    BorderInlineStartStyle(Void),
    BorderLeftStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderRightStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderTopStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    AnimationDelay(Void),
    AnimationDirection(Void),
    AnimationDuration(Void),
    AnimationFillMode(Void),
    AnimationIterationCount(Void),
    AnimationName(Void),
    AnimationPlayState(Void),
    AnimationTimingFunction(Void),
    BackgroundAttachment(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundClip(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundImage(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundOrigin(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundPositionX(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundPositionY(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundRepeat(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundSize(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageOutset(<<NonNegativeLengthOrNumberRect as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageSlice(<<BorderImageSlice as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageWidth(<<BorderImageWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderSpacing(<<BorderSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BoxShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Clip(<ClipRectOrAuto as ToComputedValue>::ComputedValue),
    Color(<<ColorPropertyValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ColumnGap(<<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ColumnWidth(<<NonNegativeLengthOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Content(<Content as ToComputedValue>::ComputedValue),
    CounterIncrement(<CounterIncrement as ToComputedValue>::ComputedValue),
    CounterReset(<CounterSetOrReset as ToComputedValue>::ComputedValue),
    Cursor(<Cursor as ToComputedValue>::ComputedValue),
    Filter(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FlexBasis(<<FlexBasis as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontFamily(<FontFamily as ToComputedValue>::ComputedValue),
    FontSize(<<FontSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    LetterSpacing(<LetterSpacing as ToComputedValue>::ComputedValue),
    LineHeight(<<LineHeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineOffset(<Length as ToComputedValue>::ComputedValue),
    Perspective(<<Perspective as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PerspectiveOrigin(<Position as ToComputedValue>::ComputedValue),
    Quotes(<Quotes as ToComputedValue>::ComputedValue),
    Rotate(<Rotate as ToComputedValue>::ComputedValue),
    Scale(<Scale as ToComputedValue>::ComputedValue),
    TextIndent(<LengthPercentage as ToComputedValue>::ComputedValue),
    TextOverflow(<TextOverflow as ToComputedValue>::ComputedValue),
    TextShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Transform(<Transform as ToComputedValue>::ComputedValue),
    TransformOrigin(<TransformOrigin as ToComputedValue>::ComputedValue),
    TransitionDelay(Void),
    TransitionDuration(Void),
    TransitionProperty(Void),
    TransitionTimingFunction(Void),
    Translate(<Translate as ToComputedValue>::ComputedValue),
    VerticalAlign(<VerticalAlign as ToComputedValue>::ComputedValue),
    WordSpacing(<WordSpacing as ToComputedValue>::ComputedValue),
    BorderImageSource(<Image as ToComputedValue>::ComputedValue),
    ListStyleImage(<Image as ToComputedValue>::ComputedValue),
    MaxBlockSize(Void),
    MaxHeight(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MaxInlineSize(Void),
    MaxWidth(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBottomLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBottomRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderEndEndRadius(Void),
    BorderEndStartRadius(Void),
    BorderStartEndRadius(Void),
    BorderStartStartRadius(Void),
    BorderTopLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingBlockEnd(Void),
    PaddingBlockStart(Void),
    PaddingBottom(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingInlineEnd(Void),
    PaddingInlineStart(Void),
    PaddingLeft(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingRight(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingTop(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BlockSize(Void),
    Height(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    InlineSize(Void),
    MinBlockSize(Void),
    MinHeight(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MinInlineSize(Void),
    MinWidth(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Width(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBlockEndWidth(Void),
    BorderBlockStartWidth(Void),
    BorderBottomWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderInlineEndWidth(Void),
    BorderInlineStartWidth(Void),
    BorderLeftWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderRightWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BackgroundColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBlockEndColor(Void),
    BorderBlockStartColor(Void),
    BorderBottomColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderInlineEndColor(Void),
    BorderInlineStartColor(Void),
    BorderLeftColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderRightColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Bottom(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    InsetBlockEnd(Void),
    InsetBlockStart(Void),
    InsetInlineEnd(Void),
    InsetInlineStart(Void),
    Left(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginBlockEnd(Void),
    MarginBlockStart(Void),
    MarginBottom(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginInlineEnd(Void),
    MarginInlineStart(Void),
    MarginLeft(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginRight(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    MarginTop(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    Right(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),
    Top(<LengthPercentageOrAuto as ToComputedValue>::ComputedValue),

}

An enum to represent a single computed value belonging to an animated property in order to be interpolated with another one. When interpolating, both values need to belong to the same property.

FIXME: We need to add a path for custom properties, but that’s trivial after this (is a similar path to that of PropertyDeclaration).

Variants

AlignContent

align-content

Tuple Fields of AlignContent

0: <SpecifiedValue as ToComputedValue>::ComputedValue

AlignItems

align-items

Tuple Fields of AlignItems

0: <SpecifiedValue as ToComputedValue>::ComputedValue

AlignSelf

align-self

Tuple Fields of AlignSelf

0: <SpecifiedValue as ToComputedValue>::ComputedValue

AspectRatio

aspect-ratio

Tuple Fields of AspectRatio

0: <AspectRatio as ToComputedValue>::ComputedValue

BackfaceVisibility

backface-visibility

Tuple Fields of BackfaceVisibility

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BorderCollapse

border-collapse

Tuple Fields of BorderCollapse

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BorderImageRepeat

border-image-repeat

Tuple Fields of BorderImageRepeat

0: <BorderImageRepeat as ToComputedValue>::ComputedValue

BoxSizing

box-sizing

Tuple Fields of BoxSizing

0: <SpecifiedValue as ToComputedValue>::ComputedValue

CaptionSide

caption-side

Tuple Fields of CaptionSide

0: <CaptionSide as ToComputedValue>::ComputedValue

Clear

clear

Tuple Fields of Clear

0: <Clear as ToComputedValue>::ComputedValue

ColumnCount

column-count

Tuple Fields of ColumnCount

0: <<ColumnCount as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

Direction

direction

Tuple Fields of Direction

0: Void

Display

display

Tuple Fields of Display

0: <Display as ToComputedValue>::ComputedValue

EmptyCells

empty-cells

Tuple Fields of EmptyCells

0: <SpecifiedValue as ToComputedValue>::ComputedValue

FlexDirection

flex-direction

Tuple Fields of FlexDirection

0: <SpecifiedValue as ToComputedValue>::ComputedValue

FlexWrap

flex-wrap

Tuple Fields of FlexWrap

0: <SpecifiedValue as ToComputedValue>::ComputedValue

Float

float

Tuple Fields of Float

0: <Float as ToComputedValue>::ComputedValue

FontStretch

font-stretch

Tuple Fields of FontStretch

0: <<FontStretch as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

FontStyle

font-style

Tuple Fields of FontStyle

0: <<FontStyle as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

FontVariantCaps

font-variant-caps

Tuple Fields of FontVariantCaps

0: <SpecifiedValue as ToComputedValue>::ComputedValue

FontWeight

font-weight

Tuple Fields of FontWeight

0: <<FontWeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

ImageRendering

image-rendering

Tuple Fields of ImageRendering

0: <SpecifiedValue as ToComputedValue>::ComputedValue

JustifyContent

justify-content

Tuple Fields of JustifyContent

0: <SpecifiedValue as ToComputedValue>::ComputedValue

ListStylePosition

list-style-position

Tuple Fields of ListStylePosition

0: <SpecifiedValue as ToComputedValue>::ComputedValue

ListStyleType

list-style-type

Tuple Fields of ListStyleType

0: <SpecifiedValue as ToComputedValue>::ComputedValue

MixBlendMode

mix-blend-mode

Tuple Fields of MixBlendMode

0: <SpecifiedValue as ToComputedValue>::ComputedValue

Opacity

opacity

Tuple Fields of Opacity

0: <Opacity as ToComputedValue>::ComputedValue

Order

order

Tuple Fields of Order

0: <Integer as ToComputedValue>::ComputedValue

OutlineStyle

outline-style

Tuple Fields of OutlineStyle

0: <OutlineStyle as ToComputedValue>::ComputedValue

OverflowWrap

overflow-wrap

Tuple Fields of OverflowWrap

0: <OverflowWrap as ToComputedValue>::ComputedValue

PointerEvents

pointer-events

Tuple Fields of PointerEvents

0: <SpecifiedValue as ToComputedValue>::ComputedValue

Position

position

Tuple Fields of Position

0: <SpecifiedValue as ToComputedValue>::ComputedValue

ServoOverflowClipBox

-servo-overflow-clip-box

Tuple Fields of ServoOverflowClipBox

0: Void

ServoTopLayer

-servo-top-layer

Tuple Fields of ServoTopLayer

0: Void

TableLayout

table-layout

Tuple Fields of TableLayout

0: <SpecifiedValue as ToComputedValue>::ComputedValue

TextAlign

text-align

Tuple Fields of TextAlign

0: <TextAlign as ToComputedValue>::ComputedValue

TextDecorationLine

text-decoration-line

Tuple Fields of TextDecorationLine

0: <TextDecorationLine as ToComputedValue>::ComputedValue

TextJustify

text-justify

Tuple Fields of TextJustify

0: <SpecifiedValue as ToComputedValue>::ComputedValue

TextRendering

text-rendering

Tuple Fields of TextRendering

0: <SpecifiedValue as ToComputedValue>::ComputedValue

TextTransform

text-transform

Tuple Fields of TextTransform

0: <TextTransform as ToComputedValue>::ComputedValue

TransformStyle

transform-style

Tuple Fields of TransformStyle

0: <TransformStyle as ToComputedValue>::ComputedValue

UnicodeBidi

unicode-bidi

Tuple Fields of UnicodeBidi

0: Void

Visibility

visibility

Tuple Fields of Visibility

0: <SpecifiedValue as ToComputedValue>::ComputedValue

WhiteSpace

white-space

Tuple Fields of WhiteSpace

0: <SpecifiedValue as ToComputedValue>::ComputedValue

WordBreak

word-break

Tuple Fields of WordBreak

0: <WordBreak as ToComputedValue>::ComputedValue

WritingMode

writing-mode

Tuple Fields of WritingMode

0: Void

ZIndex

z-index

Tuple Fields of ZIndex

0: <ZIndex as ToComputedValue>::ComputedValue

FlexGrow

flex-grow

Tuple Fields of FlexGrow

0: <<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

FlexShrink

flex-shrink

Tuple Fields of FlexShrink

0: <<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

OverflowBlock

overflow-block

Tuple Fields of OverflowBlock

0: Void

OverflowInline

overflow-inline

Tuple Fields of OverflowInline

0: Void

OverflowX

overflow-x

Tuple Fields of OverflowX

0: <Overflow as ToComputedValue>::ComputedValue

OverflowY

overflow-y

Tuple Fields of OverflowY

0: <Overflow as ToComputedValue>::ComputedValue

BorderBlockEndStyle

border-block-end-style

Tuple Fields of BorderBlockEndStyle

0: Void

BorderBlockStartStyle

border-block-start-style

Tuple Fields of BorderBlockStartStyle

0: Void

BorderBottomStyle

border-bottom-style

Tuple Fields of BorderBottomStyle

0: <BorderStyle as ToComputedValue>::ComputedValue

BorderInlineEndStyle

border-inline-end-style

Tuple Fields of BorderInlineEndStyle

0: Void

BorderInlineStartStyle

border-inline-start-style

Tuple Fields of BorderInlineStartStyle

0: Void

BorderLeftStyle

border-left-style

Tuple Fields of BorderLeftStyle

0: <BorderStyle as ToComputedValue>::ComputedValue

BorderRightStyle

border-right-style

Tuple Fields of BorderRightStyle

0: <BorderStyle as ToComputedValue>::ComputedValue

BorderTopStyle

border-top-style

Tuple Fields of BorderTopStyle

0: <BorderStyle as ToComputedValue>::ComputedValue

AnimationDelay

animation-delay

Tuple Fields of AnimationDelay

0: Void

AnimationDirection

animation-direction

Tuple Fields of AnimationDirection

0: Void

AnimationDuration

animation-duration

Tuple Fields of AnimationDuration

0: Void

AnimationFillMode

animation-fill-mode

Tuple Fields of AnimationFillMode

0: Void

AnimationIterationCount

animation-iteration-count

Tuple Fields of AnimationIterationCount

0: Void

AnimationName

animation-name

Tuple Fields of AnimationName

0: Void

AnimationPlayState

animation-play-state

Tuple Fields of AnimationPlayState

0: Void

AnimationTimingFunction

animation-timing-function

Tuple Fields of AnimationTimingFunction

0: Void

BackgroundAttachment

background-attachment

Tuple Fields of BackgroundAttachment

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BackgroundClip

background-clip

Tuple Fields of BackgroundClip

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BackgroundImage

background-image

Tuple Fields of BackgroundImage

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BackgroundOrigin

background-origin

Tuple Fields of BackgroundOrigin

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BackgroundPositionX

background-position-x

Tuple Fields of BackgroundPositionX

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BackgroundPositionY

background-position-y

Tuple Fields of BackgroundPositionY

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BackgroundRepeat

background-repeat

Tuple Fields of BackgroundRepeat

0: <SpecifiedValue as ToComputedValue>::ComputedValue

BackgroundSize

background-size

Tuple Fields of BackgroundSize

0: <<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderImageOutset

border-image-outset

Tuple Fields of BorderImageOutset

0: <<NonNegativeLengthOrNumberRect as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderImageSlice

border-image-slice

Tuple Fields of BorderImageSlice

0: <<BorderImageSlice as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderImageWidth

border-image-width

Tuple Fields of BorderImageWidth

0: <<BorderImageWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderSpacing

border-spacing

Tuple Fields of BorderSpacing

0: <<BorderSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BoxShadow

box-shadow

Tuple Fields of BoxShadow

0: <<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

Clip

clip

Tuple Fields of Clip

0: <ClipRectOrAuto as ToComputedValue>::ComputedValue

Color

color

Tuple Fields of Color

0: <<ColorPropertyValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

ColumnGap

column-gap

Tuple Fields of ColumnGap

0: <<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

ColumnWidth

column-width

Tuple Fields of ColumnWidth

0: <<NonNegativeLengthOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

Content

content

Tuple Fields of Content

0: <Content as ToComputedValue>::ComputedValue

CounterIncrement

counter-increment

Tuple Fields of CounterIncrement

0: <CounterIncrement as ToComputedValue>::ComputedValue

CounterReset

counter-reset

Tuple Fields of CounterReset

0: <CounterSetOrReset as ToComputedValue>::ComputedValue

Cursor

cursor

Tuple Fields of Cursor

0: <Cursor as ToComputedValue>::ComputedValue

Filter

filter

Tuple Fields of Filter

0: <<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

FlexBasis

flex-basis

Tuple Fields of FlexBasis

0: <<FlexBasis as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

FontFamily

font-family

Tuple Fields of FontFamily

0: <FontFamily as ToComputedValue>::ComputedValue

FontSize

font-size

Tuple Fields of FontSize

0: <<FontSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

LetterSpacing

letter-spacing

Tuple Fields of LetterSpacing

0: <LetterSpacing as ToComputedValue>::ComputedValue

LineHeight

line-height

Tuple Fields of LineHeight

0: <<LineHeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

OutlineOffset

outline-offset

Tuple Fields of OutlineOffset

0: <Length as ToComputedValue>::ComputedValue

Perspective

perspective

Tuple Fields of Perspective

0: <<Perspective as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

PerspectiveOrigin

perspective-origin

Tuple Fields of PerspectiveOrigin

0: <Position as ToComputedValue>::ComputedValue

Quotes

quotes

Tuple Fields of Quotes

0: <Quotes as ToComputedValue>::ComputedValue

Rotate

rotate

Tuple Fields of Rotate

0: <Rotate as ToComputedValue>::ComputedValue

Scale

scale

Tuple Fields of Scale

0: <Scale as ToComputedValue>::ComputedValue

TextIndent

text-indent

Tuple Fields of TextIndent

0: <LengthPercentage as ToComputedValue>::ComputedValue

TextOverflow

text-overflow

Tuple Fields of TextOverflow

0: <TextOverflow as ToComputedValue>::ComputedValue

TextShadow

text-shadow

Tuple Fields of TextShadow

0: <<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

Transform

transform

Tuple Fields of Transform

0: <Transform as ToComputedValue>::ComputedValue

TransformOrigin

transform-origin

Tuple Fields of TransformOrigin

0: <TransformOrigin as ToComputedValue>::ComputedValue

TransitionDelay

transition-delay

Tuple Fields of TransitionDelay

0: Void

TransitionDuration

transition-duration

Tuple Fields of TransitionDuration

0: Void

TransitionProperty

transition-property

Tuple Fields of TransitionProperty

0: Void

TransitionTimingFunction

transition-timing-function

Tuple Fields of TransitionTimingFunction

0: Void

Translate

translate

Tuple Fields of Translate

0: <Translate as ToComputedValue>::ComputedValue

VerticalAlign

vertical-align

Tuple Fields of VerticalAlign

0: <VerticalAlign as ToComputedValue>::ComputedValue

WordSpacing

word-spacing

Tuple Fields of WordSpacing

0: <WordSpacing as ToComputedValue>::ComputedValue

BorderImageSource

border-image-source

Tuple Fields of BorderImageSource

0: <Image as ToComputedValue>::ComputedValue

ListStyleImage

list-style-image

Tuple Fields of ListStyleImage

0: <Image as ToComputedValue>::ComputedValue

MaxBlockSize

max-block-size

Tuple Fields of MaxBlockSize

0: Void

MaxHeight

max-height

Tuple Fields of MaxHeight

0: <<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

MaxInlineSize

max-inline-size

Tuple Fields of MaxInlineSize

0: Void

MaxWidth

max-width

Tuple Fields of MaxWidth

0: <<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderBottomLeftRadius

border-bottom-left-radius

Tuple Fields of BorderBottomLeftRadius

0: <<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderBottomRightRadius

border-bottom-right-radius

Tuple Fields of BorderBottomRightRadius

0: <<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderEndEndRadius

border-end-end-radius

Tuple Fields of BorderEndEndRadius

0: Void

BorderEndStartRadius

border-end-start-radius

Tuple Fields of BorderEndStartRadius

0: Void

BorderStartEndRadius

border-start-end-radius

Tuple Fields of BorderStartEndRadius

0: Void

BorderStartStartRadius

border-start-start-radius

Tuple Fields of BorderStartStartRadius

0: Void

BorderTopLeftRadius

border-top-left-radius

Tuple Fields of BorderTopLeftRadius

0: <<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderTopRightRadius

border-top-right-radius

Tuple Fields of BorderTopRightRadius

0: <<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

PaddingBlockEnd

padding-block-end

Tuple Fields of PaddingBlockEnd

0: Void

PaddingBlockStart

padding-block-start

Tuple Fields of PaddingBlockStart

0: Void

PaddingBottom

padding-bottom

Tuple Fields of PaddingBottom

0: <<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

PaddingInlineEnd

padding-inline-end

Tuple Fields of PaddingInlineEnd

0: Void

PaddingInlineStart

padding-inline-start

Tuple Fields of PaddingInlineStart

0: Void

PaddingLeft

padding-left

Tuple Fields of PaddingLeft

0: <<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

PaddingRight

padding-right

Tuple Fields of PaddingRight

0: <<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

PaddingTop

padding-top

Tuple Fields of PaddingTop

0: <<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BlockSize

block-size

Tuple Fields of BlockSize

0: Void

Height

height

Tuple Fields of Height

0: <<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

InlineSize

inline-size

Tuple Fields of InlineSize

0: Void

MinBlockSize

min-block-size

Tuple Fields of MinBlockSize

0: Void

MinHeight

min-height

Tuple Fields of MinHeight

0: <<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

MinInlineSize

min-inline-size

Tuple Fields of MinInlineSize

0: Void

MinWidth

min-width

Tuple Fields of MinWidth

0: <<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

Width

width

Tuple Fields of Width

0: <<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderBlockEndWidth

border-block-end-width

Tuple Fields of BorderBlockEndWidth

0: Void

BorderBlockStartWidth

border-block-start-width

Tuple Fields of BorderBlockStartWidth

0: Void

BorderBottomWidth

border-bottom-width

Tuple Fields of BorderBottomWidth

0: <<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderInlineEndWidth

border-inline-end-width

Tuple Fields of BorderInlineEndWidth

0: Void

BorderInlineStartWidth

border-inline-start-width

Tuple Fields of BorderInlineStartWidth

0: Void

BorderLeftWidth

border-left-width

Tuple Fields of BorderLeftWidth

0: <<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderRightWidth

border-right-width

Tuple Fields of BorderRightWidth

0: <<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderTopWidth

border-top-width

Tuple Fields of BorderTopWidth

0: <<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

OutlineWidth

outline-width

Tuple Fields of OutlineWidth

0: <<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BackgroundColor

background-color

Tuple Fields of BackgroundColor

0: <<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderBlockEndColor

border-block-end-color

Tuple Fields of BorderBlockEndColor

0: Void

BorderBlockStartColor

border-block-start-color

Tuple Fields of BorderBlockStartColor

0: Void

BorderBottomColor

border-bottom-color

Tuple Fields of BorderBottomColor

0: <<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderInlineEndColor

border-inline-end-color

Tuple Fields of BorderInlineEndColor

0: Void

BorderInlineStartColor

border-inline-start-color

Tuple Fields of BorderInlineStartColor

0: Void

BorderLeftColor

border-left-color

Tuple Fields of BorderLeftColor

0: <<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderRightColor

border-right-color

Tuple Fields of BorderRightColor

0: <<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

BorderTopColor

border-top-color

Tuple Fields of BorderTopColor

0: <<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

OutlineColor

outline-color

Tuple Fields of OutlineColor

0: <<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue

Bottom

bottom

Tuple Fields of Bottom

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

InsetBlockEnd

inset-block-end

Tuple Fields of InsetBlockEnd

0: Void

InsetBlockStart

inset-block-start

Tuple Fields of InsetBlockStart

0: Void

InsetInlineEnd

inset-inline-end

Tuple Fields of InsetInlineEnd

0: Void

InsetInlineStart

inset-inline-start

Tuple Fields of InsetInlineStart

0: Void

Left

left

Tuple Fields of Left

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

MarginBlockEnd

margin-block-end

Tuple Fields of MarginBlockEnd

0: Void

MarginBlockStart

margin-block-start

Tuple Fields of MarginBlockStart

0: Void

MarginBottom

margin-bottom

Tuple Fields of MarginBottom

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

MarginInlineEnd

margin-inline-end

Tuple Fields of MarginInlineEnd

0: Void

MarginInlineStart

margin-inline-start

Tuple Fields of MarginInlineStart

0: Void

MarginLeft

margin-left

Tuple Fields of MarginLeft

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

MarginRight

margin-right

Tuple Fields of MarginRight

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

MarginTop

margin-top

Tuple Fields of MarginTop

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

Right

right

Tuple Fields of Right

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

Top

top

Tuple Fields of Top

0: <LengthPercentageOrAuto as ToComputedValue>::ComputedValue

Implementations

impl AnimationValue

pub fn id(&self) -> LonghandId

Returns the longhand id this animated value corresponds to.

pub fn uncompute(&self) -> PropertyDeclaration

“Uncompute” this animation value in order to be used inside the CSS cascade.

pub fn from_declaration(
decl: &PropertyDeclaration,
context: &mut Context<'_>,
extra_custom_properties: Option<&Arc<CustomPropertiesMap>>,
initial: &ComputedValues
) -> Option<Self>

Construct an AnimationValue from a property declaration.

pub fn from_computed_values(
property: LonghandId,
style: &ComputedValues
) -> Option<Self>

Get an AnimationValue for an AnimatableLonghand from a given computed values.

pub fn set_in_style_for_servo(&self, style: &mut ComputedValues)

Update style with the value of this AnimationValue.

SERVO ONLY: This doesn’t properly handle things like updating ‘em’ units when animated font-size.