Enum style::properties::animated_properties::TransitionProperty

pub enum TransitionProperty {
    All,
    BackgroundColor,
    BackgroundPositionX,
    BackgroundPositionY,
    BackgroundSize,
    BorderTopColor,
    BorderLeftColor,
    BorderBottomColor,
    BorderRightColor,
    BorderTopWidth,
    BorderLeftWidth,
    BorderBottomWidth,
    BorderRightWidth,
    BorderTopLeftRadius,
    BorderTopRightRadius,
    BorderBottomRightRadius,
    BorderBottomLeftRadius,
    VerticalAlign,
    ScrollSnapDestination,
    ScrollSnapCoordinate,
    Transform,
    Perspective,
    TransformOrigin,
    Color,
    ColumnRuleWidth,
    ColumnRuleColor,
    Opacity,
    BoxShadow,
    FontWeight,
    FontSize,
    FontSizeAdjust,
    Visibility,
    LineHeight,
    TextIndent,
    TextShadow,
    TextEmphasisColor,
    WebkitTextFillColor,
    WebkitTextStrokeColor,
    MarginTop,
    MarginLeft,
    MarginBottom,
    MarginRight,
    OutlineColor,
    OutlineWidth,
    OutlineOffset,
    PaddingTop,
    PaddingLeft,
    PaddingBottom,
    PaddingRight,
    Top,
    Left,
    Bottom,
    Right,
    ZIndex,
    FlexGrow,
    FlexShrink,
    Order,
    Width,
    MinWidth,
    MaxWidth,
    Height,
    MinHeight,
    MaxHeight,
    ObjectPosition,
    GridRowGap,
    GridColumnGap,
    TextDecorationColor,
    MaskPositionX,
    MaskPositionY,
    MaskSize,
}

A given transition property, that is either All, or an animatable property.

Variants
All

All, any animatable property changing should generate a transition.

BackgroundColor

background-color

BackgroundPositionX

background-position-x

BackgroundPositionY

background-position-y

BackgroundSize

background-size

BorderTopColor

border-top-color

BorderLeftColor

border-left-color

BorderBottomColor

border-bottom-color

BorderRightColor

border-right-color

BorderTopWidth

border-top-width

BorderLeftWidth

border-left-width

BorderBottomWidth

border-bottom-width

BorderRightWidth

border-right-width

BorderTopLeftRadius

border-top-left-radius

BorderTopRightRadius

border-top-right-radius

BorderBottomRightRadius

border-bottom-right-radius

BorderBottomLeftRadius

border-bottom-left-radius

VerticalAlign

vertical-align

ScrollSnapDestination

scroll-snap-destination

ScrollSnapCoordinate

scroll-snap-coordinate

Transform

transform

Perspective

perspective

TransformOrigin

transform-origin

Color

color

ColumnRuleWidth

column-rule-width

ColumnRuleColor

column-rule-color

Opacity

opacity

BoxShadow

box-shadow

FontWeight

font-weight

FontSize

font-size

FontSizeAdjust

font-size-adjust

Visibility

visibility

LineHeight

line-height

TextIndent

text-indent

TextShadow

text-shadow

TextEmphasisColor

text-emphasis-color

WebkitTextFillColor

-webkit-text-fill-color

WebkitTextStrokeColor

-webkit-text-stroke-color

MarginTop

margin-top

MarginLeft

margin-left

MarginBottom

margin-bottom

MarginRight

margin-right

OutlineColor

outline-color

OutlineWidth

outline-width

OutlineOffset

outline-offset

PaddingTop

padding-top

PaddingLeft

padding-left

PaddingBottom

padding-bottom

PaddingRight

padding-right

Top

top

Left

left

Bottom

bottom

Right

right

ZIndex

z-index

FlexGrow

flex-grow

FlexShrink

flex-shrink

Order

order

Width

width

MinWidth

min-width

MaxWidth

max-width

Height

height

MinHeight

min-height

MaxHeight

max-height

ObjectPosition

object-position

GridRowGap

grid-row-gap

GridColumnGap

grid-column-gap

TextDecorationColor

text-decoration-color

MaskPositionX

mask-position-x

MaskPositionY

mask-position-y

MaskSize

mask-size

Methods

impl TransitionProperty

fn each<F: FnMut(TransitionProperty) -> ()>(cb: F)

Iterates over each property that is not All.

fn parse(input: &mut Parser) -> Result<Self, ()>

Parse a transition-property value.

fn from_declaration(declaration: &PropertyDeclaration) -> Option<Self>

Get a transition property from a property declaration.