Function style::properties::cascade::cascade

pub fn cascade<E>(
    device: &Device, 
    pseudo: Option<&PseudoElement>, 
    rule_node: &StrongRuleNode, 
    guards: &StylesheetGuards, 
    parent_style: Option<&ComputedValues>, 
    parent_style_ignoring_first_line: Option<&ComputedValues>, 
    layout_parent_style: Option<&ComputedValues>, 
    visited_rules: Option<&StrongRuleNode>, 
    font_metrics_provider: &dyn FontMetricsProvider, 
    quirks_mode: QuirksMode, 
    rule_cache: Option<&RuleCache>, 
    rule_cache_conditions: &mut RuleCacheConditions, 
    element: Option<E>
) -> Arc<ComputedValues> where
    E: TElement, 

Performs the CSS cascade, computing new styles for an element from its parent style.

The arguments are:

device: Used to get the initial viewport and other external state.

rule_node: The rule node in the tree that represent the CSS rules that matched.

parent_style: The parent style, if applicable; if None, this is the root node.

Returns the computed values.

flags: Various flags.