Module style::properties::declaration_block

A property declaration block.

Structs

AnimationRules
The animation rules.

AnimationValueIterator
Iterator for AnimationValue to be generated from PropertyDeclarationBlock.

DeclarationImportanceIterator
Iterator over (PropertyDeclaration, Importance) pairs.

NormalDeclarationIterator
Iterator over PropertyDeclaration for Importance::Normal.

PropertyDeclarationBlock
Overridden declarations are skipped.

Enums

AppendableValue
A convenient enum to represent different kinds of stuff that can represent a value in the serialization of a property declaration.

DeclarationSource
Whether a given declaration comes from CSS parsing, or from CSSOM.

Importance
A declaration importance.

Functions

append_declaration_value
Append a given kind of appendable value to a serialization.

append_serialization
Append a given property and value pair to a serialization.

parse_one_declaration_into
Parse a given property declaration. Can result in multiple PropertyDeclarations when expanding a shorthand, for example.

parse_property_declaration_list
Parse a list of property declarations and return a property declaration block.

parse_style_attribute
A helper to parse the style attribute of an element, in order for this to be shared between Servo and Gecko.