Struct style::properties::declaration_block::PropertyDeclarationBlock

pub struct PropertyDeclarationBlock {
    pub declarations: Vec<(PropertyDeclaration, Importance)>,
    pub important_count: u32,
}

Overridden declarations are skipped.

Fields
declarations: Vec<(PropertyDeclaration, Importance)>

The group of declarations, along with their importance.

Only deduplicated declarations appear here.

important_count: u32

The number of entries in self.declaration with Importance::Important

Methods

impl PropertyDeclarationBlock

fn any_important(&self) -> bool

Returns wheather this block contains any declaration with !important.

This is based on the important_count counter, which should be maintained whenever declarations is changed.

fn any_normal(&self) -> bool

Returns wheather this block contains any declaration without !important.

This is based on the important_count counter, which should be maintained whenever declarations is changed.

fn get(&self,
property: PropertyDeclarationId)
-> Option<&(PropertyDeclaration, Importance)>

Get a declaration for a given property.

NOTE: This is linear time.

fn property_value_to_css<W>(&self,
property: &PropertyId,
dest: &mut W)
-> Result where W: Write

Find the value of the given property in this block and serialize it

https://dev.w3.org/csswg/cssom/#dom-cssstyledeclaration-getpropertyvalue

fn property_priority(&self, property: &PropertyId) -> Importance

https://dev.w3.org/csswg/cssom/#dom-cssstyledeclaration-getpropertypriority

fn set_parsed_declaration(&mut self,
declaration: PropertyDeclaration,
importance: Importance)

Adds or overrides the declaration for a given property in this block, without taking into account any kind of priority.

fn set_importance(&mut self,
property: &PropertyId,
new_importance: Importance)
-> bool

Set the declaration importance for a given property, if found.

Returns whether any declaration was updated.

fn remove_property(&mut self, property: &PropertyId) -> bool

https://dev.w3.org/csswg/cssom/#dom-cssstyledeclaration-removeproperty

Returns whether any declaration was actually removed.

fn single_value_to_css<W>(&self, property: &PropertyId, dest: &mut W) -> Result where W: Write

Take a declaration block known to contain a single property and serialize it.