Struct style::properties::declaration_block::PropertyDeclarationBlock

pub struct PropertyDeclarationBlock {
    declarations: ThinVec<PropertyDeclaration>,
    declarations_importance: SmallBitVec,
    property_ids: PropertyDeclarationIdSet,
}

Overridden declarations are skipped.

Fields
declarations: ThinVec<PropertyDeclaration>

The group of declarations, along with their importance.

Only deduplicated declarations appear here.

declarations_importance: SmallBitVec

The “important” flag for each declaration in declarations.

property_ids: PropertyDeclarationIdSet

The set of properties that are present in the block.

Implementations

impl PropertyDeclarationBlock

pub fn len(&self) -> usize

Returns the number of declarations in the block.

pub fn is_empty(&self) -> bool

Returns whether the block is empty.

pub fn new() -> Self

Create an empty block

pub fn with_one( declaration: PropertyDeclaration, importance: Importance, ) -> Self

Create a block with a single declaration

pub fn declarations(&self) -> &[PropertyDeclaration]

The declarations in this block

pub fn declarations_importance(&self) -> &SmallBitVec

The important flags for declarations in this block

pub fn declaration_importance_iter(&self) -> DeclarationImportanceIterator<'_>

Iterate over (PropertyDeclaration, Importance) pairs

pub fn normal_declaration_iter<'a>( &'a self, ) -> impl DoubleEndedIterator<Item = &'a PropertyDeclaration>

Iterate over PropertyDeclaration for Importance::Normal

pub fn to_animation_value_iter<'a, 'cx, 'cx_a: 'cx>( &'a self, context: &'cx mut Context<'cx_a>, style: &'a ComputedValues, default_values: &'a ComputedValues, ) -> AnimationValueIterator<'a, 'cx, 'cx_a>

Return an iterator of (AnimatableLonghand, AnimationValue).

pub fn any_important(&self) -> bool

Returns whether this block contains any declaration with !important.

This is based on the declarations_importance bit-vector, which should be maintained whenever declarations is changed.

pub fn any_normal(&self) -> bool

Returns whether this block contains any declaration without !important.

This is based on the declarations_importance bit-vector, which should be maintained whenever declarations is changed.

pub fn property_ids(&self) -> &PropertyDeclarationIdSet

Returns a PropertyDeclarationIdSet representing the properties that are changed in this block.

pub fn contains(&self, id: PropertyDeclarationId<'_>) -> bool

Returns whether this block contains a declaration of a given property id.

pub fn contains_any_reset(&self) -> bool

Returns whether this block contains any reset longhand.

pub fn get( &self, property: PropertyDeclarationId<'_>, ) -> Option<(&PropertyDeclaration, Importance)>

Get a declaration for a given property.

NOTE: This is linear time in the case of custom properties or in the case the longhand is actually in the declaration block.

pub fn shorthand_to_css( &self, shorthand: ShorthandId, dest: &mut CssStringWriter, ) -> Result

Tries to serialize a given shorthand from the declarations in this block.

pub fn property_value_to_css( &self, property: &PropertyId, dest: &mut CssStringWriter, ) -> Result

Find the value of the given property in this block and serialize it

https://drafts.csswg.org/cssom/#dom-cssstyledeclaration-getpropertyvalue

pub fn property_priority(&self, property: &PropertyId) -> Importance

https://drafts.csswg.org/cssom/#dom-cssstyledeclaration-getpropertypriority

pub fn extend( &mut self, drain: SourcePropertyDeclarationDrain<'_>, importance: Importance, ) -> bool

Adds or overrides the declaration for a given property in this block.

See the documentation of push to see what impact source has when the property is already there.

pub fn push( &mut self, declaration: PropertyDeclaration, importance: Importance, ) -> bool

Adds or overrides the declaration for a given property in this block.

Returns whether the declaration has changed.

This is only used for parsing and internal use.

pub fn prepare_for_update( &self, source_declarations: &SourcePropertyDeclaration, importance: Importance, updates: &mut SourcePropertyDeclarationUpdate, ) -> bool

Prepares updating this declaration block with the given SourcePropertyDeclaration and importance, and returns whether there is something to update.

pub fn update( &mut self, drain: SourcePropertyDeclarationDrain<'_>, importance: Importance, updates: &mut SourcePropertyDeclarationUpdate, )

Update this declaration block with the given data.

pub fn first_declaration_to_remove( &self, property: &PropertyId, ) -> Option<usize>

Returns the first declaration that would be removed by removing property.

fn remove_declaration_at(&mut self, i: usize)

Removes a given declaration at a given index.

pub fn clear(&mut self)

Clears all the declarations from this block.

pub fn remove_property( &mut self, property: &PropertyId, first_declaration: usize, )

https://drafts.csswg.org/cssom/#dom-cssstyledeclaration-removeproperty

first_declaration needs to be the result of first_declaration_to_remove.

pub fn single_value_to_css( &self, property: &PropertyId, dest: &mut CssStringWriter, computed_values: Option<&ComputedValues>, stylist: &Stylist, ) -> Result

Take a declaration block known to contain a single property and serialize it.

pub fn from_animation_value_map(animation_value_map: &AnimationValueMap) -> Self

Convert AnimationValueMap to PropertyDeclarationBlock.

pub fn has_css_wide_keyword(&self, property: &PropertyId) -> bool

Returns true if the declaration block has a CSSWideKeyword for the given property.

pub fn to_css(&self, dest: &mut CssStringWriter) -> Result

Like the method on ToCss, but without the type parameter to avoid accidentally monomorphizing this large function multiple times for different writers.

https://drafts.csswg.org/cssom/#serialize-a-css-declaration-block