Enum style::properties::LonghandId

pub enum LonghandId {
    BackgroundColor,
    BackgroundImage,
    BackgroundPositionX,
    BackgroundPositionY,
    BackgroundRepeat,
    BackgroundAttachment,
    BackgroundClip,
    BackgroundOrigin,
    BackgroundSize,
    BorderTopColor,
    BorderTopStyle,
    BorderTopWidth,
    BorderLeftColor,
    BorderLeftStyle,
    BorderLeftWidth,
    BorderBottomColor,
    BorderBottomStyle,
    BorderBottomWidth,
    BorderRightColor,
    BorderRightStyle,
    BorderRightWidth,
    BorderBlockStartColor,
    BorderBlockStartStyle,
    BorderBlockStartWidth,
    BorderBlockEndColor,
    BorderBlockEndStyle,
    BorderBlockEndWidth,
    BorderInlineStartColor,
    BorderInlineStartStyle,
    BorderInlineStartWidth,
    BorderInlineEndColor,
    BorderInlineEndStyle,
    BorderInlineEndWidth,
    BorderTopLeftRadius,
    BorderTopRightRadius,
    BorderBottomRightRadius,
    BorderBottomLeftRadius,
    BorderImageSource,
    BorderImageOutset,
    BorderImageRepeat,
    BorderImageWidth,
    BorderImageSlice,
    Display,
    Position,
    Float,
    Clear,
    ServoDisplayForHypotheticalBox,
    VerticalAlign,
    ServoOverflowClipBox,
    OverflowX,
    OverflowY,
    TransitionDuration,
    TransitionTimingFunction,
    TransitionProperty,
    TransitionDelay,
    AnimationName,
    AnimationDuration,
    AnimationTimingFunction,
    AnimationIterationCount,
    AnimationDirection,
    AnimationPlayState,
    AnimationFillMode,
    AnimationDelay,
    Transform,
    Perspective,
    PerspectiveOrigin,
    BackfaceVisibility,
    TransformStyle,
    TransformOrigin,
    Color,
    ColumnWidth,
    ColumnCount,
    ColumnGap,
    Content,
    CounterIncrement,
    CounterReset,
    Opacity,
    BoxShadow,
    Clip,
    Filter,
    MixBlendMode,
    FontFamily,
    FontStyle,
    FontVariantCaps,
    FontWeight,
    FontSize,
    FontStretch,
    Visibility,
    WritingMode,
    Direction,
    ImageRendering,
    BorderCollapse,
    EmptyCells,
    CaptionSide,
    BorderSpacing,
    LineHeight,
    TextTransform,
    TextIndent,
    OverflowWrap,
    WordBreak,
    TextJustify,
    TextAlign,
    LetterSpacing,
    WordSpacing,
    ServoTextDecorationsInEffect,
    WhiteSpace,
    TextShadow,
    TextRendering,
    ListStylePosition,
    ListStyleType,
    ListStyleImage,
    Quotes,
    MarginTop,
    MarginLeft,
    MarginBottom,
    MarginRight,
    MarginBlockStart,
    MarginBlockEnd,
    MarginInlineStart,
    MarginInlineEnd,
    OutlineColor,
    OutlineStyle,
    OutlineWidth,
    OutlineOffset,
    PaddingTop,
    PaddingLeft,
    PaddingBottom,
    PaddingRight,
    PaddingBlockStart,
    PaddingBlockEnd,
    PaddingInlineStart,
    PaddingInlineEnd,
    Cursor,
    PointerEvents,
    Top,
    Left,
    Bottom,
    Right,
    OffsetBlockStart,
    OffsetBlockEnd,
    OffsetInlineStart,
    OffsetInlineEnd,
    ZIndex,
    FlexDirection,
    FlexWrap,
    JustifyContent,
    AlignContent,
    AlignItems,
    FlexGrow,
    FlexShrink,
    AlignSelf,
    Order,
    FlexBasis,
    Width,
    MinWidth,
    MaxWidth,
    Height,
    MinHeight,
    MaxHeight,
    BlockSize,
    MinBlockSize,
    MaxBlockSize,
    InlineSize,
    MinInlineSize,
    MaxInlineSize,
    BoxSizing,
    TableLayout,
    TextOverflow,
    UnicodeBidi,
    TextDecorationLine,
}

An identifier for a given longhand property.

Variants
BackgroundColor

background-color

BackgroundImage

background-image

BackgroundPositionX

background-position-x

BackgroundPositionY

background-position-y

BackgroundRepeat

background-repeat

BackgroundAttachment

background-attachment

BackgroundClip

background-clip

BackgroundOrigin

background-origin

BackgroundSize

background-size

BorderTopColor

border-top-color

BorderTopStyle

border-top-style

BorderTopWidth

border-top-width

BorderLeftColor

border-left-color

BorderLeftStyle

border-left-style

BorderLeftWidth

border-left-width

BorderBottomColor

border-bottom-color

BorderBottomStyle

border-bottom-style

BorderBottomWidth

border-bottom-width

BorderRightColor

border-right-color

BorderRightStyle

border-right-style

BorderRightWidth

border-right-width

BorderBlockStartColor

border-block-start-color

BorderBlockStartStyle

border-block-start-style

BorderBlockStartWidth

border-block-start-width

BorderBlockEndColor

border-block-end-color

BorderBlockEndStyle

border-block-end-style

BorderBlockEndWidth

border-block-end-width

BorderInlineStartColor

border-inline-start-color

BorderInlineStartStyle

border-inline-start-style

BorderInlineStartWidth

border-inline-start-width

BorderInlineEndColor

border-inline-end-color

BorderInlineEndStyle

border-inline-end-style

BorderInlineEndWidth

border-inline-end-width

BorderTopLeftRadius

border-top-left-radius

BorderTopRightRadius

border-top-right-radius

BorderBottomRightRadius

border-bottom-right-radius

BorderBottomLeftRadius

border-bottom-left-radius

BorderImageSource

border-image-source

BorderImageOutset

border-image-outset

BorderImageRepeat

border-image-repeat

BorderImageWidth

border-image-width

BorderImageSlice

border-image-slice

Display

display

Position

position

Float

float

Clear

clear

ServoDisplayForHypotheticalBox

-servo-display-for-hypothetical-box

VerticalAlign

vertical-align

ServoOverflowClipBox

-servo-overflow-clip-box

OverflowX

overflow-x

OverflowY

overflow-y

TransitionDuration

transition-duration

TransitionTimingFunction

transition-timing-function

TransitionProperty

transition-property

TransitionDelay

transition-delay

AnimationName

animation-name

AnimationDuration

animation-duration

AnimationTimingFunction

animation-timing-function

AnimationIterationCount

animation-iteration-count

AnimationDirection

animation-direction

AnimationPlayState

animation-play-state

AnimationFillMode

animation-fill-mode

AnimationDelay

animation-delay

Transform

transform

Perspective

perspective

PerspectiveOrigin

perspective-origin

BackfaceVisibility

backface-visibility

TransformStyle

transform-style

TransformOrigin

transform-origin

Color

color

ColumnWidth

column-width

ColumnCount

column-count

ColumnGap

column-gap

Content

content

CounterIncrement

counter-increment

CounterReset

counter-reset

Opacity

opacity

BoxShadow

box-shadow

Clip

clip

Filter

filter

MixBlendMode

mix-blend-mode

FontFamily

font-family

FontStyle

font-style

FontVariantCaps

font-variant-caps

FontWeight

font-weight

FontSize

font-size

FontStretch

font-stretch

Visibility

visibility

WritingMode

writing-mode

Direction

direction

ImageRendering

image-rendering

BorderCollapse

border-collapse

EmptyCells

empty-cells

CaptionSide

caption-side

BorderSpacing

border-spacing

LineHeight

line-height

TextTransform

text-transform

TextIndent

text-indent

OverflowWrap

overflow-wrap

WordBreak

word-break

TextJustify

text-justify

TextAlign

text-align

LetterSpacing

letter-spacing

WordSpacing

word-spacing

ServoTextDecorationsInEffect

-servo-text-decorations-in-effect

WhiteSpace

white-space

TextShadow

text-shadow

TextRendering

text-rendering

ListStylePosition

list-style-position

ListStyleType

list-style-type

ListStyleImage

list-style-image

Quotes

quotes

MarginTop

margin-top

MarginLeft

margin-left

MarginBottom

margin-bottom

MarginRight

margin-right

MarginBlockStart

margin-block-start

MarginBlockEnd

margin-block-end

MarginInlineStart

margin-inline-start

MarginInlineEnd

margin-inline-end

OutlineColor

outline-color

OutlineStyle

outline-style

OutlineWidth

outline-width

OutlineOffset

outline-offset

PaddingTop

padding-top

PaddingLeft

padding-left

PaddingBottom

padding-bottom

PaddingRight

padding-right

PaddingBlockStart

padding-block-start

PaddingBlockEnd

padding-block-end

PaddingInlineStart

padding-inline-start

PaddingInlineEnd

padding-inline-end

Cursor

cursor

PointerEvents

pointer-events

Top

top

Left

left

Bottom

bottom

Right

right

OffsetBlockStart

offset-block-start

OffsetBlockEnd

offset-block-end

OffsetInlineStart

offset-inline-start

OffsetInlineEnd

offset-inline-end

ZIndex

z-index

FlexDirection

flex-direction

FlexWrap

flex-wrap

JustifyContent

justify-content

AlignContent

align-content

AlignItems

align-items

FlexGrow

flex-grow

FlexShrink

flex-shrink

AlignSelf

align-self

Order

order

FlexBasis

flex-basis

Width

width

MinWidth

min-width

MaxWidth

max-width

Height

height

MinHeight

min-height

MaxHeight

max-height

BlockSize

block-size

MinBlockSize

min-block-size

MaxBlockSize

max-block-size

InlineSize

inline-size

MinInlineSize

min-inline-size

MaxInlineSize

max-inline-size

BoxSizing

box-sizing

TableLayout

table-layout

TextOverflow

text-overflow

UnicodeBidi

unicode-bidi

TextDecorationLine

text-decoration-line

Methods

impl LonghandId

pub fn name(&self) -> &'static str

Get the name of this longhand property.

pub fn is_animatable(self) -> bool

Returns whether this property is animatable.

pub fn is_discrete_animatable(self) -> bool

Returns whether this property is animatable in a discrete way.

pub fn to_physical(&self, wm: WritingMode) -> Self

If this is a logical property, return the corresponding physical one in the given writing mode. Otherwise, return unchanged.

pub fn flags(&self) -> PropertyFlags

Returns PropertyFlags for given longhand property.

pub fn stores_complex_colors_lossily(&self) -> bool

Whether computed values of this property lossily convert any complex colors into RGBA colors.

In Gecko, there are some properties still that compute currentcolor down to an RGBA color at computed value time, instead of as StyleComplexColors. For these properties, we must return false, so that we correctly avoid caching style data in the rule tree.