Enum style::properties::PropertyDeclaration

pub enum PropertyDeclaration {
    BackgroundColor(DeclaredValue<SpecifiedValue>),
    BackgroundImage(DeclaredValue<SpecifiedValue>),
    BackgroundPosition(DeclaredValue<SpecifiedValue>),
    BackgroundRepeat(DeclaredValue<SpecifiedValue>),
    BackgroundAttachment(DeclaredValue<SpecifiedValue>),
    BackgroundClip(DeclaredValue<SpecifiedValue>),
    BackgroundOrigin(DeclaredValue<SpecifiedValue>),
    BackgroundSize(DeclaredValue<SpecifiedValue>),
    BorderTopColor(DeclaredValue<SpecifiedValue>),
    BorderRightColor(DeclaredValue<SpecifiedValue>),
    BorderBottomColor(DeclaredValue<SpecifiedValue>),
    BorderLeftColor(DeclaredValue<SpecifiedValue>),
    BorderTopStyle(DeclaredValue<SpecifiedValue>),
    BorderRightStyle(DeclaredValue<SpecifiedValue>),
    BorderBottomStyle(DeclaredValue<SpecifiedValue>),
    BorderLeftStyle(DeclaredValue<SpecifiedValue>),
    BorderTopWidth(DeclaredValue<SpecifiedValue>),
    BorderRightWidth(DeclaredValue<SpecifiedValue>),
    BorderBottomWidth(DeclaredValue<SpecifiedValue>),
    BorderLeftWidth(DeclaredValue<SpecifiedValue>),
    BorderTopLeftRadius(DeclaredValue<SpecifiedValue>),
    BorderTopRightRadius(DeclaredValue<SpecifiedValue>),
    BorderBottomRightRadius(DeclaredValue<SpecifiedValue>),
    BorderBottomLeftRadius(DeclaredValue<SpecifiedValue>),
    Display(DeclaredValue<SpecifiedValue>),
    Position(DeclaredValue<SpecifiedValue>),
    Float(DeclaredValue<SpecifiedValue>),
    Clear(DeclaredValue<SpecifiedValue>),
    ServoDisplayForHypotheticalBox(DeclaredValue<SpecifiedValue>),
    VerticalAlign(DeclaredValue<SpecifiedValue>),
    ServoOverflowClipBox(DeclaredValue<SpecifiedValue>),
    OverflowX(DeclaredValue<SpecifiedValue>),
    OverflowY(DeclaredValue<SpecifiedValue>),
    TransitionDuration(DeclaredValue<SpecifiedValue>),
    TransitionTimingFunction(DeclaredValue<SpecifiedValue>),
    TransitionProperty(DeclaredValue<SpecifiedValue>),
    TransitionDelay(DeclaredValue<SpecifiedValue>),
    Color(DeclaredValue<SpecifiedValue>),
    ColumnWidth(DeclaredValue<SpecifiedValue>),
    ColumnCount(DeclaredValue<SpecifiedValue>),
    ColumnGap(DeclaredValue<SpecifiedValue>),
    Content(DeclaredValue<SpecifiedValue>),
    CounterIncrement(DeclaredValue<SpecifiedValue>),
    CounterReset(DeclaredValue<SpecifiedValue>),
    Opacity(DeclaredValue<SpecifiedValue>),
    BoxShadow(DeclaredValue<SpecifiedValue>),
    Clip(DeclaredValue<SpecifiedValue>),
    Filter(DeclaredValue<SpecifiedValue>),
    Transform(DeclaredValue<SpecifiedValue>),
    BackfaceVisibility(DeclaredValue<SpecifiedValue>),
    TransformStyle(DeclaredValue<SpecifiedValue>),
    TransformOrigin(DeclaredValue<SpecifiedValue>),
    Perspective(DeclaredValue<SpecifiedValue>),
    PerspectiveOrigin(DeclaredValue<SpecifiedValue>),
    MixBlendMode(DeclaredValue<SpecifiedValue>),
    FontFamily(DeclaredValue<SpecifiedValue>),
    FontStyle(DeclaredValue<SpecifiedValue>),
    FontVariant(DeclaredValue<SpecifiedValue>),
    FontWeight(DeclaredValue<SpecifiedValue>),
    FontSize(DeclaredValue<SpecifiedValue>),
    FontStretch(DeclaredValue<SpecifiedValue>),
    Direction(DeclaredValue<SpecifiedValue>),
    Visibility(DeclaredValue<SpecifiedValue>),
    WritingMode(DeclaredValue<SpecifiedValue>),
    TextOrientation(DeclaredValue<SpecifiedValue>),
    ImageRendering(DeclaredValue<SpecifiedValue>),
    ServoUnderDisplayNone(DeclaredValue<SpecifiedValue>),
    BorderCollapse(DeclaredValue<SpecifiedValue>),
    EmptyCells(DeclaredValue<SpecifiedValue>),
    CaptionSide(DeclaredValue<SpecifiedValue>),
    BorderSpacing(DeclaredValue<SpecifiedValue>),
    LineHeight(DeclaredValue<SpecifiedValue>),
    TextAlign(DeclaredValue<SpecifiedValue>),
    LetterSpacing(DeclaredValue<SpecifiedValue>),
    WordSpacing(DeclaredValue<SpecifiedValue>),
    TextIndent(DeclaredValue<SpecifiedValue>),
    OverflowWrap(DeclaredValue<SpecifiedValue>),
    WordBreak(DeclaredValue<SpecifiedValue>),
    TextJustify(DeclaredValue<SpecifiedValue>),
    ServoTextDecorationsInEffect(DeclaredValue<SpecifiedValue>),
    WhiteSpace(DeclaredValue<SpecifiedValue>),
    TextShadow(DeclaredValue<SpecifiedValue>),
    TextTransform(DeclaredValue<SpecifiedValue>),
    TextRendering(DeclaredValue<SpecifiedValue>),
    ListStylePosition(DeclaredValue<SpecifiedValue>),
    ListStyleType(DeclaredValue<SpecifiedValue>),
    ListStyleImage(DeclaredValue<SpecifiedValue>),
    Quotes(DeclaredValue<SpecifiedValue>),
    MarginTop(DeclaredValue<SpecifiedValue>),
    MarginRight(DeclaredValue<SpecifiedValue>),
    MarginBottom(DeclaredValue<SpecifiedValue>),
    MarginLeft(DeclaredValue<SpecifiedValue>),
    OutlineColor(DeclaredValue<SpecifiedValue>),
    OutlineStyle(DeclaredValue<SpecifiedValue>),
    OutlineWidth(DeclaredValue<SpecifiedValue>),
    OutlineOffset(DeclaredValue<SpecifiedValue>),
    PaddingTop(DeclaredValue<SpecifiedValue>),
    PaddingRight(DeclaredValue<SpecifiedValue>),
    PaddingBottom(DeclaredValue<SpecifiedValue>),
    PaddingLeft(DeclaredValue<SpecifiedValue>),
    Cursor(DeclaredValue<SpecifiedValue>),
    PointerEvents(DeclaredValue<SpecifiedValue>),
    Top(DeclaredValue<SpecifiedValue>),
    Right(DeclaredValue<SpecifiedValue>),
    Bottom(DeclaredValue<SpecifiedValue>),
    Left(DeclaredValue<SpecifiedValue>),
    ZIndex(DeclaredValue<SpecifiedValue>),
    FlexDirection(DeclaredValue<SpecifiedValue>),
    FlexWrap(DeclaredValue<SpecifiedValue>),
    JustifyContent(DeclaredValue<SpecifiedValue>),
    AlignItems(DeclaredValue<SpecifiedValue>),
    AlignContent(DeclaredValue<SpecifiedValue>),
    FlexGrow(DeclaredValue<SpecifiedValue>),
    FlexShrink(DeclaredValue<SpecifiedValue>),
    AlignSelf(DeclaredValue<SpecifiedValue>),
    Order(DeclaredValue<SpecifiedValue>),
    FlexBasis(DeclaredValue<SpecifiedValue>),
    Width(DeclaredValue<SpecifiedValue>),
    Height(DeclaredValue<SpecifiedValue>),
    MinWidth(DeclaredValue<SpecifiedValue>),
    MaxWidth(DeclaredValue<SpecifiedValue>),
    MinHeight(DeclaredValue<SpecifiedValue>),
    MaxHeight(DeclaredValue<SpecifiedValue>),
    BoxSizing(DeclaredValue<SpecifiedValue>),
    TableLayout(DeclaredValue<SpecifiedValue>),
    TextOverflow(DeclaredValue<SpecifiedValue>),
    UnicodeBidi(DeclaredValue<SpecifiedValue>),
    TextDecoration(DeclaredValue<SpecifiedValue>),
    Custom(Name, DeclaredValue<SpecifiedValue>),
}

Variants
BackgroundColor(DeclaredValue<SpecifiedValue>)BackgroundImage(DeclaredValue<SpecifiedValue>)BackgroundPosition(DeclaredValue<SpecifiedValue>)BackgroundRepeat(DeclaredValue<SpecifiedValue>)BackgroundAttachment(DeclaredValue<SpecifiedValue>)BackgroundClip(DeclaredValue<SpecifiedValue>)BackgroundOrigin(DeclaredValue<SpecifiedValue>)BackgroundSize(DeclaredValue<SpecifiedValue>)BorderTopColor(DeclaredValue<SpecifiedValue>)BorderRightColor(DeclaredValue<SpecifiedValue>)BorderBottomColor(DeclaredValue<SpecifiedValue>)BorderLeftColor(DeclaredValue<SpecifiedValue>)BorderTopStyle(DeclaredValue<SpecifiedValue>)BorderRightStyle(DeclaredValue<SpecifiedValue>)BorderBottomStyle(DeclaredValue<SpecifiedValue>)BorderLeftStyle(DeclaredValue<SpecifiedValue>)BorderTopWidth(DeclaredValue<SpecifiedValue>)BorderRightWidth(DeclaredValue<SpecifiedValue>)BorderBottomWidth(DeclaredValue<SpecifiedValue>)BorderLeftWidth(DeclaredValue<SpecifiedValue>)BorderTopLeftRadius(DeclaredValue<SpecifiedValue>)BorderTopRightRadius(DeclaredValue<SpecifiedValue>)BorderBottomRightRadius(DeclaredValue<SpecifiedValue>)BorderBottomLeftRadius(DeclaredValue<SpecifiedValue>)Display(DeclaredValue<SpecifiedValue>)Position(DeclaredValue<SpecifiedValue>)Float(DeclaredValue<SpecifiedValue>)Clear(DeclaredValue<SpecifiedValue>)ServoDisplayForHypotheticalBox(DeclaredValue<SpecifiedValue>)VerticalAlign(DeclaredValue<SpecifiedValue>)ServoOverflowClipBox(DeclaredValue<SpecifiedValue>)OverflowX(DeclaredValue<SpecifiedValue>)OverflowY(DeclaredValue<SpecifiedValue>)TransitionDuration(DeclaredValue<SpecifiedValue>)TransitionTimingFunction(DeclaredValue<SpecifiedValue>)TransitionProperty(DeclaredValue<SpecifiedValue>)TransitionDelay(DeclaredValue<SpecifiedValue>)Color(DeclaredValue<SpecifiedValue>)ColumnWidth(DeclaredValue<SpecifiedValue>)ColumnCount(DeclaredValue<SpecifiedValue>)ColumnGap(DeclaredValue<SpecifiedValue>)Content(DeclaredValue<SpecifiedValue>)CounterIncrement(DeclaredValue<SpecifiedValue>)CounterReset(DeclaredValue<SpecifiedValue>)Opacity(DeclaredValue<SpecifiedValue>)BoxShadow(DeclaredValue<SpecifiedValue>)Clip(DeclaredValue<SpecifiedValue>)Filter(DeclaredValue<SpecifiedValue>)Transform(DeclaredValue<SpecifiedValue>)BackfaceVisibility(DeclaredValue<SpecifiedValue>)TransformStyle(DeclaredValue<SpecifiedValue>)TransformOrigin(DeclaredValue<SpecifiedValue>)Perspective(DeclaredValue<SpecifiedValue>)PerspectiveOrigin(DeclaredValue<SpecifiedValue>)MixBlendMode(DeclaredValue<SpecifiedValue>)FontFamily(DeclaredValue<SpecifiedValue>)FontStyle(DeclaredValue<SpecifiedValue>)FontVariant(DeclaredValue<SpecifiedValue>)FontWeight(DeclaredValue<SpecifiedValue>)FontSize(DeclaredValue<SpecifiedValue>)FontStretch(DeclaredValue<SpecifiedValue>)Direction(DeclaredValue<SpecifiedValue>)Visibility(DeclaredValue<SpecifiedValue>)WritingMode(DeclaredValue<SpecifiedValue>)TextOrientation(DeclaredValue<SpecifiedValue>)ImageRendering(DeclaredValue<SpecifiedValue>)ServoUnderDisplayNone(DeclaredValue<SpecifiedValue>)BorderCollapse(DeclaredValue<SpecifiedValue>)EmptyCells(DeclaredValue<SpecifiedValue>)CaptionSide(DeclaredValue<SpecifiedValue>)BorderSpacing(DeclaredValue<SpecifiedValue>)LineHeight(DeclaredValue<SpecifiedValue>)TextAlign(DeclaredValue<SpecifiedValue>)LetterSpacing(DeclaredValue<SpecifiedValue>)WordSpacing(DeclaredValue<SpecifiedValue>)TextIndent(DeclaredValue<SpecifiedValue>)OverflowWrap(DeclaredValue<SpecifiedValue>)WordBreak(DeclaredValue<SpecifiedValue>)TextJustify(DeclaredValue<SpecifiedValue>)ServoTextDecorationsInEffect(DeclaredValue<SpecifiedValue>)WhiteSpace(DeclaredValue<SpecifiedValue>)TextShadow(DeclaredValue<SpecifiedValue>)TextTransform(DeclaredValue<SpecifiedValue>)TextRendering(DeclaredValue<SpecifiedValue>)ListStylePosition(DeclaredValue<SpecifiedValue>)ListStyleType(DeclaredValue<SpecifiedValue>)ListStyleImage(DeclaredValue<SpecifiedValue>)Quotes(DeclaredValue<SpecifiedValue>)MarginTop(DeclaredValue<SpecifiedValue>)MarginRight(DeclaredValue<SpecifiedValue>)MarginBottom(DeclaredValue<SpecifiedValue>)MarginLeft(DeclaredValue<SpecifiedValue>)OutlineColor(DeclaredValue<SpecifiedValue>)OutlineStyle(DeclaredValue<SpecifiedValue>)OutlineWidth(DeclaredValue<SpecifiedValue>)OutlineOffset(DeclaredValue<SpecifiedValue>)PaddingTop(DeclaredValue<SpecifiedValue>)PaddingRight(DeclaredValue<SpecifiedValue>)PaddingBottom(DeclaredValue<SpecifiedValue>)PaddingLeft(DeclaredValue<SpecifiedValue>)Cursor(DeclaredValue<SpecifiedValue>)PointerEvents(DeclaredValue<SpecifiedValue>)Top(DeclaredValue<SpecifiedValue>)Right(DeclaredValue<SpecifiedValue>)Bottom(DeclaredValue<SpecifiedValue>)Left(DeclaredValue<SpecifiedValue>)ZIndex(DeclaredValue<SpecifiedValue>)FlexDirection(DeclaredValue<SpecifiedValue>)FlexWrap(DeclaredValue<SpecifiedValue>)JustifyContent(DeclaredValue<SpecifiedValue>)AlignItems(DeclaredValue<SpecifiedValue>)AlignContent(DeclaredValue<SpecifiedValue>)FlexGrow(DeclaredValue<SpecifiedValue>)FlexShrink(DeclaredValue<SpecifiedValue>)AlignSelf(DeclaredValue<SpecifiedValue>)Order(DeclaredValue<SpecifiedValue>)FlexBasis(DeclaredValue<SpecifiedValue>)Width(DeclaredValue<SpecifiedValue>)Height(DeclaredValue<SpecifiedValue>)MinWidth(DeclaredValue<SpecifiedValue>)MaxWidth(DeclaredValue<SpecifiedValue>)MinHeight(DeclaredValue<SpecifiedValue>)MaxHeight(DeclaredValue<SpecifiedValue>)BoxSizing(DeclaredValue<SpecifiedValue>)TableLayout(DeclaredValue<SpecifiedValue>)TextOverflow(DeclaredValue<SpecifiedValue>)UnicodeBidi(DeclaredValue<SpecifiedValue>)TextDecoration(DeclaredValue<SpecifiedValue>)Custom(Name, DeclaredValue<SpecifiedValue>)
Methods

impl PropertyDeclaration

fn name(&self) -> PropertyDeclarationName

fn value(&self) -> String

fn with_variables_from_shorthand(&self, shorthand: Shorthand) -> Option<&str>

If this is a pending-substitution value from the given shorthand, return that value

fn with_variables(&self) -> bool

Return whether this is a pending-substitution value. https://drafts.csswg.org/css-variables/#variables-in-shorthands

fn value_is_unparsed(&self) -> bool

Return whether the value is stored as it was in the CSS source, preserving whitespace (as opposed to being parsed into a more abstract data structure). This is the case of custom properties and values that contain unsubstituted variables.

fn matches(&self, name: &str) -> bool

fn parse(name: &str, context: &ParserContext, input: &mut Parser, result_list: &mut Vec<PropertyDeclaration>) -> PropertyDeclarationParseResult

fn shorthands(&self) -> &'static [Shorthand]