Enum style::properties::PropertyDeclaration

pub enum PropertyDeclaration {
    BackgroundColor(DeclaredValue<Box<SpecifiedValue>>),
    BackgroundImage(DeclaredValue<SpecifiedValue>),
    BackgroundPositionX(DeclaredValue<SpecifiedValue>),
    BackgroundPositionY(DeclaredValue<SpecifiedValue>),
    BackgroundRepeat(DeclaredValue<SpecifiedValue>),
    BackgroundAttachment(DeclaredValue<SpecifiedValue>),
    BackgroundClip(DeclaredValue<SpecifiedValue>),
    BackgroundOrigin(DeclaredValue<SpecifiedValue>),
    BackgroundSize(DeclaredValue<SpecifiedValue>),
    BackgroundBlendMode(DeclaredValue<SpecifiedValue>),
    BorderTopColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderLeftColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderBottomColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderRightColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderBlockStartColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderBlockEndColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderInlineStartColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderInlineEndColor(DeclaredValue<Box<SpecifiedValue>>),
    BorderTopStyle(DeclaredValue<SpecifiedValue>),
    BorderLeftStyle(DeclaredValue<SpecifiedValue>),
    BorderBottomStyle(DeclaredValue<SpecifiedValue>),
    BorderRightStyle(DeclaredValue<SpecifiedValue>),
    BorderBlockStartStyle(DeclaredValue<SpecifiedValue>),
    BorderBlockEndStyle(DeclaredValue<SpecifiedValue>),
    BorderInlineStartStyle(DeclaredValue<SpecifiedValue>),
    BorderInlineEndStyle(DeclaredValue<SpecifiedValue>),
    BorderTopWidth(DeclaredValue<SpecifiedValue>),
    BorderLeftWidth(DeclaredValue<SpecifiedValue>),
    BorderBottomWidth(DeclaredValue<SpecifiedValue>),
    BorderRightWidth(DeclaredValue<SpecifiedValue>),
    BorderBlockStartWidth(DeclaredValue<SpecifiedValue>),
    BorderBlockEndWidth(DeclaredValue<SpecifiedValue>),
    BorderInlineStartWidth(DeclaredValue<SpecifiedValue>),
    BorderInlineEndWidth(DeclaredValue<SpecifiedValue>),
    BorderTopLeftRadius(DeclaredValue<SpecifiedValue>),
    BorderTopRightRadius(DeclaredValue<SpecifiedValue>),
    BorderBottomRightRadius(DeclaredValue<SpecifiedValue>),
    BorderBottomLeftRadius(DeclaredValue<SpecifiedValue>),
    BoxDecorationBreak(DeclaredValue<SpecifiedValue>),
    MozFloatEdge(DeclaredValue<SpecifiedValue>),
    BorderImageSource(DeclaredValue<Box<SpecifiedValue>>),
    BorderImageOutset(DeclaredValue<SpecifiedValue>),
    BorderImageRepeat(DeclaredValue<SpecifiedValue>),
    BorderImageWidth(DeclaredValue<SpecifiedValue>),
    BorderImageSlice(DeclaredValue<SpecifiedValue>),
    Display(DeclaredValue<SpecifiedValue>),
    MozTopLayer(DeclaredValue<SpecifiedValue>),
    Position(DeclaredValue<SpecifiedValue>),
    Float(DeclaredValue<SpecifiedValue>),
    Clear(DeclaredValue<SpecifiedValue>),
    VerticalAlign(DeclaredValue<SpecifiedValue>),
    OverflowClipBox(DeclaredValue<SpecifiedValue>),
    OverflowX(DeclaredValue<SpecifiedValue>),
    OverflowY(DeclaredValue<SpecifiedValue>),
    TransitionDuration(DeclaredValue<SpecifiedValue>),
    TransitionTimingFunction(DeclaredValue<SpecifiedValue>),
    TransitionProperty(DeclaredValue<SpecifiedValue>),
    TransitionDelay(DeclaredValue<SpecifiedValue>),
    AnimationName(DeclaredValue<SpecifiedValue>),
    AnimationDuration(DeclaredValue<SpecifiedValue>),
    AnimationTimingFunction(DeclaredValue<SpecifiedValue>),
    AnimationIterationCount(DeclaredValue<SpecifiedValue>),
    AnimationDirection(DeclaredValue<SpecifiedValue>),
    AnimationPlayState(DeclaredValue<SpecifiedValue>),
    AnimationFillMode(DeclaredValue<SpecifiedValue>),
    AnimationDelay(DeclaredValue<SpecifiedValue>),
    ScrollSnapPointsY(DeclaredValue<SpecifiedValue>),
    ScrollSnapPointsX(DeclaredValue<SpecifiedValue>),
    ScrollSnapDestination(DeclaredValue<Box<SpecifiedValue>>),
    ScrollSnapCoordinate(DeclaredValue<SpecifiedValue>),
    Transform(DeclaredValue<SpecifiedValue>),
    ScrollBehavior(DeclaredValue<SpecifiedValue>),
    ScrollSnapTypeX(DeclaredValue<SpecifiedValue>),
    ScrollSnapTypeY(DeclaredValue<SpecifiedValue>),
    Isolation(DeclaredValue<SpecifiedValue>),
    PageBreakAfter(DeclaredValue<SpecifiedValue>),
    PageBreakBefore(DeclaredValue<SpecifiedValue>),
    PageBreakInside(DeclaredValue<SpecifiedValue>),
    Resize(DeclaredValue<SpecifiedValue>),
    Perspective(DeclaredValue<SpecifiedValue>),
    PerspectiveOrigin(DeclaredValue<SpecifiedValue>),
    BackfaceVisibility(DeclaredValue<SpecifiedValue>),
    TransformBox(DeclaredValue<SpecifiedValue>),
    TransformStyle(DeclaredValue<SpecifiedValue>),
    TransformOrigin(DeclaredValue<Box<SpecifiedValue>>),
    MozAppearance(DeclaredValue<SpecifiedValue>),
    MozBinding(DeclaredValue<Box<SpecifiedValue>>),
    MozOrient(DeclaredValue<SpecifiedValue>),
    Color(DeclaredValue<Box<SpecifiedValue>>),
    ColumnWidth(DeclaredValue<SpecifiedValue>),
    ColumnCount(DeclaredValue<SpecifiedValue>),
    ColumnGap(DeclaredValue<SpecifiedValue>),
    ColumnFill(DeclaredValue<SpecifiedValue>),
    ColumnRuleWidth(DeclaredValue<SpecifiedValue>),
    ColumnRuleColor(DeclaredValue<Box<SpecifiedValue>>),
    ColumnRuleStyle(DeclaredValue<SpecifiedValue>),
    Content(DeclaredValue<SpecifiedValue>),
    Opacity(DeclaredValue<SpecifiedValue>),
    BoxShadow(DeclaredValue<SpecifiedValue>),
    Filter(DeclaredValue<SpecifiedValue>),
    MixBlendMode(DeclaredValue<SpecifiedValue>),
    FontFamily(DeclaredValue<SpecifiedValue>),
    FontStyle(DeclaredValue<SpecifiedValue>),
    FontVariant(DeclaredValue<SpecifiedValue>),
    FontVariantCaps(DeclaredValue<SpecifiedValue>),
    FontWeight(DeclaredValue<SpecifiedValue>),
    FontSize(DeclaredValue<SpecifiedValue>),
    FontSizeAdjust(DeclaredValue<SpecifiedValue>),
    FontSynthesis(DeclaredValue<SpecifiedValue>),
    FontStretch(DeclaredValue<SpecifiedValue>),
    FontKerning(DeclaredValue<SpecifiedValue>),
    FontVariantPosition(DeclaredValue<SpecifiedValue>),
    Visibility(DeclaredValue<SpecifiedValue>),
    WritingMode(DeclaredValue<SpecifiedValue>),
    Direction(DeclaredValue<SpecifiedValue>),
    TextOrientation(DeclaredValue<SpecifiedValue>),
    ColorAdjust(DeclaredValue<SpecifiedValue>),
    ImageRendering(DeclaredValue<SpecifiedValue>),
    BorderCollapse(DeclaredValue<SpecifiedValue>),
    EmptyCells(DeclaredValue<SpecifiedValue>),
    CaptionSide(DeclaredValue<SpecifiedValue>),
    BorderSpacing(DeclaredValue<Box<SpecifiedValue>>),
    LineHeight(DeclaredValue<SpecifiedValue>),
    TextTransform(DeclaredValue<SpecifiedValue>),
    Hyphens(DeclaredValue<SpecifiedValue>),
    TextIndent(DeclaredValue<SpecifiedValue>),
    OverflowWrap(DeclaredValue<SpecifiedValue>),
    WordBreak(DeclaredValue<SpecifiedValue>),
    TextAlignLast(DeclaredValue<SpecifiedValue>),
    TextAlign(DeclaredValue<SpecifiedValue>),
    LetterSpacing(DeclaredValue<SpecifiedValue>),
    WordSpacing(DeclaredValue<SpecifiedValue>),
    WhiteSpace(DeclaredValue<SpecifiedValue>),
    TextShadow(DeclaredValue<SpecifiedValue>),
    TextEmphasisStyle(DeclaredValue<SpecifiedValue>),
    TextEmphasisPosition(DeclaredValue<SpecifiedValue>),
    TextEmphasisColor(DeclaredValue<Box<SpecifiedValue>>),
    MozTabSize(DeclaredValue<SpecifiedValue>),
    WebkitTextFillColor(DeclaredValue<Box<SpecifiedValue>>),
    WebkitTextStrokeColor(DeclaredValue<Box<SpecifiedValue>>),
    WebkitTextStrokeWidth(DeclaredValue<SpecifiedValue>),
    RubyAlign(DeclaredValue<SpecifiedValue>),
    RubyPosition(DeclaredValue<SpecifiedValue>),
    TextCombineUpright(DeclaredValue<SpecifiedValue>),
    TextRendering(DeclaredValue<SpecifiedValue>),
    ListStylePosition(DeclaredValue<SpecifiedValue>),
    ListStyleType(DeclaredValue<SpecifiedValue>),
    ListStyleImage(DeclaredValue<SpecifiedValue>),
    Quotes(DeclaredValue<SpecifiedValue>),
    MozImageRegion(DeclaredValue<Box<SpecifiedValue>>),
    MarginTop(DeclaredValue<SpecifiedValue>),
    MarginLeft(DeclaredValue<SpecifiedValue>),
    MarginBottom(DeclaredValue<SpecifiedValue>),
    MarginRight(DeclaredValue<SpecifiedValue>),
    MarginBlockStart(DeclaredValue<SpecifiedValue>),
    MarginBlockEnd(DeclaredValue<SpecifiedValue>),
    MarginInlineStart(DeclaredValue<SpecifiedValue>),
    MarginInlineEnd(DeclaredValue<SpecifiedValue>),
    OutlineColor(DeclaredValue<Box<SpecifiedValue>>),
    OutlineStyle(DeclaredValue<SpecifiedValue>),
    OutlineWidth(DeclaredValue<SpecifiedValue>),
    MozOutlineRadiusTopleft(DeclaredValue<SpecifiedValue>),
    MozOutlineRadiusTopright(DeclaredValue<SpecifiedValue>),
    MozOutlineRadiusBottomright(DeclaredValue<SpecifiedValue>),
    MozOutlineRadiusBottomleft(DeclaredValue<SpecifiedValue>),
    OutlineOffset(DeclaredValue<SpecifiedValue>),
    PaddingTop(DeclaredValue<SpecifiedValue>),
    PaddingLeft(DeclaredValue<SpecifiedValue>),
    PaddingBottom(DeclaredValue<SpecifiedValue>),
    PaddingRight(DeclaredValue<SpecifiedValue>),
    PaddingBlockStart(DeclaredValue<SpecifiedValue>),
    PaddingBlockEnd(DeclaredValue<SpecifiedValue>),
    PaddingInlineStart(DeclaredValue<SpecifiedValue>),
    PaddingInlineEnd(DeclaredValue<SpecifiedValue>),
    Cursor(DeclaredValue<SpecifiedValue>),
    PointerEvents(DeclaredValue<SpecifiedValue>),
    MozUserInput(DeclaredValue<SpecifiedValue>),
    MozUserModify(DeclaredValue<SpecifiedValue>),
    MozUserFocus(DeclaredValue<SpecifiedValue>),
    Top(DeclaredValue<SpecifiedValue>),
    Left(DeclaredValue<SpecifiedValue>),
    Bottom(DeclaredValue<SpecifiedValue>),
    Right(DeclaredValue<SpecifiedValue>),
    OffsetBlockStart(DeclaredValue<SpecifiedValue>),
    OffsetBlockEnd(DeclaredValue<SpecifiedValue>),
    OffsetInlineStart(DeclaredValue<SpecifiedValue>),
    OffsetInlineEnd(DeclaredValue<SpecifiedValue>),
    ZIndex(DeclaredValue<SpecifiedValue>),
    FlexDirection(DeclaredValue<SpecifiedValue>),
    FlexWrap(DeclaredValue<SpecifiedValue>),
    AlignItems(DeclaredValue<SpecifiedValue>),
    FlexGrow(DeclaredValue<SpecifiedValue>),
    FlexShrink(DeclaredValue<SpecifiedValue>),
    AlignSelf(DeclaredValue<SpecifiedValue>),
    Order(DeclaredValue<SpecifiedValue>),
    FlexBasis(DeclaredValue<SpecifiedValue>),
    Width(DeclaredValue<SpecifiedValue>),
    MinWidth(DeclaredValue<SpecifiedValue>),
    MaxWidth(DeclaredValue<SpecifiedValue>),
    Height(DeclaredValue<SpecifiedValue>),
    MinHeight(DeclaredValue<SpecifiedValue>),
    MaxHeight(DeclaredValue<SpecifiedValue>),
    BlockSize(DeclaredValue<SpecifiedValue>),
    MinBlockSize(DeclaredValue<SpecifiedValue>),
    MaxBlockSize(DeclaredValue<SpecifiedValue>),
    InlineSize(DeclaredValue<SpecifiedValue>),
    MinInlineSize(DeclaredValue<SpecifiedValue>),
    MaxInlineSize(DeclaredValue<SpecifiedValue>),
    BoxSizing(DeclaredValue<SpecifiedValue>),
    ObjectFit(DeclaredValue<SpecifiedValue>),
    ObjectPosition(DeclaredValue<Box<SpecifiedValue>>),
    GridRowStart(DeclaredValue<Box<SpecifiedValue>>),
    GridRowEnd(DeclaredValue<Box<SpecifiedValue>>),
    GridColumnStart(DeclaredValue<Box<SpecifiedValue>>),
    GridColumnEnd(DeclaredValue<Box<SpecifiedValue>>),
    GridRowGap(DeclaredValue<SpecifiedValue>),
    GridColumnGap(DeclaredValue<SpecifiedValue>),
    TableLayout(DeclaredValue<SpecifiedValue>),
    TextOverflow(DeclaredValue<Box<SpecifiedValue>>),
    UnicodeBidi(DeclaredValue<SpecifiedValue>),
    TextDecorationLine(DeclaredValue<SpecifiedValue>),
    TextDecorationStyle(DeclaredValue<SpecifiedValue>),
    TextDecorationColor(DeclaredValue<Box<SpecifiedValue>>),
    ImeMode(DeclaredValue<SpecifiedValue>),
    MozUserSelect(DeclaredValue<SpecifiedValue>),
    MozWindowDragging(DeclaredValue<SpecifiedValue>),
    TextAnchor(DeclaredValue<SpecifiedValue>),
    ColorInterpolation(DeclaredValue<SpecifiedValue>),
    ColorInterpolationFilters(DeclaredValue<SpecifiedValue>),
    FillOpacity(DeclaredValue<SpecifiedValue>),
    FillRule(DeclaredValue<SpecifiedValue>),
    ShapeRendering(DeclaredValue<SpecifiedValue>),
    StrokeLinecap(DeclaredValue<SpecifiedValue>),
    StrokeLinejoin(DeclaredValue<SpecifiedValue>),
    StrokeMiterlimit(DeclaredValue<SpecifiedValue>),
    StrokeOpacity(DeclaredValue<SpecifiedValue>),
    ClipRule(DeclaredValue<SpecifiedValue>),
    DominantBaseline(DeclaredValue<SpecifiedValue>),
    VectorEffect(DeclaredValue<SpecifiedValue>),
    StopColor(DeclaredValue<Box<SpecifiedValue>>),
    StopOpacity(DeclaredValue<SpecifiedValue>),
    FloodColor(DeclaredValue<Box<SpecifiedValue>>),
    FloodOpacity(DeclaredValue<SpecifiedValue>),
    LightingColor(DeclaredValue<Box<SpecifiedValue>>),
    MaskType(DeclaredValue<SpecifiedValue>),
    ClipPath(DeclaredValue<Box<SpecifiedValue>>),
    MaskMode(DeclaredValue<SpecifiedValue>),
    MaskRepeat(DeclaredValue<SpecifiedValue>),
    MaskPositionX(DeclaredValue<SpecifiedValue>),
    MaskPositionY(DeclaredValue<SpecifiedValue>),
    MaskClip(DeclaredValue<SpecifiedValue>),
    MaskOrigin(DeclaredValue<SpecifiedValue>),
    MaskSize(DeclaredValue<SpecifiedValue>),
    MaskComposite(DeclaredValue<SpecifiedValue>),
    MaskImage(DeclaredValue<SpecifiedValue>),
    MozBoxAlign(DeclaredValue<SpecifiedValue>),
    MozBoxFlex(DeclaredValue<SpecifiedValue>),
    MozStackSizing(DeclaredValue<SpecifiedValue>),
    Custom(Name, DeclaredValue<Box<SpecifiedValue>>),
}

Servo's representation for a property declaration.

Variants
BackgroundColor(DeclaredValue<Box<SpecifiedValue>>)

background-color

BackgroundImage(DeclaredValue<SpecifiedValue>)

background-image

BackgroundPositionX(DeclaredValue<SpecifiedValue>)

background-position-x

BackgroundPositionY(DeclaredValue<SpecifiedValue>)

background-position-y

BackgroundRepeat(DeclaredValue<SpecifiedValue>)

background-repeat

BackgroundAttachment(DeclaredValue<SpecifiedValue>)

background-attachment

BackgroundClip(DeclaredValue<SpecifiedValue>)

background-clip

BackgroundOrigin(DeclaredValue<SpecifiedValue>)

background-origin

BackgroundSize(DeclaredValue<SpecifiedValue>)

background-size

BackgroundBlendMode(DeclaredValue<SpecifiedValue>)

background-blend-mode

BorderTopColor(DeclaredValue<Box<SpecifiedValue>>)

border-top-color

BorderLeftColor(DeclaredValue<Box<SpecifiedValue>>)

border-left-color

BorderBottomColor(DeclaredValue<Box<SpecifiedValue>>)

border-bottom-color

BorderRightColor(DeclaredValue<Box<SpecifiedValue>>)

border-right-color

BorderBlockStartColor(DeclaredValue<Box<SpecifiedValue>>)

border-block-start-color

BorderBlockEndColor(DeclaredValue<Box<SpecifiedValue>>)

border-block-end-color

BorderInlineStartColor(DeclaredValue<Box<SpecifiedValue>>)

border-inline-start-color

BorderInlineEndColor(DeclaredValue<Box<SpecifiedValue>>)

border-inline-end-color

BorderTopStyle(DeclaredValue<SpecifiedValue>)

border-top-style

BorderLeftStyle(DeclaredValue<SpecifiedValue>)

border-left-style

BorderBottomStyle(DeclaredValue<SpecifiedValue>)

border-bottom-style

BorderRightStyle(DeclaredValue<SpecifiedValue>)

border-right-style

BorderBlockStartStyle(DeclaredValue<SpecifiedValue>)

border-block-start-style

BorderBlockEndStyle(DeclaredValue<SpecifiedValue>)

border-block-end-style

BorderInlineStartStyle(DeclaredValue<SpecifiedValue>)

border-inline-start-style

BorderInlineEndStyle(DeclaredValue<SpecifiedValue>)

border-inline-end-style

BorderTopWidth(DeclaredValue<SpecifiedValue>)

border-top-width

BorderLeftWidth(DeclaredValue<SpecifiedValue>)

border-left-width

BorderBottomWidth(DeclaredValue<SpecifiedValue>)

border-bottom-width

BorderRightWidth(DeclaredValue<SpecifiedValue>)

border-right-width

BorderBlockStartWidth(DeclaredValue<SpecifiedValue>)

border-block-start-width

BorderBlockEndWidth(DeclaredValue<SpecifiedValue>)

border-block-end-width

BorderInlineStartWidth(DeclaredValue<SpecifiedValue>)

border-inline-start-width

BorderInlineEndWidth(DeclaredValue<SpecifiedValue>)

border-inline-end-width

BorderTopLeftRadius(DeclaredValue<SpecifiedValue>)

border-top-left-radius

BorderTopRightRadius(DeclaredValue<SpecifiedValue>)

border-top-right-radius

BorderBottomRightRadius(DeclaredValue<SpecifiedValue>)

border-bottom-right-radius

BorderBottomLeftRadius(DeclaredValue<SpecifiedValue>)

border-bottom-left-radius

BoxDecorationBreak(DeclaredValue<SpecifiedValue>)

box-decoration-break

MozFloatEdge(DeclaredValue<SpecifiedValue>)

-moz-float-edge

BorderImageSource(DeclaredValue<Box<SpecifiedValue>>)

border-image-source

BorderImageOutset(DeclaredValue<SpecifiedValue>)

border-image-outset

BorderImageRepeat(DeclaredValue<SpecifiedValue>)

border-image-repeat

BorderImageWidth(DeclaredValue<SpecifiedValue>)

border-image-width

BorderImageSlice(DeclaredValue<SpecifiedValue>)

border-image-slice

Display(DeclaredValue<SpecifiedValue>)

display

MozTopLayer(DeclaredValue<SpecifiedValue>)

-moz-top-layer

Position(DeclaredValue<SpecifiedValue>)

position

Float(DeclaredValue<SpecifiedValue>)

float

Clear(DeclaredValue<SpecifiedValue>)

clear

VerticalAlign(DeclaredValue<SpecifiedValue>)

vertical-align

OverflowClipBox(DeclaredValue<SpecifiedValue>)

overflow-clip-box

OverflowX(DeclaredValue<SpecifiedValue>)

overflow-x

OverflowY(DeclaredValue<SpecifiedValue>)

overflow-y

TransitionDuration(DeclaredValue<SpecifiedValue>)

transition-duration

TransitionTimingFunction(DeclaredValue<SpecifiedValue>)

transition-timing-function

TransitionProperty(DeclaredValue<SpecifiedValue>)

transition-property

TransitionDelay(DeclaredValue<SpecifiedValue>)

transition-delay

AnimationName(DeclaredValue<SpecifiedValue>)

animation-name

AnimationDuration(DeclaredValue<SpecifiedValue>)

animation-duration

AnimationTimingFunction(DeclaredValue<SpecifiedValue>)

animation-timing-function

AnimationIterationCount(DeclaredValue<SpecifiedValue>)

animation-iteration-count

AnimationDirection(DeclaredValue<SpecifiedValue>)

animation-direction

AnimationPlayState(DeclaredValue<SpecifiedValue>)

animation-play-state

AnimationFillMode(DeclaredValue<SpecifiedValue>)

animation-fill-mode

AnimationDelay(DeclaredValue<SpecifiedValue>)

animation-delay

ScrollSnapPointsY(DeclaredValue<SpecifiedValue>)

scroll-snap-points-y

ScrollSnapPointsX(DeclaredValue<SpecifiedValue>)

scroll-snap-points-x

ScrollSnapDestination(DeclaredValue<Box<SpecifiedValue>>)

scroll-snap-destination

ScrollSnapCoordinate(DeclaredValue<SpecifiedValue>)

scroll-snap-coordinate

Transform(DeclaredValue<SpecifiedValue>)

transform

ScrollBehavior(DeclaredValue<SpecifiedValue>)

scroll-behavior

ScrollSnapTypeX(DeclaredValue<SpecifiedValue>)

scroll-snap-type-x

ScrollSnapTypeY(DeclaredValue<SpecifiedValue>)

scroll-snap-type-y

Isolation(DeclaredValue<SpecifiedValue>)

isolation

PageBreakAfter(DeclaredValue<SpecifiedValue>)

page-break-after

PageBreakBefore(DeclaredValue<SpecifiedValue>)

page-break-before

PageBreakInside(DeclaredValue<SpecifiedValue>)

page-break-inside

Resize(DeclaredValue<SpecifiedValue>)

resize

Perspective(DeclaredValue<SpecifiedValue>)

perspective

PerspectiveOrigin(DeclaredValue<SpecifiedValue>)

perspective-origin

BackfaceVisibility(DeclaredValue<SpecifiedValue>)

backface-visibility

TransformBox(DeclaredValue<SpecifiedValue>)

transform-box

TransformStyle(DeclaredValue<SpecifiedValue>)

transform-style

TransformOrigin(DeclaredValue<Box<SpecifiedValue>>)

transform-origin

MozAppearance(DeclaredValue<SpecifiedValue>)

-moz-appearance

MozBinding(DeclaredValue<Box<SpecifiedValue>>)

-moz-binding

MozOrient(DeclaredValue<SpecifiedValue>)

-moz-orient

Color(DeclaredValue<Box<SpecifiedValue>>)

color

ColumnWidth(DeclaredValue<SpecifiedValue>)

column-width

ColumnCount(DeclaredValue<SpecifiedValue>)

column-count

ColumnGap(DeclaredValue<SpecifiedValue>)

column-gap

ColumnFill(DeclaredValue<SpecifiedValue>)

column-fill

ColumnRuleWidth(DeclaredValue<SpecifiedValue>)

column-rule-width

ColumnRuleColor(DeclaredValue<Box<SpecifiedValue>>)

column-rule-color

ColumnRuleStyle(DeclaredValue<SpecifiedValue>)

column-rule-style

Content(DeclaredValue<SpecifiedValue>)

content

Opacity(DeclaredValue<SpecifiedValue>)

opacity

BoxShadow(DeclaredValue<SpecifiedValue>)

box-shadow

Filter(DeclaredValue<SpecifiedValue>)

filter

MixBlendMode(DeclaredValue<SpecifiedValue>)

mix-blend-mode

FontFamily(DeclaredValue<SpecifiedValue>)

font-family

FontStyle(DeclaredValue<SpecifiedValue>)

font-style

FontVariant(DeclaredValue<SpecifiedValue>)

font-variant

FontVariantCaps(DeclaredValue<SpecifiedValue>)

font-variant-caps

FontWeight(DeclaredValue<SpecifiedValue>)

font-weight

FontSize(DeclaredValue<SpecifiedValue>)

font-size

FontSizeAdjust(DeclaredValue<SpecifiedValue>)

font-size-adjust

FontSynthesis(DeclaredValue<SpecifiedValue>)

font-synthesis

FontStretch(DeclaredValue<SpecifiedValue>)

font-stretch

FontKerning(DeclaredValue<SpecifiedValue>)

font-kerning

FontVariantPosition(DeclaredValue<SpecifiedValue>)

font-variant-position

Visibility(DeclaredValue<SpecifiedValue>)

visibility

WritingMode(DeclaredValue<SpecifiedValue>)

writing-mode

Direction(DeclaredValue<SpecifiedValue>)

direction

TextOrientation(DeclaredValue<SpecifiedValue>)

text-orientation

ColorAdjust(DeclaredValue<SpecifiedValue>)

color-adjust

ImageRendering(DeclaredValue<SpecifiedValue>)

image-rendering

BorderCollapse(DeclaredValue<SpecifiedValue>)

border-collapse

EmptyCells(DeclaredValue<SpecifiedValue>)

empty-cells

CaptionSide(DeclaredValue<SpecifiedValue>)

caption-side

BorderSpacing(DeclaredValue<Box<SpecifiedValue>>)

border-spacing

LineHeight(DeclaredValue<SpecifiedValue>)

line-height

TextTransform(DeclaredValue<SpecifiedValue>)

text-transform

Hyphens(DeclaredValue<SpecifiedValue>)

hyphens

TextIndent(DeclaredValue<SpecifiedValue>)

text-indent

OverflowWrap(DeclaredValue<SpecifiedValue>)

overflow-wrap

WordBreak(DeclaredValue<SpecifiedValue>)

word-break

TextAlignLast(DeclaredValue<SpecifiedValue>)

text-align-last

TextAlign(DeclaredValue<SpecifiedValue>)

text-align

LetterSpacing(DeclaredValue<SpecifiedValue>)

letter-spacing

WordSpacing(DeclaredValue<SpecifiedValue>)

word-spacing

WhiteSpace(DeclaredValue<SpecifiedValue>)

white-space

TextShadow(DeclaredValue<SpecifiedValue>)

text-shadow

TextEmphasisStyle(DeclaredValue<SpecifiedValue>)

text-emphasis-style

TextEmphasisPosition(DeclaredValue<SpecifiedValue>)

text-emphasis-position

TextEmphasisColor(DeclaredValue<Box<SpecifiedValue>>)

text-emphasis-color

MozTabSize(DeclaredValue<SpecifiedValue>)

-moz-tab-size

WebkitTextFillColor(DeclaredValue<Box<SpecifiedValue>>)

-webkit-text-fill-color

WebkitTextStrokeColor(DeclaredValue<Box<SpecifiedValue>>)

-webkit-text-stroke-color

WebkitTextStrokeWidth(DeclaredValue<SpecifiedValue>)

-webkit-text-stroke-width

RubyAlign(DeclaredValue<SpecifiedValue>)

ruby-align

RubyPosition(DeclaredValue<SpecifiedValue>)

ruby-position

TextCombineUpright(DeclaredValue<SpecifiedValue>)

text-combine-upright

TextRendering(DeclaredValue<SpecifiedValue>)

text-rendering

ListStylePosition(DeclaredValue<SpecifiedValue>)

list-style-position

ListStyleType(DeclaredValue<SpecifiedValue>)

list-style-type

ListStyleImage(DeclaredValue<SpecifiedValue>)

list-style-image

Quotes(DeclaredValue<SpecifiedValue>)

quotes

MozImageRegion(DeclaredValue<Box<SpecifiedValue>>)

-moz-image-region

MarginTop(DeclaredValue<SpecifiedValue>)

margin-top

MarginLeft(DeclaredValue<SpecifiedValue>)

margin-left

MarginBottom(DeclaredValue<SpecifiedValue>)

margin-bottom

MarginRight(DeclaredValue<SpecifiedValue>)

margin-right

MarginBlockStart(DeclaredValue<SpecifiedValue>)

margin-block-start

MarginBlockEnd(DeclaredValue<SpecifiedValue>)

margin-block-end

MarginInlineStart(DeclaredValue<SpecifiedValue>)

margin-inline-start

MarginInlineEnd(DeclaredValue<SpecifiedValue>)

margin-inline-end

OutlineColor(DeclaredValue<Box<SpecifiedValue>>)

outline-color

OutlineStyle(DeclaredValue<SpecifiedValue>)

outline-style

OutlineWidth(DeclaredValue<SpecifiedValue>)

outline-width

MozOutlineRadiusTopleft(DeclaredValue<SpecifiedValue>)

-moz-outline-radius-topleft

MozOutlineRadiusTopright(DeclaredValue<SpecifiedValue>)

-moz-outline-radius-topright

MozOutlineRadiusBottomright(DeclaredValue<SpecifiedValue>)

-moz-outline-radius-bottomright

MozOutlineRadiusBottomleft(DeclaredValue<SpecifiedValue>)

-moz-outline-radius-bottomleft

OutlineOffset(DeclaredValue<SpecifiedValue>)

outline-offset

PaddingTop(DeclaredValue<SpecifiedValue>)

padding-top

PaddingLeft(DeclaredValue<SpecifiedValue>)

padding-left

PaddingBottom(DeclaredValue<SpecifiedValue>)

padding-bottom

PaddingRight(DeclaredValue<SpecifiedValue>)

padding-right

PaddingBlockStart(DeclaredValue<SpecifiedValue>)

padding-block-start

PaddingBlockEnd(DeclaredValue<SpecifiedValue>)

padding-block-end

PaddingInlineStart(DeclaredValue<SpecifiedValue>)

padding-inline-start

PaddingInlineEnd(DeclaredValue<SpecifiedValue>)

padding-inline-end

Cursor(DeclaredValue<SpecifiedValue>)

cursor

PointerEvents(DeclaredValue<SpecifiedValue>)

pointer-events

MozUserInput(DeclaredValue<SpecifiedValue>)

-moz-user-input

MozUserModify(DeclaredValue<SpecifiedValue>)

-moz-user-modify

MozUserFocus(DeclaredValue<SpecifiedValue>)

-moz-user-focus

Top(DeclaredValue<SpecifiedValue>)

top

Left(DeclaredValue<SpecifiedValue>)

left

Bottom(DeclaredValue<SpecifiedValue>)

bottom

Right(DeclaredValue<SpecifiedValue>)

right

OffsetBlockStart(DeclaredValue<SpecifiedValue>)

offset-block-start

OffsetBlockEnd(DeclaredValue<SpecifiedValue>)

offset-block-end

OffsetInlineStart(DeclaredValue<SpecifiedValue>)

offset-inline-start

OffsetInlineEnd(DeclaredValue<SpecifiedValue>)

offset-inline-end

ZIndex(DeclaredValue<SpecifiedValue>)

z-index

FlexDirection(DeclaredValue<SpecifiedValue>)

flex-direction

FlexWrap(DeclaredValue<SpecifiedValue>)

flex-wrap

AlignItems(DeclaredValue<SpecifiedValue>)

align-items

FlexGrow(DeclaredValue<SpecifiedValue>)

flex-grow

FlexShrink(DeclaredValue<SpecifiedValue>)

flex-shrink

AlignSelf(DeclaredValue<SpecifiedValue>)

align-self

Order(DeclaredValue<SpecifiedValue>)

order

FlexBasis(DeclaredValue<SpecifiedValue>)

flex-basis

Width(DeclaredValue<SpecifiedValue>)

width

MinWidth(DeclaredValue<SpecifiedValue>)

min-width

MaxWidth(DeclaredValue<SpecifiedValue>)

max-width

Height(DeclaredValue<SpecifiedValue>)

height

MinHeight(DeclaredValue<SpecifiedValue>)

min-height

MaxHeight(DeclaredValue<SpecifiedValue>)

max-height

BlockSize(DeclaredValue<SpecifiedValue>)

block-size

MinBlockSize(DeclaredValue<SpecifiedValue>)

min-block-size

MaxBlockSize(DeclaredValue<SpecifiedValue>)

max-block-size

InlineSize(DeclaredValue<SpecifiedValue>)

inline-size

MinInlineSize(DeclaredValue<SpecifiedValue>)

min-inline-size

MaxInlineSize(DeclaredValue<SpecifiedValue>)

max-inline-size

BoxSizing(DeclaredValue<SpecifiedValue>)

box-sizing

ObjectFit(DeclaredValue<SpecifiedValue>)

object-fit

ObjectPosition(DeclaredValue<Box<SpecifiedValue>>)

object-position

GridRowStart(DeclaredValue<Box<SpecifiedValue>>)

grid-row-start

GridRowEnd(DeclaredValue<Box<SpecifiedValue>>)

grid-row-end

GridColumnStart(DeclaredValue<Box<SpecifiedValue>>)

grid-column-start

GridColumnEnd(DeclaredValue<Box<SpecifiedValue>>)

grid-column-end

GridRowGap(DeclaredValue<SpecifiedValue>)

grid-row-gap

GridColumnGap(DeclaredValue<SpecifiedValue>)

grid-column-gap

TableLayout(DeclaredValue<SpecifiedValue>)

table-layout

TextOverflow(DeclaredValue<Box<SpecifiedValue>>)

text-overflow

UnicodeBidi(DeclaredValue<SpecifiedValue>)

unicode-bidi

TextDecorationLine(DeclaredValue<SpecifiedValue>)

text-decoration-line

TextDecorationStyle(DeclaredValue<SpecifiedValue>)

text-decoration-style

TextDecorationColor(DeclaredValue<Box<SpecifiedValue>>)

text-decoration-color

ImeMode(DeclaredValue<SpecifiedValue>)

ime-mode

MozUserSelect(DeclaredValue<SpecifiedValue>)

-moz-user-select

MozWindowDragging(DeclaredValue<SpecifiedValue>)

-moz-window-dragging

TextAnchor(DeclaredValue<SpecifiedValue>)

text-anchor

ColorInterpolation(DeclaredValue<SpecifiedValue>)

color-interpolation

ColorInterpolationFilters(DeclaredValue<SpecifiedValue>)

color-interpolation-filters

FillOpacity(DeclaredValue<SpecifiedValue>)

fill-opacity

FillRule(DeclaredValue<SpecifiedValue>)

fill-rule

ShapeRendering(DeclaredValue<SpecifiedValue>)

shape-rendering

StrokeLinecap(DeclaredValue<SpecifiedValue>)

stroke-linecap

StrokeLinejoin(DeclaredValue<SpecifiedValue>)

stroke-linejoin

StrokeMiterlimit(DeclaredValue<SpecifiedValue>)

stroke-miterlimit

StrokeOpacity(DeclaredValue<SpecifiedValue>)

stroke-opacity

ClipRule(DeclaredValue<SpecifiedValue>)

clip-rule

DominantBaseline(DeclaredValue<SpecifiedValue>)

dominant-baseline

VectorEffect(DeclaredValue<SpecifiedValue>)

vector-effect

StopColor(DeclaredValue<Box<SpecifiedValue>>)

stop-color

StopOpacity(DeclaredValue<SpecifiedValue>)

stop-opacity

FloodColor(DeclaredValue<Box<SpecifiedValue>>)

flood-color

FloodOpacity(DeclaredValue<SpecifiedValue>)

flood-opacity

LightingColor(DeclaredValue<Box<SpecifiedValue>>)

lighting-color

MaskType(DeclaredValue<SpecifiedValue>)

mask-type

ClipPath(DeclaredValue<Box<SpecifiedValue>>)

clip-path

MaskMode(DeclaredValue<SpecifiedValue>)

mask-mode

MaskRepeat(DeclaredValue<SpecifiedValue>)

mask-repeat

MaskPositionX(DeclaredValue<SpecifiedValue>)

mask-position-x

MaskPositionY(DeclaredValue<SpecifiedValue>)

mask-position-y

MaskClip(DeclaredValue<SpecifiedValue>)

mask-clip

MaskOrigin(DeclaredValue<SpecifiedValue>)

mask-origin

MaskSize(DeclaredValue<SpecifiedValue>)

mask-size

MaskComposite(DeclaredValue<SpecifiedValue>)

mask-composite

MaskImage(DeclaredValue<SpecifiedValue>)

mask-image

MozBoxAlign(DeclaredValue<SpecifiedValue>)

-moz-box-align

MozBoxFlex(DeclaredValue<SpecifiedValue>)

-moz-box-flex

MozStackSizing(DeclaredValue<SpecifiedValue>)

-moz-stack-sizing

Custom(Name, DeclaredValue<Box<SpecifiedValue>>)

A custom property declaration, with the property name and the declared value.

Methods

impl PropertyDeclaration

fn id(&self) -> PropertyDeclarationId

Given a property declaration, return the property declaration id.

fn with_variables(&self) -> bool

Return whether this is a pending-substitution value. https://drafts.csswg.org/css-variables/#variables-in-shorthands

fn value_is_unparsed(&self) -> bool

Return whether the value is stored as it was in the CSS source, preserving whitespace (as opposed to being parsed into a more abstract data structure).

This is the case of custom properties and values that contain unsubstituted variables.

fn parse(id: PropertyId,
context: &ParserContext,
input: &mut Parser,
result_list: &mut Vec<(PropertyDeclaration, Importance)>,
in_keyframe_block: bool)
-> PropertyDeclarationParseResult

The in_keyframe_block parameter controls this:

https://drafts.csswg.org/css-animations/#keyframes

The inside of accepts any CSS property except those defined in this specification, but does accept the animation-play-state property and interprets it specially.

This will not actually parse Importance values, and will always set things to Importance::Normal. Parsing Importance values is the job of PropertyDeclarationParser, we only set them here so that we don't have to reallocate

fn shorthands(&self) -> &'static [ShorthandId]

The shorthands that this longhand is part of.

fn is_animatable(&self) -> bool

Returns true if this property is one of the animable properties, false otherwise.