Enum style::properties::PropertyId

pub enum PropertyId {
    Longhand(LonghandId),
    Shorthand(ShorthandId),
    LonghandAlias(LonghandId, AliasId),
    ShorthandAlias(ShorthandId, AliasId),
    Custom(Name),
}

Servo's representation of a CSS property, that is, either a longhand, a shorthand, or a custom property.

Variants

Longhand(LonghandId)

A longhand property.

Shorthand(ShorthandId)

A shorthand property.

LonghandAlias(LonghandId, AliasId)

An alias for a longhand property.

ShorthandAlias(ShorthandId, AliasId)

An alias for a shorthand property.

Custom(Name)

A custom property.

Methods

impl PropertyId

pub fn longhand_id(&self) -> Option<LonghandId>

Return the longhand id that this property id represents.

pub fn parse_unchecked_for_testing(name: &str) -> Result<Self, ()>

Returns a given property from the given name, regardless of whether it is enabled or not, or Err(()) for unknown properties.

Do not use for non-testing purposes.

fn parse_unchecked(
property_name: &str,
use_counters: Option<&UseCounters>
) -> Result<Self, ()>

Returns a given property from the given name, regardless of whether it is enabled or not, or Err(()) for unknown properties.

pub fn parse_enabled_for_all_content(name: &str) -> Result<Self, ()>

Parses a property name, and returns an error if it's unknown or isn't enabled for all content.

pub fn parse(name: &str, context: &ParserContext) -> Result<Self, ()>

Parses a property name, and returns an error if it's unknown or isn't allowed in this context.

pub fn parse_ignoring_rule_type(
name: &str,
context: &ParserContext
) -> Result<Self, ()>

Parses a property name, and returns an error if it's unknown or isn't allowed in this context, ignoring the rule_type checks.

This is useful for parsing stuff from CSS values, for example.

pub fn is_shorthand(&self) -> bool

Returns true if the property is a shorthand or shorthand alias.

pub fn as_shorthand(&self) -> Result<ShorthandId, PropertyDeclarationId>

Given this property id, get it either as a shorthand or as a PropertyDeclarationId.

pub fn non_custom_id(&self) -> Option<NonCustomPropertyId>

Returns the NonCustomPropertyId corresponding to this property id.

fn non_custom_non_alias_id(&self) -> Option<NonCustomPropertyId>

Returns non-alias NonCustomPropertyId corresponding to this property id.

pub fn enabled_for_all_content(&self) -> bool

Whether the property is enabled for all content regardless of the stylesheet it was declared on (that is, in practice only checks prefs).

fn allowed_in(&self, context: &ParserContext) -> bool

fn allowed_in_ignoring_rule_type(&self, context: &ParserContext) -> bool

pub fn supports_type(&self, ty: u8) -> bool

Whether the property supports the given CSS type. ty should a bitflags of constants in style_traits::CssType.

pub fn collect_property_completion_keywords(&self, f: KeywordsCollectFn)

Collect supported starting word of values of this property.

See style_traits::SpecifiedValueInfo::collect_completion_keywords for more details.