Enum style::properties::PropertyId

pub enum PropertyId {
    Longhand(LonghandId),
    Shorthand(ShorthandId),
    Custom(Name),
}

Servo's representation of a CSS property, that is, either a longhand, a shorthand, or a custom property.

Variants
Longhand(LonghandId)

A longhand property.

Shorthand(ShorthandId)

A shorthand property.

Custom(Name)

A custom property.

Methods

impl PropertyId

pub fn parse(
property_name: &str,
context: Option<&PropertyParserContext>
) -> Result<Self, ()>

Returns a given property from the string s.

Returns Err(()) for unknown non-custom properties If caller wants to provide a different context, it can be provided with Some(context), if None is given, default setting for PropertyParserContext will be used. It is Origin::Author for stylesheet_origin and CssRuleType::Style for rule_type.

pub fn as_shorthand(&self) -> Result<ShorthandId, PropertyDeclarationId>

Given this property id, get it either as a shorthand or as a PropertyDeclarationId.

pub fn name(&self) -> Cow<'static, str>

Returns the name of the property without CSS escaping.