Enum style::properties::PropertyId

pub enum PropertyId {
    Longhand(LonghandId),
    Shorthand(ShorthandId),
    Custom(Name),
}

Servo's representation of a CSS property, that is, either a longhand, a shorthand, or a custom property.

Variants
Longhand(LonghandId)

A longhand property.

Shorthand(ShorthandId)

A shorthand property.

Custom(Name)

A custom property.

Methods

impl PropertyId

fn parse(s: Cow<str>) -> Result<Self, ()>

Returns a given property from the string s.

Returns Err(()) for unknown non-custom properties

fn from_nscsspropertyid(id: nsCSSPropertyID) -> Result<Self, ()>

Returns a property id from Gecko's nsCSSPropertyID.

fn to_nscsspropertyid(&self) -> Result<nsCSSPropertyID, ()>

Returns a property id from Gecko's nsCSSPropertyID.

fn as_shorthand(&self) -> Result<ShorthandId, PropertyDeclarationId>

Given this property id, get it either as a shorthand or as a PropertyDeclarationId.