Enum style::properties::PropertyId

pub enum PropertyId {
    Longhand(LonghandId),
    Shorthand(ShorthandId),
    LonghandAlias(LonghandId, AliasId),
    ShorthandAlias(ShorthandId, AliasId),
    Custom(Name),
}

Servo's representation of a CSS property, that is, either a longhand, a shorthand, or a custom property.

Variants
Longhand(LonghandId)

A longhand property.

Shorthand(ShorthandId)

A shorthand property.

LonghandAlias(LonghandId, AliasId)

An alias for a longhand property.

ShorthandAlias(ShorthandId, AliasId)

An alias for a shorthand property.

Custom(Name)

A custom property.

Methods

impl PropertyId

pub fn longhand_id(&self) -> Option<LonghandId>

Return the longhand id that this property id represents.

pub fn parse(property_name: &str) -> Result<Self, ()>

Returns a given property from the string s.

Returns Err(()) for unknown non-custom properties.

pub fn as_shorthand(&self) -> Result<ShorthandId, PropertyDeclarationId>

Given this property id, get it either as a shorthand or as a PropertyDeclarationId.

pub fn name(&self) -> Cow<'static, str>

Returns the name of the property without CSS escaping.