Enum style::properties::ShorthandId

#[repr(u16)]pub enum ShorthandId {
    Background,
    BackgroundPosition,
    BorderColor,
    BorderStyle,
    BorderWidth,
    BorderTop,
    BorderRight,
    BorderBottom,
    BorderLeft,
    BorderBlockStart,
    BorderBlockEnd,
    BorderInlineStart,
    BorderInlineEnd,
    Border,
    BorderRadius,
    BorderImage,
    BorderBlockWidth,
    BorderBlockStyle,
    BorderBlockColor,
    BorderInlineWidth,
    BorderInlineStyle,
    BorderInlineColor,
    BorderBlock,
    BorderInline,
    Overflow,
    Transition,
    Animation,
    Columns,
    Font,
    FontVariant,
    ListStyle,
    Margin,
    MarginBlock,
    MarginInline,
    Outline,
    Padding,
    PaddingBlock,
    PaddingInline,
    FlexFlow,
    Flex,
    Inset,
    InsetBlock,
    InsetInline,
    TextDecoration,
    All,
}

An identifier for a given shorthand property.

Variants

Background

background

BackgroundPosition

background-position

BorderColor

border-color

BorderStyle

border-style

BorderWidth

border-width

BorderTop

border-top

BorderRight

border-right

BorderBottom

border-bottom

BorderLeft

border-left

BorderBlockStart

border-block-start

BorderBlockEnd

border-block-end

BorderInlineStart

border-inline-start

BorderInlineEnd

border-inline-end

Border

border

BorderRadius

border-radius

BorderImage

border-image

BorderBlockWidth

border-block-width

BorderBlockStyle

border-block-style

BorderBlockColor

border-block-color

BorderInlineWidth

border-inline-width

BorderInlineStyle

border-inline-style

BorderInlineColor

border-inline-color

BorderBlock

border-block

BorderInline

border-inline

Overflow

overflow

Transition

transition

Animation

animation

Columns

columns

Font

font

FontVariant

font-variant

ListStyle

list-style

Margin

margin

MarginBlock

margin-block

MarginInline

margin-inline

Outline

outline

Padding

padding

PaddingBlock

padding-block

PaddingInline

padding-inline

FlexFlow

flex-flow

Flex

flex

Inset

inset

InsetBlock

inset-block

InsetInline

inset-inline

TextDecoration

text-decoration

All

all

Implementations

impl ShorthandId

pub fn name(&self) -> &'static str

Get the name for this shorthand property.

pub fn longhands(&self) -> NonCustomPropertyIterator<LonghandId>ⓘ

Important traits for NonCustomPropertyIterator<Item>
impl<Item> Iterator for NonCustomPropertyIterator<Item> where
Item: 'static + Copy + Into<NonCustomPropertyId>, type Item = Item;

Get the longhand ids that form this shorthand.

pub fn longhands_to_css<'a, W, I>(
&self,
declarations: I,
dest: &mut CssWriter<W>
) -> Result where
W: Write,
I: Iterator<Item = &'a PropertyDeclaration>,

Try to serialize the given declarations as this shorthand.

Returns an error if writing to the stream fails, or if the declarations do not map to a shorthand.

pub fn get_shorthand_appendable_value<'a, I>(
self,
declarations: I
) -> Option<AppendableValue<'a, I::IntoIter>> where
I: IntoIterator<Item = &'a PropertyDeclaration>,
I::IntoIter: Clone,

Finds and returns an appendable value for the given declarations.

Returns the optional appendable value.

pub fn flags(self) -> PropertyFlags

Returns PropertyFlags for the given shorthand property.

pub fn is_legacy_shorthand(self) -> bool

Returns whether this property is a legacy shorthand.

pub fn idl_name_sort_order(self) -> u32

Returns the order in which this property appears relative to other shorthands in idl-name-sorting order.

fn parse_into<'i, 't>(
&self,
declarations: &mut SourcePropertyDeclaration,
context: &ParserContext,
input: &mut Parser<'i, 't>
) -> Result<(), ParseError<'i>>