Enum style::properties::ShorthandId

pub enum ShorthandId {
    Background,
    BackgroundPosition,
    BorderColor,
    BorderStyle,
    BorderWidth,
    BorderTop,
    BorderLeft,
    BorderBottom,
    BorderRight,
    BorderBlockStart,
    BorderBlockEnd,
    BorderInlineStart,
    BorderInlineEnd,
    Border,
    BorderRadius,
    BorderImage,
    Overflow,
    Transition,
    Animation,
    ScrollSnapType,
    Columns,
    ColumnRule,
    Font,
    TextEmphasis,
    WebkitTextStroke,
    ListStyle,
    Margin,
    Mask,
    MaskPosition,
    Outline,
    MozOutlineRadius,
    Padding,
    FlexFlow,
    Flex,
    TextDecoration,
}

An identifier for a given shorthand property.

Variants
Background

background

BackgroundPosition

background-position

BorderColor

border-color

BorderStyle

border-style

BorderWidth

border-width

BorderTop

border-top

BorderLeft

border-left

BorderBottom

border-bottom

BorderRight

border-right

BorderBlockStart

border-block-start

BorderBlockEnd

border-block-end

BorderInlineStart

border-inline-start

BorderInlineEnd

border-inline-end

Border

border

BorderRadius

border-radius

BorderImage

border-image

Overflow

overflow

Transition

transition

Animation

animation

ScrollSnapType

scroll-snap-type

Columns

columns

ColumnRule

column-rule

Font

font

TextEmphasis

text-emphasis

WebkitTextStroke

-webkit-text-stroke

ListStyle

list-style

Margin

margin

Mask

mask

MaskPosition

mask-position

Outline

outline

MozOutlineRadius

-moz-outline-radius

Padding

padding

FlexFlow

flex-flow

Flex

flex

TextDecoration

text-decoration

Methods

impl ShorthandId

fn name(&self) -> &'static str

Get the name for this shorthand property.

fn longhands(&self) -> &'static [LonghandId]

Get the longhand ids that form this shorthand.

fn longhands_to_css<'a, W, I>(&self, declarations: I, dest: &mut W) -> Result where W: Write, I: Iterator<Item=&'a PropertyDeclaration>

Try to serialize the given declarations as this shorthand.

Returns an error if writing to the stream fails, or if the declarations do not map to a shorthand.

fn overflow_longhands_to_css<'a, W, I>(&self,
declarations: I,
dest: &mut W)
-> Result where W: Write, I: Iterator<Item=&'a PropertyDeclaration>

Overflow does not behave like a normal shorthand. When overflow-x and overflow-y are not of equal values, they no longer use the shared property name "overflow".

We use this function as a special-case for that.

fn serialize_shorthand_to_buffer<'a, W, I>(self,
dest: &mut W,
declarations: I,
is_first_serialization: &mut bool,
importance: Importance)
-> Result<bool, Error> where W: Write,
I: IntoIterator<Item=&'a PropertyDeclaration>,
I::IntoIter: Clone

Serializes the possible shorthand name with value to input buffer given a list of longhand declarations.

On success, returns true if the shorthand value is written, or false if no shorthand value is present.