Enum style::properties::ShorthandId

pub enum ShorthandId {
    Background,
    BackgroundPosition,
    BorderColor,
    BorderStyle,
    BorderWidth,
    BorderTop,
    BorderLeft,
    BorderBottom,
    BorderRight,
    BorderBlockStart,
    BorderBlockEnd,
    BorderInlineStart,
    BorderInlineEnd,
    Border,
    BorderRadius,
    BorderImage,
    Overflow,
    Transition,
    Animation,
    Columns,
    Font,
    FontVariant,
    ListStyle,
    Margin,
    Outline,
    Padding,
    FlexFlow,
    Flex,
    TextDecoration,
    All,
}

An identifier for a given shorthand property.

Variants
Background

background

BackgroundPosition

background-position

BorderColor

border-color

BorderStyle

border-style

BorderWidth

border-width

BorderTop

border-top

BorderLeft

border-left

BorderBottom

border-bottom

BorderRight

border-right

BorderBlockStart

border-block-start

BorderBlockEnd

border-block-end

BorderInlineStart

border-inline-start

BorderInlineEnd

border-inline-end

Border

border

BorderRadius

border-radius

BorderImage

border-image

Overflow

overflow

Transition

transition

Animation

animation

Columns

columns

Font

font

FontVariant

font-variant

ListStyle

list-style

Margin

margin

Outline

outline

Padding

padding

FlexFlow

flex-flow

Flex

flex

TextDecoration

text-decoration

All

all

Methods

impl ShorthandId

pub fn name(&self) -> &'static str

Get the name for this shorthand property.

pub fn longhands(&self) -> &'static [LonghandId]

Get the longhand ids that form this shorthand.

pub fn longhands_to_css<'a, W, I>(
&self,
declarations: I,
dest: &mut W
) -> Result where
W: Write,
I: Iterator<Item = &'a PropertyDeclaration>,

Try to serialize the given declarations as this shorthand.

Returns an error if writing to the stream fails, or if the declarations do not map to a shorthand.

pub fn get_shorthand_appendable_value<'a, I>(
self,
declarations: I
) -> Option<AppendableValue<'a, I::IntoIter>> where
I: IntoIterator<Item = &'a PropertyDeclaration>,
I::IntoIter: Clone,

Finds and returns an appendable value for the given declarations.

Returns the optional appendable value.

pub fn flags(&self) -> PropertyFlags

Returns PropertyFlags for given shorthand property.