style::properties

Module generated

The CSS properties supported by the style system. Generated from the properties.mako.rs template by build.rs

Modules

animated_properties

A module with all the code related to animated properties.

longhands

A module with all the code for longhand properties.

property_counts

A module to group various interesting property counts.

shorthands

A module with code for all the shorthand css properties, and a few serialization helpers.

style_structs

The module where all the style structs are defined.

Macros

expanded 🔒

unwrap_or_initial 🔒

Structs

AnimationCompositionIter

An iterator over the values of the animation-composition properties.

AnimationDelayIter

An iterator over the values of the animation-delay properties.

AnimationDirectionIter

An iterator over the values of the animation-direction properties.

AnimationDurationIter

An iterator over the values of the animation-duration properties.

AnimationFillModeIter

An iterator over the values of the animation-fill-mode properties.

AnimationIterationCountIter

An iterator over the values of the animation-iteration-count properties.

AnimationNameIter

An iterator over the values of the animation-name properties.

AnimationPlayStateIter

An iterator over the values of the animation-play-state properties.

AnimationTimelineIter

An iterator over the values of the animation-timeline properties.

AnimationTimingFunctionIter

An iterator over the values of the animation-timing-function properties.

ComputedValues

The struct that Servo uses to represent computed values.

ComputedValuesInner

Actual data of ComputedValues, to match up with Gecko

LogicalGroupSet

A set of logical groups.

LogicalMappingData 🔒

PropertyDeclarationVariantRepr 🔒

StyleBuilder

A type used to compute a struct with minimal overhead.

TransitionBehaviorIter

An iterator over the values of the transition-behavior properties.

TransitionDelayIter

An iterator over the values of the transition-delay properties.

TransitionDurationIter

An iterator over the values of the transition-duration properties.

TransitionPropertyIter

An iterator over the values of the transition-property properties.

TransitionTimingFunctionIter

An iterator over the values of the transition-timing-function properties.

Enums

AliasId

An identifier for a given alias property.

CountedUnknownProperty

The counted unknown property list which is used for css use counters.

LogicalGroupId

A group for properties which may override each other via logical resolution.

LogicalMappingKind 🔒

LonghandId

An identifier for a given longhand property.

PrioritaryPropertyId 🔒

PropertyDeclaration

Servo’s representation for a property declaration.

ShorthandId

An identifier for a given shorthand property.

StyleStructRef

A reference to a style struct of the parent, or our own style struct.

Constants

BOX_THRESHOLD 🔒

Statics

CASCADE_PROPERTY

A per-longhand array of functions to perform the CSS cascade on each of them, effectively doing virtual dispatch.

Traits

MaybeBoxed

Conversion with fewer impls than From/Into

Functions

restyle_damage_rebuild_box 🔒

restyle_damage_rebuild_stacking_context 🔒

restyle_damage_recalculate_overflow 🔒

restyle_damage_repaint 🔒

Type Aliases

CascadePropertyFn

A per-longhand function that performs the CSS cascade for that longhand.