style::properties::generated::animated_properties

Enum AnimationValue

#[repr(u16)]
pub enum AnimationValue {

    AlignItems(<ItemPlacement as ToComputedValue>::ComputedValue),
    AspectRatio(<<AspectRatio as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BackfaceVisibility(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BaselineSource(<BaselineSource as ToComputedValue>::ComputedValue),
    BorderCollapse(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BorderImageRepeat(<BorderImageRepeat as ToComputedValue>::ComputedValue),
    BoxSizing(<SpecifiedValue as ToComputedValue>::ComputedValue),
    CaptionSide(<CaptionSide as ToComputedValue>::ComputedValue),
    Clear(<Clear as ToComputedValue>::ComputedValue),
    ColumnCount(<<ColumnCount as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ColumnSpan(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Contain(Void),
    ContainerType(Void),
    Direction(Void),
    Display(<Display as ToComputedValue>::ComputedValue),
    EmptyCells(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FlexDirection(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FlexWrap(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Float(<Float as ToComputedValue>::ComputedValue),
    FontLanguageOverride(<FontLanguageOverride as ToComputedValue>::ComputedValue),
    FontOpticalSizing(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FontStretch(<<FontStretch as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontStyle(<<FontStyle as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontSynthesisWeight(<FontSynthesis as ToComputedValue>::ComputedValue),
    FontVariantCaps(<SpecifiedValue as ToComputedValue>::ComputedValue),
    FontWeight(<<FontWeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    GridAutoFlow(<GridAutoFlow as ToComputedValue>::ComputedValue),
    ImageRendering(<ImageRendering as ToComputedValue>::ComputedValue),
    Isolation(<SpecifiedValue as ToComputedValue>::ComputedValue),
    JustifyItems(<JustifyItems as ToComputedValue>::ComputedValue),
    ListStylePosition(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ListStyleType(<SpecifiedValue as ToComputedValue>::ComputedValue),
    MixBlendMode(<SpecifiedValue as ToComputedValue>::ComputedValue),
    ObjectFit(<SpecifiedValue as ToComputedValue>::ComputedValue),
    Opacity(<<Opacity as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Order(<<Integer as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineStyle(<OutlineStyle as ToComputedValue>::ComputedValue),
    OverflowWrap(<OverflowWrap as ToComputedValue>::ComputedValue),
    PointerEvents(<PointerEvents as ToComputedValue>::ComputedValue),
    Position(<PositionProperty as ToComputedValue>::ComputedValue),
    PositionArea(<PositionArea as ToComputedValue>::ComputedValue),
    ServoOverflowClipBox(Void),
    ServoTopLayer(Void),
    TableLayout(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextAlign(<TextAlign as ToComputedValue>::ComputedValue),
    TextAlignLast(<TextAlignLast as ToComputedValue>::ComputedValue),
    TextDecorationLine(<TextDecorationLine as ToComputedValue>::ComputedValue),
    TextDecorationStyle(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextJustify(<TextJustify as ToComputedValue>::ComputedValue),
    TextRendering(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TextTransform(<TextTransform as ToComputedValue>::ComputedValue),
    TextWrapMode(<SpecifiedValue as ToComputedValue>::ComputedValue),
    TransformStyle(<TransformStyle as ToComputedValue>::ComputedValue),
    UnicodeBidi(Void),
    Visibility(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    WhiteSpaceCollapse(<SpecifiedValue as ToComputedValue>::ComputedValue),
    WordBreak(<WordBreak as ToComputedValue>::ComputedValue),
    WritingMode(Void),
    ZIndex(<<ZIndex as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Zoom(<<Zoom as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    AlignContent(<ContentDistribution as ToComputedValue>::ComputedValue),
    JustifyContent(<ContentDistribution as ToComputedValue>::ComputedValue),
    FlexGrow(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FlexShrink(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    AlignSelf(<SelfAlignment as ToComputedValue>::ComputedValue),
    JustifySelf(<SelfAlignment as ToComputedValue>::ComputedValue),
    OverflowBlock(Void),
    OverflowInline(Void),
    OverflowX(<Overflow as ToComputedValue>::ComputedValue),
    OverflowY(<Overflow as ToComputedValue>::ComputedValue),
    BorderBlockEndStyle(Void),
    BorderBlockStartStyle(Void),
    BorderBottomStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderInlineEndStyle(Void),
    BorderInlineStartStyle(Void),
    BorderLeftStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderRightStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    BorderTopStyle(<BorderStyle as ToComputedValue>::ComputedValue),
    AnimationComposition(Void),
    AnimationDelay(Void),
    AnimationDirection(Void),
    AnimationDuration(Void),
    AnimationFillMode(Void),
    AnimationIterationCount(Void),
    AnimationName(Void),
    AnimationPlayState(Void),
    AnimationTimeline(Void),
    AnimationTimingFunction(Void),
    BackdropFilter(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BackgroundAttachment(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundClip(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundImage(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundOrigin(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundPositionX(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BackgroundPositionY(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BackgroundRepeat(<SpecifiedValue as ToComputedValue>::ComputedValue),
    BackgroundSize(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageOutset(<<NonNegativeLengthOrNumberRect as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageSlice(<<BorderImageSlice as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageWidth(<<BorderImageWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderSpacing(<<BorderSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BoxShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Clip(<<ClipRectOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ClipPath(<<ClipPath as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Color(<<ColorPropertyValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ColorScheme(<ColorScheme as ToComputedValue>::ComputedValue),
    ColumnWidth(<<NonNegativeLengthOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ContainerName(Void),
    Content(<Content as ToComputedValue>::ComputedValue),
    CounterIncrement(<CounterIncrement as ToComputedValue>::ComputedValue),
    CounterReset(<CounterReset as ToComputedValue>::ComputedValue),
    Cursor(<Cursor as ToComputedValue>::ComputedValue),
    Filter(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FlexBasis(<<FlexBasis as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontFamily(<FontFamily as ToComputedValue>::ComputedValue),
    FontSize(<<FontSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    FontVariationSettings(<<FontVariationSettings as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    GridTemplateAreas(<GridTemplateAreas as ToComputedValue>::ComputedValue),
    LetterSpacing(<<LetterSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    LineHeight(<<LineHeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MaskImage(<SpecifiedValue as ToComputedValue>::ComputedValue),
    OffsetPath(<<OffsetPath as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineOffset(<<BorderSideOffset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OverflowClipMargin(<<Length as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Perspective(<<Perspective as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Quotes(<Quotes as ToComputedValue>::ComputedValue),
    Rotate(<<Rotate as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Scale(<<Scale as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    TextIndent(<<TextIndent as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    TextOverflow(<TextOverflow as ToComputedValue>::ComputedValue),
    TextShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Transform(<<Transform as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    TransformOrigin(<<TransformOrigin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    TransitionBehavior(Void),
    TransitionDelay(Void),
    TransitionDuration(Void),
    TransitionProperty(Void),
    TransitionTimingFunction(Void),
    Translate(<<Translate as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    VerticalAlign(<<VerticalAlign as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    ViewTransitionClass(<ViewTransitionClass as ToComputedValue>::ComputedValue),
    ViewTransitionName(<ViewTransitionName as ToComputedValue>::ComputedValue),
    WillChange(Void),
    WordSpacing(<<WordSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    XLang(Void),
    ObjectPosition(<<Position as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PerspectiveOrigin(<<Position as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    GridTemplateColumns(<<GridTemplateComponent as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    GridTemplateRows(<<GridTemplateComponent as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderImageSource(<Image as ToComputedValue>::ComputedValue),
    ListStyleImage(<Image as ToComputedValue>::ComputedValue),
    GridAutoColumns(<ImplicitGridTracks as ToComputedValue>::ComputedValue),
    GridAutoRows(<ImplicitGridTracks as ToComputedValue>::ComputedValue),
    ColumnGap(<<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    RowGap(<<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    GridColumnEnd(<GridLine as ToComputedValue>::ComputedValue),
    GridColumnStart(<GridLine as ToComputedValue>::ComputedValue),
    GridRowEnd(<GridLine as ToComputedValue>::ComputedValue),
    GridRowStart(<GridLine as ToComputedValue>::ComputedValue),
    MaxBlockSize(Void),
    MaxHeight(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MaxInlineSize(Void),
    MaxWidth(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBottomLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBottomRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderEndEndRadius(Void),
    BorderEndStartRadius(Void),
    BorderStartEndRadius(Void),
    BorderStartStartRadius(Void),
    BorderTopLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Bottom(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    InsetBlockEnd(Void),
    InsetBlockStart(Void),
    InsetInlineEnd(Void),
    InsetInlineStart(Void),
    Left(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Right(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Top(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MarginBlockEnd(Void),
    MarginBlockStart(Void),
    MarginBottom(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MarginInlineEnd(Void),
    MarginInlineStart(Void),
    MarginLeft(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MarginRight(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MarginTop(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingBlockEnd(Void),
    PaddingBlockStart(Void),
    PaddingBottom(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingInlineEnd(Void),
    PaddingInlineStart(Void),
    PaddingLeft(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingRight(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    PaddingTop(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BlockSize(Void),
    Height(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    InlineSize(Void),
    MinBlockSize(Void),
    MinHeight(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    MinInlineSize(Void),
    MinWidth(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Width(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBlockEndWidth(Void),
    BorderBlockStartWidth(Void),
    BorderBottomWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderInlineEndWidth(Void),
    BorderInlineStartWidth(Void),
    BorderLeftWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderRightWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BackgroundColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderBlockEndColor(Void),
    BorderBlockStartColor(Void),
    BorderBottomColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderInlineEndColor(Void),
    BorderInlineStartColor(Void),
    BorderLeftColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderRightColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    BorderTopColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    OutlineColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    TextDecorationColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue),
    Custom(CustomAnimatedValue),

}

An enum to represent a single computed value belonging to an animated property in order to be interpolated with another one. When interpolating, both values need to belong to the same property.

Variants

AlignItems(<ItemPlacement as ToComputedValue>::ComputedValue)

align-items

AspectRatio(<<AspectRatio as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

aspect-ratio

BackfaceVisibility(<SpecifiedValue as ToComputedValue>::ComputedValue)

backface-visibility

BaselineSource(<BaselineSource as ToComputedValue>::ComputedValue)

baseline-source

BorderCollapse(<SpecifiedValue as ToComputedValue>::ComputedValue)

border-collapse

BorderImageRepeat(<BorderImageRepeat as ToComputedValue>::ComputedValue)

border-image-repeat

BoxSizing(<SpecifiedValue as ToComputedValue>::ComputedValue)

box-sizing

CaptionSide(<CaptionSide as ToComputedValue>::ComputedValue)

caption-side

Clear(<Clear as ToComputedValue>::ComputedValue)

clear

ColumnCount(<<ColumnCount as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

column-count

ColumnSpan(<SpecifiedValue as ToComputedValue>::ComputedValue)

column-span

Contain(Void)

contain

ContainerType(Void)

container-type

Direction(Void)

direction

Display(<Display as ToComputedValue>::ComputedValue)

display

EmptyCells(<SpecifiedValue as ToComputedValue>::ComputedValue)

empty-cells

FlexDirection(<SpecifiedValue as ToComputedValue>::ComputedValue)

flex-direction

FlexWrap(<SpecifiedValue as ToComputedValue>::ComputedValue)

flex-wrap

Float(<Float as ToComputedValue>::ComputedValue)

float

FontLanguageOverride(<FontLanguageOverride as ToComputedValue>::ComputedValue)

font-language-override

FontOpticalSizing(<SpecifiedValue as ToComputedValue>::ComputedValue)

font-optical-sizing

FontStretch(<<FontStretch as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-stretch

FontStyle(<<FontStyle as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-style

FontSynthesisWeight(<FontSynthesis as ToComputedValue>::ComputedValue)

font-synthesis-weight

FontVariantCaps(<SpecifiedValue as ToComputedValue>::ComputedValue)

font-variant-caps

FontWeight(<<FontWeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-weight

GridAutoFlow(<GridAutoFlow as ToComputedValue>::ComputedValue)

grid-auto-flow

ImageRendering(<ImageRendering as ToComputedValue>::ComputedValue)

image-rendering

Isolation(<SpecifiedValue as ToComputedValue>::ComputedValue)

isolation

JustifyItems(<JustifyItems as ToComputedValue>::ComputedValue)

justify-items

ListStylePosition(<SpecifiedValue as ToComputedValue>::ComputedValue)

list-style-position

ListStyleType(<SpecifiedValue as ToComputedValue>::ComputedValue)

list-style-type

MixBlendMode(<SpecifiedValue as ToComputedValue>::ComputedValue)

mix-blend-mode

ObjectFit(<SpecifiedValue as ToComputedValue>::ComputedValue)

object-fit

Opacity(<<Opacity as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

opacity

Order(<<Integer as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

order

OutlineStyle(<OutlineStyle as ToComputedValue>::ComputedValue)

outline-style

OverflowWrap(<OverflowWrap as ToComputedValue>::ComputedValue)

overflow-wrap

PointerEvents(<PointerEvents as ToComputedValue>::ComputedValue)

pointer-events

Position(<PositionProperty as ToComputedValue>::ComputedValue)

position

PositionArea(<PositionArea as ToComputedValue>::ComputedValue)

position-area

ServoOverflowClipBox(Void)

-servo-overflow-clip-box

ServoTopLayer(Void)

-servo-top-layer

TableLayout(<SpecifiedValue as ToComputedValue>::ComputedValue)

table-layout

TextAlign(<TextAlign as ToComputedValue>::ComputedValue)

text-align

TextAlignLast(<TextAlignLast as ToComputedValue>::ComputedValue)

text-align-last

TextDecorationLine(<TextDecorationLine as ToComputedValue>::ComputedValue)

text-decoration-line

TextDecorationStyle(<SpecifiedValue as ToComputedValue>::ComputedValue)

text-decoration-style

TextJustify(<TextJustify as ToComputedValue>::ComputedValue)

text-justify

TextRendering(<SpecifiedValue as ToComputedValue>::ComputedValue)

text-rendering

TextTransform(<TextTransform as ToComputedValue>::ComputedValue)

text-transform

TextWrapMode(<SpecifiedValue as ToComputedValue>::ComputedValue)

text-wrap-mode

TransformStyle(<TransformStyle as ToComputedValue>::ComputedValue)

transform-style

UnicodeBidi(Void)

unicode-bidi

Visibility(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

visibility

WhiteSpaceCollapse(<SpecifiedValue as ToComputedValue>::ComputedValue)

white-space-collapse

WordBreak(<WordBreak as ToComputedValue>::ComputedValue)

word-break

WritingMode(Void)

writing-mode

ZIndex(<<ZIndex as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

z-index

Zoom(<<Zoom as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

zoom

AlignContent(<ContentDistribution as ToComputedValue>::ComputedValue)

align-content

JustifyContent(<ContentDistribution as ToComputedValue>::ComputedValue)

justify-content

FlexGrow(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

flex-grow

FlexShrink(<<NonNegativeNumber as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

flex-shrink

AlignSelf(<SelfAlignment as ToComputedValue>::ComputedValue)

align-self

JustifySelf(<SelfAlignment as ToComputedValue>::ComputedValue)

justify-self

OverflowBlock(Void)

overflow-block

OverflowInline(Void)

overflow-inline

OverflowX(<Overflow as ToComputedValue>::ComputedValue)

overflow-x

OverflowY(<Overflow as ToComputedValue>::ComputedValue)

overflow-y

BorderBlockEndStyle(Void)

border-block-end-style

BorderBlockStartStyle(Void)

border-block-start-style

BorderBottomStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-bottom-style

BorderInlineEndStyle(Void)

border-inline-end-style

BorderInlineStartStyle(Void)

border-inline-start-style

BorderLeftStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-left-style

BorderRightStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-right-style

BorderTopStyle(<BorderStyle as ToComputedValue>::ComputedValue)

border-top-style

AnimationComposition(Void)

animation-composition

AnimationDelay(Void)

animation-delay

AnimationDirection(Void)

animation-direction

AnimationDuration(Void)

animation-duration

AnimationFillMode(Void)

animation-fill-mode

AnimationIterationCount(Void)

animation-iteration-count

AnimationName(Void)

animation-name

AnimationPlayState(Void)

animation-play-state

AnimationTimeline(Void)

animation-timeline

AnimationTimingFunction(Void)

animation-timing-function

BackdropFilter(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

backdrop-filter

BackgroundAttachment(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-attachment

BackgroundClip(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-clip

BackgroundImage(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-image

BackgroundOrigin(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-origin

BackgroundPositionX(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

background-position-x

BackgroundPositionY(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

background-position-y

BackgroundRepeat(<SpecifiedValue as ToComputedValue>::ComputedValue)

background-repeat

BackgroundSize(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

background-size

BorderImageOutset(<<NonNegativeLengthOrNumberRect as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-image-outset

BorderImageSlice(<<BorderImageSlice as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-image-slice

BorderImageWidth(<<BorderImageWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-image-width

BorderSpacing(<<BorderSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-spacing

BoxShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

box-shadow

Clip(<<ClipRectOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

clip

ClipPath(<<ClipPath as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

clip-path

Color(<<ColorPropertyValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

color

ColorScheme(<ColorScheme as ToComputedValue>::ComputedValue)

color-scheme

ColumnWidth(<<NonNegativeLengthOrAuto as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

column-width

ContainerName(Void)

container-name

Content(<Content as ToComputedValue>::ComputedValue)

content

CounterIncrement(<CounterIncrement as ToComputedValue>::ComputedValue)

counter-increment

CounterReset(<CounterReset as ToComputedValue>::ComputedValue)

counter-reset

Cursor(<Cursor as ToComputedValue>::ComputedValue)

cursor

Filter(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

filter

FlexBasis(<<FlexBasis as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

flex-basis

FontFamily(<FontFamily as ToComputedValue>::ComputedValue)

font-family

FontSize(<<FontSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-size

FontVariationSettings(<<FontVariationSettings as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

font-variation-settings

GridTemplateAreas(<GridTemplateAreas as ToComputedValue>::ComputedValue)

grid-template-areas

LetterSpacing(<<LetterSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

letter-spacing

LineHeight(<<LineHeight as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

line-height

MaskImage(<SpecifiedValue as ToComputedValue>::ComputedValue)

mask-image

OffsetPath(<<OffsetPath as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

offset-path

OutlineOffset(<<BorderSideOffset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

outline-offset

OverflowClipMargin(<<Length as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

overflow-clip-margin

Perspective(<<Perspective as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

perspective

Quotes(<Quotes as ToComputedValue>::ComputedValue)

quotes

Rotate(<<Rotate as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

rotate

Scale(<<Scale as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

scale

TextIndent(<<TextIndent as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

text-indent

TextOverflow(<TextOverflow as ToComputedValue>::ComputedValue)

text-overflow

TextShadow(<<SpecifiedValue as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

text-shadow

Transform(<<Transform as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

transform

TransformOrigin(<<TransformOrigin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

transform-origin

TransitionBehavior(Void)

transition-behavior

TransitionDelay(Void)

transition-delay

TransitionDuration(Void)

transition-duration

TransitionProperty(Void)

transition-property

TransitionTimingFunction(Void)

transition-timing-function

Translate(<<Translate as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

translate

VerticalAlign(<<VerticalAlign as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

vertical-align

ViewTransitionClass(<ViewTransitionClass as ToComputedValue>::ComputedValue)

view-transition-class

ViewTransitionName(<ViewTransitionName as ToComputedValue>::ComputedValue)

view-transition-name

WillChange(Void)

will-change

WordSpacing(<<WordSpacing as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

word-spacing

XLang(Void)

-x-lang

ObjectPosition(<<Position as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

object-position

PerspectiveOrigin(<<Position as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

perspective-origin

GridTemplateColumns(<<GridTemplateComponent as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

grid-template-columns

GridTemplateRows(<<GridTemplateComponent as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

grid-template-rows

BorderImageSource(<Image as ToComputedValue>::ComputedValue)

border-image-source

ListStyleImage(<Image as ToComputedValue>::ComputedValue)

list-style-image

GridAutoColumns(<ImplicitGridTracks as ToComputedValue>::ComputedValue)

grid-auto-columns

GridAutoRows(<ImplicitGridTracks as ToComputedValue>::ComputedValue)

grid-auto-rows

ColumnGap(<<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

column-gap

RowGap(<<NonNegativeLengthPercentageOrNormal as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

row-gap

GridColumnEnd(<GridLine as ToComputedValue>::ComputedValue)

grid-column-end

GridColumnStart(<GridLine as ToComputedValue>::ComputedValue)

grid-column-start

GridRowEnd(<GridLine as ToComputedValue>::ComputedValue)

grid-row-end

GridRowStart(<GridLine as ToComputedValue>::ComputedValue)

grid-row-start

MaxBlockSize(Void)

max-block-size

MaxHeight(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

max-height

MaxInlineSize(Void)

max-inline-size

MaxWidth(<<MaxSize as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

max-width

BorderBottomLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-left-radius

BorderBottomRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-right-radius

BorderEndEndRadius(Void)

border-end-end-radius

BorderEndStartRadius(Void)

border-end-start-radius

BorderStartEndRadius(Void)

border-start-end-radius

BorderStartStartRadius(Void)

border-start-start-radius

BorderTopLeftRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-left-radius

BorderTopRightRadius(<<BorderCornerRadius as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-right-radius

Bottom(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

bottom

InsetBlockEnd(Void)

inset-block-end

InsetBlockStart(Void)

inset-block-start

InsetInlineEnd(Void)

inset-inline-end

InsetInlineStart(Void)

inset-inline-start

Left(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

left

Right(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

right

Top(<<Inset as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

top

MarginBlockEnd(Void)

margin-block-end

MarginBlockStart(Void)

margin-block-start

MarginBottom(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

margin-bottom

MarginInlineEnd(Void)

margin-inline-end

MarginInlineStart(Void)

margin-inline-start

MarginLeft(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

margin-left

MarginRight(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

margin-right

MarginTop(<<Margin as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

margin-top

PaddingBlockEnd(Void)

padding-block-end

PaddingBlockStart(Void)

padding-block-start

PaddingBottom(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-bottom

PaddingInlineEnd(Void)

padding-inline-end

PaddingInlineStart(Void)

padding-inline-start

PaddingLeft(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-left

PaddingRight(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-right

PaddingTop(<<NonNegativeLengthPercentage as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

padding-top

BlockSize(Void)

block-size

Height(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

height

InlineSize(Void)

inline-size

MinBlockSize(Void)

min-block-size

MinHeight(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

min-height

MinInlineSize(Void)

min-inline-size

MinWidth(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

min-width

Width(<<Size as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

width

BorderBlockEndWidth(Void)

border-block-end-width

BorderBlockStartWidth(Void)

border-block-start-width

BorderBottomWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-width

BorderInlineEndWidth(Void)

border-inline-end-width

BorderInlineStartWidth(Void)

border-inline-start-width

BorderLeftWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-left-width

BorderRightWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-right-width

BorderTopWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-width

OutlineWidth(<<BorderSideWidth as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

outline-width

BackgroundColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

background-color

BorderBlockEndColor(Void)

border-block-end-color

BorderBlockStartColor(Void)

border-block-start-color

BorderBottomColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-bottom-color

BorderInlineEndColor(Void)

border-inline-end-color

BorderInlineStartColor(Void)

border-inline-start-color

BorderLeftColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-left-color

BorderRightColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-right-color

BorderTopColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

border-top-color

OutlineColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

outline-color

TextDecorationColor(<<Color as ToComputedValue>::ComputedValue as ToAnimatedValue>::AnimatedValue)

text-decoration-color

Custom(CustomAnimatedValue)

A custom property.

Implementations

impl AnimationValue

pub fn id(&self) -> PropertyDeclarationId<'_>

Returns the longhand id this animated value corresponds to.

pub fn interpolable_with(&self, other: &Self) -> bool

Returns whether this value is interpolable with another one.

pub fn uncompute(&self) -> PropertyDeclaration

Uncompute this animation value in order to be used inside the CSS cascade.

pub fn from_declaration( decl: &PropertyDeclaration, context: &mut Context<'_>, style: &ComputedValues, initial: &ComputedValues, ) -> Option<Self>

Construct an AnimationValue from a property declaration.

pub fn from_computed_values( property: PropertyDeclarationId<'_>, style: &ComputedValues, ) -> Option<Self>

Get an AnimationValue for an declaration id from a given computed values.

pub fn set_in_style_for_servo(&self, style: &mut ComputedValues)

Update style with the value of this AnimationValue.

SERVO ONLY: This doesnt properly handle things like updating em units when animated font-size.