Enum style::properties::generated::CountedUnknownProperty

#[repr(u32)]
pub enum CountedUnknownProperty {

    WebkitFontSmoothing,
    WebkitTapHighlightColor,
    Speak,
    TextSizeAdjust,
    WebkitFontFeatureSettings,
    WebkitUserDrag,
    Orphans,
    Widows,
    WebkitUserModify,
    WebkitMarginBefore,
    WebkitMarginAfter,
    WebkitMarginStart,
    WebkitColumnBreakInside,
    WebkitPaddingStart,
    WebkitMarginEnd,
    WebkitBoxReflect,
    WebkitPrintColorAdjust,
    WebkitMaskBoxImage,
    WebkitLineBreak,
    AlignmentBaseline,
    WebkitWritingMode,
    BaselineShift,
    WebkitHyphenateCharacter,
    WebkitHighlight,
    BackgroundRepeatX,
    WebkitPaddingEnd,
    BackgroundRepeatY,
    WebkitTextEmphasisColor,
    WebkitMarginTopCollapse,
    WebkitRtlOrdering,
    WebkitPaddingBefore,
    WebkitTextDecorationsInEffect,
    WebkitBorderVerticalSpacing,
    WebkitLocale,
    WebkitPaddingAfter,
    WebkitBorderHorizontalSpacing,
    ColorRendering,
    WebkitColumnBreakBefore,
    WebkitTransformOriginX,
    WebkitTransformOriginY,
    WebkitTextEmphasisPosition,
    BufferedRendering,
    WebkitTextOrientation,
    WebkitTextCombine,
    WebkitTextEmphasisStyle,
    WebkitTextEmphasis,
    WebkitMaskBoxImageWidth,
    WebkitMaskBoxImageSource,
    WebkitMaskBoxImageOutset,
    WebkitMaskBoxImageSlice,
    WebkitMaskBoxImageRepeat,
    WebkitMarginAfterCollapse,
    WebkitBorderBeforeColor,
    WebkitBorderBeforeWidth,
    WebkitPerspectiveOriginX,
    WebkitPerspectiveOriginY,
    WebkitMarginBeforeCollapse,
    WebkitBorderBeforeStyle,
    WebkitMarginBottomCollapse,
    WebkitRubyPosition,
    WebkitColumnBreakAfter,
    WebkitMarginCollapse,
    WebkitBorderBefore,
    WebkitBorderEnd,
    WebkitBorderAfter,
    WebkitBorderStart,
    WebkitMinLogicalWidth,
    WebkitLogicalHeight,
    WebkitTransformOriginZ,
    WebkitFontSizeDelta,
    WebkitLogicalWidth,
    WebkitMaxLogicalWidth,
    WebkitMinLogicalHeight,
    WebkitMaxLogicalHeight,
    WebkitBorderEndColor,
    WebkitBorderEndWidth,
    WebkitBorderStartColor,
    WebkitBorderStartWidth,
    WebkitBorderAfterColor,
    WebkitBorderAfterWidth,
    WebkitBorderEndStyle,
    WebkitBorderAfterStyle,
    WebkitBorderStartStyle,
    WebkitMaskRepeatX,
    WebkitMaskRepeatY,
    UserZoom,
    MinZoom,
    WebkitBoxDecorationBreak,
    Orientation,
    MaxZoom,
    WebkitAppRegion,
    WebkitColumnRule,
    WebkitColumnSpan,
    WebkitColumnGap,
    WebkitShapeOutside,
    WebkitColumnRuleWidth,
    WebkitColumnCount,
    WebkitOpacity,
    WebkitColumnWidth,
    WebkitShapeImageThreshold,
    WebkitColumnRuleStyle,
    WebkitColumns,
    WebkitColumnRuleColor,
    WebkitShapeMargin,
    Zoom,

}

The counted unknown property list which is used for css use counters.

FIXME: This should be just #[repr(u8)], but can’t be because of ABI issues, see https://bugs.llvm.org/show_bug.cgi?id=44228.

Variants

WebkitFontSmoothing

-webkit-font-smoothing

WebkitTapHighlightColor

-webkit-tap-highlight-color

Speak

speak

TextSizeAdjust

text-size-adjust

WebkitFontFeatureSettings

-webkit-font-feature-settings

WebkitUserDrag

-webkit-user-drag

Orphans

orphans

Widows

widows

WebkitUserModify

-webkit-user-modify

WebkitMarginBefore

-webkit-margin-before

WebkitMarginAfter

-webkit-margin-after

WebkitMarginStart

-webkit-margin-start

WebkitColumnBreakInside

-webkit-column-break-inside

WebkitPaddingStart

-webkit-padding-start

WebkitMarginEnd

-webkit-margin-end

WebkitBoxReflect

-webkit-box-reflect

WebkitPrintColorAdjust

-webkit-print-color-adjust

WebkitMaskBoxImage

-webkit-mask-box-image

WebkitLineBreak

-webkit-line-break

AlignmentBaseline

alignment-baseline

WebkitWritingMode

-webkit-writing-mode

BaselineShift

baseline-shift

WebkitHyphenateCharacter

-webkit-hyphenate-character

WebkitHighlight

-webkit-highlight

BackgroundRepeatX

background-repeat-x

WebkitPaddingEnd

-webkit-padding-end

BackgroundRepeatY

background-repeat-y

WebkitTextEmphasisColor

-webkit-text-emphasis-color

WebkitMarginTopCollapse

-webkit-margin-top-collapse

WebkitRtlOrdering

-webkit-rtl-ordering

WebkitPaddingBefore

-webkit-padding-before

WebkitTextDecorationsInEffect

-webkit-text-decorations-in-effect

WebkitBorderVerticalSpacing

-webkit-border-vertical-spacing

WebkitLocale

-webkit-locale

WebkitPaddingAfter

-webkit-padding-after

WebkitBorderHorizontalSpacing

-webkit-border-horizontal-spacing

ColorRendering

color-rendering

WebkitColumnBreakBefore

-webkit-column-break-before

WebkitTransformOriginX

-webkit-transform-origin-x

WebkitTransformOriginY

-webkit-transform-origin-y

WebkitTextEmphasisPosition

-webkit-text-emphasis-position

BufferedRendering

buffered-rendering

WebkitTextOrientation

-webkit-text-orientation

WebkitTextCombine

-webkit-text-combine

WebkitTextEmphasisStyle

-webkit-text-emphasis-style

WebkitTextEmphasis

-webkit-text-emphasis

WebkitMaskBoxImageWidth

-webkit-mask-box-image-width

WebkitMaskBoxImageSource

-webkit-mask-box-image-source

WebkitMaskBoxImageOutset

-webkit-mask-box-image-outset

WebkitMaskBoxImageSlice

-webkit-mask-box-image-slice

WebkitMaskBoxImageRepeat

-webkit-mask-box-image-repeat

WebkitMarginAfterCollapse

-webkit-margin-after-collapse

WebkitBorderBeforeColor

-webkit-border-before-color

WebkitBorderBeforeWidth

-webkit-border-before-width

WebkitPerspectiveOriginX

-webkit-perspective-origin-x

WebkitPerspectiveOriginY

-webkit-perspective-origin-y

WebkitMarginBeforeCollapse

-webkit-margin-before-collapse

WebkitBorderBeforeStyle

-webkit-border-before-style

WebkitMarginBottomCollapse

-webkit-margin-bottom-collapse

WebkitRubyPosition

-webkit-ruby-position

WebkitColumnBreakAfter

-webkit-column-break-after

WebkitMarginCollapse

-webkit-margin-collapse

WebkitBorderBefore

-webkit-border-before

WebkitBorderEnd

-webkit-border-end

WebkitBorderAfter

-webkit-border-after

WebkitBorderStart

-webkit-border-start

WebkitMinLogicalWidth

-webkit-min-logical-width

WebkitLogicalHeight

-webkit-logical-height

WebkitTransformOriginZ

-webkit-transform-origin-z

WebkitFontSizeDelta

-webkit-font-size-delta

WebkitLogicalWidth

-webkit-logical-width

WebkitMaxLogicalWidth

-webkit-max-logical-width

WebkitMinLogicalHeight

-webkit-min-logical-height

WebkitMaxLogicalHeight

-webkit-max-logical-height

WebkitBorderEndColor

-webkit-border-end-color

WebkitBorderEndWidth

-webkit-border-end-width

WebkitBorderStartColor

-webkit-border-start-color

WebkitBorderStartWidth

-webkit-border-start-width

WebkitBorderAfterColor

-webkit-border-after-color

WebkitBorderAfterWidth

-webkit-border-after-width

WebkitBorderEndStyle

-webkit-border-end-style

WebkitBorderAfterStyle

-webkit-border-after-style

WebkitBorderStartStyle

-webkit-border-start-style

WebkitMaskRepeatX

-webkit-mask-repeat-x

WebkitMaskRepeatY

-webkit-mask-repeat-y

UserZoom

user-zoom

MinZoom

min-zoom

WebkitBoxDecorationBreak

-webkit-box-decoration-break

Orientation

orientation

MaxZoom

max-zoom

WebkitAppRegion

-webkit-app-region

WebkitColumnRule

-webkit-column-rule

WebkitColumnSpan

-webkit-column-span

WebkitColumnGap

-webkit-column-gap

WebkitShapeOutside

-webkit-shape-outside

WebkitColumnRuleWidth

-webkit-column-rule-width

WebkitColumnCount

-webkit-column-count

WebkitOpacity

-webkit-opacity

WebkitColumnWidth

-webkit-column-width

WebkitShapeImageThreshold

-webkit-shape-image-threshold

WebkitColumnRuleStyle

-webkit-column-rule-style

WebkitColumns

-webkit-columns

WebkitColumnRuleColor

-webkit-column-rule-color

WebkitShapeMargin

-webkit-shape-margin

Zoom

zoom

Implementations

impl CountedUnknownProperty

pub fn parse_for_testing(property_name: &str) -> Option<Self>

Parse the counted unknown property, for testing purposes only.

pub fn bit(self) -> usize

Returns the underlying index, used for use counter.