Enum style::properties::generated::CountedUnknownProperty

#[repr(u32)]
pub enum CountedUnknownProperty {

    WebkitTapHighlightColor = 0,
    Speak = 1,
    TextSizeAdjust = 2,
    WebkitUserDrag = 3,
    Orphans = 4,
    Widows = 5,
    WebkitUserModify = 6,
    WebkitMarginBefore = 7,
    WebkitMarginAfter = 8,
    WebkitMarginStart = 9,
    WebkitColumnBreakInside = 10,
    WebkitPaddingStart = 11,
    WebkitMarginEnd = 12,
    WebkitBoxReflect = 13,
    WebkitPrintColorAdjust = 14,
    WebkitMaskBoxImage = 15,
    WebkitLineBreak = 16,
    AlignmentBaseline = 17,
    WebkitWritingMode = 18,
    BaselineShift = 19,
    WebkitHyphenateCharacter = 20,
    WebkitHighlight = 21,
    BackgroundRepeatX = 22,
    WebkitPaddingEnd = 23,
    BackgroundRepeatY = 24,
    WebkitTextEmphasisColor = 25,
    WebkitMarginTopCollapse = 26,
    WebkitRtlOrdering = 27,
    WebkitPaddingBefore = 28,
    WebkitTextDecorationsInEffect = 29,
    WebkitBorderVerticalSpacing = 30,
    WebkitLocale = 31,
    WebkitPaddingAfter = 32,
    WebkitBorderHorizontalSpacing = 33,
    ColorRendering = 34,
    WebkitColumnBreakBefore = 35,
    WebkitTransformOriginX = 36,
    WebkitTransformOriginY = 37,
    WebkitTextEmphasisPosition = 38,
    BufferedRendering = 39,
    WebkitTextOrientation = 40,
    WebkitTextCombine = 41,
    WebkitTextEmphasisStyle = 42,
    WebkitTextEmphasis = 43,
    WebkitMaskBoxImageWidth = 44,
    WebkitMaskBoxImageSource = 45,
    WebkitMaskBoxImageOutset = 46,
    WebkitMaskBoxImageSlice = 47,
    WebkitMaskBoxImageRepeat = 48,
    WebkitMarginAfterCollapse = 49,
    WebkitBorderBeforeColor = 50,
    WebkitBorderBeforeWidth = 51,
    WebkitPerspectiveOriginX = 52,
    WebkitPerspectiveOriginY = 53,
    WebkitMarginBeforeCollapse = 54,
    WebkitBorderBeforeStyle = 55,
    WebkitMarginBottomCollapse = 56,
    WebkitRubyPosition = 57,
    WebkitColumnBreakAfter = 58,
    WebkitMarginCollapse = 59,
    WebkitBorderBefore = 60,
    WebkitBorderEnd = 61,
    WebkitBorderAfter = 62,
    WebkitBorderStart = 63,
    WebkitMinLogicalWidth = 64,
    WebkitLogicalHeight = 65,
    WebkitTransformOriginZ = 66,
    WebkitFontSizeDelta = 67,
    WebkitLogicalWidth = 68,
    WebkitMaxLogicalWidth = 69,
    WebkitMinLogicalHeight = 70,
    WebkitMaxLogicalHeight = 71,
    WebkitBorderEndColor = 72,
    WebkitBorderEndWidth = 73,
    WebkitBorderStartColor = 74,
    WebkitBorderStartWidth = 75,
    WebkitBorderAfterColor = 76,
    WebkitBorderAfterWidth = 77,
    WebkitBorderEndStyle = 78,
    WebkitBorderAfterStyle = 79,
    WebkitBorderStartStyle = 80,
    WebkitMaskRepeatX = 81,
    WebkitMaskRepeatY = 82,
    UserZoom = 83,
    MinZoom = 84,
    WebkitBoxDecorationBreak = 85,
    Orientation = 86,
    MaxZoom = 87,
    WebkitAppRegion = 88,
    WebkitColumnRule = 89,
    WebkitColumnSpan = 90,
    WebkitColumnGap = 91,
    WebkitShapeOutside = 92,
    WebkitColumnRuleWidth = 93,
    WebkitColumnCount = 94,
    WebkitOpacity = 95,
    WebkitColumnWidth = 96,
    WebkitShapeImageThreshold = 97,
    WebkitColumnRuleStyle = 98,
    WebkitColumns = 99,
    WebkitColumnRuleColor = 100,
    WebkitShapeMargin = 101,

}

The counted unknown property list which is used for css use counters.

FIXME: This should be just #[repr(u8)], but can’t be because of ABI issues, see https://bugs.llvm.org/show_bug.cgi?id=44228.

Variants

WebkitTapHighlightColor = 0

-webkit-tap-highlight-color

Speak = 1

speak

TextSizeAdjust = 2

text-size-adjust

WebkitUserDrag = 3

-webkit-user-drag

Orphans = 4

orphans

Widows = 5

widows

WebkitUserModify = 6

-webkit-user-modify

WebkitMarginBefore = 7

-webkit-margin-before

WebkitMarginAfter = 8

-webkit-margin-after

WebkitMarginStart = 9

-webkit-margin-start

WebkitColumnBreakInside = 10

-webkit-column-break-inside

WebkitPaddingStart = 11

-webkit-padding-start

WebkitMarginEnd = 12

-webkit-margin-end

WebkitBoxReflect = 13

-webkit-box-reflect

WebkitPrintColorAdjust = 14

-webkit-print-color-adjust

WebkitMaskBoxImage = 15

-webkit-mask-box-image

WebkitLineBreak = 16

-webkit-line-break

AlignmentBaseline = 17

alignment-baseline

WebkitWritingMode = 18

-webkit-writing-mode

BaselineShift = 19

baseline-shift

WebkitHyphenateCharacter = 20

-webkit-hyphenate-character

WebkitHighlight = 21

-webkit-highlight

BackgroundRepeatX = 22

background-repeat-x

WebkitPaddingEnd = 23

-webkit-padding-end

BackgroundRepeatY = 24

background-repeat-y

WebkitTextEmphasisColor = 25

-webkit-text-emphasis-color

WebkitMarginTopCollapse = 26

-webkit-margin-top-collapse

WebkitRtlOrdering = 27

-webkit-rtl-ordering

WebkitPaddingBefore = 28

-webkit-padding-before

WebkitTextDecorationsInEffect = 29

-webkit-text-decorations-in-effect

WebkitBorderVerticalSpacing = 30

-webkit-border-vertical-spacing

WebkitLocale = 31

-webkit-locale

WebkitPaddingAfter = 32

-webkit-padding-after

WebkitBorderHorizontalSpacing = 33

-webkit-border-horizontal-spacing

ColorRendering = 34

color-rendering

WebkitColumnBreakBefore = 35

-webkit-column-break-before

WebkitTransformOriginX = 36

-webkit-transform-origin-x

WebkitTransformOriginY = 37

-webkit-transform-origin-y

WebkitTextEmphasisPosition = 38

-webkit-text-emphasis-position

BufferedRendering = 39

buffered-rendering

WebkitTextOrientation = 40

-webkit-text-orientation

WebkitTextCombine = 41

-webkit-text-combine

WebkitTextEmphasisStyle = 42

-webkit-text-emphasis-style

WebkitTextEmphasis = 43

-webkit-text-emphasis

WebkitMaskBoxImageWidth = 44

-webkit-mask-box-image-width

WebkitMaskBoxImageSource = 45

-webkit-mask-box-image-source

WebkitMaskBoxImageOutset = 46

-webkit-mask-box-image-outset

WebkitMaskBoxImageSlice = 47

-webkit-mask-box-image-slice

WebkitMaskBoxImageRepeat = 48

-webkit-mask-box-image-repeat

WebkitMarginAfterCollapse = 49

-webkit-margin-after-collapse

WebkitBorderBeforeColor = 50

-webkit-border-before-color

WebkitBorderBeforeWidth = 51

-webkit-border-before-width

WebkitPerspectiveOriginX = 52

-webkit-perspective-origin-x

WebkitPerspectiveOriginY = 53

-webkit-perspective-origin-y

WebkitMarginBeforeCollapse = 54

-webkit-margin-before-collapse

WebkitBorderBeforeStyle = 55

-webkit-border-before-style

WebkitMarginBottomCollapse = 56

-webkit-margin-bottom-collapse

WebkitRubyPosition = 57

-webkit-ruby-position

WebkitColumnBreakAfter = 58

-webkit-column-break-after

WebkitMarginCollapse = 59

-webkit-margin-collapse

WebkitBorderBefore = 60

-webkit-border-before

WebkitBorderEnd = 61

-webkit-border-end

WebkitBorderAfter = 62

-webkit-border-after

WebkitBorderStart = 63

-webkit-border-start

WebkitMinLogicalWidth = 64

-webkit-min-logical-width

WebkitLogicalHeight = 65

-webkit-logical-height

WebkitTransformOriginZ = 66

-webkit-transform-origin-z

WebkitFontSizeDelta = 67

-webkit-font-size-delta

WebkitLogicalWidth = 68

-webkit-logical-width

WebkitMaxLogicalWidth = 69

-webkit-max-logical-width

WebkitMinLogicalHeight = 70

-webkit-min-logical-height

WebkitMaxLogicalHeight = 71

-webkit-max-logical-height

WebkitBorderEndColor = 72

-webkit-border-end-color

WebkitBorderEndWidth = 73

-webkit-border-end-width

WebkitBorderStartColor = 74

-webkit-border-start-color

WebkitBorderStartWidth = 75

-webkit-border-start-width

WebkitBorderAfterColor = 76

-webkit-border-after-color

WebkitBorderAfterWidth = 77

-webkit-border-after-width

WebkitBorderEndStyle = 78

-webkit-border-end-style

WebkitBorderAfterStyle = 79

-webkit-border-after-style

WebkitBorderStartStyle = 80

-webkit-border-start-style

WebkitMaskRepeatX = 81

-webkit-mask-repeat-x

WebkitMaskRepeatY = 82

-webkit-mask-repeat-y

UserZoom = 83

user-zoom

MinZoom = 84

min-zoom

WebkitBoxDecorationBreak = 85

-webkit-box-decoration-break

Orientation = 86

orientation

MaxZoom = 87

max-zoom

WebkitAppRegion = 88

-webkit-app-region

WebkitColumnRule = 89

-webkit-column-rule

WebkitColumnSpan = 90

-webkit-column-span

WebkitColumnGap = 91

-webkit-column-gap

WebkitShapeOutside = 92

-webkit-shape-outside

WebkitColumnRuleWidth = 93

-webkit-column-rule-width

WebkitColumnCount = 94

-webkit-column-count

WebkitOpacity = 95

-webkit-opacity

WebkitColumnWidth = 96

-webkit-column-width

WebkitShapeImageThreshold = 97

-webkit-shape-image-threshold

WebkitColumnRuleStyle = 98

-webkit-column-rule-style

WebkitColumns = 99

-webkit-columns

WebkitColumnRuleColor = 100

-webkit-column-rule-color

WebkitShapeMargin = 101

-webkit-shape-margin

Implementations

impl CountedUnknownProperty

pub fn parse_for_testing(property_name: &str) -> Option<Self>

Parse the counted unknown property, for testing purposes only.

pub fn bit(self) -> usize

Returns the underlying index, used for use counter.