style::properties::generated

Enum LonghandId

#[repr(u16)]
pub enum LonghandId {

    AlignItems = 0,
    AspectRatio = 1,
    BackfaceVisibility = 2,
    BaselineSource = 3,
    BorderCollapse = 4,
    BorderImageRepeat = 5,
    BoxSizing = 6,
    CaptionSide = 7,
    Clear = 8,
    ColumnCount = 9,
    ColumnSpan = 10,
    Contain = 11,
    ContainerType = 12,
    Direction = 13,
    Display = 14,
    EmptyCells = 15,
    FlexDirection = 16,
    FlexWrap = 17,
    Float = 18,
    FontLanguageOverride = 19,
    FontOpticalSizing = 20,
    FontStretch = 21,
    FontStyle = 22,
    FontSynthesisWeight = 23,
    FontVariantCaps = 24,
    FontWeight = 25,
    GridAutoFlow = 26,
    ImageRendering = 27,
    Isolation = 28,
    JustifyItems = 29,
    ListStylePosition = 30,
    ListStyleType = 31,
    MixBlendMode = 32,
    ObjectFit = 33,
    Opacity = 34,
    Order = 35,
    OutlineStyle = 36,
    OverflowWrap = 37,
    PointerEvents = 38,
    Position = 39,
    PositionArea = 40,
    ServoOverflowClipBox = 41,
    ServoTopLayer = 42,
    TableLayout = 43,
    TextAlign = 44,
    TextAlignLast = 45,
    TextDecorationLine = 46,
    TextDecorationStyle = 47,
    TextJustify = 48,
    TextRendering = 49,
    TextTransform = 50,
    TextWrapMode = 51,
    TransformStyle = 52,
    UnicodeBidi = 53,
    Visibility = 54,
    WhiteSpaceCollapse = 55,
    WordBreak = 56,
    WritingMode = 57,
    ZIndex = 58,
    Zoom = 59,
    AlignContent = 60,
    JustifyContent = 61,
    FlexGrow = 62,
    FlexShrink = 63,
    AlignSelf = 64,
    JustifySelf = 65,
    OverflowBlock = 66,
    OverflowInline = 67,
    OverflowX = 68,
    OverflowY = 69,
    BorderBlockEndStyle = 70,
    BorderBlockStartStyle = 71,
    BorderBottomStyle = 72,
    BorderInlineEndStyle = 73,
    BorderInlineStartStyle = 74,
    BorderLeftStyle = 75,
    BorderRightStyle = 76,
    BorderTopStyle = 77,
    AnimationComposition = 78,
    AnimationDelay = 79,
    AnimationDirection = 80,
    AnimationDuration = 81,
    AnimationFillMode = 82,
    AnimationIterationCount = 83,
    AnimationName = 84,
    AnimationPlayState = 85,
    AnimationTimeline = 86,
    AnimationTimingFunction = 87,
    BackdropFilter = 88,
    BackgroundAttachment = 89,
    BackgroundClip = 90,
    BackgroundImage = 91,
    BackgroundOrigin = 92,
    BackgroundPositionX = 93,
    BackgroundPositionY = 94,
    BackgroundRepeat = 95,
    BackgroundSize = 96,
    BorderImageOutset = 97,
    BorderImageSlice = 98,
    BorderImageWidth = 99,
    BorderSpacing = 100,
    BoxShadow = 101,
    Clip = 102,
    ClipPath = 103,
    Color = 104,
    ColorScheme = 105,
    ColumnWidth = 106,
    ContainerName = 107,
    Content = 108,
    CounterIncrement = 109,
    CounterReset = 110,
    Cursor = 111,
    Filter = 112,
    FlexBasis = 113,
    FontFamily = 114,
    FontSize = 115,
    FontVariationSettings = 116,
    GridTemplateAreas = 117,
    LetterSpacing = 118,
    LineHeight = 119,
    MaskImage = 120,
    OffsetPath = 121,
    OutlineOffset = 122,
    OverflowClipMargin = 123,
    Perspective = 124,
    Quotes = 125,
    Rotate = 126,
    Scale = 127,
    TextIndent = 128,
    TextOverflow = 129,
    TextShadow = 130,
    Transform = 131,
    TransformOrigin = 132,
    TransitionBehavior = 133,
    TransitionDelay = 134,
    TransitionDuration = 135,
    TransitionProperty = 136,
    TransitionTimingFunction = 137,
    Translate = 138,
    VerticalAlign = 139,
    ViewTransitionClass = 140,
    ViewTransitionName = 141,
    WillChange = 142,
    WordSpacing = 143,
    XLang = 144,
    ObjectPosition = 145,
    PerspectiveOrigin = 146,
    GridTemplateColumns = 147,
    GridTemplateRows = 148,
    BorderImageSource = 149,
    ListStyleImage = 150,
    GridAutoColumns = 151,
    GridAutoRows = 152,
    ColumnGap = 153,
    RowGap = 154,
    GridColumnEnd = 155,
    GridColumnStart = 156,
    GridRowEnd = 157,
    GridRowStart = 158,
    MaxBlockSize = 159,
    MaxHeight = 160,
    MaxInlineSize = 161,
    MaxWidth = 162,
    BorderBottomLeftRadius = 163,
    BorderBottomRightRadius = 164,
    BorderEndEndRadius = 165,
    BorderEndStartRadius = 166,
    BorderStartEndRadius = 167,
    BorderStartStartRadius = 168,
    BorderTopLeftRadius = 169,
    BorderTopRightRadius = 170,
    Bottom = 171,
    InsetBlockEnd = 172,
    InsetBlockStart = 173,
    InsetInlineEnd = 174,
    InsetInlineStart = 175,
    Left = 176,
    Right = 177,
    Top = 178,
    MarginBlockEnd = 179,
    MarginBlockStart = 180,
    MarginBottom = 181,
    MarginInlineEnd = 182,
    MarginInlineStart = 183,
    MarginLeft = 184,
    MarginRight = 185,
    MarginTop = 186,
    PaddingBlockEnd = 187,
    PaddingBlockStart = 188,
    PaddingBottom = 189,
    PaddingInlineEnd = 190,
    PaddingInlineStart = 191,
    PaddingLeft = 192,
    PaddingRight = 193,
    PaddingTop = 194,
    BlockSize = 195,
    Height = 196,
    InlineSize = 197,
    MinBlockSize = 198,
    MinHeight = 199,
    MinInlineSize = 200,
    MinWidth = 201,
    Width = 202,
    BorderBlockEndWidth = 203,
    BorderBlockStartWidth = 204,
    BorderBottomWidth = 205,
    BorderInlineEndWidth = 206,
    BorderInlineStartWidth = 207,
    BorderLeftWidth = 208,
    BorderRightWidth = 209,
    BorderTopWidth = 210,
    OutlineWidth = 211,
    BackgroundColor = 212,
    BorderBlockEndColor = 213,
    BorderBlockStartColor = 214,
    BorderBottomColor = 215,
    BorderInlineEndColor = 216,
    BorderInlineStartColor = 217,
    BorderLeftColor = 218,
    BorderRightColor = 219,
    BorderTopColor = 220,
    OutlineColor = 221,
    TextDecorationColor = 222,

}

An identifier for a given longhand property.

Variants

AlignItems = 0

align-items

AspectRatio = 1

aspect-ratio

BackfaceVisibility = 2

backface-visibility

BaselineSource = 3

baseline-source

BorderCollapse = 4

border-collapse

BorderImageRepeat = 5

border-image-repeat

BoxSizing = 6

box-sizing

CaptionSide = 7

caption-side

Clear = 8

clear

ColumnCount = 9

column-count

ColumnSpan = 10

column-span

Contain = 11

contain

ContainerType = 12

container-type

Direction = 13

direction

Display = 14

display

EmptyCells = 15

empty-cells

FlexDirection = 16

flex-direction

FlexWrap = 17

flex-wrap

Float = 18

float

FontLanguageOverride = 19

font-language-override

FontOpticalSizing = 20

font-optical-sizing

FontStretch = 21

font-stretch

FontStyle = 22

font-style

FontSynthesisWeight = 23

font-synthesis-weight

FontVariantCaps = 24

font-variant-caps

FontWeight = 25

font-weight

GridAutoFlow = 26

grid-auto-flow

ImageRendering = 27

image-rendering

Isolation = 28

isolation

JustifyItems = 29

justify-items

ListStylePosition = 30

list-style-position

ListStyleType = 31

list-style-type

MixBlendMode = 32

mix-blend-mode

ObjectFit = 33

object-fit

Opacity = 34

opacity

Order = 35

order

OutlineStyle = 36

outline-style

OverflowWrap = 37

overflow-wrap

PointerEvents = 38

pointer-events

Position = 39

position

PositionArea = 40

position-area

ServoOverflowClipBox = 41

-servo-overflow-clip-box

ServoTopLayer = 42

-servo-top-layer

TableLayout = 43

table-layout

TextAlign = 44

text-align

TextAlignLast = 45

text-align-last

TextDecorationLine = 46

text-decoration-line

TextDecorationStyle = 47

text-decoration-style

TextJustify = 48

text-justify

TextRendering = 49

text-rendering

TextTransform = 50

text-transform

TextWrapMode = 51

text-wrap-mode

TransformStyle = 52

transform-style

UnicodeBidi = 53

unicode-bidi

Visibility = 54

visibility

WhiteSpaceCollapse = 55

white-space-collapse

WordBreak = 56

word-break

WritingMode = 57

writing-mode

ZIndex = 58

z-index

Zoom = 59

zoom

AlignContent = 60

align-content

JustifyContent = 61

justify-content

FlexGrow = 62

flex-grow

FlexShrink = 63

flex-shrink

AlignSelf = 64

align-self

JustifySelf = 65

justify-self

OverflowBlock = 66

overflow-block

OverflowInline = 67

overflow-inline

OverflowX = 68

overflow-x

OverflowY = 69

overflow-y

BorderBlockEndStyle = 70

border-block-end-style

BorderBlockStartStyle = 71

border-block-start-style

BorderBottomStyle = 72

border-bottom-style

BorderInlineEndStyle = 73

border-inline-end-style

BorderInlineStartStyle = 74

border-inline-start-style

BorderLeftStyle = 75

border-left-style

BorderRightStyle = 76

border-right-style

BorderTopStyle = 77

border-top-style

AnimationComposition = 78

animation-composition

AnimationDelay = 79

animation-delay

AnimationDirection = 80

animation-direction

AnimationDuration = 81

animation-duration

AnimationFillMode = 82

animation-fill-mode

AnimationIterationCount = 83

animation-iteration-count

AnimationName = 84

animation-name

AnimationPlayState = 85

animation-play-state

AnimationTimeline = 86

animation-timeline

AnimationTimingFunction = 87

animation-timing-function

BackdropFilter = 88

backdrop-filter

BackgroundAttachment = 89

background-attachment

BackgroundClip = 90

background-clip

BackgroundImage = 91

background-image

BackgroundOrigin = 92

background-origin

BackgroundPositionX = 93

background-position-x

BackgroundPositionY = 94

background-position-y

BackgroundRepeat = 95

background-repeat

BackgroundSize = 96

background-size

BorderImageOutset = 97

border-image-outset

BorderImageSlice = 98

border-image-slice

BorderImageWidth = 99

border-image-width

BorderSpacing = 100

border-spacing

BoxShadow = 101

box-shadow

Clip = 102

clip

ClipPath = 103

clip-path

Color = 104

color

ColorScheme = 105

color-scheme

ColumnWidth = 106

column-width

ContainerName = 107

container-name

Content = 108

content

CounterIncrement = 109

counter-increment

CounterReset = 110

counter-reset

Cursor = 111

cursor

Filter = 112

filter

FlexBasis = 113

flex-basis

FontFamily = 114

font-family

FontSize = 115

font-size

FontVariationSettings = 116

font-variation-settings

GridTemplateAreas = 117

grid-template-areas

LetterSpacing = 118

letter-spacing

LineHeight = 119

line-height

MaskImage = 120

mask-image

OffsetPath = 121

offset-path

OutlineOffset = 122

outline-offset

OverflowClipMargin = 123

overflow-clip-margin

Perspective = 124

perspective

Quotes = 125

quotes

Rotate = 126

rotate

Scale = 127

scale

TextIndent = 128

text-indent

TextOverflow = 129

text-overflow

TextShadow = 130

text-shadow

Transform = 131

transform

TransformOrigin = 132

transform-origin

TransitionBehavior = 133

transition-behavior

TransitionDelay = 134

transition-delay

TransitionDuration = 135

transition-duration

TransitionProperty = 136

transition-property

TransitionTimingFunction = 137

transition-timing-function

Translate = 138

translate

VerticalAlign = 139

vertical-align

ViewTransitionClass = 140

view-transition-class

ViewTransitionName = 141

view-transition-name

WillChange = 142

will-change

WordSpacing = 143

word-spacing

XLang = 144

-x-lang

ObjectPosition = 145

object-position

PerspectiveOrigin = 146

perspective-origin

GridTemplateColumns = 147

grid-template-columns

GridTemplateRows = 148

grid-template-rows

BorderImageSource = 149

border-image-source

ListStyleImage = 150

list-style-image

GridAutoColumns = 151

grid-auto-columns

GridAutoRows = 152

grid-auto-rows

ColumnGap = 153

column-gap

RowGap = 154

row-gap

GridColumnEnd = 155

grid-column-end

GridColumnStart = 156

grid-column-start

GridRowEnd = 157

grid-row-end

GridRowStart = 158

grid-row-start

MaxBlockSize = 159

max-block-size

MaxHeight = 160

max-height

MaxInlineSize = 161

max-inline-size

MaxWidth = 162

max-width

BorderBottomLeftRadius = 163

border-bottom-left-radius

BorderBottomRightRadius = 164

border-bottom-right-radius

BorderEndEndRadius = 165

border-end-end-radius

BorderEndStartRadius = 166

border-end-start-radius

BorderStartEndRadius = 167

border-start-end-radius

BorderStartStartRadius = 168

border-start-start-radius

BorderTopLeftRadius = 169

border-top-left-radius

BorderTopRightRadius = 170

border-top-right-radius

Bottom = 171

bottom

InsetBlockEnd = 172

inset-block-end

InsetBlockStart = 173

inset-block-start

InsetInlineEnd = 174

inset-inline-end

InsetInlineStart = 175

inset-inline-start

Left = 176

left

Right = 177

right

Top = 178

top

MarginBlockEnd = 179

margin-block-end

MarginBlockStart = 180

margin-block-start

MarginBottom = 181

margin-bottom

MarginInlineEnd = 182

margin-inline-end

MarginInlineStart = 183

margin-inline-start

MarginLeft = 184

margin-left

MarginRight = 185

margin-right

MarginTop = 186

margin-top

PaddingBlockEnd = 187

padding-block-end

PaddingBlockStart = 188

padding-block-start

PaddingBottom = 189

padding-bottom

PaddingInlineEnd = 190

padding-inline-end

PaddingInlineStart = 191

padding-inline-start

PaddingLeft = 192

padding-left

PaddingRight = 193

padding-right

PaddingTop = 194

padding-top

BlockSize = 195

block-size

Height = 196

height

InlineSize = 197

inline-size

MinBlockSize = 198

min-block-size

MinHeight = 199

min-height

MinInlineSize = 200

min-inline-size

MinWidth = 201

min-width

Width = 202

width

BorderBlockEndWidth = 203

border-block-end-width

BorderBlockStartWidth = 204

border-block-start-width

BorderBottomWidth = 205

border-bottom-width

BorderInlineEndWidth = 206

border-inline-end-width

BorderInlineStartWidth = 207

border-inline-start-width

BorderLeftWidth = 208

border-left-width

BorderRightWidth = 209

border-right-width

BorderTopWidth = 210

border-top-width

OutlineWidth = 211

outline-width

BackgroundColor = 212

background-color

BorderBlockEndColor = 213

border-block-end-color

BorderBlockStartColor = 214

border-block-start-color

BorderBottomColor = 215

border-bottom-color

BorderInlineEndColor = 216

border-inline-end-color

BorderInlineStartColor = 217

border-inline-start-color

BorderLeftColor = 218

border-left-color

BorderRightColor = 219

border-right-color

BorderTopColor = 220

border-top-color

OutlineColor = 221

outline-color

TextDecorationColor = 222

text-decoration-color

Implementations

impl LonghandId

pub fn shorthands(self) -> NonCustomPropertyIterator<ShorthandId>

Returns an iterator over all the shorthands that include this longhand.

pub(super) fn parse_value<'i, 't>( self, context: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<PropertyDeclaration, ParseError<'i>>

fn logical_mapping_data(self) -> Option<&'static LogicalMappingData>

Return the relevant data to map a particular logical property into physical.

pub fn to_physical(self, wm: WritingMode) -> Self

If this is a logical property, return the corresponding physical one in the given writing mode. Otherwise, return unchanged.

pub fn logical_group(self) -> Option<LogicalGroupId>

Return the logical group of this longhand property.

pub fn flags(self) -> PropertyFlags

Returns PropertyFlags for given longhand property.

impl LonghandId

pub fn name(&self) -> &'static str

Get the name of this longhand property.

pub fn inherited(self) -> bool

Returns whether the longhand property is inherited by default.

pub fn zoom_dependent(self) -> bool

Returns whether the longhand property is zoom-dependent.

pub fn ignored_when_document_colors_disabled(self) -> bool

Returns true if the property is one that is ignored when document colors are disabled.

pub fn is_or_is_longhand_of(self, non_custom: NonCustomPropertyId) -> bool

Returns whether this longhand is non_custom or is a longhand of it.

pub fn is_longhand_of(self, shorthand: ShorthandId) -> bool

Returns whether this longhand is a longhand of shorthand.

pub fn is_animatable(self) -> bool

Returns whether this property is animatable.

pub fn is_discrete_animatable(self) -> bool

Returns whether this property is animatable in a discrete way.

pub fn is_logical(self) -> bool

Return whether this property is logical.