Enum style::properties::generated::LonghandId

#[repr(u16)]
pub enum LonghandId {

    AlignContent,
    AlignItems,
    AlignSelf,
    AspectRatio,
    BackfaceVisibility,
    BorderCollapse,
    BorderImageRepeat,
    BoxSizing,
    CaptionSide,
    Clear,
    ColumnCount,
    ColumnSpan,
    ContainerType,
    Direction,
    Display,
    EmptyCells,
    FlexDirection,
    FlexWrap,
    Float,
    FontLanguageOverride,
    FontStretch,
    FontStyle,
    FontVariantCaps,
    FontWeight,
    ImageRendering,
    JustifyContent,
    ListStylePosition,
    ListStyleType,
    MixBlendMode,
    Opacity,
    Order,
    OutlineStyle,
    OverflowWrap,
    PointerEvents,
    Position,
    ServoOverflowClipBox,
    ServoTopLayer,
    TableLayout,
    TextAlign,
    TextAlignLast,
    TextDecorationLine,
    TextDecorationStyle,
    TextJustify,
    TextRendering,
    TextTransform,
    TransformStyle,
    UnicodeBidi,
    Visibility,
    WhiteSpace,
    WordBreak,
    WritingMode,
    ZIndex,
    FlexGrow,
    FlexShrink,
    OverflowBlock,
    OverflowInline,
    OverflowX,
    OverflowY,
    BorderBlockEndStyle,
    BorderBlockStartStyle,
    BorderBottomStyle,
    BorderInlineEndStyle,
    BorderInlineStartStyle,
    BorderLeftStyle,
    BorderRightStyle,
    BorderTopStyle,
    AnimationComposition,
    AnimationDelay,
    AnimationDirection,
    AnimationDuration,
    AnimationFillMode,
    AnimationIterationCount,
    AnimationName,
    AnimationPlayState,
    AnimationTimeline,
    AnimationTimingFunction,
    BackgroundAttachment,
    BackgroundClip,
    BackgroundImage,
    BackgroundOrigin,
    BackgroundPositionX,
    BackgroundPositionY,
    BackgroundRepeat,
    BackgroundSize,
    BorderImageOutset,
    BorderImageSlice,
    BorderImageWidth,
    BorderSpacing,
    BoxShadow,
    Clip,
    Color,
    ColumnGap,
    ColumnWidth,
    ContainerName,
    Content,
    CounterIncrement,
    CounterReset,
    Cursor,
    Filter,
    FlexBasis,
    FontFamily,
    FontSize,
    LetterSpacing,
    LineHeight,
    OutlineOffset,
    Perspective,
    PerspectiveOrigin,
    Quotes,
    Rotate,
    Scale,
    TextIndent,
    TextOverflow,
    TextShadow,
    Transform,
    TransformOrigin,
    TransitionDelay,
    TransitionDuration,
    TransitionProperty,
    TransitionTimingFunction,
    Translate,
    VerticalAlign,
    WordSpacing,
    XLang,
    BorderImageSource,
    ListStyleImage,
    MaxBlockSize,
    MaxHeight,
    MaxInlineSize,
    MaxWidth,
    BorderBottomLeftRadius,
    BorderBottomRightRadius,
    BorderEndEndRadius,
    BorderEndStartRadius,
    BorderStartEndRadius,
    BorderStartStartRadius,
    BorderTopLeftRadius,
    BorderTopRightRadius,
    PaddingBlockEnd,
    PaddingBlockStart,
    PaddingBottom,
    PaddingInlineEnd,
    PaddingInlineStart,
    PaddingLeft,
    PaddingRight,
    PaddingTop,
    BlockSize,
    Height,
    InlineSize,
    MinBlockSize,
    MinHeight,
    MinInlineSize,
    MinWidth,
    Width,
    BorderBlockEndWidth,
    BorderBlockStartWidth,
    BorderBottomWidth,
    BorderInlineEndWidth,
    BorderInlineStartWidth,
    BorderLeftWidth,
    BorderRightWidth,
    BorderTopWidth,
    OutlineWidth,
    BackgroundColor,
    BorderBlockEndColor,
    BorderBlockStartColor,
    BorderBottomColor,
    BorderInlineEndColor,
    BorderInlineStartColor,
    BorderLeftColor,
    BorderRightColor,
    BorderTopColor,
    OutlineColor,
    TextDecorationColor,
    Bottom,
    InsetBlockEnd,
    InsetBlockStart,
    InsetInlineEnd,
    InsetInlineStart,
    Left,
    MarginBlockEnd,
    MarginBlockStart,
    MarginBottom,
    MarginInlineEnd,
    MarginInlineStart,
    MarginLeft,
    MarginRight,
    MarginTop,
    Right,
    Top,

}

An identifier for a given longhand property.

Variants

AlignContent

align-content

AlignItems

align-items

AlignSelf

align-self

AspectRatio

aspect-ratio

BackfaceVisibility

backface-visibility

BorderCollapse

border-collapse

BorderImageRepeat

border-image-repeat

BoxSizing

box-sizing

CaptionSide

caption-side

Clear

clear

ColumnCount

column-count

ColumnSpan

column-span

ContainerType

container-type

Direction

direction

Display

display

EmptyCells

empty-cells

FlexDirection

flex-direction

FlexWrap

flex-wrap

Float

float

FontLanguageOverride

font-language-override

FontStretch

font-stretch

FontStyle

font-style

FontVariantCaps

font-variant-caps

FontWeight

font-weight

ImageRendering

image-rendering

JustifyContent

justify-content

ListStylePosition

list-style-position

ListStyleType

list-style-type

MixBlendMode

mix-blend-mode

Opacity

opacity

Order

order

OutlineStyle

outline-style

OverflowWrap

overflow-wrap

PointerEvents

pointer-events

Position

position

ServoOverflowClipBox

-servo-overflow-clip-box

ServoTopLayer

-servo-top-layer

TableLayout

table-layout

TextAlign

text-align

TextAlignLast

text-align-last

TextDecorationLine

text-decoration-line

TextDecorationStyle

text-decoration-style

TextJustify

text-justify

TextRendering

text-rendering

TextTransform

text-transform

TransformStyle

transform-style

UnicodeBidi

unicode-bidi

Visibility

visibility

WhiteSpace

white-space

WordBreak

word-break

WritingMode

writing-mode

ZIndex

z-index

FlexGrow

flex-grow

FlexShrink

flex-shrink

OverflowBlock

overflow-block

OverflowInline

overflow-inline

OverflowX

overflow-x

OverflowY

overflow-y

BorderBlockEndStyle

border-block-end-style

BorderBlockStartStyle

border-block-start-style

BorderBottomStyle

border-bottom-style

BorderInlineEndStyle

border-inline-end-style

BorderInlineStartStyle

border-inline-start-style

BorderLeftStyle

border-left-style

BorderRightStyle

border-right-style

BorderTopStyle

border-top-style

AnimationComposition

animation-composition

AnimationDelay

animation-delay

AnimationDirection

animation-direction

AnimationDuration

animation-duration

AnimationFillMode

animation-fill-mode

AnimationIterationCount

animation-iteration-count

AnimationName

animation-name

AnimationPlayState

animation-play-state

AnimationTimeline

animation-timeline

AnimationTimingFunction

animation-timing-function

BackgroundAttachment

background-attachment

BackgroundClip

background-clip

BackgroundImage

background-image

BackgroundOrigin

background-origin

BackgroundPositionX

background-position-x

BackgroundPositionY

background-position-y

BackgroundRepeat

background-repeat

BackgroundSize

background-size

BorderImageOutset

border-image-outset

BorderImageSlice

border-image-slice

BorderImageWidth

border-image-width

BorderSpacing

border-spacing

BoxShadow

box-shadow

Clip

clip

Color

color

ColumnGap

column-gap

ColumnWidth

column-width

ContainerName

container-name

Content

content

CounterIncrement

counter-increment

CounterReset

counter-reset

Cursor

cursor

Filter

filter

FlexBasis

flex-basis

FontFamily

font-family

FontSize

font-size

LetterSpacing

letter-spacing

LineHeight

line-height

OutlineOffset

outline-offset

Perspective

perspective

PerspectiveOrigin

perspective-origin

Quotes

quotes

Rotate

rotate

Scale

scale

TextIndent

text-indent

TextOverflow

text-overflow

TextShadow

text-shadow

Transform

transform

TransformOrigin

transform-origin

TransitionDelay

transition-delay

TransitionDuration

transition-duration

TransitionProperty

transition-property

TransitionTimingFunction

transition-timing-function

Translate

translate

VerticalAlign

vertical-align

WordSpacing

word-spacing

XLang

-x-lang

BorderImageSource

border-image-source

ListStyleImage

list-style-image

MaxBlockSize

max-block-size

MaxHeight

max-height

MaxInlineSize

max-inline-size

MaxWidth

max-width

BorderBottomLeftRadius

border-bottom-left-radius

BorderBottomRightRadius

border-bottom-right-radius

BorderEndEndRadius

border-end-end-radius

BorderEndStartRadius

border-end-start-radius

BorderStartEndRadius

border-start-end-radius

BorderStartStartRadius

border-start-start-radius

BorderTopLeftRadius

border-top-left-radius

BorderTopRightRadius

border-top-right-radius

PaddingBlockEnd

padding-block-end

PaddingBlockStart

padding-block-start

PaddingBottom

padding-bottom

PaddingInlineEnd

padding-inline-end

PaddingInlineStart

padding-inline-start

PaddingLeft

padding-left

PaddingRight

padding-right

PaddingTop

padding-top

BlockSize

block-size

Height

height

InlineSize

inline-size

MinBlockSize

min-block-size

MinHeight

min-height

MinInlineSize

min-inline-size

MinWidth

min-width

Width

width

BorderBlockEndWidth

border-block-end-width

BorderBlockStartWidth

border-block-start-width

BorderBottomWidth

border-bottom-width

BorderInlineEndWidth

border-inline-end-width

BorderInlineStartWidth

border-inline-start-width

BorderLeftWidth

border-left-width

BorderRightWidth

border-right-width

BorderTopWidth

border-top-width

OutlineWidth

outline-width

BackgroundColor

background-color

BorderBlockEndColor

border-block-end-color

BorderBlockStartColor

border-block-start-color

BorderBottomColor

border-bottom-color

BorderInlineEndColor

border-inline-end-color

BorderInlineStartColor

border-inline-start-color

BorderLeftColor

border-left-color

BorderRightColor

border-right-color

BorderTopColor

border-top-color

OutlineColor

outline-color

TextDecorationColor

text-decoration-color

Bottom

bottom

InsetBlockEnd

inset-block-end

InsetBlockStart

inset-block-start

InsetInlineEnd

inset-inline-end

InsetInlineStart

inset-inline-start

Left

left

MarginBlockEnd

margin-block-end

MarginBlockStart

margin-block-start

MarginBottom

margin-bottom

MarginInlineEnd

margin-inline-end

MarginInlineStart

margin-inline-start

MarginLeft

margin-left

MarginRight

margin-right

MarginTop

margin-top

Right

right

Top

top

Implementations

impl LonghandId

pub fn name(&self) -> &'static str

Get the name of this longhand property.

pub fn inherited(self) -> bool

Returns whether the longhand property is inherited by default.

pub fn shorthands(&self) -> NonCustomPropertyIterator<ShorthandId>

Returns an iterator over all the shorthands that include this longhand.

fn parse_value<'i, 't>( &self, context: &ParserContext<'_>, input: &mut Parser<'i, 't> ) -> Result<PropertyDeclaration, ParseError<'i>>

pub fn is_animatable(self) -> bool

Returns whether this property is animatable.

pub fn is_discrete_animatable(self) -> bool

Returns whether this property is animatable in a discrete way.

pub fn is_transitionable(self) -> bool

Returns whether this property is transitionable.

pub fn is_logical(self) -> bool

Return whether this property is logical.

pub fn to_physical(&self, wm: WritingMode) -> Self

If this is a logical property, return the corresponding physical one in the given writing mode.

Otherwise, return unchanged.

pub fn logical_group(&self) -> Option<LogicalGroup>

Return the logical group of this longhand property.

pub fn flags(self) -> PropertyFlags

Returns PropertyFlags for given longhand property.

pub fn ignored_when_document_colors_disabled(self) -> bool

Returns true if the property is one that is ignored when document colors are disabled.