Enum style::properties::generated::LonghandId

#[repr(u16)]
pub enum LonghandId {

    AlignContent = 0,
    AlignItems = 1,
    AlignSelf = 2,
    AspectRatio = 3,
    BackfaceVisibility = 4,
    BaselineSource = 5,
    BorderCollapse = 6,
    BorderImageRepeat = 7,
    BoxSizing = 8,
    CaptionSide = 9,
    Clear = 10,
    ColumnCount = 11,
    ColumnSpan = 12,
    ContainerType = 13,
    Direction = 14,
    Display = 15,
    EmptyCells = 16,
    FlexDirection = 17,
    FlexWrap = 18,
    Float = 19,
    FontLanguageOverride = 20,
    FontStretch = 21,
    FontStyle = 22,
    FontVariantCaps = 23,
    FontWeight = 24,
    GridAutoFlow = 25,
    ImageRendering = 26,
    JustifyContent = 27,
    JustifyItems = 28,
    JustifySelf = 29,
    ListStylePosition = 30,
    ListStyleType = 31,
    MixBlendMode = 32,
    ObjectFit = 33,
    Opacity = 34,
    Order = 35,
    OutlineStyle = 36,
    OverflowWrap = 37,
    PointerEvents = 38,
    Position = 39,
    ServoOverflowClipBox = 40,
    ServoTopLayer = 41,
    TableLayout = 42,
    TextAlign = 43,
    TextAlignLast = 44,
    TextDecorationLine = 45,
    TextDecorationStyle = 46,
    TextJustify = 47,
    TextRendering = 48,
    TextTransform = 49,
    TextWrapMode = 50,
    TransformStyle = 51,
    UnicodeBidi = 52,
    Visibility = 53,
    WhiteSpaceCollapse = 54,
    WordBreak = 55,
    WritingMode = 56,
    ZIndex = 57,
    Zoom = 58,
    FlexGrow = 59,
    FlexShrink = 60,
    OverflowBlock = 61,
    OverflowInline = 62,
    OverflowX = 63,
    OverflowY = 64,
    BorderBlockEndStyle = 65,
    BorderBlockStartStyle = 66,
    BorderBottomStyle = 67,
    BorderInlineEndStyle = 68,
    BorderInlineStartStyle = 69,
    BorderLeftStyle = 70,
    BorderRightStyle = 71,
    BorderTopStyle = 72,
    AnimationComposition = 73,
    AnimationDelay = 74,
    AnimationDirection = 75,
    AnimationDuration = 76,
    AnimationFillMode = 77,
    AnimationIterationCount = 78,
    AnimationName = 79,
    AnimationPlayState = 80,
    AnimationTimeline = 81,
    AnimationTimingFunction = 82,
    BackgroundAttachment = 83,
    BackgroundClip = 84,
    BackgroundImage = 85,
    BackgroundOrigin = 86,
    BackgroundPositionX = 87,
    BackgroundPositionY = 88,
    BackgroundRepeat = 89,
    BackgroundSize = 90,
    BorderImageOutset = 91,
    BorderImageSlice = 92,
    BorderImageWidth = 93,
    BorderSpacing = 94,
    BoxShadow = 95,
    Clip = 96,
    ClipPath = 97,
    Color = 98,
    ColorScheme = 99,
    ColumnWidth = 100,
    ContainerName = 101,
    Content = 102,
    CounterIncrement = 103,
    CounterReset = 104,
    Cursor = 105,
    Filter = 106,
    FlexBasis = 107,
    FontFamily = 108,
    FontSize = 109,
    GridTemplateAreas = 110,
    LetterSpacing = 111,
    LineHeight = 112,
    OutlineOffset = 113,
    Perspective = 114,
    Quotes = 115,
    Rotate = 116,
    Scale = 117,
    TextIndent = 118,
    TextOverflow = 119,
    TextShadow = 120,
    Transform = 121,
    TransformOrigin = 122,
    TransitionBehavior = 123,
    TransitionDelay = 124,
    TransitionDuration = 125,
    TransitionProperty = 126,
    TransitionTimingFunction = 127,
    Translate = 128,
    VerticalAlign = 129,
    WordSpacing = 130,
    XLang = 131,
    BorderImageSource = 132,
    ListStyleImage = 133,
    ObjectPosition = 134,
    PerspectiveOrigin = 135,
    GridTemplateColumns = 136,
    GridTemplateRows = 137,
    GridAutoColumns = 138,
    GridAutoRows = 139,
    ColumnGap = 140,
    RowGap = 141,
    GridColumnEnd = 142,
    GridColumnStart = 143,
    GridRowEnd = 144,
    GridRowStart = 145,
    MaxBlockSize = 146,
    MaxHeight = 147,
    MaxInlineSize = 148,
    MaxWidth = 149,
    BorderBottomLeftRadius = 150,
    BorderBottomRightRadius = 151,
    BorderEndEndRadius = 152,
    BorderEndStartRadius = 153,
    BorderStartEndRadius = 154,
    BorderStartStartRadius = 155,
    BorderTopLeftRadius = 156,
    BorderTopRightRadius = 157,
    Bottom = 158,
    InsetBlockEnd = 159,
    InsetBlockStart = 160,
    InsetInlineEnd = 161,
    InsetInlineStart = 162,
    Left = 163,
    Right = 164,
    Top = 165,
    MarginBlockEnd = 166,
    MarginBlockStart = 167,
    MarginBottom = 168,
    MarginInlineEnd = 169,
    MarginInlineStart = 170,
    MarginLeft = 171,
    MarginRight = 172,
    MarginTop = 173,
    PaddingBlockEnd = 174,
    PaddingBlockStart = 175,
    PaddingBottom = 176,
    PaddingInlineEnd = 177,
    PaddingInlineStart = 178,
    PaddingLeft = 179,
    PaddingRight = 180,
    PaddingTop = 181,
    BlockSize = 182,
    Height = 183,
    InlineSize = 184,
    MinBlockSize = 185,
    MinHeight = 186,
    MinInlineSize = 187,
    MinWidth = 188,
    Width = 189,
    BorderBlockEndWidth = 190,
    BorderBlockStartWidth = 191,
    BorderBottomWidth = 192,
    BorderInlineEndWidth = 193,
    BorderInlineStartWidth = 194,
    BorderLeftWidth = 195,
    BorderRightWidth = 196,
    BorderTopWidth = 197,
    OutlineWidth = 198,
    BackgroundColor = 199,
    BorderBlockEndColor = 200,
    BorderBlockStartColor = 201,
    BorderBottomColor = 202,
    BorderInlineEndColor = 203,
    BorderInlineStartColor = 204,
    BorderLeftColor = 205,
    BorderRightColor = 206,
    BorderTopColor = 207,
    OutlineColor = 208,
    TextDecorationColor = 209,

}

An identifier for a given longhand property.

Variants

AlignContent = 0

align-content

AlignItems = 1

align-items

AlignSelf = 2

align-self

AspectRatio = 3

aspect-ratio

BackfaceVisibility = 4

backface-visibility

BaselineSource = 5

baseline-source

BorderCollapse = 6

border-collapse

BorderImageRepeat = 7

border-image-repeat

BoxSizing = 8

box-sizing

CaptionSide = 9

caption-side

Clear = 10

clear

ColumnCount = 11

column-count

ColumnSpan = 12

column-span

ContainerType = 13

container-type

Direction = 14

direction

Display = 15

display

EmptyCells = 16

empty-cells

FlexDirection = 17

flex-direction

FlexWrap = 18

flex-wrap

Float = 19

float

FontLanguageOverride = 20

font-language-override

FontStretch = 21

font-stretch

FontStyle = 22

font-style

FontVariantCaps = 23

font-variant-caps

FontWeight = 24

font-weight

GridAutoFlow = 25

grid-auto-flow

ImageRendering = 26

image-rendering

JustifyContent = 27

justify-content

JustifyItems = 28

justify-items

JustifySelf = 29

justify-self

ListStylePosition = 30

list-style-position

ListStyleType = 31

list-style-type

MixBlendMode = 32

mix-blend-mode

ObjectFit = 33

object-fit

Opacity = 34

opacity

Order = 35

order

OutlineStyle = 36

outline-style

OverflowWrap = 37

overflow-wrap

PointerEvents = 38

pointer-events

Position = 39

position

ServoOverflowClipBox = 40

-servo-overflow-clip-box

ServoTopLayer = 41

-servo-top-layer

TableLayout = 42

table-layout

TextAlign = 43

text-align

TextAlignLast = 44

text-align-last

TextDecorationLine = 45

text-decoration-line

TextDecorationStyle = 46

text-decoration-style

TextJustify = 47

text-justify

TextRendering = 48

text-rendering

TextTransform = 49

text-transform

TextWrapMode = 50

text-wrap-mode

TransformStyle = 51

transform-style

UnicodeBidi = 52

unicode-bidi

Visibility = 53

visibility

WhiteSpaceCollapse = 54

white-space-collapse

WordBreak = 55

word-break

WritingMode = 56

writing-mode

ZIndex = 57

z-index

Zoom = 58

zoom

FlexGrow = 59

flex-grow

FlexShrink = 60

flex-shrink

OverflowBlock = 61

overflow-block

OverflowInline = 62

overflow-inline

OverflowX = 63

overflow-x

OverflowY = 64

overflow-y

BorderBlockEndStyle = 65

border-block-end-style

BorderBlockStartStyle = 66

border-block-start-style

BorderBottomStyle = 67

border-bottom-style

BorderInlineEndStyle = 68

border-inline-end-style

BorderInlineStartStyle = 69

border-inline-start-style

BorderLeftStyle = 70

border-left-style

BorderRightStyle = 71

border-right-style

BorderTopStyle = 72

border-top-style

AnimationComposition = 73

animation-composition

AnimationDelay = 74

animation-delay

AnimationDirection = 75

animation-direction

AnimationDuration = 76

animation-duration

AnimationFillMode = 77

animation-fill-mode

AnimationIterationCount = 78

animation-iteration-count

AnimationName = 79

animation-name

AnimationPlayState = 80

animation-play-state

AnimationTimeline = 81

animation-timeline

AnimationTimingFunction = 82

animation-timing-function

BackgroundAttachment = 83

background-attachment

BackgroundClip = 84

background-clip

BackgroundImage = 85

background-image

BackgroundOrigin = 86

background-origin

BackgroundPositionX = 87

background-position-x

BackgroundPositionY = 88

background-position-y

BackgroundRepeat = 89

background-repeat

BackgroundSize = 90

background-size

BorderImageOutset = 91

border-image-outset

BorderImageSlice = 92

border-image-slice

BorderImageWidth = 93

border-image-width

BorderSpacing = 94

border-spacing

BoxShadow = 95

box-shadow

Clip = 96

clip

ClipPath = 97

clip-path

Color = 98

color

ColorScheme = 99

color-scheme

ColumnWidth = 100

column-width

ContainerName = 101

container-name

Content = 102

content

CounterIncrement = 103

counter-increment

CounterReset = 104

counter-reset

Cursor = 105

cursor

Filter = 106

filter

FlexBasis = 107

flex-basis

FontFamily = 108

font-family

FontSize = 109

font-size

GridTemplateAreas = 110

grid-template-areas

LetterSpacing = 111

letter-spacing

LineHeight = 112

line-height

OutlineOffset = 113

outline-offset

Perspective = 114

perspective

Quotes = 115

quotes

Rotate = 116

rotate

Scale = 117

scale

TextIndent = 118

text-indent

TextOverflow = 119

text-overflow

TextShadow = 120

text-shadow

Transform = 121

transform

TransformOrigin = 122

transform-origin

TransitionBehavior = 123

transition-behavior

TransitionDelay = 124

transition-delay

TransitionDuration = 125

transition-duration

TransitionProperty = 126

transition-property

TransitionTimingFunction = 127

transition-timing-function

Translate = 128

translate

VerticalAlign = 129

vertical-align

WordSpacing = 130

word-spacing

XLang = 131

-x-lang

BorderImageSource = 132

border-image-source

ListStyleImage = 133

list-style-image

ObjectPosition = 134

object-position

PerspectiveOrigin = 135

perspective-origin

GridTemplateColumns = 136

grid-template-columns

GridTemplateRows = 137

grid-template-rows

GridAutoColumns = 138

grid-auto-columns

GridAutoRows = 139

grid-auto-rows

ColumnGap = 140

column-gap

RowGap = 141

row-gap

GridColumnEnd = 142

grid-column-end

GridColumnStart = 143

grid-column-start

GridRowEnd = 144

grid-row-end

GridRowStart = 145

grid-row-start

MaxBlockSize = 146

max-block-size

MaxHeight = 147

max-height

MaxInlineSize = 148

max-inline-size

MaxWidth = 149

max-width

BorderBottomLeftRadius = 150

border-bottom-left-radius

BorderBottomRightRadius = 151

border-bottom-right-radius

BorderEndEndRadius = 152

border-end-end-radius

BorderEndStartRadius = 153

border-end-start-radius

BorderStartEndRadius = 154

border-start-end-radius

BorderStartStartRadius = 155

border-start-start-radius

BorderTopLeftRadius = 156

border-top-left-radius

BorderTopRightRadius = 157

border-top-right-radius

Bottom = 158

bottom

InsetBlockEnd = 159

inset-block-end

InsetBlockStart = 160

inset-block-start

InsetInlineEnd = 161

inset-inline-end

InsetInlineStart = 162

inset-inline-start

Left = 163

left

Right = 164

right

Top = 165

top

MarginBlockEnd = 166

margin-block-end

MarginBlockStart = 167

margin-block-start

MarginBottom = 168

margin-bottom

MarginInlineEnd = 169

margin-inline-end

MarginInlineStart = 170

margin-inline-start

MarginLeft = 171

margin-left

MarginRight = 172

margin-right

MarginTop = 173

margin-top

PaddingBlockEnd = 174

padding-block-end

PaddingBlockStart = 175

padding-block-start

PaddingBottom = 176

padding-bottom

PaddingInlineEnd = 177

padding-inline-end

PaddingInlineStart = 178

padding-inline-start

PaddingLeft = 179

padding-left

PaddingRight = 180

padding-right

PaddingTop = 181

padding-top

BlockSize = 182

block-size

Height = 183

height

InlineSize = 184

inline-size

MinBlockSize = 185

min-block-size

MinHeight = 186

min-height

MinInlineSize = 187

min-inline-size

MinWidth = 188

min-width

Width = 189

width

BorderBlockEndWidth = 190

border-block-end-width

BorderBlockStartWidth = 191

border-block-start-width

BorderBottomWidth = 192

border-bottom-width

BorderInlineEndWidth = 193

border-inline-end-width

BorderInlineStartWidth = 194

border-inline-start-width

BorderLeftWidth = 195

border-left-width

BorderRightWidth = 196

border-right-width

BorderTopWidth = 197

border-top-width

OutlineWidth = 198

outline-width

BackgroundColor = 199

background-color

BorderBlockEndColor = 200

border-block-end-color

BorderBlockStartColor = 201

border-block-start-color

BorderBottomColor = 202

border-bottom-color

BorderInlineEndColor = 203

border-inline-end-color

BorderInlineStartColor = 204

border-inline-start-color

BorderLeftColor = 205

border-left-color

BorderRightColor = 206

border-right-color

BorderTopColor = 207

border-top-color

OutlineColor = 208

outline-color

TextDecorationColor = 209

text-decoration-color

Implementations

impl LonghandId

pub fn shorthands(self) -> NonCustomPropertyIterator<ShorthandId>

Returns an iterator over all the shorthands that include this longhand.

pub(super) fn parse_value<'i, 't>( self, context: &ParserContext<'_>, input: &mut Parser<'i, 't>, ) -> Result<PropertyDeclaration, ParseError<'i>>

fn logical_mapping_data(self) -> Option<&'static LogicalMappingData>

Return the relevant data to map a particular logical property into physical.

pub fn to_physical(self, wm: WritingMode) -> Self

If this is a logical property, return the corresponding physical one in the given writing mode. Otherwise, return unchanged.

pub fn logical_group(self) -> Option<LogicalGroupId>

Return the logical group of this longhand property.

pub fn flags(self) -> PropertyFlags

Returns PropertyFlags for given longhand property.

impl LonghandId

pub fn name(&self) -> &'static str

Get the name of this longhand property.

pub fn inherited(self) -> bool

Returns whether the longhand property is inherited by default.

pub fn ignored_when_document_colors_disabled(self) -> bool

Returns true if the property is one that is ignored when document colors are disabled.

pub fn is_or_is_longhand_of(self, non_custom: NonCustomPropertyId) -> bool

Returns whether this longhand is non_custom or is a longhand of it.

pub fn is_longhand_of(self, shorthand: ShorthandId) -> bool

Returns whether this longhand is a longhand of shorthand.

pub fn is_animatable(self) -> bool

Returns whether this property is animatable.

pub fn is_discrete_animatable(self) -> bool

Returns whether this property is animatable in a discrete way.

pub fn is_logical(self) -> bool

Return whether this property is logical.