style::properties::generated

Module property_counts

A module to group various interesting property counts.

Constants

ALIASES

The number of aliases.

ALL_SHORTHAND_EXPANDED

The max amount of longhands that the all shorthand will ever contain.

ANIMATABLE

The number of animatable properties.

COUNTED_UNKNOWN

The number of counted unknown properties.

LONGHANDS

The number of (non-alias) longhand properties.

LONGHANDS_AND_SHORTHANDS

The number of (non-alias) longhands and shorthands.

MAX_SHORTHAND_EXPANDED

The max number of longhands that a shorthand other than “all” expands to.

NON_CUSTOM

The number of non-custom properties.

PRIORITARY

The number of prioritary properties that we have.

SHORTHANDS

The number of (non-alias) shorthand properties.