Struct style::properties::generated::ComputedValues

pub struct ComputedValues {
    inner: ComputedValuesInner,
    pseudo: Option<PseudoElement>,
}

The struct that Servo uses to represent computed values.

This struct contains an immutable atomically-reference-counted pointer to every kind of style struct.

When needed, the structs may be copied in order to get mutated.

Fields
inner: ComputedValuesInner

The actual computed values

In Gecko the outer ComputedValues is actually a ComputedStyle, whereas ComputedValuesInner is the core set of computed values.

We maintain this distinction in servo to reduce the amount of special casing.

pseudo: Option<PseudoElement>

The pseudo-element that we’re using.

Implementations

impl ComputedValues

pub fn pseudo(&self) -> Option<&PseudoElement>

Returns the pseudo-element that this style represents.

pub fn is_pseudo_style(&self) -> bool

Returns true if this is the style for a pseudo-element.

pub fn is_display_contents(&self) -> bool

Returns whether this style’s display value is equal to contents.

pub fn rules(&self) -> &StrongRuleNode

Gets a reference to the rule node. Panic if no rule node exists.

pub fn visited_rules(&self) -> Option<&StrongRuleNode>

Returns the visited rules, if applicable.

pub fn custom_properties(&self) -> &ComputedCustomProperties

Gets a reference to the custom properties map (if one exists).

pub fn custom_properties_equal(&self, other: &Self) -> bool

Returns whether we have the same custom properties as another style.

pub fn clone_align_content(&self) -> T

Gets the computed value of a given property.

pub fn clone_align_items(&self) -> T

Gets the computed value of a given property.

pub fn clone_align_self(&self) -> T

Gets the computed value of a given property.

pub fn clone_aspect_ratio(&self) -> T

Gets the computed value of a given property.

pub fn clone_backface_visibility(&self) -> T

Gets the computed value of a given property.

pub fn clone_baseline_source(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_collapse(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_image_repeat(&self) -> T

Gets the computed value of a given property.

pub fn clone_box_sizing(&self) -> T

Gets the computed value of a given property.

pub fn clone_caption_side(&self) -> T

Gets the computed value of a given property.

pub fn clone_clear(&self) -> T

Gets the computed value of a given property.

pub fn clone_column_count(&self) -> T

Gets the computed value of a given property.

pub fn clone_column_span(&self) -> T

Gets the computed value of a given property.

pub fn clone_container_type(&self) -> T

Gets the computed value of a given property.

pub fn clone_direction(&self) -> T

Gets the computed value of a given property.

pub fn clone_display(&self) -> T

Gets the computed value of a given property.

pub fn clone_empty_cells(&self) -> T

Gets the computed value of a given property.

pub fn clone_flex_direction(&self) -> T

Gets the computed value of a given property.

pub fn clone_flex_wrap(&self) -> T

Gets the computed value of a given property.

pub fn clone_float(&self) -> T

Gets the computed value of a given property.

pub fn clone_font_language_override(&self) -> T

Gets the computed value of a given property.

pub fn clone_font_stretch(&self) -> T

Gets the computed value of a given property.

pub fn clone_font_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_font_variant_caps(&self) -> T

Gets the computed value of a given property.

pub fn clone_font_weight(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_auto_flow(&self) -> T

Gets the computed value of a given property.

pub fn clone_image_rendering(&self) -> T

Gets the computed value of a given property.

pub fn clone_justify_content(&self) -> T

Gets the computed value of a given property.

pub fn clone_justify_items(&self) -> T

Gets the computed value of a given property.

pub fn clone_justify_self(&self) -> T

Gets the computed value of a given property.

pub fn clone_list_style_position(&self) -> T

Gets the computed value of a given property.

pub fn clone_list_style_type(&self) -> T

Gets the computed value of a given property.

pub fn clone_mix_blend_mode(&self) -> T

Gets the computed value of a given property.

pub fn clone_object_fit(&self) -> T

Gets the computed value of a given property.

pub fn clone_opacity(&self) -> T

Gets the computed value of a given property.

pub fn clone_order(&self) -> T

Gets the computed value of a given property.

pub fn clone_outline_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_overflow_wrap(&self) -> T

Gets the computed value of a given property.

pub fn clone_pointer_events(&self) -> T

Gets the computed value of a given property.

pub fn clone_position(&self) -> T

Gets the computed value of a given property.

pub fn clone__servo_overflow_clip_box(&self) -> T

Gets the computed value of a given property.

pub fn clone__servo_top_layer(&self) -> T

Gets the computed value of a given property.

pub fn clone_table_layout(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_align(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_align_last(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_decoration_line(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_decoration_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_justify(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_rendering(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_transform(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_wrap_mode(&self) -> T

Gets the computed value of a given property.

pub fn clone_transform_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_unicode_bidi(&self) -> T

Gets the computed value of a given property.

pub fn clone_visibility(&self) -> T

Gets the computed value of a given property.

pub fn clone_white_space_collapse(&self) -> T

Gets the computed value of a given property.

pub fn clone_word_break(&self) -> T

Gets the computed value of a given property.

pub fn clone_writing_mode(&self) -> T

Gets the computed value of a given property.

pub fn clone_z_index(&self) -> T

Gets the computed value of a given property.

pub fn clone_zoom(&self) -> T

Gets the computed value of a given property.

pub fn clone_flex_grow(&self) -> T

Gets the computed value of a given property.

pub fn clone_flex_shrink(&self) -> T

Gets the computed value of a given property.

pub fn clone_overflow_x(&self) -> T

Gets the computed value of a given property.

pub fn clone_overflow_y(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_bottom_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_left_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_right_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_top_style(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_composition(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_delay(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_direction(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_duration(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_fill_mode(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_iteration_count(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_name(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_play_state(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_timeline(&self) -> T

Gets the computed value of a given property.

pub fn clone_animation_timing_function(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_attachment(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_clip(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_image(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_origin(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_position_x(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_position_y(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_repeat(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_size(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_image_outset(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_image_slice(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_image_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_spacing(&self) -> T

Gets the computed value of a given property.

pub fn clone_box_shadow(&self) -> T

Gets the computed value of a given property.

pub fn clone_clip(&self) -> T

Gets the computed value of a given property.

pub fn clone_clip_path(&self) -> T

Gets the computed value of a given property.

pub fn clone_color(&self) -> T

Gets the computed value of a given property.

pub fn clone_color_scheme(&self) -> T

Gets the computed value of a given property.

pub fn clone_column_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_container_name(&self) -> T

Gets the computed value of a given property.

pub fn clone_content(&self) -> T

Gets the computed value of a given property.

pub fn clone_counter_increment(&self) -> T

Gets the computed value of a given property.

pub fn clone_counter_reset(&self) -> T

Gets the computed value of a given property.

pub fn clone_cursor(&self) -> T

Gets the computed value of a given property.

pub fn clone_filter(&self) -> T

Gets the computed value of a given property.

pub fn clone_flex_basis(&self) -> T

Gets the computed value of a given property.

pub fn clone_font_family(&self) -> T

Gets the computed value of a given property.

pub fn clone_font_size(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_template_areas(&self) -> T

Gets the computed value of a given property.

pub fn clone_letter_spacing(&self) -> T

Gets the computed value of a given property.

pub fn clone_line_height(&self) -> T

Gets the computed value of a given property.

pub fn clone_outline_offset(&self) -> T

Gets the computed value of a given property.

pub fn clone_perspective(&self) -> T

Gets the computed value of a given property.

pub fn clone_quotes(&self) -> T

Gets the computed value of a given property.

pub fn clone_rotate(&self) -> T

Gets the computed value of a given property.

pub fn clone_scale(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_indent(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_overflow(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_shadow(&self) -> T

Gets the computed value of a given property.

pub fn clone_transform(&self) -> T

Gets the computed value of a given property.

pub fn clone_transform_origin(&self) -> T

Gets the computed value of a given property.

pub fn clone_transition_behavior(&self) -> T

Gets the computed value of a given property.

pub fn clone_transition_delay(&self) -> T

Gets the computed value of a given property.

pub fn clone_transition_duration(&self) -> T

Gets the computed value of a given property.

pub fn clone_transition_property(&self) -> T

Gets the computed value of a given property.

pub fn clone_transition_timing_function(&self) -> T

Gets the computed value of a given property.

pub fn clone_translate(&self) -> T

Gets the computed value of a given property.

pub fn clone_vertical_align(&self) -> T

Gets the computed value of a given property.

pub fn clone_word_spacing(&self) -> T

Gets the computed value of a given property.

pub fn clone__x_lang(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_image_source(&self) -> T

Gets the computed value of a given property.

pub fn clone_list_style_image(&self) -> T

Gets the computed value of a given property.

pub fn clone_object_position(&self) -> T

Gets the computed value of a given property.

pub fn clone_perspective_origin(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_template_columns(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_template_rows(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_auto_columns(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_auto_rows(&self) -> T

Gets the computed value of a given property.

pub fn clone_column_gap(&self) -> T

Gets the computed value of a given property.

pub fn clone_row_gap(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_column_end(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_column_start(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_row_end(&self) -> T

Gets the computed value of a given property.

pub fn clone_grid_row_start(&self) -> T

Gets the computed value of a given property.

pub fn clone_max_height(&self) -> T

Gets the computed value of a given property.

pub fn clone_max_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_bottom_left_radius(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_bottom_right_radius(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_top_left_radius(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_top_right_radius(&self) -> T

Gets the computed value of a given property.

pub fn clone_bottom(&self) -> T

Gets the computed value of a given property.

pub fn clone_left(&self) -> T

Gets the computed value of a given property.

pub fn clone_right(&self) -> T

Gets the computed value of a given property.

pub fn clone_top(&self) -> T

Gets the computed value of a given property.

pub fn clone_margin_bottom(&self) -> T

Gets the computed value of a given property.

pub fn clone_margin_left(&self) -> T

Gets the computed value of a given property.

pub fn clone_margin_right(&self) -> T

Gets the computed value of a given property.

pub fn clone_margin_top(&self) -> T

Gets the computed value of a given property.

pub fn clone_padding_bottom(&self) -> T

Gets the computed value of a given property.

pub fn clone_padding_left(&self) -> T

Gets the computed value of a given property.

pub fn clone_padding_right(&self) -> T

Gets the computed value of a given property.

pub fn clone_padding_top(&self) -> T

Gets the computed value of a given property.

pub fn clone_height(&self) -> T

Gets the computed value of a given property.

pub fn clone_min_height(&self) -> T

Gets the computed value of a given property.

pub fn clone_min_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_bottom_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_left_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_right_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_top_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_outline_width(&self) -> T

Gets the computed value of a given property.

pub fn clone_background_color(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_bottom_color(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_left_color(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_right_color(&self) -> T

Gets the computed value of a given property.

pub fn clone_border_top_color(&self) -> T

Gets the computed value of a given property.

pub fn clone_outline_color(&self) -> T

Gets the computed value of a given property.

pub fn clone_text_decoration_color(&self) -> T

Gets the computed value of a given property.

pub fn computed_or_resolved_value( &self, property_id: LonghandId, context: Option<&Context<'_>>, dest: &mut CssStringWriter, ) -> Result

Writes the (resolved or computed) value of the given longhand as a string in dest.

TODO(emilio): We should move all the special resolution from nsComputedDOMStyle to ToResolvedValue instead.

pub fn computed_or_resolved_declaration( &self, property_id: LonghandId, context: Option<&Context<'_>>, ) -> PropertyDeclaration

Returns the given longhand’s resolved value as a property declaration.

pub fn resolve_color(&self, color: Color) -> AbsoluteColor

Resolves the currentColor keyword.

Any color value from computed values (except for the ‘color’ property itself) should go through this method.

Usage example: let top_color = style.resolve_color(style.get_border().clone_border_top_color());

pub fn transition_properties<'a>(&'a self) -> TransitionPropertyIterator<'a>

Create a TransitionPropertyIterator for this styles transition properties.

impl ComputedValues

pub fn new( pseudo: Option<&PseudoElement>, custom_properties: ComputedCustomProperties, writing_mode: WritingMode, effective_zoom: Zoom, flags: ComputedValueFlags, rules: Option<StrongRuleNode>, visited_style: Option<Arc<ComputedValues>>, background: Arc<Background>, border: Arc<Border>, box_: Arc<Box>, column: Arc<Column>, counters: Arc<Counters>, effects: Arc<Effects>, font: Arc<Font>, inherited_box: Arc<InheritedBox>, inherited_table: Arc<InheritedTable>, inherited_text: Arc<InheritedText>, inherited_ui: Arc<InheritedUI>, list: Arc<List>, margin: Arc<Margin>, outline: Arc<Outline>, padding: Arc<Padding>, position: Arc<Position>, svg: Arc<SVG>, table: Arc<Table>, text: Arc<Text>, ui: Arc<UI>, ) -> Arc<Self>

Create a new refcounted ComputedValues

pub fn initial_values_with_font_override(default_font: Font) -> Arc<Self>

Get the initial computed values.

pub fn to_arc(&self) -> Arc<Self>

Converts the computed values to an Arc<> from a reference.

pub fn computed_value_to_string( &self, property: PropertyDeclarationId<'_>, ) -> String

Serializes the computed value of this property as a string.