Struct style::properties::generated::ComputedValuesInner

pub struct ComputedValuesInner {

    background: Arc<Background>,
    border: Arc<Border>,
    box_: Arc<Box>,
    column: Arc<Column>,
    counters: Arc<Counters>,
    effects: Arc<Effects>,
    font: Arc<Font>,
    inherited_box: Arc<InheritedBox>,
    inherited_table: Arc<InheritedTable>,
    inherited_text: Arc<InheritedText>,
    inherited_ui: Arc<InheritedUI>,
    list: Arc<List>,
    margin: Arc<Margin>,
    outline: Arc<Outline>,
    padding: Arc<Padding>,
    position: Arc<Position>,
    svg: Arc<SVG>,
    table: Arc<Table>,
    text: Arc<Text>,
    ui: Arc<UI>,
    custom_properties: ComputedCustomProperties,
    pub writing_mode: WritingMode,
    pub effective_zoom: Zoom,
    pub flags: ComputedValueFlags,
    pub rules: Option<StrongRuleNode>,
    visited_style: Option<Arc<ComputedValues>>,

}

Actual data of ComputedValues, to match up with Gecko

Fields
background: Arc<Background>border: Arc<Border>box_: Arc<Box>column: Arc<Column>counters: Arc<Counters>effects: Arc<Effects>font: Arc<Font>inherited_box: Arc<InheritedBox>inherited_table: Arc<InheritedTable>inherited_text: Arc<InheritedText>inherited_ui: Arc<InheritedUI>list: Arc<List>margin: Arc<Margin>outline: Arc<Outline>padding: Arc<Padding>position: Arc<Position>svg: Arc<SVG>table: Arc<Table>text: Arc<Text>ui: Arc<UI>custom_properties: ComputedCustomPropertieswriting_mode: WritingMode

The writing mode of this computed values struct.

effective_zoom: Zoom

The effective zoom value.

flags: ComputedValueFlags

A set of flags we use to store misc information regarding this style.

rules: Option<StrongRuleNode>

The rule node representing the ordered list of rules matched for this node. Can be None for default values and text nodes. This is essentially an optimization to avoid referencing the root rule node.

visited_style: Option<Arc<ComputedValues>>

The element’s computed values if visited, only computed if there’s a relevant link for this element. A element’s “relevant link” is the element being matched if it is a link or the nearest ancestor link.

Implementations

impl ComputedValuesInner

pub fn visited_style(&self) -> Option<&ComputedValues>

Returns the visited style, if any.

pub fn clone_background(&self) -> Arc<Background>

Clone the Background struct.

pub fn get_background(&self) -> &Background

Get a immutable reference to the Background struct.

pub fn mutate_background(&mut self) -> &mut Background

Get a mutable reference to the Background struct.

pub fn clone_border(&self) -> Arc<Border>

Clone the Border struct.

pub fn get_border(&self) -> &Border

Get a immutable reference to the Border struct.

pub fn mutate_border(&mut self) -> &mut Border

Get a mutable reference to the Border struct.

pub fn clone_box(&self) -> Arc<Box>

Clone the Box struct.

pub fn get_box(&self) -> &Box

Get a immutable reference to the Box struct.

pub fn mutate_box(&mut self) -> &mut Box

Get a mutable reference to the Box struct.

pub fn clone_column(&self) -> Arc<Column>

Clone the Column struct.

pub fn get_column(&self) -> &Column

Get a immutable reference to the Column struct.

pub fn mutate_column(&mut self) -> &mut Column

Get a mutable reference to the Column struct.

pub fn clone_counters(&self) -> Arc<Counters>

Clone the Counters struct.

pub fn get_counters(&self) -> &Counters

Get a immutable reference to the Counters struct.

pub fn mutate_counters(&mut self) -> &mut Counters

Get a mutable reference to the Counters struct.

pub fn clone_effects(&self) -> Arc<Effects>

Clone the Effects struct.

pub fn get_effects(&self) -> &Effects

Get a immutable reference to the Effects struct.

pub fn mutate_effects(&mut self) -> &mut Effects

Get a mutable reference to the Effects struct.

pub fn clone_font(&self) -> Arc<Font>

Clone the Font struct.

pub fn get_font(&self) -> &Font

Get a immutable reference to the Font struct.

pub fn mutate_font(&mut self) -> &mut Font

Get a mutable reference to the Font struct.

pub fn clone_inherited_box(&self) -> Arc<InheritedBox>

Clone the InheritedBox struct.

pub fn get_inherited_box(&self) -> &InheritedBox

Get a immutable reference to the InheritedBox struct.

pub fn mutate_inherited_box(&mut self) -> &mut InheritedBox

Get a mutable reference to the InheritedBox struct.

pub fn clone_inherited_table(&self) -> Arc<InheritedTable>

Clone the InheritedTable struct.

pub fn get_inherited_table(&self) -> &InheritedTable

Get a immutable reference to the InheritedTable struct.

pub fn mutate_inherited_table(&mut self) -> &mut InheritedTable

Get a mutable reference to the InheritedTable struct.

pub fn clone_inherited_text(&self) -> Arc<InheritedText>

Clone the InheritedText struct.

pub fn get_inherited_text(&self) -> &InheritedText

Get a immutable reference to the InheritedText struct.

pub fn mutate_inherited_text(&mut self) -> &mut InheritedText

Get a mutable reference to the InheritedText struct.

pub fn clone_inherited_ui(&self) -> Arc<InheritedUI>

Clone the InheritedUI struct.

pub fn get_inherited_ui(&self) -> &InheritedUI

Get a immutable reference to the InheritedUI struct.

pub fn mutate_inherited_ui(&mut self) -> &mut InheritedUI

Get a mutable reference to the InheritedUI struct.

pub fn clone_list(&self) -> Arc<List>

Clone the List struct.

pub fn get_list(&self) -> &List

Get a immutable reference to the List struct.

pub fn mutate_list(&mut self) -> &mut List

Get a mutable reference to the List struct.

pub fn clone_margin(&self) -> Arc<Margin>

Clone the Margin struct.

pub fn get_margin(&self) -> &Margin

Get a immutable reference to the Margin struct.

pub fn mutate_margin(&mut self) -> &mut Margin

Get a mutable reference to the Margin struct.

pub fn clone_outline(&self) -> Arc<Outline>

Clone the Outline struct.

pub fn get_outline(&self) -> &Outline

Get a immutable reference to the Outline struct.

pub fn mutate_outline(&mut self) -> &mut Outline

Get a mutable reference to the Outline struct.

pub fn clone_padding(&self) -> Arc<Padding>

Clone the Padding struct.

pub fn get_padding(&self) -> &Padding

Get a immutable reference to the Padding struct.

pub fn mutate_padding(&mut self) -> &mut Padding

Get a mutable reference to the Padding struct.

pub fn clone_position(&self) -> Arc<Position>

Clone the Position struct.

pub fn get_position(&self) -> &Position

Get a immutable reference to the Position struct.

pub fn mutate_position(&mut self) -> &mut Position

Get a mutable reference to the Position struct.

pub fn clone_svg(&self) -> Arc<SVG>

Clone the SVG struct.

pub fn get_svg(&self) -> &SVG

Get a immutable reference to the SVG struct.

pub fn mutate_svg(&mut self) -> &mut SVG

Get a mutable reference to the SVG struct.

pub fn clone_table(&self) -> Arc<Table>

Clone the Table struct.

pub fn get_table(&self) -> &Table

Get a immutable reference to the Table struct.

pub fn mutate_table(&mut self) -> &mut Table

Get a mutable reference to the Table struct.

pub fn clone_text(&self) -> Arc<Text>

Clone the Text struct.

pub fn get_text(&self) -> &Text

Get a immutable reference to the Text struct.

pub fn mutate_text(&mut self) -> &mut Text

Get a mutable reference to the Text struct.

pub fn clone_ui(&self) -> Arc<UI>

Clone the UI struct.

pub fn get_ui(&self) -> &UI

Get a immutable reference to the UI struct.

pub fn mutate_ui(&mut self) -> &mut UI

Get a mutable reference to the UI struct.

pub fn rules(&self) -> &StrongRuleNode

Gets a reference to the rule node. Panic if no rule node exists.

pub fn ineffective_content_property(&self) -> bool

Returns whether the “content” property for the given style is completely ineffective, and would yield an empty ::before or ::after pseudo-element.

pub fn can_be_fragmented(&self) -> bool

Whether the current style or any of its ancestors is multicolumn.

pub fn is_multicol(&self) -> bool

Whether the current style is multicolumn.

pub fn content_inline_size(&self) -> &Size

Get the logical computed inline size.

pub fn content_block_size(&self) -> &Size

Get the logical computed block size.

pub fn min_inline_size(&self) -> &Size

Get the logical computed min inline size.

pub fn min_block_size(&self) -> &Size

Get the logical computed min block size.

pub fn max_inline_size(&self) -> &MaxSize

Get the logical computed max inline size.

pub fn max_block_size(&self) -> &MaxSize

Get the logical computed max block size.

pub fn logical_padding(&self) -> LogicalMargin<&LengthPercentage>

Get the logical computed padding for this writing mode.

pub fn border_width_for_writing_mode( &self, writing_mode: WritingMode, ) -> LogicalMargin<Au>

Get the logical border width

pub fn logical_border_width(&self) -> LogicalMargin<Au>

Gets the logical computed border widths for this style.

pub fn logical_margin(&self) -> LogicalMargin<&Margin>

Gets the logical computed margin from this style.

pub fn logical_position(&self) -> LogicalMargin<&Inset>

Gets the logical position from this style.

pub fn overrides_transform_style(&self) -> bool

Return true if the effects force the transform style to be Flat

pub fn get_used_transform_style(&self) -> T

https://drafts.csswg.org/css-transforms/#grouping-property-values

pub fn transform_requires_layer(&self) -> bool

Whether given this transform value, the compositor would require a layer.