style::properties::generated::style_structs

Struct Box

pub struct Box {

    pub _moz_default_appearance: T,
    pub alignment_baseline: T,
    pub appearance: T,
    pub baseline_shift: T,
    pub baseline_source: T,
    pub clear: T,
    pub contain: T,
    pub container_name: T,
    pub container_type: T,
    pub display: T,
    pub float: T,
    pub offset_path: T,
    pub overflow_x: T,
    pub overflow_y: T,
    pub overscroll_behavior_x: T,
    pub overscroll_behavior_y: T,
    pub perspective: T,
    pub perspective_origin: T,
    pub position: T,
    pub rotate: T,
    pub scale: T,
    pub touch_action: T,
    pub transform: T,
    pub transform_origin: T,
    pub transform_style: T,
    pub translate: T,
    pub will_change: T,
    pub zoom: T,
    pub _servo_top_layer: T,
    pub scroll_behavior: T,
    pub isolation: T,
    pub backface_visibility: T,
    pub original_display: T,

}

The Box style struct.

Fields
_moz_default_appearance: T

The -moz-default-appearance computed value.

alignment_baseline: T

The alignment-baseline computed value.

appearance: T

The appearance computed value.

baseline_shift: T

The baseline-shift computed value.

baseline_source: T

The baseline-source computed value.

clear: T

The clear computed value.

contain: T

The contain computed value.

container_name: T

The container-name computed value.

container_type: T

The container-type computed value.

display: T

The display computed value.

float: T

The float computed value.

offset_path: T

The offset-path computed value.

overflow_x: T

The overflow-x computed value.

overflow_y: T

The overflow-y computed value.

overscroll_behavior_x: T

The overscroll-behavior-x computed value.

overscroll_behavior_y: T

The overscroll-behavior-y computed value.

perspective: T

The perspective computed value.

perspective_origin: T

The perspective-origin computed value.

position: T

The position computed value.

rotate: T

The rotate computed value.

scale: T

The scale computed value.

touch_action: T

The touch-action computed value.

transform: T

The transform computed value.

transform_origin: T

The transform-origin computed value.

transform_style: T

The transform-style computed value.

translate: T

The translate computed value.

will_change: T

The will-change computed value.

zoom: T

The zoom computed value.

_servo_top_layer: T

The -servo-top-layer computed value.

scroll_behavior: T

The scroll-behavior computed value.

isolation: T

The isolation computed value.

backface_visibility: T

The backface-visibility computed value.

original_display: T

The display value specified by the CSS stylesheets (without any style adjustments), which is needed for hypothetical layout boxes.

Implementations

impl Box

pub fn set__moz_default_appearance(&mut self, v: T)

Set -moz-default-appearance.

pub fn copy__moz_default_appearance_from(&mut self, other: &Self)

Set -moz-default-appearance from other struct.

pub fn reset__moz_default_appearance(&mut self, other: &Self)

Reset -moz-default-appearance from the initial struct.

pub fn clone__moz_default_appearance(&self) -> T

Get the computed value for -moz-default-appearance.

pub fn _moz_default_appearance_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for -moz-default-appearance.

pub fn set_alignment_baseline(&mut self, v: T)

Set alignment-baseline.

pub fn copy_alignment_baseline_from(&mut self, other: &Self)

Set alignment-baseline from other struct.

pub fn reset_alignment_baseline(&mut self, other: &Self)

Reset alignment-baseline from the initial struct.

pub fn clone_alignment_baseline(&self) -> T

Get the computed value for alignment-baseline.

pub fn alignment_baseline_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for alignment-baseline.

pub fn set_appearance(&mut self, v: T)

Set appearance.

pub fn copy_appearance_from(&mut self, other: &Self)

Set appearance from other struct.

pub fn reset_appearance(&mut self, other: &Self)

Reset appearance from the initial struct.

pub fn clone_appearance(&self) -> T

Get the computed value for appearance.

pub fn appearance_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for appearance.

pub fn set_baseline_shift(&mut self, v: T)

Set baseline-shift.

pub fn copy_baseline_shift_from(&mut self, other: &Self)

Set baseline-shift from other struct.

pub fn reset_baseline_shift(&mut self, other: &Self)

Reset baseline-shift from the initial struct.

pub fn clone_baseline_shift(&self) -> T

Get the computed value for baseline-shift.

pub fn baseline_shift_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for baseline-shift.

pub fn set_baseline_source(&mut self, v: T)

Set baseline-source.

pub fn copy_baseline_source_from(&mut self, other: &Self)

Set baseline-source from other struct.

pub fn reset_baseline_source(&mut self, other: &Self)

Reset baseline-source from the initial struct.

pub fn clone_baseline_source(&self) -> T

Get the computed value for baseline-source.

pub fn baseline_source_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for baseline-source.

pub fn set_clear(&mut self, v: T)

Set clear.

pub fn copy_clear_from(&mut self, other: &Self)

Set clear from other struct.

pub fn reset_clear(&mut self, other: &Self)

Reset clear from the initial struct.

pub fn clone_clear(&self) -> T

Get the computed value for clear.

pub fn clear_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for clear.

pub fn set_contain(&mut self, v: T)

Set contain.

pub fn copy_contain_from(&mut self, other: &Self)

Set contain from other struct.

pub fn reset_contain(&mut self, other: &Self)

Reset contain from the initial struct.

pub fn clone_contain(&self) -> T

Get the computed value for contain.

pub fn contain_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for contain.

pub fn set_container_name(&mut self, v: T)

Set container-name.

pub fn copy_container_name_from(&mut self, other: &Self)

Set container-name from other struct.

pub fn reset_container_name(&mut self, other: &Self)

Reset container-name from the initial struct.

pub fn clone_container_name(&self) -> T

Get the computed value for container-name.

pub fn container_name_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for container-name.

pub fn set_container_type(&mut self, v: T)

Set container-type.

pub fn copy_container_type_from(&mut self, other: &Self)

Set container-type from other struct.

pub fn reset_container_type(&mut self, other: &Self)

Reset container-type from the initial struct.

pub fn clone_container_type(&self) -> T

Get the computed value for container-type.

pub fn container_type_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for container-type.

pub fn set_display(&mut self, v: T)

Set display.

We need to keep track of the original display for hypothetical boxes, so we need to special-case this.

pub fn copy_display_from(&mut self, other: &Self)

Set display from other struct.

Same as set_display above. Thus, we need to special-case this.

pub fn reset_display(&mut self, other: &Self)

Reset display from the initial struct.

pub fn clone_display(&self) -> T

Get the computed value for display.

pub fn display_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for display.

pub fn set_float(&mut self, v: T)

Set float.

pub fn copy_float_from(&mut self, other: &Self)

Set float from other struct.

pub fn reset_float(&mut self, other: &Self)

Reset float from the initial struct.

pub fn clone_float(&self) -> T

Get the computed value for float.

pub fn float_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for float.

pub fn set_offset_path(&mut self, v: T)

Set offset-path.

pub fn copy_offset_path_from(&mut self, other: &Self)

Set offset-path from other struct.

pub fn reset_offset_path(&mut self, other: &Self)

Reset offset-path from the initial struct.

pub fn clone_offset_path(&self) -> T

Get the computed value for offset-path.

pub fn offset_path_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for offset-path.

pub fn set_overflow_x(&mut self, v: T)

Set overflow-x.

pub fn copy_overflow_x_from(&mut self, other: &Self)

Set overflow-x from other struct.

pub fn reset_overflow_x(&mut self, other: &Self)

Reset overflow-x from the initial struct.

pub fn clone_overflow_x(&self) -> T

Get the computed value for overflow-x.

pub fn overflow_x_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for overflow-x.

pub fn set_overflow_y(&mut self, v: T)

Set overflow-y.

pub fn copy_overflow_y_from(&mut self, other: &Self)

Set overflow-y from other struct.

pub fn reset_overflow_y(&mut self, other: &Self)

Reset overflow-y from the initial struct.

pub fn clone_overflow_y(&self) -> T

Get the computed value for overflow-y.

pub fn overflow_y_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for overflow-y.

pub fn set_overscroll_behavior_x(&mut self, v: T)

Set overscroll-behavior-x.

pub fn copy_overscroll_behavior_x_from(&mut self, other: &Self)

Set overscroll-behavior-x from other struct.

pub fn reset_overscroll_behavior_x(&mut self, other: &Self)

Reset overscroll-behavior-x from the initial struct.

pub fn clone_overscroll_behavior_x(&self) -> T

Get the computed value for overscroll-behavior-x.

pub fn overscroll_behavior_x_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for overscroll-behavior-x.

pub fn set_overscroll_behavior_y(&mut self, v: T)

Set overscroll-behavior-y.

pub fn copy_overscroll_behavior_y_from(&mut self, other: &Self)

Set overscroll-behavior-y from other struct.

pub fn reset_overscroll_behavior_y(&mut self, other: &Self)

Reset overscroll-behavior-y from the initial struct.

pub fn clone_overscroll_behavior_y(&self) -> T

Get the computed value for overscroll-behavior-y.

pub fn overscroll_behavior_y_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for overscroll-behavior-y.

pub fn set_perspective(&mut self, v: T)

Set perspective.

pub fn copy_perspective_from(&mut self, other: &Self)

Set perspective from other struct.

pub fn reset_perspective(&mut self, other: &Self)

Reset perspective from the initial struct.

pub fn clone_perspective(&self) -> T

Get the computed value for perspective.

pub fn perspective_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for perspective.

pub fn set_perspective_origin(&mut self, v: T)

Set perspective-origin.

pub fn copy_perspective_origin_from(&mut self, other: &Self)

Set perspective-origin from other struct.

pub fn reset_perspective_origin(&mut self, other: &Self)

Reset perspective-origin from the initial struct.

pub fn clone_perspective_origin(&self) -> T

Get the computed value for perspective-origin.

pub fn perspective_origin_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for perspective-origin.

pub fn set_position(&mut self, v: T)

Set position.

pub fn copy_position_from(&mut self, other: &Self)

Set position from other struct.

pub fn reset_position(&mut self, other: &Self)

Reset position from the initial struct.

pub fn clone_position(&self) -> T

Get the computed value for position.

pub fn position_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for position.

pub fn set_rotate(&mut self, v: T)

Set rotate.

pub fn copy_rotate_from(&mut self, other: &Self)

Set rotate from other struct.

pub fn reset_rotate(&mut self, other: &Self)

Reset rotate from the initial struct.

pub fn clone_rotate(&self) -> T

Get the computed value for rotate.

pub fn rotate_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for rotate.

pub fn set_scale(&mut self, v: T)

Set scale.

pub fn copy_scale_from(&mut self, other: &Self)

Set scale from other struct.

pub fn reset_scale(&mut self, other: &Self)

Reset scale from the initial struct.

pub fn clone_scale(&self) -> T

Get the computed value for scale.

pub fn scale_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for scale.

pub fn set_touch_action(&mut self, v: T)

Set touch-action.

pub fn copy_touch_action_from(&mut self, other: &Self)

Set touch-action from other struct.

pub fn reset_touch_action(&mut self, other: &Self)

Reset touch-action from the initial struct.

pub fn clone_touch_action(&self) -> T

Get the computed value for touch-action.

pub fn touch_action_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for touch-action.

pub fn set_transform(&mut self, v: T)

Set transform.

pub fn copy_transform_from(&mut self, other: &Self)

Set transform from other struct.

pub fn reset_transform(&mut self, other: &Self)

Reset transform from the initial struct.

pub fn clone_transform(&self) -> T

Get the computed value for transform.

pub fn transform_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transform.

pub fn set_transform_origin(&mut self, v: T)

Set transform-origin.

pub fn copy_transform_origin_from(&mut self, other: &Self)

Set transform-origin from other struct.

pub fn reset_transform_origin(&mut self, other: &Self)

Reset transform-origin from the initial struct.

pub fn clone_transform_origin(&self) -> T

Get the computed value for transform-origin.

pub fn transform_origin_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transform-origin.

pub fn set_transform_style(&mut self, v: T)

Set transform-style.

pub fn copy_transform_style_from(&mut self, other: &Self)

Set transform-style from other struct.

pub fn reset_transform_style(&mut self, other: &Self)

Reset transform-style from the initial struct.

pub fn clone_transform_style(&self) -> T

Get the computed value for transform-style.

pub fn transform_style_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transform-style.

pub fn set_translate(&mut self, v: T)

Set translate.

pub fn copy_translate_from(&mut self, other: &Self)

Set translate from other struct.

pub fn reset_translate(&mut self, other: &Self)

Reset translate from the initial struct.

pub fn clone_translate(&self) -> T

Get the computed value for translate.

pub fn translate_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for translate.

pub fn set_will_change(&mut self, v: T)

Set will-change.

pub fn copy_will_change_from(&mut self, other: &Self)

Set will-change from other struct.

pub fn reset_will_change(&mut self, other: &Self)

Reset will-change from the initial struct.

pub fn clone_will_change(&self) -> T

Get the computed value for will-change.

pub fn will_change_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for will-change.

pub fn set_zoom(&mut self, v: T)

Set zoom.

pub fn copy_zoom_from(&mut self, other: &Self)

Set zoom from other struct.

pub fn reset_zoom(&mut self, other: &Self)

Reset zoom from the initial struct.

pub fn clone_zoom(&self) -> T

Get the computed value for zoom.

pub fn zoom_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for zoom.

pub fn set__servo_top_layer(&mut self, v: T)

Set -servo-top-layer.

pub fn copy__servo_top_layer_from(&mut self, other: &Self)

Set -servo-top-layer from other struct.

pub fn reset__servo_top_layer(&mut self, other: &Self)

Reset -servo-top-layer from the initial struct.

pub fn clone__servo_top_layer(&self) -> T

Get the computed value for -servo-top-layer.

pub fn _servo_top_layer_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for -servo-top-layer.

pub fn set_scroll_behavior(&mut self, v: T)

Set scroll-behavior.

pub fn copy_scroll_behavior_from(&mut self, other: &Self)

Set scroll-behavior from other struct.

pub fn reset_scroll_behavior(&mut self, other: &Self)

Reset scroll-behavior from the initial struct.

pub fn clone_scroll_behavior(&self) -> T

Get the computed value for scroll-behavior.

pub fn scroll_behavior_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for scroll-behavior.

pub fn set_isolation(&mut self, v: T)

Set isolation.

pub fn copy_isolation_from(&mut self, other: &Self)

Set isolation from other struct.

pub fn reset_isolation(&mut self, other: &Self)

Reset isolation from the initial struct.

pub fn clone_isolation(&self) -> T

Get the computed value for isolation.

pub fn isolation_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for isolation.

pub fn set_backface_visibility(&mut self, v: T)

Set backface-visibility.

pub fn copy_backface_visibility_from(&mut self, other: &Self)

Set backface-visibility from other struct.

pub fn reset_backface_visibility(&mut self, other: &Self)

Reset backface-visibility from the initial struct.

pub fn clone_backface_visibility(&self) -> T

Get the computed value for backface-visibility.

pub fn backface_visibility_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for backface-visibility.

pub fn set_adjusted_display(&mut self, dpy: T, is_item_or_root: bool)

Sets the display property, but without touching original_display, except when the adjustment comes from root or item display fixups.

impl Box

pub fn has_transform_or_perspective(&self) -> bool

Whether there is a non-default transform or perspective style set