Struct style::properties::generated::style_structs::Box

pub struct Box {

    pub display: T,
    pub _servo_top_layer: T,
    pub position: T,
    pub float: T,
    pub clear: T,
    pub vertical_align: T,
    pub baseline_source: T,
    pub _servo_overflow_clip_box: T,
    pub overflow_y: T,
    pub overflow_x: T,
    pub transform: T,
    pub rotate: T,
    pub scale: T,
    pub translate: T,
    pub perspective: T,
    pub perspective_origin: T,
    pub backface_visibility: T,
    pub transform_style: T,
    pub transform_origin: T,
    pub container_type: T,
    pub container_name: T,
    pub zoom: T,
    pub original_display: T,

}

The Box style struct.

Fields
display: T

The display computed value.

_servo_top_layer: T

The -servo-top-layer computed value.

position: T

The position computed value.

float: T

The float computed value.

clear: T

The clear computed value.

vertical_align: T

The vertical-align computed value.

baseline_source: T

The baseline-source computed value.

_servo_overflow_clip_box: T

The -servo-overflow-clip-box computed value.

overflow_y: T

The overflow-y computed value.

overflow_x: T

The overflow-x computed value.

transform: T

The transform computed value.

rotate: T

The rotate computed value.

scale: T

The scale computed value.

translate: T

The translate computed value.

perspective: T

The perspective computed value.

perspective_origin: T

The perspective-origin computed value.

backface_visibility: T

The backface-visibility computed value.

transform_style: T

The transform-style computed value.

transform_origin: T

The transform-origin computed value.

container_type: T

The container-type computed value.

container_name: T

The container-name computed value.

zoom: T

The zoom computed value.

original_display: T

The display value specified by the CSS stylesheets (without any style adjustments), which is needed for hypothetical layout boxes.

Implementations

impl Box

pub fn set_display(&mut self, v: T)

Set display.

We need to keep track of the original display for hypothetical boxes, so we need to special-case this.

pub fn copy_display_from(&mut self, other: &Self)

Set display from other struct.

Same as set_display above. Thus, we need to special-case this.

pub fn reset_display(&mut self, other: &Self)

Reset display from the initial struct.

pub fn clone_display(&self) -> T

Get the computed value for display.

pub fn set__servo_top_layer(&mut self, v: T)

Set -servo-top-layer.

pub fn copy__servo_top_layer_from(&mut self, other: &Self)

Set -servo-top-layer from other struct.

pub fn reset__servo_top_layer(&mut self, other: &Self)

Reset -servo-top-layer from the initial struct.

pub fn clone__servo_top_layer(&self) -> T

Get the computed value for -servo-top-layer.

pub fn set_position(&mut self, v: T)

Set position.

pub fn copy_position_from(&mut self, other: &Self)

Set position from other struct.

pub fn reset_position(&mut self, other: &Self)

Reset position from the initial struct.

pub fn clone_position(&self) -> T

Get the computed value for position.

pub fn set_float(&mut self, v: T)

Set float.

pub fn copy_float_from(&mut self, other: &Self)

Set float from other struct.

pub fn reset_float(&mut self, other: &Self)

Reset float from the initial struct.

pub fn clone_float(&self) -> T

Get the computed value for float.

pub fn set_clear(&mut self, v: T)

Set clear.

pub fn copy_clear_from(&mut self, other: &Self)

Set clear from other struct.

pub fn reset_clear(&mut self, other: &Self)

Reset clear from the initial struct.

pub fn clone_clear(&self) -> T

Get the computed value for clear.

pub fn set_vertical_align(&mut self, v: T)

Set vertical-align.

pub fn copy_vertical_align_from(&mut self, other: &Self)

Set vertical-align from other struct.

pub fn reset_vertical_align(&mut self, other: &Self)

Reset vertical-align from the initial struct.

pub fn clone_vertical_align(&self) -> T

Get the computed value for vertical-align.

pub fn set_baseline_source(&mut self, v: T)

Set baseline-source.

pub fn copy_baseline_source_from(&mut self, other: &Self)

Set baseline-source from other struct.

pub fn reset_baseline_source(&mut self, other: &Self)

Reset baseline-source from the initial struct.

pub fn clone_baseline_source(&self) -> T

Get the computed value for baseline-source.

pub fn set__servo_overflow_clip_box(&mut self, v: T)

Set -servo-overflow-clip-box.

pub fn copy__servo_overflow_clip_box_from(&mut self, other: &Self)

Set -servo-overflow-clip-box from other struct.

pub fn reset__servo_overflow_clip_box(&mut self, other: &Self)

Reset -servo-overflow-clip-box from the initial struct.

pub fn clone__servo_overflow_clip_box(&self) -> T

Get the computed value for -servo-overflow-clip-box.

pub fn set_overflow_y(&mut self, v: T)

Set overflow-y.

pub fn copy_overflow_y_from(&mut self, other: &Self)

Set overflow-y from other struct.

pub fn reset_overflow_y(&mut self, other: &Self)

Reset overflow-y from the initial struct.

pub fn clone_overflow_y(&self) -> T

Get the computed value for overflow-y.

pub fn set_overflow_x(&mut self, v: T)

Set overflow-x.

pub fn copy_overflow_x_from(&mut self, other: &Self)

Set overflow-x from other struct.

pub fn reset_overflow_x(&mut self, other: &Self)

Reset overflow-x from the initial struct.

pub fn clone_overflow_x(&self) -> T

Get the computed value for overflow-x.

pub fn set_transform(&mut self, v: T)

Set transform.

pub fn copy_transform_from(&mut self, other: &Self)

Set transform from other struct.

pub fn reset_transform(&mut self, other: &Self)

Reset transform from the initial struct.

pub fn clone_transform(&self) -> T

Get the computed value for transform.

pub fn set_rotate(&mut self, v: T)

Set rotate.

pub fn copy_rotate_from(&mut self, other: &Self)

Set rotate from other struct.

pub fn reset_rotate(&mut self, other: &Self)

Reset rotate from the initial struct.

pub fn clone_rotate(&self) -> T

Get the computed value for rotate.

pub fn set_scale(&mut self, v: T)

Set scale.

pub fn copy_scale_from(&mut self, other: &Self)

Set scale from other struct.

pub fn reset_scale(&mut self, other: &Self)

Reset scale from the initial struct.

pub fn clone_scale(&self) -> T

Get the computed value for scale.

pub fn set_translate(&mut self, v: T)

Set translate.

pub fn copy_translate_from(&mut self, other: &Self)

Set translate from other struct.

pub fn reset_translate(&mut self, other: &Self)

Reset translate from the initial struct.

pub fn clone_translate(&self) -> T

Get the computed value for translate.

pub fn set_perspective(&mut self, v: T)

Set perspective.

pub fn copy_perspective_from(&mut self, other: &Self)

Set perspective from other struct.

pub fn reset_perspective(&mut self, other: &Self)

Reset perspective from the initial struct.

pub fn clone_perspective(&self) -> T

Get the computed value for perspective.

pub fn set_perspective_origin(&mut self, v: T)

Set perspective-origin.

pub fn copy_perspective_origin_from(&mut self, other: &Self)

Set perspective-origin from other struct.

pub fn reset_perspective_origin(&mut self, other: &Self)

Reset perspective-origin from the initial struct.

pub fn clone_perspective_origin(&self) -> T

Get the computed value for perspective-origin.

pub fn set_backface_visibility(&mut self, v: T)

Set backface-visibility.

pub fn copy_backface_visibility_from(&mut self, other: &Self)

Set backface-visibility from other struct.

pub fn reset_backface_visibility(&mut self, other: &Self)

Reset backface-visibility from the initial struct.

pub fn clone_backface_visibility(&self) -> T

Get the computed value for backface-visibility.

pub fn set_transform_style(&mut self, v: T)

Set transform-style.

pub fn copy_transform_style_from(&mut self, other: &Self)

Set transform-style from other struct.

pub fn reset_transform_style(&mut self, other: &Self)

Reset transform-style from the initial struct.

pub fn clone_transform_style(&self) -> T

Get the computed value for transform-style.

pub fn set_transform_origin(&mut self, v: T)

Set transform-origin.

pub fn copy_transform_origin_from(&mut self, other: &Self)

Set transform-origin from other struct.

pub fn reset_transform_origin(&mut self, other: &Self)

Reset transform-origin from the initial struct.

pub fn clone_transform_origin(&self) -> T

Get the computed value for transform-origin.

pub fn set_container_type(&mut self, v: T)

Set container-type.

pub fn copy_container_type_from(&mut self, other: &Self)

Set container-type from other struct.

pub fn reset_container_type(&mut self, other: &Self)

Reset container-type from the initial struct.

pub fn clone_container_type(&self) -> T

Get the computed value for container-type.

pub fn set_container_name(&mut self, v: T)

Set container-name.

pub fn copy_container_name_from(&mut self, other: &Self)

Set container-name from other struct.

pub fn reset_container_name(&mut self, other: &Self)

Reset container-name from the initial struct.

pub fn clone_container_name(&self) -> T

Get the computed value for container-name.

pub fn set_zoom(&mut self, v: T)

Set zoom.

pub fn copy_zoom_from(&mut self, other: &Self)

Set zoom from other struct.

pub fn reset_zoom(&mut self, other: &Self)

Reset zoom from the initial struct.

pub fn clone_zoom(&self) -> T

Get the computed value for zoom.

pub fn set_adjusted_display(&mut self, dpy: T, is_item_or_root: bool)

Sets the display property, but without touching original_display, except when the adjustment comes from root or item display fixups.