style::properties::generated::style_structs

Struct Font

pub struct Font {

    pub _moz_min_font_size_ratio: T,
    pub _x_lang: T,
    pub _x_text_scale: T,
    pub font_family: T,
    pub font_feature_settings: T,
    pub font_language_override: T,
    pub font_size: T,
    pub font_size_adjust: T,
    pub font_stretch: T,
    pub font_style: T,
    pub font_synthesis_weight: T,
    pub font_variant_alternates: T,
    pub font_variant_east_asian: T,
    pub font_variant_ligatures: T,
    pub font_variant_numeric: T,
    pub font_variation_settings: T,
    pub font_weight: T,
    pub line_height: T,
    pub math_depth: T,
    pub font_variant_caps: T,
    pub font_kerning: T,
    pub font_variant_position: T,
    pub font_optical_sizing: T,
    pub math_style: T,
    pub hash: u64,

}

The Font style struct.

Fields
_moz_min_font_size_ratio: T

The -moz-min-font-size-ratio computed value.

_x_lang: T

The -x-lang computed value.

_x_text_scale: T

The -x-text-scale computed value.

font_family: T

The font-family computed value.

font_feature_settings: T

The font-feature-settings computed value.

font_language_override: T

The font-language-override computed value.

font_size: T

The font-size computed value.

font_size_adjust: T

The font-size-adjust computed value.

font_stretch: T

The font-stretch computed value.

font_style: T

The font-style computed value.

font_synthesis_weight: T

The font-synthesis-weight computed value.

font_variant_alternates: T

The font-variant-alternates computed value.

font_variant_east_asian: T

The font-variant-east-asian computed value.

font_variant_ligatures: T

The font-variant-ligatures computed value.

font_variant_numeric: T

The font-variant-numeric computed value.

font_variation_settings: T

The font-variation-settings computed value.

font_weight: T

The font-weight computed value.

line_height: T

The line-height computed value.

math_depth: T

The math-depth computed value.

font_variant_caps: T

The font-variant-caps computed value.

font_kerning: T

The font-kerning computed value.

font_variant_position: T

The font-variant-position computed value.

font_optical_sizing: T

The font-optical-sizing computed value.

math_style: T

The math-style computed value.

hash: u64

The font hash, used for font caching.

Implementations

impl Font

pub fn set__moz_min_font_size_ratio(&mut self, v: T)

Set -moz-min-font-size-ratio.

pub fn copy__moz_min_font_size_ratio_from(&mut self, other: &Self)

Set -moz-min-font-size-ratio from other struct.

pub fn reset__moz_min_font_size_ratio(&mut self, other: &Self)

Reset -moz-min-font-size-ratio from the initial struct.

pub fn clone__moz_min_font_size_ratio(&self) -> T

Get the computed value for -moz-min-font-size-ratio.

pub fn _moz_min_font_size_ratio_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for -moz-min-font-size-ratio.

pub fn set__x_lang(&mut self, v: T)

Set -x-lang.

pub fn copy__x_lang_from(&mut self, other: &Self)

Set -x-lang from other struct.

pub fn reset__x_lang(&mut self, other: &Self)

Reset -x-lang from the initial struct.

pub fn clone__x_lang(&self) -> T

Get the computed value for -x-lang.

pub fn _x_lang_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for -x-lang.

pub fn set__x_text_scale(&mut self, v: T)

Set -x-text-scale.

pub fn copy__x_text_scale_from(&mut self, other: &Self)

Set -x-text-scale from other struct.

pub fn reset__x_text_scale(&mut self, other: &Self)

Reset -x-text-scale from the initial struct.

pub fn clone__x_text_scale(&self) -> T

Get the computed value for -x-text-scale.

pub fn _x_text_scale_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for -x-text-scale.

pub fn set_font_family(&mut self, v: T)

Set font-family.

pub fn copy_font_family_from(&mut self, other: &Self)

Set font-family from other struct.

pub fn reset_font_family(&mut self, other: &Self)

Reset font-family from the initial struct.

pub fn clone_font_family(&self) -> T

Get the computed value for font-family.

pub fn font_family_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-family.

pub fn set_font_feature_settings(&mut self, v: T)

Set font-feature-settings.

pub fn copy_font_feature_settings_from(&mut self, other: &Self)

Set font-feature-settings from other struct.

pub fn reset_font_feature_settings(&mut self, other: &Self)

Reset font-feature-settings from the initial struct.

pub fn clone_font_feature_settings(&self) -> T

Get the computed value for font-feature-settings.

pub fn font_feature_settings_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-feature-settings.

pub fn set_font_language_override(&mut self, v: T)

Set font-language-override.

pub fn copy_font_language_override_from(&mut self, other: &Self)

Set font-language-override from other struct.

pub fn reset_font_language_override(&mut self, other: &Self)

Reset font-language-override from the initial struct.

pub fn clone_font_language_override(&self) -> T

Get the computed value for font-language-override.

pub fn font_language_override_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-language-override.

pub fn set_font_size(&mut self, v: T)

Set font-size.

pub fn copy_font_size_from(&mut self, other: &Self)

Set font-size from other struct.

pub fn reset_font_size(&mut self, other: &Self)

Reset font-size from the initial struct.

pub fn clone_font_size(&self) -> T

Get the computed value for font-size.

pub fn font_size_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-size.

pub fn set_font_size_adjust(&mut self, v: T)

Set font-size-adjust.

pub fn copy_font_size_adjust_from(&mut self, other: &Self)

Set font-size-adjust from other struct.

pub fn reset_font_size_adjust(&mut self, other: &Self)

Reset font-size-adjust from the initial struct.

pub fn clone_font_size_adjust(&self) -> T

Get the computed value for font-size-adjust.

pub fn font_size_adjust_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-size-adjust.

pub fn set_font_stretch(&mut self, v: T)

Set font-stretch.

pub fn copy_font_stretch_from(&mut self, other: &Self)

Set font-stretch from other struct.

pub fn reset_font_stretch(&mut self, other: &Self)

Reset font-stretch from the initial struct.

pub fn clone_font_stretch(&self) -> T

Get the computed value for font-stretch.

pub fn font_stretch_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-stretch.

pub fn set_font_style(&mut self, v: T)

Set font-style.

pub fn copy_font_style_from(&mut self, other: &Self)

Set font-style from other struct.

pub fn reset_font_style(&mut self, other: &Self)

Reset font-style from the initial struct.

pub fn clone_font_style(&self) -> T

Get the computed value for font-style.

pub fn font_style_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-style.

pub fn set_font_synthesis_weight(&mut self, v: T)

Set font-synthesis-weight.

pub fn copy_font_synthesis_weight_from(&mut self, other: &Self)

Set font-synthesis-weight from other struct.

pub fn reset_font_synthesis_weight(&mut self, other: &Self)

Reset font-synthesis-weight from the initial struct.

pub fn clone_font_synthesis_weight(&self) -> T

Get the computed value for font-synthesis-weight.

pub fn font_synthesis_weight_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-synthesis-weight.

pub fn set_font_variant_alternates(&mut self, v: T)

Set font-variant-alternates.

pub fn copy_font_variant_alternates_from(&mut self, other: &Self)

Set font-variant-alternates from other struct.

pub fn reset_font_variant_alternates(&mut self, other: &Self)

Reset font-variant-alternates from the initial struct.

pub fn clone_font_variant_alternates(&self) -> T

Get the computed value for font-variant-alternates.

pub fn font_variant_alternates_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-variant-alternates.

pub fn set_font_variant_east_asian(&mut self, v: T)

Set font-variant-east-asian.

pub fn copy_font_variant_east_asian_from(&mut self, other: &Self)

Set font-variant-east-asian from other struct.

pub fn reset_font_variant_east_asian(&mut self, other: &Self)

Reset font-variant-east-asian from the initial struct.

pub fn clone_font_variant_east_asian(&self) -> T

Get the computed value for font-variant-east-asian.

pub fn font_variant_east_asian_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-variant-east-asian.

pub fn set_font_variant_ligatures(&mut self, v: T)

Set font-variant-ligatures.

pub fn copy_font_variant_ligatures_from(&mut self, other: &Self)

Set font-variant-ligatures from other struct.

pub fn reset_font_variant_ligatures(&mut self, other: &Self)

Reset font-variant-ligatures from the initial struct.

pub fn clone_font_variant_ligatures(&self) -> T

Get the computed value for font-variant-ligatures.

pub fn font_variant_ligatures_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-variant-ligatures.

pub fn set_font_variant_numeric(&mut self, v: T)

Set font-variant-numeric.

pub fn copy_font_variant_numeric_from(&mut self, other: &Self)

Set font-variant-numeric from other struct.

pub fn reset_font_variant_numeric(&mut self, other: &Self)

Reset font-variant-numeric from the initial struct.

pub fn clone_font_variant_numeric(&self) -> T

Get the computed value for font-variant-numeric.

pub fn font_variant_numeric_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-variant-numeric.

pub fn set_font_variation_settings(&mut self, v: T)

Set font-variation-settings.

pub fn copy_font_variation_settings_from(&mut self, other: &Self)

Set font-variation-settings from other struct.

pub fn reset_font_variation_settings(&mut self, other: &Self)

Reset font-variation-settings from the initial struct.

pub fn clone_font_variation_settings(&self) -> T

Get the computed value for font-variation-settings.

pub fn font_variation_settings_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-variation-settings.

pub fn set_font_weight(&mut self, v: T)

Set font-weight.

pub fn copy_font_weight_from(&mut self, other: &Self)

Set font-weight from other struct.

pub fn reset_font_weight(&mut self, other: &Self)

Reset font-weight from the initial struct.

pub fn clone_font_weight(&self) -> T

Get the computed value for font-weight.

pub fn font_weight_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-weight.

pub fn set_line_height(&mut self, v: T)

Set line-height.

pub fn copy_line_height_from(&mut self, other: &Self)

Set line-height from other struct.

pub fn reset_line_height(&mut self, other: &Self)

Reset line-height from the initial struct.

pub fn clone_line_height(&self) -> T

Get the computed value for line-height.

pub fn line_height_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for line-height.

pub fn set_math_depth(&mut self, v: T)

Set math-depth.

pub fn copy_math_depth_from(&mut self, other: &Self)

Set math-depth from other struct.

pub fn reset_math_depth(&mut self, other: &Self)

Reset math-depth from the initial struct.

pub fn clone_math_depth(&self) -> T

Get the computed value for math-depth.

pub fn math_depth_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for math-depth.

pub fn set_font_variant_caps(&mut self, v: T)

Set font-variant-caps.

pub fn copy_font_variant_caps_from(&mut self, other: &Self)

Set font-variant-caps from other struct.

pub fn reset_font_variant_caps(&mut self, other: &Self)

Reset font-variant-caps from the initial struct.

pub fn clone_font_variant_caps(&self) -> T

Get the computed value for font-variant-caps.

pub fn font_variant_caps_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-variant-caps.

pub fn set_font_kerning(&mut self, v: T)

Set font-kerning.

pub fn copy_font_kerning_from(&mut self, other: &Self)

Set font-kerning from other struct.

pub fn reset_font_kerning(&mut self, other: &Self)

Reset font-kerning from the initial struct.

pub fn clone_font_kerning(&self) -> T

Get the computed value for font-kerning.

pub fn font_kerning_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-kerning.

pub fn set_font_variant_position(&mut self, v: T)

Set font-variant-position.

pub fn copy_font_variant_position_from(&mut self, other: &Self)

Set font-variant-position from other struct.

pub fn reset_font_variant_position(&mut self, other: &Self)

Reset font-variant-position from the initial struct.

pub fn clone_font_variant_position(&self) -> T

Get the computed value for font-variant-position.

pub fn font_variant_position_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-variant-position.

pub fn set_font_optical_sizing(&mut self, v: T)

Set font-optical-sizing.

pub fn copy_font_optical_sizing_from(&mut self, other: &Self)

Set font-optical-sizing from other struct.

pub fn reset_font_optical_sizing(&mut self, other: &Self)

Reset font-optical-sizing from the initial struct.

pub fn clone_font_optical_sizing(&self) -> T

Get the computed value for font-optical-sizing.

pub fn font_optical_sizing_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for font-optical-sizing.

pub fn set_math_style(&mut self, v: T)

Set math-style.

pub fn copy_math_style_from(&mut self, other: &Self)

Set math-style from other struct.

pub fn reset_math_style(&mut self, other: &Self)

Reset math-style from the initial struct.

pub fn clone_math_style(&self) -> T

Get the computed value for math-style.

pub fn math_style_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for math-style.

pub fn compute_font_hash(&mut self)

Computes a font hash in order to be able to cache fonts effectively in GFX and layout.

pub fn initial_values() -> Self

Create a new Font with the initial values of all members.