Struct style::properties::generated::style_structs::Font

pub struct Font {
    pub font_family: T,
    pub font_style: T,
    pub font_variant_caps: T,
    pub font_weight: T,
    pub font_size: T,
    pub font_stretch: T,
    pub font_language_override: T,
    pub _x_lang: T,
    pub line_height: T,
    pub hash: u64,
}

The Font style struct.

Fields
font_family: T

The font-family computed value.

font_style: T

The font-style computed value.

font_variant_caps: T

The font-variant-caps computed value.

font_weight: T

The font-weight computed value.

font_size: T

The font-size computed value.

font_stretch: T

The font-stretch computed value.

font_language_override: T

The font-language-override computed value.

_x_lang: T

The -x-lang computed value.

line_height: T

The line-height computed value.

hash: u64

The font hash, used for font caching.

Implementations

impl Font

pub fn set_font_family(&mut self, v: T)

Set font-family.

pub fn copy_font_family_from(&mut self, other: &Self)

Set font-family from other struct.

pub fn reset_font_family(&mut self, other: &Self)

Reset font-family from the initial struct.

pub fn clone_font_family(&self) -> T

Get the computed value for font-family.

pub fn set_font_style(&mut self, v: T)

Set font-style.

pub fn copy_font_style_from(&mut self, other: &Self)

Set font-style from other struct.

pub fn reset_font_style(&mut self, other: &Self)

Reset font-style from the initial struct.

pub fn clone_font_style(&self) -> T

Get the computed value for font-style.

pub fn set_font_variant_caps(&mut self, v: T)

Set font-variant-caps.

pub fn copy_font_variant_caps_from(&mut self, other: &Self)

Set font-variant-caps from other struct.

pub fn reset_font_variant_caps(&mut self, other: &Self)

Reset font-variant-caps from the initial struct.

pub fn clone_font_variant_caps(&self) -> T

Get the computed value for font-variant-caps.

pub fn set_font_weight(&mut self, v: T)

Set font-weight.

pub fn copy_font_weight_from(&mut self, other: &Self)

Set font-weight from other struct.

pub fn reset_font_weight(&mut self, other: &Self)

Reset font-weight from the initial struct.

pub fn clone_font_weight(&self) -> T

Get the computed value for font-weight.

pub fn set_font_size(&mut self, v: T)

Set font-size.

pub fn copy_font_size_from(&mut self, other: &Self)

Set font-size from other struct.

pub fn reset_font_size(&mut self, other: &Self)

Reset font-size from the initial struct.

pub fn clone_font_size(&self) -> T

Get the computed value for font-size.

pub fn set_font_stretch(&mut self, v: T)

Set font-stretch.

pub fn copy_font_stretch_from(&mut self, other: &Self)

Set font-stretch from other struct.

pub fn reset_font_stretch(&mut self, other: &Self)

Reset font-stretch from the initial struct.

pub fn clone_font_stretch(&self) -> T

Get the computed value for font-stretch.

pub fn set_font_language_override(&mut self, v: T)

Set font-language-override.

pub fn copy_font_language_override_from(&mut self, other: &Self)

Set font-language-override from other struct.

pub fn reset_font_language_override(&mut self, other: &Self)

Reset font-language-override from the initial struct.

pub fn clone_font_language_override(&self) -> T

Get the computed value for font-language-override.

pub fn set__x_lang(&mut self, v: T)

Set -x-lang.

pub fn copy__x_lang_from(&mut self, other: &Self)

Set -x-lang from other struct.

pub fn reset__x_lang(&mut self, other: &Self)

Reset -x-lang from the initial struct.

pub fn clone__x_lang(&self) -> T

Get the computed value for -x-lang.

pub fn set_line_height(&mut self, v: T)

Set line-height.

pub fn copy_line_height_from(&mut self, other: &Self)

Set line-height from other struct.

pub fn reset_line_height(&mut self, other: &Self)

Reset line-height from the initial struct.

pub fn clone_line_height(&self) -> T

Get the computed value for line-height.

pub fn compute_font_hash(&mut self)

Computes a font hash in order to be able to cache fonts effectively in GFX and layout.

pub fn initial_values() -> Self

Create a new Font with the initial values of all members.