style::properties::generated::style_structs

Struct Position

pub struct Position {

    pub align_content: T,
    pub align_items: T,
    pub align_self: T,
    pub aspect_ratio: T,
    pub bottom: T,
    pub column_gap: T,
    pub flex_basis: T,
    pub flex_grow: T,
    pub flex_shrink: T,
    pub grid_auto_columns: T,
    pub grid_auto_flow: T,
    pub grid_auto_rows: T,
    pub grid_column_end: T,
    pub grid_column_start: T,
    pub grid_row_end: T,
    pub grid_row_start: T,
    pub grid_template_areas: T,
    pub grid_template_columns: T,
    pub grid_template_rows: T,
    pub height: T,
    pub justify_content: T,
    pub justify_items: T,
    pub justify_self: T,
    pub left: T,
    pub max_height: T,
    pub max_width: T,
    pub min_height: T,
    pub min_width: T,
    pub object_position: T,
    pub order: T,
    pub position_area: T,
    pub position_try_fallbacks: T,
    pub right: T,
    pub row_gap: T,
    pub top: T,
    pub width: T,
    pub z_index: T,
    pub flex_direction: T,
    pub flex_wrap: T,
    pub box_sizing: T,
    pub object_fit: T,

}

The Position style struct.

Fields
align_content: T

The align-content computed value.

align_items: T

The align-items computed value.

align_self: T

The align-self computed value.

aspect_ratio: T

The aspect-ratio computed value.

bottom: T

The bottom computed value.

column_gap: T

The column-gap computed value.

flex_basis: T

The flex-basis computed value.

flex_grow: T

The flex-grow computed value.

flex_shrink: T

The flex-shrink computed value.

grid_auto_columns: T

The grid-auto-columns computed value.

grid_auto_flow: T

The grid-auto-flow computed value.

grid_auto_rows: T

The grid-auto-rows computed value.

grid_column_end: T

The grid-column-end computed value.

grid_column_start: T

The grid-column-start computed value.

grid_row_end: T

The grid-row-end computed value.

grid_row_start: T

The grid-row-start computed value.

grid_template_areas: T

The grid-template-areas computed value.

grid_template_columns: T

The grid-template-columns computed value.

grid_template_rows: T

The grid-template-rows computed value.

height: T

The height computed value.

justify_content: T

The justify-content computed value.

justify_items: T

The justify-items computed value.

justify_self: T

The justify-self computed value.

left: T

The left computed value.

max_height: T

The max-height computed value.

max_width: T

The max-width computed value.

min_height: T

The min-height computed value.

min_width: T

The min-width computed value.

object_position: T

The object-position computed value.

order: T

The order computed value.

position_area: T

The position-area computed value.

position_try_fallbacks: T

The position-try-fallbacks computed value.

right: T

The right computed value.

row_gap: T

The row-gap computed value.

top: T

The top computed value.

width: T

The width computed value.

z_index: T

The z-index computed value.

flex_direction: T

The flex-direction computed value.

flex_wrap: T

The flex-wrap computed value.

box_sizing: T

The box-sizing computed value.

object_fit: T

The object-fit computed value.

Implementations

impl Position

pub fn set_align_content(&mut self, v: T)

Set align-content.

pub fn copy_align_content_from(&mut self, other: &Self)

Set align-content from other struct.

pub fn reset_align_content(&mut self, other: &Self)

Reset align-content from the initial struct.

pub fn clone_align_content(&self) -> T

Get the computed value for align-content.

pub fn align_content_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for align-content.

pub fn set_align_items(&mut self, v: T)

Set align-items.

pub fn copy_align_items_from(&mut self, other: &Self)

Set align-items from other struct.

pub fn reset_align_items(&mut self, other: &Self)

Reset align-items from the initial struct.

pub fn clone_align_items(&self) -> T

Get the computed value for align-items.

pub fn align_items_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for align-items.

pub fn set_align_self(&mut self, v: T)

Set align-self.

pub fn copy_align_self_from(&mut self, other: &Self)

Set align-self from other struct.

pub fn reset_align_self(&mut self, other: &Self)

Reset align-self from the initial struct.

pub fn clone_align_self(&self) -> T

Get the computed value for align-self.

pub fn align_self_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for align-self.

pub fn set_aspect_ratio(&mut self, v: T)

Set aspect-ratio.

pub fn copy_aspect_ratio_from(&mut self, other: &Self)

Set aspect-ratio from other struct.

pub fn reset_aspect_ratio(&mut self, other: &Self)

Reset aspect-ratio from the initial struct.

pub fn clone_aspect_ratio(&self) -> T

Get the computed value for aspect-ratio.

pub fn aspect_ratio_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for aspect-ratio.

pub fn set_bottom(&mut self, v: T)

Set bottom.

pub fn copy_bottom_from(&mut self, other: &Self)

Set bottom from other struct.

pub fn reset_bottom(&mut self, other: &Self)

Reset bottom from the initial struct.

pub fn clone_bottom(&self) -> T

Get the computed value for bottom.

pub fn bottom_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for bottom.

pub fn set_column_gap(&mut self, v: T)

Set column-gap.

pub fn copy_column_gap_from(&mut self, other: &Self)

Set column-gap from other struct.

pub fn reset_column_gap(&mut self, other: &Self)

Reset column-gap from the initial struct.

pub fn clone_column_gap(&self) -> T

Get the computed value for column-gap.

pub fn column_gap_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for column-gap.

pub fn set_flex_basis(&mut self, v: T)

Set flex-basis.

pub fn copy_flex_basis_from(&mut self, other: &Self)

Set flex-basis from other struct.

pub fn reset_flex_basis(&mut self, other: &Self)

Reset flex-basis from the initial struct.

pub fn clone_flex_basis(&self) -> T

Get the computed value for flex-basis.

pub fn flex_basis_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for flex-basis.

pub fn set_flex_grow(&mut self, v: T)

Set flex-grow.

pub fn copy_flex_grow_from(&mut self, other: &Self)

Set flex-grow from other struct.

pub fn reset_flex_grow(&mut self, other: &Self)

Reset flex-grow from the initial struct.

pub fn clone_flex_grow(&self) -> T

Get the computed value for flex-grow.

pub fn flex_grow_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for flex-grow.

pub fn set_flex_shrink(&mut self, v: T)

Set flex-shrink.

pub fn copy_flex_shrink_from(&mut self, other: &Self)

Set flex-shrink from other struct.

pub fn reset_flex_shrink(&mut self, other: &Self)

Reset flex-shrink from the initial struct.

pub fn clone_flex_shrink(&self) -> T

Get the computed value for flex-shrink.

pub fn flex_shrink_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for flex-shrink.

pub fn set_grid_auto_columns(&mut self, v: T)

Set grid-auto-columns.

pub fn copy_grid_auto_columns_from(&mut self, other: &Self)

Set grid-auto-columns from other struct.

pub fn reset_grid_auto_columns(&mut self, other: &Self)

Reset grid-auto-columns from the initial struct.

pub fn clone_grid_auto_columns(&self) -> T

Get the computed value for grid-auto-columns.

pub fn grid_auto_columns_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-auto-columns.

pub fn set_grid_auto_flow(&mut self, v: T)

Set grid-auto-flow.

pub fn copy_grid_auto_flow_from(&mut self, other: &Self)

Set grid-auto-flow from other struct.

pub fn reset_grid_auto_flow(&mut self, other: &Self)

Reset grid-auto-flow from the initial struct.

pub fn clone_grid_auto_flow(&self) -> T

Get the computed value for grid-auto-flow.

pub fn grid_auto_flow_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-auto-flow.

pub fn set_grid_auto_rows(&mut self, v: T)

Set grid-auto-rows.

pub fn copy_grid_auto_rows_from(&mut self, other: &Self)

Set grid-auto-rows from other struct.

pub fn reset_grid_auto_rows(&mut self, other: &Self)

Reset grid-auto-rows from the initial struct.

pub fn clone_grid_auto_rows(&self) -> T

Get the computed value for grid-auto-rows.

pub fn grid_auto_rows_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-auto-rows.

pub fn set_grid_column_end(&mut self, v: T)

Set grid-column-end.

pub fn copy_grid_column_end_from(&mut self, other: &Self)

Set grid-column-end from other struct.

pub fn reset_grid_column_end(&mut self, other: &Self)

Reset grid-column-end from the initial struct.

pub fn clone_grid_column_end(&self) -> T

Get the computed value for grid-column-end.

pub fn grid_column_end_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-column-end.

pub fn set_grid_column_start(&mut self, v: T)

Set grid-column-start.

pub fn copy_grid_column_start_from(&mut self, other: &Self)

Set grid-column-start from other struct.

pub fn reset_grid_column_start(&mut self, other: &Self)

Reset grid-column-start from the initial struct.

pub fn clone_grid_column_start(&self) -> T

Get the computed value for grid-column-start.

pub fn grid_column_start_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-column-start.

pub fn set_grid_row_end(&mut self, v: T)

Set grid-row-end.

pub fn copy_grid_row_end_from(&mut self, other: &Self)

Set grid-row-end from other struct.

pub fn reset_grid_row_end(&mut self, other: &Self)

Reset grid-row-end from the initial struct.

pub fn clone_grid_row_end(&self) -> T

Get the computed value for grid-row-end.

pub fn grid_row_end_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-row-end.

pub fn set_grid_row_start(&mut self, v: T)

Set grid-row-start.

pub fn copy_grid_row_start_from(&mut self, other: &Self)

Set grid-row-start from other struct.

pub fn reset_grid_row_start(&mut self, other: &Self)

Reset grid-row-start from the initial struct.

pub fn clone_grid_row_start(&self) -> T

Get the computed value for grid-row-start.

pub fn grid_row_start_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-row-start.

pub fn set_grid_template_areas(&mut self, v: T)

Set grid-template-areas.

pub fn copy_grid_template_areas_from(&mut self, other: &Self)

Set grid-template-areas from other struct.

pub fn reset_grid_template_areas(&mut self, other: &Self)

Reset grid-template-areas from the initial struct.

pub fn clone_grid_template_areas(&self) -> T

Get the computed value for grid-template-areas.

pub fn grid_template_areas_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-template-areas.

pub fn set_grid_template_columns(&mut self, v: T)

Set grid-template-columns.

pub fn copy_grid_template_columns_from(&mut self, other: &Self)

Set grid-template-columns from other struct.

pub fn reset_grid_template_columns(&mut self, other: &Self)

Reset grid-template-columns from the initial struct.

pub fn clone_grid_template_columns(&self) -> T

Get the computed value for grid-template-columns.

pub fn grid_template_columns_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-template-columns.

pub fn set_grid_template_rows(&mut self, v: T)

Set grid-template-rows.

pub fn copy_grid_template_rows_from(&mut self, other: &Self)

Set grid-template-rows from other struct.

pub fn reset_grid_template_rows(&mut self, other: &Self)

Reset grid-template-rows from the initial struct.

pub fn clone_grid_template_rows(&self) -> T

Get the computed value for grid-template-rows.

pub fn grid_template_rows_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for grid-template-rows.

pub fn set_height(&mut self, v: T)

Set height.

pub fn copy_height_from(&mut self, other: &Self)

Set height from other struct.

pub fn reset_height(&mut self, other: &Self)

Reset height from the initial struct.

pub fn clone_height(&self) -> T

Get the computed value for height.

pub fn height_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for height.

pub fn set_justify_content(&mut self, v: T)

Set justify-content.

pub fn copy_justify_content_from(&mut self, other: &Self)

Set justify-content from other struct.

pub fn reset_justify_content(&mut self, other: &Self)

Reset justify-content from the initial struct.

pub fn clone_justify_content(&self) -> T

Get the computed value for justify-content.

pub fn justify_content_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for justify-content.

pub fn set_justify_items(&mut self, v: T)

Set justify-items.

pub fn copy_justify_items_from(&mut self, other: &Self)

Set justify-items from other struct.

pub fn reset_justify_items(&mut self, other: &Self)

Reset justify-items from the initial struct.

pub fn clone_justify_items(&self) -> T

Get the computed value for justify-items.

pub fn justify_items_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for justify-items.

pub fn set_justify_self(&mut self, v: T)

Set justify-self.

pub fn copy_justify_self_from(&mut self, other: &Self)

Set justify-self from other struct.

pub fn reset_justify_self(&mut self, other: &Self)

Reset justify-self from the initial struct.

pub fn clone_justify_self(&self) -> T

Get the computed value for justify-self.

pub fn justify_self_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for justify-self.

pub fn set_left(&mut self, v: T)

Set left.

pub fn copy_left_from(&mut self, other: &Self)

Set left from other struct.

pub fn reset_left(&mut self, other: &Self)

Reset left from the initial struct.

pub fn clone_left(&self) -> T

Get the computed value for left.

pub fn left_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for left.

pub fn set_max_height(&mut self, v: T)

Set max-height.

pub fn copy_max_height_from(&mut self, other: &Self)

Set max-height from other struct.

pub fn reset_max_height(&mut self, other: &Self)

Reset max-height from the initial struct.

pub fn clone_max_height(&self) -> T

Get the computed value for max-height.

pub fn max_height_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for max-height.

pub fn set_max_width(&mut self, v: T)

Set max-width.

pub fn copy_max_width_from(&mut self, other: &Self)

Set max-width from other struct.

pub fn reset_max_width(&mut self, other: &Self)

Reset max-width from the initial struct.

pub fn clone_max_width(&self) -> T

Get the computed value for max-width.

pub fn max_width_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for max-width.

pub fn set_min_height(&mut self, v: T)

Set min-height.

pub fn copy_min_height_from(&mut self, other: &Self)

Set min-height from other struct.

pub fn reset_min_height(&mut self, other: &Self)

Reset min-height from the initial struct.

pub fn clone_min_height(&self) -> T

Get the computed value for min-height.

pub fn min_height_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for min-height.

pub fn set_min_width(&mut self, v: T)

Set min-width.

pub fn copy_min_width_from(&mut self, other: &Self)

Set min-width from other struct.

pub fn reset_min_width(&mut self, other: &Self)

Reset min-width from the initial struct.

pub fn clone_min_width(&self) -> T

Get the computed value for min-width.

pub fn min_width_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for min-width.

pub fn set_object_position(&mut self, v: T)

Set object-position.

pub fn copy_object_position_from(&mut self, other: &Self)

Set object-position from other struct.

pub fn reset_object_position(&mut self, other: &Self)

Reset object-position from the initial struct.

pub fn clone_object_position(&self) -> T

Get the computed value for object-position.

pub fn object_position_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for object-position.

pub fn set_order(&mut self, v: T)

Set order.

pub fn copy_order_from(&mut self, other: &Self)

Set order from other struct.

pub fn reset_order(&mut self, other: &Self)

Reset order from the initial struct.

pub fn clone_order(&self) -> T

Get the computed value for order.

pub fn order_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for order.

pub fn set_position_area(&mut self, v: T)

Set position-area.

pub fn copy_position_area_from(&mut self, other: &Self)

Set position-area from other struct.

pub fn reset_position_area(&mut self, other: &Self)

Reset position-area from the initial struct.

pub fn clone_position_area(&self) -> T

Get the computed value for position-area.

pub fn position_area_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for position-area.

pub fn set_position_try_fallbacks(&mut self, v: T)

Set position-try-fallbacks.

pub fn copy_position_try_fallbacks_from(&mut self, other: &Self)

Set position-try-fallbacks from other struct.

pub fn reset_position_try_fallbacks(&mut self, other: &Self)

Reset position-try-fallbacks from the initial struct.

pub fn clone_position_try_fallbacks(&self) -> T

Get the computed value for position-try-fallbacks.

pub fn position_try_fallbacks_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for position-try-fallbacks.

pub fn set_right(&mut self, v: T)

Set right.

pub fn copy_right_from(&mut self, other: &Self)

Set right from other struct.

pub fn reset_right(&mut self, other: &Self)

Reset right from the initial struct.

pub fn clone_right(&self) -> T

Get the computed value for right.

pub fn right_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for right.

pub fn set_row_gap(&mut self, v: T)

Set row-gap.

pub fn copy_row_gap_from(&mut self, other: &Self)

Set row-gap from other struct.

pub fn reset_row_gap(&mut self, other: &Self)

Reset row-gap from the initial struct.

pub fn clone_row_gap(&self) -> T

Get the computed value for row-gap.

pub fn row_gap_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for row-gap.

pub fn set_top(&mut self, v: T)

Set top.

pub fn copy_top_from(&mut self, other: &Self)

Set top from other struct.

pub fn reset_top(&mut self, other: &Self)

Reset top from the initial struct.

pub fn clone_top(&self) -> T

Get the computed value for top.

pub fn top_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for top.

pub fn set_width(&mut self, v: T)

Set width.

pub fn copy_width_from(&mut self, other: &Self)

Set width from other struct.

pub fn reset_width(&mut self, other: &Self)

Reset width from the initial struct.

pub fn clone_width(&self) -> T

Get the computed value for width.

pub fn width_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for width.

pub fn set_z_index(&mut self, v: T)

Set z-index.

pub fn copy_z_index_from(&mut self, other: &Self)

Set z-index from other struct.

pub fn reset_z_index(&mut self, other: &Self)

Reset z-index from the initial struct.

pub fn clone_z_index(&self) -> T

Get the computed value for z-index.

pub fn z_index_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for z-index.

pub fn set_flex_direction(&mut self, v: T)

Set flex-direction.

pub fn copy_flex_direction_from(&mut self, other: &Self)

Set flex-direction from other struct.

pub fn reset_flex_direction(&mut self, other: &Self)

Reset flex-direction from the initial struct.

pub fn clone_flex_direction(&self) -> T

Get the computed value for flex-direction.

pub fn flex_direction_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for flex-direction.

pub fn set_flex_wrap(&mut self, v: T)

Set flex-wrap.

pub fn copy_flex_wrap_from(&mut self, other: &Self)

Set flex-wrap from other struct.

pub fn reset_flex_wrap(&mut self, other: &Self)

Reset flex-wrap from the initial struct.

pub fn clone_flex_wrap(&self) -> T

Get the computed value for flex-wrap.

pub fn flex_wrap_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for flex-wrap.

pub fn set_box_sizing(&mut self, v: T)

Set box-sizing.

pub fn copy_box_sizing_from(&mut self, other: &Self)

Set box-sizing from other struct.

pub fn reset_box_sizing(&mut self, other: &Self)

Reset box-sizing from the initial struct.

pub fn clone_box_sizing(&self) -> T

Get the computed value for box-sizing.

pub fn box_sizing_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for box-sizing.

pub fn set_object_fit(&mut self, v: T)

Set object-fit.

pub fn copy_object_fit_from(&mut self, other: &Self)

Set object-fit from other struct.

pub fn reset_object_fit(&mut self, other: &Self)

Reset object-fit from the initial struct.

pub fn clone_object_fit(&self) -> T

Get the computed value for object-fit.

pub fn object_fit_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for object-fit.

pub fn get_inset(&self, s: PhysicalSide) -> &T

Gets the value of the longhand of inset on the s side`.

pub fn set_inset(&mut self, s: PhysicalSide, v: T)

Sets the value of the longhand of inset on the s side to v.