style::properties::generated::style_structs

Struct UI

pub struct UI {

    pub animation_delay: T,
    pub animation_direction: T,
    pub animation_duration: T,
    pub animation_fill_mode: T,
    pub animation_iteration_count: T,
    pub animation_name: T,
    pub animation_play_state: T,
    pub animation_range_end: T,
    pub animation_range_start: T,
    pub animation_timeline: T,
    pub animation_timing_function: T,
    pub transition_behavior: T,
    pub transition_delay: T,
    pub transition_duration: T,
    pub transition_property: T,
    pub transition_timing_function: T,
    pub user_select: T,
    pub animation_composition: T,
    pub view_transition_class: T,
    pub view_transition_name: T,
    pub scrollbar_width: T,

}

The UI style struct.

Fields
animation_delay: T

The animation-delay computed value.

animation_direction: T

The animation-direction computed value.

animation_duration: T

The animation-duration computed value.

animation_fill_mode: T

The animation-fill-mode computed value.

animation_iteration_count: T

The animation-iteration-count computed value.

animation_name: T

The animation-name computed value.

animation_play_state: T

The animation-play-state computed value.

animation_range_end: T

The animation-range-end computed value.

animation_range_start: T

The animation-range-start computed value.

animation_timeline: T

The animation-timeline computed value.

animation_timing_function: T

The animation-timing-function computed value.

transition_behavior: T

The transition-behavior computed value.

transition_delay: T

The transition-delay computed value.

transition_duration: T

The transition-duration computed value.

transition_property: T

The transition-property computed value.

transition_timing_function: T

The transition-timing-function computed value.

user_select: T

The user-select computed value.

animation_composition: T

The animation-composition computed value.

view_transition_class: T

The view-transition-class computed value.

view_transition_name: T

The view-transition-name computed value.

scrollbar_width: T

The scrollbar-width computed value.

Implementations

impl UI

pub fn set_animation_delay(&mut self, v: T)

Set animation-delay.

pub fn copy_animation_delay_from(&mut self, other: &Self)

Set animation-delay from other struct.

pub fn reset_animation_delay(&mut self, other: &Self)

Reset animation-delay from the initial struct.

pub fn clone_animation_delay(&self) -> T

Get the computed value for animation-delay.

pub fn animation_delay_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-delay.

pub fn animation_delay_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_delay_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_direction(&mut self, v: T)

Set animation-direction.

pub fn copy_animation_direction_from(&mut self, other: &Self)

Set animation-direction from other struct.

pub fn reset_animation_direction(&mut self, other: &Self)

Reset animation-direction from the initial struct.

pub fn clone_animation_direction(&self) -> T

Get the computed value for animation-direction.

pub fn animation_direction_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-direction.

pub fn animation_direction_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_direction_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_duration(&mut self, v: T)

Set animation-duration.

pub fn copy_animation_duration_from(&mut self, other: &Self)

Set animation-duration from other struct.

pub fn reset_animation_duration(&mut self, other: &Self)

Reset animation-duration from the initial struct.

pub fn clone_animation_duration(&self) -> T

Get the computed value for animation-duration.

pub fn animation_duration_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-duration.

pub fn animation_duration_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_duration_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_fill_mode(&mut self, v: T)

Set animation-fill-mode.

pub fn copy_animation_fill_mode_from(&mut self, other: &Self)

Set animation-fill-mode from other struct.

pub fn reset_animation_fill_mode(&mut self, other: &Self)

Reset animation-fill-mode from the initial struct.

pub fn clone_animation_fill_mode(&self) -> T

Get the computed value for animation-fill-mode.

pub fn animation_fill_mode_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-fill-mode.

pub fn animation_fill_mode_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_fill_mode_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_iteration_count(&mut self, v: T)

Set animation-iteration-count.

pub fn copy_animation_iteration_count_from(&mut self, other: &Self)

Set animation-iteration-count from other struct.

pub fn reset_animation_iteration_count(&mut self, other: &Self)

Reset animation-iteration-count from the initial struct.

pub fn clone_animation_iteration_count(&self) -> T

Get the computed value for animation-iteration-count.

pub fn animation_iteration_count_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-iteration-count.

pub fn animation_iteration_count_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_iteration_count_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_name(&mut self, v: T)

Set animation-name.

pub fn copy_animation_name_from(&mut self, other: &Self)

Set animation-name from other struct.

pub fn reset_animation_name(&mut self, other: &Self)

Reset animation-name from the initial struct.

pub fn clone_animation_name(&self) -> T

Get the computed value for animation-name.

pub fn animation_name_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-name.

pub fn animation_name_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_name_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_play_state(&mut self, v: T)

Set animation-play-state.

pub fn copy_animation_play_state_from(&mut self, other: &Self)

Set animation-play-state from other struct.

pub fn reset_animation_play_state(&mut self, other: &Self)

Reset animation-play-state from the initial struct.

pub fn clone_animation_play_state(&self) -> T

Get the computed value for animation-play-state.

pub fn animation_play_state_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-play-state.

pub fn animation_play_state_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_play_state_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_range_end(&mut self, v: T)

Set animation-range-end.

pub fn copy_animation_range_end_from(&mut self, other: &Self)

Set animation-range-end from other struct.

pub fn reset_animation_range_end(&mut self, other: &Self)

Reset animation-range-end from the initial struct.

pub fn clone_animation_range_end(&self) -> T

Get the computed value for animation-range-end.

pub fn animation_range_end_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-range-end.

pub fn animation_range_end_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_range_end_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_range_start(&mut self, v: T)

Set animation-range-start.

pub fn copy_animation_range_start_from(&mut self, other: &Self)

Set animation-range-start from other struct.

pub fn reset_animation_range_start(&mut self, other: &Self)

Reset animation-range-start from the initial struct.

pub fn clone_animation_range_start(&self) -> T

Get the computed value for animation-range-start.

pub fn animation_range_start_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-range-start.

pub fn animation_range_start_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_range_start_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_timeline(&mut self, v: T)

Set animation-timeline.

pub fn copy_animation_timeline_from(&mut self, other: &Self)

Set animation-timeline from other struct.

pub fn reset_animation_timeline(&mut self, other: &Self)

Reset animation-timeline from the initial struct.

pub fn clone_animation_timeline(&self) -> T

Get the computed value for animation-timeline.

pub fn animation_timeline_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-timeline.

pub fn animation_timeline_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_timeline_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_timing_function(&mut self, v: T)

Set animation-timing-function.

pub fn copy_animation_timing_function_from(&mut self, other: &Self)

Set animation-timing-function from other struct.

pub fn reset_animation_timing_function(&mut self, other: &Self)

Reset animation-timing-function from the initial struct.

pub fn clone_animation_timing_function(&self) -> T

Get the computed value for animation-timing-function.

pub fn animation_timing_function_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-timing-function.

pub fn animation_timing_function_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_timing_function_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_behavior(&mut self, v: T)

Set transition-behavior.

pub fn copy_transition_behavior_from(&mut self, other: &Self)

Set transition-behavior from other struct.

pub fn reset_transition_behavior(&mut self, other: &Self)

Reset transition-behavior from the initial struct.

pub fn clone_transition_behavior(&self) -> T

Get the computed value for transition-behavior.

pub fn transition_behavior_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transition-behavior.

pub fn transition_behavior_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_behavior_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_delay(&mut self, v: T)

Set transition-delay.

pub fn copy_transition_delay_from(&mut self, other: &Self)

Set transition-delay from other struct.

pub fn reset_transition_delay(&mut self, other: &Self)

Reset transition-delay from the initial struct.

pub fn clone_transition_delay(&self) -> T

Get the computed value for transition-delay.

pub fn transition_delay_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transition-delay.

pub fn transition_delay_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_delay_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_duration(&mut self, v: T)

Set transition-duration.

pub fn copy_transition_duration_from(&mut self, other: &Self)

Set transition-duration from other struct.

pub fn reset_transition_duration(&mut self, other: &Self)

Reset transition-duration from the initial struct.

pub fn clone_transition_duration(&self) -> T

Get the computed value for transition-duration.

pub fn transition_duration_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transition-duration.

pub fn transition_duration_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_duration_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_property(&mut self, v: T)

Set transition-property.

pub fn copy_transition_property_from(&mut self, other: &Self)

Set transition-property from other struct.

pub fn reset_transition_property(&mut self, other: &Self)

Reset transition-property from the initial struct.

pub fn clone_transition_property(&self) -> T

Get the computed value for transition-property.

pub fn transition_property_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transition-property.

pub fn transition_property_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_property_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_timing_function(&mut self, v: T)

Set transition-timing-function.

pub fn copy_transition_timing_function_from(&mut self, other: &Self)

Set transition-timing-function from other struct.

pub fn reset_transition_timing_function(&mut self, other: &Self)

Reset transition-timing-function from the initial struct.

pub fn clone_transition_timing_function(&self) -> T

Get the computed value for transition-timing-function.

pub fn transition_timing_function_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for transition-timing-function.

pub fn transition_timing_function_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_timing_function_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_user_select(&mut self, v: T)

Set user-select.

pub fn copy_user_select_from(&mut self, other: &Self)

Set user-select from other struct.

pub fn reset_user_select(&mut self, other: &Self)

Reset user-select from the initial struct.

pub fn clone_user_select(&self) -> T

Get the computed value for user-select.

pub fn user_select_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for user-select.

pub fn set_animation_composition(&mut self, v: T)

Set animation-composition.

pub fn copy_animation_composition_from(&mut self, other: &Self)

Set animation-composition from other struct.

pub fn reset_animation_composition(&mut self, other: &Self)

Reset animation-composition from the initial struct.

pub fn clone_animation_composition(&self) -> T

Get the computed value for animation-composition.

pub fn animation_composition_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for animation-composition.

pub fn animation_composition_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_composition_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_view_transition_class(&mut self, v: T)

Set view-transition-class.

pub fn copy_view_transition_class_from(&mut self, other: &Self)

Set view-transition-class from other struct.

pub fn reset_view_transition_class(&mut self, other: &Self)

Reset view-transition-class from the initial struct.

pub fn clone_view_transition_class(&self) -> T

Get the computed value for view-transition-class.

pub fn view_transition_class_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for view-transition-class.

pub fn set_view_transition_name(&mut self, v: T)

Set view-transition-name.

pub fn copy_view_transition_name_from(&mut self, other: &Self)

Set view-transition-name from other struct.

pub fn reset_view_transition_name(&mut self, other: &Self)

Reset view-transition-name from the initial struct.

pub fn clone_view_transition_name(&self) -> T

Get the computed value for view-transition-name.

pub fn view_transition_name_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for view-transition-name.

pub fn set_scrollbar_width(&mut self, v: T)

Set scrollbar-width.

pub fn copy_scrollbar_width_from(&mut self, other: &Self)

Set scrollbar-width from other struct.

pub fn reset_scrollbar_width(&mut self, other: &Self)

Reset scrollbar-width from the initial struct.

pub fn clone_scrollbar_width(&self) -> T

Get the computed value for scrollbar-width.

pub fn scrollbar_width_equals(&self, other: &Self) -> bool

Whether self and other have the same computed value for scrollbar-width.

impl UI

pub fn animation_delay_iter(&self) -> AnimationDelayIter<'_>

Iterate over the values of animation-delay.

pub fn animation_delay_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-delay.

pub fn animation_direction_iter(&self) -> AnimationDirectionIter<'_>

Iterate over the values of animation-direction.

pub fn animation_direction_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-direction.

pub fn animation_duration_iter(&self) -> AnimationDurationIter<'_>

Iterate over the values of animation-duration.

pub fn animation_duration_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-duration.

pub fn animation_fill_mode_iter(&self) -> AnimationFillModeIter<'_>

Iterate over the values of animation-fill-mode.

pub fn animation_fill_mode_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-fill-mode.

pub fn animation_iteration_count_iter(&self) -> AnimationIterationCountIter<'_>

Iterate over the values of animation-iteration-count.

pub fn animation_iteration_count_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-iteration-count.

pub fn animation_name_iter(&self) -> AnimationNameIter<'_>

Iterate over the values of animation-name.

pub fn animation_name_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-name.

pub fn animation_play_state_iter(&self) -> AnimationPlayStateIter<'_>

Iterate over the values of animation-play-state.

pub fn animation_play_state_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-play-state.

pub fn animation_range_end_iter(&self) -> AnimationRangeEndIter<'_>

Iterate over the values of animation-range-end.

pub fn animation_range_end_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-range-end.

pub fn animation_range_start_iter(&self) -> AnimationRangeStartIter<'_>

Iterate over the values of animation-range-start.

pub fn animation_range_start_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-range-start.

pub fn animation_timeline_iter(&self) -> AnimationTimelineIter<'_>

Iterate over the values of animation-timeline.

pub fn animation_timeline_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-timeline.

pub fn animation_timing_function_iter(&self) -> AnimationTimingFunctionIter<'_>

Iterate over the values of animation-timing-function.

pub fn animation_timing_function_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-timing-function.

pub fn transition_behavior_iter(&self) -> TransitionBehaviorIter<'_>

Iterate over the values of transition-behavior.

pub fn transition_behavior_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-behavior.

pub fn transition_delay_iter(&self) -> TransitionDelayIter<'_>

Iterate over the values of transition-delay.

pub fn transition_delay_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-delay.

pub fn transition_duration_iter(&self) -> TransitionDurationIter<'_>

Iterate over the values of transition-duration.

pub fn transition_duration_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-duration.

pub fn transition_property_iter(&self) -> TransitionPropertyIter<'_>

Iterate over the values of transition-property.

pub fn transition_property_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-property.

pub fn transition_timing_function_iter( &self, ) -> TransitionTimingFunctionIter<'_>

Iterate over the values of transition-timing-function.

pub fn transition_timing_function_mod( &self, index: usize, ) -> SingleComputedValue

Get a value mod index for the property transition-timing-function.

pub fn animation_composition_iter(&self) -> AnimationCompositionIter<'_>

Iterate over the values of animation-composition.

pub fn animation_composition_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-composition.

pub fn specifies_animations(&self) -> bool

Returns whether there is any animation specified with animation-name other than none.

pub fn specifies_transitions(&self) -> bool

Returns whether there are any transitions specified.

pub fn has_initial_animation_timeline(&self) -> bool

Returns whether animation-timeline is initial value. We need this information to resolve animation-duration.

pub fn animations_equals(&self, other: &Self) -> bool

Returns true if animation properties are equal between styles, but without considering keyframe data and animation-timeline.