Module style::properties::shorthands

A module with code for all the shorthand css properties, and a few serialization helpers.

Modules

_moz_outline_radius
Nonstandard (https://developer.mozilla.org/en-US/docs/Web/CSS/-moz-outline-radius)

_webkit_text_stroke
https://compat.spec.whatwg.org/#the-webkit-text-stroke

animation
https://drafts.csswg.org/css-animations/#propdef-animation

background
https://drafts.csswg.org/css-backgrounds/#the-background

background_position
https://drafts.csswg.org/css-backgrounds-4/#the-background-position

border
https://drafts.csswg.org/css-backgrounds/#border

border_block_end
https://drafts.csswg.org/css-logical-props/#propdef-border-block-end

border_block_start
https://drafts.csswg.org/css-logical-props/#propdef-border-block-start

border_bottom
https://drafts.csswg.org/css-backgrounds/#border-bottom

border_color
https://drafts.csswg.org/css-backgrounds/#border-color

border_image
https://drafts.csswg.org/css-backgrounds-3/#border-image

border_inline_end
https://drafts.csswg.org/css-logical-props/#propdef-border-inline-end

border_inline_start
https://drafts.csswg.org/css-logical-props/#propdef-border-inline-start

border_left
https://drafts.csswg.org/css-backgrounds/#border-left

border_radius
https://drafts.csswg.org/css-backgrounds/#border-radius

border_right
https://drafts.csswg.org/css-backgrounds/#border-right

border_style
https://drafts.csswg.org/css-backgrounds/#border-style

border_top
https://drafts.csswg.org/css-backgrounds/#border-top

border_width
https://drafts.csswg.org/css-backgrounds/#border-width

column_rule
https://drafts.csswg.org/css-multicol/#propdef-column-rule

columns
https://drafts.csswg.org/css-multicol/#propdef-columns

flex
https://drafts.csswg.org/css-flexbox/#flex-property

flex_flow
https://drafts.csswg.org/css-flexbox/#flex-flow-property

font
https://drafts.csswg.org/css-fonts-3/#propdef-font

list_style
https://drafts.csswg.org/css-lists/#propdef-list-style

margin
https://drafts.csswg.org/css-box/#propdef-margin

mask
https://drafts.fxtf.org/css-masking/#propdef-mask

mask_position
https://drafts.csswg.org/css-masks-4/#the-mask-position

outline
https://drafts.csswg.org/css-ui/#propdef-outline

overflow
https://drafts.csswg.org/css-overflow/#propdef-overflow

padding
https://drafts.csswg.org/css-box-3/#propdef-padding

scroll_snap_type
https://drafts.csswg.org/css-scroll-snap/#propdef-scroll-snap-type

text_decoration
https://drafts.csswg.org/css-text-decor/#propdef-text-decoration

text_emphasis
https://drafts.csswg.org/css-text-decor-3/#text-emphasis-property

transition
https://drafts.csswg.org/css-transitions/#propdef-transition

Functions

is_overflow_shorthand

parse_border

parse_four_sides
Parses a property for four different sides per CSS syntax.

serialize_four_sides