Struct style::properties::BuilderArc

#[repr(C)]

pub struct BuilderArc<T> where
    T: 'static + ?Sized,  { /* fields omitted */ }

Methods

impl<T> Arc<T>

fn new(data: T) -> Arc<T>

fn into_raw(this: Arc<T>) -> *const T

fn borrow_arc(&'a self) -> ArcBorrow<'a, T>

Produce a pointer to the data that can be converted back to an arc

fn with_raw_offset_arc<F, U>(&self, f: F) -> U where
F: FnOnce(&RawOffsetArc<T>) -> U,

Temporarily converts |self| into a bonafide RawOffsetArc and exposes it to the provided callback. The refcount is not modified.

fn heap_ptr(&self) -> *const c_void

Returns the address on the heap of the Arc itself -- not the T within it -- for memory reporting.

impl<T> Arc<T> where
T: ?Sized,

fn ptr_eq(this: &Arc<T>, other: &Arc<T>) -> bool

impl<T> Arc<T> where
T: Clone,

fn make_mut(this: &mut Arc<T>) -> &mut T

impl<T> Arc<T> where
T: ?Sized,

fn get_mut(this: &mut Arc<T>) -> Option<&mut T>

fn is_unique(&self) -> bool

impl<H, T> Arc<HeaderSlice<H, [T]>>

fn from_header_and_iter<I>(header: H, items: I) -> Arc<HeaderSlice<H, [T]>> where
I: Iterator<Item = T> + ExactSizeIterator,

Creates an Arc for a HeaderSlice using the given header struct and iterator to generate the slice. The resulting Arc will be fat.

impl<H, T> Arc<HeaderSlice<HeaderWithLength<H>, [T]>> where
H: 'static,
T: 'static,

fn into_thin(a: Arc<HeaderSlice<HeaderWithLength<H>, [T]>>) -> ThinArc<H, T>

Converts an Arc into a ThinArc. This consumes the Arc, so the refcount is not modified.

fn from_thin(a: ThinArc<H, T>) -> Arc<HeaderSlice<HeaderWithLength<H>, [T]>>

Converts a ThinArc into an Arc. This consumes the ThinArc, so the refcount is not modified.

impl<T> Arc<T> where
T: 'static,

fn into_raw_offset(a: Arc<T>) -> RawOffsetArc<T>

Converts an Arc into a RawOffsetArc. This consumes the Arc, so the refcount is not modified.

fn from_raw_offset(a: RawOffsetArc<T>) -> Arc<T>

Converts a RawOffsetArc into an Arc. This consumes the RawOffsetArc, so the refcount is not modified.