Struct style::properties::ComputedValues

pub struct ComputedValues { /* fields omitted */ }

The struct that Servo uses to represent computed values.

This struct contains an immutable atomically-reference-counted pointer to every kind of style struct.

When needed, the structs may be copied in order to get mutated.

Methods

impl ComputedValues

pub fn is_style_if_visited(&self) -> bool

Whether we're a visited style.

pub fn rules(&self) -> &StrongRuleNode

Gets a reference to the rule node. Panic if no rule node exists.

pub fn visited_style(&self) -> Option<&ComputedValues>

Returns the visited style, if any.

pub fn visited_rules(&self) -> Option<&StrongRuleNode>

Returns the visited rules, if applicable.

pub fn is_in_display_none_subtree(&self) -> bool

Returns whether we're in a display: none subtree.

pub fn custom_properties(&self) -> Option<&Arc<CustomPropertiesMap>>

Gets a reference to the custom properties map (if one exists).

impl ComputedValues

pub fn new(
_: &Device,
_: Option<&ComputedValues>,
_: Option<&PseudoElement>,
custom_properties: Option<Arc<CustomPropertiesMap>>,
writing_mode: WritingMode,
flags: ComputedValueFlags,
rules: Option<StrongRuleNode>,
visited_style: Option<Arc<ComputedValues>>,
background: Arc<Background>,
border: Arc<Border>,
box_: Arc<Box>,
color: Arc<Color>,
column: Arc<Column>,
counters: Arc<Counters>,
effects: Arc<Effects>,
font: Arc<Font>,
inheritedbox: Arc<InheritedBox>,
inheritedtable: Arc<InheritedTable>,
inheritedtext: Arc<InheritedText>,
list: Arc<List>,
margin: Arc<Margin>,
outline: Arc<Outline>,
padding: Arc<Padding>,
pointing: Arc<Pointing>,
position: Arc<Position>,
table: Arc<Table>,
text: Arc<Text>
) -> Arc<Self>

Create a new refcounted ComputedValues

pub fn initial_values() -> &'static Self

Get the initial computed values.