style::properties

Struct NonCustomPropertyId

#[repr(C)]
pub struct NonCustomPropertyId(u16);

A longhand or shorthand property.

Tuple Fields
0: u16
Implementations

impl NonCustomPropertyId

pub fn name(self) -> &'static str

Get the property name.

pub fn is_animatable(self) -> bool

Returns whether this property is animatable.

pub(super) fn enabled_for_all_content(self) -> bool

Whether this property is enabled for all content right now.

pub fn allowed_in_rule(self, rule_types: CssRuleTypes) -> bool

Returns whether a given rule allows a given property.

pub(super) fn allowed_in(self, context: &ParserContext<'_>) -> bool

pub(super) fn allowed_in_ignoring_rule_type( self, context: &ParserContext<'_>, ) -> bool

pub(super) fn supported_types(&self) -> u8

The supported types of this property. The return value should be style_traits::CssType when it can become a bitflags type.

pub(super) fn collect_property_completion_keywords( &self, f: KeywordsCollectFn<'_>, )

See PropertyId::collect_property_completion_keywords.

impl NonCustomPropertyId

pub fn bit(self) -> usize

Returns the underlying index, used for use counter.

pub fn unaliased(self) -> Self

Resolves the alias of a given property if needed.

pub fn to_property_id(self) -> PropertyId

Turns this NonCustomPropertyId into a PropertyId.

pub fn as_longhand(self) -> Option<LonghandId>

Returns a longhand id, if this property is one.

pub fn as_shorthand(self) -> Option<ShorthandId>

Returns a shorthand id, if this property is one.

pub fn as_alias(self) -> Option<AliasId>

Returns an alias id, if this property is one.

pub fn longhand_or_shorthand(self) -> Result<LonghandId, ShorthandId>

Returns either a longhand or a shorthand, resolving aliases.

pub const fn from_longhand(id: LonghandId) -> Self

Converts a longhand id into a non-custom property id.

pub const fn from_shorthand(id: ShorthandId) -> Self

Converts a shorthand id into a non-custom property id.

pub const fn from_alias(id: AliasId) -> Self

Converts an alias id into a non-custom property id.