Struct style::properties::style_structs::Border

pub struct Border {

    pub border_top_color: T,
    pub border_top_style: T,
    pub border_top_width: T,
    pub border_right_color: T,
    pub border_right_style: T,
    pub border_right_width: T,
    pub border_bottom_color: T,
    pub border_bottom_style: T,
    pub border_bottom_width: T,
    pub border_left_color: T,
    pub border_left_style: T,
    pub border_left_width: T,
    pub border_top_left_radius: T,
    pub border_top_right_radius: T,
    pub border_bottom_right_radius: T,
    pub border_bottom_left_radius: T,
    pub border_image_source: T,
    pub border_image_outset: T,
    pub border_image_repeat: T,
    pub border_image_width: T,
    pub border_image_slice: T,

}

The Border style struct.

Fields
border_top_color: T

The border-top-color computed value.

border_top_style: T

The border-top-style computed value.

border_top_width: T

The border-top-width computed value.

border_right_color: T

The border-right-color computed value.

border_right_style: T

The border-right-style computed value.

border_right_width: T

The border-right-width computed value.

border_bottom_color: T

The border-bottom-color computed value.

border_bottom_style: T

The border-bottom-style computed value.

border_bottom_width: T

The border-bottom-width computed value.

border_left_color: T

The border-left-color computed value.

border_left_style: T

The border-left-style computed value.

border_left_width: T

The border-left-width computed value.

border_top_left_radius: T

The border-top-left-radius computed value.

border_top_right_radius: T

The border-top-right-radius computed value.

border_bottom_right_radius: T

The border-bottom-right-radius computed value.

border_bottom_left_radius: T

The border-bottom-left-radius computed value.

border_image_source: T

The border-image-source computed value.

border_image_outset: T

The border-image-outset computed value.

border_image_repeat: T

The border-image-repeat computed value.

border_image_width: T

The border-image-width computed value.

border_image_slice: T

The border-image-slice computed value.

Implementations

impl Border

pub fn set_border_top_color(&mut self, v: T)

Set border-top-color.

pub fn copy_border_top_color_from(&mut self, other: &Self)

Set border-top-color from other struct.

pub fn reset_border_top_color(&mut self, other: &Self)

Reset border-top-color from the initial struct.

pub fn clone_border_top_color(&self) -> T

Get the computed value for border-top-color.

pub fn set_border_top_style(&mut self, v: T)

Set border-top-style.

pub fn copy_border_top_style_from(&mut self, other: &Self)

Set border-top-style from other struct.

pub fn reset_border_top_style(&mut self, other: &Self)

Reset border-top-style from the initial struct.

pub fn clone_border_top_style(&self) -> T

Get the computed value for border-top-style.

pub fn set_border_top_width(&mut self, v: T)

Set border-top-width.

pub fn copy_border_top_width_from(&mut self, other: &Self)

Set border-top-width from other struct.

pub fn reset_border_top_width(&mut self, other: &Self)

Reset border-top-width from the initial struct.

pub fn clone_border_top_width(&self) -> T

Get the computed value for border-top-width.

pub fn set_border_right_color(&mut self, v: T)

Set border-right-color.

pub fn copy_border_right_color_from(&mut self, other: &Self)

Set border-right-color from other struct.

pub fn reset_border_right_color(&mut self, other: &Self)

Reset border-right-color from the initial struct.

pub fn clone_border_right_color(&self) -> T

Get the computed value for border-right-color.

pub fn set_border_right_style(&mut self, v: T)

Set border-right-style.

pub fn copy_border_right_style_from(&mut self, other: &Self)

Set border-right-style from other struct.

pub fn reset_border_right_style(&mut self, other: &Self)

Reset border-right-style from the initial struct.

pub fn clone_border_right_style(&self) -> T

Get the computed value for border-right-style.

pub fn set_border_right_width(&mut self, v: T)

Set border-right-width.

pub fn copy_border_right_width_from(&mut self, other: &Self)

Set border-right-width from other struct.

pub fn reset_border_right_width(&mut self, other: &Self)

Reset border-right-width from the initial struct.

pub fn clone_border_right_width(&self) -> T

Get the computed value for border-right-width.

pub fn set_border_bottom_color(&mut self, v: T)

Set border-bottom-color.

pub fn copy_border_bottom_color_from(&mut self, other: &Self)

Set border-bottom-color from other struct.

pub fn reset_border_bottom_color(&mut self, other: &Self)

Reset border-bottom-color from the initial struct.

pub fn clone_border_bottom_color(&self) -> T

Get the computed value for border-bottom-color.

pub fn set_border_bottom_style(&mut self, v: T)

Set border-bottom-style.

pub fn copy_border_bottom_style_from(&mut self, other: &Self)

Set border-bottom-style from other struct.

pub fn reset_border_bottom_style(&mut self, other: &Self)

Reset border-bottom-style from the initial struct.

pub fn clone_border_bottom_style(&self) -> T

Get the computed value for border-bottom-style.

pub fn set_border_bottom_width(&mut self, v: T)

Set border-bottom-width.

pub fn copy_border_bottom_width_from(&mut self, other: &Self)

Set border-bottom-width from other struct.

pub fn reset_border_bottom_width(&mut self, other: &Self)

Reset border-bottom-width from the initial struct.

pub fn clone_border_bottom_width(&self) -> T

Get the computed value for border-bottom-width.

pub fn set_border_left_color(&mut self, v: T)

Set border-left-color.

pub fn copy_border_left_color_from(&mut self, other: &Self)

Set border-left-color from other struct.

pub fn reset_border_left_color(&mut self, other: &Self)

Reset border-left-color from the initial struct.

pub fn clone_border_left_color(&self) -> T

Get the computed value for border-left-color.

pub fn set_border_left_style(&mut self, v: T)

Set border-left-style.

pub fn copy_border_left_style_from(&mut self, other: &Self)

Set border-left-style from other struct.

pub fn reset_border_left_style(&mut self, other: &Self)

Reset border-left-style from the initial struct.

pub fn clone_border_left_style(&self) -> T

Get the computed value for border-left-style.

pub fn set_border_left_width(&mut self, v: T)

Set border-left-width.

pub fn copy_border_left_width_from(&mut self, other: &Self)

Set border-left-width from other struct.

pub fn reset_border_left_width(&mut self, other: &Self)

Reset border-left-width from the initial struct.

pub fn clone_border_left_width(&self) -> T

Get the computed value for border-left-width.

pub fn set_border_block_start_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-start-color given a writing mode.

pub fn copy_border_block_start_color_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-block-start-color given a writing mode.

pub fn reset_border_block_start_color(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-start-color given a writing mode.

pub fn clone_border_block_start_color(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-start-color given a writing mode.

pub fn set_border_block_start_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-start-style given a writing mode.

pub fn copy_border_block_start_style_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-block-start-style given a writing mode.

pub fn reset_border_block_start_style(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-start-style given a writing mode.

pub fn clone_border_block_start_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-start-style given a writing mode.

pub fn set_border_block_start_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-start-width given a writing mode.

pub fn copy_border_block_start_width_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-block-start-width given a writing mode.

pub fn reset_border_block_start_width(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-start-width given a writing mode.

pub fn clone_border_block_start_width(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-start-width given a writing mode.

pub fn set_border_block_end_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-end-color given a writing mode.

pub fn copy_border_block_end_color_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-block-end-color given a writing mode.

pub fn reset_border_block_end_color(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-end-color given a writing mode.

pub fn clone_border_block_end_color(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-end-color given a writing mode.

pub fn set_border_block_end_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-end-style given a writing mode.

pub fn copy_border_block_end_style_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-block-end-style given a writing mode.

pub fn reset_border_block_end_style(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-end-style given a writing mode.

pub fn clone_border_block_end_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-end-style given a writing mode.

pub fn set_border_block_end_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-block-end-width given a writing mode.

pub fn copy_border_block_end_width_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-block-end-width given a writing mode.

pub fn reset_border_block_end_width(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-block-end-width given a writing mode.

pub fn clone_border_block_end_width(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-block-end-width given a writing mode.

pub fn set_border_inline_start_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-start-color given a writing mode.

pub fn copy_border_inline_start_color_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-inline-start-color given a writing mode.

pub fn reset_border_inline_start_color(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-start-color given a writing mode.

pub fn clone_border_inline_start_color(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-start-color given a writing mode.

pub fn set_border_inline_start_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-start-style given a writing mode.

pub fn copy_border_inline_start_style_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-inline-start-style given a writing mode.

pub fn reset_border_inline_start_style(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-start-style given a writing mode.

pub fn clone_border_inline_start_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-start-style given a writing mode.

pub fn set_border_inline_start_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-start-width given a writing mode.

pub fn copy_border_inline_start_width_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-inline-start-width given a writing mode.

pub fn reset_border_inline_start_width(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-start-width given a writing mode.

pub fn clone_border_inline_start_width(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-start-width given a writing mode.

pub fn set_border_inline_end_color(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-end-color given a writing mode.

pub fn copy_border_inline_end_color_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-inline-end-color given a writing mode.

pub fn reset_border_inline_end_color(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-end-color given a writing mode.

pub fn clone_border_inline_end_color(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-end-color given a writing mode.

pub fn set_border_inline_end_style(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-end-style given a writing mode.

pub fn copy_border_inline_end_style_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-inline-end-style given a writing mode.

pub fn reset_border_inline_end_style(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-end-style given a writing mode.

pub fn clone_border_inline_end_style(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-end-style given a writing mode.

pub fn set_border_inline_end_width(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-inline-end-width given a writing mode.

pub fn copy_border_inline_end_width_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-inline-end-width given a writing mode.

pub fn reset_border_inline_end_width(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-inline-end-width given a writing mode.

pub fn clone_border_inline_end_width(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-inline-end-width given a writing mode.

pub fn set_border_top_left_radius(&mut self, v: T)

Set border-top-left-radius.

pub fn copy_border_top_left_radius_from(&mut self, other: &Self)

Set border-top-left-radius from other struct.

pub fn reset_border_top_left_radius(&mut self, other: &Self)

Reset border-top-left-radius from the initial struct.

pub fn clone_border_top_left_radius(&self) -> T

Get the computed value for border-top-left-radius.

pub fn set_border_top_right_radius(&mut self, v: T)

Set border-top-right-radius.

pub fn copy_border_top_right_radius_from(&mut self, other: &Self)

Set border-top-right-radius from other struct.

pub fn reset_border_top_right_radius(&mut self, other: &Self)

Reset border-top-right-radius from the initial struct.

pub fn clone_border_top_right_radius(&self) -> T

Get the computed value for border-top-right-radius.

pub fn set_border_bottom_right_radius(&mut self, v: T)

Set border-bottom-right-radius.

pub fn copy_border_bottom_right_radius_from(&mut self, other: &Self)

Set border-bottom-right-radius from other struct.

pub fn reset_border_bottom_right_radius(&mut self, other: &Self)

Reset border-bottom-right-radius from the initial struct.

pub fn clone_border_bottom_right_radius(&self) -> T

Get the computed value for border-bottom-right-radius.

pub fn set_border_bottom_left_radius(&mut self, v: T)

Set border-bottom-left-radius.

pub fn copy_border_bottom_left_radius_from(&mut self, other: &Self)

Set border-bottom-left-radius from other struct.

pub fn reset_border_bottom_left_radius(&mut self, other: &Self)

Reset border-bottom-left-radius from the initial struct.

pub fn clone_border_bottom_left_radius(&self) -> T

Get the computed value for border-bottom-left-radius.

pub fn set_border_start_start_radius(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-start-start-radius given a writing mode.

pub fn copy_border_start_start_radius_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-start-start-radius given a writing mode.

pub fn reset_border_start_start_radius(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-start-start-radius given a writing mode.

pub fn clone_border_start_start_radius(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-start-start-radius given a writing mode.

pub fn set_border_start_end_radius(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-start-end-radius given a writing mode.

pub fn copy_border_start_end_radius_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-start-end-radius given a writing mode.

pub fn reset_border_start_end_radius(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-start-end-radius given a writing mode.

pub fn clone_border_start_end_radius(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-start-end-radius given a writing mode.

pub fn set_border_end_start_radius(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-end-start-radius given a writing mode.

pub fn copy_border_end_start_radius_from(
&mut self,
other: &Self,
wm: WritingMode
)

Copy the appropriate physical property from another struct for border-end-start-radius given a writing mode.

pub fn reset_border_end_start_radius(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-end-start-radius given a writing mode.

pub fn clone_border_end_start_radius(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-end-start-radius given a writing mode.

pub fn set_border_end_end_radius(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for border-end-end-radius given a writing mode.

pub fn copy_border_end_end_radius_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-end-end-radius given a writing mode.

pub fn reset_border_end_end_radius(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for border-end-end-radius given a writing mode.

pub fn clone_border_end_end_radius(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for border-end-end-radius given a writing mode.

pub fn set_border_image_source(&mut self, v: T)

Set border-image-source.

pub fn copy_border_image_source_from(&mut self, other: &Self)

Set border-image-source from other struct.

pub fn reset_border_image_source(&mut self, other: &Self)

Reset border-image-source from the initial struct.

pub fn clone_border_image_source(&self) -> T

Get the computed value for border-image-source.

pub fn set_border_image_outset(&mut self, v: T)

Set border-image-outset.

pub fn copy_border_image_outset_from(&mut self, other: &Self)

Set border-image-outset from other struct.

pub fn reset_border_image_outset(&mut self, other: &Self)

Reset border-image-outset from the initial struct.

pub fn clone_border_image_outset(&self) -> T

Get the computed value for border-image-outset.

pub fn set_border_image_repeat(&mut self, v: T)

Set border-image-repeat.

pub fn copy_border_image_repeat_from(&mut self, other: &Self)

Set border-image-repeat from other struct.

pub fn reset_border_image_repeat(&mut self, other: &Self)

Reset border-image-repeat from the initial struct.

pub fn clone_border_image_repeat(&self) -> T

Get the computed value for border-image-repeat.

pub fn set_border_image_width(&mut self, v: T)

Set border-image-width.

pub fn copy_border_image_width_from(&mut self, other: &Self)

Set border-image-width from other struct.

pub fn reset_border_image_width(&mut self, other: &Self)

Reset border-image-width from the initial struct.

pub fn clone_border_image_width(&self) -> T

Get the computed value for border-image-width.

pub fn set_border_image_slice(&mut self, v: T)

Set border-image-slice.

pub fn copy_border_image_slice_from(&mut self, other: &Self)

Set border-image-slice from other struct.

pub fn reset_border_image_slice(&mut self, other: &Self)

Reset border-image-slice from the initial struct.

pub fn clone_border_image_slice(&self) -> T

Get the computed value for border-image-slice.

pub fn border_top_has_nonzero_width(&self) -> bool

Whether the border-top property has nonzero width.

pub fn border_right_has_nonzero_width(&self) -> bool

Whether the border-right property has nonzero width.

pub fn border_bottom_has_nonzero_width(&self) -> bool

Whether the border-bottom property has nonzero width.

pub fn border_left_has_nonzero_width(&self) -> bool

Whether the border-left property has nonzero width.