Struct style::properties::style_structs::Box

pub struct Box {
    pub display: T,
    pub position: T,
    pub float: T,
    pub clear: T,
    pub _servo_display_for_hypothetical_box: T,
    pub vertical_align: T,
    pub _servo_overflow_clip_box: T,
    pub overflow_x: T,
    pub overflow_y: T,
    pub transition_duration: T,
    pub transition_timing_function: T,
    pub transition_property: T,
    pub transition_delay: T,
    pub animation_name: T,
    pub animation_duration: T,
    pub animation_timing_function: T,
    pub animation_iteration_count: T,
    pub animation_direction: T,
    pub animation_play_state: T,
    pub animation_fill_mode: T,
    pub animation_delay: T,
    pub transform: T,
    pub perspective: T,
    pub perspective_origin: T,
    pub backface_visibility: T,
    pub transform_style: T,
    pub transform_origin: T,
}

The Box style struct.

Fields
display: T

The display computed value.

position: T

The position computed value.

float: T

The float computed value.

clear: T

The clear computed value.

_servo_display_for_hypothetical_box: T

The -servo-display-for-hypothetical-box computed value.

vertical_align: T

The vertical-align computed value.

_servo_overflow_clip_box: T

The -servo-overflow-clip-box computed value.

overflow_x: T

The overflow-x computed value.

overflow_y: T

The overflow-y computed value.

transition_duration: T

The transition-duration computed value.

transition_timing_function: T

The transition-timing-function computed value.

transition_property: T

The transition-property computed value.

transition_delay: T

The transition-delay computed value.

animation_name: T

The animation-name computed value.

animation_duration: T

The animation-duration computed value.

animation_timing_function: T

The animation-timing-function computed value.

animation_iteration_count: T

The animation-iteration-count computed value.

animation_direction: T

The animation-direction computed value.

animation_play_state: T

The animation-play-state computed value.

animation_fill_mode: T

The animation-fill-mode computed value.

animation_delay: T

The animation-delay computed value.

transform: T

The transform computed value.

perspective: T

The perspective computed value.

perspective_origin: T

The perspective-origin computed value.

backface_visibility: T

The backface-visibility computed value.

transform_style: T

The transform-style computed value.

transform_origin: T

The transform-origin computed value.

Methods

impl Box

pub fn set_display(&mut self, v: T)

Set display.

pub fn copy_display_from(&mut self, other: &Self)

Set display from other struct.

pub fn reset_display(&mut self, other: &Self)

Reset display from the initial struct.

pub fn clone_display(&self) -> T

Get the computed value for display.

pub fn set_position(&mut self, v: T)

Set position.

pub fn copy_position_from(&mut self, other: &Self)

Set position from other struct.

pub fn reset_position(&mut self, other: &Self)

Reset position from the initial struct.

pub fn clone_position(&self) -> T

Get the computed value for position.

pub fn set_float(&mut self, v: T)

Set float.

pub fn copy_float_from(&mut self, other: &Self)

Set float from other struct.

pub fn reset_float(&mut self, other: &Self)

Reset float from the initial struct.

pub fn clone_float(&self) -> T

Get the computed value for float.

pub fn set_clear(&mut self, v: T)

Set clear.

pub fn copy_clear_from(&mut self, other: &Self)

Set clear from other struct.

pub fn reset_clear(&mut self, other: &Self)

Reset clear from the initial struct.

pub fn clone_clear(&self) -> T

Get the computed value for clear.

pub fn set__servo_display_for_hypothetical_box(&mut self, v: T)

Set -servo-display-for-hypothetical-box.

pub fn copy__servo_display_for_hypothetical_box_from(&mut self, other: &Self)

Set -servo-display-for-hypothetical-box from other struct.

pub fn reset__servo_display_for_hypothetical_box(&mut self, other: &Self)

Reset -servo-display-for-hypothetical-box from the initial struct.

pub fn clone__servo_display_for_hypothetical_box(&self) -> T

Get the computed value for -servo-display-for-hypothetical-box.

pub fn set_vertical_align(&mut self, v: T)

Set vertical-align.

pub fn copy_vertical_align_from(&mut self, other: &Self)

Set vertical-align from other struct.

pub fn reset_vertical_align(&mut self, other: &Self)

Reset vertical-align from the initial struct.

pub fn clone_vertical_align(&self) -> T

Get the computed value for vertical-align.

pub fn set__servo_overflow_clip_box(&mut self, v: T)

Set -servo-overflow-clip-box.

pub fn copy__servo_overflow_clip_box_from(&mut self, other: &Self)

Set -servo-overflow-clip-box from other struct.

pub fn reset__servo_overflow_clip_box(&mut self, other: &Self)

Reset -servo-overflow-clip-box from the initial struct.

pub fn clone__servo_overflow_clip_box(&self) -> T

Get the computed value for -servo-overflow-clip-box.

pub fn set_overflow_x(&mut self, v: T)

Set overflow-x.

pub fn copy_overflow_x_from(&mut self, other: &Self)

Set overflow-x from other struct.

pub fn reset_overflow_x(&mut self, other: &Self)

Reset overflow-x from the initial struct.

pub fn clone_overflow_x(&self) -> T

Get the computed value for overflow-x.

pub fn set_overflow_y(&mut self, v: T)

Set overflow-y.

pub fn copy_overflow_y_from(&mut self, other: &Self)

Set overflow-y from other struct.

pub fn reset_overflow_y(&mut self, other: &Self)

Reset overflow-y from the initial struct.

pub fn clone_overflow_y(&self) -> T

Get the computed value for overflow-y.

pub fn set_transition_duration<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set transition-duration.

pub fn copy_transition_duration_from(&mut self, other: &Self)

Set transition-duration from other struct.

pub fn reset_transition_duration(&mut self, other: &Self)

Reset transition-duration from the initial struct.

pub fn clone_transition_duration(&self) -> T

Get the computed value for transition-duration.

pub fn transition_duration_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_duration_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_timing_function<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set transition-timing-function.

pub fn copy_transition_timing_function_from(&mut self, other: &Self)

Set transition-timing-function from other struct.

pub fn reset_transition_timing_function(&mut self, other: &Self)

Reset transition-timing-function from the initial struct.

pub fn clone_transition_timing_function(&self) -> T

Get the computed value for transition-timing-function.

pub fn transition_timing_function_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_timing_function_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_property<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set transition-property.

pub fn copy_transition_property_from(&mut self, other: &Self)

Set transition-property from other struct.

pub fn reset_transition_property(&mut self, other: &Self)

Reset transition-property from the initial struct.

pub fn clone_transition_property(&self) -> T

Get the computed value for transition-property.

pub fn transition_property_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_property_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transition_delay<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set transition-delay.

pub fn copy_transition_delay_from(&mut self, other: &Self)

Set transition-delay from other struct.

pub fn reset_transition_delay(&mut self, other: &Self)

Reset transition-delay from the initial struct.

pub fn clone_transition_delay(&self) -> T

Get the computed value for transition-delay.

pub fn transition_delay_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn transition_delay_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_name<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-name.

pub fn copy_animation_name_from(&mut self, other: &Self)

Set animation-name from other struct.

pub fn reset_animation_name(&mut self, other: &Self)

Reset animation-name from the initial struct.

pub fn clone_animation_name(&self) -> T

Get the computed value for animation-name.

pub fn animation_name_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_name_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_duration<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-duration.

pub fn copy_animation_duration_from(&mut self, other: &Self)

Set animation-duration from other struct.

pub fn reset_animation_duration(&mut self, other: &Self)

Reset animation-duration from the initial struct.

pub fn clone_animation_duration(&self) -> T

Get the computed value for animation-duration.

pub fn animation_duration_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_duration_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_timing_function<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-timing-function.

pub fn copy_animation_timing_function_from(&mut self, other: &Self)

Set animation-timing-function from other struct.

pub fn reset_animation_timing_function(&mut self, other: &Self)

Reset animation-timing-function from the initial struct.

pub fn clone_animation_timing_function(&self) -> T

Get the computed value for animation-timing-function.

pub fn animation_timing_function_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_timing_function_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_iteration_count<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-iteration-count.

pub fn copy_animation_iteration_count_from(&mut self, other: &Self)

Set animation-iteration-count from other struct.

pub fn reset_animation_iteration_count(&mut self, other: &Self)

Reset animation-iteration-count from the initial struct.

pub fn clone_animation_iteration_count(&self) -> T

Get the computed value for animation-iteration-count.

pub fn animation_iteration_count_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_iteration_count_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_direction<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-direction.

pub fn copy_animation_direction_from(&mut self, other: &Self)

Set animation-direction from other struct.

pub fn reset_animation_direction(&mut self, other: &Self)

Reset animation-direction from the initial struct.

pub fn clone_animation_direction(&self) -> T

Get the computed value for animation-direction.

pub fn animation_direction_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_direction_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_play_state<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-play-state.

pub fn copy_animation_play_state_from(&mut self, other: &Self)

Set animation-play-state from other struct.

pub fn reset_animation_play_state(&mut self, other: &Self)

Reset animation-play-state from the initial struct.

pub fn clone_animation_play_state(&self) -> T

Get the computed value for animation-play-state.

pub fn animation_play_state_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_play_state_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_fill_mode<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-fill-mode.

pub fn copy_animation_fill_mode_from(&mut self, other: &Self)

Set animation-fill-mode from other struct.

pub fn reset_animation_fill_mode(&mut self, other: &Self)

Reset animation-fill-mode from the initial struct.

pub fn clone_animation_fill_mode(&self) -> T

Get the computed value for animation-fill-mode.

pub fn animation_fill_mode_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_fill_mode_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_animation_delay<I>(&mut self, v: I) where
I: IntoIterator<Item = T>,
I::IntoIter: ExactSizeIterator,

Set animation-delay.

pub fn copy_animation_delay_from(&mut self, other: &Self)

Set animation-delay from other struct.

pub fn reset_animation_delay(&mut self, other: &Self)

Reset animation-delay from the initial struct.

pub fn clone_animation_delay(&self) -> T

Get the computed value for animation-delay.

pub fn animation_delay_count(&self) -> usize

If this longhand is indexed, get the number of elements.

pub fn animation_delay_at(&self, index: usize) -> SingleComputedValue

If this longhand is indexed, get the element at given index.

pub fn set_transform(&mut self, v: T)

Set transform.

pub fn copy_transform_from(&mut self, other: &Self)

Set transform from other struct.

pub fn reset_transform(&mut self, other: &Self)

Reset transform from the initial struct.

pub fn clone_transform(&self) -> T

Get the computed value for transform.

pub fn set_perspective(&mut self, v: T)

Set perspective.

pub fn copy_perspective_from(&mut self, other: &Self)

Set perspective from other struct.

pub fn reset_perspective(&mut self, other: &Self)

Reset perspective from the initial struct.

pub fn clone_perspective(&self) -> T

Get the computed value for perspective.

pub fn set_perspective_origin(&mut self, v: T)

Set perspective-origin.

pub fn copy_perspective_origin_from(&mut self, other: &Self)

Set perspective-origin from other struct.

pub fn reset_perspective_origin(&mut self, other: &Self)

Reset perspective-origin from the initial struct.

pub fn clone_perspective_origin(&self) -> T

Get the computed value for perspective-origin.

pub fn set_backface_visibility(&mut self, v: T)

Set backface-visibility.

pub fn copy_backface_visibility_from(&mut self, other: &Self)

Set backface-visibility from other struct.

pub fn reset_backface_visibility(&mut self, other: &Self)

Reset backface-visibility from the initial struct.

pub fn clone_backface_visibility(&self) -> T

Get the computed value for backface-visibility.

pub fn set_transform_style(&mut self, v: T)

Set transform-style.

pub fn copy_transform_style_from(&mut self, other: &Self)

Set transform-style from other struct.

pub fn reset_transform_style(&mut self, other: &Self)

Reset transform-style from the initial struct.

pub fn clone_transform_style(&self) -> T

Get the computed value for transform-style.

pub fn set_transform_origin(&mut self, v: T)

Set transform-origin.

pub fn copy_transform_origin_from(&mut self, other: &Self)

Set transform-origin from other struct.

pub fn reset_transform_origin(&mut self, other: &Self)

Reset transform-origin from the initial struct.

pub fn clone_transform_origin(&self) -> T

Get the computed value for transform-origin.

pub fn set_adjusted_display(&mut self, dpy: T, is_item_or_root: bool)

Sets the display property, but without touching __servo_display_for_hypothetical_box, except when the adjustment comes from root or item display fixups.

impl Box

pub fn transition_duration_iter(&self) -> TransitionDurationIter

Iterate over the values of transition-duration.

pub fn transition_duration_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-duration.

pub fn transition_timing_function_iter(&self) -> TransitionTimingFunctionIter

Iterate over the values of transition-timing-function.

pub fn transition_timing_function_mod(
&self,
index: usize
) -> SingleComputedValue

Get a value mod index for the property transition-timing-function.

pub fn transition_property_iter(&self) -> TransitionPropertyIter

Iterate over the values of transition-property.

pub fn transition_property_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-property.

pub fn transition_delay_iter(&self) -> TransitionDelayIter

Iterate over the values of transition-delay.

pub fn transition_delay_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property transition-delay.

pub fn animation_name_iter(&self) -> AnimationNameIter

Iterate over the values of animation-name.

pub fn animation_name_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-name.

pub fn animation_duration_iter(&self) -> AnimationDurationIter

Iterate over the values of animation-duration.

pub fn animation_duration_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-duration.

pub fn animation_timing_function_iter(&self) -> AnimationTimingFunctionIter

Iterate over the values of animation-timing-function.

pub fn animation_timing_function_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-timing-function.

pub fn animation_iteration_count_iter(&self) -> AnimationIterationCountIter

Iterate over the values of animation-iteration-count.

pub fn animation_iteration_count_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-iteration-count.

pub fn animation_direction_iter(&self) -> AnimationDirectionIter

Iterate over the values of animation-direction.

pub fn animation_direction_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-direction.

pub fn animation_play_state_iter(&self) -> AnimationPlayStateIter

Iterate over the values of animation-play-state.

pub fn animation_play_state_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-play-state.

pub fn animation_fill_mode_iter(&self) -> AnimationFillModeIter

Iterate over the values of animation-fill-mode.

pub fn animation_fill_mode_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-fill-mode.

pub fn animation_delay_iter(&self) -> AnimationDelayIter

Iterate over the values of animation-delay.

pub fn animation_delay_mod(&self, index: usize) -> SingleComputedValue

Get a value mod index for the property animation-delay.

pub fn specifies_animations(&self) -> bool

Returns whether there is any animation specified with animation-name other than none.

pub fn specifies_transitions(&self) -> bool

Returns whether there are any transitions specified.