Struct style::properties::style_structs::Margin

pub struct Margin {
    pub margin_top: T,
    pub margin_left: T,
    pub margin_bottom: T,
    pub margin_right: T,
    pub margin_block_start: T,
    pub margin_block_end: T,
    pub margin_inline_start: T,
    pub margin_inline_end: T,
}

The Margin style struct.

Fields
margin_top: T

The margin-top computed value.

margin_left: T

The margin-left computed value.

margin_bottom: T

The margin-bottom computed value.

margin_right: T

The margin-right computed value.

margin_block_start: T

The margin-block-start computed value.

margin_block_end: T

The margin-block-end computed value.

margin_inline_start: T

The margin-inline-start computed value.

margin_inline_end: T

The margin-inline-end computed value.

Methods

impl Margin

pub fn set_margin_top(&mut self, v: T)

Set margin-top.

pub fn copy_margin_top_from(&mut self, other: &Self)

Set margin-top from other struct.

pub fn reset_margin_top(&mut self, other: &Self)

Reset margin-top from the initial struct.

pub fn clone_margin_top(&self) -> T

Get the computed value for margin-top.

pub fn set_margin_left(&mut self, v: T)

Set margin-left.

pub fn copy_margin_left_from(&mut self, other: &Self)

Set margin-left from other struct.

pub fn reset_margin_left(&mut self, other: &Self)

Reset margin-left from the initial struct.

pub fn clone_margin_left(&self) -> T

Get the computed value for margin-left.

pub fn set_margin_bottom(&mut self, v: T)

Set margin-bottom.

pub fn copy_margin_bottom_from(&mut self, other: &Self)

Set margin-bottom from other struct.

pub fn reset_margin_bottom(&mut self, other: &Self)

Reset margin-bottom from the initial struct.

pub fn clone_margin_bottom(&self) -> T

Get the computed value for margin-bottom.

pub fn set_margin_right(&mut self, v: T)

Set margin-right.

pub fn copy_margin_right_from(&mut self, other: &Self)

Set margin-right from other struct.

pub fn reset_margin_right(&mut self, other: &Self)

Reset margin-right from the initial struct.

pub fn clone_margin_right(&self) -> T

Get the computed value for margin-right.

pub fn set_margin_block_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-block-start given a writing mode.

pub fn copy_margin_block_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-block-start given a writing mode.

pub fn reset_margin_block_start(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-block-start given a writing mode.

pub fn clone_margin_block_start(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for margin-block-start given a writing mode.

pub fn set_margin_block_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-block-end given a writing mode.

pub fn copy_margin_block_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-block-end given a writing mode.

pub fn reset_margin_block_end(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-block-end given a writing mode.

pub fn clone_margin_block_end(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for margin-block-end given a writing mode.

pub fn set_margin_inline_start(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-inline-start given a writing mode.

pub fn copy_margin_inline_start_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-inline-start given a writing mode.

pub fn reset_margin_inline_start(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-inline-start given a writing mode.

pub fn clone_margin_inline_start(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for margin-inline-start given a writing mode.

pub fn set_margin_inline_end(&mut self, v: T, wm: WritingMode)

Set the appropriate physical property for margin-inline-end given a writing mode.

pub fn copy_margin_inline_end_from(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-inline-end given a writing mode.

pub fn reset_margin_inline_end(&mut self, other: &Self, wm: WritingMode)

Copy the appropriate physical property from another struct for margin-inline-end given a writing mode.

pub fn clone_margin_inline_end(&self, wm: WritingMode) -> T

Get the computed value for the appropriate physical property for margin-inline-end given a writing mode.