Module style::properties_and_values::rule

The [@property] at-rule.

https://drafts.css-houdini.org/css-properties-values-api-1/#at-property-rule

Macros

property_descriptors 🔒

Structs

PropertyDescriptors 🔒

Data inside a @property rule.

PropertyRuleName

A custom property name wrapper that includes the -- prefix in its serialization

PropertyRuleParser 🔒

Enums

Inherits

https://drafts.css-houdini.org/css-properties-values-api-1/#inherits-descriptor

PropertyRegistrationError

Errors that can happen when registering a property.

Functions

parse_property_block

Parse the block inside a @property rule.

Type Aliases

InitialValue

Specifies the initial value of the custom property registration represented by the @property rule, controlling the property’s initial value.