Type Alias style::properties_and_values::rule::InitialValue

pub type InitialValue = Arc<SpecifiedValue>;

Specifies the initial value of the custom property registration represented by the @property rule, controlling the property’s initial value.

The SpecifiedValue is wrapped in an Arc to avoid copying when using it.

Aliased Type

struct InitialValue {
    pub(crate) p: NonNull<ArcInner<VariableValue>>,
    pub(crate) phantom: PhantomData<VariableValue>,
}

Fields
p: NonNull<ArcInner<VariableValue>>phantom: PhantomData<VariableValue>
Implementations

impl<T> Arc<T>

pub fn new(data: T) -> Arc<T>

Construct an Arc<T>

pub fn new_leaked(data: T) -> Arc<T>

Construct an intentionally-leaked arc.

pub fn into_raw(this: Arc<T>) -> *const T

Convert the Arc to a raw pointer, suitable for use across FFI

Note: This returns a pointer to the data T, which is offset in the allocation.

pub unsafe fn from_raw(ptr: *const T) -> Arc<T>

Reconstruct the Arc from a raw pointer obtained from into_raw()

Note: This raw pointer will be offset in the allocation and must be preceded by the atomic count.

pub unsafe fn from_raw_addrefed(ptr: *const T) -> Arc<T>

Like from_raw, but returns an addrefed arc instead.

pub unsafe fn new_static<F>(alloc: F, data: T) -> Arc<T>where F: FnOnce(Layout) -> *mut u8,

Create a new static Arc (one that won’t reference count the object) and place it in the allocation provided by the specified alloc function.

alloc must return a pointer into a static allocation suitable for storing data with the Layout passed into it. The pointer returned by alloc will not be freed.

pub fn borrow_arc<'a>(&'a self) -> ArcBorrow<'a, T>

Produce a pointer to the data that can be converted back to an Arc. This is basically an &Arc<T>, without the extra indirection. It has the benefits of an &T but also knows about the underlying refcount and can be converted into more Arc<T>s if necessary.

pub fn heap_ptr(&self) -> *const c_void

Returns the address on the heap of the Arc itself – not the T within it – for memory reporting.

If this is a static reference, this returns null.

impl<T> Arc<T>where T: ?Sized,

pub fn mark_as_intentionally_leaked(&self)

Marks this Arc as intentionally leaked for the purposes of refcount logging.

It’s a logic error to call this more than once, but it’s not unsafe, as it’d just report negative leaks.

pub fn ptr_eq(this: &Arc<T>, other: &Arc<T>) -> bool

Test pointer equality between the two Arcs, i.e. they must be the same allocation

impl<T> Arc<T>where T: Clone,

pub fn make_mut(this: &mut Arc<T>) -> &mut T

Makes a mutable reference to the Arc, cloning if necessary

This is functionally equivalent to Arc::make_mut from the standard library.

If this Arc is uniquely owned, make_mut() will provide a mutable reference to the contents. If not, make_mut() will create a new Arc with a copy of the contents, update this to point to it, and provide a mutable reference to its contents.

This is useful for implementing copy-on-write schemes where you wish to avoid copying things if your Arc is not shared.

impl<T> Arc<T>where T: ?Sized,

pub fn get_mut(this: &mut Arc<T>) -> Option<&mut T>

Provides mutable access to the contents if the Arc is uniquely owned.

pub fn is_static(&self) -> bool

Whether or not the Arc is a static reference.

pub fn is_unique(&self) -> bool

Whether or not the Arc is uniquely owned (is the refcount 1?) and not a static reference.