Module style::queries::feature_expression

Parsing for query feature expressions, like (foo: bar) or (width >= 400px).

Structs

QueryFeatureExpression

A feature expression contains a reference to the feature, the value the query contained, and the range to evaluate.

Enums

FeatureType

Whether we’re parsing a media or container query feature.

LegacyRange 🔒

The kind of matching that should be performed on a feature value.

Operator 🔒

The operator that was specified in this feature.

QueryExpressionValue

A value found or expected in a expression.

QueryFeatureExpressionKind 🔒

Functions

consume_operation_or_colon 🔒

disabled_by_pref 🔒