Module style::restyle_hints

Restyle hints: an optimization to avoid unnecessarily matching selectors.

Structs

DependencySet
A set of dependencies for a given stylist.

RestyleHint
When the ElementState of an element (like IN_HOVER_STATE) changes, certain pseudo-classes (like :hover) may require us to restyle that element, its siblings, and/or its descendants. Similarly, when various attributes of an element change, we may also need to restyle things with id, class, and attribute selectors. Doing this conservatively is expensive, and so we use RestyleHints to short-circuit work we know is unnecessary.

Constants

RESTYLE_DESCENDANTS
Rerun selector matching on all of the element's descendants.

RESTYLE_LATER_SIBLINGS
Rerun selector matching on all later siblings of the element and all of their descendants.

RESTYLE_SELF
Rerun selector matching on the element.

RESTYLE_STYLE_ATTRIBUTE
Don't re-run selector-matching on the element, only the style attribute has changed, and this change didn't have any other dependencies.

Traits

ElementSnapshot
In order to compute restyle hints, we perform a selector match against a list of partial selectors whose rightmost simple selector may be sensitive to the thing being changed. We do this matching twice, once for the element as it exists now and once for the element as it existed at the time of the last restyle. If the results of the selector match differ, that means that the given partial selector is sensitive to the change, and we compute a restyle hint based on its combinator.