Module style::rule_tree

The rule tree.

Modules

core

level

map

source

unsafe_box

Structs

RuleTree

The rule tree, the structure servo uses to preserve the results of selector matching.

SelfAndAncestors

An iterator over a rule node and its ancestors.

ShadowCascadeOrder

A counter to track how many shadow root rules deep we are. This is used to handle:

StrongRuleNode

A strong reference to a rule node.

StyleSource

A style source for the rule node. It can either be a CSS style rule or a declaration block.

Enums

CascadeLevel

The cascade level these rules are relevant at, as per13.

Constants

RULE_NODE_SIZE

Used for some size assertions.