Module style::rule_tree::core

Structs

ChildKey 🔒

RuleNode 🔒

A node in the rule tree.

RuleTree

The rule tree, the structure servo uses to preserve the results of selector matching.

StrongRuleNode

A strong reference to a rule node.

WeakRuleNode 🔒

Constants

RULE_TREE_GC_INTERVAL 🔒

The number of RuleNodes added to the free list before we will consider doing a GC when calling maybe_gc(). (The value is copied from Gecko, where it likely did not result from a rigorous performance analysis.)

Functions

log_drop 🔒

log_new 🔒