Enum style::rule_tree::CascadeLevel

#[repr(u8)]

pub enum CascadeLevel {
    UANormal,
    PresHints,
    UserNormal,
    AuthorNormal,
    StyleAttributeNormal,
    Animations,
    AuthorImportant,
    StyleAttributeImportant,
    UserImportant,
    UAImportant,
    Transitions,
}

The cascade level these rules are relevant at, as per1.

The order of variants declared here is significant, and must be in ascending order of precedence.

Variants
UANormal

Normal User-Agent rules.

PresHints

Presentational hints.

UserNormal

User normal rules.

AuthorNormal

Author normal rules.

StyleAttributeNormal

Style attribute normal rules.

Animations

CSS animations and script-generated animations.

AuthorImportant

Author-supplied important rules.

StyleAttributeImportant

Style attribute important rules.

UserImportant

User important rules.

UAImportant

User-agent important rules.

Transitions

Transitions

Methods

impl CascadeLevel

fn is_unique_per_element(&self) -> bool

Returns whether this cascade level is unique per element, in which case we can replace the path in the cascade without fear.

fn is_important(&self) -> bool

Returns whether this cascade level represents important rules of some sort.

fn importance(&self) -> Importance

Returns the importance relevant for this rule. Pretty similar to is_important.