Enum style::rule_tree::CascadeLevel

#[repr(u8)]

pub enum CascadeLevel {
    UANormal,
    PresHints,
    UserNormal,
    AuthorNormal,
    StyleAttributeNormal,
    SMILOverride,
    Animations,
    AuthorImportant,
    StyleAttributeImportant,
    UserImportant,
    UAImportant,
    Transitions,
}

The cascade level these rules are relevant at, as per1.

The order of variants declared here is significant, and must be in ascending order of precedence.

Variants
UANormal

Normal User-Agent rules.

PresHints

Presentational hints.

UserNormal

User normal rules.

AuthorNormal

Author normal rules.

StyleAttributeNormal

Style attribute normal rules.

SMILOverride

SVG SMIL animations.

Animations

CSS animations and script-generated animations.

AuthorImportant

Author-supplied important rules.

StyleAttributeImportant

Style attribute important rules.

UserImportant

User important rules.

UAImportant

User-agent important rules.

Transitions

Transitions

NB: If this changes from being last, change from_byte below.

Methods

impl CascadeLevel

pub unsafe fn from_byte(byte: u8) -> Self

Converts a raw byte to a CascadeLevel.

pub fn guard<'a>(
&self,
guards: &'a StylesheetGuards<'a>
) -> &'a SharedRwLockReadGuard<'a>

Select a lock guard for this level

pub fn is_unique_per_element(&self) -> bool

Returns whether this cascade level is unique per element, in which case we can replace the path in the cascade without fear.

pub fn is_important(&self) -> bool

Returns whether this cascade level represents important rules of some sort.

pub fn importance(&self) -> Importance

Returns the importance relevant for this rule. Pretty similar to is_important.

pub fn is_animation(&self) -> bool

Returns whether this cascade level represents an animation rules.