Struct style::rule_tree::StrongRuleNode

pub struct StrongRuleNode {
    p: UnsafeBox<RuleNode>,
}

A strong reference to a rule node.

Fields
p: UnsafeBox<RuleNode>
Implementations

impl StrongRuleNode

fn new(n: Box<RuleNode>) -> Self

unsafe fn from_unsafe_box(p: UnsafeBox<RuleNode>) -> Self

unsafe fn downgrade(&self) -> WeakRuleNode

pub fn parent(&self) -> Option<&StrongRuleNode>

Get the parent rule node of this rule node.

pub(super) fn ensure_child(
&self,
root: &StrongRuleNode,
source: StyleSource,
level: CascadeLevel
) -> StrongRuleNode

pub fn style_source(&self) -> Option<&StyleSource>

Get the style source corresponding to this rule node. May return None if it’s the root node, which means that the node hasn’t matched any rules.

pub fn cascade_level(&self) -> CascadeLevel

The cascade level for this node

pub fn importance(&self) -> Importance

Get the importance that this rule node represents.

pub unsafe fn has_children_for_testing(&self) -> bool

Returns whether this node has any child, only intended for testing purposes.

pub(super) fn dump<W: Write>(
&self,
guards: &StylesheetGuards<'_>,
writer: &mut W,
indent: usize
)

impl StrongRuleNode

pub fn self_and_ancestors(&self) -> SelfAndAncestors<'_>

Get an iterator for this rule node and its ancestors.

pub fn has_animation_or_transition_rules(&self) -> bool

Returns true if there is either animation or transition level rule.

pub fn get_properties_overriding_animations(
&self,
guards: &StylesheetGuards<'_>
) -> (LonghandIdSet, bool)

Get a set of properties whose CascadeLevel are higher than Animations but not equal to Transitions.

If there are any custom properties, we set the boolean value of the returned tuple to true.