style::rule_tree::unsafe_box

Struct UnsafeBox

pub(super) struct UnsafeBox<T> {
    inner: ManuallyDrop<Box<T>>,
}

An unsafe box, derefs to T.

Fields
inner: ManuallyDrop<Box<T>>
Implementations

impl<T> UnsafeBox<T>

pub(super) fn from_box(value: Box<T>) -> Self

Creates a new unsafe box.

pub(super) unsafe fn from_raw(ptr: *mut T) -> Self

Creates a new box from a pointer.

Safety

The input should point to a valid T.

pub(super) unsafe fn clone(this: &Self) -> Self

Creates a new unsafe box from an existing one.

Safety

There is no refcounting or whatever else in an unsafe box, so this operation can lead to double frees.

pub(crate) unsafe fn deref_mut(this: &mut Self) -> &mut T

Returns a mutable reference to the inner value of this unsafe box.

Safety

Given Self::clone, nothing prevents anyone from creating multiple mutable references to the inner value, which is completely UB.

pub(super) unsafe fn drop(this: &mut Self)

Drops the inner value of this unsafe box.

Safety

Given this doesnโt consume the unsafe box itself, this has the same safety caveats as ManuallyDrop::drop.