style::scoped_tls

Struct ScopedTLS

pub struct ScopedTLS<'scope, T: Send> {
    pool: Option<&'scope ThreadPool>,
    slots: [RefCell<Option<T>>; 6],
}

A scoped TLS set, that is alive during the 'scope lifetime.

We use this on Servo to construct thread-local contexts, but clear them once we’re done with restyling.

Note that the cleanup is done on the thread that owns the scoped TLS, thus the Send bound.

Fields
pool: Option<&'scope ThreadPool>slots: [RefCell<Option<T>>; 6]
Implementations

impl<'scope, T: Send> ScopedTLS<'scope, T>

pub fn new(pool: Option<&'scope ThreadPool>) -> Self

Create a new scoped TLS that will last as long as this rayon threadpool reference.

pub fn current_thread_index(&self) -> usize

Returns the index corresponding to the calling thread in the thread pool.

pub fn borrow(&self) -> Ref<'_, Option<T>>

Return an immutable reference to the Option<T> that this thread owns.

pub fn borrow_mut(&self) -> RefMut<'_, Option<T>>

Return a mutable reference to the Option<T> that this thread owns.

pub fn ensure<F: FnOnce(&mut Option<T>)>(&self, f: F) -> RefMut<'_, T>

Ensure that the current data this thread owns is initialized, or initialize it using f. We want ensure() to be fast and inline, and we want to inline the memmove that initializes the Option. But we don’t want to inline space for the entire large T struct in our stack frame. That’s why we hand f a mutable borrow to write to instead of just having it return a T.

pub fn slots(&mut self) -> &mut [RefCell<Option<T>>]

Returns the slots. Safe because if we have a mut reference the tls can’t be referenced by any other thread.