Module style::selector_map

A data structure to efficiently index structs containing selectors by local name, ids and hash.

Structs

MaybeCaseInsensitiveHashMap

Wrapper for PrecomputedHashMap that does ASCII-case-insensitive lookup in quirks mode.

PrecomputedHasher

A hasher implementation that doesn’t hash anything, because it expects its input to be a suitable u32 hash.

SelectorMap

Map element data to selector-providing objects for which the last simple selector starts with them.

Enums

Bucket

Traits

SelectorMapEntry

A trait to abstract over a given selector map entry.

Functions

find_bucket

Searches a compound selector from left to right, and returns the appropriate bucket for it.

specific_bucket_for

Type Definitions

DisjointBuckets

PrecomputedHashMap

A simple alias for a hashmap using PrecomputedHasher.

PrecomputedHashSet

A simple alias for a hashset using PrecomputedHasher.