style::selector_map

Trait SelectorMapEntry

pub trait SelectorMapEntry: Sized + Clone {
    // Required method
    fn selector(&self) -> SelectorIter<'_, SelectorImpl>;
}

A trait to abstract over a given selector map entry.

Required Methods

fn selector(&self) -> SelectorIter<'_, SelectorImpl>

Gets the selector we should use to index in the selector map.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.