Trait style::selector_map::SelectorMapEntry

pub trait SelectorMapEntry: Sized + Clone {
    fn selector(&self) -> SelectorIter<'_, SelectorImpl>;
}

A trait to abstract over a given selector map entry.

Required methods

fn selector(&self) -> SelectorIter<'_, SelectorImpl>

Gets the selector we should use to index in the selector map.