Trait style::selector_map::SelectorMapEntry

pub trait SelectorMapEntry: Sized + Clone {
    // Required method
    fn selector(&self) -> SelectorIter<'_, SelectorImpl>;
}

A trait to abstract over a given selector map entry.

Required Methods

fn selector(&self) -> SelectorIter<'_, SelectorImpl>

Gets the selector we should use to index in the selector map.

Object Safety

This trait is not object safe.