Trait style::selector_map::SelectorMapEntry

pub trait SelectorMapEntry: Sized + Clone {
    fn selector(&self) -> SelectorIter<SelectorImpl>;
}

A trait to abstract over a given selector map entry.

Required Methods

fn selector(&self) -> SelectorIter<SelectorImpl>

Gets the selector we should use to index in the selector map.