Struct style::selector_matching::Stylist

pub struct Stylist<Impl: SelectorImplExt> {
    pub device: Device,
    viewport_constraints: Option<ViewportConstraints>,
    quirks_mode: bool,
    is_device_dirty: bool,
    element_map: PerPseudoElementSelectorMap<Impl>,
    pseudos_map: HashMap<Impl::PseudoElement, PerPseudoElementSelectorMap<Impl>, BuildHasherDefault<FnvHasher>>,
    precomputed_pseudo_element_decls: HashMap<Impl::PseudoElement, Vec<DeclarationBlock>, BuildHasherDefault<FnvHasher>>,
    rules_source_order: usize,
    state_deps: DependencySet<Impl>,
}

This structure holds all the selectors and device characteristics for a given document. The selectors are converted into Rules (defined in rust-selectors), and introduced in a SelectorMap depending on the pseudo-element (see PerPseudoElementSelectorMap), stylesheet origin (see PerOriginSelectorMap), and priority (see the normal and important fields in PerOriginSelectorMap).

This structure is effectively created once per pipeline, in the LayoutThread corresponding to that pipeline.

The stylist is parameterized on SelectorImplExt, a trait that extends selectors::parser::SelectorImpl, and that allows to customise what pseudo-classes and pseudo-elements are parsed. This is actually either ServoSelectorImpl, the implementation used by Servo's layout system in regular builds, or GeckoSelectorImpl, the implementation used in the geckolib port.

Fields
device: Device

Device that the stylist is currently evaluating against.

viewport_constraints: Option<ViewportConstraints>

Viewport constraints based on the current device.

quirks_mode: bool

If true, the quirks-mode stylesheet is applied.

is_device_dirty: bool

If true, the device has changed, and the stylist needs to be updated.

element_map: PerPseudoElementSelectorMap<Impl>

The current selector maps, after evaluating media rules against the current device.

pseudos_map: HashMap<Impl::PseudoElement, PerPseudoElementSelectorMap<Impl>, BuildHasherDefault<FnvHasher>>

The selector maps corresponding to a given pseudo-element (depending on the implementation)

precomputed_pseudo_element_decls: HashMap<Impl::PseudoElement, Vec<DeclarationBlock>, BuildHasherDefault<FnvHasher>>

Applicable declarations for a given non-eagerly cascaded pseudo-element. These are eagerly computed once, and then used to resolve the new computed values on the fly on layout.

rules_source_order: usize state_deps: DependencySet<Impl>

Selector dependencies used to compute restyle hints.

Methods

impl<Impl: SelectorImplExt> Stylist<Impl>

fn new(device: Device) -> Stylist<Impl>

fn update(&mut self, doc_stylesheets: &[Arc<Stylesheet<Impl>>], stylesheets_changed: bool) -> bool where Impl: 'static

fn add_stylesheet(&mut self, stylesheet: &Stylesheet<Impl>)

fn precomputed_values_for_pseudo(&self, pseudo: &Impl::PseudoElement, parent: Option<&Arc<Impl::ComputedValues>>) -> Option<Arc<Impl::ComputedValues>>

Computes the style for a given "precomputed" pseudo-element, taking the universal rules and applying them.

fn lazily_compute_pseudo_element_style<E>(&self, element: &E, pseudo: &Impl::PseudoElement, parent: &Arc<Impl::ComputedValues>) -> Option<Arc<Impl::ComputedValues>> where E: Element<Impl=Impl> + PresentationalHintsSynthetizer

fn compute_restyle_hint<E>(&self, element: &E, snapshot: &ElementSnapshot, current_state: ElementState) -> RestyleHint where E: Element<Impl=Impl> + Clone

fn set_device(&mut self, device: Device, stylesheets: &[Arc<Stylesheet<Impl>>])

fn viewport_constraints(&self) -> &Option<ViewportConstraints>

fn set_quirks_mode(&mut self, enabled: bool)

fn push_applicable_declarations<E, V>(&self, element: &E, parent_bf: Option<&BloomFilter>, style_attribute: Option<&PropertyDeclarationBlock>, pseudo_element: Option<&Impl::PseudoElement>, applicable_declarations: &mut V) -> bool where E: Element<Impl=Impl> + PresentationalHintsSynthetizer, V: VecLike<DeclarationBlock>

Returns the applicable CSS declarations for the given element. This corresponds to ElementRuleCollector in WebKit.

The returned boolean indicates whether the style is shareable; that is, whether the matched selectors are simple enough to allow the matching logic to be reduced to the logic in css::matching::PrivateMatchMethods::candidate_element_allows_for_style_sharing.

fn is_device_dirty(&self) -> bool