Struct style::selector_matching::Stylist

pub struct Stylist {
    pub device: Device,
    // some fields omitted
}

Fields

device

Methods

impl Stylist

fn new(device: Device) -> Stylist

fn update(&mut self) -> bool

fn set_device(&mut self, device: Device)

fn add_quirks_mode_stylesheet(&mut self)

fn add_stylesheet(&mut self, stylesheet: Stylesheet)

fn push_applicable_declarations<'a, N, V>(&self, element: &N, parent_bf: &Option<Box<BloomFilter>>, style_attribute: Option<&PropertyDeclarationBlock>, pseudo_element: Option<PseudoElement>, applicable_declarations: &mut V) -> bool where N: TNode<'a>, Element: TElementAttributes, V: VecLike<DeclarationBlock>

Returns the applicable CSS declarations for the given element. This corresponds to ElementRuleCollector in WebKit.

The returned boolean indicates whether the style is shareable; that is, whether the matched selectors are simple enough to allow the matching logic to be reduced to the logic in css::matching::PrivateMatchMethods::candidate_element_allows_for_style_sharing.