Module style::selector_parser

The pseudo-classes and pseudo-elements supported by the style system.

Reexports

pub use gecko::selector_parser::*;

pub use gecko::snapshot::GeckoElementSnapshot as Snapshot;

pub use gecko::restyle_damage::GeckoRestyleDamage as RestyleDamage;

Structs

SelectorParser
Servo's selector parser.

Enums

PseudoElementCascadeType
This enumeration determines if a pseudo-element is eagerly cascaded or not.

Traits

ElementExt
An extension to rust-selector's Element trait.

Functions

attr_equals_selector_is_shareable
Checks whether we can share style if an element matches a given attribute-selector that checks for equality ([attr_name="attr_value"]) easily.

attr_exists_selector_is_shareable
Checks whether we can share style if an element matches a given attribute-selector that checks for existence ([attr_name]) easily.

Type Definitions

AttrValue
A convenient alias for the type that represents an attribute value used for selector parser implementation.

PreExistingComputedValues
A type that represents the previous computed values needed for restyle damage calculation.