Module style::selector_parser

The pseudo-classes and pseudo-elements supported by the style system.

Reexports

pub use servo::selector_parser::*;

pub use servo::selector_parser::ServoElementSnapshot as Snapshot;

pub use servo::restyle_damage::ServoRestyleDamage as RestyleDamage;

Structs

PerPseudoElementMap
A per-pseudo map, from a given pseudo to a T.

SelectorParser
Servo's selector parser.

Enums

Direction
Values for the :dir() pseudo class

PseudoElementCascadeType
This enumeration determines if a pseudo-element is eagerly cascaded or not.

Type Definitions

AttrValue
A convenient alias for the type that represents an attribute value used for selector parser implementation.