Struct style::selector_parser::SelectorParser

pub struct SelectorParser<'a> {
    pub stylesheet_origin: Origin,
    pub namespaces: &'a Namespaces,
    pub url_data: Option<&'a UrlExtraData>,
}

Servo's selector parser.

Fields
stylesheet_origin: Origin

The origin of the stylesheet we're parsing.

namespaces: &'a Namespaces

The namespace set of the stylesheet.

url_data: Option<&'a UrlExtraData>

The extra URL data of the stylesheet, which is used to look up whether we are parsing a chrome:// URL style sheet.

Methods

impl<'a> SelectorParser<'a>

pub fn parse_author_origin_no_namespace(
input: &str
) -> Result<SelectorList<SelectorImpl>, ParseError>

Parse a selector list with an author origin and without taking into account namespaces.

This is used for some DOM APIs like querySelector.

pub fn in_user_agent_stylesheet(&self) -> bool

Whether we're parsing selectors in a user-agent stylesheet.

pub fn chrome_rules_enabled(&self) -> bool

Whether we're parsing selectors in a stylesheet that has chrome privilege.