Module style::servo::media_queries

Servo’s media-query device and expression representation.

Structs

Device

A device is a structure that represents the current media a given document is displayed in.

Enums

ForcedColors

Possible values for the forced-colors media query. https://drafts.csswg.org/mediaqueries-5/#forced-colors

Scan 🔒

Statics

MEDIA_FEATURES

A list with all the media features that Servo supports.

Traits

FontMetricsProvider

A trait used to query font metrics in clients of Stylo. This is used by Device to query font metrics in a way that is specific to the client using Stylo.

Functions

eval_device_pixel_ratio 🔒

https://compat.spec.whatwg.org/#css-media-queries-webkit-device-pixel-ratio

eval_resolution 🔒

https://drafts.csswg.org/mediaqueries-4/#resolution

eval_scan 🔒

https://drafts.csswg.org/mediaqueries-4/#scan

eval_width 🔒

https://drafts.csswg.org/mediaqueries-4/#width