style::servo::media_queries

Struct Device

pub struct Device {

    media_type: MediaType,
    viewport_size: Size2D<f32, CSSPixel>,
    device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>,
    quirks_mode: QuirksMode,
    root_font_size: AtomicU32,
    root_line_height: AtomicU32,
    used_root_font_size: AtomicBool,
    used_root_line_height: AtomicBool,
    used_font_metrics: AtomicBool,
    used_viewport_units: AtomicBool,
    prefers_color_scheme: PrefersColorScheme,
    environment: CssEnvironment,
    font_metrics_provider: Box<dyn FontMetricsProvider>,
    default_computed_values: Arc<ComputedValues>,

}

A device is a structure that represents the current media a given document is displayed in.

This is the struct against which media queries are evaluated.

Fields
media_type: MediaType

The current media type used by de device.

viewport_size: Size2D<f32, CSSPixel>

The current viewport size, in CSS pixels.

device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>

The current device pixel ratio, from CSS pixels to device pixels.

quirks_mode: QuirksMode

The current quirks mode.

root_font_size: AtomicU32

The font size of the root element This is set when computing the style of the root element, and used for rem units in other elements

When computing the style of the root element, there can’t be any other style being computed at the same time, given we need the style of the parent to compute everything else. So it is correct to just use a relaxed atomic here.

root_line_height: AtomicU32

Line height of the root element, used for rlh units in other elements.

used_root_font_size: AtomicBool

Whether any styles computed in the document relied on the root font-size by using rem units.

used_root_line_height: AtomicBool

Whether any styles computed in the document relied on the root line-height by using rlh units.

used_font_metrics: AtomicBool

Whether any styles computed in the document relied on font metrics.

used_viewport_units: AtomicBool

Whether any styles computed in the document relied on the viewport size.

prefers_color_scheme: PrefersColorScheme

Whether the user prefers light mode or dark mode

environment: CssEnvironment

The CssEnvironment object responsible of getting CSS environment variables.

font_metrics_provider: Box<dyn FontMetricsProvider>

An implementation of a trait which implements support for querying font metrics.

default_computed_values: Arc<ComputedValues>

The default computed values for this Device.

Implementations

impl Device

pub fn new( media_type: MediaType, quirks_mode: QuirksMode, viewport_size: Size2D<f32, CSSPixel>, device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>, font_metrics_provider: Box<dyn FontMetricsProvider>, default_computed_values: Arc<ComputedValues>, prefers_color_scheme: PrefersColorScheme, ) -> Device

Trivially construct a new Device.

pub fn environment(&self) -> &CssEnvironment

Get the relevant environment to resolve env() functions.

pub fn default_computed_values(&self) -> &ComputedValues

Return the default computed values for this device.

pub fn root_font_size(&self) -> CSSPixelLength

Get the font size of the root element (for rem)

pub fn set_root_font_size(&self, size: f32)

Set the font size of the root element (for rem), in zoom-independent CSS pixels.

pub fn root_line_height(&self) -> CSSPixelLength

Get the line height of the root element (for rlh)

pub fn set_root_line_height(&self, size: f32)

Set the line height of the root element (for rlh), in zoom-independent CSS pixels.

pub fn calc_line_height( &self, font: &Font, _writing_mode: WritingMode, _element: Option<()>, ) -> NonNegativeLength

Returns the computed line-height for the font in a given computed values instance.

If you pass down an element, then the used line-height is returned.

pub fn quirks_mode(&self) -> QuirksMode

Get the quirks mode of the current device.

pub fn set_body_text_color(&self, _color: AbsoluteColor)

Sets the body text color for the “inherit color from body” quirk.

https://quirks.spec.whatwg.org/#the-tables-inherit-color-from-body-quirk

pub fn base_size_for_generic(&self, generic: GenericFontFamily) -> Length

Gets the base size given a generic font family.

pub fn animation_name_may_be_referenced(&self, _: &KeyframesName) -> bool

Whether a given animation name may be referenced from style.

pub fn used_root_font_size(&self) -> bool

Returns whether we ever looked up the root font size of the Device.

pub fn used_root_line_height(&self) -> bool

Returns whether we ever looked up the root line-height of the device.

pub fn used_font_metrics(&self) -> bool

Returns whether font metrics have been queried.

pub fn viewport_size(&self) -> Size2D<f32, CSSPixel>

Get the viewport size on this Device.

pub fn set_viewport_size(&mut self, viewport_size: Size2D<f32, CSSPixel>)

Set the viewport size on this Device.

Note that this does not update any associated Stylist. For this you must call Stylist::media_features_change_changed_style and Stylist::force_stylesheet_origins_dirty.

pub fn au_viewport_size(&self) -> UntypedSize2D<Au>

Returns the viewport size of the current device in app units, needed, among other things, to resolve viewport units.

pub fn au_viewport_size_for_viewport_unit_resolution( &self, _: ViewportVariant, ) -> UntypedSize2D<Au>

Like the above, but records that we’ve used viewport units.

pub fn used_viewport_units(&self) -> bool

Whether viewport units were used since the last device change.

pub fn app_units_per_device_pixel(&self) -> i32

Returns the number of app units per device pixel we’re using currently.

pub fn device_pixel_ratio_ignoring_full_zoom( &self, ) -> Scale<f32, CSSPixel, DevicePixel>

Returns the device pixel ratio, ignoring the full zoom factor.

pub fn device_pixel_ratio(&self) -> Scale<f32, CSSPixel, DevicePixel>

Returns the device pixel ratio.

pub fn set_device_pixel_ratio( &mut self, device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>, )

Set a new device pixel ratio on this Device.

Note that this does not update any associated Stylist. For this you must call Stylist::media_features_change_changed_style and Stylist::force_stylesheet_origins_dirty.

pub fn scrollbar_inline_size(&self) -> CSSPixelLength

Gets the size of the scrollbar in CSS pixels.

pub fn query_font_metrics( &self, vertical: bool, font: &Font, base_size: CSSPixelLength, flags: QueryFontMetricsFlags, ) -> FontMetrics

Queries font metrics using the FontMetricsProvider interface.

pub fn media_type(&self) -> MediaType

Return the media type of the current device.

pub fn forced_colors(&self) -> ForcedColors

Returns whether document colors are enabled.

pub fn default_background_color(&self) -> AbsoluteColor

Returns the default background color.

pub fn default_color(&self) -> AbsoluteColor

Returns the default foreground color.

pub fn set_color_scheme(&mut self, new_color_scheme: PrefersColorScheme)

Set the PrefersColorScheme value on this Device.

Note that this does not update any associated Stylist. For this you must call Stylist::media_features_change_changed_style and Stylist::force_stylesheet_origins_dirty.

pub fn color_scheme(&self) -> PrefersColorScheme

Returns the color scheme of this Device.

pub(crate) fn is_dark_color_scheme(&self, _: ColorSchemeFlags) -> bool

pub fn safe_area_insets(&self) -> SideOffsets2D<f32, CSSPixel>

Returns safe area insets

pub fn is_supported_mime_type(&self, mime_type: &str) -> bool

Returns true if the given MIME type is supported

pub fn chrome_rules_enabled_for_document(&self) -> bool

Return whether the document is a chrome document.