Struct style::servo::media_queries::Device

pub struct Device {
    media_type: MediaType,
    viewport_size: Size2D<f32, CSSPixel>,
    device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>,
    root_font_size: AtomicIsize,
    used_root_font_size: AtomicBool,
    used_viewport_units: AtomicBool,
    environment: CssEnvironment,
}

A device is a structure that represents the current media a given document is displayed in.

This is the struct against which media queries are evaluated.

Fields
media_type: MediaType

The current media type used by de device.

viewport_size: Size2D<f32, CSSPixel>

The current viewport size, in CSS pixels.

device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>

The current device pixel ratio, from CSS pixels to device pixels.

root_font_size: AtomicIsize

The font size of the root element This is set when computing the style of the root element, and used for rem units in other elements

When computing the style of the root element, there can't be any other style being computed at the same time, given we need the style of the parent to compute everything else. So it is correct to just use a relaxed atomic here.

used_root_font_size: AtomicBool

Whether any styles computed in the document relied on the root font-size by using rem units.

used_viewport_units: AtomicBool

Whether any styles computed in the document relied on the viewport size.

environment: CssEnvironment

The CssEnvironment object responsible of getting CSS environment variables.

Implementations

impl Device

pub fn new(
media_type: MediaType,
viewport_size: Size2D<f32, CSSPixel>,
device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>
) -> Device

Trivially construct a new Device.

pub fn environment(&self) -> &CssEnvironment

Get the relevant environment to resolve env() functions.

pub fn default_computed_values(&self) -> &ComputedValues

Return the default computed values for this device.

pub fn root_font_size(&self) -> Au

Get the font size of the root element (for rem)

pub fn set_root_font_size(&self, size: Au)

Set the font size of the root element (for rem)

pub fn set_body_text_color(&self, _color: RGBA)

Sets the body text color for the "inherit color from body" quirk.

https://quirks.spec.whatwg.org/#the-tables-inherit-color-from-body-quirk

pub fn animation_name_may_be_referenced(&self, _: &KeyframesName) -> bool

Whether a given animation name may be referenced from style.

pub fn used_root_font_size(&self) -> bool

Returns whether we ever looked up the root font size of the Device.

pub fn au_viewport_size(&self) -> UntypedSize2D<Au>

Returns the viewport size of the current device in app units, needed, among other things, to resolve viewport units.

pub fn au_viewport_size_for_viewport_unit_resolution(&self) -> UntypedSize2D<Au>

Like the above, but records that we've used viewport units.

pub fn used_viewport_units(&self) -> bool

Whether viewport units were used since the last device change.

pub fn device_pixel_ratio(&self) -> Scale<f32, CSSPixel, DevicePixel>

Returns the device pixel ratio.

pub fn account_for_viewport_rule(&mut self, constraints: &ViewportConstraints)

Take into account a viewport rule taken from the stylesheets.

pub fn media_type(&self) -> MediaType

Return the media type of the current device.

pub fn use_document_colors(&self) -> bool

Returns whether document colors are enabled.

pub fn default_background_color(&self) -> RGBA

Returns the default background color.

pub fn default_color(&self) -> RGBA

Returns the default color color.

pub fn safe_area_insets(&self) -> SideOffsets2D<f32, CSSPixel>

Returns safe area insets