Struct style::servo::media_queries::Device

pub struct Device { /* fields omitted */ }

A device is a structure that represents the current media a given document is displayed in.

This is the struct against which media queries are evaluated.

Methods

impl Device

pub fn new(
media_type: MediaType,
viewport_size: TypedSize2D<f32, CSSPixel>,
device_pixel_ratio: TypedScale<f32, CSSPixel, DevicePixel>
) -> Device

Trivially construct a new Device.

pub fn default_computed_values(&self) -> &ComputedValues

Return the default computed values for this device.

pub fn root_font_size(&self) -> Au

Get the font size of the root element (for rem)

pub fn set_root_font_size(&self, size: Au)

Set the font size of the root element (for rem)

pub fn set_body_text_color(&self, _color: RGBA)

Sets the body text color for the "inherit color from body" quirk.

https://quirks.spec.whatwg.org/#the-tables-inherit-color-from-body-quirk

pub fn used_root_font_size(&self) -> bool

Returns whether we ever looked up the root font size of the Device.

pub fn au_viewport_size(&self) -> Size2D<Au>

Returns the viewport size of the current device in app units, needed, among other things, to resolve viewport units.

pub fn au_viewport_size_for_viewport_unit_resolution(&self) -> Size2D<Au>

Like the above, but records that we've used viewport units.

pub fn used_viewport_units(&self) -> bool

Whether viewport units were used since the last device change.

pub fn device_pixel_ratio(&self) -> TypedScale<f32, CSSPixel, DevicePixel>

Returns the device pixel ratio.

pub fn account_for_viewport_rule(&mut self, constraints: &ViewportConstraints)

Take into account a viewport rule taken from the stylesheets.

pub fn media_type(&self) -> MediaType

Return the media type of the current device.

pub fn use_document_colors(&self) -> bool

Returns whether document colors are enabled.

pub fn default_background_color(&self) -> RGBA

Returns the default background color.