style::servo

Module restyle_damage

The restyle damage is a hint that tells layout which kind of operations may be needed in presence of incremental style changes.

Structs

ServoRestyleDamage

Major phases of layout that need to be run due to the damage to a node during restyling. In addition to the 4 bytes used for that, the rest of the u16 is exposed as an extension point for users of the crate to add their own custom types of damage that correspond to the layout system they are implementing.

Functions

augmented_restyle_damage_rebuild_box 🔒

augmented_restyle_damage_rebuild_stacking_context 🔒

compute_damage 🔒