style::servo

Module selector_parser

Servo’s selector parser.

Structs

CustomState

The type used to store the state argument to the :state pseudo-class.

ExtraMatchingData

A set of extra data to carry along with the matching context, either for selector-matching or invalidation.

SelectorImpl

The abstract struct we implement the selector parser implementation on top of.

ServoElementSnapshot

Servo’s version of an element snapshot.

SnapshotMap

A map from elements to snapshots for the Servo style back-end.

Enums

NonTSPseudoClass

A non tree-structural pseudo-class. See https://drafts.csswg.org/selectors-4/#structural-pseudos

PseudoElement

A pseudo-element, both public and private.

Constants

EAGER_PSEUDO_COUNT

The number of eager pseudo-elements. Keep this in sync with cascade_type.

PSEUDO_COUNT

The count of all pseudo-elements.

Functions

extended_filtering

Returns whether the language is matched, as defined by RFC 4647.

Type Aliases

Lang

The type used for storing :lang arguments.