style::servo::selector_parser

Enum NonTSPseudoClass

pub enum NonTSPseudoClass {

    Active,
    AnyLink,
    Autofill,
    Checked,
    CustomState(CustomState),
    Default,
    Defined,
    Disabled,
    Enabled,
    Focus,
    FocusWithin,
    FocusVisible,
    Fullscreen,
    Hover,
    InRange,
    Indeterminate,
    Invalid,
    Lang(Lang),
    Link,
    Modal,
    MozMeterOptimum,
    MozMeterSubOptimum,
    MozMeterSubSubOptimum,
    Optional,
    OutOfRange,
    PlaceholderShown,
    PopoverOpen,
    ReadOnly,
    ReadWrite,
    Required,
    ServoNonZeroBorder,
    Target,
    UserInvalid,
    UserValid,
    Valid,
    Visited,

}

A non tree-structural pseudo-class. See https://drafts.csswg.org/selectors-4/#structural-pseudos

Variants

Active

AnyLink

Autofill

Checked

CustomState(CustomState)

The :state` pseudo-class.

Default

Defined

Disabled

Enabled

Focus

FocusWithin

FocusVisible

Fullscreen

Hover

InRange

Indeterminate

Invalid

Lang(Lang)

Link

Modal

MozMeterOptimum

MozMeterSubOptimum

MozMeterSubSubOptimum

Optional

OutOfRange

PlaceholderShown

PopoverOpen

ReadOnly

ReadWrite

Required

ServoNonZeroBorder

Target

UserInvalid

UserValid

Valid

Visited

Implementations

impl NonTSPseudoClass

pub fn state_flag(&self) -> ElementState

Gets a given state flag for this pseudo-class. This is used to do selector matching, and itโs set from the DOM.

pub fn document_state_flag(&self) -> DocumentState

Get the document state flag associated with a pseudo-class, if any.

pub fn needs_cache_revalidation(&self) -> bool

Returns true if the given pseudoclass should trigger style sharing cache revalidation.