style::servo::selector_parser

Enum PseudoElement

#[repr(usize)]
pub enum PseudoElement {

    After = 0,
    Before = 1,
    Selection = 2,
    Backdrop = 3,
    DetailsSummary = 4,
    DetailsContent = 5,
    Marker = 6,
    ColorSwatch = 7,
    Placeholder = 8,
    ServoTextControlInnerContainer = 9,
    ServoTextControlInnerEditor = 10,
    ServoAnonymousBox = 11,
    ServoAnonymousTable = 12,
    ServoAnonymousTableCell = 13,
    ServoAnonymousTableRow = 14,
    ServoTableGrid = 15,
    ServoTableWrapper = 16,

}

A pseudo-element, both public and private.

NB: If you add to this list, be sure to update each_simple_pseudo_element too.

Variants

After = 0

Before = 1

Selection = 2

Backdrop = 3

DetailsSummary = 4

DetailsContent = 5

Marker = 6

ColorSwatch = 7

Placeholder = 8

ServoTextControlInnerContainer = 9

ServoTextControlInnerEditor = 10

ServoAnonymousBox = 11

ServoAnonymousTable = 12

ServoAnonymousTableCell = 13

ServoAnonymousTableRow = 14

ServoTableGrid = 15

ServoTableWrapper = 16

Implementations

impl PseudoElement

pub fn eager_index(&self) -> usize

Gets the canonical index of this eagerly-cascaded pseudo-element.

pub fn index(&self) -> usize

An index for this pseudo-element to be indexed in an enumerated array.

pub fn pseudo_none_array<T>() -> [Option<T>; 17]

An array of None, one per pseudo-element.

pub fn from_eager_index(i: usize) -> Self

Creates a pseudo-element from an eager index.

pub fn is_before_or_after(&self) -> bool

Whether the current pseudo element is ::before or ::after.

pub fn is_unknown_webkit_pseudo_element(&self) -> bool

Whether this is an unknown ::-webkit- pseudo-element.

pub fn is_marker(&self) -> bool

Whether this pseudo-element is the ::marker pseudo.

pub fn is_selection(&self) -> bool

Whether this pseudo-element is the ::selection pseudo.

pub fn is_before(&self) -> bool

Whether this pseudo-element is the ::before pseudo.

pub fn is_after(&self) -> bool

Whether this pseudo-element is the ::after pseudo.

pub fn is_first_letter(&self) -> bool

Whether the current pseudo element is :first-letter

pub fn is_first_line(&self) -> bool

Whether the current pseudo element is :first-line

pub fn is_color_swatch(&self) -> bool

Whether this pseudo-element is representing the color swatch inside an <input> element.

pub fn is_eager(&self) -> bool

Whether this pseudo-element is eagerly-cascaded.

pub fn is_lazy(&self) -> bool

Whether this pseudo-element is lazily-cascaded.

pub fn is_anon_box(&self) -> bool

Whether this pseudo-element is for an anonymous box.

pub fn skip_item_display_fixup(&self) -> bool

Whether this pseudo-element skips flex/grid container display-based fixup.

pub fn is_precomputed(&self) -> bool

Whether this pseudo-element is precomputed.

pub fn cascade_type(&self) -> PseudoElementCascadeType

Returns which kind of cascade type has this pseudo.

See the documentation for PseudoElementCascadeType for how we choose which cascade type to use.

Note: Keep eager pseudos in sync with EAGER_PSEUDO_COUNT and EMPTY_PSEUDO_ARRAY in style/data.rs

pub fn canonical(&self) -> PseudoElement

Covert non-canonical pseudo-element to canonical one, and keep a canonical one as it is.

pub fn pseudo_info(&self)

Stub, only Gecko needs this

pub fn property_restriction(&self) -> Option<PropertyFlags>

Property flag that properties must have to apply to this pseudo-element.

pub fn should_exist(&self, style: &ComputedValues) -> bool

Whether this pseudo-element should actually exist if it has the given styles.