Struct style::servo::selector_parser::ServoElementSnapshot

pub struct ServoElementSnapshot {
    pub state: Option<ElementState>,
    pub attrs: Option<Vec<(AttrIdentifier, AttrValue)>>,
    pub is_html_element_in_html_document: bool,
    pub class_changed: bool,
    pub id_changed: bool,
    pub other_attributes_changed: bool,
}

Servo's version of an element snapshot.

Fields
state: Option<ElementState>

The stored state of the element.

attrs: Option<Vec<(AttrIdentifier, AttrValue)>>

The set of stored attributes and its values.

is_html_element_in_html_document: bool

Whether this element is an HTML element in an HTML document.

class_changed: bool

Whether the class attribute changed or not.

id_changed: bool

Whether the id attribute changed or not.

other_attributes_changed: bool

Whether other attributes other than id or class changed or not.

Methods

impl ServoElementSnapshot

pub fn new(is_html_element_in_html_document: bool) -> Self

Create an empty element snapshot.

pub fn id_changed(&self) -> bool

Returns whether the id attribute changed or not.

pub fn class_changed(&self) -> bool

Returns whether the class attribute changed or not.

pub fn other_attr_changed(&self) -> bool

Returns whether other attributes other than id or class changed or not.

impl ServoElementSnapshot

pub fn attr_matches(
&self,
ns: &NamespaceConstraint<&Namespace>,
local_name: &LocalName,
operation: &AttrSelectorOperation<&String>
) -> bool

selectors::Element::attr_matches