Type Alias style::servo::url::SpecifiedUrl

pub type SpecifiedUrl = CssUrl;

A specified url() value for servo.

Aliased Type

struct SpecifiedUrl {
    original: Option<Arc<String>>,
    resolved: Option<Arc<Url>>,
}

Fields
original: Option<Arc<String>>

The original URI. This might be optional since we may insert computed values of images into the cascade directly, and we don’t bother to convert their serialization.

Refcounted since cloning this should be cheap and data: uris can be really large.

resolved: Option<Arc<Url>>

The resolved value for the url, if valid.