Module style::shared_lock

Different objects protected by the same lock

Modules

compile_time_assert

Structs

DeepCloneParams

Parameters needed for deep clones.

Locked

Data protect by a shared lock.

SharedRwLock

A shared read/write lock that can protect multiple objects.

SharedRwLockReadGuard

Proof that a shared lock was obtained for reading (servo).

SharedRwLockWriteGuard

Proof that a shared lock was obtained for writing (servo).

StylesheetGuards

Guards for a document

Traits

DeepCloneWithLock

A trait to do a deep clone of a given CSS type. Gets a lock and a read guard, in order to be able to read and clone nested structures.

ToCssWithGuard

Like ToCss, but with a lock guard given by the caller, and with the writer specified concretely rather than with a parameter.