Trait style::shared_lock::DeepCloneWithLock

pub trait DeepCloneWithLock: Sized {
    // Required method
    fn deep_clone_with_lock(
        &self,
        lock: &SharedRwLock,
        guard: &SharedRwLockReadGuard<'_>,
    ) -> Self;
}

A trait to do a deep clone of a given CSS type. Gets a lock and a read guard, in order to be able to read and clone nested structures.

Required Methods

fn deep_clone_with_lock( &self, lock: &SharedRwLock, guard: &SharedRwLockReadGuard<'_>, ) -> Self

Deep clones this object.

Object Safety

This trait is not object safe.