Trait style::shared_lock::ToCssWithGuard

pub trait ToCssWithGuard {
    fn to_css<W>(&self, guard: &SharedRwLockReadGuard, dest: &mut W) -> Result
    where
        W: Write;

    fn to_css_string(&self, guard: &SharedRwLockReadGuard) -> String { ... }
}

Like ToCss, but with a lock guard given by the caller.

Required Methods

fn to_css<W>(&self, guard: &SharedRwLockReadGuard, dest: &mut W) -> Result where
W: Write,

Serialize self in CSS syntax, writing to dest, using the given lock guard.

Provided Methods

fn to_css_string(&self, guard: &SharedRwLockReadGuard) -> String

Serialize self in CSS syntax using the given lock guard and return a string.

(This is a convenience wrapper for to_css and probably should not be overridden.)