style::sharing::checks

Function revalidate

pub fn revalidate<E>(
    target: &mut StyleSharingTarget<E>,
    candidate: &mut StyleSharingCandidate<E>,
    shared_context: &SharedStyleContext<'_>,
    bloom: &StyleBloom<E>,
    selector_caches: &mut SelectorCaches,
) -> bool
where
    E: TElement,

Whether a given element and a candidate match the same set of “revalidation” selectors.

Revalidation selectors are those that depend on the DOM structure, like :first-child, etc, or on attributes that we don’t check off-hand (pretty much every attribute selector except id and class.