Struct style::sharing::StyleSharingTarget

pub struct StyleSharingTarget<E: TElement> { /* fields omitted */ }

An element we want to test against the style sharing cache.

Methods

impl<E: TElement> StyleSharingTarget<E>

pub fn new(element: E) -> Self

Trivially construct a new StyleSharingTarget to test against the cache.

pub fn share_style_if_possible(
&mut self,
context: &mut StyleContext<E>
) -> Option<ResolvedElementStyles>

Attempts to share a style with another node.

pub fn take_validation_data(&mut self) -> ValidationData

Gets the validation data used to match against this target, if any.