Module style::str

String utils for attributes and similar stuff.

Statics

HTML_SPACE_CHARACTERS
A "space character" according to:

Functions

char_is_whitespace
Whether a character is a HTML whitespace character.

is_ascii_digit
Character is ascii digit

is_whitespace
Whether all the string is HTML whitespace.

read_exponent
Reads an exponent from an iterator over chars, for example e100.

read_fraction
Read a decimal fraction.

read_numbers
Read a set of ascii digits and read them into a number.

split_commas
Split a string on commas.

split_html_space_chars
Split a string on HTML whitespace.

starts_with_ignore_ascii_case
Returns true if a given string has a given prefix with case-insensitive match.

str_join
Join a set of strings with a given delimiter join.

string_as_ascii_lowercase
Returns an ascii lowercase version of a string, only allocating if needed.

Type Definitions

StaticCharVec
A static slice of characters.

StaticStringVec
A static slice of strs.