Struct style::style_adjuster::StyleAdjuster

pub struct StyleAdjuster<'a, 'b: 'a> {
    style: &'a mut StyleBuilder<'b>,
}

A struct that implements all the adjustment methods.

NOTE(emilio): If new adjustments are introduced that depend on reset properties of the parent, you may need tweaking the ChildCascadeRequirement code in matching.rs.

NOTE(emilio): Also, if new adjustments are introduced that break the following invariant:

Given same tag name, namespace, rules and parent style, two elements would end up with exactly the same style.

Then you need to adjust the lookup_by_rules conditions in the sharing cache.

Fields
style: &'a mut StyleBuilder<'b>
Implementations

impl<'a, 'b: 'a> StyleAdjuster<'a, 'b>

pub fn new(style: &'a mut StyleBuilder<'b>) -> Self

Trivially constructs a new StyleAdjuster.

fn adjust_for_top_layer(&mut self)

https://fullscreen.spec.whatwg.org/#new-stacking-layer

Any position value other than ‘absolute’ and ‘fixed’ are computed to ‘absolute’ if the element is in a top layer.

fn adjust_for_position(&mut self)

CSS 2.1 section 9.7:

If ‘position’ has the value ‘absolute’ or ‘fixed’, […] the computed value of ‘float’ is ‘none’.

fn skip_item_display_fixup<E>(&self, element: Option<E>) -> bool where
E: TElement,

Whether we should skip any item-based display property blockification on this element.

fn blockify_if_necessary<E>(
&mut self,
layout_parent_style: &ComputedValues,
element: Option<E>
) where
E: TElement,

Apply the blockification rules based on the table in CSS 2.2 section 9.7. https://drafts.csswg.org/css2/visuren.html#dis-pos-flo A ::marker pseudo-element with ‘list-style-position:outside’ needs to have its ‘display’ blockified, unless the ::marker is for an inline list-item (for which ‘list-style-position:outside’ behaves as ‘inside’). https://drafts.csswg.org/css-lists-3/#list-style-position-property

fn set_bits(&mut self)

Compute a few common flags for both text and element’s style.

fn adjust_for_writing_mode(&mut self, layout_parent_style: &ComputedValues)

https://drafts.csswg.org/css-writing-modes-3/#block-flow:

If a box has a different writing-mode value than its containing block:

   - If the box has a specified display of inline, its display
     computes to inline-block. [CSS21]

This matches the adjustment that Gecko does, not exactly following the spec. See also:

https://lists.w3.org/Archives/Public/www-style/2017Mar/0045.html https://github.com/servo/servo/issues/15754

fn adjust_for_alignment(&mut self, layout_parent_style: &ComputedValues)

This implements an out-of-date spec. The new spec moves the handling of this to layout, which Gecko implements but Servo doesn’t.

See https://github.com/servo/servo/issues/15229

fn adjust_for_border_width(&mut self)

The initial value of border-*-width may be changed at computed value time.

This is moved to properties.rs for convenience.

fn adjust_for_outline(&mut self)

The initial value of outline-width may be changed at computed value time.

fn adjust_for_overflow(&mut self)

CSS overflow-x and overflow-y require some fixup as well in some cases. https://drafts.csswg.org/css-overflow-3/#overflow-properties “Computed value: as specified, except with visible/clip computing to auto/hidden (respectively) if one of overflow-x or overflow-y is neither visible nor clip.”

fn adjust_for_text_decorations_in_effect(&mut self)

Computes the used text decoration for Servo.

FIXME(emilio): This is a layout tree concept, should move away from style, since otherwise we’re going to have the same subtle bugs WebKit and Blink have with this very same thing.

fn adjust_for_visited<E>(&mut self, element: Option<E>) where
E: TElement,

Computes the RELEVANT_LINK_VISITED flag based on the parent style and on whether we’re a relevant link.

NOTE(emilio): We don’t do this for text styles, which is… dubious, but Gecko doesn’t seem to do it either. It’s extremely easy to do if needed though.

FIXME(emilio): This isn’t technically a style adjustment thingie, could it move somewhere else?

pub fn adjust<E>(
&mut self,
layout_parent_style: &ComputedValues,
element: Option<E>
) where
E: TElement,

Adjusts the style to account for various fixups that don’t fit naturally into the cascade.

When comparing to Gecko, this is similar to the work done by ComputedStyle::ApplyStyleFixups, plus some parts of nsStyleSet::GetContext.