Struct style::style_adjuster::StyleAdjuster

pub struct StyleAdjuster<'a, 'b: 'a> { /* fields omitted */ }

A struct that implements all the adjustment methods.

NOTE(emilio): If new adjustments are introduced that depend on reset properties of the parent, you may need tweaking the ChildCascadeRequirement code in matching.rs.

Methods

impl<'a, 'b: 'a> StyleAdjuster<'a, 'b>

pub fn new(style: &'a mut StyleBuilder<'b>) -> Self

Trivially constructs a new StyleAdjuster.

pub fn set_bits(&mut self)

Compute a few common flags for both text and element's style.

pub fn adjust(
&mut self,
layout_parent_style: &ComputedValues,
flags: CascadeFlags
)

Adjusts the style to account for various fixups that don't fit naturally into the cascade.

When comparing to Gecko, this is similar to the work done by nsStyleContext::ApplyStyleFixups, plus some parts of nsStyleSet::GetContext.