Struct style::style_resolver::StyleResolverForElement

pub struct StyleResolverForElement<'a, 'ctx, 'le, E>
where
    E: TElement + MatchMethods + 'le,
    'ctx: 'a,
    'le: 'ctx,
{
    element: E,
    context: &'a mut StyleContext<'ctx, E>,
    rule_inclusion: RuleInclusion,
    pseudo_resolution: PseudoElementResolution,
    _marker: PhantomData<&'le E>,
}

A struct that takes care of resolving the style of a given element.

Fields
element: Econtext: &'a mut StyleContext<'ctx, E>rule_inclusion: RuleInclusionpseudo_resolution: PseudoElementResolution_marker: PhantomData<&'le E>
Implementations

impl<'a, 'ctx, 'le, E> StyleResolverForElement<'a, 'ctx, 'le, E>
where E: TElement + MatchMethods + 'le, 'ctx: 'a, 'le: 'ctx,

pub fn new( element: E, context: &'a mut StyleContext<'ctx, E>, rule_inclusion: RuleInclusion, pseudo_resolution: PseudoElementResolution, ) -> Self

Trivially construct a new StyleResolverForElement.

pub fn resolve_primary_style( &mut self, parent_style: Option<&ComputedValues>, layout_parent_style: Option<&ComputedValues>, include_starting_style: IncludeStartingStyle, ) -> PrimaryStyle

Resolve just the style of a given element.

fn cascade_primary_style( &mut self, inputs: CascadeInputs, parent_style: Option<&ComputedValues>, layout_parent_style: Option<&ComputedValues>, include_starting_style: IncludeStartingStyle, may_have_starting_style: bool, ) -> PrimaryStyle

pub fn resolve_style( &mut self, parent_style: Option<&ComputedValues>, layout_parent_style: Option<&ComputedValues>, ) -> ResolvedElementStyles

Resolve the style of a given element, and all its eager pseudo-elements.

pub fn resolve_style_with_default_parents(&mut self) -> ResolvedElementStyles

Resolve an element’s styles with the default inheritance parent/layout parents.

pub fn cascade_style_and_visited_with_default_parents( &mut self, inputs: CascadeInputs, ) -> ResolvedStyle

Cascade a set of rules, using the default parent for inheritance.

pub fn cascade_style_and_visited_for_pseudo_with_default_parents( &mut self, inputs: CascadeInputs, pseudo: &PseudoElement, primary_style: &PrimaryStyle, ) -> ResolvedStyle

Cascade a set of rules for pseudo element, using the default parent for inheritance.

fn cascade_style_and_visited( &mut self, inputs: CascadeInputs, parent_style: Option<&ComputedValues>, layout_parent_style: Option<&ComputedValues>, pseudo: Option<&PseudoElement>, ) -> ResolvedStyle

pub fn cascade_styles_with_default_parents( &mut self, inputs: ElementCascadeInputs, may_have_starting_style: bool, ) -> ResolvedElementStyles

Cascade the element and pseudo-element styles with the default parents.

fn resolve_pseudo_style( &mut self, pseudo: &PseudoElement, originating_element_style: &PrimaryStyle, layout_parent_style: Option<&ComputedValues>, ) -> Option<ResolvedStyle>

fn match_primary( &mut self, visited_handling: VisitedHandlingMode, include_starting_style: IncludeStartingStyle, ) -> MatchingResults

fn match_pseudo( &mut self, originating_element_style: &ComputedValues, pseudo_element: &PseudoElement, visited_handling: VisitedHandlingMode, ) -> Option<MatchingResults>

pub fn resolve_starting_style(&mut self) -> PrimaryStyle

Resolve the starting style.

pub fn after_change_style( &mut self, primary_style: &Arc<ComputedValues>, ) -> Option<Arc<ComputedValues>>

If there is no transition rule in the ComputedValues, it returns None.