Struct style::style_resolver::StyleResolverForElement

pub struct StyleResolverForElement<'a, 'ctx, 'le, E> where
    'ctx: 'a,
    'le: 'ctx,
    E: TElement + MatchMethods + 'le,  { /* fields omitted */ }

A struct that takes care of resolving the style of a given element.

Methods

impl<'a, 'ctx, 'le, E> StyleResolverForElement<'a, 'ctx, 'le, E> where
'ctx: 'a,
'le: 'ctx,
E: TElement + MatchMethods + 'le,

pub fn new(
element: E,
context: &'a mut StyleContext<'ctx, E>,
rule_inclusion: RuleInclusion,
pseudo_resolution: PseudoElementResolution
) -> Self

Trivially construct a new StyleResolverForElement.

pub fn resolve_primary_style(
&mut self,
parent_style: Option<&ComputedValues>,
layout_parent_style: Option<&ComputedValues>
) -> PrimaryStyle

Resolve just the style of a given element.

pub fn resolve_style(
&mut self,
parent_style: Option<&ComputedValues>,
layout_parent_style: Option<&ComputedValues>
) -> ResolvedElementStyles

Resolve the style of a given element, and all its eager pseudo-elements.

pub fn resolve_style_with_default_parents(&mut self) -> ResolvedElementStyles

Resolve an element's styles with the default inheritance parent/layout parents.

pub fn cascade_style_and_visited_with_default_parents(
&mut self,
inputs: CascadeInputs
) -> ResolvedStyle

Cascade a set of rules, using the default parent for inheritance.

pub fn cascade_styles_with_default_parents(
&mut self,
inputs: ElementCascadeInputs
) -> ResolvedElementStyles

Cascade the element and pseudo-element styles with the default parents.