Module style::stylesheet_set

A centralized set of stylesheets for a document.

Structs

AuthorStylesheetFlusher
A struct to flush an author style sheet collection.

AuthorStylesheetSet
The set of stylesheets effective for a given Shadow Root.

DocumentStylesheetFlusher
A struct to iterate over the different stylesheets to be flushed.

DocumentStylesheetSet
The set of stylesheets effective for a given document.

SheetCollection

SheetCollectionFlusher
A flusher struct for a given collection, that takes care of returning the appropriate stylesheets that need work.

StylesheetCollectionIterator
A iterator over the stylesheets of a list of entries in the StylesheetSet.

StylesheetIterator
An iterator over the flattened view of the stylesheet collections.

StylesheetSetEntry
Entry for a StylesheetSet.

Enums

DataValidity
The validity of the data in a given cascade origin.

SheetRebuildKind
The type of rebuild that we need to do for a given stylesheet.