Module style::stylesheet_set

A centralized set of stylesheets for a document.

Structs

PerOriginFlusher
A flusher struct for a given origin, that takes care of returning the appropriate stylesheets that need work.

StylesheetCollectionIterator
A iterator over the stylesheets of a list of entries in the StylesheetSet.

StylesheetFlusher
A struct to iterate over the different stylesheets to be flushed.

StylesheetIterator
An iterator over the flattened view of the stylesheet collections.

StylesheetSet
The set of stylesheets effective for a given document.

Enums

OriginValidity
The validity of the data in a given cascade origin.

SheetRebuildKind
The type of rebuild that we need to do for a given stylesheet.