Module style::stylesheets

Style sheets and their CSS rules.

Structs

CssRules
A list of CSS rules.

ImportRule
The @import at-rule.

KeyframesRule
A @keyframes rule.

MediaRule

MemoryHoleReporter
Error reporter which silently forgets errors

NamespaceRule

Namespaces
A set of namespaces applying to a given stylesheet.

StyleRule

Stylesheet
The structure servo uses to represent a stylesheet.

SupportsRule
An @supports rule

UserAgentStylesheets
This structure holds the user-agent and user stylesheets.

Enums

CssRule
A CSS rule.

CssRuleType

Origin
Each style rule has an origin, which determines where it enters the cascade.

RulesMutateError

SingleRuleParseError

State

Traits

StylesheetLoader
The stylesheet loader is the abstraction used to trigger network requests for @import rules.