Module style::stylesheets

Style sheets and their CSS rules.

Reexports

pub use self::counter_style_rule::CounterStyleRule;

pub use self::font_face_rule::FontFaceRule;

pub use self::font_feature_values_rule::FontFeatureValuesRule;

pub use self::import_rule::ImportRule;

pub use self::keyframes_rule::KeyframesRule;

pub use self::origin::Origin;

pub use self::origin::OriginSet;

pub use self::origin::OriginSetIterator;

pub use self::origin::PerOrigin;

pub use self::origin::PerOriginIter;

pub use self::supports_rule::SupportsRule;

pub use self::viewport_rule::ViewportRule;

Modules

font_feature_values_rule
The @font-feature-values at-rule.

import_rule
The @import at-rule.

keyframes_rule
Keyframes: https://drafts.csswg.org/css-animations/#keyframes

origin
CSS cascade origins.

supports_rule
@supports rules

viewport_rule
The @viewport at-rule and meta element.

Structs

AllRules
A filter that processes all the rules in a rule list.

CssRules
A list of CSS rules.

DocumentRule
A @-moz-document rule

DocumentStyleSheet
A simple wrapper over an Arc<Stylesheet>, with pointer comparison, and suitable for its use in a StylesheetSet.

EffectiveRules
A struct that represents the condition that a rule applies to the document.

MediaRule
An @media urle.

NamespaceRule
A @namespace rule.

Namespaces
A set of namespaces applying to a given stylesheet.

PageRule
A @page rule.

RulesIterator
An iterator over a list of rules.

StyleRule
A style rule, with selectors and declarations.

Stylesheet
The structure servo uses to represent a stylesheet.

StylesheetContents
The contents of a given stylesheet. This effectively maps to a StyleSheetInner in Gecko.

TopLevelRuleParser
The parser for the top-level rules in a stylesheet.

UserAgentStylesheets
This structure holds the user-agent and user stylesheets.

Enums

CssRule
A CSS rule.

CssRuleType

RulesMutateError

SingleRuleParseError

State
The current state of the parser.

Traits

CssRulesHelpers
A trait to implement helpers for Arc<Locked<CssRules>>.

NestedRuleIterationCondition
RulesIterator.

StylesheetInDocument
A trait to represent a given stylesheet in a document.

StylesheetLoader
The stylesheet loader is the abstraction used to trigger network requests for @import rules.

Type Definitions

UrlExtraData
Extra data that the backend may need to resolve url values.