style::stylesheets

Module font_feature_values_rule

The @font-feature-values at-rule.

Macros

font_feature_values_blocks 🔒

Structs

FFVDeclaration

A @font-feature-values block declaration. It is <ident>: <integer>+. This struct can take 3 value types.

FFVDeclarationsParser 🔒

@font-feature-values inside block parser. Parses a list of FFVDeclaration. (<ident>: <integer>+)

FontFeatureValuesRule

The @font-feature-values at-rule.

FontFeatureValuesRuleParser 🔒

Parser for FontFeatureValuesRule. Parses all blocks { } = @stylistic | @historical-forms | @styleset | @character-variant | @swash | @ornaments | @annotation

PairValues

A @font-feature-values block declaration value that keeps one or two values.

SingleValue

A @font-feature-values block declaration value that keeps one value.

VectorValues

A @font-feature-values block declaration value that keeps a list of values.

Enums

BlockType 🔒

Keeps the information about block type like @swash, @styleset etc.

Functions

parse_family_name_list

Parses a list of FamilyNames.

update_or_push 🔒

Updates with new value if same ident exists, otherwise pushes to the vector.