Module style::stylesheets::font_feature_values_rule

The @font-feature-values at-rule.

Structs

FFVDeclaration
A @font-feature-values block declaration. It is <ident>: <integer>+. This struct can take 3 value types. - SingleValue is to keep just one unsigned integer value. - PairValues is to keep one or two unsigned integer values. - VectorValues is to keep a list of unsigned integer values.

FontFeatureValuesRule
The @font-feature-values at-rule.

PairValues
A @font-feature-values block declaration value that keeps one or two values.

SingleValue
A @font-feature-values block declaration value that keeps one value.

VectorValues
A @font-feature-values block declaration value that keeps a list of values.

Functions

parse_family_name_list
Parses a list of FamilyNames.