Struct style::stylesheets::font_feature_values_rule::FontFeatureValuesRule

pub struct FontFeatureValuesRule {
    pub family_names: Vec<FamilyName>,
    pub swash: Vec<FFVDeclaration<SingleValue>>,
    pub stylistic: Vec<FFVDeclaration<SingleValue>>,
    pub ornaments: Vec<FFVDeclaration<SingleValue>>,
    pub annotation: Vec<FFVDeclaration<SingleValue>>,
    pub character_variant: Vec<FFVDeclaration<PairValues>>,
    pub styleset: Vec<FFVDeclaration<VectorValues>>,
    pub source_location: SourceLocation,
}

The @font-feature-values at-rule.

Fields
family_names: Vec<FamilyName>

Font family list for @font-feature-values rule. Family names cannot contain generic families. FamilyName also accepts only non-generic names.

swash: Vec<FFVDeclaration<SingleValue>>

A @swash blocksck. Specifies a feature name that will work with the swash() functional notation of font-variant-alternates.

stylistic: Vec<FFVDeclaration<SingleValue>>

A @stylistic block. Specifies a feature name that will work with the annotation() functional notation of font-variant-alternates.

ornaments: Vec<FFVDeclaration<SingleValue>>

A @ornaments block. Specifies a feature name that will work with the ornaments() ] functional notation of font-variant-alternates.

annotation: Vec<FFVDeclaration<SingleValue>>

A @annotation block. Specifies a feature name that will work with the stylistic() functional notation of font-variant-alternates.

character_variant: Vec<FFVDeclaration<PairValues>>

A @character-variant block. Specifies a feature name that will work with the styleset() functional notation of font-variant-alternates. The value can be a pair.

styleset: Vec<FFVDeclaration<VectorValues>>

A @styleset block. Specifies a feature name that will work with the character-variant() functional notation of font-variant-alternates. The value can be a list.

source_location: SourceLocation

The line and column of the rule's source code.

Methods

impl FontFeatureValuesRule

pub fn parse<R>(
context: &ParserContext,
error_context: &ParserErrorContext<R>,
input: &mut Parser,
family_names: Vec<FamilyName>,
location: SourceLocation
) -> FontFeatureValuesRule where
R: ParseErrorReporter,

Parses a FontFeatureValuesRule.

pub fn font_family_to_css<W>(&self, dest: &mut W) -> Result where
W: Write,

Prints font family names.

pub fn value_to_css<W>(&self, dest: &mut W) -> Result where
W: Write,

Prints inside of @font-feature-values block.

pub fn len(&self) -> usize

Returns length of all at-rules.