Module style::stylesheets::keyframes_rule

Keyframes: https://drafts.csswg.org/css-animations/#keyframes

Structs

Keyframe

A keyframe.

KeyframeDeclarationParser

KeyframeListParser

Parses a keyframes list, like: 0%, 50% { width: 50%; }

KeyframePercentage

A number from 0 to 1, indicating the percentage of the animation when this keyframe should run.

KeyframeSelector

A keyframes selector is a list of percentages or from/to symbols, which are converted at parse time to percentages.

KeyframesAnimation

This structure represents a list of animation steps computed from the list of keyframes, in order.

KeyframesRule

A @keyframes rule.

KeyframesStep

A single step from a keyframe animation.

Enums

KeyframesStepValue

A keyframes step value. This can be a synthetised keyframes animation, that is, one autogenerated from the current computed values, or a list of declarations to apply.

Functions

get_animated_properties

Get all the animated properties in a keyframes animation.

parse_keyframe_list

Parses a keyframe list from CSS input.