Struct style::stylesheets::rule_parser::TopLevelRuleParser

pub struct TopLevelRuleParser<'a> {
    pub shared_lock: &'a SharedRwLock,
    pub loader: Option<&'a dyn StylesheetLoader>,
    pub context: ParserContext<'a>,
    pub state: State,
    pub dom_error: Option<RulesMutateError>,
    pub namespaces: &'a mut Namespaces,
    pub insert_rule_context: Option<InsertRuleContext<'a>>,
    pub allow_import_rules: AllowImportRules,
}

The parser for the top-level rules in a stylesheet.

Fields
shared_lock: &'a SharedRwLock

A reference to the lock we need to use to create rules.

loader: Option<&'a dyn StylesheetLoader>

A reference to a stylesheet loader if applicable, for @import rules.

context: ParserContext<'a>

The top-level parser context.

This won’t contain any namespaces, and only nested parsers created with ParserContext::new_with_rule_type will.

state: State

The current stajkj/te of the parser.

dom_error: Option<RulesMutateError>

Whether we have tried to parse was invalid due to being in the wrong place (e.g. an @import rule was found while in the Body state). Reset to false when take_had_hierarchy_error is called.

namespaces: &'a mut Namespaces

The namespace map we use for parsing. Needs to start as Some(), and will be taken out after parsing namespace rules, and that reference will be moved to ParserContext.

insert_rule_context: Option<InsertRuleContext<'a>>

The info we need insert a rule in a list.

allow_import_rules: AllowImportRules

Whether @import rules will be allowed.

Implementations

impl<'b> TopLevelRuleParser<'b>

fn nested<'a: 'b>(&'a self) -> NestedRuleParser<'a, 'b>

pub fn state(&self) -> State

Returns the current state of the parser.

fn check_state(&mut self, new_state: State) -> bool

Checks whether we can parse a rule that would transition us to new_state.

This is usually a simple branch, but we may need more bookkeeping if doing insertRule from CSSOM.