style::stylesheets

Module rules_iterator

An iterator over a list of rules.

Structs

AllRules

A filter that processes all the rules in a rule list.

EffectiveRules

A struct that represents the condition that a rule applies to the document.

RulesIterator

An iterator over a list of rules.

Traits

NestedRuleIterationCondition

RulesIterator.

Type Aliases

EffectiveRulesIterator

An iterator over all the effective rules of a stylesheet.