Module style::stylesheets::rules_iterator

An iterator over a list of rules.

Structs

AllRules
A filter that processes all the rules in a rule list.

EffectiveRules
A struct that represents the condition that a rule applies to the document.

RulesIterator
An iterator over a list of rules.

Traits

NestedRuleIterationCondition
RulesIterator.

Type Definitions

EffectiveRulesIterator
An iterator over all the effective rules of a stylesheet.