Trait style::stylesheets::rules_iterator::NestedRuleIterationCondition

pub trait NestedRuleIterationCondition {
    // Required methods
    fn process_import(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        rule: &ImportRule,
    ) -> bool;
    fn process_media(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        rule: &MediaRule,
    ) -> bool;
    fn process_document(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        rule: &DocumentRule,
    ) -> bool;
    fn process_supports(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        rule: &SupportsRule,
    ) -> bool;
}

RulesIterator.

Required Methods

fn process_import( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, rule: &ImportRule, ) -> bool

Whether we should process the nested rules in a given @import rule.

fn process_media( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, rule: &MediaRule, ) -> bool

Whether we should process the nested rules in a given @media rule.

fn process_document( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, rule: &DocumentRule, ) -> bool

Whether we should process the nested rules in a given @-moz-document rule.

fn process_supports( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, rule: &SupportsRule, ) -> bool

Whether we should process the nested rules in a given @supports rule.

Object Safety

This trait is not object safe.