Struct style::stylesheets::Stylesheet

pub struct Stylesheet<Impl: SelectorImpl> {
    pub rules: Vec<CSSRule<Impl>>,
    pub media: Option<MediaQueryList>,
    pub origin: Origin,
    pub dirty_on_viewport_size_change: bool,
}

Fields
rules: Vec<CSSRule<Impl>>

List of rules in the order they were found (important for cascading order)

media: Option<MediaQueryList>

List of media associated with the Stylesheet, if any.

origin: Origin dirty_on_viewport_size_change: bool
Methods

impl<Impl: SelectorImpl> Stylesheet<Impl>

fn from_bytes_iter<I: Iterator<Item=Vec<u8>>>(input: I, base_url: Url, protocol_encoding_label: Option<&str>, environment_encoding: Option<EncodingRef>, origin: Origin, error_reporter: Box<ParseErrorReporter + Send>, extra_data: ParserContextExtraData) -> Stylesheet<Impl>

fn from_bytes(bytes: &[u8], base_url: Url, protocol_encoding_label: Option<&str>, environment_encoding: Option<EncodingRef>, origin: Origin, error_reporter: Box<ParseErrorReporter + Send>, extra_data: ParserContextExtraData) -> Stylesheet<Impl>

fn from_str(css: &str, base_url: Url, origin: Origin, error_reporter: Box<ParseErrorReporter + Send>, extra_data: ParserContextExtraData) -> Stylesheet<Impl>

fn set_media(&mut self, media: Option<MediaQueryList>)

Set the MediaQueryList associated with the style-sheet.

fn is_effective_for_device(&self, device: &Device) -> bool

Returns whether the style-sheet applies for the current device depending on the associated MediaQueryList.

Always true if no associated MediaQueryList exists.

fn rules(&self) -> Rules<Impl>

Return an iterator over all the rules within the style-sheet.

fn effective_rules<'a>(&'a self, device: &'a Device) -> Rules<'a, Impl>

Return an iterator over the effective rules within the style-sheet, as according to the supplied Device.

If a condition does not hold, its associated conditional group rule and nested rules will be skipped. Use rules if all rules need to be examined.