Struct style::stylesheets::Stylesheet

pub struct Stylesheet {
    pub rules: Arc<RwLock<CssRules>>,
    pub media: Arc<RwLock<MediaList>>,
    pub origin: Origin,
    pub base_url: ServoUrl,
    pub namespaces: RwLock<Namespaces>,
    pub dirty_on_viewport_size_change: AtomicBool,
    pub disabled: AtomicBool,
}

The structure servo uses to represent a stylesheet.

Fields
rules: Arc<RwLock<CssRules>>

List of rules in the order they were found (important for cascading order)

media: Arc<RwLock<MediaList>>

List of media associated with the Stylesheet.

origin: Origin

The origin of this stylesheet.

base_url: ServoUrl

The base url this stylesheet should use.

namespaces: RwLock<Namespaces>

The namespaces that apply to this stylesheet.

dirty_on_viewport_size_change: AtomicBool

Whether this stylesheet would be dirty when the viewport size changes.

disabled: AtomicBool

Whether this stylesheet should be disabled.

Methods

impl Stylesheet

fn from_bytes(bytes: &[u8],
base_url: ServoUrl,
protocol_encoding_label: Option<&str>,
environment_encoding: Option<EncodingRef>,
origin: Origin,
media: MediaList,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)
-> Stylesheet

Parse a stylesheet from a set of bytes, potentially received over the network.

Takes care of decoding the network bytes and forwards the resulting string to Stylesheet::from_str.

fn update_from_bytes(existing: &Stylesheet,
bytes: &[u8],
protocol_encoding_label: Option<&str>,
environment_encoding: Option<EncodingRef>,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)

Updates an empty stylesheet with a set of bytes that reached over the network.

fn update_from_str(existing: &Stylesheet,
css: &str,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)

Updates an empty stylesheet from a given string of text.

fn from_str(css: &str,
base_url: ServoUrl,
origin: Origin,
media: MediaList,
stylesheet_loader: Option<&StylesheetLoader>,
error_reporter: Box<ParseErrorReporter + Send>,
extra_data: ParserContextExtraData)
-> Stylesheet

Creates an empty stylesheet and parses it with a given base url, origin and media.

Effectively creates a new stylesheet and forwards the hard work to Stylesheet::update_from_str.

fn dirty_on_viewport_size_change(&self) -> bool

Whether this stylesheet can be dirty on viewport size change.

fn inserted_has_viewport_percentages(&self, has_viewport_percentages: bool)

When CSSOM inserts a rule or declaration into this stylesheet, it needs to call this method with the return value of cssparser::Parser::seen_viewport_percentages.

FIXME: actually make these calls

Note: when removing a rule or declaration that contains a viewport percentage, to keep the flag accurate we’d need to iterator through the rest of the stylesheet to check for other such values.

Instead, we conservatively assume there might be some. Restyling will some some more work than necessary, but give correct results.

fn is_effective_for_device(&self, device: &Device) -> bool

Returns whether the style-sheet applies for the current device depending on the associated MediaList.

Always true if no associated MediaList exists.

fn effective_rules<F>(&self, device: &Device, f: F) where F: FnMut(&CssRule)

Return an iterator over the effective rules within the style-sheet, as according to the supplied Device.

If a condition does not hold, its associated conditional group rule and nested rules will be skipped. Use rules if all rules need to be examined.

fn disabled(&self) -> bool

Returns whether the stylesheet has been explicitly disabled through the CSSOM.

fn set_disabled(&self, disabled: bool) -> bool

Records that the stylesheet has been explicitly disabled through the CSSOM.

Returns whether the the call resulted in a change in disabled state.

Disabled stylesheets remain in the document, but their rules are not added to the Stylist.

impl Stylesheet

fn effective_style_rules<F>(&self, device: &Device, f: F) where F: FnMut(&StyleRule)

fn effective_media_rules<F>(&self, device: &Device, f: F) where F: FnMut(&MediaRule)

fn effective_font_face_rules<F>(&self, device: &Device, f: F) where F: FnMut(&FontFaceRule)

fn effective_viewport_rules<F>(&self, device: &Device, f: F) where F: FnMut(&ViewportRule)

fn effective_keyframes_rules<F>(&self, device: &Device, f: F) where F: FnMut(&KeyframesRule)

fn effective_supports_rules<F>(&self, device: &Device, f: F) where F: FnMut(&SupportsRule)