style::stylesheets

Struct StylesheetContents

pub struct StylesheetContents {
    pub rules: Arc<Locked<CssRules>>,
    pub origin: Origin,
    pub url_data: UrlExtraData,
    pub namespaces: Namespaces,
    pub quirks_mode: QuirksMode,
    pub source_map_url: Option<String>,
    pub source_url: Option<String>,
    pub use_counters: UseCounters,
    _forbid_construction: (),
}

The contents of a given stylesheet. This effectively maps to a StyleSheetInner in Gecko.

Fields
rules: Arc<Locked<CssRules>>

List of rules in the order they were found (important for cascading order)

origin: Origin

The origin of this stylesheet.

url_data: UrlExtraData

The url data this stylesheet should use.

namespaces: Namespaces

The namespaces that apply to this stylesheet.

quirks_mode: QuirksMode

The quirks mode of this stylesheet.

source_map_url: Option<String>

This stylesheet’s source map URL.

source_url: Option<String>

This stylesheet’s source URL.

use_counters: UseCounters

The use counters of the original stylesheet.

_forbid_construction: ()

We don’t want to allow construction outside of this file, to guarantee that all contents are created with Arc<>.

Implementations

impl StylesheetContents

pub fn from_str( css: &str, url_data: UrlExtraData, origin: Origin, shared_lock: &SharedRwLock, stylesheet_loader: Option<&dyn StylesheetLoader>, error_reporter: Option<&dyn ParseErrorReporter>, quirks_mode: QuirksMode, allow_import_rules: AllowImportRules, sanitization_data: Option<&mut SanitizationData>, ) -> Arc<Self>

Parse a given CSS string, with a given url-data, origin, and quirks mode.

pub fn from_shared_data( rules: Arc<Locked<CssRules>>, origin: Origin, url_data: UrlExtraData, quirks_mode: QuirksMode, ) -> Arc<Self>

Creates a new StylesheetContents with the specified pre-parsed rules, origin, URL data, and quirks mode.

Since the rules have already been parsed, and the intention is that this function is used for read only User Agent style sheets, an empty namespace map is used, and the source map and source URLs are set to None.

An empty namespace map should be fine, as it is only used for parsing, not serialization of existing selectors. Since UA sheets are read only, we should never need the namespace map.

pub fn rules<'a, 'b: 'a>( &'a self, guard: &'b SharedRwLockReadGuard<'_>, ) -> &'a [CssRule]

Returns a reference to the list of rules.

pub fn iter_rules<'a, 'b, C, CMM>( &'a self, device: &'a Device, custom_media: CMM, guard: &'a SharedRwLockReadGuard<'b>, ) -> RulesIterator<'a, 'b, C, CMM>
where C: NestedRuleIterationCondition, CMM: Deref<Target = CustomMediaMap>,

Return an iterator using the condition C.

pub fn effective_rules<'a, 'b, CMM: Deref<Target = CustomMediaMap>>( &'a self, device: &'a Device, custom_media: CMM, guard: &'a SharedRwLockReadGuard<'b>, ) -> EffectiveRulesIterator<'a, 'b, CMM>

Return an iterator over the effective rules within the style-sheet, as according to the supplied Device.

pub fn deep_clone( &self, lock: &SharedRwLock, url_data: Option<&UrlExtraData>, guard: &SharedRwLockReadGuard<'_>, ) -> Arc<Self>

Perform a deep clone, of this stylesheet, with an explicit URL data if needed.