Struct style::stylesheets::TopLevelRuleParser

pub struct TopLevelRuleParser<'a, R: 'a> {
    pub stylesheet_origin: Origin,
    pub shared_lock: &'a SharedRwLock,
    pub loader: Option<&'a StylesheetLoader>,
    pub context: ParserContext<'a>,
    pub error_context: ParserErrorContext<'a, R>,
    pub state: State,
    pub had_hierarchy_error: bool,
    pub namespaces: &'a mut Namespaces,
}

The parser for the top-level rules in a stylesheet.

Fields
stylesheet_origin: Origin

The origin of the stylesheet we're parsing.

shared_lock: &'a SharedRwLock

A reference to the lock we need to use to create rules.

loader: Option<&'a StylesheetLoader>

A reference to a stylesheet loader if applicable, for @import rules.

context: ParserContext<'a>

The top-level parser context.

This won't contain any namespaces, and only nested parsers created with ParserContext::new_with_rule_type will.

error_context: ParserErrorContext<'a, R>

The context required for reporting parse errors.

state: State

The current state of the parser.

had_hierarchy_error: bool

Whether we have tried to parse was invalid due to being in the wrong place (e.g. an @import rule was found while in the Body state). Reset to false when take_had_hierarchy_error is called.

namespaces: &'a mut Namespaces

The namespace map we use for parsing. Needs to start as Some(), and will be taken out after parsing namespace rules, and that reference will be moved to ParserContext.

Methods

impl<'b, R> TopLevelRuleParser<'b, R>

pub fn state(&self) -> State

Returns the current state of the parser.

pub fn take_had_hierarchy_error(&mut self) -> bool

Returns whether we previously tried to parse a rule that was invalid due to being in the wrong place (e.g. an @import rule was found after a regular style rule). The state of this flag is reset when this function is called.