Module style::stylesheets::stylesheet

Structs

DocumentStyleSheet

A simple wrapper over an Arc<Stylesheet>, with pointer comparison, and suitable for its use in a StylesheetSet.

Namespaces

A set of namespaces applying to a given stylesheet.

SanitizationData

A struct to hold the data relevant to style sheet sanitization.

Stylesheet

The structure servo uses to represent a stylesheet.

StylesheetContents

The contents of a given stylesheet. This effectively maps to a StyleSheetInner in Gecko.

UserAgentStylesheets

This structure holds the user-agent and user stylesheets.

Enums

AllowImportRules

Whether @import rules are allowed.

SanitizationKind

The kind of sanitization to use when parsing a stylesheet.

Traits

StylesheetInDocument

A trait to represent a given stylesheet in a document.