Trait style::stylesheets::stylesheet::StylesheetInDocument

pub trait StylesheetInDocument: Debug {

    fn origin(&self, guard: &SharedRwLockReadGuard<'_>) -> Origin;

    fn quirks_mode(&self, guard: &SharedRwLockReadGuard<'_>) -> QuirksMode;

    fn enabled(&self) -> bool;

    fn media<'a>(
        &'a self, 
        guard: &'a SharedRwLockReadGuard<'_>
    ) -> Option<&'a MediaList>;

    fn rules<'a, 'b: 'a>(
        &'a self, 
        guard: &'b SharedRwLockReadGuard<'_>
    ) -> &'a [CssRule];

    fn iter_rules<'a, 'b, C>(
        &'a self, 
        device: &'a Device, 
        guard: &'a SharedRwLockReadGuard<'b>
    ) -> RulesIterator<'a, 'b, C>
    where
        C: NestedRuleIterationCondition,
    { ... }

    fn is_effective_for_device(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>
    ) -> bool { ... }

    fn effective_rules<'a, 'b>(
        &'a self, 
        device: &'a Device, 
        guard: &'a SharedRwLockReadGuard<'b>
    ) -> EffectiveRulesIterator<'a, 'b> { ... }

    fn effective_style_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&StyleRule),
    { ... }

    fn effective_media_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&MediaRule),
    { ... }

    fn effective_font_face_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&FontFaceRule),
    { ... }

    fn effective_font_face_feature_values_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&FontFeatureValuesRule),
    { ... }

    fn effective_counter_style_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&CounterStyleRule),
    { ... }

    fn effective_viewport_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&ViewportRule),
    { ... }

    fn effective_keyframes_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&KeyframesRule),
    { ... }

    fn effective_supports_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&SupportsRule),
    { ... }

    fn effective_page_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&PageRule),
    { ... }

    fn effective_document_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard<'_>, 
        f: F
    )
    where
        F: FnMut(&DocumentRule),
    { ... }

}

A trait to represent a given stylesheet in a document.

Required methods

fn origin(&self, guard: &SharedRwLockReadGuard<'_>) -> Origin

Get the stylesheet origin.

fn quirks_mode(&self, guard: &SharedRwLockReadGuard<'_>) -> QuirksMode

Get the stylesheet quirks mode.

fn enabled(&self) -> bool

Get whether this stylesheet is enabled.

fn media<'a>(
&'a self,
guard: &'a SharedRwLockReadGuard<'_>
) -> Option<&'a MediaList>

Get the media associated with this stylesheet.

fn rules<'a, 'b: 'a>(
&'a self,
guard: &'b SharedRwLockReadGuard<'_>
) -> &'a [CssRule]

Returns a reference to the list of rules in this stylesheet.

Provided methods

fn iter_rules<'a, 'b, C>(
&'a self,
device: &'a Device,
guard: &'a SharedRwLockReadGuard<'b>
) -> RulesIterator<'a, 'b, C> where
C: NestedRuleIterationCondition,

Return an iterator using the condition C.

fn is_effective_for_device(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>
) -> bool

Returns whether the style-sheet applies for the current device.

fn effective_rules<'a, 'b>(
&'a self,
device: &'a Device,
guard: &'a SharedRwLockReadGuard<'b>
) -> EffectiveRulesIterator<'a, 'b>

Return an iterator over the effective rules within the style-sheet, as according to the supplied Device.

fn effective_style_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&StyleRule),

fn effective_media_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&MediaRule),

fn effective_font_face_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&FontFaceRule),

fn effective_font_face_feature_values_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&FontFeatureValuesRule),

fn effective_counter_style_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&CounterStyleRule),

fn effective_viewport_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&ViewportRule),

fn effective_keyframes_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&KeyframesRule),

fn effective_supports_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&SupportsRule),

fn effective_page_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&PageRule),

fn effective_document_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard<'_>,
f: F
) where
F: FnMut(&DocumentRule),