style::stylesheets

Trait NestedRuleIterationCondition

pub trait NestedRuleIterationCondition {
    // Required methods
    fn process_import(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        custom_media_map: &CustomMediaMap,
        rule: &ImportRule,
    ) -> bool;
    fn process_media(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        custom_media_map: &CustomMediaMap,
        rule: &MediaRule,
    ) -> bool;
    fn process_document(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        rule: &DocumentRule,
    ) -> bool;
    fn process_supports(
        guard: &SharedRwLockReadGuard<'_>,
        device: &Device,
        quirks_mode: QuirksMode,
        rule: &SupportsRule,
    ) -> bool;
}

RulesIterator.

Required Methods

fn process_import( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, custom_media_map: &CustomMediaMap, rule: &ImportRule, ) -> bool

Whether we should process the nested rules in a given @import rule.

fn process_media( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, custom_media_map: &CustomMediaMap, rule: &MediaRule, ) -> bool

Whether we should process the nested rules in a given @media rule.

fn process_document( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, rule: &DocumentRule, ) -> bool

Whether we should process the nested rules in a given @-moz-document rule.

fn process_supports( guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, rule: &SupportsRule, ) -> bool

Whether we should process the nested rules in a given @supports rule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.