Trait style::stylesheets::StylesheetInDocument

pub trait StylesheetInDocument: Debug {
    // Required methods
    fn enabled(&self) -> bool;
    fn media<'a>(
        &'a self,
        guard: &'a SharedRwLockReadGuard<'_>,
    ) -> Option<&'a MediaList>;
    fn contents(&self) -> &StylesheetContents;
    fn implicit_scope_root(&self) -> Option<ImplicitScopeRoot>;

    // Provided methods
    fn rules<'a, 'b: 'a>(
        &'a self,
        guard: &'b SharedRwLockReadGuard<'_>,
    ) -> &'a [CssRule] { ... }
    fn iter_rules<'a, 'b, C>(
        &'a self,
        device: &'a Device,
        guard: &'a SharedRwLockReadGuard<'b>,
    ) -> RulesIterator<'a, 'b, C> 
       where C: NestedRuleIterationCondition { ... }
    fn is_effective_for_device(
        &self,
        device: &Device,
        guard: &SharedRwLockReadGuard<'_>,
    ) -> bool { ... }
    fn effective_rules<'a, 'b>(
        &'a self,
        device: &'a Device,
        guard: &'a SharedRwLockReadGuard<'b>,
    ) -> EffectiveRulesIterator<'a, 'b> { ... }
}

A trait to represent a given stylesheet in a document.

Required Methods

fn enabled(&self) -> bool

Get whether this stylesheet is enabled.

fn media<'a>( &'a self, guard: &'a SharedRwLockReadGuard<'_>, ) -> Option<&'a MediaList>

Get the media associated with this stylesheet.

fn contents(&self) -> &StylesheetContents

Returns a reference to the contents of the stylesheet.

fn implicit_scope_root(&self) -> Option<ImplicitScopeRoot>

Return the implicit scope root for this stylesheet, if one exists.

Provided Methods

fn rules<'a, 'b: 'a>( &'a self, guard: &'b SharedRwLockReadGuard<'_>, ) -> &'a [CssRule]

Returns a reference to the list of rules in this stylesheet.

fn iter_rules<'a, 'b, C>( &'a self, device: &'a Device, guard: &'a SharedRwLockReadGuard<'b>, ) -> RulesIterator<'a, 'b, C>
where C: NestedRuleIterationCondition,

Return an iterator using the condition C.

fn is_effective_for_device( &self, device: &Device, guard: &SharedRwLockReadGuard<'_>, ) -> bool

Returns whether the style-sheet applies for the current device.

fn effective_rules<'a, 'b>( &'a self, device: &'a Device, guard: &'a SharedRwLockReadGuard<'b>, ) -> EffectiveRulesIterator<'a, 'b>

Return an iterator over the effective rules within the style-sheet, as according to the supplied Device.

Object Safety

This trait is not object safe.