Trait style::stylesheets::StylesheetInDocument

pub trait StylesheetInDocument {
    fn contents(&self, guard: &SharedRwLockReadGuard) -> &StylesheetContents;

    fn media<'a>(
        &'a self, 
        guard: &'a SharedRwLockReadGuard
    ) -> Option<&'a MediaList>;

    fn enabled(&self) -> bool;

    fn origin(&self, guard: &SharedRwLockReadGuard) -> Origin { ... }

    fn quirks_mode(&self, guard: &SharedRwLockReadGuard) -> QuirksMode { ... }

    fn is_effective_for_device(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard
    ) -> bool { ... }

    fn iter_rules<'a, 'b, C>(
        &'a self, 
        device: &'a Device, 
        guard: &'a SharedRwLockReadGuard<'b>
    ) -> RulesIterator<'a, 'b, C>
    where
        C: NestedRuleIterationCondition,
    { ... }

    fn effective_rules<'a, 'b>(
        &'a self, 
        device: &'a Device, 
        guard: &'a SharedRwLockReadGuard<'b>
    ) -> RulesIterator<'a, 'b, EffectiveRules> { ... }

    fn effective_style_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&StyleRule),
    { ... }

    fn effective_media_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&MediaRule),
    { ... }

    fn effective_font_face_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&FontFaceRule),
    { ... }

    fn effective_font_face_feature_values_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&FontFeatureValuesRule),
    { ... }

    fn effective_counter_style_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&CounterStyleRule),
    { ... }

    fn effective_viewport_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&ViewportRule),
    { ... }

    fn effective_keyframes_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&KeyframesRule),
    { ... }

    fn effective_supports_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&SupportsRule),
    { ... }

    fn effective_page_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&PageRule),
    { ... }

    fn effective_document_rules<F>(
        &self, 
        device: &Device, 
        guard: &SharedRwLockReadGuard, 
        f: F
    )
    where
        F: FnMut(&DocumentRule),
    { ... }
}

A trait to represent a given stylesheet in a document.

Required Methods

fn contents(&self, guard: &SharedRwLockReadGuard) -> &StylesheetContents

Get the contents of this stylesheet.

fn media<'a>(
&'a self,
guard: &'a SharedRwLockReadGuard
) -> Option<&'a MediaList>

Get the media associated with this stylesheet.

fn enabled(&self) -> bool

Get whether this stylesheet is enabled.

Provided Methods

fn origin(&self, guard: &SharedRwLockReadGuard) -> Origin

Get the stylesheet origin.

fn quirks_mode(&self, guard: &SharedRwLockReadGuard) -> QuirksMode

Get the stylesheet quirks mode.

fn is_effective_for_device(
&self,
device: &Device,
guard: &SharedRwLockReadGuard
) -> bool

Returns whether the style-sheet applies for the current device.

fn iter_rules<'a, 'b, C>(
&'a self,
device: &'a Device,
guard: &'a SharedRwLockReadGuard<'b>
) -> RulesIterator<'a, 'b, C> where
C: NestedRuleIterationCondition,

Return an iterator using the condition C.

fn effective_rules<'a, 'b>(
&'a self,
device: &'a Device,
guard: &'a SharedRwLockReadGuard<'b>
) -> RulesIterator<'a, 'b, EffectiveRules>

Return an iterator over the effective rules within the style-sheet, as according to the supplied Device.

fn effective_style_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&StyleRule),

fn effective_media_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&MediaRule),

fn effective_font_face_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&FontFaceRule),

fn effective_font_face_feature_values_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&FontFeatureValuesRule),

fn effective_counter_style_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&CounterStyleRule),

fn effective_viewport_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&ViewportRule),

fn effective_keyframes_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&KeyframesRule),

fn effective_supports_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&SupportsRule),

fn effective_page_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&PageRule),

fn effective_document_rules<F>(
&self,
device: &Device,
guard: &SharedRwLockReadGuard,
f: F
) where
F: FnMut(&DocumentRule),