Module style::stylist

Selector matching.

Structs

ApplicableDeclarationBlock
A property declaration together with its precedence among rules of equal specificity so that we can sort them.

Rule
A rule, that wraps a style rule, but represents a single selector of the rule.

SelectorMap
Map element data to Rules whose last simple selector starts with them.

Stylist
This structure holds all the selectors and device characteristics for a given document. The selectors are converted into Rules (defined in rust-selectors), and introduced in a SelectorMap depending on the pseudo-element (see PerPseudoElementSelectorMap), and stylesheet origin (see the fields of PerPseudoElementSelectorMap).

Type Definitions

FnvHashMap
A HashMap using a default FNV hasher.