Module style::stylist

Selector matching.

Structs

ExtraStyleData
This struct holds data which users of Stylist may want to extract from stylesheets which can be done at the same time as updating.

ExtraStyleDataIterator
An iterator over the different ExtraStyleData.

Rule
A rule, that wraps a style rule, but represents a single selector of the rule.

Stylist
This structure holds all the selectors and device characteristics for a given document. The selectors are converted into Rules and sorted into SelectorMaps keyed off stylesheet origin and pseudo-element (see CascadeData).

Enums

RuleInclusion
What cascade levels to include when styling elements.

Functions

needs_revalidation_for_testing
A function to be able to test the revalidation stuff.

Type Definitions

StylistSheet
The type of the stylesheets that the stylist contains.