style::stylist

Struct CascadeData

pub struct CascadeData {

    normal_rules: GenericElementAndPseudoRules<SelectorMap<Rule>>,
    featureless_host_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>,
    slotted_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>,
    part_rules: Option<Box<GenericElementAndPseudoRules<PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>>>,
    invalidation_map: InvalidationMap,
    relative_selector_invalidation_map: InvalidationMap,
    additional_relative_selector_invalidation_map: AdditionalRelativeSelectorInvalidationMap,
    attribute_dependencies: PrecomputedHashSet<LocalName>,
    nth_of_class_dependencies: PrecomputedHashSet<Atom>,
    nth_of_attribute_dependencies: PrecomputedHashSet<LocalName>,
    nth_of_custom_state_dependencies: PrecomputedHashSet<AtomIdent>,
    state_dependencies: ElementState,
    nth_of_state_dependencies: ElementState,
    document_state_dependencies: DocumentState,
    mapped_ids: PrecomputedHashSet<Atom>,
    nth_of_mapped_ids: PrecomputedHashSet<Atom>,
    selectors_for_cache_revalidation: SelectorMap<RevalidationSelectorAndHashes>,
    animations: LayerOrderedMap<KeyframesAnimation>,
    custom_property_registrations: LayerOrderedMap<Arc<PropertyRegistration>>,
    custom_media: CustomMediaMap,
    layer_id: FxHashMap<LayerName, LayerId>,
    layers: SmallVec<[CascadeLayer; 1]>,
    container_conditions: SmallVec<[ContainerConditionReference; 1]>,
    scope_conditions: SmallVec<[ScopeConditionReference; 1]>,
    scope_subject_map: ScopeSubjectMap,
    effective_media_query_results: EffectiveMediaQueryResults,
    extra_data: ExtraStyleData,
    rules_source_order: u32,
    num_selectors: usize,
    num_declarations: usize,

}

Data resulting from performing the CSS cascade that is specific to a given origin.

FIXME(emilio): Consider renaming and splitting in CascadeData and InvalidationData? That’d make clear_cascade_data() clearer.

Fields
normal_rules: GenericElementAndPseudoRules<SelectorMap<Rule>>

The data coming from normal style rules that apply to elements at this cascade level.

featureless_host_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>

The :host pseudo rules that are the rightmost selector (without accounting for pseudo-elements), or :scope rules that may match the featureless host.

slotted_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>

The data coming from ::slotted() pseudo-element rules.

We need to store them separately because an element needs to match ::slotted() pseudo-element rules in different shadow roots.

In particular, we need to go through all the style data in all the containing style scopes starting from the closest assigned slot.

part_rules: Option<Box<GenericElementAndPseudoRules<PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>>>

The data coming from ::part() pseudo-element rules.

We need to store them separately because an element needs to match ::part() pseudo-element rules in different shadow roots.

invalidation_map: InvalidationMap

The invalidation map for these rules.

relative_selector_invalidation_map: InvalidationMap

The relative selector equivalent of the invalidation map.

additional_relative_selector_invalidation_map: AdditionalRelativeSelectorInvalidationMapattribute_dependencies: PrecomputedHashSet<LocalName>

The attribute local names that appear in attribute selectors. Used to avoid taking element snapshots when an irrelevant attribute changes. (We don’t bother storing the namespace, since namespaced attributes are rare.)

nth_of_class_dependencies: PrecomputedHashSet<Atom>

The classes that appear in the selector list of :nth-child(… of ). Used to avoid restyling siblings of an element when an irrelevant class changes.

nth_of_attribute_dependencies: PrecomputedHashSet<LocalName>

The attributes that appear in the selector list of :nth-child(… of ). Used to avoid restyling siblings of an element when an irrelevant attribute changes.

nth_of_custom_state_dependencies: PrecomputedHashSet<AtomIdent>

The custom states that appear in the selector list of :nth-child(… of ). Used to avoid restyling siblings of an element when an irrelevant custom state changes.

state_dependencies: ElementState

The element state bits that are relied on by selectors. Like attribute_dependencies, this is used to avoid taking element snapshots when an irrelevant element state bit changes.

nth_of_state_dependencies: ElementState

The element state bits that are relied on by selectors that appear in the selector list of :nth-child(… of ).

document_state_dependencies: DocumentState

The document state bits that are relied on by selectors. This is used to tell whether we need to restyle the entire document when a document state bit changes.

mapped_ids: PrecomputedHashSet<Atom>

The ids that appear in the rightmost complex selector of selectors (and hence in our selector maps). Used to determine when sharing styles is safe: we disallow style sharing for elements whose id matches this filter, and hence might be in one of our selector maps.

nth_of_mapped_ids: PrecomputedHashSet<Atom>

The IDs that appear in the selector list of :nth-child(… of ). Used to avoid restyling siblings of an element when an irrelevant ID changes.

selectors_for_cache_revalidation: SelectorMap<RevalidationSelectorAndHashes>

Selectors that require explicit cache revalidation (i.e. which depend on state that is not otherwise visible to the cache, like attributes or tree-structural state like child index and pseudos).

animations: LayerOrderedMap<KeyframesAnimation>

A map with all the animations at this CascadeData’s origin, indexed by name.

custom_property_registrations: LayerOrderedMap<Arc<PropertyRegistration>>

A map with all the layer-ordered registrations from style at this CascadeData’s origin, indexed by name.

custom_media: CustomMediaMap

Custom media query registrations.

layer_id: FxHashMap<LayerName, LayerId>

A map from cascade layer name to layer order.

layers: SmallVec<[CascadeLayer; 1]>

The list of cascade layers, indexed by their layer id.

container_conditions: SmallVec<[ContainerConditionReference; 1]>

The list of container conditions, indexed by their id.

scope_conditions: SmallVec<[ScopeConditionReference; 1]>

The list of scope conditions, indexed by their id.

scope_subject_map: ScopeSubjectMap

Map of unique selectors on scope start selectors’ subjects.

effective_media_query_results: EffectiveMediaQueryResults

Effective media query results cached from the last rebuild.

extra_data: ExtraStyleData

Extra data, like different kinds of rules, etc.

rules_source_order: u32

A monotonically increasing counter to represent the order on which a style rule appears in a stylesheet, needed to sort them by source order.

num_selectors: usize

The total number of selectors.

num_declarations: usize

The total number of declarations.

Implementations

impl CascadeData

pub fn new() -> Self

Creates an empty CascadeData.

pub fn rebuild<'a, S>( &mut self, device: &Device, quirks_mode: QuirksMode, collection: SheetCollectionFlusher<'_, S>, guard: &SharedRwLockReadGuard<'_>, ) -> Result<(), AllocErr>
where S: StylesheetInDocument + PartialEq + 'static,

Rebuild the cascade data from a given SheetCollection, incrementally if possible.

pub fn custom_media_map(&self) -> &CustomMediaMap

Returns the custom media query map.

pub fn invalidation_map(&self) -> &InvalidationMap

Returns the invalidation map.

pub fn relative_selector_invalidation_map(&self) -> &InvalidationMap

Returns the relative selector invalidation map.

pub fn relative_invalidation_map_attributes( &self, ) -> &AdditionalRelativeSelectorInvalidationMap

Returns the relative selector invalidation map data.

pub fn has_state_dependency(&self, state: ElementState) -> bool

Returns whether the given ElementState bit is relied upon by a selector of some rule.

pub fn has_nth_of_custom_state_dependency(&self, state: &AtomIdent) -> bool

Returns whether the given Custom State is relied upon by a selector of some rule in the selector list of :nth-child(… of ).

pub fn has_nth_of_state_dependency(&self, state: ElementState) -> bool

Returns whether the given ElementState bit is relied upon by a selector of some rule in the selector list of :nth-child(… of ).

pub fn might_have_attribute_dependency(&self, local_name: &LocalName) -> bool

Returns whether the given attribute might appear in an attribute selector of some rule.

pub fn might_have_nth_of_id_dependency(&self, id: &Atom) -> bool

Returns whether the given ID might appear in an ID selector in the selector list of :nth-child(… of ).

pub fn might_have_nth_of_class_dependency(&self, class: &Atom) -> bool

Returns whether the given class might appear in a class selector in the selector list of :nth-child(… of ).

pub fn might_have_nth_of_attribute_dependency( &self, local_name: &LocalName, ) -> bool

Returns whether the given attribute might appear in an attribute selector in the selector list of :nth-child(… of ).

pub fn normal_rules( &self, pseudo_elements: &[PseudoElement], ) -> Option<&SelectorMap<Rule>>

Returns the normal rule map for a given pseudo-element.

pub fn featureless_host_rules( &self, pseudo_elements: &[PseudoElement], ) -> Option<&SelectorMap<Rule>>

Returns the featureless pseudo rule map for a given pseudo-element.

pub fn any_featureless_host_rules(&self) -> bool

Whether there’s any featureless rule that could match in this scope.

pub fn slotted_rules( &self, pseudo_elements: &[PseudoElement], ) -> Option<&SelectorMap<Rule>>

Returns the slotted rule map for a given pseudo-element.

pub fn any_slotted_rule(&self) -> bool

Whether there’s any ::slotted rule that could match in this scope.

pub fn part_rules( &self, pseudo_elements: &[PseudoElement], ) -> Option<&PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>

Returns the parts rule map for a given pseudo-element.

pub fn any_part_rule(&self) -> bool

Whether there’s any ::part rule that could match in this scope.

fn layer_order_for(&self, id: LayerId) -> LayerOrder

pub(crate) fn container_condition_matches<E>( &self, id: ContainerConditionId, stylist: &Stylist, element: E, context: &mut MatchingContext<'_, E::Impl>, ) -> bool
where E: TElement,

pub(crate) fn find_scope_proximity_if_matching<E: TElement>( &self, rule: &Rule, element: E, context: &mut MatchingContext<'_, E::Impl>, ) -> ScopeProximity

fn did_finish_rebuild(&mut self)

fn shrink_maps_if_needed(&mut self)

fn compute_layer_order(&mut self)

fn collect_applicable_media_query_results_into<S>( device: &Device, stylesheet: &S, guard: &SharedRwLockReadGuard<'_>, results: &mut Vec<MediaListKey>, contents_list: &mut Vec<StylesheetContentsPtr>, custom_media_map: &mut CustomMediaMap, )
where S: StylesheetInDocument + 'static,

Collects all the applicable media query results into results.

This duplicates part of the logic in add_stylesheet, which is a bit unfortunate.

FIXME(emilio): With a bit of smartness in media_feature_affected_matches, we could convert EffectiveMediaQueryResults into a vector without too much effort.

fn add_styles( &mut self, selectors: &SelectorList<SelectorImpl>, declarations: &Arc<Locked<PropertyDeclarationBlock>>, ancestor_selectors: Option<&SelectorList<SelectorImpl>>, containing_rule_state: &ContainingRuleState, replaced_selectors: Option<&mut SmallVec<[Selector<SelectorImpl>; 4]>>, guard: &SharedRwLockReadGuard<'_>, rebuild_kind: SheetRebuildKind, precomputed_pseudo_element_decls: Option<&mut PerPseudoElementMap<Vec<ApplicableDeclarationBlock>>>, quirks_mode: QuirksMode, collected_scope_dependencies: Option<&mut Vec<Dependency>>, ) -> Result<(), AllocErr>

fn add_rule_list<S>( &mut self, rules: Iter<'_, CssRule>, device: &Device, quirks_mode: QuirksMode, stylesheet: &S, sheet_index: usize, guard: &SharedRwLockReadGuard<'_>, rebuild_kind: SheetRebuildKind, containing_rule_state: &mut ContainingRuleState, precomputed_pseudo_element_decls: Option<&mut PerPseudoElementMap<Vec<ApplicableDeclarationBlock>>>, ) -> Result<(), AllocErr>
where S: StylesheetInDocument + 'static,

fn add_stylesheet<S>( &mut self, device: &Device, quirks_mode: QuirksMode, stylesheet: &S, sheet_index: usize, guard: &SharedRwLockReadGuard<'_>, rebuild_kind: SheetRebuildKind, precomputed_pseudo_element_decls: Option<&mut PerPseudoElementMap<Vec<ApplicableDeclarationBlock>>>, ) -> Result<(), AllocErr>
where S: StylesheetInDocument + 'static,

pub fn media_feature_affected_matches<S>( &self, stylesheet: &S, guard: &SharedRwLockReadGuard<'_>, device: &Device, quirks_mode: QuirksMode, ) -> bool
where S: StylesheetInDocument + 'static,

Returns whether all the media-feature affected values matched before and match now in the given stylesheet.

pub fn custom_property_registrations( &self, ) -> &LayerOrderedMap<Arc<PropertyRegistration>>

Returns the custom properties map.

fn revalidate_scopes<E: TElement>( &self, element: &E, matching_context: &mut MatchingContext<'_, E::Impl>, result: &mut ScopeRevalidationResult, )

fn clear_cascade_data(&mut self)

Clears the cascade data, but not the invalidation data.

fn clear(&mut self)