Struct style::stylist::CascadeData

pub struct CascadeData {

    normal_rules: GenericElementAndPseudoRules<SelectorMap<Rule>>,
    host_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>,
    slotted_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>,
    part_rules: Option<Box<GenericElementAndPseudoRules<PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>>>,
    invalidation_map: InvalidationMap,
    attribute_dependencies: PrecomputedHashSet<LocalName>,
    state_dependencies: ElementState,
    document_state_dependencies: DocumentState,
    mapped_ids: PrecomputedHashSet<Atom>,
    selectors_for_cache_revalidation: SelectorMap<RevalidationSelectorAndHashes>,
    animations: PrecomputedHashMap<Atom, KeyframesAnimation>,
    effective_media_query_results: EffectiveMediaQueryResults,
    extra_data: ExtraStyleData,
    rules_source_order: u32,
    num_selectors: usize,
    num_declarations: usize,

}

Data resulting from performing the CSS cascade that is specific to a given origin.

FIXME(emilio): Consider renaming and splitting in CascadeData and InvalidationData? That’d make clear_cascade_data() clearer.

Fields
normal_rules: GenericElementAndPseudoRules<SelectorMap<Rule>>

The data coming from normal style rules that apply to elements at this cascade level.

host_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>

The :host pseudo rules that are the rightmost selector (without accounting for pseudo-elements).

slotted_rules: Option<Box<GenericElementAndPseudoRules<SelectorMap<Rule>>>>

The data coming from ::slotted() pseudo-element rules.

We need to store them separately because an element needs to match ::slotted() pseudo-element rules in different shadow roots.

In particular, we need to go through all the style data in all the containing style scopes starting from the closest assigned slot.

part_rules: Option<Box<GenericElementAndPseudoRules<PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>>>

The data coming from ::part() pseudo-element rules.

We need to store them separately because an element needs to match ::part() pseudo-element rules in different shadow roots.

invalidation_map: InvalidationMap

The invalidation map for these rules.

attribute_dependencies: PrecomputedHashSet<LocalName>

The attribute local names that appear in attribute selectors. Used to avoid taking element snapshots when an irrelevant attribute changes. (We don’t bother storing the namespace, since namespaced attributes are rare.)

state_dependencies: ElementState

The element state bits that are relied on by selectors. Like attribute_dependencies, this is used to avoid taking element snapshots when an irrelevant element state bit changes.

document_state_dependencies: DocumentState

The document state bits that are relied on by selectors. This is used to tell whether we need to restyle the entire document when a document state bit changes.

mapped_ids: PrecomputedHashSet<Atom>

The ids that appear in the rightmost complex selector of selectors (and hence in our selector maps). Used to determine when sharing styles is safe: we disallow style sharing for elements whose id matches this filter, and hence might be in one of our selector maps.

selectors_for_cache_revalidation: SelectorMap<RevalidationSelectorAndHashes>

Selectors that require explicit cache revalidation (i.e. which depend on state that is not otherwise visible to the cache, like attributes or tree-structural state like child index and pseudos).

animations: PrecomputedHashMap<Atom, KeyframesAnimation>

A map with all the animations at this CascadeData’s origin, indexed by name.

effective_media_query_results: EffectiveMediaQueryResults

Effective media query results cached from the last rebuild.

extra_data: ExtraStyleData

Extra data, like different kinds of rules, etc.

rules_source_order: u32

A monotonically increasing counter to represent the order on which a style rule appears in a stylesheet, needed to sort them by source order.

num_selectors: usize

The total number of selectors.

num_declarations: usize

The total number of declarations.

Implementations

impl CascadeData

pub fn new() -> Self

Creates an empty CascadeData.

pub fn rebuild<'a, S>(
&mut self,
device: &Device,
quirks_mode: QuirksMode,
collection: SheetCollectionFlusher<'_, S>,
guard: &SharedRwLockReadGuard<'_>
) -> Result<(), FailedAllocationError> where
S: StylesheetInDocument + ToMediaListKey + PartialEq + 'static,

Rebuild the cascade data from a given SheetCollection, incrementally if possible.

pub fn invalidation_map(&self) -> &InvalidationMap

Returns the invalidation map.

pub fn has_state_dependency(&self, state: ElementState) -> bool

Returns whether the given ElementState bit is relied upon by a selector of some rule.

pub fn might_have_attribute_dependency(&self, local_name: &LocalName) -> bool

Returns whether the given attribute might appear in an attribute selector of some rule.

pub fn normal_rules(
&self,
pseudo: Option<&PseudoElement>
) -> Option<&SelectorMap<Rule>>

Returns the normal rule map for a given pseudo-element.

pub fn host_rules(
&self,
pseudo: Option<&PseudoElement>
) -> Option<&SelectorMap<Rule>>

Returns the host pseudo rule map for a given pseudo-element.

pub fn any_host_rules(&self) -> bool

Whether there’s any host rule that could match in this scope.

pub fn slotted_rules(
&self,
pseudo: Option<&PseudoElement>
) -> Option<&SelectorMap<Rule>>

Returns the slotted rule map for a given pseudo-element.

pub fn any_slotted_rule(&self) -> bool

Whether there’s any ::slotted rule that could match in this scope.

pub fn part_rules(
&self,
pseudo: Option<&PseudoElement>
) -> Option<&PrecomputedHashMap<Atom, SmallVec<[Rule; 1]>>>

Returns the parts rule map for a given pseudo-element.

pub fn any_part_rule(&self) -> bool

Whether there’s any ::part rule that could match in this scope.

fn collect_applicable_media_query_results_into<S>(
device: &Device,
stylesheet: &S,
guard: &SharedRwLockReadGuard<'_>,
results: &mut EffectiveMediaQueryResults
) where
S: StylesheetInDocument + ToMediaListKey + 'static,

Collects all the applicable media query results into results.

This duplicates part of the logic in add_stylesheet, which is a bit unfortunate.

FIXME(emilio): With a bit of smartness in media_feature_affected_matches, we could convert EffectiveMediaQueryResults into a vector without too much effort.

fn add_stylesheet<S>(
&mut self,
device: &Device,
quirks_mode: QuirksMode,
stylesheet: &S,
guard: &SharedRwLockReadGuard<'_>,
rebuild_kind: SheetRebuildKind,
precomputed_pseudo_element_decls: Option<&mut PerPseudoElementMap<Vec<ApplicableDeclarationBlock>>>
) -> Result<(), FailedAllocationError> where
S: StylesheetInDocument + ToMediaListKey + 'static,

pub fn media_feature_affected_matches<S>(
&self,
stylesheet: &S,
guard: &SharedRwLockReadGuard<'_>,
device: &Device,
quirks_mode: QuirksMode
) -> bool where
S: StylesheetInDocument + ToMediaListKey + 'static,

Returns whether all the media-feature affected values matched before and match now in the given stylesheet.

fn clear_cascade_data(&mut self)

Clears the cascade data, but not the invalidation data.

fn clear(&mut self)