Struct style::stylist::SelectorMap

pub struct SelectorMap {
    pub id_hash: FnvHashMap<Atom, Vec<Rule>>,
    pub class_hash: FnvHashMap<Atom, Vec<Rule>>,
    pub local_name_hash: FnvHashMap<LocalName, Vec<Rule>>,
    pub lower_local_name_hash: FnvHashMap<LocalName, Vec<Rule>>,
    pub other_rules: Vec<Rule>,
    pub empty: bool,
}

Map element data to Rules whose last simple selector starts with them.

e.g., "p > img" would go into the set of Rules corresponding to the element "img" "a .foo .bar.baz" would go into the set of Rules corresponding to the class "bar"

Because we match Rules right-to-left (i.e., moving up the tree from an element), we need to compare the last simple selector in the Rule with the element.

So, if an element has ID "id1" and classes "foo" and "bar", then all the rules it matches will have their last simple selector starting either with "#id1" or with ".foo" or with ".bar".

Hence, the union of the rules keyed on each of element's classes, ID, element name, etc. will contain the Rules that actually match that element.

TODO: Tune the initial capacity of the HashMap

Fields
id_hash: FnvHashMap<Atom, Vec<Rule>>

A hash from an ID to rules which contain that ID selector.

class_hash: FnvHashMap<Atom, Vec<Rule>>

A hash from a class name to rules which contain that class selector.

local_name_hash: FnvHashMap<LocalName, Vec<Rule>>

A hash from local name to rules which contain that local name selector.

lower_local_name_hash: FnvHashMap<LocalName, Vec<Rule>>

Same as local_name_hash, but keys are lower-cased. For HTML elements in HTML documents.

other_rules: Vec<Rule>

Rules that don't have ID, class, or element selectors.

empty: bool

Whether this hash is empty.

Methods

impl SelectorMap

fn new() -> Self

Trivially constructs an empty SelectorMap.

fn get_all_matching_rules<E, V>(&self,
element: &E,
parent_bf: Option<&BloomFilter>,
matching_rules_list: &mut V,
relations: &mut StyleRelations,
flags: &mut ElementSelectorFlags,
cascade_level: CascadeLevel) where E: Element<Impl=SelectorImpl>, V: VecLike<ApplicableDeclarationBlock>

Append to rule_list all Rules in self that match element.

Extract matching rules as per element's ID, classes, tag name, etc.. Sort the Rules at the end to maintain cascading order.

fn get_universal_rules(&self,
cascade_level: CascadeLevel,
important_cascade_level: CascadeLevel)
-> Vec<ApplicableDeclarationBlock>

Append to rule_list all universal Rules (rules with selector *|*) in self sorted by specificity and source order.

fn insert(&mut self, rule: Rule)

Insert rule into the correct hash. Order in which to try: id_hash, class_hash, local_name_hash, other_rules.

fn get_id_name(rule: &Rule) -> Option<Atom>

Retrieve the first ID name in Rule, or None otherwise.

fn get_class_name(rule: &Rule) -> Option<Atom>

Retrieve the FIRST class name in Rule, or None otherwise.

fn get_local_name(rule: &Rule) -> Option<LocalNameSelector<SelectorImpl>>

Retrieve the name if it is a type selector, or None otherwise.