Struct style::stylist::Stylist

pub struct Stylist {
    pub device: Arc<Device>,
    pub rule_tree: RuleTree,
    // some fields omitted
}

This structure holds all the selectors and device characteristics for a given document. The selectors are converted into Rules (defined in rust-selectors), and introduced in a SelectorMap depending on the pseudo-element (see PerPseudoElementSelectorMap), and stylesheet origin (see the fields of PerPseudoElementSelectorMap).

This structure is effectively created once per pipeline, in the LayoutThread corresponding to that pipeline.

Fields
device: Arc<Device>

Device that the stylist is currently evaluating against.

This field deserves a bigger comment due to the different use that Gecko and Servo give to it (that we should eventually unify).

With Gecko, the device is never changed. Gecko manually tracks whether the device data should be reconstructed, and "resets" the state of the device.

On Servo, on the other hand, the device is a really cheap representation that is recreated each time some constraint changes and calling set_device.

In both cases, the device is actually owned by the Stylist, and it's only an Arc so we can implement add_stylesheet more idiomatically.

rule_tree: RuleTree

The rule tree, that stores the results of selector matching.

FIXME(emilio): Not pub!

Methods

impl Stylist

fn new(device: Device) -> Self

Construct a new Stylist, using a given Device.

fn update(&mut self,
doc_stylesheets: &[Arc<Stylesheet>],
ua_stylesheets: Option<&UserAgentStylesheets>,
stylesheets_changed: bool)
-> bool

Update the stylist for the given document stylesheets, and optionally with a set of user agent stylesheets.

This method resets all the style data each time the stylesheets change (which is indicated by the stylesheets_changed parameter), or the device is dirty, which means we need to re-evaluate media queries.

fn precomputed_values_for_pseudo(&self,
pseudo: &PseudoElement,
parent: Option<&Arc<ComputedValues>>,
default: &Arc<ComputedValues>,
inherit_all: bool)
-> ComputedStyle

Computes the style for a given "precomputed" pseudo-element, taking the universal rules and applying them.

If inherit_all is true, then all properties are inherited from the parent; otherwise, non-inherited properties are reset to their initial values. The flow constructor uses this flag when constructing anonymous flows.

fn lazily_compute_pseudo_element_style<E>(&self,
element: &E,
pseudo: &PseudoElement,
parent: &Arc<ComputedValues>,
default: &Arc<ComputedValues>)
-> Option<ComputedStyle> where E: TElement + Debug + PresentationalHintsSynthetizer

Computes a pseudo-element style lazily during layout.

This can only be done for a certain set of pseudo-elements, like :selection.

Check the documentation on lazy pseudo-elements in docs/components/style.md

fn viewport_constraints(&self) -> Option<&ViewportConstraints>

Returns the viewport constraints that apply to this document because of a @viewport rule.

fn set_quirks_mode(&mut self, enabled: bool)

Sets the quirks mode of the document.

fn push_applicable_declarations<E, V>(&self,
element: &E,
parent_bf: Option<&BloomFilter>,
style_attribute: Option<&Arc<RwLock<PropertyDeclarationBlock>>>,
animation_rules: AnimationRules,
pseudo_element: Option<&PseudoElement>,
applicable_declarations: &mut V,
flags: &mut ElementSelectorFlags)
-> StyleRelations where E: TElement + Debug + PresentationalHintsSynthetizer,
V: Push<ApplicableDeclarationBlock> + VecLike<ApplicableDeclarationBlock>

Returns the applicable CSS declarations for the given element.

This corresponds to ElementRuleCollector in WebKit.

The returned StyleRelations indicate hints about which kind of rules have matched.

fn is_device_dirty(&self) -> bool

Return whether the device is dirty, that is, whether the screen size or media type have changed (for now).

fn animations(&self) -> &FnvHashMap<Atom, KeyframesAnimation>

Returns the map of registered @keyframes animations.

fn match_same_not_common_style_affecting_attributes_rules<E>(&self,
element: &E,
candidate: &E)
-> bool where E: TElement

Whether two elements match the same not-common style-affecting attribute rules.

This is used to test elements and candidates in the style-sharing candidate cache.

fn rule_tree_root(&self) -> StrongRuleNode

Returns the rule root node.

fn match_same_sibling_affecting_rules<E>(&self,
element: &E,
candidate: &E)
-> bool where E: TElement

Returns whether two elements match the same sibling-affecting rules.

This is also for the style sharing candidate cache.

fn compute_restyle_hint<E>(&self,
element: &E,
snapshot: &Snapshot)
-> RestyleHint where E: TElement

Given an element, and a snapshot that represents a previous state of the element, compute the appropriate restyle hint, that is, the kind of restyle we need to do.