Module style::traversal

Traversing the DOM tree; the bloom filter.

Structs

PerLevelTraversalData
A per-traversal-level chunk of data. This is sent down by the traversal, and currently only holds the dom depth for the bloom filter.

PreTraverseToken
We use this structure, rather than just returning a boolean from pre_traverse, to enfore that callers process root invalidations before starting the traversal.

Statics

IS_SERVO_NONINCREMENTAL_LAYOUT
A global variable holding the state of is_servo_nonincremental_layout(). See #22854.

Traits

DomTraversal
A DOM Traversal trait, that is used to generically implement styling for Gecko and Servo.

Functions

clear_descendant_data⚠
Clear style data for all the subtree under root (but not for root itself).

clear_state_after_traversing

compute_style

is_servo_nonincremental_layout

note_children

notify_paint_worklet

recalc_style_at
Calculates the style for a single node.

resolve_style
Manually resolve style by sequentially walking up the parent chain to the first styled Element, ignoring pending restyles. The resolved style is made available via a callback, and can be dropped by the time this function returns in the display:none subtree case.