Module style::traversal

Traversing the DOM tree; the bloom filter.

Structs

PerLevelTraversalData
A per-traversal-level chunk of data. This is sent down by the traversal, and currently only holds the dom depth for the bloom filter.

PreTraverseToken
This structure exists to enforce that callers invoke pre_traverse, and also to pass information from the pre-traversal into the primary traversal.

Enums

LogBehavior
Enum to prevent duplicate logging.

TraversalDriver
The kind of traversals we could perform.

Traits

DomTraversal
A DOM Traversal trait, that is used to generically implement styling for Gecko and Servo.

Functions

clear_descendant_data
Clear style data for all the subtree under el.

recalc_style_at
Calculates the style for a single node.

resolve_style
Manually resolve style by sequentially walking up the parent chain to the first styled Element, ignoring pending restyles. The resolved style is made available via a callback, and can be dropped by the time this function returns in the display:none subtree case.