Function style::traversal::resolve_style

pub fn resolve_style<E>(
    context: &mut StyleContext<'_, E>,
    element: E,
    rule_inclusion: RuleInclusion,
    pseudo: Option<&PseudoElement>,
    undisplayed_style_cache: Option<&mut UndisplayedStyleCache>,
) -> ElementStyles
where
    E: TElement,

Manually resolve style by sequentially walking up the parent chain to the first styled Element, ignoring pending restyles. The resolved style is made available via a callback, and can be dropped by the time this function returns in the display:none subtree case.