Function style::traversal::resolve_style

pub fn resolve_style<E>(
    context: &mut StyleContext<E>, 
    element: E, 
    rule_inclusion: RuleInclusion, 
    ignore_existing_style: bool, 
    pseudo: Option<&PseudoElement>
) -> ElementStyles where
    E: TElement, 

Manually resolve style by sequentially walking up the parent chain to the first styled Element, ignoring pending restyles. The resolved style is made available via a callback, and can be dropped by the time this function returns in the display:none subtree case.