Trait style::traversal::DomTraversal

pub trait DomTraversal<E: TElement>: Sync {
    type ThreadLocalContext: Send + BorrowMut<ThreadLocalStyleContext<E>>;
    fn process_preorder(&self,
                        data: &mut PerLevelTraversalData,
                        thread_local: &mut Self::ThreadLocalContext,
                        node: E::ConcreteNode);
    fn process_postorder(&self,
                         thread_local: &mut Self::ThreadLocalContext,
                         node: E::ConcreteNode);
    unsafe fn ensure_element_data(element: &E) -> &AtomicRefCell<ElementData>;
    unsafe fn clear_element_data(element: &E);
    fn shared_context(&self) -> &SharedStyleContext;
    fn create_thread_local_context(&self) -> Self::ThreadLocalContext;
    fn is_parallel(&self) -> bool;

    fn needs_postorder_traversal() -> bool { ... }
    fn pre_traverse(root: E,
                    stylist: &Stylist,
                    unstyled_children_only: bool)
                    -> PreTraverseToken { ... }
    fn text_node_needs_traversal(node: E::ConcreteNode) -> bool { ... }
    fn node_needs_traversal(node: E::ConcreteNode) -> bool { ... }
    fn should_traverse_children(&self,
                                thread_local: &mut ThreadLocalStyleContext<E>,
                                parent: E,
                                parent_data: &ElementData,
                                log: LogBehavior)
                                -> bool { ... }
    fn traverse_children<F>(&self,
                            thread_local: &mut Self::ThreadLocalContext,
                            parent: E,
                            f: F) where F: FnMut(&mut Self::ThreadLocalContext, E::ConcreteNode) { ... }
}

A DOM Traversal trait, that is used to generically implement styling for Gecko and Servo.

Associated Types

type ThreadLocalContext: Send + BorrowMut<ThreadLocalStyleContext<E>>

The thread-local context, used to store non-thread-safe stuff that needs to be used in the traversal, and of which we use one per worker, like the bloom filter, for example.

Required Methods

fn process_preorder(&self,
data: &mut PerLevelTraversalData,
thread_local: &mut Self::ThreadLocalContext,
node: E::ConcreteNode)

Process node on the way down, before its children have been processed.

fn process_postorder(&self,
thread_local: &mut Self::ThreadLocalContext,
node: E::ConcreteNode)

Process node on the way up, after its children have been processed.

This is only executed if needs_postorder_traversal returns true.

unsafe fn ensure_element_data(element: &E) -> &AtomicRefCell<ElementData>

Ensures the existence of the ElementData, and returns it. This can't live on TNode because of the trait-based separation between Servo's script and layout crates.

This is only safe to call in top-down traversal before processing the children of |element|.

unsafe fn clear_element_data(element: &E)

Clears the ElementData attached to this element, if any.

This is only safe to call in top-down traversal before processing the children of |element|.

fn shared_context(&self) -> &SharedStyleContext

Return the shared style context common to all worker threads.

fn create_thread_local_context(&self) -> Self::ThreadLocalContext

Creates a thread-local context.

fn is_parallel(&self) -> bool

Whether we're performing a parallel traversal.

NB: We do this check on runtime. We could guarantee correctness in this regard via the type system via a TraversalDriver trait for this trait, that could be one of two concrete types. It's not clear whether the potential code size impact of that is worth it.

Provided Methods

fn needs_postorder_traversal() -> bool

Boolean that specifies whether a bottom up traversal should be performed.

If it's false, then process_postorder has no effect at all.

fn pre_traverse(root: E,
stylist: &Stylist,
unstyled_children_only: bool)
-> PreTraverseToken

Must be invoked before traversing the root element to determine whether a traversal is needed. Returns a token that allows the caller to prove that the call happened.

The unstyled_children_only parameter is used in Gecko to style newly- appended children without restyling the parent.

fn text_node_needs_traversal(node: E::ConcreteNode) -> bool

Returns true if traversal should visit a text node. The style system never processes text nodes, but Servo overrides this to visit them for flow construction when necessary.

fn node_needs_traversal(node: E::ConcreteNode) -> bool

Returns true if traversal is needed for the given node and subtree.

fn should_traverse_children(&self,
thread_local: &mut ThreadLocalStyleContext<E>,
parent: E,
parent_data: &ElementData,
log: LogBehavior)
-> bool

Returns true if traversal of this element's children is allowed. We use this to cull traversal of various subtrees.

This may be called multiple times when processing an element, so we pass a parameter to keep the logs tidy.

fn traverse_children<F>(&self,
thread_local: &mut Self::ThreadLocalContext,
parent: E,
f: F) where F: FnMut(&mut Self::ThreadLocalContext, E::ConcreteNode)

Helper for the traversal implementations to select the children that should be enqueued for processing.