style

Module use_counters

Various stuff for CSS property use counters.

Macros

use_counters_methods 🔒

Structs

CountedUnknownPropertyUseCounters

One bit per each non-custom CSS property.

CustomUseCounters

One bit for each custom use counter.

NonCustomPropertyUseCounters

One bit per each non-custom CSS property.

UseCounters

The use-counter data related to a given document we want to store.

Enums

CustomUseCounter

A custom style use counter that we may want to record.

Constants

BITS_PER_ENTRY 🔒