Module style::values

Common values used in CSS.

Modules

computed
Computed values.

specified
Specified values.

Structs

Auto

None_

Normal

Parser
A CSS parser that borrows its &str input, yields Tokens, and keeps track of nested blocks and functions.

RGBA
A color with red, green, blue, and alpha components.

Enums

Either
A struct representing one of two kinds of values.

Constants

FONT_MEDIUM_PX
The default font size.

Traits

HasViewportPercentage
A trait used to query whether this value has viewport units.

Type Definitions

CSSFloat
A CSS float value.