Module style::values::animated

Animated values.

Some values, notably colors, cannot be interpolated directly with their computed values and need yet another intermediate representation. This module’s raison d’être is to ultimately contain all these types.

Modules

color

Animated types for CSS colors.

effects

Animated types for CSS values related to effects.

font

Animation implementation for various font-related types.

grid

Animation implementation for various grid-related types.

svg

Animation implementations for various SVG-related types.

transform

Animated types for transform.

Macros

trivial_to_animated_value

Enums

Procedure

An animation procedure.

PropertyCategory

The category a property falls into for ordering purposes.

Traits

Animate

Animate from one value to another.

ToAnimatedValue

Conversion between computed values and intermediate values for animations.

ToAnimatedZero

Returns a value similar to self that represents zero.

Functions

animate_multiplicative_factor

A helper function to animate two multiplicative factor.

compare_property_priority

A comparator to sort PropertyIds such that physical longhands are sorted before logical longhands and shorthands, shorthands with fewer components are sorted before shorthands with more components, and otherwise shorthands are sorted by IDL name as defined by [Web Animations][property-order].