Module style::values::animated

Animated values.

Some values, notably colors, cannot be interpolated directly with their computed values and need yet another intermediate representation. This module's raison d'être is to ultimately contain all these types.

Modules

color
Animated types for CSS colors.

effects
Animated types for CSS values related to effects.

Enums

Procedure
An animation procedure.

Traits

Animate
Animate from one value to another.

AnimatedValueAsComputed
Marker trait for computed values with the same representation during animations.

ToAnimatedValue
Conversion between computed values and intermediate values for animations.

ToAnimatedZero
Returns a value similar to self that represents zero.