Trait style::values::animated::Animate

pub trait Animate: Sized {
    fn animate(&self, other: &Self, procedure: Procedure) -> Result<Self, ()>;
}

Animate from one value to another.

This trait is derivable with #[derive(Animate)]. The derived implementation uses a match expression with identical patterns for both self and other, calling Animate::animate on each fields of the values. If a field is annotated with #[animation(constant)], the two values should be equal or an error is returned.

If a variant is annotated with #[animation(error)], the corresponding match arm is not generated.

If the two values are not similar, an error is returned unless a fallback function has been specified through #[animate(fallback)].

Required Methods

fn animate(&self, other: &Self, procedure: Procedure) -> Result<Self, ()>

Animate a value towards another one, given an animation procedure.