Trait style::values::animated::ToAnimatedValue

pub trait ToAnimatedValue {
    type AnimatedValue;
    fn to_animated_value(self) -> Self::AnimatedValue;

    fn from_animated_value(animated: Self::AnimatedValue) -> Self;
}

Conversion between computed values and intermediate values for animations.

Notably, colors are represented as four floats during animations.

This trait is derivable with #[derive(ToAnimatedValue)].

Associated Types

type AnimatedValue

The type of the animated value.

Required methods

fn to_animated_value(self) -> Self::AnimatedValue

Converts this value to an animated value.

fn from_animated_value(animated: Self::AnimatedValue) -> Self

Converts back an animated value into a computed value.