Module style::values::computed

Computed values.

Re-exports

pub use self::angle::Angle;

pub use self::background::BackgroundRepeat;

pub use self::background::BackgroundSize;

pub use self::basic_shape::FillRule;

pub use self::border::BorderCornerRadius;

pub use self::border::BorderRadius;

pub use self::border::BorderSpacing;

pub use self::border::BorderImageRepeat;

pub use self::border::BorderImageSideWidth;

pub use self::border::BorderImageSlice;

pub use self::border::BorderImageWidth;

pub use self::box_::AnimationIterationCount;

pub use self::box_::AnimationName;

pub use self::box_::Contain;

pub use self::box_::Appearance;

pub use self::box_::BreakBetween;

pub use self::box_::BreakWithin;

pub use self::box_::Clear;

pub use self::box_::Float;

pub use self::box_::Display;

pub use self::box_::Overflow;

pub use self::box_::OverflowAnchor;

pub use self::box_::TransitionProperty;

pub use self::box_::OverflowClipBox;

pub use self::box_::OverscrollBehavior;

pub use self::box_::Perspective;

pub use self::box_::Resize;

pub use self::box_::ScrollSnapAlign;

pub use self::box_::ScrollSnapAxis;

pub use self::box_::ScrollSnapStrictness;

pub use self::box_::ScrollSnapType;

pub use self::box_::TouchAction;

pub use self::box_::VerticalAlign;

pub use self::box_::WillChange;

pub use self::color::Color;

pub use self::color::ColorOrAuto;

pub use self::color::ColorPropertyValue;

pub use self::column::ColumnCount;

pub use self::counters::Content;

pub use self::counters::ContentItem;

pub use self::counters::CounterIncrement;

pub use self::counters::CounterSetOrReset;

pub use self::easing::TimingFunction;

pub use self::effects::BoxShadow;

pub use self::effects::Filter;

pub use self::effects::SimpleShadow;

pub use self::flex::FlexBasis;

pub use self::font::FontFamily;

pub use self::font::FontLanguageOverride;

pub use self::font::FontStyle;

pub use self::font::FontFeatureSettings;

pub use self::font::FontVariantLigatures;

pub use self::font::FontVariantNumeric;

pub use self::font::FontSize;

pub use self::font::FontSizeAdjust;

pub use self::font::FontStretch;

pub use self::font::FontSynthesis;

pub use self::font::FontVariantAlternates;

pub use self::font::FontWeight;

pub use self::font::FontVariantEastAsian;

pub use self::font::FontVariationSettings;

pub use self::font::MozScriptLevel;

pub use self::font::MozScriptMinSize;

pub use self::font::MozScriptSizeMultiplier;

pub use self::font::XLang;

pub use self::font::XTextZoom;

pub use self::image::Gradient;

pub use self::image::Image;

pub use self::image::LineDirection;

pub use self::image::MozImageRect;

pub use self::length::CSSPixelLength;

pub use self::length::ExtremumLength;

pub use self::length::NonNegativeLength;

pub use self::length::Length;

pub use self::length::LengthOrNumber;

pub use self::length::LengthPercentage;

pub use self::length::NonNegativeLengthOrNumber;

pub use self::length::LengthOrAuto;

pub use self::length::LengthPercentageOrAuto;

pub use self::length::MaxSize;

pub use self::length::Size;

pub use self::length::NonNegativeLengthPercentage;

pub use self::length::NonNegativeLengthPercentageOrAuto;

pub use self::list::MozListReversed;

pub use self::list::Quotes;

pub use self::motion::OffsetPath;

pub use self::motion::OffsetRotate;

pub use self::outline::OutlineStyle;

pub use self::percentage::NonNegativePercentage;

pub use self::percentage::Percentage;

pub use self::position::AspectRatio;

pub use self::position::GridAutoFlow;

pub use self::position::GridTemplateAreas;

pub use self::position::MasonryAutoFlow;

pub use self::position::Position;

pub use self::position::PositionOrAuto;

pub use self::position::ZIndex;

pub use self::rect::NonNegativeLengthOrNumberRect;

pub use self::resolution::Resolution;

pub use self::svg::MozContextProperties;

pub use self::svg::SVGLength;

pub use self::svg::SVGOpacity;

pub use self::svg::SVGPaint;

pub use self::svg::SVGPaintKind;

pub use self::svg::SVGPaintOrder;

pub use self::svg::SVGStrokeDashArray;

pub use self::svg::SVGWidth;

pub use self::text::TextUnderlinePosition;

pub use self::text::InitialLetter;

pub use self::text::LetterSpacing;

pub use self::text::LineBreak;

pub use self::text::LineHeight;

pub use self::text::OverflowWrap;

pub use self::text::TextOverflow;

pub use self::text::WordBreak;

pub use self::text::WordSpacing;

pub use self::text::TextAlign;

pub use self::text::TextAlignLast;

pub use self::text::TextEmphasisPosition;

pub use self::text::TextEmphasisStyle;

pub use self::text::TextDecorationLength;

pub use self::text::TextDecorationSkipInk;

pub use self::time::Time;

pub use self::transform::Rotate;

pub use self::transform::Scale;

pub use self::transform::Transform;

pub use self::transform::TransformOperation;

pub use self::transform::TransformOrigin;

pub use self::transform::TransformStyle;

pub use self::transform::Translate;

pub use self::ui::Cursor;

pub use self::ui::MozForceBrokenImageIcon;

pub use self::ui::UserSelect;

pub use super::specified::TextTransform;

pub use super::specified::BorderStyle;

pub use super::specified::TextDecorationLine;

pub use super::Auto;

pub use super::Either;

pub use super::None_;

Modules

angle
Computed angles.

background
Computed types for CSS values related to backgrounds.

basic_shape
CSS handling for the computed value of basic-shapes

border
Computed types for CSS values related to borders.

box_
Computed types for box properties.

color
Computed color values.

column
Computed types for the column properties.

counters
Computed values for counter properties

easing
Computed types for CSS Easing functions.

effects
Computed types for CSS values related to effects.

flex
Computed types for CSS values related to flexbox.

font
Computed values for font properties

image
CSS handling for the computed value of images

length
<length> computed values, and related ones.

length_percentage
<length-percentage> computed values, and related ones.

list
list computed values.

motion
Computed types for CSS values that are related to motion path.

outline
Computed values for outline properties

percentage
Computed percentages.

position
CSS handling for the computed value of position values.

rect
Computed types for CSS borders.

resolution
Resolution values:

svg
Computed types for SVG properties.

text
Computed types for text properties.

time
Computed time values.

transform
Computed types for CSS values that are related to transformations.

ui
Computed values for UI properties

url
Common handling for the computed value CSS url() values.

Structs

Au
An App Unit, the fundamental unit of length in Servo. Usually 1/60th of a pixel (see AU_PER_PX)

ComputedVecIter
An iterator over a slice of computed values

Context
A Context is all the data a specified value could ever need to compute itself and be transformed to a computed value.

Enums

AngleOrPercentage

NumberOrPercentage

Traits

ToComputedValue
A trait to represent the conversion between computed and specified values.

Type Definitions

ClipRect
rect(...) | auto

ClipRectOrAuto
rect(...) | auto

GreaterThanOrEqualToOneNumber
A wrapper of Number, but the value >= 1.

GridLine
The computed value of a <grid-line>.

GridTemplateComponent
<grid-template-rows> | <grid-template-columns>

ImplicitGridTracks
The computed value of a grid <track-size>+

Integer
A <integer> value.

NonNegativeNumber
A wrapper of Number, but the value >= 0.

NonNegativeNumberOrPercentage
A non-negative .

Number
A <number> value.

Opacity
A type used for opacity.

PositiveInteger
A wrapper of Integer, but only accept a value >= 1.

PositiveIntegerOrNone
A computed positive <integer> value or none.

TrackBreadth
The computed value of a grid <track-breadth>

TrackList
The computed value of a grid <track-list> (could also be <auto-track-list> or <explicit-track-list>)

TrackSize
The computed value of a grid <track-size>

ZeroToOneNumber
A wrapper of Number, but the value between 0 and 1