Module style::values::computed

Computed values.

Reexports

pub use self::image::AngleOrCorner;

pub use self::image::EndingShape as GradientShape;

pub use self::image::Gradient;

pub use self::image::GradientKind;

pub use self::image::Image;

pub use self::image::LengthOrKeyword;

pub use self::image::LengthOrPercentageOrKeyword;

pub use super::Auto;

pub use super::Either;

pub use super::None_;

pub use super::specified::Angle;

pub use super::specified::BorderStyle;

pub use super::specified::GridLine;

pub use super::specified::Time;

pub use super::specified::UrlOrNone;

pub use super::specified::url::UrlExtraData;

pub use self::length::CalcLengthOrPercentage;

pub use self::length::Length;

pub use self::length::LengthOrNumber;

pub use self::length::LengthOrPercentage;

pub use self::length::LengthOrPercentageOrAuto;

pub use self::length::LengthOrPercentageOrAutoOrContent;

pub use self::length::LengthOrPercentageOrNone;

pub use self::length::LengthOrNone;

pub use self::position::Position;

Modules

basic_shape
CSS handling for the computed value of basic-shapes

image
CSS handling for the computed value of images

length
<length> computed values, and related ones.

position
CSS handling for the computed value of positions

Structs

BorderRadiusSize

ClipRect
A computed cliprect for clip and image-region

Context
A Context is all the data a specified value could ever need to compute itself and be transformed to a computed value.

Shadow

Enums

CSSColor
A value.

Traits

ComputedValueAsSpecified
A marker trait to represent that the specified value is also the computed value.

ToComputedValue
A trait to represent the conversion between computed and specified values.

Type Definitions

ClipRectOrAuto
rect(...) | auto

Number
A <number> value.

Opacity
A type used for opacity.