Module style::values::computed

Reexports

pub use super::specified::{BorderStyle, Time};

Structs

ColorStop
Computed values for one color stop in a linear gradient.

Context

LengthAndPercentage
The sum of a series of lengths and a percentage. This is used in calc() and other things that effectively work like it (e.g. transforms).

LinearGradient
Computed values for a CSS linear gradient.

Enums

CSSColor
A value.

Image
Computed values for an image according to CSS-IMAGES.

LengthOrPercentage

LengthOrPercentageOrAuto

LengthOrPercentageOrNone

Traits

ComputedValueAsSpecified

ToComputedValue

Type Definitions

Length