Module style::values::computed

Computed values.

Reexports

pub use properties::animated_properties::TransitionProperty;

pub use self::angle::Angle;

pub use self::background::BackgroundSize;

pub use self::background::BackgroundRepeat;

pub use self::border::BorderImageSlice;

pub use self::border::BorderImageWidth;

pub use self::border::BorderImageSideWidth;

pub use self::border::BorderRadius;

pub use self::border::BorderCornerRadius;

pub use self::border::BorderSpacing;

pub use self::font::FontSize;

pub use self::font::FontSizeAdjust;

pub use self::font::FontSynthesis;

pub use self::font::FontWeight;

pub use self::font::FontVariantAlternates;

pub use self::font::FontFamily;

pub use self::font::FontLanguageOverride;

pub use self::font::FontVariantSettings;

pub use self::font::FontVariantEastAsian;

pub use self::font::FontVariantLigatures;

pub use self::font::FontVariantNumeric;

pub use self::font::FontFeatureSettings;

pub use self::font::MozScriptLevel;

pub use self::font::MozScriptMinSize;

pub use self::font::MozScriptSizeMultiplier;

pub use self::font::XTextZoom;

pub use self::font::XLang;

pub use self::box_::AnimationIterationCount;

pub use self::box_::AnimationName;

pub use self::box_::OverscrollBehavior;

pub use self::box_::OverflowClipBox;

pub use self::box_::ScrollSnapType;

pub use self::box_::VerticalAlign;

pub use self::box_::WillChange;

pub use self::color::Color;

pub use self::color::ColorPropertyValue;

pub use self::color::RGBAColor;

pub use self::effects::BoxShadow;

pub use self::effects::Filter;

pub use self::effects::SimpleShadow;

pub use self::flex::FlexBasis;

pub use self::image::Gradient;

pub use self::image::GradientItem;

pub use self::image::Image;

pub use self::image::ImageLayer;

pub use self::image::LineDirection;

pub use self::image::MozImageRect;

pub use self::rect::LengthOrNumberRect;

pub use super::Auto;

pub use super::Either;

pub use super::None_;

pub use super::specified::BorderStyle;

pub use super::specified::TextDecorationLine;

pub use self::length::CalcLengthOrPercentage;

pub use self::length::Length;

pub use self::length::LengthOrNone;

pub use self::length::LengthOrNumber;

pub use self::length::LengthOrPercentage;

pub use self::length::LengthOrPercentageOrAuto;

pub use self::length::LengthOrPercentageOrNone;

pub use self::length::MaxLength;

pub use self::length::MozLength;

pub use self::length::CSSPixelLength;

pub use self::length::NonNegativeLength;

pub use self::length::NonNegativeLengthOrPercentage;

pub use self::list::ListStyleImage;

pub use self::list::Quotes;

pub use self::outline::OutlineStyle;

pub use self::percentage::Percentage;

pub use self::position::Position;

pub use self::position::GridAutoFlow;

pub use self::position::GridTemplateAreas;

pub use self::svg::SVGLength;

pub use self::svg::SVGOpacity;

pub use self::svg::SVGPaint;

pub use self::svg::SVGPaintKind;

pub use self::svg::SVGStrokeDashArray;

pub use self::svg::SVGWidth;

pub use self::table::XSpan;

pub use self::text::InitialLetter;

pub use self::text::LetterSpacing;

pub use self::text::LineHeight;

pub use self::text::TextAlign;

pub use self::text::TextOverflow;

pub use self::text::WordSpacing;

pub use self::time::Time;

pub use self::transform::TimingFunction;

pub use self::transform::Transform;

pub use self::transform::TransformOperation;

pub use self::transform::TransformOrigin;

pub use self::ui::MozForceBrokenImageIcon;

Modules

angle
Computed angles.

background
Computed types for CSS values related to backgrounds.

basic_shape
CSS handling for the computed value of basic-shapes

border
Computed types for CSS values related to borders.

box_
Computed types for box properties.

color
Computed color values.

effects
Computed types for CSS values related to effects.

flex
Computed types for CSS values related to flexbox.

font
Computed values for font properties

image
CSS handling for the computed value of images

length
<length> computed values, and related ones.

list
list computed values.

outline
Computed values for outline properties

percentage
Computed percentages.

position
CSS handling for the computed value of position values.

rect
Computed types for CSS borders.

svg
Computed types for SVG properties.

table
Computed types for table properties.

text
Computed types for text properties.

time
Computed time values.

transform
Computed types for CSS values that are related to transformations.

ui
Computed values for UI properties

Structs

Au
An App Unit, the fundamental unit of length in Servo. Usually 1/60th of a pixel (see AU_PER_PX)

ClipRect
A computed cliprect for clip and image-region

ComputedVecIter
An iterator over a slice of computed values

Context
A Context is all the data a specified value could ever need to compute itself and be transformed to a computed value.

Enums

ComputedUrl
The computed value of a CSS url(), resolved relative to the stylesheet URL.

NumberOrPercentage

Traits

ToComputedValue
A trait to represent the conversion between computed and specified values.

Type Definitions

ClipRectOrAuto
rect(...) | auto

ColorOrAuto
| auto

GreaterThanOrEqualToOneNumber
A wrapper of Number, but the value >= 1.

GridLine
The computed value of a <grid-line>.

GridTemplateComponent
<grid-template-rows> | <grid-template-columns>

Integer
A <integer> value.

IntegerOrAuto
| auto

LengthOrPercentageOrNumber
| |

NonNegativeLengthOrPercentageOrNumber
NonNegativeLengthOrPercentage | NonNegativeNumber

NonNegativeNumber
A wrapper of Number, but the value >= 0.

Number
A <number> value.

Opacity
A type used for opacity.

PositiveInteger
A wrapper of Integer, but only accept a value >= 1.

PositiveIntegerOrAuto
PositiveInteger | auto

TrackBreadth
The computed value of a grid <track-breadth>

TrackList
The computed value of a grid <track-list> (could also be <auto-track-list> or <explicit-track-list>)

TrackSize
The computed value of a grid <track-size>

UrlOrNone
|