style::values::computed::color

Type Alias Color

pub type Color = GenericColor<Percentage>;

A computed value for <color>.

Aliased Type

#[repr(C)]
pub enum Color {
    Absolute(AbsoluteColor),
    ColorFunction(Box<ColorFunction<GenericColor<Percentage>>>),
    CurrentColor,
    ColorMix(Box<GenericColorMix<GenericColor<Percentage>, Percentage>>),
    ContrastColor(Box<GenericColor<Percentage>>),
}

Variants

Absolute(AbsoluteColor)

The actual numeric color.

ColorFunction(Box<ColorFunction<GenericColor<Percentage>>>)

A unresolvable color.

CurrentColor

The CurrentColor keyword.

ColorMix(Box<GenericColorMix<GenericColor<Percentage>, Percentage>>)

The color-mix() function.

ContrastColor(Box<GenericColor<Percentage>>)

The contrast-color() function.

Implementations

impl Color

pub const TRANSPARENT_BLACK: Self

A fully transparent color.

pub const BLACK: Self

An opaque black color.

pub const WHITE: Self

An opaque white color.

pub fn from_color_mix(color_mix: ColorMix) -> Self

Create a new computed Color from a given color-mix, simplifying it to an absolute color if possible.

pub fn resolve_to_absolute( &self, current_color: &AbsoluteColor, ) -> AbsoluteColor

Combine this complex color with the given foreground color into an absolute color.

fn contrast_ratio(a: &AbsoluteColor, b: &AbsoluteColor) -> f32