Type Alias style::values::computed::color::Color

pub type Color = GenericColor<Percentage>;

A computed value for <color>.

Aliased Type

enum Color {
    Absolute(AbsoluteColor),
    CurrentColor,
    ColorMix(Box<GenericColorMix<GenericColor<Percentage>, Percentage>>),
}

Variants

Absolute(AbsoluteColor)

The actual numeric color.

CurrentColor

The CurrentColor keyword.

ColorMix(Box<GenericColorMix<GenericColor<Percentage>, Percentage>>)

The color-mix() function.

Implementations

impl Color

pub const TRANSPARENT_BLACK: Self = _

A fully transparent color.

pub const BLACK: Self = _

An opaque black color.

pub const WHITE: Self = _

An opaque white color.

pub fn from_color_mix(color_mix: ColorMix) -> Self

Create a new computed Color from a given color-mix, simplifying it to an absolute color if possible.

pub fn resolve_to_absolute( &self, current_color: &AbsoluteColor, ) -> AbsoluteColor

Combine this complex color with the given foreground color into an absolute color.