Module style::values::computed::font

Computed values for font properties

Re-exports

pub use crate::values::computed::Length as MozScriptMinSize;

pub use crate::values::specified::font::FontSynthesis;

pub use crate::values::specified::font::MozScriptSizeMultiplier;

pub use crate::values::specified::font::XLang;

pub use crate::values::specified::font::XTextZoom;

Structs

FamilyName
The name of a font family of choice

FontFamily
Specifies a prioritized list of font family names or generic family names.

FontFamilyList
A list of SingleFontFamily

FontLanguageOverride
font-language-override can only have a single three-letter OpenType "language system" tag, so we should be able to compute it and store it as a 32-bit integer (see http://www.microsoft.com/typography/otspec/languagetags.htm).

FontSize
The computed value of font-size

FontStretch
A value for the font-stretch property per:

FontStyleAngle
A wrapper over an Angle, that handles clamping to the appropriate range for font-style animation.

FontWeight
A value for the font-weight property per:

Enums

FontFamilyNameSyntax
Font family names must either be given quoted as strings, or unquoted as a sequence of one or more identifiers.

FontSizeAdjust
Preserve the readability of text when font fallback occurs

GenericFontFamily
A generic font-family name.

SingleFontFamily
A set of faces that vary in weight, width or slope.

Type Definitions

FontFeatureSettings
Use FontSettings as computed type of FontFeatureSettings.

FontStyle
The computed value of font-style.

FontVariantAlternates
Use VariantAlternatesList as computed type of FontVariantAlternates

FontVariantEastAsian
Use VariantEastAsian as computed type of FontVariantEastAsian

FontVariantLigatures
Use VariantLigatures as computed type of FontVariantLigatures

FontVariantNumeric
Use VariantNumeric as computed type of FontVariantNumeric

FontVariationSettings
The computed value for font-variation-settings.

MozScriptLevel
The computed value of the -moz-script-level property.