Module style::values::computed::font

Computed values for font properties

Re-exports

pub use crate::values::computed::Length as MozScriptMinSize;

pub use crate::values::specified::font::MozScriptSizeMultiplier;

pub use crate::values::specified::font::FontPalette;

pub use crate::values::specified::font::FontSynthesis;

pub use crate::values::specified::font::FontVariantAlternates;

pub use crate::values::specified::font::FontVariantEastAsian;

pub use crate::values::specified::font::FontVariantLigatures;

pub use crate::values::specified::font::FontVariantNumeric;

pub use crate::values::specified::font::XLang;

pub use crate::values::specified::font::XTextScale;

pub use crate::values::specified::Integer as SpecifiedInteger;

pub use crate::values::specified::Number as SpecifiedNumber;

Macros

static_font_family 🔒

Structs

FamilyName

The name of a font family of choice.

FixedPoint

Generic template for font property type classes that use a fixed-point internal representation with FRACTION_BITS for the fractional part.

FontFamily

Specifies a prioritized list of font family names or generic family names.

FontFamilyList

A list of font families.

FontLanguageOverride

font-language-override can only have a single 1-4 ASCII character OpenType “language system” tag, so we should be able to compute it and store it as a 32-bit integer (see http://www.microsoft.com/typography/otspec/languagetags.htm).

FontSize

The computed value of font-size

FontStretch

A value for the font-stretch property per:

FontStyle

The computed value of font-style.

FontWeight

A value for the font-weight property per:

Enums

FontFamilyNameSyntax

Font family names must either be given quoted as strings, or unquoted as a sequence of one or more identifiers.

GenericFontFamily

A generic font-family name.

SingleFontFamily

A set of faces that vary in weight, width or slope. cbindgen:derive-mut-casts=true

Constants

FONT_STRETCH_FRACTION_BITS

font-stretch is a percentage relative to normal.

FONT_STYLE_FRACTION_BITS

Use a signed 8.8 fixed-point value (representable range -128.0..128)

FONT_WEIGHT_FRACTION_BITS

font-weight: range 1..1000, fractional values permitted; keywords ‘normal’, ‘bold’ aliased to 400, 700 respectively.

Functions

dedup_font_settings 🔒

system_ui_enabled 🔒

Type Aliases

FontFeatureSettings

Use FontSettings as computed type of FontFeatureSettings.

FontSizeAdjust

Preserve the readability of text when font fallback occurs.

FontSizeAdjustFactor

A factor for one of the font-size-adjust metrics, which may be either a number or the from-font keyword.

FontStretchFixedPoint

This is an alias which is useful mostly as a cbindgen / C++ inference workaround.

FontStyleFixedPoint

This is an alias which is useful mostly as a cbindgen / C++ inference workaround.

FontVariationSettings

The computed value for font-variation-settings.

FontWeightFixedPoint

This is an alias which is useful mostly as a cbindgen / C++ inference workaround.

MathDepth

The computed value of the math-depth property.