Module style::values::computed::font

Computed values for font properties

Reexports

pub use values::computed::Length as MozScriptMinSize;

pub use values::specified::font::XTextZoom;

pub use values::specified::font::XLang;

pub use values::specified::font::MozScriptSizeMultiplier;

pub use values::specified::font::FontSynthesis;

pub use values::specified::font::FontVariantSettings;

Structs

FamilyName
The name of a font family of choice

FontFamily
Specifies a prioritized list of font family names or generic family names.

FontFamilyList
A list of SingleFontFamily

FontLanguageOverride
font-language-override can only have a single three-letter OpenType "language system" tag, so we should be able to compute it and store it as a 32-bit integer (see http://www.microsoft.com/typography/otspec/languagetags.htm).

FontSize
The computed value of font-size

FontWeight
As of CSS Fonts Module Level 3, only the following values are valid: 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900

KeywordInfo
Additional information for keyword-derived font sizes.

Enums

FamilyNameSyntax
Font family names must either be given quoted as strings, or unquoted as a sequence of one or more identifiers.

FontSizeAdjust
Preserve the readability of text when font fallback occurs

SingleFontFamily
A set of faces that vary in weight, width or slope.

Type Definitions

FontFeatureSettings
Use FontSettings as computed type of FontFeatureSettings

FontVariantAlternates
Use VariantAlternatesList as computed type of FontVariantAlternates

FontVariantEastAsian
Use VariantEastAsian as computed type of FontVariantEastAsian

FontVariantLigatures
Use VariantLigatures as computed type of FontVariantLigatures

FontVariantNumeric
Use VariantNumeric as computed type of FontVariantNumeric

MozScriptLevel
The computed value of the -moz-script-level property.