Enum style::values::computed::length::LengthOrPercentage

pub enum LengthOrPercentage {
    Length(Length),
    Percentage(Percentage),
    Calc(CalcLengthOrPercentage),
}

Variants
Length(Length)Percentage(Percentage)Calc(CalcLengthOrPercentage)
Methods

impl LengthOrPercentage

pub fn zero() -> LengthOrPercentage

pub fn one() -> LengthOrPercentage

1px length value for SVG defaults

pub fn is_definitely_zero(&self) -> bool

Returns true if the computed value is absolute 0 or 0%.

(Returns false for calc() values, even if ones that may resolve to zero.)

pub fn to_hash_key(&self) -> (Au, NotNaN<f32>)

pub fn to_used_value(&self, containing_length: Au) -> Au

Returns the used value.

pub fn to_pixel_length(&self, containing_length: Au) -> Length

Returns the used value as CSSPixelLength.

pub fn clamp_to_non_negative(self) -> Self

Returns the clamped non-negative values.