style::values::computed::motion

Type Alias OffsetPathFunction

pub type OffsetPathFunction = GenericOffsetPathFunction<BasicShape, RayFunction, ComputedUrl>;

The computed value of .

Aliased Type

#[repr(C, u8)]
pub enum OffsetPathFunction {
    Ray(GenericRayFunction<Angle, GenericPosition<LengthPercentage, LengthPercentage>>),
    Url(ComputedUrl),
    Shape(GenericBasicShape<Angle, GenericPosition<LengthPercentage, LengthPercentage>, LengthPercentage, GenericInsetRect<LengthPercentage>>),
}

Variants

Ray(GenericRayFunction<Angle, GenericPosition<LengthPercentage, LengthPercentage>>)

ray() function, which defines a path in the polar coordinate system. Use Box<> to make sure the size of offset-path is not too large.

Url(ComputedUrl)

A URL reference to an SVG shape element. If the URL does not reference a shape element, this behaves as path(“m 0 0”) instead.

Shape(GenericBasicShape<Angle, GenericPosition<LengthPercentage, LengthPercentage>, LengthPercentage, GenericInsetRect<LengthPercentage>>)

The value.