style::values::computed

Module position

CSS handling for the computed value of position values.

Re-exports

pub use crate::values::specified::position::AnchorName;

pub use crate::values::specified::position::AnchorScope;

pub use crate::values::specified::position::DashedIdentAndOrTryTactic;

pub use crate::values::specified::position::GridAutoFlow;

pub use crate::values::specified::position::GridTemplateAreas;

pub use crate::values::specified::position::MasonryAutoFlow;

pub use crate::values::specified::position::PositionAnchor;

pub use crate::values::specified::position::PositionArea;

pub use crate::values::specified::position::PositionAreaAxis;

pub use crate::values::specified::position::PositionAreaKeyword;

pub use crate::values::specified::position::PositionAreaType;

pub use crate::values::specified::position::PositionTryFallbacks;

pub use crate::values::specified::position::PositionTryFallbacksTryTactic;

pub use crate::values::specified::position::PositionTryFallbacksTryTacticKeyword;

pub use crate::values::specified::position::PositionTryOrder;

pub use crate::values::specified::position::PositionVisibility;

Traits

TryTacticAdjustment 🔒

Perform the adjustment of a given value for a given try tactic, as per: https://drafts.csswg.org/css-anchor-position-1/#swap-due-to-a-try-tactic

Functions

block_or_inline_to_inferred 🔒

inferred_to_block 🔒

inferred_to_inline 🔒

Type Aliases

AnchorFunction

The computed value of an anchor() function.

AnchorSide

The computed value of anchor side.

AspectRatio

A computed value for the aspect-ratio property.

HorizontalPosition

The computed value of a CSS horizontal position.

Inset

A computed type for inset properties.

Position

The computed value of a CSS <position>

PositionOrAuto

The computed value of an auto | <position>

VerticalPosition

The computed value of a CSS vertical position.

ZIndex

A computed value for the z-index property.