Struct style::values::computed::Context

pub struct Context<'a> {
    pub is_root_element: bool,
    pub viewport_size: Size2D<Au>,
    pub inherited_style: &'a ComputedValues,
    pub style: ComputedValues,
    pub font_metrics_provider: Option<&'a FontMetricsProvider>,
}

Fields
is_root_element: bool viewport_size: Size2D<Au> inherited_style: &'a ComputedValues style: ComputedValues

Values access through this need to be in the properties "computed early": color, text-decoration, font-size, display, position, float, border-*-style, outline-style, font-family, writing-mode...

font_metrics_provider: Option<&'a FontMetricsProvider>
Methods

impl<'a> Context<'a>

fn is_root_element(&self) -> bool

fn viewport_size(&self) -> Size2D<Au>

fn inherited_style(&self) -> &ComputedValues

fn style(&self) -> &ComputedValues

fn mutate_style(&mut self) -> &mut ComputedValues

fn initial(viewport_size: Size2D<Au>, is_root_element: bool) -> Self

Creates a dummy computed context for use in multiple places, like evaluating media queries.