Struct style::values::computed::Context

pub struct Context<'a> {
    pub is_root_element: bool,
    pub viewport_size: Size2D<Au>,
    pub inherited_style: &'a ComputedValues,
    pub style: ComputedValues,
    pub font_metrics_provider: Option<&'a FontMetricsProvider>,
}

A Context is all the data a specified value could ever need to compute itself and be transformed to a computed value.

Fields
is_root_element: bool

Whether the current element is the root element.

viewport_size: Size2D<Au>

The current viewport size.

inherited_style: &'a ComputedValues

The style we're inheriting from.

style: ComputedValues

Values access through this need to be in the properties "computed early": color, text-decoration, font-size, display, position, float, border-*-style, outline-style, font-family, writing-mode...

font_metrics_provider: Option<&'a FontMetricsProvider>

A font metrics provider, used to access font metrics to implement font-relative units.

TODO(emilio): This should be required, see #14079.

Methods

impl<'a> Context<'a>

fn is_root_element(&self) -> bool

Whether the current element is the root element.

fn viewport_size(&self) -> Size2D<Au>

The current viewport size.

fn inherited_style(&self) -> &ComputedValues

The style we're inheriting from.

fn style(&self) -> &ComputedValues

The current style. Note that only "eager" properties should be accessed from here, see the comment in the member.

fn mutate_style(&mut self) -> &mut ComputedValues

A mutable reference to the current style.