Struct style::values::computed::Context

pub struct Context<'a> {
    pub builder: StyleBuilder<'a>,
    pub cached_system_font: Option<()>,
    pub in_media_query: bool,
    pub in_container_query: bool,
    pub quirks_mode: QuirksMode,
    pub for_smil_animation: bool,
    pub container_info: Option<ContainerInfo>,
    pub for_non_inherited_property: bool,
    pub rule_cache_conditions: RefCell<&'a mut RuleCacheConditions>,
    container_size_query: RefCell<ContainerSizeQuery<'a>>,
}

A Context is all the data a specified value could ever need to compute itself and be transformed to a computed value.

Fields
builder: StyleBuilder<'a>

Values accessed through this need to be in the properties “computed early”: color, text-decoration, font-size, display, position, float, border-*-style, outline-style, font-family, writing-mode…

cached_system_font: Option<()>

A dummy option for servo so initializing a computed::Context isn’t painful.

TODO(emilio): Make constructors for Context, and drop this.

in_media_query: bool

Whether or not we are computing the media list in a media query.

in_container_query: bool

Whether or not we are computing the container query condition.

quirks_mode: QuirksMode

The quirks mode of this context.

for_smil_animation: bool

Whether this computation is being done for a SMIL animation.

This is used to allow certain properties to generate out-of-range values, which SMIL allows.

container_info: Option<ContainerInfo>

Returns the container information to evaluate a given container query.

for_non_inherited_property: bool

Whether we’re computing a value for a non-inherited property. False if we are computed a value for an inherited property or not computing for a property at all (e.g. in a media query evaluation).

rule_cache_conditions: RefCell<&'a mut RuleCacheConditions>

The conditions to cache a rule node on the rule cache.

FIXME(emilio): Drop the refcell.

container_size_query: RefCell<ContainerSizeQuery<'a>>

Container size query for this context.

Implementations

impl<'a> Context<'a>

pub fn get_container_size_query(&self) -> ContainerSizeQueryResult

Lazily evaluate the container size query, returning the result.

pub fn for_media_query_evaluation<F, R>( device: &Device, quirks_mode: QuirksMode, f: F ) -> Rwhere F: FnOnce(&Context<'_>) -> R,

Creates a suitable context for media query evaluation, in which font-relative units compute against the system_font, and executes f with it.

pub fn for_container_query_evaluation<F, R>( device: &Device, stylist: Option<&Stylist>, container_info_and_style: Option<(ContainerInfo, Arc<ComputedValues>)>, container_size_query: ContainerSizeQuery<'_>, f: F ) -> Rwhere F: FnOnce(&Context<'_>) -> R,

Creates a suitable context for container query evaluation for the style specified.

pub fn new( builder: StyleBuilder<'a>, quirks_mode: QuirksMode, rule_cache_conditions: &'a mut RuleCacheConditions, container_size_query: ContainerSizeQuery<'a> ) -> Self

Creates a context suitable for more general cases.

pub fn new_for_animation( builder: StyleBuilder<'a>, for_smil_animation: bool, quirks_mode: QuirksMode, rule_cache_conditions: &'a mut RuleCacheConditions, container_size_query: ContainerSizeQuery<'a> ) -> Self

Creates a context suitable for computing animations.

pub fn device(&self) -> &Device

The current device.

pub fn query_font_metrics( &self, base_size: FontBaseSize, orientation: FontMetricsOrientation, retrieve_math_scales: bool ) -> FontMetrics

Queries font metrics.

pub fn viewport_size_for_viewport_unit_resolution( &self, variant: ViewportVariant ) -> Size2D<Au>

The current viewport size, used to resolve viewport units.

pub fn in_media_or_container_query(&self) -> bool

Whether we’re in a media or container query.

pub fn default_style(&self) -> &ComputedValues

The default computed style we’re getting our reset style from.

pub fn style(&self) -> &StyleBuilder<'_>

The current style.

pub fn maybe_zoom_text(&self, size: CSSPixelLength) -> CSSPixelLength

(Servo doesn’t do text-zoom)