Module style::values::computed::text

Computed types for text properties.

Re-exports

pub use crate::values::specified::text::TextAlignLast;

pub use crate::values::specified::text::TextUnderlinePosition;

pub use crate::values::specified::LineBreak;

pub use crate::values::specified::OverflowWrap;

pub use crate::values::specified::WordBreak;

pub use crate::values::specified::TextDecorationLine;

pub use crate::values::specified::TextEmphasisPosition;

pub use crate::values::specified::TextDecorationSkipInk;

pub use crate::values::specified::TextTransform;

Structs

LetterSpacing
A computed value for the letter-spacing property.

TextDecorationsInEffect
A struct that represents the used value of the text-decoration property.

TextOverflow
text-overflow. When the specified value only has one side, that's the "second" side, and the sides are logical, so "second" means "end". The start side is Clip in that case.

Enums

TextEmphasisStyle
Computed value for the text-emphasis-style property

Type Definitions

InitialLetter
A computed value for the initial-letter property.

LineHeight
A computed value for the line-height property.

TextAlign
The computed value of text-align.

TextDecorationLength
Implements type for text-decoration-thickness property.

WordSpacing
A computed value for the word-spacing property.