Type Definition style::values::computed::ClipRect

type ClipRect = GenericClipRect<LengthOrAuto>;

rect(...) | auto

Implementations

impl ClipRect

pub fn for_border_rect<T: Copy + From<Length> + Add<Output = T> + Sub<Output = T>, U>(
&self,
border_box: Rect<T, U>
) -> Rect<T, U>

Given a border box, resolves the clip rect against the border box in the same space the border box is in