Module style::values::generics

Generic types that share their serialization implementations for both specified and computed values.

Re-exports

pub use self::GenericClipRect as ClipRect;

pub use self::GenericClipRectOrAuto as ClipRectOrAuto;

Modules

background
Generic types for CSS values related to backgrounds.

basic_shape
CSS handling for the basic-shape types that are generic over their ToCss implementations.

border
Generic types for CSS values related to borders.

box_
Generic types for box properties.

calc
Calc expressions.

color
Generic types for color properties.

column
Generic types for the column properties.

counters
Generic types for counters-related CSS values.

easing
Generic types for CSS Easing Functions. https://drafts.csswg.org/css-easing/#timing-functions

effects
Generic types for CSS values related to effects.

flex
Generic types for CSS values related to flexbox.

font
Generic types for font stuff.

grid
Generic types for the handling of grids.

image
Generic types for the handling of images.

length
Generic types for CSS values related to length.

motion
Generic types for CSS Motion Path.

position
Generic types for CSS handling of specified and computed values of position

rect
Generic types for CSS values that are composed of four sides.

size
Generic type for CSS properties that are composed by two dimensions.

svg
Generic types for CSS values in SVG

text
Generic types for text properties.

transform
Generic types for CSS values that are related to transformations.

ui
Generic values for UI properties.

url
Generic types for url properties.

Structs

GenericClipRect
A clip rect for clip and image-region

GreaterThanOrEqualToOne
A wrapper of greater-than-or-equal-to-one values.

NonNegative
A wrapper of Non-negative values.

ZeroToOne
A wrapper of values between zero and one.

Enums

CounterStyle
https://drafts.csswg.org/css-counter-styles/#typedef-counter-style

GenericClipRectOrAuto
Either a clip-rect or auto.

SymbolsType
https://drafts.csswg.org/css-counter-styles/#typedef-symbols-type

Functions

is_symbolic