Module style::values::generics

Generic types that share their serialization implementations for both specified and computed values.

Modules

background
Generic types for CSS values related to backgrounds.

basic_shape
CSS handling for the basic-shape types that are generic over their ToCss implementations.

border
Generic types for CSS values related to borders.

box_
Generic types for box properties.

effects
Generic types for CSS values related to effects.

flex
Generic types for CSS values related to flexbox.

grid
Generic types for the handling of grids.

image
Generic types for the handling of images.

position
Generic types for CSS handling of specified and computed values of position

rect
Generic types for CSS values that are composed of four sides.

size
Generic type for CSS properties that are composed by two dimensions.

svg
Generic types for CSS values in SVG

text
Generic types for text properties.

transform
Generic types for CSS values that are related to transformations.

Structs

FontSettingTag
A settings tag, defined by a four-character tag and a setting value

FontSettingTagFloat
A number value to be used for font-variation-settings

FontSettingTagInt
An integer that can also parse "on" and "off", for font-feature-settings

GreaterThanOrEqualToOne
A wrapper of greater-than-or-equal-to-one values.

NonNegative
A wrapper of Non-negative values.

Enums

CounterStyleOrNone
https://drafts.csswg.org/css-counter-styles/#typedef-counter-style

FontSettings
A font settings value for font-variation-settings or font-feature-settings

SymbolsType