Module style::values::generics::basic_shape

CSS handling for the basic-shape types that are generic over their ToCss implementations.

Re-exports

pub use self::GenericClipPath as ClipPath;

pub use self::GenericShapeOutside as ShapeOutside;

pub use self::GenericBasicShape as BasicShape;

pub use self::GenericInsetRect as InsetRect;

pub use self::GenericShapeRadius as ShapeRadius;

pub use self::GenericPolygon as Polygon;

pub use self::GenericShapeCommand as ShapeCommand;

Structs

Circle

https://drafts.csswg.org/css-shapes/#funcdef-circle

CoordinatePair

Defines a pair of coordinates, representing a rightward and downward offset, respectively, from a specified reference point. Percentages are resolved against the width or height, respectively, of the reference box. https://drafts.csswg.org/css-shapes-2/#typedef-shape-coordinate-pair

Ellipse

https://drafts.csswg.org/css-shapes/#funcdef-ellipse

GenericInsetRect

https://drafts.csswg.org/css-shapes/#funcdef-inset

GenericPolygon

A generic type for representing the polygon() function

Path

The path function.

PolygonCoord

Coordinates for Polygon.

Shape

The shape function defined in css-shape-2. shape() = shape(? from , #)

Enums

ArcSize

This indicates that the larger or smaller, respectively, of the two possible arcs must be chosen. https://drafts.csswg.org/css-shapes-2/#typedef-shape-arc-size

ArcSweep

This indicates that the arc that is traced around the ellipse clockwise or counter-clockwise from the center. https://drafts.csswg.org/css-shapes-2/#typedef-shape-arc-sweep

ByTo

This indicates the command is absolute or relative. https://drafts.csswg.org/css-shapes-2/#typedef-shape-by-to

FillRule

GenericBasicShape

The .

GenericClipPath

A value for the clip-path property.

GenericPathOrShapeFunction

path() function or shape() function.

GenericShapeCommand

This is a more general shape(path) command type, for both shape() and path().

GenericShapeOutside

A value for the shape-outside property.

GenericShapeRadius

https://drafts.csswg.org/css-shapes/#typedef-shape-radius

ShapeBox

https://drafts.csswg.org/css-shapes-1/#typedef-shape-box

ShapeGeometryBox

https://drafts.fxtf.org/css-masking-1/#typedef-geometry-box

Functions

is_default 🔒

is_default_box_for_clip_path 🔒

Skip the serialization if the author omits the box or specifies border-box.