style::values::generics

Module calc

Calc expressions.

Re-exports

pub use self::GenericCalcNode as CalcNode;

Macros

compare_helpers 🔒

Structs

CalcUnits

Expected units we allow parsing within a calc() expression.

Enums

ArgumentLevel 🔒

The level of any argument being serialized in to_css_impl.

GenericCalcNode

A generic node in a calc expression.

MinMaxOp

Whether we’re a min or max function.

ModRemOp

Whether we’re a mod or rem function.

PositivePercentageBasis

For percentage resolution, sometimes we can’t assume that the percentage basis is positive (so we don’t know whether a percentage is larger than another).

RoundingStrategy

The strategy used in round()

SortKey

This determines the order in which we serialize members of a calc() sum.

Traits

CalcNodeLeaf

A trait that represents all the stuff a valid leaf of a calc expression.

Type Aliases

GenericCalcAnchorFunction

anchor() function used in math functions.

GenericCalcAnchorSizeFunction

anchor-size() function used in math functions.