Enum style::values::generics::calc::GenericCalcNode

#[repr(u8)]
pub enum GenericCalcNode<L> {
    Leaf(L),
    Sum(OwnedSlice<GenericCalcNode<L>>),
    MinMax(OwnedSlice<GenericCalcNode<L>>, MinMaxOp),
    Clamp {
        min: Box<GenericCalcNode<L>>,
        center: Box<GenericCalcNode<L>>,
        max: Box<GenericCalcNode<L>>,
    },
}

A generic node in a calc expression.

FIXME: This would be much more elegant if we used Self in the types below, but we can’t because of https://github.com/serde-rs/serde/issues/1565.

FIXME: The following annotations are to workaround an LLVM inlining bug, see bug 1631929.

cbindgen:destructor-attributes=MOZ_NEVER_INLINE cbindgen:copy-constructor-attributes=MOZ_NEVER_INLINE cbindgen:eq-attributes=MOZ_NEVER_INLINE

Variants

Leaf(L)

A leaf node.

Sum(OwnedSlice<GenericCalcNode<L>>)

A sum node, representing a + b + c where a, b, and c are the arguments.

MinMax(OwnedSlice<GenericCalcNode<L>>, MinMaxOp)

A min or max function.

Clamp

Fields

min: Box<GenericCalcNode<L>>

The minimum value.

center: Box<GenericCalcNode<L>>

The central value.

max: Box<GenericCalcNode<L>>

The maximum value.

A clamp() function.

Implementations

impl<L: CalcNodeLeaf> CalcNode<L>

pub fn negate(&mut self)

Negates the node.

fn sort_key(&self) -> SortKey

fn try_sum_in_place(&mut self, other: &Self) -> Result<(), ()>

Tries to merge one sum to another, that is, perform x + y.

pub fn map_leaves<O, F>(&self, map: F) -> CalcNode<O> where
O: CalcNodeLeaf,
F: FnMut(&L) -> O,

Convert this CalcNode into a CalcNode with a different leaf kind.

fn map_leaves_internal<O, F>(&self, map: &mut F) -> CalcNode<O> where
O: CalcNodeLeaf,
F: FnMut(&L) -> O,

pub fn resolve<O>(
&self,
leaf_to_output_fn: impl FnMut(&L) -> Result<O, ()>
) -> Result<O, ()> where
O: PartialOrd + PartialEq + Add<Output = O> + Zero,

Resolves the expression returning a value of O, given a function to turn a leaf into the relevant value.

fn resolve_internal<O, F>(&self, leaf_to_output_fn: &mut F) -> Result<O, ()> where
O: PartialOrd + PartialEq + Add<Output = O> + Zero,
F: FnMut(&L) -> Result<O, ()>,

fn is_negative_leaf(&self) -> bool

pub fn mul_by(&mut self, scalar: f32)

Multiplies the node by a scalar.

pub fn visit_depth_first(&mut self, f: impl FnMut(&mut Self))

Visits all the nodes in this calculation tree recursively, starting by the leaves and bubbling all the way up.

This is useful for simplification, but can also be used for validation and such.

fn visit_depth_first_internal(&mut self, f: &mut impl FnMut(&mut Self))

pub fn simplify_and_sort_direct_children(&mut self)

Simplifies and sorts the calculation of a given node. All the nodes below it should be simplified already, this only takes care of simplifying directly nested nodes. So, probably should always be used in combination with visit_depth_first().

This is only needed if it’s going to be preserved after parsing (so, for <length-percentage>). Otherwise we can just evaluate it using resolve(), and we’ll come up with a simplified value anyways.

pub fn simplify_and_sort(&mut self)

Simplifies and sorts the kids in the whole calculation subtree.

fn to_css_impl<W>(
&self,
dest: &mut CssWriter<'_, W>,
is_outermost: bool
) -> Result where
W: Write,

impl GenericCalcNode<Leaf>

fn parse_one<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

Tries to parse a single element in the expression, that is, a <length>, <angle>, <time>, <percentage>, according to expected_unit.

May return a “complex” CalcNode, in the presence of a parenthesized expression, for example.

fn parse<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

Parse a top-level calc expression, with all nested sub-expressions.

This is in charge of parsing, for example, 2 + 3 * 100%.

fn parse_argument<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

fn parse_product<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

Parse a top-level calc expression, and all the products that may follow, and stop as soon as a non-product expression is found.

This should parse correctly:

2

2 * 2

2 * 2 + 2 (but will leave the + 2 unparsed).

fn into_length_or_percentage(
self,
clamping_mode: AllowedNumericType
) -> Result<CalcLengthPercentage, ()>

Tries to simplify this expression into a <length> or <percentage> value.

fn to_time(&self) -> Result<Time, ()>

Tries to simplify this expression into a <time> value.

fn to_angle(&self) -> Result<Angle, ()>

Tries to simplify this expression into an Angle value.

fn to_number(&self) -> Result<CSSFloat, ()>

Tries to simplify this expression into a <number> value.

fn to_percentage(&self) -> Result<CSSFloat, ()>

Tries to simplify this expression into a <percentage> value.

pub fn math_function<'i>(
name: &CowRcStr<'i>,
location: SourceLocation
) -> Result<MathFunction, ParseError<'i>>

Given a function name, and the location from where the token came from, return a mathematical function corresponding to that name or an error.

pub fn parse_integer<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<CSSInteger, ParseError<'i>>

Convenience parsing function for integers.

pub fn parse_length_or_percentage<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
clamping_mode: AllowedNumericType,
function: MathFunction
) -> Result<CalcLengthPercentage, ParseError<'i>>

Convenience parsing function for <length> | <percentage>.

pub fn parse_percentage<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<CSSFloat, ParseError<'i>>

Convenience parsing function for percentages.

pub fn parse_length<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
clamping_mode: AllowedNumericType,
function: MathFunction
) -> Result<CalcLengthPercentage, ParseError<'i>>

Convenience parsing function for <length>.

pub fn parse_number<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<CSSFloat, ParseError<'i>>

Convenience parsing function for <number>.

pub fn parse_angle<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<Angle, ParseError<'i>>

Convenience parsing function for <angle>.

pub fn parse_time<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<Time, ParseError<'i>>

Convenience parsing function for <time>.

pub fn parse_number_or_percentage<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<NumberOrPercentage, ParseError<'i>>

Convenience parsing function for <number> or <percentage>.

pub fn parse_angle_or_number<'i, 't>(
context: &ParserContext<'_>,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<AngleOrNumber, ParseError<'i>>

Convenience parsing function for <number> or <angle>.