Trait style::values::generics::calc::CalcNodeLeaf

pub trait CalcNodeLeaf: Clone + Sized + PartialOrd + PartialEq + ToCss {
    fn is_negative(&self) -> bool;

    fn try_sum_in_place(&mut self, other: &Self) -> Result<(), ()>;

    fn mul_by(&mut self, scalar: f32);

    fn simplify(&mut self);

    fn sort_key(&self) -> SortKey;

    fn negate(&mut self) { ... }
}

A trait that represents all the stuff a valid leaf of a calc expression.

Required methods

fn is_negative(&self) -> bool

Whether this value is known-negative.

fn try_sum_in_place(&mut self, other: &Self) -> Result<(), ()>

Tries to merge one sum to another, that is, perform x + y.

fn mul_by(&mut self, scalar: f32)

Multiplies the leaf by a given scalar number.

fn simplify(&mut self)

Canonicalizes the expression if necessary.

fn sort_key(&self) -> SortKey

Returns the sort key for simplification.

Loading content...
Provided methods

fn negate(&mut self)

Negates the leaf.

Loading content...