style::values::generics

Module grid

Generic types for the handling of grids.

Re-exports

pub use self::GenericGridLine as GridLine;

pub use self::GenericTrackBreadth as TrackBreadth;

pub use self::GenericTrackSize as TrackSize;

pub use self::GenericImplicitGridTracks as ImplicitGridTracks;

pub use self::GenericTrackRepeat as TrackRepeat;

pub use self::GenericTrackListValue as TrackListValue;

pub use self::GenericTrackList as TrackList;

pub use self::GenericNameRepeat as NameRepeat;

pub use self::GenericLineNameListValue as LineNameListValue;

pub use self::GenericLineNameList as LineNameList;

pub use self::GenericGridTemplateComponent as GridTemplateComponent;

Structs

GenericGridLine

A <grid-line> type.

GenericImplicitGridTracks

A <track-size>+. We use the empty slice as auto, and always parse auto as an empty slice. This means it’s impossible to have a slice containing only one auto item.

GenericLineNameList

The <line-name-list> for subgrids.

GenericNameRepeat

The <name-repeat> for subgrids.

GenericTrackList

A grid <track-list> type.

GenericTrackRepeat

The structure containing <line-names> and <track-size> values.

Enums

GenericGridTemplateComponent

Variants for <grid-template-rows> | <grid-template-columns>

GenericLineNameListValue

A single value for <line-names> or <name-repeat>.

GenericTrackBreadth

A track breadth for explicit grid track sizing. It’s generic solely to avoid re-implementing it for the computed type.

GenericTrackListValue

Track list values. Can be or

GenericTrackSize

A <track-size> type for explicit grid track sizing. Like <track-breadth>, this is generic only to avoid code bloat. It only takes <length-percentage>

RepeatCount

The initial argument of the repeat function.

Constants

MAX_GRID_LINE

See above.

MIN_GRID_LINE

These are the limits that we choose to clamp grid line numbers to. http://drafts.csswg.org/css-grid/#overlarge-grids line_num is clamped to this range at parse time.

Functions

concat_serialize_idents

Helper function for serializing identifiers with a prefix and suffix, used for serializing (in grid).