Module style::values::generics::grid

Generic types for the handling of grids.

Structs

GridLine
A <grid-line> type.

LineNameList
The <line-name-list> for subgrids.

TrackList
A grid <track-list> type.

TrackRepeat
The structure containing <line-names> and <track-size> values.

Enums

GridTemplateComponent
Variants for <grid-template-rows> | <grid-template-columns> Subgrid deferred to Level 2 spec due to lack of implementation. But it's implemented in gecko, so we have to as well.

RepeatCount
The initial argument of the repeat function.

TrackBreadth
A track breadth for explicit grid track sizing. It's generic solely to avoid re-implementing it for the computed type.

TrackKeyword

TrackListType
The type of a <track-list> as determined during parsing.

TrackListValue
Track list values. Can be or

TrackSize
A <track-size> type for explicit grid track sizing. Like <track-breadth>, this is generic only to avoid code bloat. It only takes <length-percentage>

Functions

concat_serialize_idents
Helper function for serializing identifiers with a prefix and suffix, used for serializing (in grid).