Module style::values::generics::image

Generic types for the handling of images.

Re-exports

pub use self::GenericImage as Image;

pub use self::GenericCrossFade as CrossFade;

pub use self::GenericCrossFadeElement as CrossFadeElement;

pub use self::GenericCrossFadeImage as CrossFadeImage;

pub use self::GenericImageSet as ImageSet;

pub use self::GenericImageSetItem as ImageSetItem;

pub use self::GenericGradient as Gradient;

pub use self::GenericEndingShape as EndingShape;

pub use self::GenericCircle as Circle;

pub use self::GenericEllipse as Ellipse;

pub use self::GenericGradientItem as GradientItem;

Structs

ColorStop

A color stop. https://drafts.csswg.org/css-images/#typedef-color-stop-list

GenericCrossFade

https://drafts.csswg.org/css-images-4/#cross-fade-function

GenericCrossFadeElement

An optional percent and a cross fade image.

GenericImageSet

https://drafts.csswg.org/css-images-4/#image-set-notation

GenericImageSetItem

An optional percent and a cross fade image.

GradientFlags

State flags stored on each variant of a Gradient.

PaintWorklet

Specified values for a paint worklet. https://drafts.css-houdini.org/css-paint-api/

Enums

GenericCircle

A circle shape.

GenericCrossFadeImage

An image or a color. cross-fade takes either when blending images together.

GenericEllipse

An ellipse shape.

GenericEndingShape

A radial gradient’s ending shape.

GenericGradient

A CSS gradient. https://drafts.csswg.org/css-images/#gradients

GenericGradientItem

A gradient item. https://drafts.csswg.org/css-images-4/#color-stop-syntax

GenericImage

An <image> | none value.

GradientCompatMode

Whether we used the modern notation or the compatibility -webkit, -moz prefixes.

ShapeExtent

https://drafts.csswg.org/css-images/#typedef-extent-keyword

Traits

LineDirection

The direction of a linear gradient.