Module style::values::generics::position

Generic types for CSS handling of specified and computed values of position

Re-exports

pub use self::GenericPosition as Position;

pub use self::GenericPositionOrAuto as PositionOrAuto;

pub use self::GenericZIndex as ZIndex;

pub use self::GenericAspectRatio as AspectRatio;

Structs

GenericAspectRatio
A generic value for the aspect-ratio property, the value is auto || <ratio>.

GenericPosition
A generic type for representing a CSS position.

Ratio
A generic value for the <ratio> value.

Enums

GenericPositionOrAuto
A generic type for representing an Auto | <position>. This is used by for now. https://drafts.fxtf.org/motion-1/#offset-anchor-property cbindgen:private-default-tagged-enum-constructor=false

GenericZIndex
A generic value for the z-index property.

PreferredRatio
Ratio or None.

Traits

PositionComponent
Implements a method that checks if the position is centered.