Module style::values::generics::transform

Generic types for CSS values that are related to transformations.

Structs

Matrix
A generic 2D transformation matrix.

Matrix3D

Transform
A value of the transform property

TransformOrigin
A generic transform origin.

Enums

StepPosition

TimingFunction
A generic timing function.

TimingKeyword

TransformOperation
A single operation in the list of a transform value

Traits

ToAbsoluteLength
Convert a length type into the absolute lengths.

ToMatrix
Support the conversion to a 3d matrix.

Functions

create_perspective_matrix
Return the transform matrix from a perspective length.

get_normalized_vector_and_angle
Return the normalized direction vector and its angle for Rotate3D.