Module style::values::specified

Structs

Angle

BorderRadiusSize

CSSColor

CSSRGBA

CalcLengthOrPercentage

CalcProductNode

CalcSumNode

CharacterWidth

ColorStop
Specified values for one color stop in a linear gradient.

LinearGradient
Specified values for a CSS linear gradient.

Number

Opacity

Percentage

SimplifiedSumNode

Time
A time in seconds according to CSS-VALUES  6.2.

Enums

AngleOrCorner
Specified values for an angle or a corner in a linear gradient.

BorderStyle

CalcUnit

CalcValueNode

FontRelativeLength

HorizontalDirection

Image
Specified values for an image according to CSS-IMAGES.

Length

LengthOrNone

LengthOrPercentage

LengthOrPercentageOrAuto

LengthOrPercentageOrAutoOrContent

LengthOrPercentageOrNone

PositionComponent

SimplifiedValueNode

VerticalDirection

ViewportPercentageLength

Constants

AU_PER_CM

AU_PER_IN

AU_PER_MM

AU_PER_PC

AU_PER_PT

AU_PER_PX

AU_PER_Q

RAD_PER_DEG

RAD_PER_GRAD

RAD_PER_TURN

_IMPL_DESERIALIZE_FOR_Angle

_IMPL_DESERIALIZE_FOR_BorderStyle

_IMPL_DESERIALIZE_FOR_HorizontalDirection

_IMPL_DESERIALIZE_FOR_VerticalDirection

_IMPL_SERIALIZE_FOR_Angle

_IMPL_SERIALIZE_FOR_BorderStyle

_IMPL_SERIALIZE_FOR_HorizontalDirection

_IMPL_SERIALIZE_FOR_VerticalDirection

Functions

parse_border_radius

parse_border_width

parse_integer

parse_number

parse_one_color_stop