Module style::values::specified

Specified values.

TODO(emilio): Enhance docs.

Re-exports

pub use self::angle::AllowUnitlessZeroAngle;

pub use self::angle::Angle;

pub use self::background::BackgroundRepeat;

pub use self::background::BackgroundSize;

pub use self::basic_shape::FillRule;

pub use self::border::BorderCornerRadius;

pub use self::border::BorderImageSlice;

pub use self::border::BorderImageWidth;

pub use self::border::BorderImageRepeat;

pub use self::border::BorderImageSideWidth;

pub use self::border::BorderRadius;

pub use self::border::BorderSideWidth;

pub use self::border::BorderSpacing;

pub use self::border::BorderStyle;

pub use self::box_::AnimationIterationCount;

pub use self::box_::AnimationName;

pub use self::box_::Contain;

pub use self::box_::Display;

pub use self::box_::Appearance;

pub use self::box_::BreakBetween;

pub use self::box_::BreakWithin;

pub use self::box_::Clear;

pub use self::box_::Float;

pub use self::box_::Overflow;

pub use self::box_::OverflowAnchor;

pub use self::box_::OverflowClipBox;

pub use self::box_::OverscrollBehavior;

pub use self::box_::Perspective;

pub use self::box_::Resize;

pub use self::box_::ScrollSnapAlign;

pub use self::box_::ScrollSnapAxis;

pub use self::box_::ScrollSnapStrictness;

pub use self::box_::ScrollSnapType;

pub use self::box_::TouchAction;

pub use self::box_::TransitionProperty;

pub use self::box_::VerticalAlign;

pub use self::box_::WillChange;

pub use self::color::Color;

pub use self::color::ColorOrAuto;

pub use self::color::ColorPropertyValue;

pub use self::column::ColumnCount;

pub use self::counters::Content;

pub use self::counters::ContentItem;

pub use self::counters::CounterIncrement;

pub use self::counters::CounterSetOrReset;

pub use self::easing::TimingFunction;

pub use self::effects::BoxShadow;

pub use self::effects::Filter;

pub use self::effects::SimpleShadow;

pub use self::flex::FlexBasis;

pub use self::font::FontFamily;

pub use self::font::FontLanguageOverride;

pub use self::font::FontStyle;

pub use self::font::FontFeatureSettings;

pub use self::font::FontVariantLigatures;

pub use self::font::FontVariantNumeric;

pub use self::font::FontSize;

pub use self::font::FontSizeAdjust;

pub use self::font::FontSizeKeyword;

pub use self::font::FontStretch;

pub use self::font::FontSynthesis;

pub use self::font::FontVariantAlternates;

pub use self::font::FontWeight;

pub use self::font::FontVariantEastAsian;

pub use self::font::FontVariationSettings;

pub use self::font::MathDepth;

pub use self::font::MozScriptMinSize;

pub use self::font::MozScriptSizeMultiplier;

pub use self::font::XLang;

pub use self::font::XTextZoom;

pub use self::image::EndingShape as GradientEndingShape;

pub use self::image::Gradient;

pub use self::image::Image;

pub use self::image::MozImageRect;

pub use self::length::AbsoluteLength;

pub use self::length::CalcLengthPercentage;

pub use self::length::CharacterWidth;

pub use self::length::FontRelativeLength;

pub use self::length::Length;

pub use self::length::LengthOrNumber;

pub use self::length::NonNegativeLengthOrNumber;

pub use self::length::LengthOrAuto;

pub use self::length::LengthPercentage;

pub use self::length::LengthPercentageOrAuto;

pub use self::length::MaxSize;

pub use self::length::Size;

pub use self::length::NoCalcLength;

pub use self::length::ViewportPercentageLength;

pub use self::length::NonNegativeLength;

pub use self::length::NonNegativeLengthPercentage;

pub use self::length::NonNegativeLengthPercentageOrAuto;

pub use self::list::MozListReversed;

pub use self::list::Quotes;

pub use self::motion::OffsetPath;

pub use self::motion::OffsetRotate;

pub use self::outline::OutlineStyle;

pub use self::page::Orientation;

pub use self::page::PageSize;

pub use self::page::PaperSize;

pub use self::percentage::Percentage;

pub use self::position::AspectRatio;

pub use self::position::GridAutoFlow;

pub use self::position::GridTemplateAreas;

pub use self::position::MasonryAutoFlow;

pub use self::position::Position;

pub use self::position::PositionOrAuto;

pub use self::position::PositionComponent;

pub use self::position::ZIndex;

pub use self::ratio::Ratio;

pub use self::rect::NonNegativeLengthOrNumberRect;

pub use self::resolution::Resolution;

pub use self::svg::MozContextProperties;

pub use self::svg::SVGLength;

pub use self::svg::SVGOpacity;

pub use self::svg::SVGPaint;

pub use self::svg::SVGPaintOrder;

pub use self::svg::SVGStrokeDashArray;

pub use self::svg::SVGWidth;

pub use self::svg_path::SVGPathData;

pub use self::text::TextAlignLast;

pub use self::text::TextUnderlinePosition;

pub use self::text::InitialLetter;

pub use self::text::LetterSpacing;

pub use self::text::LineBreak;

pub use self::text::LineHeight;

pub use self::text::TextAlign;

pub use self::text::OverflowWrap;

pub use self::text::TextEmphasisPosition;

pub use self::text::TextEmphasisStyle;

pub use self::text::WordBreak;

pub use self::text::TextAlignKeyword;

pub use self::text::TextDecorationLine;

pub use self::text::TextOverflow;

pub use self::text::WordSpacing;

pub use self::text::TextDecorationLength;

pub use self::text::TextDecorationSkipInk;

pub use self::text::TextTransform;

pub use self::time::Time;

pub use self::transform::Rotate;

pub use self::transform::Scale;

pub use self::transform::Transform;

pub use self::transform::TransformOrigin;

pub use self::transform::TransformStyle;

pub use self::transform::Translate;

pub use self::ui::Cursor;

pub use self::ui::MozForceBrokenImageIcon;

pub use self::ui::UserSelect;

pub use super::generics::grid::GridTemplateComponent as GenericGridTemplateComponent;

Modules

angle

Specified angles.

background

Specified types for CSS values related to backgrounds.

basic_shape

CSS handling for the specified value of basic-shapes

border

Specified types for CSS values related to borders.

box_

Specified types for box properties.

calc

Calc expressions.

color

Specified color values.

column

Specified types for the column properties.

counters

Specified types for counter properties.

easing

Specified types for CSS Easing functions.

effects

Specified types for CSS values related to effects.

flex

Specified types for CSS values related to flexbox.

font

Specified values for font properties

grid

CSS handling for the computed value of grids

image

CSS handling for the specified value of images

length

Length values.

list

list specified values.

motion

Specified types for CSS values that are related to motion path.

outline

Specified values for outline properties

page

Specified @page at-rule properties

percentage

Specified percentages.

position

CSS handling for the specified value of positions

ratio

Specified types for .

rect

Specified types for CSS borders.

resolution

Resolution values:

source_size_list

https://html.spec.whatwg.org/multipage/#source-size-list

svg

Specified types for SVG properties.

svg_path

Specified types for SVG Path.

table

Specified types for CSS values related to tables.

text

Specified types for text properties.

time

Specified time values.

transform

Specified types for CSS values that are related to transformations.

ui

Specified types for UI properties.

url

Common handling for the specified value CSS url() values.

Structs

Attr

An attr(…) rule

Integer

A specified <integer>, optionally coming from a calc() expression.

Number

A CSS <number> specified value.

Opacity

The value of Opacity is , which is “ | ”. However, we serialize the specified value as number, so it’s ok to store the Opacity as Number.

Enums

AllowQuirks

Whether quirks are allowed in this context.

AngleOrPercentage

| https://drafts.csswg.org/css-values/#typedef-angle-percentage

NumberOrPercentage

|

Functions

get_namespace_for_prefix

Get the Namespace for a given prefix from the namespace map.

parse_number_with_clamping_mode

Parse a <number> value, with a given clamping mode.

Type Definitions

ClipRect

rect(…)

ClipRectOrAuto

rect(…) | auto

GreaterThanOrEqualToOneNumber

A Number which is >= 1.0.

GridLine

The specified value of a <grid-line>.

GridTemplateComponent

<grid-template-rows> | <grid-template-columns>

ImplicitGridTracks

The specified value of a grid <track-size>+

NonNegativeNumber

A Number which is >= 0.0.

NonNegativeNumberOrPercentage

A non-negative | .

PositiveInteger

A wrapper of Integer, with value >= 1.

PositiveIntegerOrNone

A specified positive <integer> value or none.

TrackBreadth

The specified value of a grid <track-breadth>

TrackList

The specified value of a grid <track-list> (could also be <auto-track-list> or <explicit-track-list>)

TrackSize

The specified value of a grid <track-size>