Module style::values::specified

Structs

Angle

CSSColor

CSSRGBA

CharacterWidth

ColorStop
Specified values for one color stop in a linear gradient.

LengthAndPercentage
The sum of a series of lengths and a percentage. This is used in calc() and other things that effectively work like it (e.g. transforms).

LinearGradient
Specified values for a CSS linear gradient.

Time
A time in seconds according to CSS-VALUES  6.2.

Enums

AngleOrCorner
Specified values for an angle or a corner in a linear gradient.

BorderStyle

FontRelativeLength

HorizontalDirection

Image
Specified values for an image according to CSS-IMAGES.

Length

LengthOrPercentage

LengthOrPercentageOrAuto

LengthOrPercentageOrNone

PositionComponent

VerticalDirection

ViewportPercentageLength

Functions

parse_border_width