style::values::specified

Module animation

Specified types for properties related to animations and transitions.

Structs

AnimationName

A value for the animation-name property.

ScrollFunction

The scroll() notation. https://drafts.csswg.org/scroll-animations-1/#scroll-notation

TimelineName

The typedef of scroll-timeline-name or view-timeline-name.

ViewTransitionClass

The view-transition-class: none | <custom-ident>+.

Enums

AnimationComposition

https://drafts.csswg.org/css-animations-2/#animation-composition

AnimationDirection

https://drafts.csswg.org/css-animations/#propdef-animation-direction

AnimationFillMode

https://drafts.csswg.org/css-animations/#propdef-animation-fill-mode

AnimationIterationCount

https://drafts.csswg.org/css-animations/#animation-iteration-count

AnimationPlayState

https://drafts.csswg.org/css-animations/#animation-play-state

ScrollAxis

A value for the used in scroll(), or a value for {scroll|view}-timeline-axis.

Scroller

A value for the used in scroll().

TransitionBehavior

A specified value for .

TransitionProperty

A given transition property, that is either All, a longhand or shorthand property, or an unsupported or custom property.

ViewTransitionName

The view-transition-name: none | <custom-ident> | match-element.

Type Aliases

AnimationDuration

A specified value for the animation-duration property.

AnimationTimeline

A specified value for the animation-timeline property.

ViewTimelineInset

A specified value for the view-timeline-inset property.