style::values::specified

Module border

Specified types for CSS values related to borders.

Structs

BorderImageRepeat

The specified value for the border-image-repeat property.

BorderSideOffset

A specified value for outline-offset.

BorderSideWidth

A specified value for a single side of the border-width property. The difference between this and LineWidth is whether we snap to device pixels or not.

Enums

BorderImageRepeatKeyword

A single border-image-repeat keyword.

BorderStyle

A specified value for a single side of a border-style property.

LineWidth

https://drafts.csswg.org/css-backgrounds-3/#typedef-line-width

Functions

serialize_directional_border

Serializes a border shorthand value composed of width/style/color.

snap_as_border_width 🔒

Type Aliases

BorderCornerRadius

A specified value for the border-*-radius longhand properties.

BorderImageSideWidth

A specified value for a single side of a border-image-width property.

BorderImageSlice

A specified value for the border-image-slice property.

BorderImageWidth

A specified value for the border-image-width property.

BorderRadius

A specified value for the border-radius property.

BorderSpacing

A specified value for the border-spacing longhand properties.