Module style::values::specified::border

Specified types for CSS values related to borders.

Structs

BorderImageRepeat
The specified value for the border-image-repeat property.

Enums

BorderImageRepeatKeyword
A single border-image-repeat keyword.

BorderSideWidth
A specified value for a single side of the border-width property.

BorderStyle
A specified value for a single side of a border-style property.

Constants

_IMPL_NUM_FromPrimitive_FOR_BorderStyle

Type Definitions

BorderCornerRadius
A specified value for the border-*-radius longhand properties.

BorderImageSideWidth
A specified value for a single side of a border-image-width property.

BorderImageSlice
A specified value for the border-image-slice property.

BorderImageWidth
A specified value for the border-image-width property.

BorderRadius
A specified value for the border-radius property.

BorderSpacing
A specified value for the border-spacing longhand properties.