Module style::values::specified::box_

Specified types for box properties.

Structs

AnimationName
A value for the animation-name property.

Contain
Constants for contain: https://drafts.csswg.org/css-contain/#contain-property

Display

ScrollSnapAlign
https://drafts.csswg.org/css-scroll-snap-1/#scroll-snap-align

ScrollSnapType
https://drafts.csswg.org/css-scroll-snap-1/#scroll-snap-type

TouchAction
Values for the touch-action property. These constants match Gecko's NS_STYLE_TOUCH_ACTION_* constants.

WillChange
Provides a rendering hint to the user agent, stating what kinds of changes the author expects to perform on the element.

WillChangeBits
The change bits that we care about.

Enums

Appearance
The value for the appearance property.

BreakBetween
A kind of break between two boxes.

BreakWithin
A kind of break within a box.

Clear
https://drafts.csswg.org/css-box/#propdef-clear

DisplayInside

DisplayOutside
Defines an element’s display type, which consists of the two basic qualities of how an element generates boxes https://drafts.csswg.org/css-display/#propdef-display

Float
https://drafts.csswg.org/css-box/#propdef-float

Overflow
The value for the overflow-x / overflow-y properties.

OverflowAnchor

OverflowClipBox

OverscrollBehavior

Resize
https://drafts.csswg.org/css-ui/#propdef-resize

ScrollSnapAlignKeyword
Specified value of scroll-snap-align keyword value.

ScrollSnapAxis
https://drafts.csswg.org/css-scroll-snap-1/#snap-axis

ScrollSnapStrictness
https://drafts.csswg.org/css-scroll-snap-1/#snap-strictness

TransitionProperty
A given transition property, that is either All, a longhand or shorthand property, or an unsupported or custom property.

Constants

_IMPL_NUM_FromPrimitive_FOR_Display

_IMPL_NUM_FromPrimitive_FOR_DisplayInside

_IMPL_NUM_FromPrimitive_FOR_DisplayOutside

_IMPL_NUM_FromPrimitive_FOR_ScrollSnapAlignKeyword

Functions

change_bits_for_longhand

change_bits_for_maybe_property

flexbox_enabled

is_valid_inside_for_list_item
Test a Result for same values as above.

parse_display_inside
= flow | flow-root | table | flex | grid | ruby https://drafts.csswg.org/css-display/#typedef-display-inside

parse_display_inside_for_list_item
(flow | flow-root)?

parse_display_outside
= block | inline | run-in https://drafts.csswg.org/css-display/#typedef-display-outside

parse_list_item
Parse list-item.

Type Definitions

AnimationIterationCount
https://drafts.csswg.org/css-animations/#animation-iteration-count

Perspective
A specified value for the perspective property.

VerticalAlign
A specified value for the vertical-align property.