Module style::values::specified::box_

Specified types for box properties.

Structs

Contain

Constants for contain: https://drafts.csswg.org/css-contain/#contain-property

ContainerName

https://drafts.csswg.org/css-contain-3/#container-name

Display

ScrollSnapAlign

https://drafts.csswg.org/css-scroll-snap-1/#scroll-snap-align

ScrollSnapType

https://drafts.csswg.org/css-scroll-snap-1/#scroll-snap-type

ScrollbarGutter

Values for scrollbar-gutter: https://drafts.csswg.org/css-overflow-3/#scrollbar-gutter-property

TouchAction

Values for the touch-action property.

WillChange

Provides a rendering hint to the user agent, stating what kinds of changes the author expects to perform on the element.

WillChangeBits

The change bits that we care about.

Enums

Appearance

The value for the appearance property.

BaselineSource

A specified value for the baseline-source property. https://drafts.csswg.org/css-inline-3/#baseline-source

BreakBetween

A kind of break between two boxes.

BreakWithin

A kind of break within a box.

Clear

https://drafts.csswg.org/css2/#propdef-clear

ContainerType

https://drafts.csswg.org/css-contain-3/#container-type

ContentVisibility

https://drafts.csswg.org/css-contain-2/#content-visibility

DisplayInside

DisplayKeyword 🔒

DisplayOutside

Defines an element’s display type, which consists of the two basic qualities of how an element generates boxes https://drafts.csswg.org/css-display/#propdef-display

Float

https://drafts.csswg.org/css-box/#propdef-float

Overflow

The value for the overflow-x / overflow-y properties.

OverflowAnchor

OverflowClipBox

OverscrollBehavior

Resize

https://drafts.csswg.org/css-ui/#propdef-resize

ScrollSnapAlignKeyword

Specified value of scroll-snap-align keyword value.

ScrollSnapAxis

https://drafts.csswg.org/css-scroll-snap-1/#snap-axis

ScrollSnapStop

ScrollSnapStrictness

https://drafts.csswg.org/css-scroll-snap-1/#snap-strictness

Zoom

A specified value for the zoom property.

Functions

change_bits_for_longhand 🔒

change_bits_for_maybe_property 🔒

flexbox_enabled 🔒

grid_enabled 🔒

Type Aliases

ContainIntrinsicSize

A specified value for the contain-intrinsic-size property.

LineClamp

A specified value for the line-clamp property.

Perspective

A specified value for the perspective property.

VerticalAlign

A specified value for the vertical-align property.