Type Definition style::values::specified::calc::CalcNode

type CalcNode = GenericCalcNode<Leaf>;

A calc node representation for specified values.

Implementations

impl CalcNode

fn parse_one<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

Tries to parse a single element in the expression, that is, a <length>, <angle>, <time>, <percentage>, according to expected_unit.

May return a "complex" CalcNode, in the presence of a parenthesized expression, for example.

fn parse<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

Parse a top-level calc expression, with all nested sub-expressions.

This is in charge of parsing, for example, 2 + 3 * 100%.

fn parse_argument<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

fn parse_product<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
expected_unit: CalcUnit
) -> Result<Self, ParseError<'i>>

Parse a top-level calc expression, and all the products that may follow, and stop as soon as a non-product expression is found.

This should parse correctly:

2

2 * 2

2 * 2 + 2 (but will leave the + 2 unparsed).

fn into_length_or_percentage(
self,
clamping_mode: AllowedNumericType
) -> Result<CalcLengthPercentage, ()>

Tries to simplify this expression into a <length> or <percentage> value.

fn to_time(&self) -> Result<Time, ()>

Tries to simplify this expression into a <time> value.

fn to_angle(&self) -> Result<Angle, ()>

Tries to simplify this expression into an Angle value.

fn to_number(&self) -> Result<CSSFloat, ()>

Tries to simplify this expression into a <number> value.

fn to_percentage(&self) -> Result<CSSFloat, ()>

Tries to simplify this expression into a <percentage> value.

pub fn math_function<'i>(
name: &CowRcStr<'i>,
location: SourceLocation
) -> Result<MathFunction, ParseError<'i>>

Given a function name, and the location from where the token came from, return a mathematical function corresponding to that name or an error.

pub fn parse_integer<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<CSSInteger, ParseError<'i>>

Convenience parsing function for integers.

pub fn parse_length_or_percentage<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
clamping_mode: AllowedNumericType,
function: MathFunction
) -> Result<CalcLengthPercentage, ParseError<'i>>

Convenience parsing function for <length> | <percentage>.

pub fn parse_percentage<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<CSSFloat, ParseError<'i>>

Convenience parsing function for percentages.

pub fn parse_length<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
clamping_mode: AllowedNumericType,
function: MathFunction
) -> Result<CalcLengthPercentage, ParseError<'i>>

Convenience parsing function for <length>.

pub fn parse_number<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<CSSFloat, ParseError<'i>>

Convenience parsing function for <number>.

pub fn parse_angle<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<Angle, ParseError<'i>>

Convenience parsing function for <angle>.

pub fn parse_time<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<Time, ParseError<'i>>

Convenience parsing function for <time>.

pub fn parse_number_or_percentage<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<NumberOrPercentage, ParseError<'i>>

Convenience parsing function for <number> or <percentage>.

pub fn parse_angle_or_number<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
function: MathFunction
) -> Result<AngleOrNumber, ParseError<'i>>

Convenience parsing function for <number> or <angle>.