Enum style::values::specified::color::Color

pub enum Color {
    CurrentColor,
    Numeric {
        parsed: RGBA,
        authored: Option<Box<str>>,
    },
    Complex(ComputedColor),
}

Specified color value

Variants
CurrentColor

The 'currentColor' keyword

Numeric

A specific RGBA color

Fields of Numeric

parsed: RGBA

Parsed RGBA color

authored: Option<Box<str>>

Authored representation

Complex(ComputedColor)

A complex color value from computed value

Methods

impl Color

pub fn currentcolor() -> Color

Returns currentcolor value.

pub fn transparent() -> Color

Returns transparent value.

pub fn rgba(rgba: RGBA) -> Self

Returns a numeric RGBA color value.

pub fn parse_quirky<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>,
allow_quirks: AllowQuirks
) -> Result<Self, ParseError<'i>>

Parse a color, with quirks.

https://quirks.spec.whatwg.org/#the-hashless-hex-color-quirk

pub fn is_non_transparent(&self) -> bool

Returns false if the color is completely transparent, and true otherwise.

impl Color

pub fn to_computed_color(
&self,
_context: Option<&Context>
) -> Option<ComputedColor>

Converts this Color into a ComputedColor.

If context is None, and the specified color requires data from the context to resolve, then None is returned.