Module style::values::specified::font

Specified values for font properties

Macros

impl_variant_east_asian

impl_variant_ligatures

impl_variant_numeric

system_font_methods

Structs

FontSynthesis

Whether user agents are allowed to synthesize bold or oblique font faces when a font family lacks bold or italic faces

KeywordInfo

Additional information for keyword-derived font sizes.

MozScriptMinSize

Specifies the minimum font size allowed due to changes in scriptlevel. Ref: https://wiki.mozilla.org/MathML:mstyle

MozScriptSizeMultiplier

Specifies the multiplier to be used to adjust font size due to changes in scriptlevel.

VariantAlternatesList

List of Variant Alternates

VariantAlternatesParsingFlags

Flags of variant alternates in bit

VariantEastAsian

Vairants for east asian variant

VariantLigatures

Variants of ligatures

VariantNumeric

Vairants of numeric values

XLang

Internal property that reflects the lang attribute

XTextZoom

text-zoom. Enable if true, disable if false

Enums

AbsoluteFontWeight

An absolute font-weight value for a @font-face rule.

FontFamily

Specifies a prioritized list of font family names or generic family names.

FontFeatureSettings

Define initial settings that apply when the font defined by an @font-face rule is rendered.

FontLanguageOverride

Allows authors to explicitly specify the language system of the font, overriding the language system implied by the content language

FontSize

A specified font-size value

FontSizeAdjust

Preserve the readability of text when font fallback occurs

FontSizeKeyword

CSS font keywords

FontStretch

A value for the font-stretch property.

FontStretchKeyword

A keyword value for font-stretch.

FontStyle

The specified value of the font-style property.

FontVariantAlternates

Control over the selection of these alternate glyphs

FontVariantEastAsian

Allows control of glyph substitution and sizing in East Asian text.

FontVariantLigatures

Ligatures and contextual forms are ways of combining glyphs to produce more harmonized forms

FontVariantNumeric

Specifies control over numerical forms.

FontVariationSettings

Define initial settings that apply when the font defined by an @font-face rule is rendered.

FontWeight

A specified font-weight value.

MathDepth

A value for the math-depth property. https://mathml-refresh.github.io/mathml-core/#the-math-script-level-property

VariantAlternates

Set of variant alternates

Constants

DEFAULT_FONT_STYLE_OBLIQUE_ANGLE_DEGREES

The default angle for font-style: oblique.

DEFAULT_SCRIPT_MIN_SIZE_PT

DEFAULT_SCRIPT_SIZE_MULTIPLIER

FONT_MEDIUM_PX

The default font size.

FONT_STYLE_OBLIQUE_MAX_ANGLE_DEGREES

From https://drafts.csswg.org/css-fonts-4/#valdef-font-style-oblique-angle:

FONT_STYLE_OBLIQUE_MIN_ANGLE_DEGREES

The minimum angle value that font-style: oblique should compute to.

LARGER_FONT_SIZE_RATIO

This is the ratio applied for font-size: larger and smaller by both Firefox and Chrome

MAX_FONT_WEIGHT

The maximum font-weight value per:

MIN_FONT_WEIGHT

The minimum font-weight value per:

Functions

parse_one_feature_value

Type Definitions

SpecifiedFontFeatureSettings

This property provides low-level control over OpenType or TrueType font features.

SpecifiedFontStyle

The specified value of the font-style property, without the system font crap.

SpecifiedFontVariationSettings

This property provides low-level control over OpenType or TrueType font variations.