Module style::values::specified::font

Specified values for font properties

Structs

FontSynthesis
Whether user agents are allowed to synthesize bold or oblique font faces when a font family lacks bold or italic faces

MozScriptMinSize
Specifies the minimum font size allowed due to changes in scriptlevel. Ref: https://wiki.mozilla.org/MathML:mstyle

MozScriptSizeMultiplier
Specifies the multiplier to be used to adjust font size due to changes in scriptlevel.

VariantAlternatesList
List of Variant Alternates

VariantEastAsian
Vairants for east asian variant

VariantLigatures
Variants of ligatures

VariantNumeric
Vairants of numeric values

XLang
Internal property that reflects the lang attribute

XTextZoom
text-zoom. Enable if true, disable if false

Enums

FontFamily
Specifies a prioritized list of font family names or generic family names

FontFeatureSettings
Define initial settings that apply when the font defined by an @font-face rule is rendered.

FontLanguageOverride
Allows authors to explicitly specify the language system of the font, overriding the language system implied by the content language

FontSize
A specified font-size value

FontSizeAdjust
Preserve the readability of text when font fallback occurs

FontVariantAlternates
Control over the selection of these alternate glyphs

FontVariantEastAsian
Allows control of glyph substitution and sizing in East Asian text.

FontVariantLigatures
Ligatures and contextual forms are ways of combining glyphs to produce more harmonized forms

FontVariantNumeric
Specifies control over numerical forms.

FontWeight
A specified font-weight value

KeywordSize
CSS font keywords

MozScriptLevel
Changes the scriptlevel in effect for the children. Ref: https://wiki.mozilla.org/MathML:mstyle

VariantAlternates
Set of variant alternates

Constants

FONT_MEDIUM_PX
The default font size.

Type Definitions

FontVariantSettings
This property provides low-level control over OpenType or TrueType font variations.