Type Definition style::values::specified::image::Image

type Image = Image<Gradient, MozImageRect, SpecifiedImageUrl>;

Specified values for an image according to CSS-IMAGES. https://drafts.csswg.org/css-images/#image-values

Implementations

impl Image

pub fn for_cascade(url: ServoUrl) -> Self

Creates an already specified image value from an already resolved URL for insertion in the cascade.

pub fn parse_with_cors_anonymous<'i, 't>(
context: &ParserContext,
input: &mut Parser<'i, 't>
) -> Result<Image, ParseError<'i>>

Provides an alternate method for parsing that associates the URL with anonymous CORS headers.

FIXME(emilio): It'd be nicer for this to pass a CorsMode parameter to a shared function instead.