Module style::values::specified::length

Length values.

Reexports

pub use super::image::AngleOrCorner;

pub use super::image::ColorStop;

pub use super::image::EndingShape as GradientEndingShape;

pub use super::image::Gradient;

pub use super::image::GradientKind;

pub use super::image::HorizontalDirection;

pub use super::image::Image;

pub use super::image::LengthOrKeyword;

pub use super::image::LengthOrPercentageOrKeyword;

pub use super::image::SizeKeyword;

pub use super::image::VerticalDirection;

Structs

CalcLengthOrPercentage

CalcProductNode
A calc product expression node.

CalcSumNode
A calc sum expression node.

CharacterWidth
HTML5 "character width", as defined in HTML5  14.5.4.

Percentage
A percentage value.

Enums

CalcUnit

CalcValueNode
A value inside a Calc expression.

FontRelativeLength
A font relative length.

Length
An extension to NoCalcLength to parse calc expressions. This is commonly used for the <length> values.

LengthOrPercentage
A length or a percentage value.

LengthOrPercentageOrAuto
TODO(emilio): Do the Length and Percentage variants make any sense with CalcLengthOrPercentage?

LengthOrPercentageOrAutoOrContent
Either a <length> or a <percentage> or the auto keyword or the content keyword.

LengthOrPercentageOrNone
TODO(emilio): Do the Length and Percentage variants make any sense with CalcLengthOrPercentage?

NoCalcLength
A <length> without taking calc expressions into account

ViewportPercentageLength
A viewport-relative length.

Type Definitions

LengthOrAuto
Either a <length> or the auto keyword.

LengthOrNone
Either a <length> or the none keyword.

LengthOrNormal
Either a <length> or the normal keyword.

LengthOrNumber
Either a <length> or a <number>.