Enum style::values::specified::length::NoCalcLength

pub enum NoCalcLength {
    Absolute(AbsoluteLength),
    FontRelative(FontRelativeLength),
    ViewportPercentage(ViewportPercentageLength),
    ContainerRelative(ContainerRelativeLength),
    ServoCharacterWidth(CharacterWidth),
}

A <length> without taking calc expressions into account

https://drafts.csswg.org/css-values/#lengths

Variants

Absolute(AbsoluteLength)

An absolute length

https://drafts.csswg.org/css-values/#absolute-length

FontRelative(FontRelativeLength)

A font-relative length:

https://drafts.csswg.org/css-values/#font-relative-lengths

ViewportPercentage(ViewportPercentageLength)

A viewport-relative length.

https://drafts.csswg.org/css-values/#viewport-relative-lengths

ContainerRelative(ContainerRelativeLength)

A container query length.

https://drafts.csswg.org/css-contain-3/#container-lengths

ServoCharacterWidth(CharacterWidth)

HTML5 “character width”, as defined in HTML5  14.5.4.

This cannot be specified by the user directly and is only generated by Stylist::synthesize_rules_for_legacy_attributes().

Implementations

impl NoCalcLength

pub fn to_computed_value_with_base_size( &self, context: &Context<'_>, base_size: FontBaseSize, line_height_base: LineHeightBase, ) -> Length

Computes a length with a given font-relative base size.

impl NoCalcLength

pub fn unitless_value(&self) -> CSSFloat

Return the unitless, raw value.

fn unit(&self) -> &'static str

pub fn is_negative(&self) -> bool

Returns whether the value of this length without unit is less than zero.

pub fn is_zero(&self) -> bool

Returns whether the value of this length without unit is equal to zero.

pub fn is_infinite(&self) -> bool

Returns whether the value of this length without unit is infinite.

pub fn is_nan(&self) -> bool

Returns whether the value of this length without unit is NaN.

pub fn should_zoom_text(&self) -> bool

Whether text-only zoom should be applied to this length.

Generally, font-dependent/relative units don’t get text-only-zoomed, because the font they’re relative to should be zoomed already.

pub fn parse_dimension( context: &ParserContext<'_>, value: CSSFloat, unit: &str, ) -> Result<Self, ()>

Parse a given absolute or relative dimension.

pub(crate) fn try_op<O>(&self, other: &Self, op: O) -> Result<Self, ()>
where O: Fn(f32, f32) -> f32,

pub(crate) fn map(&self, op: impl FnMut(f32) -> f32) -> Self

pub fn to_computed_pixel_length_without_context(&self) -> Result<CSSFloat, ()>

Get a px value without context (so only absolute units can be handled).

pub fn from_px(px_value: CSSFloat) -> NoCalcLength

Get an absolute length from a px value.