Module style::values::specified::position

CSS handling for the specified value of positions

Structs

AnchorName

https://drafts.csswg.org/css-anchor-position-1/#propdef-anchor-name

DashedIdentAndOrTryTactic

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-try-fallbacks ||

GridAutoFlow

Controls how the auto-placement algorithm works specifying exactly how auto-placed items get flowed into the grid: [ row | column ] || dense https://drafts.csswg.org/css-grid-2/#grid-auto-flow-property

MasonryAutoFlow

Controls how the Masonry layout algorithm works specifying exactly how auto-placed items get flowed in the masonry axis.

NamedArea

Not associated with any particular grid item, but can be referenced from the grid-placement properties.

PositionArea

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-area

PositionTryFallbacks

https://drafts.csswg.org/css-anchor-position-1/#position-try-fallbacks

PositionTryFallbacksTryTactic

Changes for the automatically-generated position option. Note that this is order-dependent - e.g. flip-start flip-inline != flip-inline flip-start.

PositionVisibility

Specified keyword values for the position-visibility property.

TemplateAreas

https://drafts.csswg.org/css-grid/#named-grid-area

TemplateAreasArc

Arc type for Arc<TemplateAreas>

TemplateAreasParser

Parser for grid template areas.

TemplateAreasTokenizer 🔒

Tokenize the string into a list of the tokens, using longest-match semantics

UnsignedRange

A range of rows or columns. Using this instead of std::ops::Range for FFI purposes.

Enums

AnchorScope

https://drafts.csswg.org/css-anchor-position-1/#propdef-scope

GridTemplateAreas

This property specifies named grid areas.

HorizontalPositionKeyword

A keyword for the X direction.

MasonryItemOrder

Masonry auto-placement algorithm item sorting option.

MasonryPlacement

Masonry auto-placement algorithm packing.

PositionAnchor

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-anchor

PositionAreaKeyword

Possible values for the position-area preperty’s keywords. https://drafts.csswg.org/css-anchor-position-1/#propdef-position-area

PositionComponent

The specified value of a component of a CSS <position>.

PositionTryFallbacksItem

https://drafts.csswg.org/css-anchor-position-1/#propdef-position-try-fallbacks [ [ || ] | <‘position-area’> ]

PositionTryFallbacksTryTacticKeyword

How to swap values for the automatically-generated position tactic.

PositionTryOrder

https://drafts.csswg.org/css-anchor-position-1/#position-try-order-property

VerticalPositionKeyword

A keyword for the Y direction.

Traits

Side

Represents a side, either horizontal or vertical, of a CSS position.

Functions

is_compatible_pairing 🔒

is_item_order_definite_first 🔒

is_name_code_point 🔒

is_pack_with_non_default_order 🔒

Type Aliases

AnchorFunction

A specified value for anchor() function.

AspectRatio

A specified value for the aspect-ratio property.

HorizontalPosition

The specified value of a horizontal position.

Inset

A specified value for inset types.

Position

The specified value of a CSS <position>

PositionOrAuto

The specified value of an auto | <position>.

VerticalPosition

The specified value of a vertical position.

ZIndex

A specified value for the z-index property.