Module style::values::specified::position

CSS handling for the specified value of positions

Structs

GridAutoFlow

Controls how the auto-placement algorithm works specifying exactly how auto-placed items get flowed into the grid

MasonryAutoFlow

Controls how the Masonry layout algorithm works specifying exactly how auto-placed items get flowed in the masonry axis.

NamedArea

Not associated with any particular grid item, but can be referenced from the grid-placement properties.

TemplateAreas

https://drafts.csswg.org/css-grid/#named-grid-area

TemplateAreasArc

Arc type for Arc<TemplateAreas>

TemplateAreasTokenizer

Tokenize the string into a list of the tokens, using longest-match semantics

UnsignedRange

A range of rows or columns. Using this instead of std::ops::Range for FFI purposes.

Enums

GridTemplateAreas

This property specifies named grid areas.

HorizontalPositionKeyword

A keyword for the X direction.

MasonryItemOrder

Masonry auto-placement algorithm item sorting option.

MasonryPlacement

Masonry auto-placement algorithm packing.

PositionComponent

The specified value of a component of a CSS <position>.

VerticalPositionKeyword

A keyword for the Y direction.

Traits

Side

Represents a side, either horizontal or vertical, of a CSS position.

Functions

is_item_order_definite_first

is_name_code_point

is_pack_with_non_default_order

Type Definitions

AspectRatio

A specified value for the aspect-ratio property.

HorizontalPosition

The specified value of a horizontal position.

Position

The specified value of a CSS <position>

PositionOrAuto

The specified value of an auto | <position>.

VerticalPosition

The specified value of a vertical position.

ZIndex

A specified value for the z-index property.