style::values::specified::position

Enum PositionAreaKeyword

#[repr(u8)]
pub enum PositionAreaKeyword {

    None = 40,
    Center = 42,
    SpanAll = 47,
    Start = 33,
    End = 36,
    SpanStart = 35,
    SpanEnd = 38,
    Left = 1,
    Right = 4,
    Top = 9,
    Bottom = 12,
    XStart = 17,
    XEnd = 20,
    YStart = 25,
    YEnd = 28,
    BlockStart = 49,
    BlockEnd = 52,
    InlineStart = 57,
    InlineEnd = 60,
    SpanLeft = 3,
    SpanRight = 6,
    SpanTop = 11,
    SpanBottom = 14,
    SpanXStart = 19,
    SpanXEnd = 22,
    SpanYStart = 27,
    SpanYEnd = 30,
    SpanBlockStart = 51,
    SpanBlockEnd = 54,
    SpanInlineStart = 59,
    SpanInlineEnd = 62,
    SelfStart = 97,
    SelfEnd = 100,
    SpanSelfStart = 99,
    SpanSelfEnd = 102,
    SelfXStart = 81,
    SelfXEnd = 84,
    SelfYStart = 89,
    SelfYEnd = 92,
    SelfBlockStart = 113,
    SelfBlockEnd = 116,
    SelfInlineStart = 121,
    SelfInlineEnd = 124,
    SpanSelfXStart = 83,
    SpanSelfXEnd = 86,
    SpanSelfYStart = 91,
    SpanSelfYEnd = 94,
    SpanSelfBlockStart = 115,
    SpanSelfBlockEnd = 118,
    SpanSelfInlineStart = 123,
    SpanSelfInlineEnd = 126,

}

Possible values for the position-area property’s keywords. Represented by [0z xxx yyy], where z means “self wm resolution”, xxxx is the axis (as in PositionAreaAxis) and yyy is the PositionAreaTrack https://drafts.csswg.org/css-anchor-position-1/#propdef-position-area

Variants

None = 40

Center = 42

SpanAll = 47

Start = 33

End = 36

SpanStart = 35

SpanEnd = 38

Left = 1

Right = 4

Top = 9

Bottom = 12

XStart = 17

XEnd = 20

YStart = 25

YEnd = 28

BlockStart = 49

BlockEnd = 52

InlineStart = 57

InlineEnd = 60

SpanLeft = 3

SpanRight = 6

SpanTop = 11

SpanBottom = 14

SpanXStart = 19

SpanXEnd = 22

SpanYStart = 27

SpanYEnd = 30

SpanBlockStart = 51

SpanBlockEnd = 54

SpanInlineStart = 59

SpanInlineEnd = 62

SelfStart = 97

SelfEnd = 100

SpanSelfStart = 99

SpanSelfEnd = 102

SelfXStart = 81

SelfXEnd = 84

SelfYStart = 89

SelfYEnd = 92

SelfBlockStart = 113

SelfBlockEnd = 116

SelfInlineStart = 121

SelfInlineEnd = 124

SpanSelfXStart = 83

SpanSelfXEnd = 86

SpanSelfYStart = 91

SpanSelfYEnd = 94

SpanSelfBlockStart = 115

SpanSelfBlockEnd = 118

SpanSelfInlineStart = 123

SpanSelfInlineEnd = 126

Implementations

impl PositionAreaKeyword

pub fn parse<'i, 't>(input: &mut Parser<'i, 't>) -> Result<Self, ParseError<'i>>

Parse this keyword.

pub fn from_ident(ident: &str) -> Result<Self, ()>

Parse this keyword from a string slice.

impl PositionAreaKeyword

pub fn none() -> Self

Returns the ‘none’ value.

pub fn is_none(&self) -> bool

Returns true if this is the none keyword.

pub fn self_wm(self) -> bool

Whether we’re one of the self-wm keywords.

pub fn axis(self) -> PositionAreaAxis

Get this keyword’s axis.

pub fn with_axis(self, axis: PositionAreaAxis) -> Self

Returns this keyword but with the axis swapped by the argument.

pub fn with_inferred_axis(self, axis: PositionAreaAxis) -> Self

If this keyword uses an inferred axis, replaces it.

pub fn track(self) -> Option<PositionAreaTrack>

Get this keyword’s track, or None if we’re the None keyword.

fn group_type(self) -> PositionAreaType

fn to_physical( self, cb_wm: WritingMode, self_wm: WritingMode, inferred_axis: LogicalAxis, ) -> Self

fn flip_track(self) -> Self

pub fn to_self_alignment(self) -> Option<AlignFlags>

Returns a value for the self-alignment properties in order to resolve normal.

https://drafts.csswg.org/css-anchor-position/#position-area-alignment