Module style::values::specified::text

Specified types for text properties.

Structs

TextDecorationLine
Specified keyword values for the text-decoration-line property.

TextEmphasisPosition
Specified value of text-emphasis-position property.

TextOverflow
text-overflow. Specifies rendering when inline content overflows its line box edge.

TextTransform
Specified value of the text-transform property, stored in two parts: the case-related transforms (mutually exclusive, only one may be in effect), and others (non-exclusive).

TextTransformOther
Specified keyword values for non-case transforms in the text-transform property. (Non-exclusive.)

TextUnderlinePosition
Specified keyword values for the text-underline-position property. (Non-exclusive, but not all combinations are allowed: the spec grammar gives auto | [ from-font | under ] || [ left | right ].) https://drafts.csswg.org/css-text-decor-4/#text-underline-position-property

Enums

LineBreak
Values for the line-break property.

OverflowWrap
Values for the overflow-wrap property.

TextAlign
Specified value of text-align property.

TextAlignKeyword
Specified value of text-align keyword value.

TextAlignLast
Specified and computed value of text-align-last.

TextDecorationSkipInk
Implements text-decoration-skip-ink which takes the keywords auto | none | all

TextEmphasisFillMode
Fill mode for the text-emphasis-style property

TextEmphasisHorizontalWritingModeValue
The allowed horizontal values for the text-emphasis-position property.

TextEmphasisShapeKeyword
Shape keyword for the text-emphasis-style property

TextEmphasisStyle
Specified value of text-emphasis-style property.

TextEmphasisVerticalWritingModeValue
The allowed vertical values for the text-emphasis-position property.

TextOverflowSide
A generic value for the text-overflow property.

TextTransformCase
Specified keyword values for case transforms in the text-transform property. (These are exclusive.)

WordBreak
Values for the word-break property.

Constants

_IMPL_NUM_FromPrimitive_FOR_TextAlignKeyword

_IMPL_NUM_FromPrimitive_FOR_TextAlignLast

Functions

fill_mode_is_default_and_shape_exists

Type Definitions

InitialLetter
A specified type for the initial-letter property.

LetterSpacing
A specified value for the letter-spacing property.

LineHeight
A specified value for the line-height property.

TextDecorationLength
Implements type for text-decoration-thickness property

WordSpacing
A specified value for the word-spacing property.