style::values::specified

Module transform

Specified types for CSS values that are related to transformations.

Re-exports

pub use crate::values::generics::transform::TransformStyle;

Enums

AllowUnitlessPerspective

Whether to allow unitless values for perspective in prefixed transform properties.

OriginComponent

The specified value of a component of a CSS <transform-origin>.

TransformBox

The specified value of transform-box. https://drafts.csswg.org/css-transforms-1/#transform-box

Functions

all_transform_boxes_are_enabled ๐

Type Aliases

Rotate

A specified CSS rotate

Scale

A specified CSS scale

Transform

A specified CSS transform

TransformOperation

A single operation in a specified CSS transform

TransformOrigin

The specified value of a CSS <transform-origin>

Translate

A specified CSS translate