Type Alias style::values::specified::transform::TransformOperation

pub type TransformOperation = TransformOperation<Angle, Number, Length, Integer, LengthPercentage>;

A single operation in a specified CSS transform

Aliased Type

enum TransformOperation {

    Matrix(GenericMatrix<Number>),
    Matrix3D(GenericMatrix3D<Number>),
    Skew(Angle, Angle),
    SkewX(Angle),
    SkewY(Angle),
    Translate(LengthPercentage, LengthPercentage),
    TranslateX(LengthPercentage),
    TranslateY(LengthPercentage),
    TranslateZ(Length),
    Translate3D(LengthPercentage, LengthPercentage, Length),
    Scale(Number, Number),
    ScaleX(Number),
    ScaleY(Number),
    ScaleZ(Number),
    Scale3D(Number, Number, Number),
    Rotate(Angle),
    RotateX(Angle),
    RotateY(Angle),
    RotateZ(Angle),
    Rotate3D(Number, Number, Number, Angle),
    Perspective(GenericPerspectiveFunction<Length>),
    InterpolateMatrix {
        from_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>,
        to_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>,
        progress: Percentage,
    },
    AccumulateMatrix {
        from_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>,
        to_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>,
        count: Integer,
    },

}

Variants

Matrix(GenericMatrix<Number>)

Represents a 2D 2x3 matrix.

Matrix3D(GenericMatrix3D<Number>)

Represents a 3D 4x4 matrix.

Skew(Angle, Angle)

A 2D skew.

If the second angle is not provided it is assumed zero.

Syntax can be skew(angle) or skew(angle, angle)

SkewX(Angle)

skewX(angle)

SkewY(Angle)

skewY(angle)

Translate(LengthPercentage, LengthPercentage)

translate(x, y) or translate(x)

TranslateX(LengthPercentage)

translateX(x)

TranslateY(LengthPercentage)

translateY(y)

TranslateZ(Length)

translateZ(z)

Translate3D(LengthPercentage, LengthPercentage, Length)

translate3d(x, y, z)

Scale(Number, Number)

A 2D scaling factor.

Syntax can be scale(factor) or scale(factor, factor)

ScaleX(Number)

scaleX(factor)

ScaleY(Number)

scaleY(factor)

ScaleZ(Number)

scaleZ(factor)

Scale3D(Number, Number, Number)

scale3D(factorX, factorY, factorZ)

Rotate(Angle)

Describes a 2D Rotation.

In a 3D scene rotate(angle) is equivalent to rotateZ(angle).

RotateX(Angle)

Rotation in 3D space around the x-axis.

RotateY(Angle)

Rotation in 3D space around the y-axis.

RotateZ(Angle)

Rotation in 3D space around the z-axis.

Rotate3D(Number, Number, Number, Angle)

Rotation in 3D space.

Generalization of rotateX, rotateY and rotateZ.

Perspective(GenericPerspectiveFunction<Length>)

Specifies a perspective projection matrix.

Part of CSS Transform Module Level 2 and defined at  13.1. 3D Transform Function.

The value must be greater than or equal to zero.

InterpolateMatrix

A intermediate type for interpolation of mismatched transform lists.

Fields

from_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>

to_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>

progress: Percentage

AccumulateMatrix

A intermediate type for accumulation of mismatched transform lists.

Fields

from_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>

to_list: GenericTransform<GenericTransformOperation<Angle, Number, Length, Integer, LengthPercentage>>

count: Integer