Struct style::values::specified::url::SpecifiedUrl

pub struct SpecifiedUrl { /* fields omitted */ }

A specified url() value.

Methods

impl SpecifiedUrl

fn parse_from_string<'a>(url: Cow<'a, str>,
context: &ParserContext)
-> Result<Self, ()>

Try to parse a URL from a string value that is a valid CSS token for a URL.

Only returns Err for Gecko, in the case we can't construct a URLExtraData.

fn extra_data(&self) -> &UrlExtraData

Get this URL's extra data.

fn url(&self) -> Option<&ServoUrl>

Returns the resolved url if it was valid.

fn as_str(&self) -> &str

Return the resolved url as string, or the empty string if it's invalid.

TODO(emilio): Should we return the original one if needed?

fn as_slice_components(&self) -> Result<(*const u8, usize), (*const u8, usize)>

Little helper for Gecko's ffi.

fn for_cascade(url: ServoUrl, extra_data: UrlExtraData) -> Self

Creates an already specified url value from an already resolved URL for insertion in the cascade.